% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtree.R
\name{addtree}
\alias{addtree}
\title{Additive Tree training preocedure}
\usage{
addtree(
  x,
  y,
  catPredictors = NULL,
  depthLimit = 8,
  learning.rate = 1,
  gamma = 0,
  update = "exponential",
  min.update = ifelse(update == "polynomial", 10, 1000),
  weights = NULL,
  autoweights = FALSE,
  min.hessian = 0.01,
  min.membership = 0,
  steps.past.min.membership = 2,
  rpart.params = NULL,
  save.rpart = FALSE,
  verbose = TRUE,
  trace = 1
)
}
\arguments{
\item{x}{Matrix / Data frame of features}

\item{y}{Integer {-1, 1}: Vector of binary outcomes}

\item{catPredictors}{Optional: Logical vector indicating categorical features}

\item{depthLimit}{Integer: Maximum depth of tree to grow}

\item{learning.rate}{learning rate for the Newton Raphson step that updates
the function values of the node for \code{update = "exponential"}}

\item{gamma}{Float (0, 1): Acceleration factor}

\item{update}{Character: "Exponential" or "Polynomial"}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Trains a binary decision tree classifier using the Additive Tree
}
\details{
Outcome must be factor with two levels, the first level is the 'positive'
class

lambda = gamma/(1 - gamma)
}
\references{
Valdes Gilmer, Luna Jose, Eaton Eric, Ungar Lyle, Simone Charles
and Solberg Timothy. MediBoost: a Patient Stratification Tool for
Interpretable Decision Making in the Era of Precision Medicine.
}
\author{
E.D. Gennatas (R),
Gilmer Valdes, Jose Marcio Luna Castaneda (original MATLAB)
}
\keyword{internal}
