% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTreeLeaves.R
\name{addTreeLeavesRC}
\alias{addTreeLeavesRC}
\title{\pkg{rtemis internal}: Low-level Asymmetric Additive Tree procedure}
\usage{
addTreeLeavesRC(x, y, x.valid = NULL, y.valid = NULL, weights = NULL,
  max.leaves = 5, select.leaves.smooth = TRUE, gamma = 0.1,
  min.update = 10, alpha = 0, lambda = 0.01, lambda.seq = NULL,
  minobsinnode = 2, minobsinnode.lin = 10, learning.rate = 1,
  part.minsplit = 2, part.xval = 0, part.max.depth = 1,
  part.cp = 0, .rho = TRUE, rho.max = 1000, lin.type = c("glmnet",
  "cv.glmnet", "glm"), cv.glmnet.nfolds = 5,
  cv.glmnet.lambda = "lambda.min", loss.fn = if (is.factor(y))
  class.loss else msew, verbose = TRUE, plot.tune.error = FALSE,
  trace = 0, n.cores = future::availableCores())
}
\arguments{
\item{x}{Data frame}

\item{max.leaves}{Integer: Total number of terminal nodes to reach. 1 is a special case where no split is performed and a linear
model is trained. Otherwise, this should be an even number as each split introduces two children nodes.
Note: this is total N of nodes in the tree, with the root uncounted, not the number of terminal nodes.}

\item{loss.fn}{Function with arguments y, Fval, weights. Allows you to define a custom loss function.}
}
\description{
Train an Asymmetric Additive Tree for Classification & Regression
}
\details{
With \code{max.nodes = 0}, the model is logistic regression trained according to \code{lin.type}
(i.e. standard or regularized using glmnet, etc.)
Note that lambda is treated differently by \code{glmnet::glmnet} and \code{MASS::lm.ridge}
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
