% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{anyConstant}
\alias{anyConstant}
\title{Check for constant columns}
\usage{
anyConstant(x)
}
\arguments{
\item{x}{Input Data Frame}
}
\description{
Checks if any column of a data frame have zero variance
}
\author{
E.D. Gennatas
}
