% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{as.boost}
\alias{as.boost}
\title{\code{as.boost} Place model in \link{boost} structure}
\usage{
as.boost(
  object,
  x = NULL,
  y = NULL,
  x.valid = NULL,
  y.valid = NULL,
  learning.rate = 1,
  init = 0,
  apply.lr = TRUE,
  tolerance = 1e-05,
  tolerance.valid = 1e-05
)
}
\arguments{
\item{object}{rtMod model}

\item{x}{Data.frame, optional: if provided, use to calculate fitted values of new boost object}

\item{y}{Float, vector: Outcome}

\item{x.valid}{Data.frame; optional: Validation data}

\item{y.valid}{Float, vector; optional: Validation outcome}

\item{learning.rate}{Float: Learning rate for new boost object. Default = 1}

\item{init}{Float: Initial value for new boost object. Default = 0}

\item{apply.lr}{Logical: Only considered is \code{x = NULL}. If TRUE, new boost object's fitted values will
be object$fitted * learning.rate, otherwise object$fitted}

\item{tolerance}{Float: error tolerance for new boost object. See \link{boost}}

\item{tolerance.valid}{Float: error tolerance for validation set. See \link{boost}}
}
\description{
\code{as.boost} Place model in \link{boost} structure
}
\author{
E.D. Gennatas
}
