% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartLinBoostTV.R
\name{as.cartLinBoostTV}
\alias{as.cartLinBoostTV}
\title{Place model in \code{cartLinBoostTV} structure}
\usage{
as.cartLinBoostTV(
  object,
  x,
  y = NULL,
  x.valid = NULL,
  y.valid = NULL,
  learning.rate = 1,
  init = 0,
  apply.lr = TRUE
)
}
\arguments{
\item{object}{rtMod model}

\item{learning.rate}{Float: Learning rate for new boost object.}

\item{init}{Float: Initial value for new boost object.}

\item{apply.lr}{Logical: Only considered is \code{x = NULL}. If TRUE, new boost object's fitted values will
be object$fitted * learning.rate, otherwise object$fitted}
}
\description{
Place model in \code{cartLinBoostTV} structure
}
\author{
E.D. Gennatas
}
