% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTreeLeaves.R
\name{as.data.tree.addTreeLeavesRC}
\alias{as.data.tree.addTreeLeavesRC}
\title{Convert \link{addTreeLeavesRC} to \code{data.tree} object}
\usage{
as.data.tree.addTreeLeavesRC(object)
}
\arguments{
\item{object}{\link{addTreeLeavesRC} object}
}
\description{
Convert \link{addTreeLeavesRC} to \code{data.tree} object
}
