% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shytreegamleaves.R
\name{as.data.tree.shytreegamleaves}
\alias{as.data.tree.shytreegamleaves}
\title{Convert \link{shytreegamleaves} to \code{data.tree} object}
\usage{
as.data.tree.shytreegamleaves(object)
}
\arguments{
\item{object}{\link{shytreegamleaves} object}
}
\description{
Convert \link{shytreegamleaves} to \code{data.tree} object
}
