% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shytreeLeaves.R
\name{as.data.tree.shytreeLeavesRC}
\alias{as.data.tree.shytreeLeavesRC}
\title{Convert \link{shytreeLeavesRC} to \code{data.tree} object}
\usage{
as.data.tree.shytreeLeavesRC(object)
}
\arguments{
\item{object}{\link{shytreeLeavesRC} object}
}
\description{
Convert \link{shytreeLeavesRC} to \code{data.tree} object
}
