% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{as.hytboost2}
\alias{as.hytboost2}
\title{\code{as.hytboost} Place model in \link{hytboost} structure}
\usage{
as.hytboost2(
  object,
  x,
  y,
  learning.rate = 0.1,
  init.learning.rate = learning.rate,
  init = 0,
  apply.lr = TRUE
)
}
\description{
\code{as.hytboost} Place model in \link{hytboost} structure
}
\author{
Efstathios D. Gennatas
}
