% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboostnow.R
\name{as.hytboostnow}
\alias{as.hytboostnow}
\title{\code{as.hytboostnow} Place model in \link{hytboostnow} structure}
\usage{
as.hytboostnow(
  object,
  x,
  y,
  x.valid,
  y.valid,
  learning.rate = 0.1,
  init.learning.rate = learning.rate,
  init = 0,
  apply.lr = TRUE
)
}
\description{
\code{as.hytboostnow} Place model in \link{hytboostnow} structure
}
\author{
Efstathios D. Gennatas
}
