% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modError.R
\name{auc}
\alias{auc}
\title{Area under the Curve}
\usage{
auc(prob, labels, verbose = FALSE)
}
\arguments{
\item{prob}{Float, Vector: Probabilities or model scores (e.g. c(.32, .75, .63), etc)}

\item{labels}{True labels of outcomes (e.g. c(0, 1, 1))}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Get the Area under the ROC curve to assess classifier performance using \code{pROC}
}
\details{
Consider looking at Balanced Accuracy and F1 as well
}
