% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_pairs.R
\name{auc_pairs}
\alias{auc_pairs}
\title{Area under the Curve by pairwise concordance}
\usage{
auc_pairs(estimated.score, true.labels, verbose = TRUE)
}
\arguments{
\item{estimated.score}{Float, Vector: Probabilities or model scores
(e.g. c(.32, .75, .63), etc)}

\item{true.labels}{True labels of outcomes (e.g. c(0, 1, 1))}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Get the Area under the ROC curve to assess classifier performance using pairwise concordance
}
\details{
The first level of \code{true.labels} must be the positive class, and high numbers in
\code{estimated.score} should correspond to the positive class.
}
\examples{
\dontrun{
true.labels <- factor(c("a", "a", "a", "b", "b", "b", "b"))
estimated.score <- c(0.7, 0.55, 0.45, 0.25, 0.6, 0.7, 0.2)
auc_pairs(estimated.score, true.labels, verbose = TRUE)
}
}
