% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_pairs.R
\name{auPRc}
\alias{auPRc}
\title{Area under Precision-Recall curve by average precision}
\usage{
auPRc(estimated.score, true.labels, thresholds = NULL)
}
\description{
Area under Precision-Recall curve by average precision
}
