% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmDB.R
\name{available_draw}
\alias{available_draw}
\title{Available Draw Functions}
\usage{
available_draw(verbosity = 1L)
}
\value{
NULL, invisibly.
}
\description{
Print available draw functions for visualization.
}
\examples{
available_draw()
}
\author{
EDG
}
