% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{available_themes}
\alias{available_themes}
\title{Print available \pkg{rtemis} themes}
\usage{
available_themes()
}
\value{
Called for its side effect of printing available themes.
}
\description{
Print available \pkg{rtemis} themes
}
