% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modError.R
\name{bacc}
\alias{bacc}
\title{Balanced Accuracy}
\usage{
bacc(true, predicted, harmonize = FALSE, verbose = TRUE)
}
\arguments{
\item{true}{True labels}

\item{predicted}{Estimated labels}

\item{harmonize}{Logical: passed to \link{sensitivity} and \link{specificity}, which use \link{factorHarmonize}.
Default = FALSE}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Balanced Accuracy of a binary classifier
}
\details{
BAcc = .5 * (Sensitivity + Specificity)
}
