% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagLearn.R
\name{bagLearn}
\alias{bagLearn}
\title{\pkg{rtemis} internal: Bootstrap Aggregating}
\usage{
bagLearn(x, y, mod, resample.rtset = rtset.bag.resample(),
  weights = NULL, params = list(), verbose = TRUE,
  res.verbose = FALSE, n.cores = NULL)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{mod}{String: \pkg{rtemis} model. See \link{modSelect} gives available models}

\item{resample.rtset}{List: output of \link{rtset} (or a list of same structure)}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{params}{List of named elements, each is a single value}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{res.verbose}{Logical: Will be passed to each \code{mod}'s \code{verbose} argument}

\item{n.cores}{Integer: Number of cores to use}
}
\description{
Train an \pkg{rtemis} model on a set of resamples.
}
\details{
This is meant to be called using \link{bag} instead of directly
}
\author{
Efstathios D. Gennatas
}
