% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.HYTREE.R
\name{betas.hytree}
\alias{betas.hytree}
\title{Extract coefficients from Additive Tree leaves}
\usage{
betas.hytree(object, newdata, verbose = FALSE, trace = 0)
}
\arguments{
\item{object}{\code{hytree} object}

\item{newdata}{matrix/data.frame of features}

\item{verbose}{Logical: If TRUE, print output to console}

\item{trace}{Integer {0:2} Increase verbosity}
}
\description{
Extract coefficients from Additive Tree leaves
}
\author{
Efstathios D. Gennatas
}
