% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LIHAD.R
\name{betas.lihad}
\alias{betas.lihad}
\title{Extract coefficients from Additive Tree leaves}
\usage{
betas.lihad(object, newdata, verbose = FALSE, trace = 0)
}
\arguments{
\item{object}{\code{lihad} object}

\item{newdata}{matrix/data.frame of features}

\item{verbose}{Logical: If TRUE, print output to console}

\item{trace}{Integer {0:2} Increase verbosity}
}
\description{
Extract coefficients from Additive Tree leaves
}
\author{
E.D. Gennatas
}
