% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot2factor.R
\name{binmat2vec}
\alias{binmat2vec}
\title{Binary matrix times character vector}
\usage{
binmat2vec(x, labels = colnames(x))
}
\arguments{
\item{x}{A binary matrix or data.frame}

\item{labels}{Character vector length equal to \code{ncol(x)}}
}
\value{
a character vector
}
\description{
Binary matrix times character vector
}
