% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Resampling}
\usage{
bootstrap(x, n.resamples = 10, seed = NULL)
}
\arguments{
\item{x}{Input vector}

\item{seed}{Integer: If provided, set seed for reproducibility. Default = NULL}

\item{n.resample}{Integer: Number of resamples to make. Default = 10}
}
\description{
Bootstrap Resampling
}
\author{
Efstathios D. Gennatas
}
