% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_error.R
\name{brier_score}
\alias{brier_score}
\title{Brier Score}
\usage{
brier_score(true, estimated.prob)
}
\arguments{
\item{true}{Numeric vector, {0, 1}: True labels}

\item{estimated.prob}{Numeric vector, [0, 1]: Estimated probabilities}
}
\description{
Calculate the Brier Score for classification:
}
\details{
\deqn{BS = \frac{1}{N} \sum_{i=1}^{N} (y_i - p_i)^2}{BS = 1/N * sum_{i=1}^{N} (y_i - p_i)^2}
}
\author{
E.D. Gennatas
}
