% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_CMeans.R
\name{c_CMeans}
\alias{c_CMeans}
\title{Fuzzy C-means Clustering}
\usage{
c_CMeans(
  x,
  k = 2,
  iter.max = 100,
  dist = "euclidean",
  method = "cmeans",
  m = 2,
  rate.par = NULL,
  weights = 1,
  control = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer: Number of clusters to get. Default = 2}

\item{iter.max}{Integer: Maximum number of iterations. Default = 100}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'. Default = "euclidean"}

\item{method}{Character: "cmeans" - fuzzy c-means clustering; "ufcl": on-line update. Default = "cmeans"}

\item{m}{Float (>1): Degree of fuzzification. Default = 2}

\item{rate.par}{Float (0, 1): Learning rate for the online variant. (Default = .3)}

\item{weights}{Float (>0): Case weights}

\item{control}{List of control parameters. See \code{e1071::cmeans}}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{...}{Additional parameters to be passed to \code{e1071::cmeans}}
}
\value{
\code{rtClust} object
}
\description{
Perform \href{https://en.wikipedia.org/wiki/Fuzzy_clustering}{fuzzy C-means clustering} using \code{e1071::cmeans}
}
\seealso{
Other Clustering: 
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAM}()},
\code{\link{c_PAMK}()},
\code{\link{c_SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
