% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_DBSCAN.R
\name{c_DBSCAN}
\alias{c_DBSCAN}
\title{Density-based spatial clustering of applications with noise}
\usage{
c_DBSCAN(
  x,
  x.test = NULL,
  eps = 1,
  minPts = NCOL(x) + 1,
  weights = NULL,
  borderPoints = TRUE,
  search = c("kdtree", "linear", "dist"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{eps}{Numeric: Radius of the epsilon neighborhood}

\item{minPts}{Integer: Number of minimum points required in the eps
neighborhood for core points (including the point itself).}

\item{weights}{Numeric vector: Data points' weights. Needed for weighted
clustering.}

\item{borderPoints}{Logical: If TRUE, assign border points to clusters,
otherwise they are considered noise}

\item{search}{Character: "kdtree", "linear" or "dist": nearest neighbor
search strategy}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform \href{https://en.wikipedia.org/wiki/DBSCAN}{DBSCAN} clustering
}
\details{
See \code{dbscan::dbscan} for info on how to choose \code{eps} and
\code{minPts}
}
\seealso{
Other Clustering: 
\code{\link{c_CMeans}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAM}()},
\code{\link{c_PAMK}()},
\code{\link{c_SPEC}()}
}
\author{
Efstathios D. Gennatas
}
\concept{Clustering}
