% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_EMC.R
\name{c_EMC}
\alias{c_EMC}
\title{Expectation Maximization Clustering}
\usage{
c_EMC(
  x,
  x.test = NULL,
  k = 2,
  lab = NULL,
  EMC = EMCluster::.EMC,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{lab}{Vector, length \code{NROW(x)}: Labels for semi-supervised clustering}

\item{EMC}{List of control parameters for \code{EMCluster::emcluster}. Default = \code{EMCluster::.EMC}}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{EMCluster::emcluster}}
}
\description{
Perform clustering by
\href{https://en.wikipedia.org/wiki/Expectation\%E2\%80\%93maximization_algorithm}{EM}
using \code{EMCluster::emcluster}
}
\details{
First, \code{EMCluster::simple.init(x, nclass = k)} is run,
followed by \code{EMCluster::emcluster(x, emobj = emobj, assign.class = TRUE, ...)}

This can be very slow.
}
\seealso{
Other Clustering: 
\code{\link{c_CMeans}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAM}()},
\code{\link{c_PAMK}()},
\code{\link{c_SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
