% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_HOPACH.R
\name{c_HOPACH}
\alias{c_HOPACH}
\title{Hierarchical Ordered Partitioning and Collapsing Hybrid}
\usage{
c_HOPACH(
  x,
  dmat = NULL,
  metric = c("cosangle", "abscosangle", "euclid", "abseuclid", "cor", "abscor"),
  k = 15,
  kmax = 9,
  khigh = 9,
  trace = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{dmat}{Matrix (numeric, no missing values) or \code{hdist} object of pairwise distances.
If NULL, it is computed based on \code{metric}}

\item{metric}{Character: Dissimilarity metric to be used. Options: "cosangle", "abscosangle", "euclid",
"abseuclid", "cor", "abscor"}

\item{k}{Integer, (0:15]: Maximum number of levels}

\item{kmax}{Integer, [1:9]: Maximum number of children at each node in the tree}

\item{khigh}{Integer, [1:9]: Maximum number of children at each nod ein the tree when computing the
the Mean/Median Split Silhouette. Usually same as \code{kmax}}

\item{trace}{Integer: If trace > 0, print messages during HOPACH run. Default = 0}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{...}{Additional parameters to pass to \code{cluster::hopach}}
}
\description{
Perform
\href{https://www.sciencedirect.com/science/article/abs/pii/S0378375802003889}{HOPACH clustering}
using \code{hopach::hopach}
}
\seealso{
Other Clustering: 
\code{\link{c_CMeans}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HARDCL}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAM}()},
\code{\link{c_PAMK}()},
\code{\link{c_SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
