% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_KMEANS.R
\name{c_KMEANS}
\alias{c_KMEANS}
\title{K-means Clustering}
\usage{
c_KMEANS(x, x.test = NULL, k = 2, dist = "euclidean", verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform K-means clustering using \code{flexclust::cclust}
}
\seealso{
Other Clustering: 
\code{\link{c_CMEANS}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_MEANSHIFT}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAMK}()},
\code{\link{c_PAM}()},
\code{\link{c_SPEC}()},
\code{\link{u_H2OKMEANS}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
