% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_MEANSHIFT.R
\name{c_MEANSHIFT}
\alias{c_MEANSHIFT}
\title{Mean Shift Clustering}
\usage{
c_MEANSHIFT(
  x,
  nNeighbors = NROW(x),
  algorithm = c("LINEAR", "KDTREE"),
  kernelType = c("NORMAL", "EPANECHNIKOV", "BIWEIGHT"),
  bandwidth = rep(1, NCOL(x)),
  alpha = 0,
  iterations = 10,
  epsilon = 1e-08,
  epsilonCluster = 1e-04,
  parameters = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform Mean Shift clustering using \code{meanShiftR::meanShift}
}
\seealso{
Other Clustering: 
\code{\link{c_CMEANS}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMEANS}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMEANS}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAMK}()},
\code{\link{c_PAM}()},
\code{\link{c_SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
