% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_NGAS.R
\name{c_NGAS}
\alias{c_NGAS}
\title{Neural Gas Clustering}
\usage{
c_NGAS(x, x.test = NULL, k = 2, dist = "euclidean", verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\value{
\code{rtClust} object
}
\description{
Perform \href{https://en.wikipedia.org/wiki/Neural_gas}{Neural Gas clustering} using \code{flexclust::cclust}
}
\seealso{
Other Clustering: 
\code{\link{c_CMeans}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_PAM}()},
\code{\link{c_PAMK}()},
\code{\link{c_SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
