% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_PAM.R
\name{c_PAM}
\alias{c_PAM}
\title{Partitioning Around Medoids}
\usage{
c_PAM(
  x,
  k = 2,
  diss = FALSE,
  metric = "euclidean",
  do.swap = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{diss}{Logical: If TRUE, \code{x} should be a \code{dist} or dissimilarity matrix.
Otherwise, \code{x} should be a matrix of cases by features. Default = FALSE}

\item{metric}{Character: Dissimilarity metric to be used. Options: 'euclidean', 'manhattan'}

\item{do.swap}{Logical: If TRUE, perform the swap phase (See \code{cluster::pam}), as in the
original PAM algorithm. This is computationally intensive and can be skipped. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{cluster::pam}}
}
\description{
Perform
\href{https://en.wikipedia.org/wiki/K-medoids#Partitioning_Around_Medoids_(PAM)}{PAM clustering}
using \code{cluster::pam}
}
\seealso{
Other Clustering: 
\code{\link{c_CMeans}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAMK}()},
\code{\link{c_SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
