% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_SPEC.R
\name{c_SPEC}
\alias{c_SPEC}
\title{Spectral Clustering}
\usage{
c_SPEC(
  x,
  k = 2,
  kernel = "rbfdot",
  kpar = "automatic",
  nystrom.red = FALSE,
  nystrom.sample = dim(x)[1]/6,
  iterations = 200,
  mod.sample = 0.75,
  na.action = na.omit,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{kernel}{Character: Kernel to use: "rbfdot", "polydot", "vanilladot", tanhdot", "laplacedot", "besseldot",
"anovadot", "splinedot", "stringdot"}

\item{kpar}{String OR List: "automatic", "local" OR list with: sigma (for "rbfdor", "laplacedot");
degree, scale, offset (for "polydot");
scale, offset (for "tanhdot");
sigma, order, degree (for "besseldot");
sigma, degree (for "anovadot");
length, lambda, normalized (for "stringdot")}

\item{nystrom.red}{Logical: if TRUE, use nystrom method to calculate eigenvectors (Default = FALSE)}

\item{nystrom.sample}{Integer: Number of points to use for estimating the eigenvalues when \code{nystrom.red = TRUE}
Default = \code{dim(x)[1]/6}}

\item{iterations}{Integer: Number of iterations allowed}

\item{mod.sample}{Float (0, 1): Proportion of data to use when estimating sigma. Default = .75}

\item{na.action}{Function: Action to perform on NA (Default = \code{na.omit})}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform \href{https://en.wikipedia.org/wiki/Spectral_clustering}{Spectral Clustering}
using \code{kernlab::specc}
}
\seealso{
Other Clustering: 
\code{\link{c_CMeans}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HARDCL}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAM}()},
\code{\link{c_PAMK}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
