% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate.Classification}
\alias{calibrate.Classification}
\title{Calibrate Binary Classification Models}
\usage{
calibrate.Classification(
  x,
  predicted_probabilities,
  true_labels,
  algorithm = "isotonic",
  hyperparameters = NULL,
  verbosity = 1L,
  ...
)
}
\arguments{
\item{x}{Classification object.}

\item{predicted_probabilities}{Numeric vector: Predicted probabilities.}

\item{true_labels}{Factor: True class labels.}

\item{algorithm}{Character: Algorithm to use to train calibration model.}

\item{hyperparameters}{Hyperparameters object: Setup using one of \verb{setup_*} functions.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Not used}
}
\description{
The goal of calibration is to adjust the predicted probabilities of a binary classification
model so that they better reflect the true probabilities (i.e. empirical risk) of the positive
class.
}
\details{
Important: The calibration model's training data should be different from the classification
model's training data.
}
\author{
EDG
}
