% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate.ClassificationCV}
\alias{calibrate.ClassificationCV}
\title{Calibrate Cross-validated Binary Classification Models}
\arguments{
\item{x}{ClassificationCV object: The model will be returned with the calibration model and
calibration metrics added.}

\item{algorithm}{Character: Algorithm to use to train calibration model.}

\item{hyperparameters}{Hyperparameters object: Setup using one of \verb{setup_*} functions.}

\item{resampler_parameters}{ResamplerParameters}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Not used}
}
\description{
The goal of calibration is to adjust the predicted probabilities of a binary classification
model so that they better reflect the true probabilities (i.e. empirical risk) of the positive
class.
}
\author{
EDG
}
