% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_cv.R
\name{calibrate_cv}
\alias{calibrate_cv}
\title{Calibrate cross-validated model}
\usage{
calibrate_cv(
  mod,
  alg = "gam",
  learn.params = list(),
  which.repeat = 1,
  seed = NULL,
  verbosity = 1
)
}
\arguments{
\item{mod}{\code{rtModCV} object returned by \link{train_cv}}

\item{alg}{Character: "gam" or "glm", algorithm to use for calibration}

\item{learn.params}{List: List of parameters to pass to the learning algorithm}

\item{which.repeat}{Integer: Which repeat to use for calibration}

\item{seed}{Integer: seed for the resampling}

\item{verbosity}{Integer: 0: silent, > 0: print messages}
}
\value{
List: Calibrated models, test-set labels, test set performance metrics,
estimated probabilities (uncalibrated), calibrated probabilities,
}
\description{
Calibrate cross-validated model trained using \link{train_cv}
}
\details{
This is a work in progress to be potentially incorporated into \link{train_cv}
}
\author{
E.D. Gennatas
}
