% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_slide.R
\name{calibrate_slide}
\alias{calibrate_slide}
\title{Calibrate predicted probabilities using GAM}
\usage{
calibrate_slide(
  true_labels,
  est_prob,
  n_windows = NULL,
  pos_class_idi = 1,
  k = 9,
  verbose = TRUE
)
}
\arguments{
\item{true_labels}{Factor with true class labels}

\item{est_prob}{Numeric vector with predicted probabilities}

\item{n_windows}{Integer: Number of calibration windows}

\item{pos_class_idi}{Integer: Index of the positive class}

\item{k}{Integer: GAM degrees of freedom}

\item{verbose}{Logical: If TRUE, printe messages to the console}
}
\value{
List with mod: fitted GAM model to be used for calibration of new data
calibrated_prob: Numeric vector with calibrated probabilities
}
\description{
Calibrate predicted probabilities using a generalized additive model (GAM).
}
\details{
This is meant for experimentation.
}
\examples{
\dontrun{
data(segment_logistic, package = "probably")
# Plot the calibration curve of the original predictions
dplot3_calibration(
    true_labels = segment_logistic$Class,
    est_prob = segment_logistic$.pred_poor,
    n_windows = 10,
    pos_class_idi = 2
)

# Plot the calibration curve of the calibrated predictions
dplot3_calibration(
    true_labels = segment_logistic$Class,
    est_prob = calibrate(segment_logistic$Class, segment_logistic$.pred_poor),
    n_windows = 10,
    pos_class_idi = 2
)
}
}
\author{
EDG
}
