% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartLite.R
\name{cartLite}
\alias{cartLite}
\title{Bare bones decision tree derived from \code{rpart}}
\usage{
cartLite(
  x,
  y,
  weights = NULL,
  minsplit = 2,
  minbucket = 1,
  cp = 0,
  maxcompete = 0,
  usesurrogate = 2,
  surrogatestyle = 0,
  maxdepth = 3,
  xval = 0,
  save.fitted = FALSE,
  trace = 0,
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{minsplit}{\link{gS} Integer: Minimum number of cases that must belong in a
node before considering a split.}

\item{minbucket}{\link{gS} Integer: Minimum number of cases allowed in a child
node.}

\item{cp}{\link{gS} Float: Complexity threshold for allowing a split.}

\item{maxcompete}{Integer: The number of competitor splits saved in the
output}

\item{usesurrogate}{See \code{rpart::rpart.control}}

\item{surrogatestyle}{See \code{rpart::rpart.control}}

\item{maxdepth}{\link{gS} Integer: Maximum depth of tree.}

\item{xval}{Integer: Number of cross-validations}

\item{save.fitted}{Logical: If TRUE, save fitted values in output. Default = FALSE}
}
\description{
A super-stripped down decision tree for when space and performance are critical
}
\author{
E.D. Gennatas
}
\keyword{internal}
