% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catrange}
\alias{catrange}
\title{Print range of continuous variable}
\usage{
catrange(x, ddSci = TRUE, decimal_places = 1, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{ddSci}{Logical: If TRUE, use \link{ddSci} or range.}

\item{decimal_places}{Integer: Number of decimal place to use if \code{ddSci = TRUE}.}

\item{na.rm}{Logical: passed to \code{base::range}}
}
\value{
Called for its side effect of printing the range of \code{x}.
}
\description{
Print range of continuous variable
}
\author{
EDG
}
