% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_print.R
\name{catsize}
\alias{catsize}
\title{Print Size}
\usage{
catsize(x, name = NULL, verbosity = 1L, newline = TRUE)
}
\arguments{
\item{x}{R object (usually that inherits from matrix or data.frame)}

\item{name}{Character: Name of input object}

\item{verbosity}{Integer: Verbosity level.}

\item{newline}{Logical: If TRUE, end with new line character.}
}
\value{
vector of NROW, NCOL invisibly
}
\description{
Get \code{NCOL(x)} and \code{NROW{x}}
}
\examples{
\dontrun{
catsize(iris)
}
}
\author{
EDG
}
