% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Concatenate Vectors}
\usage{
cc(...)
}
\arguments{
\item{...}{Two or more vectors of any type}
}
\value{
Concatenation of \code{...}
}
\description{
Concatenate that maintains factors
}
\details{
A \code{c()} replacement that maintains factors as factors because it doesn't make sense not to.
If all inputs are factors, they are converted to character, concatenated and converted to factor again.
Otherwise, they are passed to \code{c()}
}
\author{
Efstathios D. Gennatas
}
