% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data2.R
\name{check_data2}
\alias{check_data2}
\title{Check Data v2}
\usage{
check_data2(
  x,
  name = NULL,
  recommend = TRUE,
  font.family = "'Inter', sans-serif",
  color = "#e7e7e7",
  background.color = "#121212",
  verbose = TRUE
)
}
\arguments{
\item{x}{data.frame, data.table or similar structure}

\item{name}{Character: Dataset name}

\item{recommend}{Logical: If TRUE, output recommendations at the end.}

\item{font.family}{Character: Font family to use.}

\item{color}{Character: Text color.}

\item{background.color}{Character: Background color.}

\item{verbose}{Logical: If TRUE, print output in HTML viewer.}

\item{class}{Character: CSS class to assign to div output.}
}
\description{
Check Data v2
}
\examples{
\dontrun{
n <- 1000
x <- rnormmat(n, 50, return.df = TRUE)
x$char1 <- sample(letters, n, TRUE)
x$char2 <- sample(letters, n, TRUE)
x$fct <- factor(sample(letters, n, TRUE))
x <- rbind(x, x[1, ])
x$const <- 99L
x[sample(nrow(x), 20), 3] <- NA
x[sample(nrow(x), 20), 10] <- NA
check_data2(x)
}

}
\author{
E.D. Gennatas
}
