% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST.R
\name{check_data_rest}
\alias{check_data_rest}
\title{Check Data for REST API}
\usage{
check_data_rest(x)
}
\arguments{
\item{x}{data.frame, data.table or similar structure.}

\item{...}{Additional arguments to pass to \code{check_data()}.}
}
\value{
\code{CheckData} object.
}
\description{
Check Data for REST API
}
\examples{
n <- 1000
x <- rnormmat(n, 50, return_df = TRUE)
x$char1 <- sample(letters, n, TRUE)
x$char2 <- sample(letters, n, TRUE)
x$fct <- factor(sample(letters, n, TRUE))
x <- rbind(x, x[1, ])
x$const <- 99L
x[sample(nrow(x), 20), 3] <- NA
x[sample(nrow(x), 20), 10] <- NA
x$fct[30:35] <- NA
check_data_rest(x)
}
\author{
EDG
}
