% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_dependencies}
\alias{check_dependencies}
\title{\pkg{rtemis} internal: Dependencies check}
\usage{
check_dependencies(..., verbosity = 0L)
}
\arguments{
\item{...}{List or vector of strings defining namespaces to be checked}

\item{verbosity}{Integer: Verbosity level.
Note: An error will always printed if dependencies are missing.
Setting this to FALSE stops it from printing
"Dependencies check passed".}
}
\description{
Checks if dependencies can be loaded; names missing dependencies if not.
}
\author{
EDG
}
\keyword{internal}
