% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_io.R
\name{check_files}
\alias{check_files}
\title{Check file(s) exist}
\usage{
check_files(paths, verbose = TRUE, pad = 0)
}
\arguments{
\item{paths}{Character vector of paths}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{pad}{Integer: Number of spaces to pad to the left}
}
\description{
Check file(s) exist
}
\author{
E.D. Gennatas
}
