% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData_html}
\alias{checkData_html}
\title{Check Data (HTML output)}
\usage{
checkData_html(
  x,
  name = NULL,
  str = FALSE,
  recommend = TRUE,
  reportCases.thres = NULL,
  reportFeatures.thres = NULL,
  verbose = TRUE
)
}
\description{
Check Data (HTML output)
}
