% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData_live.R
\name{checkData_live}
\alias{checkData_live}
\title{Check Data (HTML output)}
\usage{
checkData_live(
  x,
  name = NULL,
  recommend = TRUE,
  font.family = "Fira Sans",
  color = "#e7e7e7",
  background.color = "#121212",
  class = "checkData",
  verbose = TRUE
)
}
\arguments{
\item{name}{Character: Dataset name}

\item{recommend}{Logical: If TRUE, output recommendations at the end. 
Default = TRUE}

\item{font.family}{Character: Font family to use. Default = "Fira Sans"}

\item{color}{Character: Text color. Default = "#e7e7e7"}

\item{background.color}{Character: Background color. Default = "transparent"}

\item{class}{Character: CSS class to assign to div output. 
Default = "checkData"}

\item{verbose}{Logical: If TRUE, print output in HTML viewer.
Default = TRUE}
}
\description{
Check Data (HTML output)
}
\examples{
\dontrun{
checkData_live(iris)
}
}
\author{
E.D. Gennatas
}
