% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_earlystop.R
\name{checkpoint_earlystop}
\alias{checkpoint_earlystop}
\title{Early stopping check}
\usage{
checkpoint_earlystop(
  x,
  absolute.threshold = NA,
  relative.threshold = NA,
  minimize = TRUE,
  relativeVariance.threshold = NA,
  n.steps = 10,
  combine.relative.thresholds = "AND",
  min.steps = 50,
  na.response = c("stop", "continue"),
  verbose = TRUE
)
}
\arguments{
\item{x}{Float, vector: Input - this would normally be the loss at each iteration}

\item{absolute.threshold}{Float: If set and the last value of \code{x} is less than or equal to this
(if \code{minimize = TRUE}) or greater than or equal to this (if \code{minimize = FALSE}), then return
\code{stop = TRUE}. See output under Value. Default = NA}

\item{relative.threshold}{Float: If set, checks if the relative change from the first to last value of \code{x}
exceeds this number. i.e. if set to .9 and \code{minimize = TRUE}, if there is a 90 percent drop from \code{x[1]} to
\code{x[length(x)]}, then the function returns \code{stop = TRUE}. If \code{minimize = FALSE}, then checks if there
is a 90 percent increase, accordingly.}

\item{minimize}{Logical: See \code{absolute.threshold}. Default = TRUE}

\item{relativeVariance.threshold}{Float: If relative variance over last \code{n.steps} is less than or equal to this,
return \code{stop = TRUE}. See output under Value}

\item{n.steps}{Integer; > 1: Calculate relative variance over this many last values of \code{x}}

\item{combine.relative.thresholds}{Character: "AND" or "OR": How to combine the criteria \code{relative.threshold} and
\code{relativeVariance.threshold}. Default = "AND", which means both must be TRUE to stop. The scenario is you might
want to check relastiveVariance threshold only after a certain amount of learning has taken place, which you
can't predict with \code{min.steps} but would rather quantify with \code{relative.threshold}.}

\item{min.steps}{Integer: Do not calculate relativeVariance unless \code{x} is at least this length}

\item{na.response}{Character: "stop" or "continue": what should happen if the last value of \code{x} is \code{NA}}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\value{
List with the following items:
\itemize{
\item \code{last.value} Float: Last value of \code{x}
\item \code{relativeVariance} Float: relative variance of last \code{n.steps}
\item \code{check.thresh} Logical: TRUE, if absolute threshold was reached
\item \code{check.rvar }Logical: TRUE, if relative variance threshold was reached
\item \code{stop} Logical: TRUE, if either criterion was met - absolute threshold or relativeVariance.threshold
}
}
\description{
Returns list with relative variance over n.steps, absolute.threshold, last value,
and logical "stop", if conditions are met and training should stop.
The final stop decision is:
\code{check.thresh | (check.rthresh & check.rvar)} if \code{combine.relative.thresholds = "AND"}
or
\code{check.thresh | (check.rthresh | check.rvar)} if \code{combine.relative.thresholds = "OR"}
}
\author{
E.D. Gennatas
}
