% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_harmonograph.R
\name{chill}
\alias{chill}
\title{Chill}
\usage{
chill(sleep = 0.5, text = NULL, max = 1000)
}
\arguments{
\item{sleep}{Float: Time in seconds between drawings. Default = .5}

\item{text}{Character: Text to display}

\item{max}{Integer: Max times to repeat. Default = 1000}
}
\description{
Relax. Use Ctrl-C to exit (but try to stay relaxed)
}
