% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.harmonograph.R
\name{chillalittle}
\alias{chillalittle}
\title{Chill a little}
\usage{
chillalittle(sleep = 0.3)
}
\arguments{
\item{sleep}{Float: Time in seconds between drawings. Default = .3}
}
\description{
Relax. Use Ctrl-C to exit (but stay relaxed)
}
