% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{choose_theme}
\alias{choose_theme}
\title{Select an rtemis theme}
\usage{
choose_theme(
  x = c("white", "whitegrid", "whiteigrid", "black", "blackgrid", "blackigrid",
    "darkgray", "darkgraygrid", "darkgrayigrid", "lightgraygrid", "mediumgraygrid"),
  parameters = NULL
)
}
\arguments{
\item{x}{Character: Name of theme to select. If not defined, will use \code{getOption("rtemis_theme", "whitegrid")}.}

\item{parameters}{List: Optional list of parameters to override the defaults.}
}
\value{
Theme object.
}
\description{
Select an rtemis theme
}
\author{
EDG
}
