% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_error.R
\name{class_error}
\alias{class_error}
\title{Classification Error}
\usage{
class_error(
  true,
  estimated,
  estimated.prob = NULL,
  calc.auc = TRUE,
  calc.brier = TRUE,
  auc.method = c("pROC", "ROCR", "auc_pairs"),
  trace = 0
)
}
\arguments{
\item{true}{Factor: True labels}

\item{estimated}{Factor: Estimated values}

\item{estimated.prob}{Numeric vector: Estimated probabilities}

\item{calc.auc}{Logical: If TRUE, calculate AUC. May be slow in very large datasets.}

\item{calc.brier}{Logical: If TRUE, calculate Brier Score}

\item{auc.method}{Character: "pROC", "ROCR", "auc_pairs": Method to use, passed to
\link{auc}.}

\item{trace}{Integer: If > 0, print  diagnostic messages. Default = 0}
}
\value{
S3 object of type "class_error"
}
\description{
Calculates Classification Metrics
}
\details{
Note that auc.method = "pROC" is the only one that will output an AUC even if
one or more estimated probabilities are NA.
}
\examples{
\dontrun{
true <- factor(c("a", "a", "a", "b", "b", "b", "b", "b", "b", "b"))
estimated <- factor(c("a", "a", "b", "b", "a", "a", "b", "b", "a", "a"))
estimated.prob <- c(0.7, 0.55, 0.45, 0.25, 0.6, 0.7, 0.2, .37, .57, .61)

class_error(true, estimated, estimated.prob, auc.method = "pROC")
class_error(true, estimated, estimated.prob, auc.method = "ROCR")
class_error(true, estimated, estimated.prob, auc.method = "auc_pairs")
}
}
\author{
E.D. Gennatas
}
