% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_supervised.R
\name{class_imbalance}
\alias{class_imbalance}
\title{Class Imbalance}
\usage{
class_imbalance(x)
}
\arguments{
\item{x}{Vector, factor: Outcome.}
}
\value{
Numeric.
}
\description{
Calculate class imbalance as given by:
\deqn{I = K\cdot\sum_{i=1}^K (n_i/N - 1/K)^2}{I = K * sum(n_i/N - 1/K)^2}
where \eqn{K} is the number of classes, and \eqn{n_i} is the number of
instances of class \eqn{i}
}
\author{
EDG
}
