% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Clean column names}
\usage{
clean_colnames(x)
}
\arguments{
\item{x}{Character vector or matrix with colnames or any object with \code{names()} method.}
}
\value{
Character vector.
}
\description{
Clean column names by replacing all spaces and punctuation with a single underscore
}
\examples{
\dontrun{
clean_colnames(iris)
}
}
\author{
EDG
}
