% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strng.R
\name{clean_factor_levels}
\alias{clean_factor_levels}
\title{Clean factor levels of data.frame}
\usage{
clean_factor_levels(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
Finds all factors in a data.frame and cleans factor levels to include
only underscore symbols
}
\author{
E.D. Gennatas
}
