% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{clust}
\alias{clust}
\title{Clustering with \pkg{rtemis}}
\usage{
clust(x, clust = "kmeans", x.test = NULL, verbose = TRUE, ...)
}
\arguments{
\item{x}{Numeric matrix / data frame: Input data}

\item{clust}{Character: Decomposition algorithm name, e.g. "nmf"
(case-insensitive)}

\item{x.test}{Numeric matrix / Data frame: Testing set data if supported by
\code{clust}}

\item{verbose}{Logical: if TRUE, print messages to screen}

\item{...}{Additional arguments to be passed to clusterer \code{clust}}
}
\value{
\code{rtClust} object
}
\description{
Convenience function to perform any \pkg{rtemis} clustering
}
\author{
E.D. Gennatas
}
