% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.ADDT.R
\name{coef.addTree}
\alias{coef.addTree}
\title{Extract coefficients from Additive Tree leaves}
\usage{
\method{coef}{addTree}(object, newdata, verbose = FALSE, trace = 0,
  ...)
}
\arguments{
\item{object}{\code{addTree} object}

\item{newdata}{matrix/data.frame of features}

\item{verbose}{Logical: If TRUE, print output to console}

\item{trace}{Integer {0:2} Increase verbosity}

\item{...}{Not used}
}
\description{
Extract coefficients from Additive Tree leaves
}
\author{
Efstathios D. Gennatas
}
