% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{col2grayscale}
\alias{col2grayscale}
\title{Color to Grayscale}
\usage{
col2grayscale(x, what = c("color", "decimal"))
}
\arguments{
\item{x}{Color to convert to grayscale}

\item{what}{Character: "color" returns a hexadecimal color,
"decimal" returns a decimal between 0 and 1}
}
\value{
Character: color hex code.
}
\description{
Convert a color to grayscale
}
\details{
Uses the NTSC grayscale conversion:
0.299 * R + 0.587 * G + 0.114 * B
}
\examples{
\dontrun{
col2grayscale("red")
col2grayscale("red", "dec")
}
}
\author{
EDG
}
