% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtPalette.R
\name{col2hex}
\alias{col2hex}
\title{Convert R color to hexadecimal code}
\usage{
col2hex(color)
}
\arguments{
\item{color}{Color(s) that R understands}
}
\description{
Convert any color that R understands into the corresponding hexadecimal code
}
\examples{
col2hex(c("gray50", "skyblue"))
}
\author{
Efstathios D. Gennatas
}
