% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{colMax}
\alias{colMax}
\title{Collapse data.frame to vector by getting column max}
\usage{
colMax(x, na.rm = TRUE)
}
\arguments{
\item{x}{Matrix or Data frame input}

\item{na.rm}{Logical: passed to \code{max}, If TRUE, ignore NA values,
otherwise if NA is present in any column, NA will be returned.}
}
\description{
Collapse data.frame to vector by getting column max
}
\author{
E.D. Gennatas
}
