% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{color_fade}
\alias{color_fade}
\title{Fade color to black or white}
\usage{
color_fade(x, add = -0.2, mul = 1)
}
\arguments{
\item{x}{Color input}

\item{add}{Numeric [0, 1]: Add this number to each RGB channel}

\item{mul}{Numeric: Multiply each RGB channel by this number}
}
\value{
Color in hex notation
}
\description{
Fade color to black or white
}
\author{
E.D. Gennatas
}
