% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{color_mean}
\alias{color_mean}
\title{Average colors}
\usage{
color_mean(x, space = c("RGB", "HSV"))
}
\arguments{
\item{x}{Color vector}

\item{space}{Character: RGB  or HSV; space to average in}
}
\description{
Average colors
}
\examples{
\dontrun{
color_mean(c("red", "blue")) |> previewcolor()
color_mean(c("red", "blue"), "HSV") |> previewcolor()
}
}
\author{
E.D. Gennatas
}
