% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{color_order}
\alias{color_order}
\title{Order colors}
\usage{
color_order(x, start_with = 1, order_by = c("similarity", "dissimilarity"))
}
\arguments{
\item{x}{Vector of colors}

\item{start_with}{Integer: Which color to output in first position}

\item{order_by}{Character: "similarity" or "dissimilarity"}
}
\description{
Order colors by RGB distance
}
\author{
E.D. Gennatas
}
