% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{color_separate}
\alias{color_separate}
\title{Separate colors}
\usage{
color_separate(x, start_with = 1)
}
\arguments{
\item{x}{Vector of colors}

\item{start_with}{Integer: Which color to output in first position}
}
\description{
Separate colors by RGB distance
}
\details{
Starting with the first color defined by \code{start_with}, the next color
is chosen to be max distance from all preceding colors
}
\author{
E.D. Gennatas
}
