% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{color_sqdist}
\alias{color_sqdist}
\title{Squared Color Distance}
\usage{
color_sqdist(x, y)
}
\arguments{
\item{x}{Color}

\item{y}{Color}
}
\description{
Get the squared RGB distance between two colors
}
\examples{
color_sqdist("red", "green")
color_sqdist("#16A0AC", "#FA6E1E")

}
\author{
E.D. Gennatas
}
