% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorAdjust.R
\name{colorAdjust}
\alias{colorAdjust}
\title{Adjust HSV Color}
\usage{
colorAdjust(color, alpha = NULL, hue = 0, sat = 0, val = 0)
}
\arguments{
\item{color}{Input color. Any format that grDevices::col2rgb() recognizes}

\item{alpha}{Numeric: Scale alpha by this amount. Future: replace with absolute setting}

\item{hue}{Float: How much hue to add to \code{color}}

\item{sat}{Float: How much saturation to add to \code{color}}

\item{val}{Float: How much to increase value of \code{color} by}
}
\value{
Adjusted color
}
\description{
Modify alpha, hue, saturation and value (HSV) of a color
}
\author{
E.D. Gennatas
}
