% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorGrad.x.R
\name{colorGrad.x}
\alias{colorGrad.x}
\title{Color gradient for continuous variable}
\usage{
colorGrad.x(x, color = c("gray20", "#18A3AC"), space = "Lab")
}
\arguments{
\item{x}{Float, vector}

\item{color}{Color, vector, length 2}

\item{space}{Character: "rgb" or "Lab", Default = "Lab"}
}
\description{
Color gradient for continuous variable
}
\author{
E.D. Gennatas
}
