% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{colorgrad_x}
\alias{colorgrad_x}
\title{Color gradient for continuous variable}
\usage{
colorgrad_x(x, color = c("gray20", "#18A3AC"), space = "Lab")
}
\arguments{
\item{x}{Float, vector}

\item{color}{Color, vector, length 2}

\item{space}{Character: "rgb" or "Lab".}
}
\description{
Color gradient for continuous variable
}
\author{
EDG
}
