% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{colorgradient.x}
\alias{colorgradient.x}
\title{Color gradient for continuous variable}
\usage{
colorgradient.x(
  x,
  symmetric = FALSE,
  lo_col = "#0290EE",
  mid_col = "#1A1A1A",
  hi_col = "#FFBD4F",
  space = "Lab"
)
}
\arguments{
\item{x}{Float, vector}

\item{symmetric}{Logical: If TRUE, make symmetric gradient between
\code{-max(abs(x))} and \code{max(abs(x))}}

\item{lo_col}{Low color}

\item{mid_col}{Middle color}

\item{hi_col}{High color}

\item{space}{Character: "rgb" or "Lab".}
}
\description{
Color gradient for continuous variable
}
\examples{
\dontrun{
x <- seq(-10, 10, length.out = 51)
previewcolor(colorgradient.x(x))
x <- sort(rnorm(40))
previewcolor(colorgradient.x(x, mid_col = "white"))
# Notice how most values are near zero therefore almost white
}
}
\author{
EDG
}
