% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOp.R
\name{colorOp}
\alias{colorOp}
\title{Simple Color Operations}
\usage{
colorOp(col, fn = c("invert", "mean"), space = c("HSV", "RGB"))
}
\arguments{
\item{col}{Input color(s)}

\item{fn}{String. Function to perform. Options: "invert", "mean"}

\item{space}{Colorspace to operate in. Options: "HSV", "RGB"}
}
\value{
Color
}
\description{
Invert a color or calculate the mean of two colors in HSV or RGB space.
This may be useful in creating colors for plots
}
\details{
The average of two colors in RGB space will often pass through gray,
which is likely undesirable. Averaging in HSV space, better for most applications.
}
\author{
Efstathios D. Gennatas
}
