% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPreview.R
\name{colorPreview}
\alias{colorPreview}
\title{Color Preview}
\usage{
colorPreview(col, main = NULL, names = TRUE, names.y = 1.045,
  srt = 45, alpha = 1, ylim = c(0.95, 1.1), cex = 12, pch = 18,
  bg = "white", mar = c(0, 1.7, 1, 2.8), main.col = "black",
  main.line = 0, filename = NULL, par.reset = TRUE,
  pdf.width = length(col) * 0.5, pdf.height = 2)
}
\arguments{
\item{col}{Input color. Hexadecimal or any string that R recognizes.
(See all with \code{colors()})}

\item{main}{Character: Plot title. Default = NULL}

\item{names}{Logical: If TRUE, add l\code{names(col)} if not NULL. Default = TRUE}

\item{names.y}{Float: y position to place \code{names}. Adjust as needed. Default = 1.045}

\item{srt}{Float: String rotation in degrees. Default = 45}

\item{alpha}{Float: Alpha level. Default = 1}

\item{ylim}{Float, vector, length 2: y axis limits. Default = c(0.95, 1.1)}

\item{cex}{Float: Character expansion factor. Default = 26}

\item{pch}{Integer or single character: The \code{pch} parameter of \code{par}}

\item{bg}{Background color}

\item{mar}{Float, vector, length 4: Plot margins. Default = c(0, 1.7, 0.6, 2.8)}

\item{main.line}{Float: Line to plot \code{main} Default = 0}

\item{filename}{String, optional: Path to file to save plot}

\item{par.reset}{Logical: If TRUE, reset \code{par} before exiting}

\item{pdf.width}{Float: PDF output width}

\item{pdf.height}{Float: PDF output height}
}
\description{
Very simple preview of colors
}
\author{
Efstathios D. Gennatas
}
