% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols2list.R
\name{cols2list}
\alias{cols2list}
\title{Convert data frame columns to list elements}
\usage{
cols2list(x)
}
\arguments{
\item{x}{Input: Will be coerced to data.frame, then each column will become an element of a list}
}
\description{
Convenience function to create a list out of data frame columns
}
\author{
E.D. Gennatas
}
