% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config}
\alias{create_config}
\title{Create rtemis configuration file}
\usage{
create_config(
  data_path,
  target = NULL,
  binclass_posid = 1,
  alg = "lightgbm",
  train.params = NULL,
  inner.resampling = setup.resample(resampler = "cv", n.resamples = 5),
  outer.resampling = setup.resample(resampler = "cv", n.resamples = 10),
  config.path = "rtemis-config.json",
  model.outdir = NULL,
  allow.overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data_path}{Character: Path to data file. Can be any file recognized by \link{read}, commonly CSV,
Excel, or RDS.}

\item{target}{Character: Name of the target variable in the data. If not specified, the last
column of \code{data} will be used.}

\item{binclass_posid}{Integer: Position of the positive class in a binary classification problem.}

\item{alg}{Character: Algorithm to use. Any of \code{select_learn()}.}

\item{train.params}{List: Parameters for the training algorithm.}

\item{inner.resampling}{List: Resampling method for the inner loop, i.e. hyperparameter tuning,
a.k.a. model selection. Set using \link{setup.resample}}

\item{outer.resampling}{List: Resampling method for the outer loop, i.e. testing. Set using
\link{setup.resample}}

\item{config.path}{Character: Path to save configuration file.}

\item{model.outdir}{Character: Directory to save trained model and associated files. If NULL,
the directory of \code{config.path} will be used.}

\item{allow.overwrite}{Logical: If TRUE, overwrite existing config file.}

\item{verbose}{Logical: If TRUE, print messages}
}
\value{
config as list, invisibly.
}
\description{
Defines a complete predictive modeling pipeline and saves it as a JSON file.
}
\author{
EDG
}
