% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.CUR.R
\name{d.CUR}
\alias{d.CUR}
\title{CUR Decomposition}
\usage{
d.CUR(x, c = dim(x)[2], r = dim(x)[1], k = NULL, sv = NULL,
  method = "highest.ranks", matrix.return = TRUE,
  error.return = FALSE, scale = TRUE, center = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{x}{Numeric matrix / data.frame: Input data}

\item{c}{Integer: Number of columns to be selected}

\item{r}{Integer: Number of rows to be selected}

\item{k}{Integer: Rank of decomposition (Creates \code{k} components)}

\item{sv}{The SVD of \code{x}, if already available}

\item{method}{String: "random", "top.scores", "ortho.top.scores",
"exact.num.random", "highest.ranks" (Default). See \code{rCUR::CUR("method")}}

\item{matrix.return}{Logical: if TRUE, matrices C, U, and R are returned. If FALSE,
U is not computed, which can be expensive, if r and c are large. Default = TRUE}

\item{error.return}{Logical: if TRUE, the Frobenius norm of the difference between the original matrix and the
CUR approximation is returned. Effective only if \code{matrix.return = TRUE}. Default = FALSE}

\item{scale}{Logical: If TRUE, scale input}

\item{center}{Logical: If TRUE, center input}

\item{verbose}{Logical: If TRUE, print messages to output}

\item{...}{Additional parameters to be passed to \code{rCUR::CUR}}
}
\value{
\link{rtDecom} object
}
\description{
Performs a CUR decomposition
}
\details{
Note that \code{k} here does not correspond with \code{k} in the other decomposition functions.
Use \code{c} to determine dimensionality of resulting decomposition
}
\seealso{
Other Decomposition: \code{\link{d.H2OAE}},
  \code{\link{d.H2OGLRM}}, \code{\link{d.ICA}},
  \code{\link{d.ISOMAP}}, \code{\link{d.KPCA}},
  \code{\link{d.LLE}}, \code{\link{d.MDS}},
  \code{\link{d.NMF}}, \code{\link{d.PCA}},
  \code{\link{d.SPCA}}, \code{\link{d.SVD}},
  \code{\link{d.TSNE}}, \code{\link{d.UMAP}}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
