% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.ICA.R
\name{d.ICA}
\alias{d.ICA}
\title{Independent Component Analysis}
\usage{
d.ICA(x, k = 3, package = c("fastICA", "ica"), alg.type = "parallel",
  maxit = 100, scale = TRUE, center = TRUE, verbose = TRUE,
  trace = 0, ...)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{package}{String: Which package to use for ICA. "fastICA" will use \code{fastICA::fastICA},
"ica" will use \code{ica::fastica}. Default = "fastICA".
Note: only \code{fastICA} works with \code{k = 1}}

\item{alg.type}{String: For \code{package = "fastICA"}, "parallel" or "deflation". Default = "parallel"}

\item{maxit}{Integer: Maximum N of iterations}

\item{scale}{Logical: If TRUE, scale input data before decomposition. Default = TRUE}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}.
Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{trace}{Integer: If > 0, print messages during ICA run. Default = 0}

\item{...}{Additional parameters to be passed to \code{fastICA::fastICA} or \code{ica::icafast}}
}
\value{
\link{rtDecom} object
}
\description{
Perform ICA decomposition using the fastICA algorithm in \code{fastICA::fastICA} or \code{ica::fastica}
}
\details{
Project scaled variables to ICA components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
fastICA will be applied to the transpose of the n x p matrix.
fastICA will fail if there are any NA values or constant features: remove them using \link{preprocess}
}
\seealso{
Other Decomposition: \code{\link{d.CUR}},
  \code{\link{d.H2OAE}}, \code{\link{d.H2OGLRM}},
  \code{\link{d.ISOMAP}}, \code{\link{d.KPCA}},
  \code{\link{d.LLE}}, \code{\link{d.MDS}},
  \code{\link{d.NMF}}, \code{\link{d.PCA}},
  \code{\link{d.SPCA}}, \code{\link{d.SVD}},
  \code{\link{d.TSNE}}, \code{\link{d.UMAP}}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
