% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.ISOMAP.R
\name{d.ISOMAP}
\alias{d.ISOMAP}
\title{Isomap}
\usage{
d.ISOMAP(x, k = 2, dist.method = "euclidean", nsd = 0,
  path = c("shortest", "extended"), verbose = TRUE,
  n.cores = rtCores, ...)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{dist.method}{Character: Distance calculation method. See \code{vegan::vegdist}}

\item{nsd}{Integer: Number of shortest dissimilarities retained}

\item{path}{Character: The \code{path} argument of \code{vegan::isomap}}

\item{verbose}{Logical: If TRUE, print messages to output}

\item{n.cores}{Integer: Number of cores to use}

\item{...}{Additional parameters to be passed to \code{vegan::isomap}}
}
\value{
\link{rtDecom} object
}
\description{
Perform ISOMAP decomposition using \code{vegan::isomap}
}
\details{
Project scaled variables to ISOMAP components
Input must be n by p,
where n represents number of cases,
and p represents number of features.
ISOMAP will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: \code{\link{d.CUR}},
  \code{\link{d.H2OAE}}, \code{\link{d.H2OGLRM}},
  \code{\link{d.ICA}}, \code{\link{d.KPCA}},
  \code{\link{d.LLE}}, \code{\link{d.MDS}},
  \code{\link{d.NMF}}, \code{\link{d.PCA}},
  \code{\link{d.SPCA}}, \code{\link{d.SVD}},
  \code{\link{d.TSNE}}, \code{\link{d.UMAP}}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
