% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.KPCA.R
\name{d.KPCA}
\alias{d.KPCA}
\title{Kernel Principal Component Analysis}
\usage{
d.KPCA(x, x.test = NULL, k = 2, th = 1e-04, kernel = "rbfdot",
  kpar = NULL, scale = TRUE, center = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{Input data}

\item{x.test}{Optional test set. Will be projected on to KPCA basis}

\item{k}{Integer vector of length 1 or greater. N of components to return
If set to 0, \code{th} determines eigenvalue below which PCs are ignored}

\item{th}{Threshold for eigenvalue below which PCs are ignored if \code{k} is set to 0}

\item{kernel}{String: Type of kernel to use. See \code{kernlab::kpca}}

\item{kpar}{List of hyperparameters: See \code{kernlab::kpca("kpar")}}

\item{scale}{Logical: If TRUE, scale input data before projecting}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{fastKPCA::fastKPCA}}
}
\value{
\link{rtDecom} object
}
\description{
Calculates kernel PCA decomposition and projections using \code{kernlab::kpca}
}
\details{
Project scaled variables to KPCA components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
KPCA will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: \code{\link{d.CUR}},
  \code{\link{d.H2OAE}}, \code{\link{d.H2OGLRM}},
  \code{\link{d.ICA}}, \code{\link{d.ISOMAP}},
  \code{\link{d.LLE}}, \code{\link{d.MDS}},
  \code{\link{d.NMF}}, \code{\link{d.PCA}},
  \code{\link{d.SPCA}}, \code{\link{d.SVD}},
  \code{\link{d.TSNE}}, \code{\link{d.UMAP}}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
