% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.MDS.R
\name{d.MDS}
\alias{d.MDS}
\title{Multidimensional Scaling}
\usage{
d.MDS(x, k = 2, dist.method = c("euclidean", "maximum", "manhattan",
  "canberra", "binary", "minkowski"), eig = FALSE, add = FALSE,
  x.ret = FALSE, scale = TRUE, center = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{dist.method}{String: method to use to calculate distance. See \code{stats::dist("method")}}

\item{eig}{Logical: If TRUE, return eigenvalues. Default = FALSE}

\item{add}{Logical: If TRUE, an additive constant \code{c*} will be computed and added to the
non-diagonal dissimilarities, which makes the Euclidean. Default = FALSE}

\item{x.ret}{Logical: If TRUE, return the doubly centered symmetric distance matrix. Default = FALSE}

\item{scale}{Logicall: If TRUE, scale input data before decomposition. Default = TRUE}

\item{center}{Logicall: If TRUE, also center input data if \code{scale} is \code{TRUE}.
Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{svd}}
}
\value{
\link{rtDecom} object
}
\description{
Perform MDS decomposition using \code{stats:cmdscale}
}
\details{
Project scaled variables to MDS components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
fastMDS will be applied to the transpose of the n x p matrix.
fastMDS will fail if there are any NA values or constant features: remove them using \link{preprocess}
}
\seealso{
Other Decomposition: \code{\link{d.CUR}},
  \code{\link{d.H2OAE}}, \code{\link{d.H2OGLRM}},
  \code{\link{d.ICA}}, \code{\link{d.ISOMAP}},
  \code{\link{d.KPCA}}, \code{\link{d.LLE}},
  \code{\link{d.NMF}}, \code{\link{d.PCA}},
  \code{\link{d.SPCA}}, \code{\link{d.SVD}},
  \code{\link{d.TSNE}}, \code{\link{d.UMAP}}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
