% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.NMF.R
\name{d.NMF}
\alias{d.NMF}
\title{Non-negative Matrix Factorization (NMF)}
\usage{
d.NMF(x, x.test = NULL, k = 2, method = "brunet", nrun = 30,
  scale = TRUE, center = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{Input data}

\item{x.test}{Optional test set. Will be projected on to NMF basis}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{method}{NMF method. Defaults to "brunet". See \code{NMF::nmf}}

\item{nrun}{Integer: Number of runs to perform}

\item{scale}{Logical: If TRUE, scale input data before projecting}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{NMF::nmf}}
}
\value{
\link{rtDecom} object
}
\description{
Calculate NMF decomposition and projections
}
\details{
Project scaled variables to NMF bases.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
NMF will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: \code{\link{d.CUR}},
  \code{\link{d.H2OAE}}, \code{\link{d.H2OGLRM}},
  \code{\link{d.ICA}}, \code{\link{d.ISOMAP}},
  \code{\link{d.KPCA}}, \code{\link{d.LLE}},
  \code{\link{d.MDS}}, \code{\link{d.PCA}},
  \code{\link{d.SPCA}}, \code{\link{d.SVD}},
  \code{\link{d.TSNE}}, \code{\link{d.UMAP}}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
