% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.PCA.R
\name{d.PCA}
\alias{d.PCA}
\title{Principal Component Analysis}
\usage{
d.PCA(x, x.test = NULL, k = NULL, scale = TRUE, center = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set. Will be projected on to PCA basis}

\item{k}{Integer: Number of right singular vectors to compute (\code{svd}'s \code{nv})}

\item{scale}{Logical: If TRUE, scale input data before doing SVD}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{PCA::PCA}}
}
\value{
\link{rtDecom} object
}
\description{
Perform PCA decomposition using \code{stats::prcomp}
}
\details{
Same solution as \link{d.SVD}. d.PCA runs \code{prcomp}, which has useful
\code{summary} output
}
\seealso{
Other Decomposition: \code{\link{d.CUR}},
  \code{\link{d.H2OAE}}, \code{\link{d.H2OGLRM}},
  \code{\link{d.ICA}}, \code{\link{d.ISOMAP}},
  \code{\link{d.KPCA}}, \code{\link{d.LLE}},
  \code{\link{d.MDS}}, \code{\link{d.NMF}},
  \code{\link{d.SPCA}}, \code{\link{d.SVD}},
  \code{\link{d.TSNE}}, \code{\link{d.UMAP}}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
