% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.SPCA.R
\name{d.SPCA}
\alias{d.SPCA}
\title{Sparse Principal Component Analysis}
\usage{
d.SPCA(
  x,
  x.test = NULL,
  k = 1,
  nz = 0.5 * NCOL(x),
  nneg = TRUE,
  method = c("cumulative", "vanilla"),
  scale = TRUE,
  center = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set. Will be projected on to SPCA basis}

\item{k}{Integer vector of length 1 or greater. N of components to return
If set to 0, \code{th} determines eigenvalue below which PCs are ignored}

\item{nz}{Integer: Upper bound on non-zero loadings. See \code{nsprcomp::nscumcomp("k")}}

\item{nneg}{Logical: If TRUE, calculate non-negative loadings only. Default = TRUE}

\item{method}{Character: "cumulative" or "vanilla" sparse PCA. Default = "cumulative"}

\item{scale}{LogSPCAl: If TRUE, scale input data before projecting. Default = TRUE}

\item{center}{LogSPCAl: If TRUE, also center input data if \code{scale} is \code{TRUE}. Default = FALSE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{fastSPCA::fastSPCA}}
}
\value{
\link{rtDecom} object
}
\description{
Perform sparse and/or non-negative PCA or cumulative PCA decomposition
  using \code{nsprcomp::nsprcomp} or \code{nsprcomp::nscumcomp} respectively
}
\details{
Project scaled variables to sparse and/or non-negative PCA components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
SPCA will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: 
\code{\link{d.CUR}()},
\code{\link{d.H2OAE}()},
\code{\link{d.H2OGLRM}()},
\code{\link{d.ICA}()},
\code{\link{d.ISOMAP}()},
\code{\link{d.KPCA}()},
\code{\link{d.LLE}()},
\code{\link{d.MDS}()},
\code{\link{d.NMF}()},
\code{\link{d.PCA}()},
\code{\link{d.SVD}()},
\code{\link{d.TSNE}()},
\code{\link{d.UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
