% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.SVD.R
\name{d.SVD}
\alias{d.SVD}
\title{Singular Value Decomposition}
\usage{
d.SVD(
  x,
  x.test = NULL,
  k = 2,
  nu = 0,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set matrix. Will be projected on to SVD bases}

\item{k}{Integer: Number of right singular vectors to compute (\code{svd}'s \code{nv})}

\item{nu}{Integer: Number of left singular vectors to compute}

\item{scale}{Logical: If TRUE, scale input data before doing SVD. Default = TRUE}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{svd}}
}
\value{
\link{rtDecom} object
}
\description{
Perform SVD decomposition using \code{base::svd}
}
\details{
Same solution as \link{d.PCA}
}
\seealso{
Other Decomposition: 
\code{\link{d.CUR}()},
\code{\link{d.H2OAE}()},
\code{\link{d.H2OGLRM}()},
\code{\link{d.ICA}()},
\code{\link{d.ISOMAP}()},
\code{\link{d.KPCA}()},
\code{\link{d.LLE}()},
\code{\link{d.MDS}()},
\code{\link{d.NMF}()},
\code{\link{d.PCA}()},
\code{\link{d.SPCA}()},
\code{\link{d.TSNE}()},
\code{\link{d.UMAP}()}
}
\author{
Efstathios D. Gennatas
}
\concept{Decomposition}
