% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.UMAP.R
\name{d.UMAP}
\alias{d.UMAP}
\title{Uniform Manifold Approximation and Projection}
\usage{
d.UMAP(
  x,
  x.test = NULL,
  k = 2,
  n.neighbors = 15,
  init = "spectral",
  metric = c("euclidean", "cosine", "manhattan", "hamming", "categorical"),
  epochs = NULL,
  learning.rate = 1,
  scale = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set matrix. Will be projected on to UMAP bases}

\item{k}{Integer: Number of projections}

\item{n.neighbors}{Integer: Number of keighbors}

\item{init}{Character: Initialization type. See \code{uwot::umap "init"}}

\item{metric}{Character: Distance metric to use: "euclidean", "cosine", "manhattan", "hamming", "categorical"
Default = "euclidean"}

\item{epochs}{Integer: Number of epochs}

\item{learning.rate}{Float: Learning rate. Default = 1}

\item{scale}{Logical: If TRUE, scale input data before doing UMAP. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{uwot::umap}}
}
\value{
\link{rtDecom} object
}
\description{
Perform UMAP decomposition using github package \code{jlmelville/uwot}
}
\seealso{
Other Decomposition: 
\code{\link{d.CUR}()},
\code{\link{d.H2OAE}()},
\code{\link{d.H2OGLRM}()},
\code{\link{d.ICA}()},
\code{\link{d.ISOMAP}()},
\code{\link{d.KPCA}()},
\code{\link{d.LLE}()},
\code{\link{d.MDS}()},
\code{\link{d.NMF}()},
\code{\link{d.PCA}()},
\code{\link{d.SPCA}()},
\code{\link{d.SVD}()},
\code{\link{d.TSNE}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
