% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_CUR.R
\name{d_CUR}
\alias{d_CUR}
\title{CUR Decomposition}
\usage{
d_CUR(
  x,
  c = dim(x)[2],
  r = dim(x)[1],
  k = NULL,
  sv = NULL,
  method = "highest.ranks",
  matrix.return = TRUE,
  error.return = FALSE,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric matrix / data.frame: Input data}

\item{c}{Integer: Number of columns to be selected}

\item{r}{Integer: Number of rows to be selected}

\item{k}{Integer: Rank of decomposition (Creates \code{k} components)}

\item{sv}{The SVD of \code{x}, if already available}

\item{method}{Character: "random", "top.scores", "ortho.top.scores",
"exact.num.random", "highest.ranks" (Default). See \code{rCUR::CUR("method")}}

\item{matrix.return}{Logical: if TRUE, matrices C, U, and R are returned.
If FALSE, U is not computed, which can be expensive if \code{r} and \code{c}
are large.}

\item{error.return}{Logical: if TRUE, the Frobenius norm of the difference
between the original matrix and the CUR approximation is returned.
Effective only if \code{matrix.return = TRUE}.}

\item{scale}{Logical: If TRUE, scale input}

\item{center}{Logical: If TRUE, center input}

\item{verbose}{Logical: If TRUE, print messages to output}

\item{...}{Additional parameters to be passed to \code{rCUR::CUR}}
}
\value{
\link{rtDecom} object
}
\description{
Performs CUR decomposition using \code{rCUR::CUR}
}
\details{
Note that \code{k} here does not correspond with \code{k} in the other
decomposition functions.
Use \code{c} to determine dimensionality of resulting decomposition
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_ISOMAP}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
