% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_H2OGLRM.R
\name{d_H2OGLRM}
\alias{d_H2OGLRM}
\title{Generalized Low-Rank Models (GLRM) on H2O}
\usage{
d_H2OGLRM(
  x,
  x.test = NULL,
  x.valid = NULL,
  k = 3,
  ip = "localhost",
  port = 54321,
  transform = "NONE",
  loss = "Quadratic",
  regularization.x = "None",
  regularization.y = "None",
  gamma.x = 0,
  gamma.y = 0,
  max_iterations = 1000,
  max_updates = 2 * max_iterations,
  init_step_size = 1,
  min_step_size = 1e-04,
  seed = -1,
  init = "PlusPlus",
  svd.method = "Randomized",
  verbose = TRUE,
  print.plot = TRUE,
  plot.theme = rtTheme,
  n.cores = rtCores,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{x.test}{Optional test set. Will be projected on to NMF basis}

\item{x.valid}{Optional validation set}

\item{k}{Integer: Rank of decomposition}

\item{ip}{Character: IP address of H2O server. Default = "localhost"}

\item{port}{Integer: Port number for server. Default = 54321}

\item{transform}{Character: Transformation of input prior to decomposition}

\item{loss}{Character: Numeric loss function: "Quadratic", "Absolute", "Huber", "Poisson", "Hinge", "Logistic",
"Periodic". Default = "Quadratic"}

\item{regularization.x}{Character: Regularization function for X matrix: "None", "Quadratic", "L2", "L1",
"NonNegative", "OneSparse", "UnitOneSparse", "Simplex". Default = "None"}

\item{regularization.y}{Character: Regularization function for Y matrix: "None", "Quadratic", "L2", "L1",
"NonNegative", "OneSparse", "UnitOneSparse", "Simplex". Default = "None"}

\item{gamma.x}{Float: Regularization weight on X matrix. Default = 0}

\item{gamma.y}{Float: Regularization weight on Y matrix. Default = 0}

\item{max_iterations}{Integer: Maximum number of iterations. Default = 1000}

\item{max_updates}{Integer: Maximum number of iterations. Default = 2 * \code{max_iterations}}

\item{init_step_size}{Float: Initial step size. Default = 1}

\item{min_step_size}{Float: Minimum step size. Default = .0001}

\item{seed}{Integer: Seed for random number generator. Default = -1 (time-based)}

\item{init}{Character: Initialization mode: "Random", "SVD", "PlusPlus", "User". Default = "PlusPlus"}

\item{svd.method}{Character: SVD method for initialization: "GramSVD", "Power", "Randomized". Default = "Randomized"}

\item{verbose}{Logical: If TRUE, print console messages}

\item{print.plot}{Logical: If TRUE, print objective score against iteration number}

\item{plot.theme}{Character: Theme to pass to \link{mplot3_xy} if \code{print.plot = TRUE}}

\item{n.cores}{Integer: Number of cores to use}

\item{...}{Additional parameters to be passed to \code{h2o::h2o.glrm}}
}
\value{
\code{rtDecom} object
}
\description{
Perform GLRM decomposition using \code{h2o::h2o.glrm}
Given Input matrix \code{A}:
\verb{A(m x n) = X(m x k) \\\%*\\\% Y(k x n)}
}
\details{
Learn more about GLRM from the H2O tutorial
\url{https://github.com/h2oai/h2o-tutorials/blob/master/tutorials/glrm/glrm-tutorial.md}
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
