% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ICA.R
\name{d_ICA}
\alias{d_ICA}
\title{Independent Component Analysis}
\usage{
d_ICA(
  x,
  k = 3,
  package = c("fastICA", "ica"),
  alg.type = "parallel",
  maxit = 100,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  trace = 0,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{package}{Character: Which package to use for ICA. "fastICA" will
use \code{fastICA::fastICA},
"ica" will use \code{ica::fastica}. Default = "fastICA".
Note: only \code{fastICA} works with \code{k = 1}}

\item{alg.type}{Character: For \code{package = "fastICA"}, "parallel" or
"deflation".}

\item{maxit}{Integer: Maximum N of iterations}

\item{scale}{Logical: If TRUE, scale input data before decomposition.}

\item{center}{Logical: If TRUE, also center input data if \code{scale}
is \code{TRUE}.}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{trace}{Integer: If > 0, print messages during ICA run. Default = 0}

\item{...}{Additional parameters to be passed to \code{fastICA::fastICA}
or \code{ica::icafast}}
}
\value{
\code{rtDecom} object
}
\description{
Perform ICA decomposition using the fastICA algorithm in \code{fastICA::fastICA} or
\code{ica::fastica}
}
\details{
Project scaled variables to ICA components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
fastICA will be applied to the transpose of the n x p matrix.
fastICA will fail if there are any NA values or constant features: remove them using
\link{preprocess}
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
