% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_Isomap.R
\name{d_Isomap}
\alias{d_Isomap}
\title{Isomap}
\usage{
d_Isomap(
  x,
  k = 2,
  dist.method = "euclidean",
  nsd = 0,
  path = c("shortest", "extended"),
  center = TRUE,
  scale = TRUE,
  verbose = TRUE,
  n.cores = rtCores,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{dist.method}{Character: Distance calculation method. See \code{vegan::vegdist}}

\item{nsd}{Integer: Number of shortest dissimilarities retained}

\item{path}{Character: The \code{path} argument of \code{vegan::isomap}}

\item{center}{Logical: If TRUE, center data prior to decomposition. Default = TRUE}

\item{scale}{Logical: If TRUE, scale data prior to decomposition. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to output}

\item{n.cores}{Integer: Number of cores to use}

\item{...}{Additional parameters to be passed to \code{vegan::isomap}}
}
\value{
\code{rtDecom} object
}
\description{
Perform ISOMAP decomposition using \code{vegan::isomap}
}
\details{
Project scaled variables to ISOMAP components
Input must be n by p,
where n represents number of cases,
and p represents number of features.
ISOMAP will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
