% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_KPCA.R
\name{d_KPCA}
\alias{d_KPCA}
\title{Kernel Principal Component Analysis}
\usage{
d_KPCA(
  x,
  x.test = NULL,
  k = 2,
  th = 1e-04,
  kernel = "rbfdot",
  kpar = NULL,
  center = TRUE,
  scale = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{x.test}{Optional test set. Will be projected on to KPCA basis}

\item{k}{Integer vector of length 1 or greater. N of components to return
If set to 0, \code{th} determines eigenvalue below which PCs are ignored}

\item{th}{Threshold for eigenvalue below which PCs are ignored if \code{k} is set to 0}

\item{kernel}{Character: Type of kernel to use. See \code{kernlab::kpca}}

\item{kpar}{List of hyperparameters: See \code{kernlab::kpca("kpar")}}

\item{center}{Logical: If TRUE, center data prior to decomposition. Default = TRUE}

\item{scale}{Logical: If TRUE, scale data prior to decomposition. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{fastKPCA::fastKPCA}}
}
\value{
\code{rtDecom} object
}
\description{
Perform kernel PCA decomposition using \code{kernlab::kpca}
}
\details{
Project scaled variables to KPCA components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
KPCA will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
