% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_LLE.R
\name{d_LLE}
\alias{d_LLE}
\title{Locally Linear Embedding}
\usage{
d_LLE(
  x,
  k = 2,
  nn = 0,
  nn.min = 1,
  nn.max = 20,
  plot.calcnn = FALSE,
  id = FALSE,
  iLLE = FALSE,
  nnk = TRUE,
  reg = 2,
  v = 0.9,
  verbose = TRUE,
  n.cores = 1,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{nn}{Integer: Number of neighbors. If Set to 0 (default), will use \code{lle::calc_k} to
estimate optimal number}

\item{nn.min}{Integer: Minimum n of neighbors to consider in search, used if \code{nn = 0}}

\item{nn.max}{Integer: Maximum n of enighbors to consider in search, used if \code{nn = 0}}

\item{plot.calcnn}{Logical: If TRUE, print plot after estimation of number of neighbors. Default = FALSE}

\item{id}{Logical: If TRUE, calculate \code{k} (the intrinsic dimension)}

\item{iLLE}{Logical: If TRUE, use the improved LLE algorithm; see Details in \code{lle::lle}
Notice: It causes warnings for matrix dimensions (check \code{lle} code)}

\item{nnk}{Logical: If TRUE, use k nearest neighbors method; otherwise, epsilon environment neighbourhood
will be used}

\item{reg}{Integer {1, 2, 3}: Regularization methods: See \code{lle::lle("reg")}}

\item{v}{Float: Threshold value for intrinsic dimension estimation. Suggestion for noiseless
data: .99, for noisy data: .9. Default = .9}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{n.cores}{Integer: Number of cores to use. Default = 1. At some point using more than one cores stopped working.
The \code{lle} package has not been updated since February 2015 - we will switch to a different implementation soon}

\item{...}{Additional parameters to be passed to \code{LLE::LLE}}
}
\value{
\link{rtDecom} object
}
\description{
Perform LLE decomposition using \code{lle::lle}
}
\details{
Project scaled variables to LLE components
Input must be n by p,
where n represents number of cases,
and p represents number of features.
LLE will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: 
\code{\link{d_CUR}()},
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
