% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_MDS.R
\name{d_MDS}
\alias{d_MDS}
\title{Multidimensional Scaling}
\usage{
d_MDS(
  x,
  k = 2,
  dist.method = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  eig = FALSE,
  add = FALSE,
  x.ret = FALSE,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{dist.method}{Character: method to use to calculate distance. See \code{stats::dist("method")}}

\item{eig}{Logical: If TRUE, return eigenvalues. Default = FALSE}

\item{add}{Logical: If TRUE, an additive constant \verb{c*} will be computed and added to the
non-diagonal dissimilarities, which makes the Euclidean. Default = FALSE}

\item{x.ret}{Logical: If TRUE, return the doubly centered symmetric distance matrix. Default = FALSE}

\item{scale}{Logical: If TRUE, scale input data before decomposition. Default = TRUE}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}.
Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{svd}}
}
\value{
\code{rtDecom} object
}
\description{
Perform MDS decomposition using \code{stats:cmdscale}
}
\details{
Project scaled variables to MDS components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
fastMDS will be applied to the transpose of the n x p matrix.
fastMDS will fail if there are any NA values or constant features: remove them using \link{preprocess}
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
