% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_NMF.R
\name{d_NMF}
\alias{d_NMF}
\title{Non-negative Matrix Factorization (NMF)}
\usage{
d_NMF(
  x,
  x.test = NULL,
  k = 2,
  method = "brunet",
  nrun = 30,
  scale = TRUE,
  center = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input data}

\item{x.test}{Optional test set. Will be projected on to NMF basis}

\item{k}{Integer vector of length 1 or greater. Rank of decomposition}

\item{method}{NMF method. Defaults to "brunet". See \code{NMF::nmf}}

\item{nrun}{Integer: Number of runs to perform}

\item{scale}{Logical: If TRUE, scale input data before projecting}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{NMF::nmf}}
}
\value{
\code{rtDecom} object
}
\description{
Perform NMF decomposition using \code{NMF::nmf}
}
\details{
Project scaled variables to NMF bases.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
NMF will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
