% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_PCA.R
\name{d_PCA}
\alias{d_PCA}
\title{Principal Component Analysis}
\usage{
d_PCA(
  x,
  x.test = NULL,
  k = NULL,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set. Will be projected on to PCA basis}

\item{k}{Integer: Number of right singular vectors to compute (\code{svd}'s \code{nv})}

\item{scale}{Logical: If TRUE, scale input data before doing SVD}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{PCA::PCA}}
}
\value{
\code{rtDecom} object
}
\description{
Perform PCA decomposition using \code{stats::prcomp}
}
\details{
Same solution as \link{d_SVD}. d_PCA runs \code{prcomp}, which has useful
\code{summary} output
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
