% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_SPCA.R
\name{d_SPCA}
\alias{d_SPCA}
\title{Sparse Principal Component Analysis}
\usage{
d_SPCA(
  x,
  x.test = NULL,
  k = 1,
  nz = floor(0.5 * NCOL(x)),
  nneg = FALSE,
  gamma = 0,
  method = c("cumulative", "vanilla"),
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set. Will be projected on to SPCA basis}

\item{k}{Integer vector of length 1 or greater. N of components to return
If set to 0, \code{th} determines eigenvalue below which PCs are ignored}

\item{nz}{Integer: Upper bound on non-zero loadings. See \code{nsprcomp::nscumcomp("k")}}

\item{nneg}{Logical: If TRUE, calculate non-negative loadings only. Default = FALSE}

\item{gamma}{Float (>0): Penalty on the divergence from otrhonormality of the pseudo-rotation
matrix. Default = 0, i.e. no penalty. May need to increase with collinear features.}

\item{method}{Character: "cumulative" or "vanilla" sparse PCA. Default = "cumulative"}

\item{scale}{Logical: If TRUE, scale input data before projecting. Default = TRUE}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}. Default = FALSE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{fastSPCA::fastSPCA}}
}
\value{
\code{rtDecom} object
}
\description{
Perform sparse and/or non-negative PCA or cumulative PCA decomposition
using \code{nsprcomp::nsprcomp} or \code{nsprcomp::nscumcomp} respectively
}
\details{
Project scaled variables to sparse and/or non-negative PCA components.
Input must be n by p,
where n represents number of cases,
and p represents number of features.
SPCA will be applied to the transpose of the n x p matrix.
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
