% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_SVD.R
\name{d_SVD}
\alias{d_SVD}
\title{Singular Value Decomposition}
\usage{
d_SVD(
  x,
  x.test = NULL,
  k = 2,
  nu = 0,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set matrix. Will be projected on to SVD bases}

\item{k}{Integer: Number of right singular vectors to compute (\code{svd}'s \code{nv})}

\item{nu}{Integer: Number of left singular vectors to compute}

\item{scale}{Logical: If TRUE, scale input data before doing SVD. Default = TRUE}

\item{center}{Logical: If TRUE, also center input data if \code{scale} is \code{TRUE}. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{svd}}
}
\value{
\code{rtDecom} object
}
\description{
Perform SVD decomposition using \code{base::svd}
}
\details{
Same solution as \link{d_PCA}
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
