% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_UMAP.R
\name{d_UMAP}
\alias{d_UMAP}
\title{Uniform Manifold Approximation and Projection (UMAP)}
\usage{
d_UMAP(
  x,
  x.test = NULL,
  k = 2,
  n.neighbors = 15,
  init = "spectral",
  metric = c("euclidean", "cosine", "manhattan", "hamming", "categorical"),
  epochs = NULL,
  learning.rate = 1,
  scale = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{x.test}{Optional test set matrix. Will be projected on to UMAP bases}

\item{k}{Integer: Number of projections}

\item{n.neighbors}{Integer: Number of keighbors}

\item{init}{Character: Initialization type. See \verb{uwot::umap "init"}}

\item{metric}{Character: Distance metric to use: "euclidean", "cosine",
"manhattan", "hamming", "categorical". Default = "euclidean"}

\item{epochs}{Integer: Number of epochs}

\item{learning.rate}{Float: Learning rate. Default = 1}

\item{scale}{Logical: If TRUE, scale input data before doing UMAP.
Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}

\item{...}{Additional parameters to be passed to \code{uwot::umap}}
}
\value{
\code{rtDecom} object
}
\description{
Perform UMAP decomposition using \code{uwot::umap}
}
\details{
Updated 2023-12-09: See \href{https://github.com/jlmelville/uwot/issues/115}{GitHub issue}
and \href{https://github.com/bwlewis/irlba/issues/70#issuecomment-1826900769}{related comment}
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_LLE}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
