% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2bsplinemat.R
\name{dat2bsplinemat}
\alias{dat2bsplinemat}
\title{B-Spline matrix from dataset}
\usage{
dat2bsplinemat(
  x,
  df = NULL,
  knots = NULL,
  degree = 3L,
  intercept = FALSE,
  Boundary.knots = range(x, na.rm = TRUE),
  return.deriv = FALSE,
  as.data.frame = TRUE
)
}
\arguments{
\item{x}{data.frame: Input}

\item{df}{Integer: Degrees of freedom. See \code{splines::bSpline}}

\item{knots}{Float, vector: Internal breakpoints. See \code{splines::bSpline}}

\item{degree}{Integer (>0): Degree of the piecewise polynomial. See \code{splines::bSpline}}

\item{intercept}{Logical: If TRUE, an intercept is included. Default = FALSE}

\item{Boundary.knots}{Float, vector (length = 2): Boundary points to anchor the spline basis.}

\item{return.deriv}{Logical: If TRUE, return list containing a data frame with the splines and another
data frame with their derivatives}

\item{as.data.frame}{Logical: If TRUE, return data.frame, otherwise matrix. Default = TRUE
See \code{splines::bSpline}}
}
\value{
If \code{return.deriv=F}, a data frame where each original feature is replaced with its basis set or a list,
otherwise a list containing a data frame with splines and a data frame with their derivatives
}
\description{
Convert a dataset to its b-spline basis set
}
\author{
E.D. Gennatas
}
