% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2poly.R
\name{dat2poly}
\alias{dat2poly}
\title{Create n-degree polynomial from data frame}
\usage{
dat2poly(
  dat,
  method = c("simple", "poly"),
  degree = 2,
  raw = FALSE,
  as.data.frame = TRUE
)
}
\arguments{
\item{dat}{Numeric, matrix / data.frame: Input}

\item{method}{Character: "simple", "poly". "simple": raise each column of \code{dat} up to \code{degree}.
"poly": use \code{stats::poly} with arguments \code{degree} and \code{raw}}

\item{degree}{Integer: degree of polynomials to create. Default = 2}

\item{raw}{Logical: If TRUE, create simple polynomial, not orthogonalized. Default = FALSE}

\item{as.data.frame}{Logical: If TRUE, return data.frame. Default = TRUE}
}
\description{
This is a convenience function that will take each column of the input and calculate 1:degree powers and concatenate
into a data.frame of dimensions \code{n * (degree * p)} given an \code{n * p} input
}
\author{
E.D. Gennatas
}
