% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{dataPrepare}
\alias{dataPrepare}
\title{\code{rtemis-internals}: \code{dataPrepare}}
\usage{
dataPrepare(x, y, x.test = NULL, y.test = NULL, x.valid = NULL,
  y.valid = NULL, ipw = FALSE, ipw.type = 2, upsample = FALSE,
  upsample.seed = NULL, removeDots = FALSE, .preprocess = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{(Optional) Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{(Optional) Numeric vector of testing set outcome}

\item{x.valid}{Matrix / Data frame: Validation set features}

\item{y.valid}{Vector: Validation outcome}

\item{ipw}{Logical: If TRUE, return class weights for inverse probability weighting
(for Classification)}

\item{ipw.type}{{1, 2}: 1:}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Caution: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{upsample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{removeDots}{Logical: If TRUE, replace dots in variable names with underscores.
Some algorithms do not work with variable names containing dots (SparkML)}

\item{.preprocess}{List: Preprocessing settings. Set with \link{rtset.preprocess}}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\description{
Prepare data for \pkg{rtemis} modeling
}
