% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemisInit.R
\name{dataSummary}
\alias{dataSummary}
\title{\code{rtemis-internals}: \code{dataSummary}}
\usage{
dataSummary(x, y, x.test = NULL, y.test = NULL, type = NULL, testSet = TRUE)
}
\description{
Print input data dimensions and test dimensions match
}
\keyword{internal}
