% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2factor.R
\name{date2factor}
\alias{date2factor}
\title{Date to factor time bin}
\usage{
date2factor(
  x,
  time_bin = c("year", "quarter", "month", "day"),
  make_bins = c("range", "present"),
  bin_range = range(x),
  ordered = FALSE
)
}
\arguments{
\item{x}{Date vector}

\item{time_bin}{Character: "year", "quarter", "month", or "day"}

\item{make_bins}{Character: "range" or "preseent". If "range" the factor levels will include all
time periods define by \code{time_bin} within \code{bin_range}. This means factor levels can be
empty. Otherwise, if "present", factor levels only include time periods present in data.}

\item{bin_range}{Date, vector, length 2: Range of dates to make levels for. Defaults to range of
input dates \code{x}}

\item{ordered}{Logical: If TRUE, factor output is ordered. Default = FALSE}
}
\value{
factor of time periods
}
\description{
Convert Date to time bin factor.
}
\details{
Order of levels will be chronological (important e.g. for plotting)
Additionally, can output ordered factor with \code{ordered = TRUE}
}
\examples{
\dontrun{
library(data.table)
startDate <- as.Date("2018-01-01")
endDate <- as.Date("2020-12-31")
time <- sample(seq(startDate, endDate, length.out = 100))
date2factor(time)
date2factor(time, "quarter")
date2factor(time, "month")
date2factor(time, "day")
# range vs present
x <- sample(seq(as.Date("2018-01-01"), as.Date("2021-01-01"), by = 1), 10)
date2factor(x, time_bin = "quarter", make_bins = "present")
date2factor(x, time_bin = "quarter", make_bins = "range")
}
}
\author{
E.D. Gennatas
}
