% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2factor.R
\name{date2ym}
\alias{date2ym}
\title{Date to year-month factor}
\usage{
date2ym(x, ordered = FALSE)
}
\arguments{
\item{x}{Date vector}

\item{ordered}{Logical: If TRUE, return ordered factor. Default = FALSE}
}
\description{
Date to year-month factor
}
\author{
E.D. Gennatas
}
