% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2factor.R
\name{date2yq}
\alias{date2yq}
\title{Date to year-quarter factor}
\usage{
date2yq(x, ordered = FALSE)
}
\arguments{
\item{x}{Date vector}

\item{ordered}{Logical: If TRUE, return ordered factor. Default = FALSE}
}
\description{
Date to year-quarter factor
}
\author{
E.D. Gennatas
}
