% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ops.R
\name{dates2features}
\alias{dates2features}
\title{Extract features from dates}
\usage{
dates2features(
  dates,
  features = c("weekday", "month", "year"),
  drop.dates = TRUE
)
}
\arguments{
\item{dates}{Date vector}

\item{features}{Character vector: features to extract}

\item{drop.dates}{Logical: If TRUE, drop original date column}
}
\value{
data.table with extracted features
}
\description{
Extract features from dates
}
\details{
weekday and month will be extracted as factors, year as integer
}
\author{
EDG
}
