% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_duckdb.R
\name{ddb_collect}
\alias{ddb_collect}
\title{Collect a lazy-read duckdb table}
\usage{
ddb_collect(sql, progress = TRUE, returnobj = c("data.frame", "data.table"))
}
\arguments{
\item{sql}{Character: DuckDB SQL query, usually output of
\link{ddb_data} with \code{collect = FALSE}}

\item{progress}{Logical: If TRUE, show progress bar}

\item{returnobj}{Character: data.frame or data.table: class of object to return}
}
\value{
\code{data.frame} or \code{data.table}.
}
\description{
Collect a table read with \code{ddb_data(x, collect = FALSE)}
}
\examples{
\dontrun{
sql <- ddb_data("/Data/iris.csv", collect = FALSE)
ir <- ddb_ollect(sql)
}
}
\author{
EDG
}
