% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp}
\alias{decomp}
\title{Perform Data Decomposition}
\usage{
decomp(x, algorithm = "ICA", parameters = NULL, verbosity = 1L)
}
\arguments{
\item{x}{Matrix or data frame: Input data.}

\item{algorithm}{Character: Decomposition algorithm.}

\item{parameters}{DecompositionParameters: Algorithm-specific parameters.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Decomposition object.
}
\description{
Perform linear or non-linear decomposition of numeric data.
}
\author{
EDG
}
