% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depCheck.R
\name{depCheck}
\alias{depCheck}
\title{\pkg{rtemis} internal: Dependencies check}
\usage{
depCheck(..., verbose = TRUE)
}
\arguments{
\item{...}{List or vector of strings defining namespaces to be checked}

\item{verbose}{Logical. Defaults to TRUE, disable to eliminate unwanted console output}
}
\value{
TRUE if dependencies can be loaded, otherwise FALSE
}
\description{
Checks if dependencies can be loaded; names missing dependencies if not.
}
\author{
Efstathios D. Gennatas
}
