% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_check.R
\name{dependency_check}
\alias{dependency_check}
\title{\pkg{rtemis} internal: Dependencies check}
\usage{
dependency_check(..., verbose = FALSE)
}
\arguments{
\item{...}{List or vector of strings defining namespaces to be checked}

\item{verbose}{Logical. If TRUE, print messages to consolde.
Note: An error will always printed if dependencies are missing.
Setting this to FALSE stops it from printing
"Dependencies check passed".}
}
\description{
Checks if dependencies can be loaded; names missing dependencies if not.
}
\author{
E.D. Gennatas
}
