% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{df_movecolumn}
\alias{df_movecolumn}
\title{Move data frame column}
\usage{
df_movecolumn(x, from, to = ncol(x))
}
\arguments{
\item{x}{data.frame.}

\item{from}{String or Integer: Define which column holds the vector you want to move.}

\item{to}{Integer: Define which column number you want the vector to be moved to.
Default = \code{ncol(x)} i.e. the last column.}
}
\value{
data.frame
}
\description{
Move data frame column
}
\examples{
\dontrun{
mtcars_hp <- df_movecolumn(mtcars, "hp")
}
}
\author{
EDG
}
