% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{df_nunique_perfeat}
\alias{df_nunique_perfeat}
\title{Unique values per feature}
\usage{
df_nunique_perfeat(x, excludeNA = FALSE)
}
\arguments{
\item{x}{matrix or data frame input}

\item{excludeNA}{Logical: If TRUE, exclude NA values from unique count.}
}
\value{
Vector, integer of length \code{NCOL(x)} with number of unique
values per column/feature
}
\description{
Get number of unique values per features
}
\examples{
df_nunique_perfeat(iris)
}
\author{
EDG
}
