% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.addtree.R
\name{dplot3.addtree}
\alias{dplot3.addtree}
\title{\code{mplot3} ADDTREE trees}
\usage{
dplot3.addtree(
  addtree,
  col.positive = "#F48024DD",
  col.negative = "#18A3ACDD",
  node.col = "#666666",
  node.shape = "none",
  node.labels = TRUE,
  node.labels.pct.pos = NULL,
  pos.name = NULL,
  edge.col = "#999999",
  edge.width = 2,
  edge.labels = FALSE,
  arrowhead = "vee",
  layout = "dot",
  drop.leaves = FALSE,
  rankdir = "TB",
  splines = "polyline",
  fontname = "helvetica",
  bg.color = "white",
  overlap = "false",
  prune = NULL,
  prune.empty.leaves = TRUE,
  remove.bad.parents = FALSE,
  filename = NULL
)
}
\description{
Plot ADDTREE trees using \code{data.tree::plot.Node}
}
\author{
Efstathios D. Gennatas
}
