% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.bar.R
\name{dplot3.bar}
\alias{dplot3.bar}
\title{Interactive Barplots}
\usage{
dplot3.bar(
  x,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  alpha = 0.8,
  theme = getOption("rt.theme", "darkgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  barmode = c("group", "relative", "stack", "overlay"),
  group.names = NULL,
  order.by.val = FALSE,
  ylim = NULL,
  hovernames = NULL,
  feature.names = NULL,
  font.size = 16,
  legend = TRUE,
  legend.col = NULL,
  margin = list(b = 50, l = 50, t = 50, r = 20),
  padding = 0,
  displayModeBar = TRUE,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  ...
)
}
\arguments{
\item{x}{data.frame: Input where rows are groups (can be 1 row), columns are features}

\item{main}{Character: Plot title. Default = NULL}

\item{xlab}{Character: x-axis label. Default = NULL}

\item{ylab}{Character: y-axis label. Default = NULL}

\item{col}{Color, vector: Color for bars. Default NULL, which will draw colors from \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for bar colors. Default = .8}

\item{theme}{Character: THeme to use: "light", "dark", "lightgrid", "darkgrid". Default = "lightgrid"}

\item{palette}{Character: Name of \pkg{rtemis} palette to use. Default = "rtCol1". Only used if \code{col = NULL}}

\item{barmode}{Character: Type of bar plot to make: "group", "relative", "stack", "overlay". Default = "group". Use
"relative" for stacked bars, wich handles negative values correctly, unlike "stack", as of writing.}

\item{group.names}{Character, vector, length = NROW(x): Group names. Default = NULL, which uses \code{rownames(x)}}

\item{order.by.val}{Logical: If TRUE, order bars by increasing value.
Only be use for single group data. Default = NULL}

\item{ylim}{Float, vector, length 2: y-axis limits.}

\item{hovernames}{Character, vector: Optional character vector to show on hover over each bar.}

\item{feature.names}{Character, vector, length = NCOL(x): Feature names. Default = NULL, which uses
\code{colnames(x)}}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{legend}{Logical: If TRUE, draw legend. Default = TRUE}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by theme}

\item{margin}{Named list: plot margins. Default = \code{list(b = 50, l = 50, t = 50, r = 20)}}

\item{padding}{Integer: N pixels to pad plot. Default = 0}

\item{filename}{Character: Path to file to save static plot. Default = NULL}

\item{file.width}{Integer: File width in pixels for when \code{filename} is set. Default = 500}

\item{file.height}{Integer: File height in pixels for when \code{filename} is set. Default = 500}

\item{bg}{Color: Background color. Default = "white"}

\item{plot.bg}{Color: Background color for plot area. Default = "white"}

\item{font.family}{String: Font family to use. Default = "Helvetica Neue"}

\item{main.col}{Color: Title color. Default = NULL, determined by theme}

\item{axes.col}{Color: Axes color. Default = NULL, determined, by theme}

\item{labs.col}{Color: Labels' color. Default = NULL, determined by theme}
}
\description{
Draw interactive barplots using \code{plotly}
}
\examples{
\dontrun{
dplot3.bar(VADeaths)
}
}
\author{
E.D. Gennatas
}
