% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.box.R
\name{dplot3.box}
\alias{dplot3.box}
\title{Interactive Boxplots & Violin plots}
\usage{
dplot3.box(
  x,
  type = c("box", "violin"),
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  alpha = 0.6,
  bg = NULL,
  plot.bg = NULL,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  boxmode = c("group", "relative", "stack", "overlay"),
  boxpoints = "outliers",
  quartilemethod = "linear",
  width = 0,
  violin.box = TRUE,
  group.names = NULL,
  order.by.fn = NULL,
  font.size = 16,
  legend = FALSE,
  legend.col = NULL,
  legend.xy = NULL,
  margin = list(t = 35, pad = 0),
  automargin.x = TRUE,
  automargin.y = TRUE,
  displayModeBar = TRUE,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  ...
)
}
\arguments{
\item{x}{Vector or List of vectors: Input}

\item{main}{Character: Plot title. Default = NULL}

\item{xlab}{Character: x-axis label. Default = NULL}

\item{ylab}{Character: y-axis label. Default = NULL}

\item{col}{Color, vector: Color for boxes. Default NULL, which will draw colors from \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for box colors. Default = .8}

\item{bg}{Color: Background color. Default = "white"}

\item{plot.bg}{Color: Background color for plot area. Default = "white"}

\item{theme}{Character: THeme to use: "light", "dark", "lightgrid", "darkgrid". Default = "lightgrid"}

\item{palette}{Character: Name of \pkg{rtemis} palette to use. Default = "rtCol1". Only used if \code{col = NULL}}

\item{boxmode}{Character: Type of box plot to make: "group", "relative", "stack", "overlay". Default = "group". Use
"relative" for stacked boxes, which handles negative values correctly, unlike "stack", as of writing.}

\item{boxpoints}{Character or FALSE: "all", "suspectedoutliers", "outliers"
See \url{https://plotly.com/r/box-plots/#choosing-the-algorithm-for-computing-quartiles}}

\item{quartilemethod}{Character: "linear", "exclusive", "inclusive"}

\item{group.names}{Character, vector, length = NROW(x): Group names. Default = NULL, which uses \code{rownames(x)}}

\item{order.by.fn}{Logical: If TRUE, order boxes by increasing value of this function
(e.g. median). Default = NULL}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{legend}{Logical: If TRUE, draw legend. Default = TRUE}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by theme}

\item{legend.xy}{Float, vector, length 2: Relative x, y position for legend. Default = NULL, which places
the legend top right beside the plot area. For example, c(0, 1) places the legend top left within the plot area}

\item{margin}{Named list: plot margins. Default = \code{list(t = 35)}}

\item{feature.names}{Character, vector, length = NCOL(x): Feature names. Default = NULL, which uses
\code{colnames(x)}}
}
\description{
Draw interactive boxplots or violin plots using \pkg{plotly}
}
\details{
Any non-numeric variable in \code{x} will be removed
}
\examples{
\dontrun{
dplot3.box(iris[, 1:4])
}
}
\author{
E.D. Gennatas
}
