% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.graphjs.R
\name{dplot3.graphjs}
\alias{dplot3.graphjs}
\title{Plot network using \pkg{threejs::graphjs}}
\usage{
dplot3.graphjs(
  net,
  vertex.size = 1,
  vertex.col = NULL,
  vertex.label.col = NULL,
  vertex.label.alpha = 0.66,
  vertex.frame.col = NA,
  vertex.label = NULL,
  vertex.shape = "circle",
  edge.col = NULL,
  edge.alpha = 0.5,
  edge.curved = 0.35,
  edge.width = 2,
  layout = c("fr", "dh", "drl", "gem", "graphopt", "kk", "lgl", "mds", "sugiyama"),
  coords = NULL,
  layout_params = list(),
  cluster = NULL,
  groups = NULL,
  cluster_params = list(),
  cluster_mark_groups = TRUE,
  cluster_color_vertices = FALSE,
  main = "",
  theme = getOption("rt.theme", "lightgrid"),
  theme_extra_args = list(),
  palette = getOption("rt.palette", "rtCol1"),
  mar = rep(0, 4),
  par.reset = TRUE,
  filename = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{net}{\code{igraph} network}

\item{vertex.size}{Numeric: Vertex size}

\item{vertex.col}{Color for vertices}

\item{vertex.label.col}{Color for vertex labels}

\item{vertex.frame.col}{Color for vertex border (frame)}

\item{vertex.label}{Character vector: Vertex labels. Default = NULL, which will keep existing
names in \code{net} if any. Set to NA to avoid printing vertex labels}

\item{vertex.shape}{Character, vector, length 1 or N nodes: Vertex shape. See \code{graphjs("vertex.shape")}. Default = "circle"}

\item{edge.col}{Color for edges}

\item{edge.alpha}{Numeric: Transparency for edges}

\item{edge.curved}{Numeric: Curvature of edges. Default = .35}

\item{edge.width}{Numeric: Edge thickness}

\item{layout}{Character: one of: "fr", "dh", "drl", "gem", "graphopt", "kk", "lgl", "mds",
"sugiyama", corresponding to all the available layouts in \pkg{igraph}}

\item{coords}{Output of precomputed \pkg{igraph} layout. If provided, \code{layout} is ignored}

\item{layout_params}{List of parameters to pass to \code{layout} function}

\item{cluster}{Characer: one of: "edge_betweenness", "fast_greedy", "infomap", "label_prop",
"leading_eigen", "louvain", "optimal", "spinglass", "walktrap", corresponding to all the
available \pkg{igraph} clustering functions}

\item{groups}{Output of precomputed \pkg{igraph} clustering. If provided, \code{cluster} is
ignored}

\item{cluster_params}{List of parameters to pass to \code{cluster} function}

\item{cluster_mark_groups}{Logical: If TRUE, draw polygons to indicate clusters, if \code{groups}
or \code{cluster} defined}

\item{cluster_color_vertices}{Logical: If TRUE, color vertices by cluster membership}

\item{theme}{\pkg{rtemis} theme to use}

\item{theme_extra_args}{List of extra arguments to pass to the theme function defined by
\code{theme}. This argument is used when the extra args (...) are passed the plotting function
(in this case \code{igraph::plot.igraph}) and not to the theme function}

\item{mar}{Numeric vector, length 4: \code{par}'s margin argument}

\item{par.reset}{Logical: If TRUE, reset par before exiting. Default = TRUE}

\item{filename}{Character: If provided, save plot to this filepath}

\item{verbose}{Logical, If TRUE, print messages to console. Default = TRUE}

\item{...}{Extra arguments to pass to \code{igraph::plot.igraph()}}
}
\description{
Interactive plotting of an \pkg{igraph} net using \pkg{threejs}
}
\author{
E.D. Gennatas
}
