% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.heatmap.R
\name{dplot3.heatmap}
\alias{dplot3.heatmap}
\title{Dynamic Heatmap}
\usage{
dplot3.heatmap(z, Rowv = TRUE, Colv = TRUE, cluster = FALSE,
  symm = FALSE, cellnote = NULL, colorGrad.n = 101, colors = NULL,
  space = "rgb", lo = "#18A3AC", lomid = NULL, mid = "white",
  midhi = NULL, hi = "#F48024", k.row = NA, k.col = NA,
  grid.gap = 0, limits = NULL, margins = NULL, main = NULL,
  xlab = NULL, ylab = NULL, key.title = NULL,
  showticklabels = NULL, ...)
}
\arguments{
\item{z}{Input matrix}

\item{Rowv}{Logical or dendrogram.
If Logical: Compute dendrogram and reorder rows. Defaults to FALSE
If dendrogram: use as is, without reordering
See more at \code{heatmaply::heatmaply("Rowv")}}

\item{Colv}{Logical or dendrogram.
If Logical: Compute dendrogram and reorder columns. Defaults to FALSE
If dendrogram: use as is, without reordering
See more at \code{heatmaply::heatmaply("Colv")}}

\item{cluster}{Logical: If TRUE, set \code{Rowv} and \code{Colv} to TRUE}

\item{symm}{Logical: If TRUE, treat \code{x} symmetrically - \code{x} must be a square matrix. Default = FALSE}

\item{cellnote}{Matrix with values to be desplayed on hover. Defaults to \code{ddSci(z)}}

\item{colors}{String: Acts as a shortcut to defining \code{lo}, \code{mid}, etc for a number of defaults:
"french", "penn", "grnblkred",}

\item{space}{String: Which colorspace to use. Option: "rgb", or "Lab". Default = "rgb".
Recommendation: If \code{mid} is "white" or "black" (default), use "rgb", otherwise "Lab"}

\item{lo}{Color for low end}

\item{lomid}{Color for low-mid}

\item{mid}{Color for middle of the range or "mean", which will result in \code{colorOp(c(lo, hi), "mean")}.
If \code{mid = NA}, then only \code{lo} and \code{hi} are used to create the color gradient.}

\item{midhi}{Color for middle-high}

\item{hi}{Color for high end}

\item{k.row}{Integer: Number of desired number of groups by which to color dendrogram branches in the rows.
Default = NA (determined automatically). See \code{heatmaply::heatmaply("k_row")}}

\item{k.col}{Integer: Number of desired number of groups by which to color dendrogram branches in the columns.
Default = NA (determined automatically). See \code{heatmaply::heatmaply("k_col")}}

\item{grid.gap}{Integer: Space between cells. Default = 0 (no space)}

\item{limits}{Float, length 2: Determine color range. Default = NULL, which automatically centers values around 0}

\item{margins}{Float, length 4: Heatmap margins. Default = c(30, 30, 30, 30)}

\item{main}{String: Plot title}

\item{xlab}{String: x-axis label}

\item{ylab}{String: y-axis label}

\item{key.title}{String: Title for the color key. Default = NULL (no title)}

\item{...}{Additional arguments to be passed to \code{heatmaply::heatmaply}}
}
\description{
Draw a dynamic heatmap using \code{heatmaply}
}
\author{
Efstathios D. Gennatas
}
