% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.leaflet.R
\name{dplot3.leaflet}
\alias{dplot3.leaflet}
\title{Plot interactive choropleth map using \pkg{leaflet}}
\usage{
dplot3.leaflet(
  dat,
  fillOpacity = 1,
  palette = NULL,
  color.mapping = c("Numeric", "Bin"),
  col.lo = "#0290EE",
  col.hi = "#FE4AA3",
  col.na = "#303030",
  col.highlight = "#FE8A4F",
  col.interpolate = c("linear", "spline"),
  col.bins = 21,
  domain = NULL,
  weight = 0.5,
  color = "black",
  alpha = 1,
  bg.tile.provider = leaflet::providers$Stamen.TonerBackground,
  bg.tile.alpha = 0.67,
  fg.tile.provider = leaflet::providers$Stamen.TonerLabels,
  legend.position = c("topright", "bottomright", "bottomleft", "topleft"),
  legend.alpha = 0.8,
  legend.title = NULL,
  init.lng = -98.5418083333333,
  init.lat = 39.2074138888889,
  init.zoom = 3,
  stroke = TRUE
)
}
\arguments{
\item{dat}{data.frame where first column is county FIPS codes and second column is values to be
plotted. County names can be provided on a third column and will appear as hover-over information
along with values from column 2}

\item{fillOpacity}{Float: Opacity for fill colors. Default = 1}

\item{palette}{Character: Color palette to use}

\item{color.mapping}{Character: "Numeric" or "Bin"}

\item{col.lo}{Overlay color mapped to lowest value}

\item{col.hi}{Overaly color mapped to highest value}

\item{col.na}{Color mappes to NA values}

\item{col.highlight}{Hover border color. Default = "#FE8A4F" (orange)}

\item{col.interpolate}{Character: "linear" or "spline"}

\item{col.bins}{Integer: Number of color bins to create if \code{color.mapping = "Bin"}.
Default = 21}

\item{domain}{Limits for mapping colors to values. Default = NULL and set to range}

\item{weight}{Float: Weight of county border lines. Default = .5}

\item{color}{Color of county border lines. Default = "black"}

\item{alpha}{Float: Overaly transparency. Default = 1}

\item{bg.tile.provider}{Background tile (below overlay colors), one of \code{leaflet::providers}}

\item{bg.tile.alpha}{Float: Background tile transparency. Default = .67}

\item{fg.tile.provider}{Foreground tile (above overlay colors), one of \code{leaflet::providers}}

\item{legend.position}{Character: One of: "topright", "bottomright", "bottomleft", "topleft".
Default = "topright"}

\item{legend.alpha}{Float: Legend box transparency. Default = .8}

\item{legend.title}{Character: Defaults to \code{colnames(dat)[2]}}

\item{init.lng}{Float: Center map around this longitude (in decimal form).
Default = -98.54180833333334 (US geographic center)}

\item{init.lat}{Float: Center map around this latitude (in decimal form).
Default = 39.207413888888894 (US geographic center)}

\item{init.zoom}{Integer: Initial zoom level (depends on device, i.e. window, size). Default = 3}

\item{stroke}{Logical: If TRUE, draw polygon borders. Default = TRUE}
}
\description{
Plot interactive choropleth map using \pkg{leaflet}
}
\author{
E.D. Gennatas
}
