% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.linad.R
\name{dplot3.linad}
\alias{dplot3.linad}
\title{Plot a Linear Additive Tree trained by \code{s.LINAD} using \strong{visNetwork}}
\usage{
dplot3.linad(
  x,
  main = NULL,
  bg = "#FFFFFF",
  shape = "box",
  nodelabels = TRUE,
  ncases.inlabels = TRUE,
  rules.on.edges = FALSE,
  log = FALSE,
  top = NULL,
  root.col = "#202020",
  node.col = "#5a5a5a",
  leaf.col = "#178CCB",
  edge.col = "#848484",
  edge.width = 4,
  arrow.scale = 0.7,
  arrow.middle = FALSE,
  col.highlight = "#FE4AA3",
  node.font.col = NULL,
  edge.font.col = "#000000",
  sort.coefs = FALSE,
  height = NULL,
  width = NULL,
  levelSeparation = 100,
  tree.font.size = 22,
  edgethickness.by.ncases = FALSE,
  font.family = "Lato",
  tooltip.coefs = TRUE,
  tooltip.delay = 50,
  table.font.size = "16px",
  table.dat.padding = "0px",
  table.lo.col = "#0290EE",
  table.hi.col = "#FE4AA3",
  dragNodes = FALSE,
  zoomView = FALSE,
  nodeSpacing = 150,
  blockShifting = TRUE,
  edgeMinimization = TRUE,
  parentCentralization = TRUE,
  direction = "UD",
  trace = 0
)
}
\arguments{
\item{main}{Character: Title. Default = NULL}

\item{bg}{Background color. Default = "#FFFFFF" (white)}

\item{shape}{Character: Node shape; one of: "square", "triangle", "box", "circle", "dot", "star", "ellipse", "database",
"text", "diamond". Default = "box"}

\item{nodelabels}{Logical: If TRUE, inlcude node labels. Default = TRUE}

\item{ncases.inlabels}{Logical: If TRUE, include number of cases with the node labels. Default = TRUE}

\item{rules.on.edges}{Logical: If TRUE, display rules on edges instead of nodes. Default = FALSE}

\item{node.col}{Color for nodes. Default = #7F7F7F" (some gray)}

\item{leaf.col}{Color for leaf nodes. Default = "#18A3AC" (teal)}

\item{edge.col}{Color for edges. Default = "#848484" (another gray)}

\item{col.highlight}{Color for surrounding edges when node is selected. Default = "#F48024" (orange)}

\item{node.font.col}{Color for node labels. Default varies by \code{shape}, black or white depending if
\code{visNetwork} draws labels on node or underneath}

\item{edge.font.col}{Color for edge labels. Default = "#000000" (black)}

\item{sort.coefs}{Logical: If TRUE, sort each coefs table. Default = FALSE}

\item{height}{Float: Height for \code{visNetwork}. Default = NULL, i.e. auto}

\item{width}{Float: Width for \code{visNetwork}. Default = NULL, i.e. auto}

\item{levelSeparation}{Float: N of pixels to separate tree levels. Default = 100}

\item{tree.font.size}{Integer: Font size for tree labels. Default = 22}

\item{font.family}{Character: Font to use throughout. Default = 'Helvetica Neue', because otherwise it will fail on a
number of external viewers, but feel free to play around, esp. within RStudio}

\item{tooltip.coefs}{Logical: If TRUE, show html coefficient tables on hover over nodes. This was placed here before
a custom html table creation function was made to replace some impossibly slow alternatives.}

\item{tooltip.delay}{Float: Delay (in milliseconds) on mouse over before showing tooltip. Default = 50}

\item{table.font.size}{Character: Font size for html coefficient on-hover tables. Default = "14px"}

\item{table.dat.padding}{Ignore, has no visible effect. Otherwise, Character: html table padding. Default = "0px"}

\item{table.lo.col}{Color for lowest coefficient values (negative). Default = "#80FFFF" (light blue)}

\item{table.hi.col}{Color for highest coefficient values (positive). Default = "#FFBE00" (light orange)}

\item{trace}{Integer: If > 0, print info to console (not particularly informative). Default = 0}

\item{tree}{\code{s.LINAD} tree}

\item{edgethickness.by.ncases.}{Logical: If TRUE, scale edge thickness by number of cases with weight = 1}
}
\description{
Plot a Linear Additive Tree trained by \code{s.LINAD} using \strong{visNetwork}
}
