% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.shytree.R
\name{dplot3.shytree}
\alias{dplot3.shytree}
\title{Plot a stepwise hybrid tree trained by \code{s.SHYTREE} using \code{plotly}}
\usage{
dplot3.shytree(
  tree,
  main = NULL,
  bg = "#FFFFFF",
  shape = "box",
  nodelabels = TRUE,
  ncases.inlabels = TRUE,
  rules.on.edges = FALSE,
  node.col = "#7F7F7F",
  leaf.col = "#18A3AC",
  edge.col = "#848484",
  col.highlight = "#F48024",
  node.font.col = NULL,
  edge.font.col = "#000000",
  sort.coefs = FALSE,
  height = NULL,
  width = NULL,
  levelSeparation = 100,
  tree.font.size = 22,
  edgethickness.by.ncases = FALSE,
  font.family = "Helvetica Neue",
  tooltip.coefs = TRUE,
  table.font.size = "14px",
  table.dat.padding = "0px",
  table.lo.col = "#80FFFF",
  table.hi.col = "#FFBE00",
  trace = 0
)
}
\arguments{
\item{tree}{\code{s.SHYTREE} tree}

\item{main}{String: Title. Default = NULL}

\item{bg}{Background color. Default = "#FFFFFF" (white)}

\item{shape}{String: Node shape; one of: "square", "triangle", "box", "circle", "dot", "star", "ellipse", "database",
"text", "diamond". Default = "box"}

\item{nodelabels}{Logical: If TRUE, inlcude node labels. Default = TRUE}

\item{ncases.inlabels}{Logical: If TRUE, include number of cases with the node labels. Default = TRUE}

\item{rules.on.edges}{Logical: If TRUE, display rules on edges instead of nodes. Default = FALSE}

\item{node.col}{Color for nodes. Default = #7F7F7F" (some gray)}

\item{leaf.col}{Color for leaf nodes. Default = "#18A3AC" (teal)}

\item{edge.col}{Color for edges. Default = "#848484" (another gray)}

\item{col.highlight}{Color for surrounding edges when node is selected. Default = "#F48024" (orange)}

\item{node.font.col}{Color for node labels. Default varies by \code{shape}, black or white depending if
\code{visNetwork} draws labels on node or underneath}

\item{edge.font.col}{Color for edge labels. Default = "#000000" (black)}

\item{sort.coefs}{Logical: If TRUE, sort each coefs table. Default = FALSE}

\item{height}{Float: Height for \code{visNetwork}. Default = NULL, i.e. auto}

\item{width}{Float: Width for \code{visNetwork}. Default = NULL, i.e. auto}

\item{levelSeparation}{Float: N of pixels to separate tree levels. Default = 100}

\item{tree.font.size}{Integer: Font size for tree labels. Default = 22}

\item{font.family}{String: Font to use throughout. Default = 'Helvetica Neue', because otherwise it will fail on a
number of external viewers, but feel free to play around, esp. within RStudio}

\item{tooltip.coefs}{Logical: If TRUE, show html coefficient tables on hover over nodes. This was placed here before
a custom html table creation function was made to replace some impossibly slow alternatives.}

\item{table.font.size}{String: Font size for html coefficient on-hover tables. Default = "14px"}

\item{table.dat.padding}{Ignore, has no visible effect. Otherwise, String: html table padding. Default = "0px"}

\item{table.lo.col}{Color for lowest coefficient values (negative). Default = "#80FFFF" (light blue)}

\item{table.hi.col}{Color for highest coefficient values (positive). Default = "#80FFFF" (light orange)}

\item{trace}{Integer: If > 0, print info to console (not particularly informative). Default = 0}

\item{edgethickness.by.ncases.}{Logical: If TRUE, scale edge thickness by number of cases with weight = 1}
}
\description{
Plot a stepwise hybrid tree trained by \code{s.SHYTREE} using \code{plotly}
}
