% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.shytreecoef.R
\name{dplot3.shytreecoef}
\alias{dplot3.shytreecoef}
\title{Coefficient heatmap for The Hybrid Tree}
\usage{
dplot3.shytreecoef(
  tree,
  Rowv = FALSE,
  Colv = FALSE,
  lo = "#2B27F1",
  mid = "black",
  hi = "#FFBE00"
)
}
\description{
Coefficient heatmap for The Hybrid Tree
}
\author{
Efstathios D. Gennatas
}
