% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.table.R
\name{dplot3.table}
\alias{dplot3.table}
\title{Simple HTML table}
\usage{
dplot3.table(
  x,
  .ddSci = TRUE,
  main = NULL,
  main.col = "black",
  main.x = 0,
  main.xanchor = "auto",
  fill.col = "#18A3AC",
  table.bg = "white",
  bg = "white",
  line.col = "white",
  lwd = 1,
  header.font.col = "white",
  table.font.col = "gray20",
  font.size = 14,
  font.family = "Helvetica Neue",
  margin = list(l = 0, r = 5, t = 30, b = 0, pad = 0)
)
}
\arguments{
\item{x}{data.frame: Table to draw}

\item{main}{Character: Table tile. Default = NULL}

\item{main.col}{Color: Title color. Default = "black"}

\item{main.x}{Float [0, 1]: Align title: 0: left, .5: center, 1: right. Default = 0}

\item{main.xanchor}{Character: "auto", "left", "right": plotly's layout xanchor for title. Default = "auto"}

\item{fill.col}{Color: Used to fill header with column names and first column with row names.
Default = "#18A3AC"  (teal)}

\item{table.bg}{Colot: Table background. Default = "white"}

\item{bg}{Color: Background. Default = "white"}

\item{line.col}{Color: Line color. Default = "white"}

\item{lwd}{Float: Line width. Default = 1}

\item{header.font.col}{Color: Header font color. Default = "white"}

\item{table.font.col}{Color: Table font color. Default = "gray20"}

\item{font.size}{Integer: Font size. Default = 14}

\item{font.family}{Character: Font family. Default = "Helvetica Neue"}

\item{margin}{List: plotly's margins. Default = \code{list(l = 0, r = 5, t = 30, b = 0, pad = 0)}}

\item{ddSci}{Logical: If TRUE, apply \link{ddSci} to numeric columns. Default = TRUE}
}
\description{
Draw an html table using \code{plotly}
}
\author{
E.D. Gennatas
}
