% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.varimp.R
\name{dplot3.varimp}
\alias{dplot3.varimp}
\title{Interactive Variable Importance Plot}
\usage{
dplot3.varimp(
  varimp,
  names = NULL,
  main = "Variable Importance",
  xlab = "",
  ylab = "",
  col = colorAdjust(colorGrad(length(names), "penn"), 0.8),
  mar = c(50, 110, 50, 50),
  pad = 10,
  font.family = "Helvetica Neue",
  font.size = 14,
  axis.font.size = 14,
  font.color = "000",
  showlegend = TRUE
)
}
\arguments{
\item{varimp}{Vector, float: Variable importance of features}

\item{names}{Vector, string: Names of features}

\item{main}{Character: Plot title}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{col}{Vector, colors: Single value, or multiple values to define bar (feature) color(s)}

\item{mar}{Vector, numeric, length 4: Plot margins in pixels (NOT inches).
Default = c(50, 110, 50, 50)}

\item{pad}{Integer: Pad plot by this many pixels. Default = 10}

\item{font.family}{Character: Font to use. Default = "Open Sans"}

\item{font.size}{Integer: Overall font size to use (essentially for the title at this point).
Default = 14}

\item{axis.font.size}{Integer: Font size to use for axis labels and tick labels
(Seems not to be in same scale as \code{font.size} for some reason. Experiment!)}

\item{font.color}{Color for all text}

\item{showlegend}{Logical: If TRUE, show legend}
}
\description{
Plot variable importance using \code{plotly}
}
\details{
A simple \code{plotly} wrapper to plot horizontal barplots, sorted by value, which can be used
to visualize variable importance, model coefficients, etc.
}
\author{
Efstathios D. Gennatas
}
