% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.xyz.R
\name{dplot3.xyz}
\alias{dplot3.xyz}
\title{Interactive 3D Plots}
\usage{
dplot3.xyz(
  x,
  y = NULL,
  z = NULL,
  fit = NULL,
  cluster = NULL,
  cluster.params = list(k = 2),
  group = NULL,
  formula = NULL,
  rsq = TRUE,
  mode = "markers",
  order.on.x = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  col = NULL,
  alpha = 0.8,
  bg = NULL,
  plot.bg = NULL,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  axes.square = FALSE,
  group.names = NULL,
  font.size = 16,
  font.alpha = 0.8,
  font.col = NULL,
  font.family = "Helvetica Neue",
  marker.col = NULL,
  fit.col = NULL,
  fit.alpha = 0.7,
  fit.lwd = 2.5,
  main.col = NULL,
  axes.col = NULL,
  labs.col = NULL,
  grid.col = NULL,
  grid.lwd = 1,
  grid.alpha = 0.8,
  tick.col = NULL,
  tick.font.size = 12,
  spike.col = NULL,
  legend = NULL,
  legend.xy = c(0, 1),
  legend.xanchor = "left",
  legend.yanchor = "auto",
  legend.orientation = "v",
  legend.col = NULL,
  legend.bg = "#FFFFFF00",
  legend.border.col = "#FFFFFF00",
  legend.borderwidth = 0,
  legend.group.gap = 0,
  margin = list(t = 30, b = 0, l = 0, r = 0),
  zerolines = TRUE,
  mod.params = list(),
  width = NULL,
  height = NULL,
  padding = 0,
  trace = 0,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  ...
)
}
\arguments{
\item{x}{Numeric, vector/data.frame/list: x-axis data. If y is NULL and \code{NCOL(x) > 1}, first two columns used as
\code{x} and \code{y}, respectively}

\item{y}{Numeric, vector/data.frame/list: y-axis data}

\item{z}{Numeric, vector/data.frame/list: z-axis data}

\item{fit}{Character: \pkg{rtemis} model to calculate y ~ x fit. Options: see \code{modSelect}
Can also be Logical, which will give a GAM fit if TRUE. If you specify "NLA", the activation function should
be passed as a string.}

\item{cluster}{Character: Clusterer name. Will cluster \code{data.frame(x, y)} and
pass result to \code{group}. Run \link{clustSelect} for options}

\item{cluster.params}{List: Names list of parameters to pass to the \code{cluster} function}

\item{group}{Vector: will be converted to factor.
If \code{data} is provided, name of variable, unquoted.}

\item{formula}{Formula: Provide a formula to be solved using \link{s.NLS}. If provided, \code{fit} is
forced to \code{'nls'}. e.g. y ~ b * m ^ x for a power curve. Note: \code{nls} is prone to errors
and warnings, but is powerful. Use single letters for parameter names, no numbers.}

\item{rsq}{Logical: If TRUE, print R-squared values in legend if \code{fit} is set}

\item{mode}{Character, vector: "markers", "lines", "markers+lines". Default = "markers"}

\item{order.on.x}{Logical: If TRUE, order \code{x} and \code{y} on \code{x}. Default = NULL, which becomes
\code{TRUE} if \code{mode} includes lines.}

\item{main}{Character: Plot title}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{zlab}{Character: z-axis label}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{palette}{Vector of colors, or Character defining a builtin palette - get all options with \code{rtPalette()}}

\item{axes.square}{Logical: If TRUE: draw a square plot to fill the graphic device.
Default = FALSE. Note: If TRUE, the device size at time of call is captured and height and width are set so as
to draw the largest square available. This means that resizing the device window will not automatically resize the
plot.}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{fit.col}{Color: Color of the fit line.}

\item{fit.lwd}{Float: Fit line width}

\item{legend}{Logical: If TRUE, draw legend. Default = NULL, which will be set to TRUE if there are more than 1
groups, or \code{fit} is set}

\item{legend.orientation}{Character: "v": vertical, or "h": horizontal. Default = "v"}

\item{margin}{Numeric, named list: Margins for top, bottom, left, right.
Default = \code{list(t = 30, b = 0, l = 0, r = 0}}

\item{zerolines}{Logical: If TRUE: draw lines at y = 0. Default = FALSE}

\item{width}{Float: Force plot size to this width. Default = NULL, i.e. fill available space}

\item{height}{Float: Force plot size to this height. Default = NULL, i.e. fill available space}

\item{trace}{Integer: If > 0, pass \code{verbose = TRUE} to the cluster and fit functions, if used. Default = 0}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{...}{Additional arguments to be passed to theme function}
}
\description{
Draw interactive 3D plots using \code{plotly}
}
\examples{
\dontrun{
dplot3.xyz(iris, group = iris$Species, theme = "darkgrid")
}
}
\author{
Efstathios D. Gennatas
}
