% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_addtree.R
\name{dplot3_addtree}
\alias{dplot3_addtree}
\title{Plot AddTree trees}
\usage{
dplot3_addtree(
  addtree,
  col.positive = "#F48024DD",
  col.negative = "#18A3ACDD",
  node.col = "#666666",
  node.shape = "none",
  node.labels = TRUE,
  node.labels.pct.pos = NULL,
  pos.name = NULL,
  edge.col = "#999999",
  edge.width = 2,
  edge.labels = FALSE,
  arrowhead = "vee",
  layout = "dot",
  drop.leaves = FALSE,
  rankdir = "TB",
  splines = "polyline",
  fontname = "helvetica",
  bg.color = "#ffffff",
  overlap = "false",
  prune = NULL,
  prune.empty.leaves = TRUE,
  remove.bad.parents = FALSE,
  filename = NULL
)
}
\arguments{
\item{col.positive}{Color for outcome positive.}

\item{col.negative}{Color for negative outcome.}

\item{node.col}{Color for non-terminal leaves.}

\item{edge.col}{Color for edges.}

\item{bg.color}{Background color.}

\item{filename}{Character: Path to filename to save PDF
Requires packages \code{DiagrammeRsvg} and \code{rsvg}. Default = NULL}
}
\description{
Plot AddTree trees trained with \link{s_AddTree} using \code{data.tree::plot.Node}
}
\details{
Edge info and styles have been removed because of problems with \code{DiagrammeR}
}
\author{
E.D. Gennatas
}
