% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_calibration.R, R/dplot3_calibration1.R
\name{dplot3_calibration}
\alias{dplot3_calibration}
\title{Draw calibration plot}
\usage{
dplot3_calibration(
  true_labels,
  est_prob,
  n_windows = 10,
  pos_class_idi = 1,
  xlab = "Mean estimated probability",
  ylab = "Empirical risk",
  mode = "markers+lines",
  ...
)

dplot3_calibration(
  true_labels,
  est_prob,
  n_windows = 10,
  pos_class_idi = 1,
  xlab = "Mean estimated probability",
  ylab = "Empirical risk",
  mode = "markers+lines",
  ...
)
}
\arguments{
\item{true_labels}{Factor with true class labels}

\item{est_prob}{Numeric vector with predicted probabilities}

\item{n_windows}{Integer: Number of windows to split the data into}

\item{pos_class_idi}{Integer: Index of the positive class}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{mode}{Character: Plot mode}

\item{...}{Additional arguments passed to \link{dplot3_xy}}
}
\description{
Draw calibration plot

Draw calibration plot
}
\examples{
\dontrun{
data(segment_logistic, package = "probably")

# Plot the calibration curve of the original predictions
dplot3_calibration(
    true_labels = segment_logistic$Class,
    est_prob = segment_logistic$.pred_poor,
    n_windows = 10,
    pos_class_idi = 2
)

# Plot the calibration curve of the calibrated predictions
dplot3_calibration(
    true_labels = segment_logistic$Class,
    est_prob = calibrate(
        segment_logistic$Class,
        segment_logistic$.pred_poor
    )$fitted.values,
    n_windows = 10,
    pos_class_idi = 2
)
}
\dontrun{
data(segment_logistic, package = "probably")

# Plot the calibration curve of the original predictions
dplot3_calibration(
    true_labels = segment_logistic$Class,
    est_prob = segment_logistic$.pred_poor,
    n_windows = 10,
    pos_class_idi = 2
)

# Plot the calibration curve of the calibrated predictions
dplot3_calibration(
    true_labels = segment_logistic$Class,
    est_prob = calibrate(segment_logistic$Class,
                         segment_logistic$.pred_poor)$fitted.values,
    n_windows = 10,
    pos_class_idi = 2
)
}
}
\author{
EDG
}
