% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_cart.R
\name{dplot3_cart}
\alias{dplot3_cart}
\title{Plot \code{rpart} decision trees}
\usage{
dplot3_cart(
  object,
  col.positive = "#F48024DD",
  col.negative = "#18A3ACDD",
  col.lo = "#80ffff",
  col.mid = "gray20",
  col.hi = "#F4A0FF",
  node.col = "#666666",
  node.shape = "none",
  node.labels = TRUE,
  node.cond = TRUE,
  node.prob = TRUE,
  node.estimate = NULL,
  node.n = TRUE,
  edge.col = "#999999",
  edge.width = 2,
  edge.labels = FALSE,
  arrowhead = "vee",
  layout = "dot",
  drop.leaves = FALSE,
  rankdir = "TB",
  splines = "polyline",
  fontname = "helvetica",
  bg.color = "white",
  overlap = "false",
  prune = FALSE,
  rpart.cp = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Either \code{rpart} object or \code{rtMod} object trained with
\link{s_CART}}

\item{col.positive}{Color for outcome positive.}

\item{col.negative}{Color for negative outcome.}

\item{col.lo}{Low color for estimated outcome}

\item{col.mid}{Middle color for estimated outcome}

\item{col.hi}{High color for estimated outcome}

\item{node.col}{Color for non-terminal leaves.}

\item{node.shape}{Shape of node. Default = "none"}

\item{node.labels}{Logical: If TRUE, print the node labels.}

\item{node.cond}{Logical: If TRUE, print the splitting condition inside each
node.}

\item{node.prob}{Logical: If TRUE, print the probability estimate for the
first class of the outcome inside each node.}

\item{node.estimate}{Logical: If TRUE, print the estimated outcome level
inside each node.}

\item{node.n}{Logical: If TRUE, print the number of cases (from training
data) that matched this condition}

\item{edge.col}{Color for edges.}

\item{edge.width}{Width of edges.}

\item{edge.labels}{Logical: If TRUE, print the splitting condition on the
edge.}

\item{arrowhead}{Character: Arrowhead shape.}

\item{layout}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{drop.leaves}{Logical: If TRUE, position leaves at the bottom of the plot.}

\item{rankdir}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{splines}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{fontname}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{bg.color}{Background color.}

\item{overlap}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{prune}{Logical: If TRUE, prune tree using \code{rpart::prune.rpart}}

\item{rpart.cp}{Numeric: Complexity parameter for pruning. If NULL, no
pruning is performed.}

\item{verbose}{Logical: If TRUE, print messages.}
}
\description{
Plot \code{rpart} decision trees using \code{data.tree::plot.Node}
}
\details{
If you want to show split conditions as edge labels (\code{edge.labels = TRUE}),
it is recommened to set \code{rankdir = "LR"} and \code{node.cond = FALSE}.
Edge labels in graphviz are shown to the right of
the edge when \code{rankdir = "TB"} and above when \code{rankdir = "LR"}.
}
\author{
E.D. Gennatas
}
