% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_conf.R
\name{dplot3_conf}
\alias{dplot3_conf}
\title{Plot confusion matrix}
\usage{
dplot3_conf(
  x,
  true.col = "#72CDF4",
  false.col = "#FEB2E0",
  pos.class = rtenv$binclasspos,
  font.size = 18,
  main = NULL,
  main.y = 1,
  main.yanchor = "bottom",
  theme = rtTheme,
  margin = list(l = 20, r = 5, b = 5, t = 20),
  filename = NULL,
  file.width = 500,
  file.height = 500,
  file.scale = 1,
  ...
)
}
\arguments{
\item{x}{Confusion matrix where rows are the reference and columns are the estimated classes}

\item{true.col}{Color for true positives & true negatives}

\item{false.col}{Color for false positives & false negatives}

\item{pos.class}{Integer: Index of factor level to treat as the positive class}

\item{font.size}{Integer: font size}

\item{main}{Character: plot title}

\item{main.y}{Numeric: y position of the title}

\item{main.yanchor}{Character: y anchor of the title}

\item{theme}{List or Character: Either the output of a \verb{theme_*()} function or the name of  a
theme. Use \code{themes()} to get available theme names. Theme functions are of the form
\verb{theme_<name>}.}

\item{margin}{List: Plot margins}

\item{filename}{Character: Path to file to save static plot. Default = NULL}

\item{file.width}{Integer: File width in pixels for when \code{filename} is
set.}

\item{file.height}{Integer: File height in pixels for when \code{filename}
is set.}

\item{file.scale}{Numeric: If saving to file, scale plot by this number}

\item{...}{Additional arguments passed to theme function.}
}
\value{
A plotly object
}
\description{
Plot confusion matrix
}
\examples{
\dontrun{
true_labels <- factor(c("a", "b", "a", "a", "b", "b", "b", "b"))
predicted_labels <- factor(c("a", "a", "b", "a", "b", "b", "a", "a"))
error <- mod_error(true_labels, predicted_labels)
dplot3_conf(error$ConfusionMatrix)
}
}
\author{
EDG
}
