% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_xy.R
\name{dplot3_fit}
\alias{dplot3_fit}
\title{True vs. Predicted Plot}
\usage{
dplot3_fit(x, y, fit = "gam", se.fit = TRUE, ...)
}
\arguments{
\item{x}{Numeric, vector/data.frame/list: True values. If y is NULL and
\code{NCOL(x) > 1}, first two columns used as \code{x} and \code{y}, respectively}

\item{y}{Numeric, vector/data.frame/list: Predicted values}

\item{fit}{Character: \pkg{rtemis} model to calculate \code{y ~ x} fit.
Options: see \code{select_learn()}
Can also be Logical, which will give a GAM fit if TRUE. If you specify
"NLA", the activation function should be passed as a string.}

\item{se.fit}{Logical: If TRUE, draw the standard error of the fit}

\item{...}{Additional arguments passed to \link{dplot3_xy}}
}
\description{
A \code{dplot3_xy} wrapper for plotting true vs. predicted values
}
\examples{
\dontrun{
x <- rnorm(500)
y <- x + rnorm(500)
dplot3_fit(x, y)
}
}
\author{
EDG
}
