% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_graphd3.R
\name{dplot3_graphd3}
\alias{dplot3_graphd3}
\title{Plot graph using \pkg{networkD3}}
\usage{
dplot3_graphd3(
  net,
  groups = NULL,
  color.scale = NULL,
  edge.col = NULL,
  node.col = NULL,
  node.alpha = 0.5,
  edge.alpha = 0.33,
  zoom = TRUE,
  legend = FALSE,
  palette = rtPalette,
  theme = rtTheme,
  ...
)
}
\arguments{
\item{net}{\pkg{igraph} network}

\item{groups}{Vector, length n nodes indicating group/cluster/community membership of nodes in
\code{net}}

\item{color.scale}{D3 colorscale (e.g. \code{networkD3::JS("d3.scaleOrdinal(d3.schemeCategory20b);")})}

\item{edge.col}{Color for edges}

\item{node.col}{Color for nodes}

\item{node.alpha}{Float [0, 1]: Node opacity. Default = .5}

\item{edge.alpha}{Float [0, 1]: Edge opacity. Default = .33}

\item{zoom}{Logical: If TRUE, graph is zoomable. Default = TRUE}

\item{legend}{Logical: If TRUE, display legend for groups}

\item{palette}{Vector of colors, or Character defining a builtin palette - get options with
\code{rtpalette()}}

\item{theme}{rtemis theme to use}

\item{...}{Additional arguments to pass to \code{networkD3}}
}
\description{
Plot graph using \pkg{networkD3}
}
\author{
E.D. Gennatas
}
