% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_linad.R
\name{dplot3_linad}
\alias{dplot3_linad}
\title{Plot a Linear Additive Tree trained by \link{s_LINAD} using
\emph{visNetwork}}
\usage{
dplot3_linad(
  x,
  main = NULL,
  bg = "#FFFFFF",
  shape = "box",
  nodelabels = TRUE,
  ncases.inlabels = TRUE,
  rules.on.edges = FALSE,
  top = NULL,
  root.col = "#202020",
  node.col = "#5a5a5a",
  leaf.col = "#178CCB",
  edge.col = "#848484",
  edge.width = 4,
  arrow.scale = 0.7,
  arrow.middle = FALSE,
  col.highlight = "#FE4AA3",
  node.font.col = NULL,
  edge.font.col = "#000000",
  sort.coefs = FALSE,
  height = NULL,
  width = NULL,
  levelSeparation = 100,
  tree.font.size = 22,
  edgethickness.by.ncases = FALSE,
  font.family = "Lato",
  uselog = FALSE,
  tooltip.coefs = TRUE,
  tooltip.delay = 50,
  table.font.size = "16px",
  table.dat.padding = "0px",
  table.lo.col = "#0290EE",
  table.hi.col = "#FE4AA3",
  dragNodes = FALSE,
  zoomView = FALSE,
  nodeSpacing = 150,
  blockShifting = TRUE,
  edgeMinimization = TRUE,
  parentCentralization = TRUE,
  direction = "UD",
  trace = 0
)
}
\arguments{
\item{x}{\code{rtMod} object trained using \link{s_LINAD}}

\item{main}{Character: Title.}

\item{bg}{Background color.}

\item{shape}{Character: Node shape; one of: "square", "triangle", "box", "circle",
"dot", "star", "ellipse", "database", "text", "diamond".}

\item{nodelabels}{Logical: If TRUE, inlcude node labels.}

\item{ncases.inlabels}{Logical: If TRUE, include number of cases with the
node labels.}

\item{rules.on.edges}{Logical: If TRUE, display rules on edges instead of
nodes.}

\item{top}{Integer: If not NULL, only show the top \code{top} coefficients.}

\item{root.col}{Color for root node.}

\item{node.col}{Color for nodes.}

\item{leaf.col}{Color for leaf nodes.}

\item{edge.col}{Color for edges.}

\item{edge.width}{Numeric: Width for edges.}

\item{arrow.scale}{Numeric: Scale factor for arrows.}

\item{arrow.middle}{Logical: If TRUE, draw arrows in the middle of edges.}

\item{col.highlight}{Color for surrounding edges when node is selected.}

\item{node.font.col}{Color for node labels. Default varies by \code{shape},
black or white depending if
\code{visNetwork} draws labels on node or underneath}

\item{edge.font.col}{Color for edge labels.}

\item{sort.coefs}{Logical: If TRUE, sort each coefs table.}

\item{height}{Numeric: Height for \code{visNetwork}. Default = NULL, i.e. auto}

\item{width}{Numeric: Width for \code{visNetwork}. Default = NULL, i.e. auto}

\item{levelSeparation}{Numeric: N of pixels to separate tree levels.}

\item{tree.font.size}{Integer: Font size for tree labels. Default = 22}

\item{edgethickness.by.ncases}{Logical: If TRUE, scale edge thickness by
number of cases with weight = 1}

\item{font.family}{Character: Font to use throughout.
Default = 'Helvetica Neue', because otherwise it may fail on a
number of external viewers.}

\item{uselog}{Logical: If TRUE, use log10 scale for coefficient colors.}

\item{tooltip.coefs}{Logical: If TRUE, show html coefficient tables on hover
over nodes. This was placed here before a custom html table creation
function was made to replace some impossibly slow alternatives.}

\item{tooltip.delay}{Numeric: Delay (in milliseconds) on mouse over before
showing tooltip.}

\item{table.font.size}{Character: Font size for html coefficient on-hover
tables.}

\item{table.dat.padding}{Ignore, has no visible effect. Otherwise,
Character: html table padding.}

\item{table.lo.col}{Color for lowest coefficient values (negative)}

\item{table.hi.col}{Color for highest coefficient values (positive).}

\item{dragNodes}{Logical: If TRUE, allow dragging nodes.}

\item{zoomView}{Logical: If TRUE, allow zooming.}

\item{nodeSpacing}{Numeric: Spacing between nodes.}

\item{blockShifting}{Logical: If TRUE, allow block shifting.}

\item{edgeMinimization}{Logical: If TRUE, minimize edge length.}

\item{parentCentralization}{Logical: If TRUE, centralize parent nodes.}

\item{direction}{Character: Direction of tree. One of: "UD", "DU", "LR",
"RL".}

\item{trace}{Integer: If > 0, print info to console (not particularly
informative).}
}
\description{
Plot a Linear Additive Tree trained by \link{s_LINAD} using
\emph{visNetwork}
}
\author{
E.D. Gennatas
}
