% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_pie.R
\name{dplot3_pie}
\alias{dplot3_pie}
\title{Interactive Pie Chart}
\usage{
dplot3_pie(
  x,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  alpha = 0.8,
  bg = NULL,
  plot.bg = NULL,
  theme = getOption("rt.theme", "black"),
  palette = rtPalette,
  category.names = NULL,
  textinfo = "label+percent",
  font.size = 16,
  labs.col = NULL,
  legend = TRUE,
  legend.col = NULL,
  sep.col = NULL,
  margin = list(b = 50, l = 50, t = 50, r = 20),
  padding = 0,
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  filename = NULL,
  file.width = 500,
  file.height = 500,
  file.scale = 1,
  ...
)
}
\arguments{
\item{x}{data.frame: Input: Either a) 1 numeric column with categories defined by rownames, or
b) two columns, the first is category names, the second numeric or c) a numeric vector with categories defined using
the \code{category.names} argument}

\item{main}{Character: Plot title. Default = NULL, which results in colnames(x)[1],}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{col}{Color, vector: Color for bars. Default NULL, which will draw
colors from \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for bar colors. Default = .8}

\item{bg}{Background color}

\item{plot.bg}{Plot background color}

\item{theme}{Character: "light", "dark". Default = \code{getOption("rt.theme", "light")}}

\item{palette}{Character: Name of \pkg{rtemis} palette to use.
Default = "rtCol1". Only used if \code{col = NULL}}

\item{category.names}{Character, vector, length = NROW(x): Category names. Default = NULL, which uses
either \code{rownames(x)}, or the first column of \code{x} if \code{ncol(x) = 2}}

\item{textinfo}{Character: Info to show over each slince: "label", "percent", "label+percent" Default = "label+percent"}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{labs.col}{Color of labels}

\item{legend}{Logical: If TRUE, draw legend. Default = NULL, and will be
turned on if there is more than one feature present}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by
theme}

\item{sep.col}{Separator color}

\item{margin}{Named list: plot margins.}

\item{padding}{Integer: N pixels to pad plot.}

\item{displayModeBar}{Logical: If TRUE, show plotly's modebar}

\item{modeBar.file.format}{Character: "svg", "png", "jpeg", "pdf" / any
output file type supported by plotly and your system}

\item{filename}{Character: Path to file to save static plot. Default = NULL}

\item{file.width}{Integer: File width in pixels for when \code{filename} is
set.}

\item{file.height}{Integer: File height in pixels for when \code{filename}
is set.}

\item{file.scale}{Numeric: If saving to file, scale plot by this number}

\item{...}{Additional arguments passed to theme}
}
\description{
Draw interactive pie charts using \code{plotly}
}
\examples{
\dontrun{
dplot3_pie(VADeaths[, 1, drop = F])
}
}
\author{
E.D. Gennatas
}
