% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_shytreecoef.R
\name{dplot3_shytreecoef}
\alias{dplot3_shytreecoef}
\title{Coefficient heatmap for The Hybrid Tree}
\usage{
dplot3_shytreecoef(
  tree,
  Rowv = FALSE,
  Colv = FALSE,
  leaf.col = "#18A3AC",
  lo = "#2B27F1",
  mid = "black",
  hi = "#FFBE00"
)
}
\description{
Coefficient heatmap for The Hybrid Tree
}
\author{
E.D. Gennatas
}
