% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_spectrogram.R
\name{dplot3_spectrogram}
\alias{dplot3_spectrogram}
\title{Interactive Spectrogram}
\usage{
dplot3_spectrogram(
  x,
  y,
  z,
  colorGrad.n = 101,
  colors = NULL,
  xlab = "Time",
  ylab = "Frequency",
  zlab = "Power",
  zmin = NULL,
  zmax = NULL,
  zauto = TRUE,
  colorscale = "Jet",
  colorbar.y = 0.5,
  colorbar.yanchor = "middle",
  colorbar.xpad = 0,
  colorbar.ypad = 0,
  colorbar.len = 0.75,
  colorbar.title.side = "bottom",
  showgrid = FALSE,
  space = "rgb",
  lo = "#18A3AC",
  lomid = NULL,
  mid = NULL,
  midhi = NULL,
  hi = "#F48024",
  k_row = 1,
  k_col = 1,
  grid.gap = 0,
  limits = NULL,
  margins = c(50, 50, 50, 50),
  main = NULL,
  key.title = NULL,
  showticklabels = NULL,
  theme = rtTheme,
  row_side_colors = NULL,
  row_side_palette = NULL,
  col_side_colors = NULL,
  col_side_palette = NULL,
  font.size = NULL,
  padding = 0,
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  filename = NULL,
  file.width = 500,
  file.height = 500,
  file.scale = 1,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{colorGrad.n}{Integer: Number of distinct colors to generate using \link{colorGrad}. Default = 101}

\item{colors}{Character: Acts as a shortcut to defining \code{lo}, \code{mid}, etc for a number of defaults:
"french", "penn", "grnblkred",}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{space}{Character: Which colorspace to use. Option: "rgb", or "Lab". Default = "rgb".
Recommendation: If \code{mid} is "white" or "black" (default), use "rgb", otherwise "Lab"}

\item{lo}{Color for low end}

\item{lomid}{Color for low-mid}

\item{mid}{Color for middle of the range or "mean", which will result in \code{colorOp(c(lo, hi), "mean")}.
If \code{mid = NA}, then only \code{lo} and \code{hi} are used to create the color gradient.}

\item{midhi}{Color for middle-high}

\item{hi}{Color for high end}

\item{grid.gap}{Integer: Space between cells. Default = 0 (no space)}

\item{limits}{Float, length 2: Determine color range. Default = NULL, which automatically centers values around 0}

\item{margins}{Float, length 4: Heatmap margins.}

\item{main}{Character: Plot title}

\item{theme}{Character: "light", "dark"}

\item{font.size}{Numeric: Font size}

\item{padding}{Numeric: Padding between cells}

\item{displayModeBar}{Logical: If TRUE, display the plotly mode bar}

\item{modeBar.file.format}{Character: File format for image exports from the mode bar}

\item{filename}{String (Optional: Path to file to save colorbar}

\item{file.width}{Numeric: Width of exported image}

\item{file.height}{Numeric: Height of exported image}

\item{file.scale}{Numeric: Scale of exported image}

\item{...}{Additional arguments to be passed to \code{heatmaply::heatmaply}}

\item{cellnote}{Matrix with values to be desplayed on hover. Defaults to \code{ddSci(x)}}
}
\description{
Draw interactive spectrograms using \code{plotly}
}
\details{
To set custom colors, use a minimum of \code{lo} and \code{hi}, optionnaly also
\code{lomid}, \code{mid}, \code{midhi} colors and set \code{colorscale = NULL}.
}
\author{
E.D. Gennatas
}
