% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_varimp.R
\name{dplot3_varimp}
\alias{dplot3_varimp}
\title{Interactive Variable Importance Plot}
\usage{
dplot3_varimp(
  x,
  names = NULL,
  main = NULL,
  xlab = "Variable Importance",
  ylab = "",
  plot.top = 1,
  labelify = TRUE,
  col = NULL,
  alpha = 1,
  palette = NULL,
  mar = NULL,
  font.size = 16,
  axis.font.size = 14,
  theme = rtTheme,
  showlegend = TRUE,
  ...
)
}
\arguments{
\item{x}{Vector, numeric: Input}

\item{names}{Vector, string: Names of features}

\item{main}{Character: main title}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{plot.top}{Float or Integer: If <= 1, plot this percent highest absolute values, otherwise plot this many top values.
i.e.: \code{plot.top = .2} will print the top 20\% highest values, and \code{plot.top = 20} will plot the top 20
highest values}

\item{labelify}{Logical: If TRUE convert \code{names(x)} using \link{labelify}. Default = TRUE}

\item{col}{Vector, colors: Single value, or multiple values to define bar
(feature) color(s)}

\item{alpha}{Numeric: Transparency}

\item{palette}{Character: Name of \pkg{rtemis} palette to use.}

\item{mar}{Vector, numeric, length 4: Plot margins in pixels (NOT inches).
Default = c(50, 110, 50, 50)}

\item{font.size}{Integer: Overall font size to use (essentially for the
title at this point).
Default = 14}

\item{axis.font.size}{Integer: Font size to use for axis labels and tick labels
(Seems not to be in same scale as \code{font.size} for some reason.
Experiment!)}

\item{theme}{Output of an rtemis theme function (list of parameters) or theme
name. Use \code{themes()} to print available themes.}

\item{showlegend}{Logical: If TRUE, show legend}

\item{...}{Additional arguments passed to theme}
}
\description{
Plot variable importance using \code{plotly}
}
\details{
A simple \code{plotly} wrapper to plot horizontal barplots, sorted by value,
which can be used to visualize variable importance, model coefficients, etc.
}
\examples{
# made-up data
x <- rnorm(10)
names(x) <- paste0("Feature_", seq(x))
dplot3_varimp(x)
}
\author{
E.D. Gennatas
}
