% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_volcano1.R
\name{dplot3_volcano1}
\alias{dplot3_volcano1}
\title{Volcano Plot}
\usage{
dplot3_volcano1(
  x,
  pvals,
  xnames = NULL,
  x.thresh = 0,
  p.thresh = 0.05,
  p.transform = function(x) -log10(x),
  p.adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  legend.lo = NULL,
  legend.hi = NULL,
  label.lo = "Low",
  label.hi = "High",
  xlab = NULL,
  ylab = NULL,
  margin = list(b = 65, l = 65, t = 50, r = 10, pad = 0),
  xlim = NULL,
  ylim = NULL,
  alpha = 0.5,
  theme = getOption("rt.theme"),
  font.size = 16,
  palette = list("#18A3AC", "#7f7f7f", "#F48024"),
  legend.x.lo = NULL,
  legend.x.hi = NULL,
  legend.y = 0.97,
  annotate.n = 7,
  ay.lo = NULL,
  ay.hi = NULL,
  annotate = TRUE,
  annotate.alpha = 0.7,
  hovertext = NULL,
  displayModeBar = FALSE,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector: Input values, e.g. log2 fold change, coefficients, etc.}

\item{pvals}{Numeric vector: p-values}

\item{p.thresh}{Numeric: p-value threshold of significance. Default = .05}

\item{p.transform}{function. Default = \code{\(x) -log10(x)}}

\item{p.adjust.method}{Character: p-value adjustment method.
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"
Default = "holm". Use "none" for raw p-values.}

\item{legend.lo}{Character: Legend to annotate significant points below the
\code{x.thresh}}

\item{legend.hi}{Character: Legend to annotate significant points above the
\code{x.thresh}}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{xlim}{Numeric vector, length 2: x-axis limits}

\item{ylim}{Numeric vector, length 2: y-axis limits}

\item{alpha}{Numeric: point transparency}
}
\description{
Volcano Plot
}
\examples{
\dontrun{
set.seed(2019)
x <- rnormmat(500, 500)
y <- x[, 3] + x[, 5] + x[, 9] + x[, 15] + rnorm(500)
mod <- massGLM(y, x)
dplot3_volcano(mod$summary$`Coefficient y`, mod$summary$`p_value y`)
}

}
\author{
E.D. Gennatas
}
