% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_x.R
\name{dplot3_x}
\alias{dplot3_x}
\title{Interactive Univariate Plots}
\usage{
dplot3_x(
  x,
  type = c("density", "histogram"),
  mode = c("overlap", "ridge"),
  group = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  alpha = 0.75,
  plot.bg = NULL,
  theme = rtTheme,
  palette = rtPalette,
  axes.square = FALSE,
  group.names = NULL,
  font.size = 16,
  font.alpha = 0.8,
  legend = NULL,
  legend.xy = c(0, 1),
  legend.col = NULL,
  legend.bg = "#FFFFFF00",
  legend.border.col = "#FFFFFF00",
  bargap = 0.05,
  vline = NULL,
  vline.col = theme$fg,
  vline.width = 1,
  vline.dash = "dot",
  text = NULL,
  text.x = 1,
  text.xref = "paper",
  text.xanchor = "left",
  text.y = 1,
  text.yref = "paper",
  text.yanchor = "top",
  text.col = theme$fg,
  margin = list(b = 65, l = 65, t = 50, r = 10, pad = 0),
  automargin.x = TRUE,
  automargin.y = TRUE,
  zerolines = FALSE,
  density.kernel = "gaussian",
  density.bw = "SJ",
  histnorm = c("", "density", "percent", "probability", "probability density"),
  histfunc = c("count", "sum", "avg", "min", "max"),
  hist.n.bins = 20,
  barmode = "overlay",
  ridge.sharex = TRUE,
  ridge.y.labs = FALSE,
  ridge.order.on.mean = TRUE,
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  width = NULL,
  height = NULL,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  file.scale = 1,
  ...
)
}
\arguments{
\item{x}{Numeric, vector / data.frame /list: Input. If not a vector, each
column of or each element}

\item{type}{Character: "density" or "histogram"}

\item{mode}{Character: "overlap", "ridge". How to plot different groups;
on the same axes ("overlap"), or on separate plots with the same x-axis
("ridge")}

\item{group}{Vector: Will be converted to factor; levels define group
members. Default = NULL}

\item{main}{Character: Main plot title.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{col}{Color, vector: Color for bars. Default NULL, which will draw
colors from \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for bar colors. Default = .8}

\item{plot.bg}{Color: Background color for plot area}

\item{theme}{List or Character: Either the output of a \verb{theme_*()} function or the name of  a
theme. Use \code{themes()} to get available theme names. Theme functions are of the form
\verb{theme_<name>}.}

\item{palette}{Character: Name of \pkg{rtemis} palette to use.
Default = "rtCol1". Only used if \code{col = NULL}}

\item{axes.square}{Logical: If TRUE: draw a square plot to fill the graphic
device. Default = FALSE. Note: If TRUE, the device size at time of call is
captured and height and width are set so as to draw the largest square
available. This means that resizing the device window will not automatically
resize the plot.}

\item{group.names}{Character, vector, length = NROW(x): Group names.
Default = NULL, which uses \code{rownames(x)}}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{font.alpha}{Float: Alpha transparency for font.}

\item{legend}{Logical: If TRUE, draw legend. Default = NULL, which will be
set to TRUE if x is a list of more than 1 element}

\item{legend.xy}{Float, vector, length 2: Relative x, y position for legend.
Default = c(0, 1), which places the legend top left within the plot area.
Set to NULL to place legend top right beside the plot area}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by
theme}

\item{legend.bg}{Color: Background color for legend}

\item{legend.border.col}{Color: Border color for legend}

\item{bargap}{Float: The gap between adjacent histogram bars in plot
fraction.}

\item{vline}{Float, vector: If defined, draw a vertical line at this x
value(s). Default = NULL}

\item{vline.col}{Color for \code{vline}. Default = theme$fg}

\item{vline.width}{Float: Width for \code{vline}. Default = 1}

\item{vline.dash}{Character: Type of line to draw: "solid", "dot", "dash",
"longdash", "dashdot",
or "longdashdot"}

\item{text}{Character: If defined, add this text over the plot}

\item{text.x}{Float: x-coordinate for \code{text}}

\item{text.xref}{Character: "x": \code{text.x} refers to plot's x-axis;
"paper": \code{text.x} refers to plotting area from 0-1}

\item{text.xanchor}{Character: "auto", "left", "center", "right"}

\item{text.y}{Float: y-coordinate for \code{text}}

\item{text.yref}{Character: "y": \code{text.y} refers to plot's y-axis;
"paper": \code{text.y} refers to plotting area from 0-1}

\item{text.yanchor}{Character: "auto", "top", "middle", "bottom"}

\item{text.col}{Color for \code{text}. Default = theme$fg}

\item{margin}{Named list: plot margins.}

\item{automargin.x}{Logical: If TRUE, automatically set x-axis amrgins}

\item{automargin.y}{Logical: If TRUE, automatically set y-axis amrgins}

\item{zerolines}{Logical: If TRUE, draw lines at y = 0.}

\item{density.kernel}{Character: Kernel to use for density estimation.}

\item{density.bw}{Character: Bandwidth to use for density estimation.}

\item{histnorm}{Character: NULL, "percent", "probability", "density",
"probability density"}

\item{histfunc}{Character: "count", "sum", "avg", "min", "max".}

\item{hist.n.bins}{Integer: Number of bins to use if type = "histogram".}

\item{barmode}{Character: Type of bar plot to make: "group", "relative",
"stack", "overlay". Default = "group". Use
"relative" for stacked bars, wich handles negative values correctly,
unlike "stack", as of writing.}

\item{ridge.sharex}{Logical: If TRUE, draw single x-axis when
\code{mode = "ridge"}.}

\item{ridge.y.labs}{Lofical: If TRUE, show individual y labs when
\code{mode = "ridge"}.}

\item{ridge.order.on.mean}{Logical: If TRUE, order groups by mean value
when \code{mode = "ridge"}.
Turn to FALSE, if, for example, groups are ordered by date or similar.}

\item{displayModeBar}{Logical: If TRUE, show plotly's modebar}

\item{modeBar.file.format}{Character: "svg", "png", "jpeg", "pdf" / any
output file type supported by plotly and your system}

\item{width}{Float: Force plot size to this width. Default = NULL, i.e. fill
available space}

\item{height}{Float: Force plot size to this height. Default = NULL, i.e.
fill available space}

\item{filename}{Character: Path to file to save static plot.}

\item{file.width}{Integer: File width in pixels for when \code{filename} is
set.}

\item{file.height}{Integer: File height in pixels for when \code{filename}
is set.}

\item{file.scale}{Numeric: If saving to file, scale plot by this number}

\item{...}{Additional arguments passed to theme function.}
}
\value{
A plotly object
}
\description{
Draw interactive univariate plots using \code{plotly}
}
\details{
If input is data.frame, non-numeric variables will be removed
}
\examples{
\dontrun{
dplot3_x(iris)
dplot3_x(split(iris$Sepal.Length, iris$Species), xlab = "Sepal Length")
}

}
\author{
E.D. Gennatas
}
