% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drange.R
\name{drange}
\alias{drange}
\title{Set Dynamic Range}
\usage{
drange(x, lo = 0, hi = 1, byCol = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame: Input}

\item{lo}{Target range minimum. Defaults to 0}

\item{hi}{Target range maximum. Defaults to 1}

\item{byCol}{Logical: If TRUE: if \code{x} is matrix, \code{drange} each
column separately}
}
\description{
\verb{rtemis preproc}: Adjusts the dynamic range of a vector or matrix input.
By default normalizes to 0-1 range.
}
\examples{
x <- runif(20, -10, 10)
x <- drange(x)
}
\author{
E.D. Gennatas
}
