% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_conf.R
\name{draw_confusion}
\alias{draw_confusion}
\title{Plot confusion matrix}
\usage{
draw_confusion(
  x,
  xlab = "Predicted",
  ylab = "Reference",
  true_col = "#72CDF4",
  false_col = "#FEB2E0",
  font_size = 18,
  main = NULL,
  main_y = 1,
  main_yanchor = "bottom",
  theme = choose_theme(),
  margin = list(l = 20, r = 5, b = 5, t = 20),
  filename = NULL,
  file_width = 500,
  file_height = 500,
  file_scale = 1
)
}
\arguments{
\item{x}{\code{ClassificationMetrics} object produced by \link{classification_metrics} or confusion matrix
where rows are the reference and columns are the estimated classes. For binary classification,
the first row and column are the positive class.}

\item{xlab}{Character: x-axis label. Default is "Predicted".}

\item{ylab}{Character: y-axis label. Default is "Reference".}

\item{true_col}{Color for true positives & true negatives.}

\item{false_col}{Color for false positives & false negatives.}

\item{font_size}{Integer: font size.}

\item{main}{Character: plot title.}

\item{main_y}{Numeric: y position of the title.}

\item{main_yanchor}{Character: y anchor of the title.}

\item{theme}{Theme object.}

\item{margin}{List: Plot margins.}

\item{filename}{Character: file name to save the plot. Default is NULL.}

\item{file_width}{Numeric: width of the file. Default is 500.}

\item{file_height}{Numeric: height of the file. Default is 500.}

\item{file_scale}{Numeric: scale of the file. Default is 1.}
}
\value{
\code{plotly} object.
}
\description{
Plot confusion matrix
}
\examples{
\dontrun{
# Assume positive class is "b"
true_labels <- factor(c("a", "a", "a", "b", "b", "b", "b", "b", "b", "b"))
predicted_labels <- factor(c("a", "b", "a", "b", "b", "a", "b", "b", "b", "a"))
predicted_prob <- c(0.3, 0.55, 0.45, 0.75, 0.57, 0.3, 0.8, 0.63, 0.62, 0.39)
metrics <- classification_metrics(true_labels, predicted_labels, predicted_prob)
draw_confusion(metrics)
}
}
\author{
EDG
}
