% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dist.R
\name{draw_dist}
\alias{draw_dist}
\title{Draw Distributions using Histograms and Density Plots}
\usage{
draw_dist(
  x,
  type = c("density", "histogram"),
  mode = c("overlap", "ridge"),
  group = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  alpha = 0.75,
  plot_bg = NULL,
  theme = choose_theme(),
  palette = rtemis_palette,
  axes_square = FALSE,
  group_names = NULL,
  font_size = 16,
  font_alpha = 0.8,
  legend = NULL,
  legend_xy = c(0, 1),
  legend_col = NULL,
  legend_bg = "#FFFFFF00",
  legend_border_col = "#FFFFFF00",
  bargap = 0.05,
  vline = NULL,
  vline_col = theme[["fg"]],
  vline_width = 1,
  vline_dash = "dot",
  text = NULL,
  text_x = 1,
  text_xref = "paper",
  text_xanchor = "left",
  text_y = 1,
  text_yref = "paper",
  text_yanchor = "top",
  text_col = theme[["fg"]],
  margin = list(b = 65, l = 65, t = 50, r = 10, pad = 0),
  automargin_x = TRUE,
  automargin_y = TRUE,
  zerolines = FALSE,
  density_kernel = "gaussian",
  density_bw = "SJ",
  histnorm = c("", "density", "percent", "probability", "probability density"),
  histfunc = c("count", "sum", "avg", "min", "max"),
  hist_n_bins = 20,
  barmode = "overlay",
  ridge_sharex = TRUE,
  ridge_y_labs = FALSE,
  ridge_order_on_mean = TRUE,
  displayModeBar = TRUE,
  modeBar_file_format = "svg",
  width = NULL,
  height = NULL,
  filename = NULL,
  file_width = 500,
  file_height = 500,
  file_scale = 1
)
}
\arguments{
\item{x}{Numeric vector / data.frame / list: Input. If not a vector, each column / each element is drawn.}

\item{type}{Character: "density" or "histogram".}

\item{mode}{Character: "overlap", "ridge". How to plot different groups; on the same axes ("overlap"), or on separate plots with the same x-axis ("ridge").}

\item{group}{Vector: Will be converted to factor; levels define group members.}

\item{main}{Character: Main title for the plot.}

\item{xlab}{Character: Label for the x-axis.}

\item{ylab}{Character: Label for the y-axis.}

\item{col}{Color: Colors for the plot.}

\item{alpha}{Numeric: Alpha transparency for plot elements.}

\item{plot_bg}{Color: Background color for plot area.}

\item{theme}{Theme object.}

\item{palette}{Character: Color palette to use.}

\item{axes_square}{Logical: If TRUE, draw a square plot to fill the graphic device. Default = FALSE.}

\item{group_names}{Character: Names for the groups.}

\item{font_size}{Numeric: Font size for plot text.}

\item{font_alpha}{Numeric: Alpha transparency for font.}

\item{legend}{Logical: If TRUE, draw legend. Default = NULL, which will be set to TRUE if x is a list of more than 1 element.}

\item{legend_xy}{Numeric, vector, length 2: Relative x, y position for legend. Default = c(0, 1).}

\item{legend_col}{Color: Color for the legend text.}

\item{legend_bg}{Color: Background color for legend.}

\item{legend_border_col}{Color: Border color for legend.}

\item{bargap}{Numeric: The gap between adjacent histogram bars in plot fraction.}

\item{vline}{Numeric, vector: If defined, draw a vertical line at this x value(s).}

\item{vline_col}{Color: Color for \code{vline}.}

\item{vline_width}{Numeric: Width for \code{vline}.}

\item{vline_dash}{Character: Type of line to draw: "solid", "dot", "dash", "longdash", "dashdot", or "longdashdot".}

\item{text}{Character: If defined, add this text over the plot.}

\item{text_x}{Numeric: x-coordinate for \code{text}.}

\item{text_xref}{Character: "x": \code{text_x} refers to plot's x-axis; "paper": \code{text_x} refers to plotting area from 0-1.}

\item{text_xanchor}{Character: "auto", "left", "center", "right".}

\item{text_y}{Numeric: y-coordinate for \code{text}.}

\item{text_yref}{Character: "y": \code{text_y} refers to plot's y-axis; "paper": \code{text_y} refers to plotting area from 0-1.}

\item{text_yanchor}{Character: "auto", "top", "middle", "bottom".}

\item{text_col}{Color: Color for \code{text}.}

\item{margin}{List: Margins for the plot.}

\item{automargin_x}{Logical: If TRUE, automatically adjust x-axis margins.}

\item{automargin_y}{Logical: If TRUE, automatically adjust y-axis margins.}

\item{zerolines}{Logical: If TRUE, draw lines at y = 0.}

\item{density_kernel}{Character: Kernel to use for density estimation.}

\item{density_bw}{Character: Bandwidth to use for density estimation.}

\item{histnorm}{Character: NULL, "percent", "probability", "density", "probability density".}

\item{histfunc}{Character: "count", "sum", "avg", "min", "max".}

\item{hist_n_bins}{Integer: Number of bins to use if type = "histogram".}

\item{barmode}{Character: Barmode for histogram. Default = "overlay".}

\item{ridge_sharex}{Logical: If TRUE, draw single x-axis when \code{mode = "ridge"}.}

\item{ridge_y_labs}{Logical: If TRUE, show individual y labels when \code{mode = "ridge"}.}

\item{ridge_order_on_mean}{Logical: If TRUE, order groups by mean value when \code{mode = "ridge"}.}

\item{displayModeBar}{Logical: If TRUE, display the mode bar.}

\item{modeBar_file_format}{Character: File format for mode bar. Default = "svg".}

\item{width}{Numeric: Force plot size to this width. Default = NULL, i.e. fill available space.}

\item{height}{Numeric: Force plot size to this height. Default = NULL, i.e. fill available space.}

\item{filename}{Character: Path to file to save static plot.}

\item{file_width}{Integer: File width in pixels for when \code{filename} is set.}

\item{file_height}{Integer: File height in pixels for when \code{filename} is set.}

\item{file_scale}{Numeric: If saving to file, scale plot by this number.}
}
\value{
A plotly object
}
\description{
Draw Distributions using Histograms and Density Plots using \code{plotly}.
}
\details{
If input is data.frame, non-numeric variables will be removed.
}
\examples{
\dontrun{
draw_dist(iris)
draw_dist(split(iris[["Sepal.Length"]], iris[["Species"]]), xlab = "Sepal Length")
}

}
\author{
EDG
}
