% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_scatter.R
\name{draw_fit}
\alias{draw_fit}
\title{True vs. Predicted Plot}
\usage{
draw_fit(
  x,
  y,
  fit = "gam",
  se_fit = TRUE,
  axes_square = TRUE,
  diagonal = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric, vector/data.frame/list: True values. If y is NULL and
\code{NCOL(x) > 1}, first two columns used as \code{x} and \code{y}, respectively}

\item{y}{Numeric, vector/data.frame/list: Predicted values}

\item{fit}{Character: Fit method.}

\item{se_fit}{Logical: If TRUE, include standard error of the fit.}

\item{axes_square}{Logical: If TRUE, draw a square plot.}

\item{diagonal}{Logical: If TRUE, add diagonal line.}

\item{...}{Additional arguments passed to \link{draw_scatter}}
}
\description{
A \code{draw_scatter} wrapper for plotting true vs. predicted values
}
\examples{
\dontrun{
x <- rnorm(500)
y <- x + rnorm(500)
draw_fit(x, y)
}
}
\author{
EDG
}
