% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_graphd3.R
\name{draw_graphD3}
\alias{draw_graphD3}
\title{Plot graph using \pkg{networkD3}}
\usage{
draw_graphD3(
  net,
  groups = NULL,
  color_scale = NULL,
  edge_col = NULL,
  node_col = NULL,
  node_alpha = 0.5,
  edge_alpha = 0.33,
  zoom = TRUE,
  legend = FALSE,
  palette = rtemis_palette,
  theme = choose_theme(),
  ...
)
}
\arguments{
\item{net}{\pkg{igraph} network.}

\item{groups}{Vector, length n nodes indicating group/cluster/community membership of nodes in \code{net}.}

\item{color_scale}{D3 colorscale (e.g. \code{networkD3::JS("d3.scaleOrdinal(d3.schemeCategory20b);")}).}

\item{edge_col}{Color for edges.}

\item{node_col}{Color for nodes.}

\item{node_alpha}{Float [0, 1]: Node opacity.}

\item{edge_alpha}{Float [0, 1]: Edge opacity.}

\item{zoom}{Logical: If TRUE, graph is zoomable.}

\item{legend}{Logical: If TRUE, display legend for groups.}

\item{palette}{Vector of colors, or Character defining a builtin palette - get options with \code{rtpalette()}.}

\item{theme}{Theme object.}

\item{...}{Additional arguments to pass to \code{networkD3}.}
}
\value{
\code{forceNetwork} object.
}
\description{
Plot graph using \pkg{networkD3}
}
\author{
EDG
}
