% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_graphjs.R
\name{draw_graphjs}
\alias{draw_graphjs}
\title{Plot network using \pkg{threejs::graphjs}}
\usage{
draw_graphjs(
  net,
  vertex_size = 1,
  vertex_col = NULL,
  vertex_label_col = NULL,
  vertex_label_alpha = 0.66,
  vertex_frame_col = NA,
  vertex_label = NULL,
  vertex_shape = "circle",
  edge_col = NULL,
  edge_alpha = 0.5,
  edge_curved = 0.35,
  edge_width = 2,
  layout = c("fr", "dh", "drl", "gem", "graphopt", "kk", "lgl", "mds", "sugiyama"),
  coords = NULL,
  layout_params = list(),
  cluster = NULL,
  groups = NULL,
  cluster_params = list(),
  cluster_mark_groups = TRUE,
  cluster_color_vertices = FALSE,
  main = "",
  theme = choose_theme(),
  palette = rtemis_palette,
  mar = rep(0, 4),
  par_reset = TRUE,
  filename = NULL,
  verbosity = 1L,
  ...
)
}
\arguments{
\item{net}{\code{igraph} network.}

\item{vertex_size}{Numeric: Vertex size.}

\item{vertex_col}{Color for vertices.}

\item{vertex_label_col}{Color for vertex labels.}

\item{vertex_label_alpha}{Numeric: Transparency for \code{vertex_label_col}.}

\item{vertex_frame_col}{Color for vertex border (frame).}

\item{vertex_label}{Character vector: Vertex labels. Default = NULL, which will keep existing names in \code{net} if any. Set to NA to avoid printing vertex labels.}

\item{vertex_shape}{Character, vector, length 1 or N nodes: Vertex shape. See \code{graphjs("vertex_shape")}.}

\item{edge_col}{Color for edges.}

\item{edge_alpha}{Numeric: Transparency for edges.}

\item{edge_curved}{Numeric: Curvature of edges.}

\item{edge_width}{Numeric: Edge thickness.}

\item{layout}{Character: one of: "fr", "dh", "drl", "gem", "graphopt", "kk", "lgl", "mds", "sugiyama", corresponding to all the available layouts in \pkg{igraph}.}

\item{coords}{Output of precomputed \pkg{igraph} layout. If provided, \code{layout} is ignored.}

\item{layout_params}{List of parameters to pass to \code{layout} function.}

\item{cluster}{Character: one of: "edge_betweenness", "fast_greedy", "infomap", "label_prop", "leading_eigen", "louvain", "optimal", "spinglass", "walktrap", corresponding to all the available \pkg{igraph} clustering functions.}

\item{groups}{Output of precomputed \pkg{igraph} clustering. If provided, \code{cluster} is ignored.}

\item{cluster_params}{List of parameters to pass to \code{cluster} function.}

\item{cluster_mark_groups}{Logical: If TRUE, draw polygons to indicate clusters, if \code{groups} or \code{cluster} are defined.}

\item{cluster_color_vertices}{Logical: If TRUE, color vertices by cluster membership.}

\item{main}{Character: Main title.}

\item{theme}{Theme object.}

\item{palette}{Color vector or name of rtemis palette.}

\item{mar}{Numeric vector, length 4: \code{par}'s margin argument.}

\item{par_reset}{Logical: If TRUE, reset par before exiting.}

\item{filename}{Character: If provided, save plot to this filepath.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Extra arguments to pass to \code{igraph::plot.igraph()}.}
}
\value{
\code{scatterplotThree} object.
}
\description{
Interactive plotting of an \pkg{igraph} net using \pkg{threejs}.
}
\author{
EDG
}
