% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heat.R
\name{draw_heat}
\alias{draw_heat}
\title{Heatmap with \code{plotly}}
\usage{
draw_heat(
  z,
  x = NULL,
  y = NULL,
  title = NULL,
  col = penn_heat(21),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  transpose = TRUE
)
}
\arguments{
\item{z}{Input matrix.}

\item{x}{Vectors for x-axis.}

\item{y}{Vectors for y-axis.}

\item{title}{Plot title.}

\item{col}{Set of colors to make gradient from.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{zlab}{z value label.}

\item{transpose}{Logical: If TRUE, transpose matrix.}
}
\value{
A plotly heatmap object.
}
\description{
Draw a heatmap using \code{plotly}.
}
\author{
EDG
}
