% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Interactive Heatmaps}
\usage{
draw_heatmap(
  x,
  Rowv = TRUE,
  Colv = TRUE,
  cluster = FALSE,
  symm = FALSE,
  cellnote = NULL,
  colorgrad_n = 101,
  colors = NULL,
  space = "rgb",
  lo = "#18A3AC",
  lomid = NULL,
  mid = NULL,
  midhi = NULL,
  hi = "#F48024",
  k_row = 1,
  k_col = 1,
  grid_gap = 0,
  limits = NULL,
  margins = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  key_title = NULL,
  showticklabels = NULL,
  colorbar_len = 0.7,
  plot_method = "plotly",
  theme = choose_theme(),
  row_side_colors = NULL,
  row_side_palette = NULL,
  col_side_colors = NULL,
  col_side_palette = NULL,
  font_size = NULL,
  padding = 0,
  displayModeBar = TRUE,
  modeBar_file_format = "svg",
  filename = NULL,
  file_width = 500,
  file_height = 500,
  file_scale = 1,
  ...
)
}
\arguments{
\item{x}{Input matrix.}

\item{Rowv}{Logical or dendrogram. If Logical: Compute dendrogram and reorder rows. Defaults to FALSE. If dendrogram: use as is, without reordering. See more at \code{heatmaply::heatmaply("Rowv")}.}

\item{Colv}{Logical or dendrogram. If Logical: Compute dendrogram and reorder columns. Defaults to FALSE. If dendrogram: use as is, without reordering. See more at \code{heatmaply::heatmaply("Colv")}.}

\item{cluster}{Logical: If TRUE, set \code{Rowv} and \code{Colv} to TRUE.}

\item{symm}{Logical: If TRUE, treat \code{x} symmetrically - \code{x} must be a square matrix.}

\item{cellnote}{Matrix with values to be displayed on hover. Defaults to \code{ddSci(x)}.}

\item{colorgrad_n}{Integer: Number of colors in gradient. Default = 101.}

\item{colors}{Character vector: Colors to use in gradient.}

\item{space}{Character: Color space to use. Default = "rgb".}

\item{lo}{Character: Color for low values. Default = "#18A3AC".}

\item{lomid}{Character: Color for low-mid values.}

\item{mid}{Character: Color for mid values.}

\item{midhi}{Character: Color for mid-high values.}

\item{hi}{Character: Color for high values. Default = "#F48024".}

\item{k_row}{Integer: Number of desired number of groups by which to color dendrogram branches in the rows. Default = 1.}

\item{k_col}{Integer: Number of desired number of groups by which to color dendrogram branches in the columns. Default = 1.}

\item{grid_gap}{Integer: Space between cells. Default = 0 (no space).}

\item{limits}{Float, length 2: Determine color range. Default = NULL, which automatically centers values around 0.}

\item{margins}{Float, length 4: Heatmap margins.}

\item{main}{Character: Main title.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{key_title}{Character: Title for the color key.}

\item{showticklabels}{Logical: If TRUE, show tick labels.}

\item{colorbar_len}{Numeric: Length of the colorbar.}

\item{plot_method}{Character: Plot method to use. Default = "plotly".}

\item{theme}{Theme object.}

\item{row_side_colors}{Data frame: Column names will be label names, cells should be label colors. See \code{heatmaply::heatmaply("row_side_colors")}.}

\item{row_side_palette}{Color palette function. See \code{heatmaply::heatmaply("row_side_palette")}.}

\item{col_side_colors}{Data frame: Column names will be label names, cells should be label colors. See \code{heatmaply::heatmaply("col_side_colors")}.}

\item{col_side_palette}{Color palette function. See \code{heatmaply::heatmaply("col_side_palette")}.}

\item{font_size}{Numeric: Font size.}

\item{padding}{Numeric: Padding between cells.}

\item{displayModeBar}{Logical: If TRUE, display the plotly mode bar.}

\item{modeBar_file_format}{Character: File format for image exports from the mode bar.}

\item{filename}{Character: File name to save the plot.}

\item{file_width}{Numeric: Width of exported image.}

\item{file_height}{Numeric: Height of exported image.}

\item{file_scale}{Numeric: Scale of exported image.}

\item{...}{Additional arguments to be passed to \code{heatmaply::heatmaply}.}
}
\value{
A heatmaply object.
}
\description{
Draw interactive heatmaps using \code{heatmaply}.
}
\examples{
\dontrun{
x <- rnormmat(200, 20)
xcor <- cor(x)
draw_heatmap(xcor)
}
}
\author{
EDG
}
