% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_leaflet.R
\name{draw_leaflet}
\alias{draw_leaflet}
\title{Plot interactive choropleth map using \pkg{leaflet}}
\usage{
draw_leaflet(
  fips,
  values,
  names = NULL,
  fillOpacity = 1,
  palette = NULL,
  color_mapping = c("Numeric", "Bin"),
  col_lo = "#0290EE",
  col_hi = "#FE4AA3",
  col_na = "#303030",
  col_highlight = "#FE8A4F",
  col_interpolate = c("linear", "spline"),
  col_bins = 21,
  domain = NULL,
  weight = 0.5,
  color = "black",
  alpha = 1,
  bg_tile_provider = leaflet::providers[["Stamen.TonerBackground"]],
  bg_tile_alpha = 0.67,
  fg_tile_provider = leaflet::providers[["Stamen.TonerLabels"]],
  legend_position = c("topright", "bottomright", "bottomleft", "topleft"),
  legend_alpha = 0.8,
  legend_title = NULL,
  init_lng = -98.5418083333333,
  init_lat = 39.2074138888889,
  init_zoom = 3,
  stroke = TRUE
)
}
\arguments{
\item{fips}{Character vector of FIPS codes. (If numeric, it will be appropriately zero-padded).}

\item{values}{Values to map to \code{fips}.}

\item{names}{Character vector: Optional county names to appear on hover along \code{values}.}

\item{fillOpacity}{Float: Opacity for fill colors.}

\item{palette}{Character: Color palette to use.}

\item{color_mapping}{Character: "Numeric" or "Bin".}

\item{col_lo}{Overlay color mapped to lowest value.}

\item{col_hi}{Overlay color mapped to highest value.}

\item{col_na}{Color mapped to NA values.}

\item{col_highlight}{Hover border color.}

\item{col_interpolate}{Character: "linear" or "spline".}

\item{col_bins}{Integer: Number of color bins to create if \code{color_mapping = "Bin"}.}

\item{domain}{Limits for mapping colors to values. Default = NULL and set to range.}

\item{weight}{Float: Weight of county border lines.}

\item{color}{Color of county border lines.}

\item{alpha}{Float: Overlay transparency.}

\item{bg_tile_provider}{Background tile (below overlay colors), one of \code{leaflet::providers}.}

\item{bg_tile_alpha}{Float: Background tile transparency.}

\item{fg_tile_provider}{Foreground tile (above overlay colors), one of \code{leaflet::providers}.}

\item{legend_position}{Character: One of: "topright", "bottomright", "bottomleft", "topleft".}

\item{legend_alpha}{Float: Legend box transparency.}

\item{legend_title}{Character: Defaults to name of \code{values} variable.}

\item{init_lng}{Float: Center map around this longitude (in decimal form). Default = -98.54180833333334 (US geographic center).}

\item{init_lat}{Float: Center map around this latitude (in decimal form). Default = 39.207413888888894 (US geographic center).}

\item{init_zoom}{Integer: Initial zoom level (depends on device, i.e. window, size).}

\item{stroke}{Logical: If TRUE, draw polygon borders.}
}
\value{
\code{leaflet} object.
}
\description{
Plot interactive choropleth map using \pkg{leaflet}
}
\examples{
\dontrun{
fips <- c(06075, 42101)
population <- c(874961, 1579000)
names <- c("SF", "Philly")
draw_leaflet(fips, population, names)
}
}
\author{
EDG
}
