% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_protein.R
\name{draw_protein}
\alias{draw_protein}
\title{Plot the amino acid sequence with annotations}
\usage{
draw_protein(
  x,
  site = NULL,
  region = NULL,
  ptm = NULL,
  clv = NULL,
  variant = NULL,
  disease_variants = NULL,
  n_per_row = NULL,
  main = NULL,
  main_xy = c(0.055, 0.975),
  main_xref = "paper",
  main_yref = "paper",
  main_xanchor = "middle",
  main_yanchor = "top",
  layout = c("simple", "grid", "1curve", "2curve"),
  show_markers = TRUE,
  show_labels = TRUE,
  font_size = 18,
  label_col = NULL,
  scatter_mode = "markers+lines",
  marker_size = 28,
  marker_col = NULL,
  marker_alpha = 1,
  marker_symbol = "circle",
  line_col = NULL,
  line_alpha = 1,
  line_width = 2,
  show_full_names = TRUE,
  region_scatter_mode = "markers+lines",
  region_style = 3,
  region_marker_size = marker_size,
  region_marker_alpha = 0.6,
  region_marker_symbol = "circle",
  region_line_dash = "solid",
  region_line_shape = "line",
  region_line_smoothing = 1,
  region_line_width = 1,
  region_line_alpha = 0.6,
  theme = rtemis_theme,
  region_palette = rtemis_palette,
  region_outline_only = FALSE,
  region_outline_pad = 2,
  region_pad = 0.35,
  region_fill_alpha = 0.1666666,
  region_fill_shape = "line",
  region_fill_smoothing = 1,
  bpadcx = 0.5,
  bpadcy = 0.5,
  site_marker_size = marker_size,
  site_marker_symbol = marker_symbol,
  site_marker_alpha = 1,
  site_border_width = 1.5,
  site_palette = rtemis_palette,
  variant_col = "#FA6E1E",
  disease_variant_col = "#E266AE",
  showlegend_ptm = TRUE,
  ptm_col = NULL,
  ptm_symbol = "circle",
  ptm_offset = 0.12,
  ptm_pad = 0.35,
  ptm_marker_size = marker_size/4.5,
  clv_col = NULL,
  clv_symbol = "triangle-down",
  clv_offset = 0.12,
  clv_pad = 0.35,
  clv_marker_size = marker_size/4,
  annotate_position_every = 10,
  annotate_position_alpha = 0.5,
  annotate_position_ay = -0.4 * marker_size,
  position_font_size = font_size - 6,
  legend_xy = c(0.97, 0.954),
  legend_xanchor = "left",
  legend_yanchor = "top",
  legend_orientation = "v",
  legend_col = NULL,
  legend_bg = "#FFFFFF00",
  legend_border_col = "#FFFFFF00",
  legend_borderwidth = 0,
  legend_group_gap = 0,
  margin = list(b = 0, l = 0, t = 0, r = 0, pad = 0),
  showgrid_x = FALSE,
  showgrid_y = FALSE,
  automargin_x = TRUE,
  automargin_y = TRUE,
  xaxis_autorange = TRUE,
  yaxis_autorange = "reversed",
  scaleanchor_y = "x",
  scaleratio_y = 1,
  hoverlabel_align = "left",
  displayModeBar = TRUE,
  modeBar_file_format = "svg",
  scrollZoom = TRUE,
  filename = NULL,
  file_width = 1320,
  file_height = 990,
  file_scale = 1,
  width = NULL,
  height = NULL,
  verbosity = 1L,
  ...
)
}
\arguments{
\item{x}{Character vector: amino acid sequence (1-letter abbreviations) OR
\code{a3} object OR Character: path to JSON file OR Character: UniProt accession number.}

\item{site}{Named list of lists with indices of sites. These will be
highlighted by coloring the border of markers.}

\item{region}{Named list of lists with indices of regions. These will be
highlighted by coloring the markers and lines of regions using the
\code{palette} colors.}

\item{ptm}{List of post-translational modifications.}

\item{clv}{List of cleavage sites.}

\item{variant}{List of variant information.}

\item{disease_variants}{List of disease variant information.}

\item{n_per_row}{Integer: Number of amino acids to show per row.}

\item{main}{Character: Main title.}

\item{main_xy}{Numeric vector, length 2: x and y coordinates for title.
e.g. if \code{main_xref} and \code{main_yref} are \code{"paper"}:
\code{c(0.055, .975)} is top left, \code{c(.5, .975)} is top and
middle.}

\item{main_xref}{Character: xref for title.}

\item{main_yref}{Character: yref for title.}

\item{main_xanchor}{Character: xanchor for title.}

\item{main_yanchor}{Character: yanchor for title.}

\item{layout}{Character: "1curve", "grid": type of layout to use.}

\item{show_markers}{Logical: If TRUE, show amino acid markers.}

\item{show_labels}{Logical: If TRUE, annotate amino acids with elements.}

\item{font_size}{Integer: Font size for labels.}

\item{label_col}{Color for labels.}

\item{scatter_mode}{Character: Mode for scatter plot.}

\item{marker_size}{Integer: Size of markers.}

\item{marker_col}{Color for markers.}

\item{marker_alpha}{Numeric: Alpha for markers.}

\item{marker_symbol}{Character: Symbol for markers.}

\item{line_col}{Color for lines.}

\item{line_alpha}{Numeric: Alpha for lines.}

\item{line_width}{Numeric: Width for lines.}

\item{show_full_names}{Logical: If TRUE, show full names of amino acids.}

\item{region_scatter_mode}{Character: Mode for scatter plot.}

\item{region_style}{Integer: Style for regions.}

\item{region_marker_size}{Integer: Size of region markers.}

\item{region_marker_alpha}{Numeric: Alpha for region markers.}

\item{region_marker_symbol}{Character: Symbol for region markers.}

\item{region_line_dash}{Character: Dash for region lines.}

\item{region_line_shape}{Character: Shape for region lines.}

\item{region_line_smoothing}{Numeric: Smoothing for region lines.}

\item{region_line_width}{Numeric: Width for region lines.}

\item{region_line_alpha}{Numeric: Alpha for region lines.}

\item{theme}{Character: Theme to use: Run \code{themes()} for available themes.}

\item{region_palette}{Named list of colors for regions.}

\item{region_outline_only}{Logical: If TRUE, only show outline of regions.}

\item{region_outline_pad}{Numeric: Padding for region outline.}

\item{region_pad}{Numeric: Padding for region.}

\item{region_fill_alpha}{Numeric: Alpha for region fill.}

\item{region_fill_shape}{Character: Shape for region fill.}

\item{region_fill_smoothing}{Numeric: Smoothing for region fill.}

\item{bpadcx}{Numeric: Padding for region border.}

\item{bpadcy}{Numeric: Padding for region border.}

\item{site_marker_size}{Integer: Size of site markers.}

\item{site_marker_symbol}{Character: Symbol for site markers.}

\item{site_marker_alpha}{Numeric: Alpha for site markers.}

\item{site_border_width}{Numeric: Width for site borders.}

\item{site_palette}{Named list of colors for sites.}

\item{variant_col}{Color for variants.}

\item{disease_variant_col}{Color for disease variants.}

\item{showlegend_ptm}{Logical: If TRUE, show legend for PTMs.}

\item{ptm_col}{Named list of colors for PTMs.}

\item{ptm_symbol}{Character: Symbol for PTMs.}

\item{ptm_offset}{Numeric: Offset for PTMs.}

\item{ptm_pad}{Numeric: Padding for PTMs.}

\item{ptm_marker_size}{Integer: Size of PTM markers.}

\item{clv_col}{Color for cleavage site annotations.}

\item{clv_symbol}{Character: Symbol for cleavage site annotations.}

\item{clv_offset}{Numeric: Offset for cleavage site annotations.}

\item{clv_pad}{Numeric: Padding for cleavage site annotations.}

\item{clv_marker_size}{Integer: Size of cleavage site annotation markers.}

\item{annotate_position_every}{Integer: Annotate every nth position.}

\item{annotate_position_alpha}{Numeric: Alpha for position annotations.}

\item{annotate_position_ay}{Numeric: Y offset for position annotations.}

\item{position_font_size}{Integer: Font size for position annotations.}

\item{legend_xy}{Numeric vector, length 2: x and y coordinates for legend.}

\item{legend_xanchor}{Character: xanchor for legend.}

\item{legend_yanchor}{Character: yanchor for legend.}

\item{legend_orientation}{Character: Orientation for legend.}

\item{legend_col}{Color for legend.}

\item{legend_bg}{Color for legend background.}

\item{legend_border_col}{Color for legend border.}

\item{legend_borderwidth}{Numeric: Width for legend border.}

\item{legend_group_gap}{Numeric: Gap between legend groups.}

\item{margin}{List: Margin settings.}

\item{showgrid_x}{Logical: If TRUE, show x grid.}

\item{showgrid_y}{Logical: If TRUE, show y grid.}

\item{automargin_x}{Logical: If TRUE, use automatic margin for x axis.}

\item{automargin_y}{Logical: If TRUE, use automatic margin for y axis.}

\item{xaxis_autorange}{Logical: If TRUE, use automatic range for x axis.}

\item{yaxis_autorange}{Character: If TRUE, use automatic range for y axis.}

\item{scaleanchor_y}{Character: Scale anchor for y axis.}

\item{scaleratio_y}{Numeric: Scale ratio for y axis.}

\item{hoverlabel_align}{Character: Alignment for hover label.}

\item{displayModeBar}{Logical: If TRUE, display mode bar.}

\item{modeBar_file_format}{Character: File format for mode bar.}

\item{scrollZoom}{Logical: If TRUE, enable scroll zoom.}

\item{filename}{Character: File name to save plot.}

\item{file_width}{Integer: Width for saved file.}

\item{file_height}{Integer: Height for saved file.}

\item{file_scale}{Numeric: Scale for saved file.}

\item{width}{Integer: Width for plot.}

\item{height}{Integer: Height for plot.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments to pass to the theme function.}
}
\value{
A plotly object
}
\description{
Plot the amino acid sequence with annotations
}
\examples{
\dontrun{
tau <- seqinr::read.fasta("https://rest.uniprot.org/uniprotkb/P10636.fasta",
  seqtype = "AA"
)
draw_protein(as.character(tau[[1]]))

# or directly using the UniProt accession number:
draw_protein("P10636")
}
}
\author{
EDG
}
