% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pvals.R
\name{draw_pvals}
\alias{draw_pvals}
\title{Barplot p-values using \link{draw_bar}}
\usage{
draw_pvals(
  x,
  xnames = NULL,
  yname = NULL,
  p_adjust_method = "none",
  pval_hline = 0.05,
  hline_col = "#FE4AA3",
  hline_dash = "dash",
  ...
)
}
\arguments{
\item{x}{Float, vector: p-values.}

\item{xnames}{Character, vector: feature names.}

\item{yname}{Character: outcome name.}

\item{p_adjust_method}{Character: method for \link{p.adjust}.}

\item{pval_hline}{Float: Significance level at which to plot horizontal line.}

\item{hline_col}{Color for \code{pval_hline}.}

\item{hline_dash}{Character: type of line to draw.}

\item{...}{Additional arguments passed to \link{draw_bar}.}
}
\value{
A plotly object.
}
\description{
Plot 1 - p-values as a barplot
}
\examples{
\dontrun{
draw_pvals(c(0.01, 0.02, 0.03), xnames = c("Feature1", "Feature2", "Feature3"))
}
}
\author{
EDG
}
