% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_roc.R
\name{draw_roc}
\alias{draw_roc}
\title{Draw ROC curve}
\usage{
draw_roc(
  true_labels,
  predicted_prob,
  main = NULL,
  theme = rtemis_theme,
  col = rtpalette(rtemis_palette),
  legend = TRUE,
  legend_xy = c(1, 0),
  legend_xanchor = "right",
  legend_yanchor = "bottom",
  xlim = c(0, 1),
  ylim = c(0, 1),
  diagonal = TRUE,
  diagonal_col = "#808080",
  axes_square = TRUE,
  filename = NULL,
  ...
)
}
\arguments{
\item{true_labels}{Factor: True outcome labels.}

\item{predicted_prob}{Numeric vector \link{0, 1}: Predicted probabilities for the positive class (i.e. second level of outcome).}

\item{main}{Character: Main title for the plot.}

\item{theme}{Theme object.}

\item{col}{Color vector.}

\item{legend}{Logical: If TRUE, draw legend.}

\item{legend_xy}{Numeric vector: Position of the legend in the form c(x, y).}

\item{legend_xanchor}{Character: X anchor for the legend.}

\item{legend_yanchor}{Character: Y anchor for the legend.}

\item{xlim}{Numeric vector: Limits for the x-axis.}

\item{ylim}{Numeric vector: Limits for the y-axis.}

\item{diagonal}{Logical: If TRUE, draw diagonal line.}

\item{diagonal_col}{Character: Color for the diagonal line.}

\item{axes_square}{Logical: If TRUE, make axes square.}

\item{filename}{Character: If provided, save the plot to this file.}

\item{...}{Additional arguments passed to \link{draw_scatter}.}
}
\value{
plotly object
}
\description{
Draw ROC curve
}
\author{
EDG
}
