% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_spectrogram.R
\name{draw_spectrogram}
\alias{draw_spectrogram}
\title{Interactive Spectrogram}
\usage{
draw_spectrogram(
  x,
  y,
  z,
  colorgrad_n = 101,
  colors = NULL,
  xlab = "Time",
  ylab = "Frequency",
  zlab = "Power",
  hover_xlab = xlab,
  hover_ylab = ylab,
  hover_zlab = zlab,
  zmin = NULL,
  zmax = NULL,
  zauto = TRUE,
  hoverlabel_align = "right",
  colorscale = "Jet",
  colorbar_y = 0.5,
  colorbar_yanchor = "middle",
  colorbar_xpad = 0,
  colorbar_ypad = 0,
  colorbar_len = 0.75,
  colorbar_title_side = "bottom",
  showgrid = FALSE,
  space = "rgb",
  lo = "#18A3AC",
  lomid = NULL,
  mid = NULL,
  midhi = NULL,
  hi = "#F48024",
  grid_gap = 0,
  limits = NULL,
  main = NULL,
  key_title = NULL,
  showticklabels = NULL,
  theme = choose_theme(),
  font_size = NULL,
  padding = 0,
  displayModeBar = TRUE,
  modeBar_file_format = "svg",
  filename = NULL,
  file_width = 500,
  file_height = 500,
  file_scale = 1,
  ...
)
}
\arguments{
\item{x}{Numeric: Time.}

\item{y}{Numeric: Frequency.}

\item{z}{Numeric: Power.}

\item{colorgrad_n}{Integer: Number of colors in the gradient.}

\item{colors}{Character: Custom colors for the gradient.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{zlab}{Character: z-axis label.}

\item{hover_xlab}{Character: x-axis label for hover.}

\item{hover_ylab}{Character: y-axis label for hover.}

\item{hover_zlab}{Character: z-axis label for hover.}

\item{zmin}{Numeric: Minimum value for color scale.}

\item{zmax}{Numeric: Maximum value for color scale.}

\item{zauto}{Logical: If TRUE, automatically set zmin and zmax.}

\item{hoverlabel_align}{Character: Alignment of hover labels.}

\item{colorscale}{Character: Color scale.}

\item{colorbar_y}{Numeric: Y position of colorbar.}

\item{colorbar_yanchor}{Character: Y anchor of colorbar.}

\item{colorbar_xpad}{Numeric: X padding of colorbar.}

\item{colorbar_ypad}{Numeric: Y padding of colorbar.}

\item{colorbar_len}{Numeric: Length of colorbar.}

\item{colorbar_title_side}{Character: Side of colorbar title.}

\item{showgrid}{Logical: If TRUE, show grid.}

\item{space}{Character: Color space for gradient.}

\item{lo}{Character: Low color for gradient.}

\item{lomid}{Character: Low-mid color for gradient.}

\item{mid}{Character: Mid color for gradient.}

\item{midhi}{Character: Mid-high color for gradient.}

\item{hi}{Character: High color for gradient.}

\item{grid_gap}{Integer: Space between cells.}

\item{limits}{Numeric, length 2: Determine color range. Default = NULL, which automatically centers values around 0.}

\item{main}{Character: Main title.}

\item{key_title}{Character: Title of the key.}

\item{showticklabels}{Logical: If TRUE, show tick labels.}

\item{theme}{Theme object.}

\item{font_size}{Numeric: Font size.}

\item{padding}{Numeric: Padding between cells.}

\item{displayModeBar}{Logical: If TRUE, display the plotly mode bar.}

\item{modeBar_file_format}{Character: File format for image exports from the mode bar.}

\item{filename}{Character: Filename to save the plot. Default is NULL.}

\item{file_width}{Numeric: Width of exported image.}

\item{file_height}{Numeric: Height of exported image.}

\item{file_scale}{Numeric: Scale of exported image.}

\item{...}{Additional arguments to be passed to \code{heatmaply::heatmaply}.}
}
\value{
\code{plotly} object.
}
\description{
Draw interactive spectrograms using \code{plotly}
}
\details{
To set custom colors, use a minimum of \code{lo} and \code{hi}, optionally also
\code{lomid}, \code{mid}, \code{midhi} colors and set \code{colorscale = NULL}.
}
\author{
EDG
}
