% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_survfit.R
\name{draw_survfit}
\alias{draw_survfit}
\title{Draw a survfit object}
\usage{
draw_survfit(
  x,
  mode = "lines",
  symbol = "cross",
  line_shape = "hv",
  xlim = NULL,
  ylim = NULL,
  xlab = "Time",
  ylab = "Survival",
  main = NULL,
  legend_xy = c(1, 1),
  legend_xanchor = "right",
  legend_yanchor = "top",
  theme = choose_theme(),
  nrisk_table = FALSE,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{\code{survfit} object created by \link[survival:survfit]{survival::survfit}.}

\item{mode}{Character, vector: "markers", "lines", "markers+lines".}

\item{symbol}{Character: Symbol to use for the points.}

\item{line_shape}{Character: Line shape for line plots. Options: "linear", "hv", "vh", "hvh", "vhv".}

\item{xlim}{Numeric vector of length 2: x-axis limits.}

\item{ylim}{Numeric vector of length 2: y-axis limits.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{main}{Character: Main title.}

\item{legend_xy}{Numeric: Position of legend.}

\item{legend_xanchor}{Character: X anchor for legend.}

\item{legend_yanchor}{Character: Y anchor for legend.}

\item{theme}{Theme object.}

\item{nrisk_table}{Logical: If \code{TRUE}, subplot a table of the number at risk at each time point.}

\item{filename}{Character: Filename to save plot.}

\item{...}{Additional arguments passed to \link{draw_scatter}.}
}
\value{
\code{plotly} object.
}
\description{
Draw a \code{survfit} object using \link{draw_scatter}.
}
\examples{
\dontrun{
# Get the lung dataset
data(cancer, package = "survival")
sf1 <- survival::survfit(survival::Surv(time, status) ~ 1, data = lung)
draw_survfit(sf1)
sf2 <- survival::survfit(survival::Surv(time, status) ~ sex, data = lung)
draw_survfit(sf2)
# with N at risk table
draw_survfit(sf2)
}
}
\author{
EDG
}
