% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_varimp.R
\name{draw_varimp}
\alias{draw_varimp}
\title{Interactive Variable Importance Plot}
\usage{
draw_varimp(
  x,
  names = NULL,
  main = NULL,
  type = c("bar", "line"),
  xlab = NULL,
  ylab = NULL,
  plot_top = 1,
  orientation = "v",
  line_width = 12,
  labelify = TRUE,
  col = NULL,
  alpha = 1,
  palette = rtemis_palette,
  mar = NULL,
  font_size = 16,
  axis_font_size = 14,
  theme = choose_theme(),
  showlegend = TRUE,
  filename = NULL,
  file_width = 500,
  file_height = 500,
  file_scale = 1
)
}
\arguments{
\item{x}{Numeric vector: Input.}

\item{names}{Vector, string: Names of features.}

\item{main}{Character: Main title.}

\item{type}{Character: "bar" or "line".}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{plot_top}{Integer: Plot this many top features.}

\item{orientation}{Character: "h" or "v".}

\item{line_width}{Numeric: Line width.}

\item{labelify}{Logical: If TRUE, labelify feature names.}

\item{col}{Vector, colors: Single value, or multiple values to define bar
(feature) color(s).}

\item{alpha}{Numeric: Transparency.}

\item{palette}{Character: Name of \pkg{rtemis} palette to use.}

\item{mar}{Vector, numeric, length 4: Plot margins in pixels (NOT inches).}

\item{font_size}{Integer: Overall font size to use (essentially for the
title at this point).}

\item{axis_font_size}{Integer: Font size to use for axis labels and tick labels.}

\item{theme}{Theme object.}

\item{showlegend}{Logical: If TRUE, show legend.}

\item{filename}{Character: Path to save the plot image.}

\item{file_width}{Numeric: Width of the saved plot image.}

\item{file_height}{Numeric: Height of the saved plot image.}

\item{file_scale}{Numeric: Scale of the saved plot image.}
}
\value{
A \code{plotly} object.
}
\description{
Plot variable importance using \code{plotly}
}
\details{
A simple \code{plotly} wrapper to plot horizontal barplots, sorted by value,
which can be used to visualize variable importance, model coefficients, etc.
}
\examples{
\dontrun{
# made-up data
x <- rnorm(10)
names(x) <- paste0("Feature_", seq(x))
draw_varimp(x)
draw_varimp(x, orientation = "v")
}
}
\author{
EDG
}
