% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_check_unique}
\alias{dt_check_unique}
\title{Check if all levels in a column are unique}
\usage{
dt_check_unique(x, on)
}
\arguments{
\item{x}{data.frame or data.table}

\item{on}{Integer or character: column to check}
}
\description{
Check if all levels in a column are unique
}
\author{
E.D. Gennatas
}
