% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_get_duplicates}
\alias{dt_get_duplicates}
\title{Get index of duplicate values}
\usage{
dt_get_duplicates(x, on)
}
\arguments{
\item{x}{data.frame or data.table}

\item{on}{Integer or character: column to check}
}
\description{
Get index of duplicate values
}
\author{
E.D. Gennatas
}
