% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_index_attr}
\alias{dt_index_attr}
\title{Index columns by attribute name & value}
\usage{
dt_index_attr(x, name, value)
}
\arguments{
\item{x}{data.frame or compatible}

\item{name}{Character: Name of attribute}

\item{value}{Character: Value of attribute}
}
\description{
Index columns by attribute name & value
}
\author{
E.D. Gennatas
}
