% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_inspect_type}
\alias{dt_inspect_type}
\title{Inspect column types}
\usage{
dt_inspect_type(x, cols = NULL, verbosity = 1L)
}
\arguments{
\item{x}{data.table}

\item{cols}{Character vector: columns to inspect.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Character vector.
}
\description{
Will attempt to identify columns that should be numeric but are either character or
factor by running \link{inspect_type} on each column.
}
\author{
EDG
}
