% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_keybin_reshape}
\alias{dt_keybin_reshape}
\title{Long to wide key-value reshaping}
\usage{
dt_keybin_reshape(
  x,
  id_name,
  key_name,
  positive = 1,
  negative = 0,
  xname = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{data.table} object}

\item{id_name}{Character: Name of column in \code{x} that defines the IDs
identifying individual rows}

\item{key_name}{Character: Name of column in \code{x} that holds the key}

\item{positive}{Numeric or Character: Used to fill id ~ key combination
present in the long format input \code{x}}

\item{negative}{Numeric or Character: Used to fill id ~ key combination
NOT present in the long format input \code{x}}

\item{xname}{Character: Name of \code{x} to be used in messages}

\item{verbose}{Logical: If TRUE, print messages to the console}
}
\description{
Reshape a long format \code{data.table} using key-value pairs with
\code{data.table::dcast}
}
\examples{
\dontrun{
x <- data.table(
  ID = rep(1:3, each = 2),
  Dx = c("A", "C", "B", "C", "D", "A")
)
dt_keybin_reshape(x, id_name = "ID", key_name = "Dx")
}
}
\author{
E.D. Gennatas
}
