% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_merge}
\alias{dt_merge}
\title{Merge data.tables}
\usage{
dt_merge(
  left,
  right,
  on = NULL,
  left_on = NULL,
  right_on = NULL,
  how = "left",
  left_name = NULL,
  right_name = NULL,
  left_suffix = NULL,
  right_suffix = NULL,
  verbosity = 1L,
  ...
)
}
\arguments{
\item{left}{data.table}

\item{right}{data.table}

\item{on}{Character: Name of column to join on.}

\item{left_on}{Character: Name of column on left table.}

\item{right_on}{Character: Name of column on right table.}

\item{how}{Character: Type of join: "inner", "left", "right", "outer".}

\item{left_name}{Character: Name of left table.}

\item{right_name}{Character: Name of right table.}

\item{left_suffix}{Character: If provided, add this suffix to all left column names,
excluding on/left_on.}

\item{right_suffix}{Character: If provided, add this suffix to all right column names,
excluding on/right_on.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments to be passed to \code{data.table::merge}.}
}
\value{
Merged data.table.
}
\description{
Merge data.tables
}
\author{
EDG
}
