% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_names_by_attr}
\alias{dt_names_by_attr}
\title{List column names by attribute}
\usage{
dt_names_by_attr(x, attribute, exact = TRUE, sorted = TRUE)
}
\arguments{
\item{x}{data.table}

\item{attribute}{Character: name of attribute}

\item{exact}{Logical: If TRUE, use exact matching}

\item{sorted}{Logical: If TRUE, sort the output}
}
\value{
Character vector.
}
\description{
List column names by attribute
}
\author{
EDG
}
