% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_names_by_class}
\alias{dt_names_by_class}
\title{List column names by class}
\usage{
dt_names_by_class(x, sorted = TRUE, item_format = highlight, maxlength = 24)
}
\arguments{
\item{x}{data.table}

\item{sorted}{Logical: If TRUE, sort the output}

\item{item_format}{Function: Function to format each item}

\item{maxlength}{Integer: Maximum number of items to print}
}
\value{
\code{NULL}, invisibly.
}
\description{
List column names by class
}
\author{
EDG
}
