% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_nunique_perfeat}
\alias{dt_nunique_perfeat}
\title{Number of unique values per feature}
\usage{
dt_nunique_perfeat(x, excludeNA = FALSE, verbosity = 1L)
}
\arguments{
\item{x}{data.table}

\item{excludeNA}{Logical: If TRUE, exclude NA values.}

\item{verbosity}{Integer: If > 0, print output to console.}
}
\value{
Named integer vector of length \code{NCOL(x)} with number of unique values per column/feature, invisibly.
}
\description{
Number of unique values per feature
}
\author{
EDG
}
