% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_set_autotypes}
\alias{dt_set_autotypes}
\title{Set column types automatically}
\usage{
dt_set_autotypes(x, cols = NULL, verbosity = 1L)
}
\arguments{
\item{x}{data.table}

\item{cols}{Character vector: columns to work on. If not defined, will work on all
columns}

\item{verbosity}{Integer: Verbosity level.}
}
\description{
This function inspects a data.table and attempts to identify columns that should be
numeric but have been read in as character, because one or more fields contain
non-numeric characters
}
\author{
EDG
}
