% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_set_clean_all}
\alias{dt_set_clean_all}
\title{Clean column names and factor levels in-place}
\usage{
dt_set_clean_all(x, prefix_digits = NA)
}
\arguments{
\item{x}{data.table}

\item{prefix_digits}{Character: prefix to add to names beginning with a
digit. Set to NA to skip

Note: If \code{x} is not a data.table, it will be converted in place, i.e. the original
object will be modified.}
}
\description{
Clean column names and factor levels in-place
}
\author{
E.D. Gennatas
}
