% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_set_cleanfactorlevels}
\alias{dt_set_cleanfactorlevels}
\title{Clean factor levels of data.table in-place}
\usage{
dt_set_cleanfactorlevels(x, prefix_digits = NA)
}
\arguments{
\item{x}{data.table}

\item{prefix_digits}{Character: If not NA, add this prefix to all factor levels that
are numbers}
}
\value{
Nothing, modifies \code{x} in-place.
}
\description{
Finds all factors in a data.table and cleans factor levels to include
only underscore symbols
}
\examples{
\dontrun{
x <- as.data.table(iris)
levels(x$Species) <- c("setosa:iris", "versicolor$iris", "virginica iris")
dt_set_cleanfactorlevels(x)
x
}
}
\author{
EDG
}
