% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_set_logical2factor}
\alias{dt_set_logical2factor}
\title{Convert data.table logical columns to factor with custom labels in-place}
\usage{
dt_set_logical2factor(x, cols = NULL, labels = c("False", "True"))
}
\arguments{
\item{x}{data.table}

\item{cols}{Integer or character: columns to convert, if NULL, operates on all
logical columns}

\item{labels}{Character: labels for factor levels}
}
\description{
Convert data.table logical columns to factor with custom labels in-place
}
\examples{
\dontrun{
library(data.table)
x <- data.table(a = 1:5, b = c(T, F, F, F, T))
x
dt_set_logical2factor(x)
x
z <- data.table(alpha = 1:5, beta = c(T, F, T, F, T))
z
dt_set_logical2factor(z, cols = "beta", labels = c("No", "Yes"))
z
w <- data.table(mango = 1:5, banana = c(F, F, T, T, F))
dt_set_logical2factor(w, cols = 2, labels = c("Ugh", "Huh"))
}
}
\author{
E.D. Gennatas
}
