% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_type_inspect}
\alias{dt_type_inspect}
\title{Inspect column types}
\usage{
dt_type_inspect(x)
}
\arguments{
\item{x}{data.table}
}
\description{
Will attempt to identify columns that should be numeric but have been read in as
character by running \link{type_inspect} on each column.
}
\author{
E.D. Gennatas
}
