% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlystop.R
\name{earlystop}
\alias{earlystop}
\title{Early stopping}
\usage{
earlystop(
  x,
  window = 10,
  window_decrease_pct_min = 0.01,
  total_decrease_pct_max = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Numeric vector: loss at each iteration}

\item{window}{Integer: Number of steps to consider}

\item{window_decrease_pct_min}{Float: Stop if improvement is less than this percent over last \code{window} steps}

\item{total_decrease_pct_max}{Float: Stop if improvement from first to last step exceeds this percent. If defined, overrides \code{window_decrease_pct_min}}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Check loss vector for early stopping criteria:
- either total percent decrease from starting error (e.g. if predictions started at expectation)
- or minimum percent decrease (relative to the first value of the vector) over a window of last n steps
}
\details{
If the first loss value was set to be the loss when yhat = mean(y) (e.g. in boosting), then
\code{total_decrease_pct_max} corresponds to R-squared and \code{window_decrease_pct_min} to percent R-squared
improvement over \code{window} last steps.
}
\author{
E.D. Gennatas
}
