% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mse}
\alias{mse}
\alias{msew}
\alias{rmse}
\title{Weighted MSE}
\usage{
mse(x, y, na.rm = TRUE)

msew(x, y, weights = rep(1, length(y)), na.rm = TRUE)

rmse(x, y, na.rm = TRUE)
}
\arguments{
\item{weights}{Float, vector: Case weights}
}
\description{
Weighted MSE
}
\keyword{internal}
