% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{expand.boost}
\alias{expand.boost}
\title{Expand boosting series}
\usage{
expand.boost(
  object,
  x,
  y = NULL,
  x.valid = NULL,
  y.valid = NULL,
  x.test = NULL,
  y.test = NULL,
  mod = NULL,
  resid = NULL,
  mod.params = NULL,
  max.iter = 10,
  learning.rate = NULL,
  case.p = 1,
  prefix = NULL,
  verbose = TRUE,
  trace = 0,
  print.error.plot = "final",
  print.plot = FALSE
)
}
\arguments{
\item{object}{\link{boost} object}

\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.valid}{Data.frame; optional: Validation data}

\item{y.valid}{Float, vector; optional: Validation outcome}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{mod}{Character: Algorithm to train base learners, for options, see
\link{select_learn}. Default = "cart"}

\item{resid}{Float, vector, length = length(y): Residuals to work on. Do
not change unless you know what you're doing. Default = NULL, for regular
boosting}

\item{mod.params}{Named list of arguments for \code{mod}}

\item{max.iter}{Integer: Maximum number of iterations (additive steps) to
perform. Default = 10}

\item{learning.rate}{Float (0, 1] Learning rate for the additive steps}

\item{case.p}{Float (0, 1]: Train each iteration using this perceent of
cases. Default = 1, i.e. use all cases}

\item{prefix}{Internal}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If > 0, print diagnostic info to console}

\item{print.error.plot}{String or Integer: "final" plots a training and
validation (if available) error curve at the end of training. If integer,
plot training and validation error curve every this many iterations during
training. "none" for no plot.}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}.}
}
\description{
Expand a \link{boost} object by adding more iterations
}
\author{
E.D. Gennatas
}
