% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{expand.hytboost}
\alias{expand.hytboost}
\title{Expand boosting series}
\usage{
expand.hytboost(
  object,
  x,
  y = NULL,
  x.valid = NULL,
  y.valid = NULL,
  resid = NULL,
  mod.params = NULL,
  max.iter = 10,
  learning.rate = NULL,
  case.p = 1,
  cxrcoef = FALSE,
  prefix = NULL,
  verbose = TRUE,
  trace = 0,
  print.error.plot = "final"
)
}
\arguments{
\item{object}{\link{boost} object}
}
\description{
Add iterations to a \link{boost} object
}
\author{
E.D. Gennatas
}
