% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotly.R
\name{export_plotly}
\alias{export_plotly}
\title{Export plotly plot to file}
\usage{
export_plotly(
  x,
  filename,
  width = 600,
  height = 600,
  scale = 1,
  import_kaleido = TRUE,
  verbosity = 1L
)
}
\arguments{
\item{x}{plotly object.}

\item{filename}{Character: Filename to save the plot to.}

\item{width}{Numeric: Width of the exported image in pixels.}

\item{height}{Numeric: Height of the exported image in pixels.}

\item{scale}{Numeric: Scale factor for the exported image.}

\item{import_kaleido}{Logical: If TRUE, attempts to import kaleido for exporting plotly plots.}

\item{verbosity}{Integer: Verbosity level.}
}
\description{
Export plotly plot to file
}
\author{
EDG
}
