% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{F1 score}
\usage{
f1(sensitivity, ppv)
}
\arguments{
\item{sensitivity}{Float [0, 1]: The model Sensitivity aka Recall}

\item{ppv}{Float [0, 1]: The model Positive Predictive Value aka Precision}
}
\description{
Calculate the F1 score for classification:
}
\details{
\deqn{F1 = 2 \frac{Sensitivity \cdot PPV}{Sensitivity + PPV}}{F1 = 2 * (Sensitivity * Specificity)/(Sensitivity + Specificity)}
}
\author{
Efstathios D. Gennatas
}
