% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{factor_harmonize}
\alias{factor_harmonize}
\title{Factor harmonize}
\usage{
factor_harmonize(reference, x, verbosity = 1L)
}
\arguments{
\item{reference}{Reference factor.}

\item{x}{Input factor.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Factor: x with levels in the same order as reference.
}
\description{
Factor harmonize
}
\author{
EDG
}
\keyword{internal}
