% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_NA2missing.R
\name{factor_NA2missing}
\alias{factor_NA2missing}
\title{Factor NA to "missing" level}
\usage{
factor_NA2missing(x, na_level_name = "missing")
}
\arguments{
\item{x}{Factor}

\item{na_level_name}{Character: Name of new level to create that will be assigned to all current
NA values. Default = "missing"}
}
\description{
Set NA values of a factor vector to a new level indicating missingness
}
\examples{
x <- factor(sample(letters[1:3], 100, TRUE))
x[sample(1:100, 10)] <- NA
xm <- factor_NA2missing(x)
}
\author{
E.D. Gennatas
}
