% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modError.R
\name{factorHarmonize}
\alias{factorHarmonize}
\title{Factor harmonize}
\usage{
factorHarmonize(reference, x, verbose = TRUE)
}
\arguments{
\item{reference}{Reference factor}

\item{x}{Input factor}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Factor harmonize
}
