% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\title{Factor Analysis}
\usage{
factorize(
  x,
  n_factors = NULL,
  method = "minres",
  rotation = "oblimin",
  scores = "regression",
  cor = "cor",
  fa_n_iter = 100,
  omega_method = "minres",
  omega_rotation = c("oblimin", "simplimax", "promax", "cluster", "target"),
  omega_n_iter = 1,
  x_name = NULL,
  print_plot = TRUE,
  do_pa = TRUE,
  do_fa = TRUE,
  do_bifactor = TRUE,
  do_hclust = FALSE,
  verbosity = 1L,
  ...
)
}
\arguments{
\item{x}{Data. Will be coerced to data frame}

\item{n_factors}{Integer: If NULL, will be estimated using parallel analysis}

\item{method}{Character: Factor analysis method:
"minres": minimum residual (OLS), "wls": weighted least squares (WLS);
"gls": generalized weighted least squares (GLS); "pa": principal factor solution;
"ml": maximum likelihood;
"minchi": minimize the sample size weighted chi square when treating pairwise correlations with
different number of subjects per pair; "minrank": minimum rank factor analysis.}

\item{rotation}{Character: Rotation methods.
No rotation: "none";
Orthogonal: "varimax", "quartimax", "bentlerT", "equamax", "varimin", "geominT", "bifactor";
Oblique: "promax", "oblimin", "simplimax", "bentlerQ, "geominQ", "biquartimin", "cluster".}

\item{scores}{Character: Factor score estimation method. Options: "regression",
"Thurstone": simple regression, "tenBerge": correlation-preserving, "Anderson", "Barlett".}

\item{cor}{Character: Correlation method:
"cor": Pearson correlation, "cov": Covariance, "tet": tetrachoric, "poly": polychoric,
"mixed": mixed cor for a mixture of tetrachorics, polychorics, Pearsons, biserials, and
polyserials, "Yuleb": Yulebonett, "Yuleq" and "YuleY": Yule coefficients}

\item{fa_n_iter}{Integer: Number of iterations for factor analysis.}

\item{omega_method}{Character: Factor analysis method for the bifactor analysis. Same options as
\code{method}}

\item{omega_rotation}{Character: Rotation method for bifactor analysis:
"oblimin", "simplimax", "promax", "cluster", "target".}

\item{omega_n_iter}{Integer: Number of iterations for bifactor analysis.}

\item{x_name}{Character: Name your dataset. Used for plotting}

\item{print_plot}{Logical: If TRUE, print plots along the way.}

\item{do_pa}{Logical: If TRUE, perform parallel analysis.}

\item{do_fa}{Logical: If TRUE, perform factor analysis.}

\item{do_bifactor}{Logical: If TRUE, perform bifactor analysis.}

\item{do_hclust}{Logical: If TRUE, perform hierarchical cluster analysis.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments to pass to \code{psych::fa}}
}
\value{
List with the following elements:
\describe{
\item{parallel_analysis}{Results from parallel analysis}
\item{factor_analysis}{Results from factor analysis}
\item{factor_scores}{Factor scores}
\item{bifactor_analysis}{Results from bifactor analysis}
\item{bifactor_scores}{Bifactor scores}
\item{hclust}{Results from hierarchical cluster analysis}
}
}
\description{
Perform parallel analysis, factor analysis, bifactor analysis
and hierarchical clustering.
}
\details{
Consult \code{psych::fa} for more information on the parameters.
}
\author{
EDG
}
