% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_describe.R
\name{fct_describe}
\alias{fct_describe}
\title{Decribe factor}
\usage{
fct_describe(x, max_n = 5, return_ordered = TRUE)
}
\arguments{
\item{x}{factor}

\item{max_n}{Integer: Return counts for up to this many levels}

\item{return_ordered}{Logical: If TRUE, return levels ordered by count, otherwise
return in level order}
}
\value{
Character with level counts
}
\description{
Outputs a single character with names and counts of each level of the input factor
}
\examples{
\dontrun{
# Small number of levels
fct_describe(iris$Species)

# Large number of levels: show top n by count
x <- factor(sample(letters, 1000, TRUE))
fct_describe(x)
fct_describe(x, 3)
}

}
\author{
E.D. Gennatas
}
