% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{feature_matrix}
\alias{feature_matrix}
\title{Convert tabular data to feature matrix}
\usage{
feature_matrix(x)
}
\arguments{
\item{x}{tabular data: Input data to convert to a feature matrix.}
}
\value{
Matrix with features. Factors are one-hot encoded, if present.
}
\description{
Convert a tabular dataset to a matrix, one-hot encoding factors, if present.
}
\details{
This is a convenience function that uses  \code{\link[=features]{features()}}, \code{\link[=preprocess]{preprocess()}}, \code{as.matrix()}.
}
\examples{
# reorder columns so that we have a categorical feature
x <- set_outcome(iris, "Sepal.Length")
feature_matrix(x) |> head()
}
\author{
EDG
}
