% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_init.R
\name{feature_names}
\alias{feature_names}
\title{Get feature names}
\usage{
feature_names(x)
}
\arguments{
\item{x}{tabular data.}
}
\value{
Character vector of feature names.
}
\description{
Returns all column names except the last one
}
\details{
This applied to tabular datasets used for supervised learning in rtemis,
where, by convention, the last column is the outcome variable and all other columns
are features.
}
\examples{
feature_names(iris)
}
\author{
EDG
}
