% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{filter_order}
\alias{filter_order}
\title{Filter order}
\usage{
filter_order(x, idl, decreasing = FALSE)
}
\arguments{
\item{x}{Input vector}

\item{idl}{Logical vector: Index of elements to filter}

\item{decreasing}{Logical: If TRUE, sort in descending order}
}
\description{
Filter order
}
\examples{
\dontrun{
x <- rnorm(10)
x
x[filter_order(x, x < 0)]
}
}
\author{
E.D. Gennatas
}
\keyword{internal}
