% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_ops.R
\name{fmt_thresh}
\alias{fmt_thresh}
\alias{fmt_thresh_right}
\title{Format rule thresholds}
\usage{
fmt_thresh(catsplit, feature, threshold, factor_levels)

fmt_thresh_right(catsplit, feature, threshold, factor_levels, cat_type)
}
\arguments{
\item{catsplit}{Logical: If TRUE, feature is categorical}

\item{feature}{Character: feature name}

\item{threshold}{Character: threshold as reported by lightgbm}

\item{factor_levels}{Named list of factor levels. Names should correspond to training
set column names.}
}
\description{
Format rule thresholds
}
\keyword{internal}
