% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rules.R
\name{format_LightRuleFit_rules}
\alias{format_LightRuleFit_rules}
\title{Format LightRuleFit rules}
\usage{
format_LightRuleFit_rules(x, space_after_comma = FALSE, decimal_places = NULL)
}
\arguments{
\item{x}{Vector, string: Logical expressions}

\item{space_after_comma}{Logical: If TRUE, place spaces after commas.}

\item{decimal_places}{Integer: Limit all floats (numbers of the form 9.9) to this many
decimal places}
}
\description{
Converts R-executable logical expressions to a more human-friendly format
}
\author{
EDG
}
