% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwhm2sigma.R
\name{fwhm2sigma}
\alias{fwhm2sigma}
\title{FWHM to Sigma}
\usage{
fwhm2sigma(fwhm)
}
\arguments{
\item{fwhm}{FWHM value}
}
\value{
sigma
}
\description{
Convert Full width at half maximum values to sigma
}
\examples{
fwhm2sigma(8)
# FWHM of 8 is equivalent to sigma = 3.397287
}
\author{
E.D. Gennatas
}
