% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm2table.R
\name{gam2table}
\alias{gam2table}
\title{Collect summary table (p-values) from list of massGAMs with same predictors,
different outcome ("massy")}
\usage{
gam2table(mods, modnames = NULL)
}
\arguments{
\item{x}{list of \link{gam} models}

\item{xnames}{Character, vector: names of models}

\item{include_anova_pvals}{Integer: 1 or 3; to output ANOVA I or III p-vals. NA to not}
}
\value{
\code{data.table} with glm summaries
}
\description{
Collect summary table (p-values) from list of massGAMs with same predictors,
different outcome ("massy")
}
\author{
E.D. Gennatas
}
