% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{get-names}
\alias{get-names}
\alias{getfactornames}
\title{Get factor/numeric/logical/character names from data.frame/data.table}
\usage{
getfactornames(x)
}
\arguments{
\item{x}{data.frame or data.table (or data.frame-compatible object)}
}
\value{
Character vector of column names of x with the specified class.
}
\description{
Get factor/numeric/logical/character names from data.frame/data.table
}
\author{
EDG
}
