% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_factor_levels}
\alias{get_factor_levels}
\title{Get factor levels from data.frame or similar}
\usage{
get_factor_levels(x)
}
\arguments{
\item{x}{data.frame or data.table.}
}
\value{
Named list of factor levels. Names correspond to column names.
}
\description{
Get factor levels from data.frame or similar
}
\author{
EDG
}
