% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_init.R
\name{get_factor_names}
\alias{get_factor_names}
\title{Get factor names}
\usage{
get_factor_names(x)
}
\arguments{
\item{x}{tabular data.}
}
\value{
Character vector of factor names.
}
\description{
Get factor names
}
\details{
This applied to tabular datasets used for supervised learning in rtemis,
where, by convention, the last column is the outcome variable and all other columns
are features.
}
\examples{
get_factor_names(iris)
}
\author{
EDG
}
