% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm2table.R
\name{get_gam_pvals}
\alias{get_gam_pvals}
\title{Get GAM model's p-values for parametric and spline terms}
\usage{
get_gam_pvals(m, warn = FALSE)
}
\description{
Get GAM model's p-values for parametric and spline terms
}
