% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ops.R
\name{get_holidays}
\alias{get_holidays}
\title{Get holidays from date vector}
\usage{
get_holidays(dates, holidays = c("LaborDay", "NewYearsDay", "ChristmasDay"))
}
\arguments{
\item{dates}{Date vector}

\item{holidays}{Character vector: holidays to extract}
}
\value{
Factor of length \code{length(dates)} with levels "Not Holiday", "Holiday"
}
\description{
Get holidays from date vector
}
