% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mode}
\alias{get_mode}
\title{Get the mode of a factor or integer}
\usage{
get_mode(x, na.rm = TRUE, getlast = TRUE, retain_class = TRUE)
}
\arguments{
\item{x}{Vector, factor or integer: Input data.}

\item{na.rm}{Logical: If TRUE, exclude NAs (using \code{na.exclude(x)}).}

\item{getlast}{Logical: If TRUE, get the last value in case of ties.}

\item{retain_class}{Logical: If TRUE, output is always same class as input.}
}
\value{
The mode of \code{x}
}
\description{
Returns the mode of a factor or integer
}
\examples{
\dontrun{
x <- c(9, 3, 4, 4, 0, 2, 2, NA)
get_mode(x)
x <- c(9, 3, 2, 2, 0, 4, 4, NA)
get_mode(x)
get_mode(x, getlast = FALSE)
}
}
\author{
EDG
}
