% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\name{get_palette}
\alias{get_palette}
\title{Get Color Palette}
\usage{
get_palette(palette = NULL, verbosity = 1L)
}
\arguments{
\item{palette}{Character: Name of palette to return. Default = NULL: available palette
names are printed and no palette is returned.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Character vector of colors for the specified palette, or invisibly returns
list of available palettes if \code{palette = NULL}.
}
\description{
\code{get_palette()} returns a color palette (character vector of colors).
Without arguments, prints names of available color palettes.
Each palette is a named list of hexadecimal color definitions which can be used with
any graphics function.
}
\examples{
# Print available palettes
get_palette()
# Get the Imperial palette
get_palette("imperial")
}
\author{
EDG
}
