% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rules.R
\name{get_rules}
\alias{get_rules}
\title{Get RuleFit rules}
\usage{
get_rules(
  mod,
  formatted = FALSE,
  collapse = TRUE,
  collapse.keep.names = FALSE,
  collapse.unique = TRUE
)
}
\arguments{
\item{mod}{Model created by \link{s_RuleFit} or \link{s_LightRuleFit}}

\item{formatted}{Logical: If TRUE, return human-readable rules, otherwise return
R-parsable rules}

\item{collapse}{Logical: If TRUE, collapse all rules to a single character vector}

\item{collapse.keep.names}{Logical: If TRUE, keep names when collapsing (will
be able to tell which run each rule came from). However, has no effect if
\code{collapse.unique = TRUE}, as \code{unique()} removes names.}

\item{collapse.unique}{Logical: If TRUE and \code{collapse = TRUE}, will return only
unique rules}
}
\description{
Get rules generated by \link{s_RuleFit} or \link{s_LightRuleFit}
}
\author{
ED Gennatas
}
