% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_ops.R
\name{get_vars_from_rules}
\alias{get_vars_from_rules}
\title{Extract variable names from rules}
\usage{
get_vars_from_rules(rules, unique = FALSE)
}
\arguments{
\item{rules}{Character vector: Rules.}

\item{unique}{Logical: If TRUE, return only unique variables.}
}
\value{
Character vector: Variable names.
}
\description{
Extract variable names from rules
}
\author{
E.D. Gennatas
}
