% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{getMode}
\alias{getMode}
\title{Get the mode of a factor or integer}
\usage{
getMode(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector, factor or integer}

\item{na.rm}{Logical: If TRUE, remove any NA values, if present, before finding mode.
Default = TRUE}
}
\value{
The mode of \code{x}
}
\description{
Returns the mode of a factor or integer
}
\author{
E.D. Gennatas
}
