% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getName.R
\name{getName}
\alias{getName}
\title{\pkg{rtemis} internal: Get Variable Name from Arguments}
\usage{
getName(x, alt = "x", max.nchar = 20)
}
\arguments{
\item{x}{Variable whose name you want to extract}

\item{alt}{Character: If name derived from \code{deparse(substitute(x))} exceeds \code{max.nchar} characters, use this name instead}

\item{max.nchar}{Integer: Maximum N of characters to allow for name}
}
\description{
Get the name of the variable passed as argument, limit number of characters in case of failure
}
\details{
One way to test is to use \link{learn} with x.name = NULL, y.name = NULL
}
\author{
Efstathios D Gennatas
}
\keyword{internal}
