% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_ops.R
\name{getnames}
\alias{getnames}
\alias{getnumericnames}
\alias{getlogicalnames}
\alias{getcharacternames}
\alias{getdatenames}
\title{Get names by string matching}
\usage{
getnames(
  x,
  pattern = NULL,
  starts_with = NULL,
  ends_with = NULL,
  ignore.case = TRUE
)

getnumericnames(x)

getlogicalnames(x)

getcharacternames(x)

getdatenames(x)
}
\arguments{
\item{x}{object with \code{names()} method}

\item{pattern}{Character: pattern to match anywhere in names of x}

\item{starts_with}{Character: pattern to match in the beginning of names of x}

\item{ends_with}{Character: pattern to match at the end of names of x}

\item{ignore.case}{Logical: If TRUE, well, ignore case. Default = TRUE}
}
\description{
Get names by string matching
}
\author{
E.D. Gennatas
}
