% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_ops.R
\name{getnamesandtypes}
\alias{getnamesandtypes}
\title{Get data.frame names and types}
\usage{
getnamesandtypes(x)
}
\arguments{
\item{x}{data.frame / data.table or similar}
}
\value{
character vector of column names with attribute "type" holding the class of each
column
}
\description{
Get data.frame names and types
}
