% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_NLS.R
\name{getTerms}
\alias{getTerms}
\title{Get terms of a formula}
\usage{
getTerms(formula)
}
\arguments{
\item{formula}{formula with more than x & y, e.g. \code{y ~ b * m ^ x}}
}
\description{
Get terms of a formula
}
