% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ggplot2.R
\name{ggtheme_dark}
\alias{ggtheme_dark}
\title{\pkg{rtemis} \code{ggplot2} dark theme}
\usage{
ggtheme_dark(
  base_size = 14,
  base_family = "Helvetica Neue",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  axis.text.size.rel = 1,
  legend.key.fill = NA,
  legend.text.size.rel = 1,
  legend.position = "right",
  strip.background.fill = "gray25"
)
}
\arguments{
\item{base_size}{Float: Base font size.}

\item{base_family}{Character: Font family.}

\item{base_line_size}{Float: Line size.}

\item{base_rect_size}{Float: Size for rect elements.}

\item{axis.text.size.rel}{Float: Relative size for axis text.}

\item{legend.key.fill}{Color: Fill color for legend.}

\item{legend.text.size.rel}{Float: Relative size for legend text.}

\item{legend.position}{Character: Legend position, "top", "bottom", "right", "left" Default  = "right"}

\item{strip.background.fill}{Color: Fill color from facet labels.}
}
\description{
\pkg{rtemis} \code{ggplot2} dark theme
}
\examples{
\dontrun{
(p <- ggplot(iris, aes(Sepal.Length, Petal.Length, color = Species)) +
  geom_point() +
  ggtheme_light())
}
}
\author{
EDG
}
