% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ggplot2.R
\name{ggtheme_light}
\alias{ggtheme_light}
\title{\pkg{rtemis} \code{ggplot2} light theme}
\usage{
ggtheme_light(
  base_size = 14,
  base_family = "Helvetica Neue",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  axis_text_size_rel = 1,
  legend_key_fill = NA,
  legend_text_size_rel = 1,
  legend_position = "right",
  strip_background_fill = "grey85"
)
}
\arguments{
\item{base_size}{Float: Base font size.}

\item{base_family}{Character: Font family.}

\item{base_line_size}{Float: Line size.}

\item{base_rect_size}{Float: Size for rect elements.}

\item{axis_text_size_rel}{Float: Relative size for axis text.}

\item{legend_key_fill}{Color: Fill color for legend.}

\item{legend_text_size_rel}{Float: Relative size for legend text.}

\item{legend_position}{Character: Legend position, "top", "bottom", "right", "left" Default  = "right"}

\item{strip_background_fill}{Color: Fill color from facet labels.}
}
\value{
\code{ggplot2}, \code{theme} object.
}
\description{
\pkg{rtemis} \code{ggplot2} light theme
}
\examples{
\dontrun{
(p <- ggplot(iris, aes(Sepal.Length, Petal.Length, color = Species)) +
  geom_point() +
  ggtheme_light())
}
}
\author{
EDG
}
