% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm2table.R
\name{glm2table}
\alias{glm2table}
\title{Collect summary tablew from list of massGLMs with same predictors, different outcome}
\usage{
glm2table(x, xnames = NULL)
}
\arguments{
\item{x}{list of \link{glm} models}

\item{xnames}{Character, vector: names of models, i.e. normally name of}
}
\value{
\code{data.table} with glm summaries
}
\description{
Collect summary tablew from list of massGLMs with same predictors, different outcome
}
\author{
E.D. Gennatas
}
