% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric.
}
\description{
Geometric mean
}
\examples{
\dontrun{
x <- c(1, 3, 5)
mean(x)
gmean(x)
# same as, but a little faster than:
exp(mean(log(x)))
}
}
\author{
EDG
}
