% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Bayesian Gaussian Processes [R]}
\usage{
gp(
  x,
  y,
  new.x = NULL,
  x.name = "x",
  y.name = "y",
  print.plot = TRUE,
  lwd = 3,
  cex = 1.2,
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix of features, i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{new.x}{(Optional) Numeric vector or matrix of new set of features
Must have same set of columns as \code{x}}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{print.plot}{Logical: if TRUE, draw plot when done}

\item{lwd}{Line width for plotting}

\item{cex}{Character expansion factor for plotting}

\item{par.reset}{Logical. Reset \code{par} to its original state}

\item{...}{Additional arguments to be passed to tgp::bgp}
}
\description{
Fit a gaussian process
}
\author{
E.D. Gennatas
}
