% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot3.map.R
\name{gplot3.map}
\alias{gplot3.map}
\title{Plot US state or county choropleth map}
\usage{
gplot3.map(
  dat,
  regions = c("states", "state", "counties", "county"),
  include = NULL,
  exclude = NULL,
  limits = NULL,
  trans = "identity",
  col.lo = "#0290EE",
  col.mid = "black",
  col.hi = "#FE4AA3",
  col.na = "grey40",
  colorscale.midpoint = 0,
  colorbar.height = 0.15,
  colorbar.width = 0.03,
  border.width = 0.1,
  main = NULL,
  legend.title = NULL,
  scale.accuracy = 0.1,
  labelify = TRUE,
  theme = rt_gtheme_map(),
  labels = FALSE,
  col.labels = "gray50",
  filename = NULL,
  file.width = 7,
  file.height = 5,
  ...
)
}
\arguments{
\item{dat}{data.frame with 2 columns: fips, value}

\item{regions}{Vector, character: "states", "state", "counties", "county"}

\item{include}{Vector, character: Names of states or counties to include}

\item{exclude}{Vector, character: Names of states or counties to exclude}

\item{col.lo}{Color: Color mapped ot lowest negative values. Default = "#0290EE"}

\item{col.mid}{Color: Color mapped to value \code{colorscale.midpoint}, only used if
\code{colorscale.midpoint} is not NULL. If set to NA, \code{colorscale.midpoint} will be
set to NULL. Default = "black".}

\item{col.hi}{Color: Color mapped to highest positive values. Default = "#FE4AA3"}

\item{col.na}{Color: Color mapped to NA values. Default = "gray40"}

\item{colorscale.midpoint}{Float: Midpoint for colorscale. Default = 0}

\item{colorbar.height}{FLoat: Colorbar height, will be used as
\code{ggplot2::unit(colorbar.height, "npc")}}

\item{colorbar.width}{FLoat: Colorbar width, will be used as
\code{ggplot2::unit(colorbar.width, "npc")}}

\item{border.width}{Float: Map border width. Default = .1}

\item{main}{Character: Main title. Default = NULL}

\item{legend.title}{Character: Legend title. Default = NULL}

\item{theme}{\strong{ggplot2} theme to use. Default = \code{rt_gtheme_map()}}

\item{labels}{Logical: If TRUE, label states. Default = FALSE}

\item{col.labels}{Color for \code{labels}. Default = "gray50"}
}
\description{
Plot US state or county choropleth map using \strong{ggplot2} based on
\code{usmap::plot_usmap()}
}
\author{
E.D. Gennatas
}
