% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_ops.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Inverse of \code{\%in\%}}
\usage{
x \%notin\% set
}
\arguments{
\item{x}{vector}

\item{set}{vector}
}
\value{
logical vector
}
\description{
Inverse of \code{\%in\%}
}
\examples{
3 \%notin\% c(2, 4, 6) # TRUE
4 \%notin\% 3:5 # FALSE
}
