% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{graph_node_metrics}
\alias{graph_node_metrics}
\title{Node-wise (i.e. vertex-wise) graph metrics}
\usage{
graph_node_metrics(x, verbosity = 1L)
}
\arguments{
\item{x}{\pkg{igraph} network.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{data.frame}.
}
\description{
Node-wise (i.e. vertex-wise) graph metrics
}
\examples{
\dontrun{
datcor <- cor(rnormmat(20, 20, seed = 2021))
datcor[sample(seq(datcor), 250)] <- 0
x <- igraph::graph_from_adjacency_matrix(
  adjmatrix = datcor,
  mode = "lower",
  weighted = TRUE,
  diag = FALSE
)

graph_node_metrics(x)
}
}
\author{
EDG
}
