% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearchLearn.R
\name{gridCheck}
\alias{gridCheck}
\title{\pkg{rtemis} internal: Grid check}
\usage{
gridCheck(...)
}
\arguments{
\item{...}{Parameters; will be converted to a list}
}
\description{
Checks if grid search needs to be performed.
All tunable parameters should be passed to this function, individually or as a list. If any
argument has more than one assigned values, the function returns TRUE, otherwise FALSE. This can
be used to check whether \link{gridSearchLearn} must be run.
}
\details{
The idea is that if you know which parameter values you want to use, you define them directly
  e.g. \code{alpha = 0, lambda = .2}.
If you don't know, you enter the set of values to be tested,
  e.g. \code{alpha = c(0, .5, 1), lambda = seq(.1, 1, .1)}.
}
