% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtinit.R
\name{gridSummary}
\alias{gridSummary}
\title{\code{rtemis-internals}: \code{gridSummary}}
\usage{
gridSummary(...)
}
\description{
Pretty print list of parameters with more than one value, 
which will be used in a grid search
}
\keyword{internal}
