% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gt_table}
\alias{gt_table}
\title{Greater-than Table}
\usage{
gt_table(x = list(), x_name = NULL, na.rm = TRUE, verbosity = 1L)
}
\arguments{
\item{x}{List of vectors of same length.}

\item{x_name}{Character: Name of measure being compared.}

\item{na.rm}{Passed to \code{sum} to handle missing values.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Data frame of N times each vector is greater than the others.
}
\description{
Compare vectors element-wise, and tabulate N times each vector is greater than the others
}
\author{
EDG
}
