% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtTable.R
\name{gtTable}
\alias{gtTable}
\title{Greater-than Table}
\usage{
gtTable(x = list(), x.name = NULL, na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{List of vectors of same length}

\item{x.name}{Character: Name of measure being compared}

\item{na.rm}{Passed to \code{sum} to handle missing values}

\item{verbose}{Logical: If TRUE, write output to console}
}
\description{
Compare vectors element-wise, and tabulate N times each vector is greater than the others
}
\author{
E.D. Gennatas
}
