% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{htest}
\alias{htest}
\title{Basic Bivariate Hypothesis Testing and Plotting}
\usage{
htest(
  y,
  group = NULL,
  x = NULL,
  yname = NULL,
  groupname = NULL,
  xname = NULL,
  test = c("t.test", "wilcox.test", "aov", "kruskal.test", "chisq.test", "fisher.test",
    "cor.test", "pearson", "kendall", "spearman", "ks"),
  print.plot = TRUE,
  plot.args = list(),
  theme = getOption("rt.theme", "lightgrid"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{y}{Float, vector: Outcome of interest}

\item{group}{Factor: Groups to compare}

\item{x}{Float, vector: Second outcome for correlation tests}

\item{yname}{Character: y variable name}

\item{groupname}{Character: group variable name}

\item{xname}{Character: x variable name}

\item{test}{Character: Test to use; one of:
\itemize{
\item Continuous outcome by group: "t.test", "wilcox.test", "aov", "kruskal.test"
\item Categorical outcome by group: "chisq.test", "fisher.test", "cor.test"
\item Two continuous variables: "pearson", "kendall", "spearman"
}}

\item{print.plot}{Logical: If TRUE, print plot. Default = TRUE}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'darkgrid')}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}

\item{...}{Additional arguments to pass to test call}
}
\description{
Basic Bivariate Hypothesis Testing and Plotting
}
\examples{
\dontrun{
# t.test, wilcoxon
y <- c(rnorm(200, 2, 1.2), rnorm(300, 2.5, 1.4))
group <- c(rep(1, 200), rep(2, 300))

ht_ttest <- htest(y, group, test = "t.test")
ht_wilcoxon <- htest(y, group, test = "wilcox.test")

# aov, kruskal
y <- c(rnorm(200, 2, 1.2), rnorm(300, 2.5, 1.4), rnorm(100, 2.3, 1.1))
group <- c(rep(1, 200), rep(2, 300), rep(3, 100))

ht_aov <- htest(y, group, test = "aov")
ht_kruskal <- htest(y, group, test = "kruskal.test")

# chisq, fisher
y <- c(sample(c(1, 2), 100, T, c(.7, .3)), sample(c(1, 2), 100, T, c(.35, .65)))
group <- c(rep(1, 100), rep(2, 100))
ht_chisq <- htest(y, group, test = "chisq")
ht_fisher <- htest(y, group, test = "fisher")

# cor.test
x <- rnorm(300)
y <- x*.3 + rnorm(300)
ht_pearson <- htest(x = x, y = y, test = "pearson")
ht_kendall <- htest(x = x, y = y, test = "kendall")
ht_kendall <- htest(x = x, y = y, test = "spearman")
}
}
\author{
E.D. Gennatas
}
