% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{hytboost}
\alias{hytboost}
\title{\pkg{rtemis} internal: Boosting of Hybrid Trees}
\usage{
hytboost(
  x,
  y,
  x.valid = NULL,
  y.valid = NULL,
  resid = NULL,
  boost.obj = NULL,
  mod.params = list(),
  case.p = 1,
  learning.rate = 0.1,
  max.iter = 10,
  init = mean(y),
  cxrcoef = FALSE,
  print.progress.every = 5,
  print.error.plot = "final",
  base.verbose = FALSE,
  verbose = TRUE,
  trace = 0,
  prefix = NULL,
  print.plot = TRUE,
  plot.theme = "darkgrid",
  plot.type = "l",
  ...
)
}
\arguments{
\item{x}{Data frame: Input features}

\item{y}{Vector: Output}

\item{mod.params}{Named list of arguments for \code{mod}}

\item{learning.rate}{Float (0, 1] Learning rate for the additive steps}

\item{max.iter}{Integer: Maximum number of iterations (additive steps) to perform. Default = 10}

\item{init}{Float: Initial value for prediction. Default = mean(y)}

\item{cxrcoef}{Logical: If TRUE, pass \code{cxr = TRUE, cxrcoef = TRUE} to \link{predict.hytreeRaw}}

\item{print.error.plot}{String or Integer: "final" plots a training and validation (if available) error curve at the
end of training. If integer, plot training and validation error curve every this many iterations
during training
for each base learner}

\item{base.verbose}{Logical: \code{verbose} argument passed to learner}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If > 0, print diagnostic info to console}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}. Default = TRUE}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{...}{Additional parameters to be passed to learner}

\item{mod}{Algorithm to boost, for options, see \link{modSelect}}

\item{tolerance}{Float: If training error <= this value, training stops}

\item{tolerance.valid}{Float: If validation error <= this value, training stops}
}
\value{
\code{hytboost} object
}
\description{
Boosted Hybrid Trees. This is lower-level than \code{s.*} functions
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
