% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytreenow.R
\name{hytreenow}
\alias{hytreenow}
\title{\pkg{rtemis} internal: Low-level Hybrid Tree procedure}
\usage{
hytreenow(
  x,
  y,
  max.depth = 5,
  alpha = 0,
  lambda = 0.1,
  lambda.seq = NULL,
  minobsinnode = 2,
  minobsinnode.lin = 10,
  shrinkage = 1,
  part.minsplit = 2,
  part.xval = 0,
  part.max.depth = 1,
  part.cp = 0,
  part.minbucket = 5,
  init = mean(y),
  lin.type = c("glmnet", "cv.glmnet", "lm.ridge", "glm"),
  cv.glmnet.nfolds = 5,
  cv.glmnet.lambda = "lambda.min",
  verbose = FALSE,
  trace = 0
)
}
\arguments{
\item{x}{data.frame}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{max.depth}{[gS] Integer: Max depth of additive tree. Default = 3}

\item{alpha}{[gS] Float: \code{lincoef} alpha Overrides \code{lincoef.params} alpha}

\item{lambda}{[gS] Float: \code{lincoef} lambda. Overrides \code{lincoef.params} lambda}

\item{minobsinnode}{[gS] Integer: Minimum N observations needed in node, before considering splitting}

\item{part.max.depth}{Integer: Max depth for each tree model within the additive tree}

\item{part.cp}{[gS] Float: Minimum complexity needed to allow split by \code{rpart}. Default = 0}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If higher than 0, will print more information to the console. Default = 0}
}
\description{
Train a Hybrid Tree for Regression
}
\details{
Note that lambda is treated differently by \code{glmnet::glmnet} and \code{MASS::lm.ridge}
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
