% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{ifNotNull}
\alias{ifNotNull}
\title{Say No to \code{NULL}}
\usage{
ifNotNull(x, defType)
}
\arguments{
\item{x}{Input of any type, may be NULL}

\item{defType}{If \code{x} is NULL, return empty vector of this type. Options: list, numeric,
character, integer}
}
\description{
Returns the input, unless it is NULL, in which case it returns an empty vector / list, etc
of defined type
}
\details{
This can be useful when creating S4, RC, or R6 objects
}
\author{
Efstathios D. Gennatas
}
