% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCasesByRules.R
\name{indexCasesByRules}
\alias{indexCasesByRules}
\title{Index cases by rules}
\usage{
indexCasesByRules(x, rules, verbose = TRUE)
}
\arguments{
\item{x}{Matrix / data frame: Input features}

\item{rules}{Vector, string: Rules (MUST be string, not factor)}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\description{
Get an index of which cases match which rule - meant for cases where
}
\author{
Efstathios D. Gennatas
}
