% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_check.R
\name{inherits_test}
\alias{inherits_test}
\title{Test class of object}
\usage{
inherits_test(x, cl)
}
\arguments{
\item{x}{Object to check}

\item{cl}{Character: class to check against}
}
\description{
Test class of object
}
\author{
EDG
}
\keyword{internal}
