% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_project_dir}
\alias{init_project_dir}
\title{Initialize Project Directory}
\usage{
init_project_dir(path, output_dir = "Out", verbosity = 1L)
}
\arguments{
\item{path}{Character: Path to initialize project directory in.}

\item{output_dir}{Character: Name of output directory to create.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Character: the path where the project directory was initialized, invisibly.
}
\description{
Initializes Directory Structure: "R", "Data", "Results"
}
\author{
EDG
}
