% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_init.R
\name{intro}
\alias{intro}
\title{\code{rtemis-internals}: \code{intro}}
\usage{
intro(
  message = NULL,
  verbose = TRUE,
  logFile = NULL,
  call.depth = 1,
  newline.pre = FALSE
)
}
\description{
Intro
}
\details{
Starts function execution timer and opens log file.
Pairs with \code{outro}. Unfortunately the errors and warnings (the stderr; sink(type = "message"))
cannot get split to file and console, so we keep in console only
}
\keyword{internal}
