% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logit}
\usage{
invlogit(x)
}
\arguments{
\item{x}{Float: Input data}
}
\value{
The inverse logit of the input
}
\description{
Inverse Logit
}
\author{
Efstathios D. Gennatas
}
