% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{is.constant}
\alias{is.constant}
\title{Check if vector is constant}
\usage{
is.constant(x)
}
\arguments{
\item{x}{Vector: Input}
}
\description{
Check if vector is constant
}
\author{
Efstathios D. Gennatas
}
