% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{is.discrete}
\alias{is.discrete}
\title{Check if variable is discrete (factor or integer)}
\usage{
is.discrete(x)
}
\arguments{
\item{x}{Input}
}
\description{
Check if variable is discrete (factor or integer)
}
\author{
E.D. Gennatas
}
