% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_check.R
\name{is_check}
\alias{is_check}
\title{Check type of object}
\usage{
is_check(x, fn)
}
\arguments{
\item{x}{Object to check}

\item{fn}{Function to check against, any \verb{is.*} function, e.g. \code{is.character}}
}
\value{
Logical
}
\description{
Check type of object
}
\examples{
\dontrun{
is_check("papaya", is.character) # TRUE
is_check(c(1, 2.5, 3.2), is.integer) # FALSE
is_check(iris, is.list) # TRUE
}
}
\author{
EDG
}
\keyword{internal}
