% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{is_constant}
\alias{is_constant}
\title{Check if vector is constant}
\usage{
is_constant(x, skip_missing = FALSE)
}
\arguments{
\item{x}{Vector: Input}

\item{skip_missing}{Logical: If TRUE, skip NA values before testing}
}
\description{
Check if vector is constant
}
\examples{
\dontrun{
x <- rep(9, 1000000)
is_constant(x)
x[10] <- NA
is_constant(x)
is_constant(x, skip_missing = TRUE)
}
}
\author{
E.D. Gennatas
}
