% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_check.R
\name{is_test}
\alias{is_test}
\title{Test type of object}
\usage{
is_test(x, fn)
}
\arguments{
\item{x}{Object to check}

\item{fn}{Function to check against, any \verb{is.*} function, e.g. \code{is.character}}
}
\value{
NULL (invisibly)
}
\description{
Test type of object
}
\author{
EDG
}
\keyword{internal}
