% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JL_splitlin.R
\name{JL_splitlin}
\alias{JL_splitlin}
\title{Split where lines would minimize error}
\usage{
JL_splitlin(x, y, wts = rep(1, length(y)), lambda = 0.1, verbose = TRUE)
}
\description{
Split where lines would minimize error
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
