% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyval_reshape.R
\name{keyval_reshape}
\alias{keyval_reshape}
\title{Long to wide key-value reshaping}
\usage{
keyval_reshape(
  x,
  id_name,
  key_name,
  value_name,
  positive = 1,
  negative = 0,
  xname = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{data.table} object}

\item{id_name}{Character: Name of column in \code{x} that defines the IDs
identifying individual rows}

\item{key_name}{Character: Name of column in \code{x} that holds the key}

\item{value_name}{Character: Name of column in \code{x} thatholds the values
that correspond to the key}

\item{positive}{Numeric or Character: Used to fill id ~ key combination 
present in the long format input \code{x}}

\item{negative}{Numeric or Character: Used to fill id ~ key combination
NOT present in the long format input \code{x}}

\item{verbose}{Logical: If TRUE, print messages to the console}
}
\description{
Reshape a long format \code{data.table} using key-value pairs with
\code{data.table:dcast}
}
\author{
E.D. Gennatas
}
