% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{kfold}
\alias{kfold}
\title{K-fold Resampling}
\usage{
kfold(
  x,
  k = 10,
  stratify.var = NULL,
  strat.n.bins = 4,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Input Vector}

\item{k}{Integer: Number of folds. Default = 10}

\item{stratify.var}{Numeric vector (optional): Variable used for stratification. Defaults to \code{y}}

\item{strat.n.bins}{Integer: Number of groups to use for stratification for
\code{resampler = "strat.sub" / "strat.boot"}}

\item{seed}{Integer: (Optional) Set seed for random number generator, in order to make output reproducible.
See \code{?base::set.seed}}

\item{verbose}{Logical: If TRUE, print messages to screen}
}
\description{
K-fold Resampling
}
\author{
Efstathios D. Gennatas
}
