% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2niftis.R
\name{labels2niftis}
\alias{labels2niftis}
\title{Write Matrix to Multiple Nifti Files}
\usage{
labels2niftis(
  datamat,
  labeledNifti,
  prefix,
  verbose = TRUE,
  n.cores = parallel::detectCores()
)
}
\arguments{
\item{datamat}{Matrix / Data Frame: Input}

\item{labeledNifti}{Character: Path to labeled nifti file}

\item{prefix}{Character: File output prefix}

\item{verbose}{Logical: If TRUE, print messages to output}

\item{n.cores}{Integer: Number of cores to use}
}
\description{
Writes matrix /data frame to nifti files columnwise.
Each column of the matrix should correspond to the labels of a nifti file.
i.e. nrow of matrix = n of labels in labeledNifti
Niftis are written in parallel using \code{parallel::parApply}
}
\author{
E.D. Gennatas
}
