% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn}
\alias{learn}
\title{Supervised Learning with \pkg{rtemis}}
\usage{
learn(
  x,
  y = NULL,
  mod = "ppr",
  x.test = NULL,
  y.test = NULL,
  x.name = NULL,
  y.name = NULL,
  args = list(),
  question = NULL,
  verbose = TRUE,
  print.plot = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{mod}{Character: Learner to use. To get list of options, run \code{algSelect()}}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{args}{Optional list of parameters to be passed to learner}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}.}

\item{...}{Additional arguments to be passed to learner}
}
\description{
Train any \pkg{rtemis} model
}
\details{
\code{args} and \code{...} allow you to either pass named arguments, or a list of arguments (or both)
}
\author{
E.D. Gennatas
}
