% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algSelect.R
\name{learnSelect}
\alias{learnSelect}
\title{Select \pkg{rtemis} Learner}
\usage{
learnSelect(alg, fn = FALSE, desc = FALSE)
}
\arguments{
\item{alg}{Character: Model name. Case insensitive. e.g. "XGB" for xgboost}

\item{fn}{Logical: If TRUE, return function, otherwise name of function. Defaults to FALSE}

\item{desc}{Logical: If TRUE, return full name / description of algorithm \code{alg}}
}
\value{
function or name of function (see param \code{fn}) or full name of algorithm (\code{desc})
}
\description{
Accepts learner name (supports abbreviations) and returns \pkg{rtemis} function name or
the function itself.
If run with no parameters, prints list of available algorithms.
}
\author{
E.D. Gennatas
}
