% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_ops.R
\name{lgb2rules}
\alias{lgb2rules}
\title{Convert LightGBM Booster to set of rules}
\usage{
lgb2rules(Booster, n_iter = NULL, xnames, factor_levels)
}
\arguments{
\item{n_iter}{Integer: Number of trees to convert to rules}

\item{xnames}{Character vector: Names of features}

\item{x}{LightGBM Booster object}
}
\value{
Character vector of rules
}
\description{
Convert LightGBM Booster to set of rules
}
\keyword{internal}
