% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtree.R
\name{likelihoodMediboostChooseFeat}
\alias{likelihoodMediboostChooseFeat}
\title{\pkg{rtemis} internal: likelihoodMediboostChooseFeat
Selects a feature with maximum information gain and provides the decision values
and column index for the chosen feature}
\usage{
likelihoodMediboostChooseFeat(
  x,
  y,
  catPredictors,
  funcValue,
  weights,
  colIdx,
  verbose = TRUE
)
}
\description{
\pkg{rtemis} internal: likelihoodMediboostChooseFeat
Selects a feature with maximum information gain and provides the decision values
and column index for the chosen feature
}
\keyword{internal}
