% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtree.R
\name{likelihoodMediboostSplitNode}
\alias{likelihoodMediboostSplitNode}
\title{\pkg{rtemis} internal: likelihoodMediboostSplitNode}
\usage{
likelihoodMediboostSplitNode(
  x,
  y,
  weights,
  catPredictors,
  obsValues,
  nodeValue,
  colIdx,
  depth,
  depthLimit,
  learning.rate,
  gamma,
  membership,
  min.membership,
  steps.past.min.membership,
  min.mem.counter,
  update,
  min.update = ifelse(update == "exponential", 1000, 10),
  hessian,
  min.hessian,
  save.rpart,
  verbose = TRUE
)
}
\description{
Recursive function that returns a node structure under the MediBoost paradigm
}
\keyword{internal}
