% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{list2csv}
\alias{list2csv}
\title{Write list elements to CSV files}
\usage{
list2csv(x, outdir)
}
\arguments{
\item{x}{List containing R objects to be written to CSV (e.g. data.frames,
matrices, etc.)}

\item{outdir}{Character: Path to output directory}
}
\value{
Nothing, writes CSV files to \code{outdir}.
}
\description{
Write list elements to CSV files
}
\author{
EDG
}
