% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadedPackageVersions.R
\name{loadedPackageVersions}
\alias{loadedPackageVersions}
\title{Get version of all loaded packages (namespaces)}
\usage{
loadedPackageVersions()
}
\value{
Data frame with columns "Package_Name" and "Version"
}
\description{
Get version of all loaded packages (namespaces)
}
\author{
Efstathios D Gennatas
}
