% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(x, x0 = 0, L = 1, k = 1)
}
\arguments{
\item{x}{Float: Input}

\item{x0}{x-value of the midpoint.}

\item{L}{maximum value.}

\item{k}{steepness of the curve.}
}
\description{
Logistic function
}
