% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{logloss}
\alias{logloss}
\title{Log Loss for a binary classifier}
\usage{
logloss(true, estimated.prob)
}
\arguments{
\item{true}{Factor: True labels. First level is the positive case}

\item{estimated.prob}{Float, vector: Estimated probabilities}
}
\description{
Log Loss for a binary classifier
}
\author{
E.D. Gennatas
}
