% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{lotri2edgeList}
\alias{lotri2edgeList}
\title{Connectivity Matrix to Edge List}
\usage{
lotri2edgeList(A, filename = NULL, verbosity = 1L)
}
\arguments{
\item{A}{Square matrix}

\item{filename}{Character: Path for csv file. Defaults to "conmat2edgelist.csv"}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Data frame with columns: NodeA, NodeB, Weight
}
\description{
Turn the lower triangle of a connectivity matrix (e.g. correlation matrix or similar)
to an edge list of the form:
Source, Target, Weight
}
\details{
The output can be read, for example, into gephi
}
\author{
EDG
}
