% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_supervised.R
\name{make_formula}
\alias{make_formula}
\title{Make formula}
\usage{
make_formula(x, output = "character")
}
\arguments{
\item{x}{data.frame}
}
\value{
character
}
\description{
Makes a formula from a data.frame assuming the last column is the outcome
}
\author{
EDG
}
\keyword{internal}
