% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_key.R
\name{make_key}
\alias{make_key}
\title{Make key from data.table id - description columns}
\usage{
make_key(x, code_name, description_name, filename = NULL)
}
\arguments{
\item{x}{Input data.table}

\item{code_name}{Character: Name of column name that holds codes}

\item{description_name}{Character: Name of column that holds descriptions}

\item{filename}{Character: Path to file to save CSV with key}
}
\description{
Make key from data.table id - description columns
}
\author{
E.D. Gennatas
}
