% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{make_path}
\alias{make_path}
\title{Expand, normalize, concatenate, clean path}
\usage{
make_path(..., expand_path = TRUE)
}
\arguments{
\item{...}{Character: Parts of path to concatenate.}

\item{expand_path}{Logical: If TRUE, expand concatenated path using \link{path.expand}.}
}
\value{
Character: Path.
}
\description{
Expand, normalize, concatenate, clean path
}
\author{
EDG
}
