% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_Supervised.R
\name{make_SupervisedRes}
\alias{make_SupervisedRes}
\title{Make SupervisedRes}
\usage{
make_SupervisedRes(
  algorithm,
  type,
  models,
  preprocessor,
  hyperparameters,
  tuner_parameters,
  outer_resampler,
  y_training,
  y_test,
  predicted_training,
  predicted_test,
  predicted_prob_training,
  predicted_prob_test,
  se_training = NULL,
  se_test = NULL,
  xnames = character(),
  varimp = NULL,
  question = character(),
  extra = NULL
)
}
\description{
Make SupervisedRes
}
\author{
EDG
}
\keyword{internal}
