% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massCART.R
\name{massCART}
\alias{massCART}
\title{Mass-univariate CART prediction and variable importance}
\usage{
massCART(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  metric = NULL,
  minsplit = 2,
  minbucket = round(minsplit/3),
  cp = 0.01,
  maxcompete = 0,
  maxsurrogate = 0,
  usesurrogate = 2,
  surrogatestyle = 0,
  maxdepth = 22,
  xval = 0,
  ipw = FALSE,
  upsample = FALSE,
  downsample = FALSE,
  resample.seed = NULL,
  n.cores = 1,
  parallel.type = ifelse(.Platform$OS.type == "unix", "fork", "psock"),
  save.mod = FALSE,
  grid.print.plot = FALSE,
  verbose = TRUE,
  grid.verbose = TRUE,
  print.plot = FALSE,
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{metric}{Character: Metric to minimize, or maximize if \code{maximize = TRUE} during grid search.
Default = NULL, which results in "Balanced Accuracy" for Classification,
"MSE" for Regression, and "Coherence" for Survival Analysis.}

\item{minsplit}{[gS] Integer: Minimum number of cases that must belong in a node before considering a split.
Default = 2}

\item{minbucket}{[gS] Integer: Minimum number of cases allowed in a child node. Default = round(minsplit/3)}

\item{cp}{[gS] Float: Complexity threshold for allowing a split. Default = .01}

\item{maxcompete}{Integer: The number of competitor splits saved in the output Default = 0}

\item{maxsurrogate}{Integer: The number of surrogate splits retained in the output (See \code{rpart::rpart.control}).
Default = 0}

\item{maxdepth}{[gS] Integer: Maximum depth of tree. Default = 20}

\item{ipw}{Logical: If TRUE, apply inverse probability weighting (for Classification only).
Note: If \code{weights} are provided, \code{ipw} is not used. Default = TRUE}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Caution: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{n.cores}{Integer: Number of cores to use}

\item{save.mod}{Logical: If TRUE, save all output to an RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{grid.verbose}{Logical: Passed to \link{gridSearchLearn}}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}. Default = TRUE}
}
\description{
Predict outcome from each predictor separately and rank by percent Variance explained
or Classification Accuracy
}
\author{
E.D. Gennatas
}
