% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R
\name{massGLM}
\alias{massGLM}
\title{Mass-univariate GLM Analysis}
\usage{
massGLM(x, y, scale_y = NULL, center_y = NULL, verbosity = 1L)
}
\arguments{
\item{x}{data.frame or similar: Predictor variables}

\item{y}{data.frame or similar: Each column is a different outcome. The function will train one
GLM for each column of \code{y}.}

\item{scale_y}{Logical: If TRUE, scale each column of \code{y} to have mean 0 and sd 1.}

\item{center_y}{Logical: If TRUE, center each column of \code{y} to have mean 0.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{MassGLM} object.
}
\description{
Mass-univariate GLM Analysis
}
\examples{
\dontrun{
# x: outcome of interest as first column, optional covariates in the other columns
# y: features whose association with x we want to study
set.seed(2022)
y <- data.table(rnormmat(500, 40))
x <- data.table(
  x1 = y[[3]] - y[[5]] + y[[14]] + rnorm(500),
  x2 = y[[21]] + rnorm(500)
)
massmod <- massGLM(x, y)
plot(massmod)
}
}
\author{
EDG
}
