% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massUni.R
\name{massUni}
\alias{massUni}
\title{Mass-univariate Analysis}
\usage{
massUni(
  x,
  y,
  mod = "gam",
  save.mods = FALSE,
  verbose = TRUE,
  n.cores = rtCores,
  ...
)
}
\arguments{
\item{x}{Matrix / data frame of features}

\item{y}{Matrix / data frame of outcomes}

\item{mod}{\pkg{rtemis} algorithm to use. Options: run \code{select_learn()}}

\item{save.mods}{Logical: If TRUE, save fitted models}

\item{verbose}{Logical: If TRUE, print messages during run}

\item{n.cores}{Integer: Number of cores to use}

\item{...}{Arguments to be passed to \code{mod}}
}
\description{
Run a mass-univariate analysis: same features (predictors)
on multiple outcomes
}
\author{
E.D. Gennatas
}
