% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchcases.R
\name{matchcases}
\alias{matchcases}
\title{Match cases by covariates}
\usage{
matchcases(
  target,
  pool,
  n.matches = 1,
  target.id = NULL,
  pool.id = NULL,
  exactmatch.factors = TRUE,
  exactmatch.cols = NULL
)
}
\arguments{
\item{target}{data.frame you are matching to}

\item{pool}{data.frame where you are looking for matches from}

\item{target.id}{String: Column name in \code{target} that holds unique cases IDs. Default = NULL,
in which case integer case numbers will be used}

\item{pool.id}{String: Same as \code{target.id} for \code{pool}}

\item{exactmatch.factors}{Logical: If TRUE, selected cases will have to exactly match factors
available in \code{target}}

\item{exactmatch.cols}{String: Names of columns of variables that should be matched exactly}
}
\description{
Match cases by exact match and by distance between covariates
}
\examples{
target <- data.frame(Sex = factor(c(1, 1, 0, 0)),
                     Handedness = factor(c(1, 1, 0, 1)),
                     Age = c(21, 27, 39, 24),
                     Var = c(.7, .8, .9, .6))

set.seed(2021)
pool <- data.frame(Sex = factor(sample(c(0, 1), 50, TRUE)),
                   Handedness = factor(sample(c(0, 1), 50, TRUE, c(.1, .9))),
                   Age = sample(16:42, 50, TRUE),
                   Var = rnorm(50))
mc <- matchcases(target, pool, 2)
}
\author{
E.D. Gennatas
}
