% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCasesByRules.R
\name{matchCasesByRules}
\alias{matchCasesByRules}
\title{Match Rules to Cases}
\usage{
matchCasesByRules(x, rules, verbose = TRUE)
}
\arguments{
\item{x}{Matrix / data frame: Input features}

\item{rules}{Character vector: Rules}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\value{
cases-by-rules matrix (binary; 1: match, 0: no match)
}
\description{
Match Rules to Cases
}
\author{
E.D. Gennatas
}
\keyword{internal}
