% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediboost.R
\name{mediboost}
\alias{mediboost}
\title{Additive Tree}
\usage{
mediboost(
  x,
  y,
  catPredictors = NULL,
  depthLimit = 8,
  learningRate = 1,
  gamma = 0,
  update = "exponential",
  min.update = ifelse(update == "polynomial", 10, 1000),
  weights = NULL,
  autoweights = FALSE,
  min.hessian = 0.01,
  min.membership = 0,
  steps.past.min.membership = 2,
  rpart.parms = NULL,
  save.rpart = FALSE,
  verbose = TRUE,
  trace = 1
)
}
\arguments{
\item{x}{Matrix / Data frame of features}

\item{y}{Integer {-1, 1}: Vector of binary outcomes}

\item{catPredictors}{Optional: Logical vector indicating categorical features}

\item{depthLimit}{Integer: Maximum depth of tree to grow}

\item{learningRate}{learning rate for the Newton Raphson step that updates the
function values of the node for \code{update = "exponential"}}

\item{gamma}{Float (0, 1): Accelaration factor}

\item{update}{Character: "Exponential" or "Polynomial"}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Trains a binary decision tree classifier using MediBoost Tree-Structured Boosting
}
\details{
Outcome must be factor with two levels, the first level is the 'positive' class

lambda = gamma/(1 - gamma)
}
\references{
Valdes Gilmer, Luna Jose, Eaton Eric, Ungar Lyle, Simone Charles
and Solberg Timothy. MediBoost: a Patient Stratification Tool for
Interpretable Decision Making in the Era of Precision Medicine.
}
\author{
Efstathios D. Gennatas (R), Gilmer Valdes, Jose Marcio Luna Castaneda (original MATLAB)
}
\keyword{internal}
