% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergelongtreatment.R
\name{mergelongtreatment}
\alias{mergelongtreatment}
\title{Merge panel data treatment and outcome data}
\usage{
mergelongtreatment(
  x,
  group_varnames,
  time_varname = "Date",
  start_date,
  end_date,
  interval_days = 14,
  verbose = TRUE,
  trace = 1
)
}
\arguments{
\item{x}{Named list: Long form datasets to merge. Will be converted to \code{data.table}}

\item{group_varnames}{Vector, character: Variable names to merge by, in order. If first is present on
a given pair of datasets, merge on that, otherwise try the next in line.}

\item{time_varname}{Character: Name of column that should be present in all datasets containing
time information. Default = "Date"}

\item{start_date}{Date or characcter: Start date for final dataset in format "YYYY-MM-DD"}

\item{end_date}{Date or character: End dat for final dataset in format "YYYY-MM-DD"}

\item{interval_days}{Integer: Starting with \code{start_date} create timepoints every this many
days. Default = 14}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}

\item{trace}{Integer: If > 0 print additional info to console. Default = 1}
}
\value{
Merged \strong{data.table}
}
\description{
Merge long format treatment and outcome data from multiple sources with possibly hierarchical
matching IDs using \strong{data.table}
}
