% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_mod.R
\name{meta_mod}
\alias{meta_mod}
\title{Meta Models for Regression (Model Stacking)}
\usage{
meta_mod(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  base.mods = c("mars", "ranger"),
  base.params = vector("list", length(base.mods)),
  base.resample.params = setup.resample(resampler = "kfold", n.resamples = 4),
  meta.mod = "gam",
  meta.params = list(),
  x.name = NULL,
  y.name = NULL,
  save.base.res = TRUE,
  save.base.full = FALSE,
  col = NULL,
  se.lty = 3,
  print.base.plot = FALSE,
  print.plot = TRUE,
  plot.fitted = NULL,
  plot.predicted = NULL,
  plot.theme = rtTheme,
  question = NULL,
  verbose.base.res.mods = FALSE,
  verbose.base.mods = FALSE,
  verbose = TRUE,
  trace = 0,
  base.n.cores = 1,
  n.cores = rtCores,
  save.mod = FALSE,
  outdir = NULL,
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix of features, i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{(Optional) Numeric vector or matrix of validation set features
must have set of columns as \code{x}}

\item{y.test}{(Optional) Numeric vector of validation set outcomes}

\item{base.mods}{Character vector: Two or more base learners. Options: \link{select_learn}}

\item{base.params}{List of length equal to N of \code{base.mods}. Each element should be a list of arguments to pass
to the corresponding base mod}

\item{meta.mod}{String. Meta learner. Options: \link{select_learn}}

\item{x.name}{Character: Name for predictor set. (What kind of data is it?)}

\item{y.name}{Character: Name for outcome}

\item{se.lty}{How to plot standard errors. If a number, it corresponds to par("lty") line types and is
plotted with lines(). If "solid", a transparent polygon is plotted using polygon()}

\item{resampler}{String. Resampling method to use. Options: "bootstrap", "kfold", "strat.boot", "strat.sub"}
}
\description{
Train a meta model from the output of base learners trained using different learners (algorithms)
}
\details{
This is included mainly for educational purposes.
\itemize{
\item Train a set of base learners on resamples of the training set x
\item Train a meta learner to map bases' validation set predictions to outcomes
\item Train base learners on full training set x
\item Use the meta learner to predict test set outcome y.test from testing set (x.test)
}
}
\author{
E.D. Gennatas
}
