% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_ops.R
\name{mgetnames}
\alias{mgetnames}
\title{Get names by string matching multiple patterns}
\usage{
mgetnames(
  x,
  pattern = NULL,
  starts_with = NULL,
  ends_with = NULL,
  ignore.case = TRUE,
  return.index = FALSE
)
}
\arguments{
\item{x}{Character vector or object with \code{names()} method}

\item{pattern}{Character vector: pattern(s) to match anywhere in names of x}

\item{starts_with}{Character: pattern to match in the beginning of names of x}

\item{ends_with}{Character: pattern to match at the end of names of x}

\item{ignore.case}{Logical: If TRUE, well, ignore case. Default = TRUE}

\item{return.index}{Logical: If TRUE, return integer index of matches instead of names}
}
\value{
Character vector of matched names or integer index
}
\description{
Get names by string matching multiple patterns
}
\author{
E.D. Gennatas
}
