% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhist.R
\name{mhist}
\alias{mhist}
\title{Histograms}
\usage{
mhist(
  x,
  breaks = "Sturges",
  measure = c("density", "counts"),
  lwd = 3,
  xlim = NULL,
  ylim = NULL,
  plot.axes = FALSE,
  xaxis = TRUE,
  yaxis = TRUE,
  xaxis.line = 0,
  yaxis.line = 0,
  xlab = NULL,
  ylab = measure,
  xaxs = "r",
  yaxs = "r",
  box = FALSE,
  grid = FALSE,
  col = pennCol$lighterBlue,
  horiz = FALSE,
  main = "",
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{Input vector}

\item{breaks}{See \code{hist("breaks")} Default = "Sturges"}

\item{measure}{Character: "density"(Default), "counts"}

\item{lwd}{Float: Line width}

\item{xlim}{Vector, length 2: x-axis limits}

\item{ylim}{Vector, length 2: y-axis limits}

\item{plot.axes}{Logical: If TRUE, draws plot axes. Separate from \code{xaxis} and \code{yaxis}}

\item{xaxis}{Logical: If TRUE, draws x-axis}

\item{yaxis}{Logical: If TRUE, draws y-axis}

\item{xaxis.line}{Float: Number of lines into the margin to position \code{xaxis}. See \code{axis("line")}}

\item{yaxis.line}{Float: Number of lines into the margin to position \code{yaxis}. See \code{axis("line")}}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{xaxs}{Character: 'r' (Default): Extends x-axis range by 4 percent at each end, 'i': Does not extend x-axis range}

\item{yaxs}{Character: 'r' (Default): Extends y-axis range by 4 percent at each end, 'i': Does not extend y-axis range}

\item{box}{Logical: If TRUE, draws a box around plot}

\item{grid}{Logical: If TRUE, draws a grid}

\item{col}{Color to use for histogram lines}

\item{horiz}{Logical: If TRUE, switches x and y axes. Important: Provide all other arguments as if for a
non-rotated plot - i.e. \code{xlab} will become the y-axis label}

\item{main}{Character: Main title}

\item{add}{Logical: If TRUE, add histogram to existing plot (Caution: make sure the axes line up!)}

\item{...}{Additional arguments to be passed to \code{graphics::plot}}
}
\description{
Draws a histogram using lines.
}
\details{
Using \code{horiz = TRUE}, you can draw vertical histograms (as used by \code{mplot3_xym})
}
\author{
E.D. Gennatas
}
