% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{mod_error}
\alias{mod_error}
\title{Error Metrics for Supervised Learning}
\usage{
mod_error(
  true,
  estimated,
  estimated.prob = NULL,
  type = NULL,
  rho = FALSE,
  tau = FALSE,
  na.rm = TRUE,
  verbosity = 0
)
}
\arguments{
\item{true}{Vector: True values}

\item{estimated}{Vector: Estimated values}

\item{estimated.prob}{Vector: Estimated probabilities for Classification,
if available.}

\item{type}{Character: "Regression", "Classification", or "Survival".
If not provided, will be set to Regression if y is numeric.}

\item{rho}{Logical: If TRUE, calculate Spearman's rho.}

\item{tau}{Logical: If TRUE, calculate Kendall's tau.
This can be slow for long vectors}

\item{na.rm}{Logical: Passed to \code{mean} and \code{range} functions.}

\item{verbosity}{Integer: If > 0, print messages to console.}
}
\value{
Object of class \code{mod_error}
}
\description{
Calculate error metrics for pair of vector, e.g. true and estimated values from a model
}
\details{
In regression, NRMSE = RMSE / range(observed)
}
\author{
E.D. Gennatas
}
