% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSelect.R
\name{modSelect}
\alias{modSelect}
\title{Select \pkg{rtemis} Learner}
\usage{
modSelect(mod, fn = FALSE, desc = FALSE)
}
\arguments{
\item{mod}{String: Model name. Case insensitive. e.g. "XGB" for xgboost}

\item{fn}{Logical: If TRUE, return function, otherwise name of function. Defaults to FALSE}

\item{desc}{Logical: If TRUE, return full name / description of algorithm \code{mod}}
}
\value{
function or name of function (see param \code{fn}) or full name of algorithm (\code{desc})
}
\description{
Accepts learner name (supports abbreviations) and returns \pkg{rtemis} function name or
  the function itself.
  If run with no parameters, prints list of available algorithms.
}
\author{
Efstathios D. Gennatas
}
