% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.missing.R
\name{mplot.missing}
\alias{mplot.missing}
\title{Plot missingness}
\usage{
mplot.missing(
  x,
  feat.names = NULL,
  case.names = NULL,
  main = NULL,
  col.present = "#18A3AC",
  col.missing = "#F48024",
  feat.names.y = 0,
  names.srt = 90,
  x.tick.labs.adj = 0,
  case.names.x = 0.25,
  case.names.every = NULL,
  y.srt = 0,
  mar = c(3, 3.5, 5.5, 1),
  oma = c(0.5, 0.5, 0.5, 0.5),
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{Data matrix or data.frame}

\item{feat.names}{Character: Feature names. Defaults to \code{colnames(x)}}

\item{case.names}{Character: Case names. Defaults to \code{rownames(x)}}
}
\description{
Plot missingness
}
\examples{
\dontrun{
dat <- iris
dat[c(1, 5, 17:20, 110, 115, 140), 1] <-
dat[c(12, 15, 55, 73, 100:103), 2] <-
dat[sample(1:150, 25), 4] <- NA
mplot.missing(dat)
}

}
