% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.box.R
\name{mplot3.box}
\alias{mplot3.box}
\title{\code{mplot3}: Boxplot}
\usage{
mplot3.box(
  x,
  col = NULL,
  alpha = 0.66,
  border = NULL,
  border.alpha = 1,
  space = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  boxwex = 0.5,
  horizontal = FALSE,
  main = NULL,
  names.arg = NULL,
  axisnames = FALSE,
  xnames = NULL,
  xnames.at = NULL,
  xnames.y = NULL,
  xnames.font = 1,
  xnames.adj = NULL,
  xnames.pos = NULL,
  xnames.srt = NULL,
  order.by.fn = NULL,
  legend = FALSE,
  legend.names = NULL,
  legend.position = "topright",
  legend.inset = c(0, 0),
  mar = NULL,
  pty = "m",
  cex.axis = cex,
  cex.names = cex,
  yaxis = TRUE,
  ylim.pad = 0,
  theme = getOption("rt.theme", "lightgrid"),
  labelify = TRUE,
  autolabel = letters,
  na.rm = TRUE,
  palette = getOption("rt.palette", "rtCol1"),
  par.reset = TRUE,
  pdf.width = 6,
  pdf.height = 6,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{Vector, data.frame or list: Each data.frame column or list element will be drawn as a box}

\item{col}{Vector of colors to use}

\item{alpha}{Float: Alpha to be applied to \code{col}}

\item{border}{Color for lines around boxes}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{horizontal}{Logical: If TRUE, draw horizontal boxplot(s). Default = FALSR}

\item{main}{Character: Plot title}

\item{order.by.fn}{String: "mean", "median" or any function that outputs a single number: E
stimate function on each vector and order boxes (when input is data.frame or list) by ascending
order. Default = NULL, i.e. no reordering}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills graphics device size. Default = "m"
(See \code{par("pty")})}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{na.rm}{Logical: If TRUE, remove NA values, otherwise function will give error.
Default = TRUE}

\item{palette}{Vector of colors, or Character defining a builtin palette - get options with
\code{rtPalette()}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting. Default = TRUE}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename} is set). Default = 6}

\item{pdf.height}{Float: Height in inches for pdf output. Default = 6}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{...}{Additional arguments to \code{graphics::boxplot}}
}
\description{
Draw boxplots of a vector (single box), data.frame (one box per column) or list
(one box per element - good for variable of different length)
}
\details{
Note that argument `xnames` refers to the x-axis labels below each box. If not specified, these
are inferred from the input when possible. Argument `xlab` is a single label for the x-axis as
per usual and often omitted if `xnames` suffice.
}
\examples{
\dontrun{
## vector
x <- rnorm(500)
mplot3.box(x)

## data.frame
x <- data.frame(alpha = rnorm(50), beta = rnorm(50), gamma = rnorm(50))
mplot3.box(x)

## list - allows different length vectors
x <- list(alpha = rnorm(50),
          beta = rnorm(80, 4, 1.5),
          gamma = rnorm(30, -3, .5))
mplot3.box(x)
}
}
\author{
E.D. Gennatas
}
