% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.cart.R
\name{mplot3.cart}
\alias{mplot3.cart}
\title{\code{mplot3}: \code{data.tree}  trees}
\usage{
mplot3.cart(
  object,
  col.positive = "#F48024DD",
  col.negative = "#18A3ACDD",
  col.lo = ucsfCol$teal,
  col.mid = "gray20",
  col.hi = ucsfCol$orange,
  node.col = "#666666",
  node.shape = "none",
  node.labels = TRUE,
  node.cond = TRUE,
  node.prob = TRUE,
  node.estimate = NULL,
  node.n = TRUE,
  edge.col = "#999999",
  edge.width = 2,
  edge.labels = FALSE,
  arrowhead = "vee",
  layout = "dot",
  drop.leaves = FALSE,
  rankdir = "TB",
  splines = "polyline",
  fontname = "helvetica",
  bg.color = "white",
  overlap = "false",
  prune = FALSE,
  prune.empty.leaves = TRUE,
  remove.bad.parents = TRUE,
  rpart.cp = NULL,
  verbose = TRUE
)
}
\arguments{
\item{node.cond}{Logical: If TRUE, print the splitting condition inside each node. Default = TRUE}

\item{node.prob}{Logical: If TRUE, print the probability estimate for the first class of the outcome
inside each node. Default = TRUE}

\item{node.estimate}{Logical: If TRUE, print the estimated outcome level inside each node.
Default = FALSE}

\item{node.n}{Logical: If TRUE, print the number of cases (from training data) that matched this condition}
}
\description{
Plot \code{data.tree} trees using \code{data.tree::plot.Node}
}
\details{
If you want to show split conditions as edge labels (\code{edge.labels = TRUE}),
it is recommened to set \code{rankdir = "LR"} and \code{node.cond = FALSE}.
Edge labels in graphviz are shown to the right of
the edge when \code{rankdir = "TB"} and above when \code{rankdir = "LR"}.
}
\author{
Efstathios D. Gennatas
}
