% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.conf.R
\name{mplot3.conf}
\alias{mplot3.conf}
\title{Plot confusion matrix}
\usage{
mplot3.conf(object, main = "auto", xlab = "Reference",
  ylab = "Estimated", plot.metrics = TRUE, mod.name = NULL,
  oma = c(0, 0, 0, 0), main.height = 1, dim.lab = 1, dim.in = 4,
  dim.out = -1, font.in = 2, font.out = 1, cex.main = 1.8,
  cex.in = 1.8, cex.lab = 1.8, cex.lab2 = 1.8, cex.lab3 = 1.5,
  cex.out = 1.4, col.main = "auto", col.lab = "auto",
  col.text.out = "auto", col.bg = "auto", col.bg.out1 = "auto",
  col.bg.out2 = "auto", col.text.hi = "auto", col.text.lo = "auto",
  theme = getOption("rt.theme", "light"), mid.col = "auto",
  hi.color.pos = "#18A3AC", hi.color.neg = "#716FB2",
  par.reset = TRUE, pdf.width = 4.5, pdf.height = 4.5,
  filename = NULL, ...)
}
\arguments{
\item{object}{Either 1. a classification \code{rtMod}, b. a \code{caret::confusionMatrix} object, or c. a matrix / 
data.frame / table}

\item{main}{String: Plot title}

\item{xlab}{String: x-axis label}

\item{ylab}{String: y-axis label}

\item{mod.name}{String: Name of the algorithm used to make predictions. If NULL, will look for 
\code{object$mod.name}. Default = NULL}

\item{dim.lab}{Float: Height for labels}

\item{dim.in}{Float: Width and height for confusion matrix cells}

\item{dim.out}{Float: Height for metrics cells}

\item{font.in}{Integer: The \code{font} parameter for confusion matrix cells}

\item{font.out}{Integer: The \code{font} parameter for metrics cells}

\item{cex.in}{Float: The \code{cex} parameter for confusion matrix cells}

\item{cex.lab}{Float: The \code{cex} parameter for first line of label cells}

\item{cex.lab2}{Float: The \code{cex} parameter for second line of label cells}

\item{cex.out}{Float: The \code{cex} parameter for metrics cells}

\item{col.text.out}{Color for metrics cells' text}

\item{theme}{String: "light", or "dark"}

\item{hi.color.pos}{Color: The hi color for correct classification.}

\item{hi.color.neg}{Color: The hi color for missclassification}

\item{par.reset}{Logical: If TRUE, reset par before exit. Default = TRUE}

\item{pdf.width}{Float: PDF width, if \code{filename} is set}

\item{pdf.height}{Float: PDF height, if \code{filename} is set}

\item{filename}{String: If specified, save plot to this path. Default = NULL}

\item{...}{Not used}

\item{mar}{Numeric, vector, length 4: Overall margins}

\item{col.bg.out}{Color for metrics cells' background}

\item{mid.color}{Color: The mid color for the confusion matrix. Default = "white" for theme = "light", 
"black" for "dark"}
}
\value{
List of metrics, invisibly
}
\description{
Plots confusion matrix and classification metrics
}
\author{
Efstathios D. Gennatas
}
