% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.graph.R
\name{mplot3.graph}
\alias{mplot3.graph}
\title{Plot \code{igraph} networks}
\usage{
mplot3.graph(
  net,
  vertex.size = 18,
  vertex.col = NULL,
  vertex.label.col = "#ffffff",
  vertex.frame.col = NA,
  edge.col = NULL,
  edge.alpha = 0.2,
  edge.curved = 0.5,
  layout = c("fr", "dh", "drl", "gem", "graphopt", "kk", "lgl", "mds", "sugiyama"),
  layout_params = list(),
  cluster = NULL,
  cluster_params = list(),
  cluster_mark_groups = TRUE,
  cluster_color_nodes = FALSE,
  theme = getOption("rt.theme", "lightgrid"),
  par.reset = TRUE,
  filename = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{net}{\code{igraph} network}
}
\description{
Plot \code{igraph} networks
}
\author{
E.D. Gennatas
}
