% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.heatmap.R
\name{mplot3.heatmap}
\alias{mplot3.heatmap}
\title{\code{mplot3} Heatmap (\code{image}; modified \code{heatmap})}
\usage{
mplot3.heatmap(x, colorGrad.n = 101, colorGrad.col = NULL,
  lo = "#18A3AC", lomid = NULL, mid = "white", midhi = NULL,
  hi = "#F48024", space = "rgb", colorbar = TRUE, cb.n = 21,
  cb.title = NULL, cb.cex = 1, cb.title.cex = 1, cb.add.mar = NULL,
  Rowv = NA, Colv = if (symm) Rowv else NA, distfun = dist,
  hclustfun = hclust, reorderfun = function(d, w) reorder(d, w),
  add.expr, symm = FALSE, revC = identical(Colv, "Rowv"),
  scale = "none", na.rm = TRUE, margins = c(5, 5), ColSideColors,
  RowSideColors, cexRow = 0.2 + 1/log10(nr), cexCol = 0.2 +
  1/log10(nc), labRow = NULL, labCol = NULL, labCol.las = NULL,
  main = "", main.adj = 0, main.line = NA, xlab = NULL,
  ylab = NULL, xlab.line = NULL, ylab.line = NULL, bg = "white",
  col.axis = "black", keep.dendro = FALSE,
  verbose = getOption("verbose"), zlim = NULL, autorange = TRUE,
  filename = NULL, par.reset = TRUE, pdf.width = 7, pdf.height = 7,
  ...)
}
\arguments{
\item{colorGrad.n}{Integer. Number of distinct colors to generate using \link{colorGrad}}

\item{colorGrad.col}{Color. See \link{colorGrad}}

\item{colorbar}{Logical. Should colorbar appear next to heatmap?}

\item{cb.n}{Integer. Number of steps in colorbar. 21 gives 10 above and 10 below midline.
If midline is zero, this corresponds to 10 percent increments / decrements}
}
\description{
Customized heatmap with optional colorbar
}
\details{
The main difference from the original \code{heatmap} is the addition of a colorbar on the side.
This is achieved with \link{colorGrad}.
Other differences:
- Dendrograms are not drawn by default. Set \code{Rowv = T} and \code{Colv = T} to get them.
- Column labels are only drawn perpendicular to the x-axis if any one is
  longer than two characters.
}
\author{
Efstathios D Gennatas modified from original \code{stats::heatmap}
by Andy Liaw, R. Gentleman, M. Maechler, W. Huber
}
