% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.lolli.R
\name{mplot3.lolli}
\alias{mplot3.lolli}
\title{\code{mplot3} Lollipop Plot}
\usage{
mplot3.lolli(
  x,
  order.on.x = TRUE,
  plot.top = 1,
  orientation = c("horizontal", "vertical"),
  xnames = NULL,
  points = TRUE,
  segments = TRUE,
  main = NULL,
  col = NULL,
  cex = 1.2,
  matching.segment.col = FALSE,
  segment.alpha = 0.333,
  lty = 3,
  lwd = 2,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  autolabel = letters,
  par.reset = TRUE,
  pdf.width = 6,
  pdf.height = 6,
  mar = c(2.5, 3, 2, 1),
  pty = "m",
  pch = 16,
  x.axis.at = NULL,
  y.axis.at = NULL,
  xlab = NULL,
  ylab = NULL,
  label.las = 1,
  label.padj = 0.5,
  xaxs = "r",
  yaxs = "r",
  xlab.adj = 0.5,
  ylab.adj = 0.5,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{Float, vector: Input data}

\item{order.on.x}{Logical: If TRUE, order by value of \code{x}. Default = TRUE}

\item{plot.top}{Float or Integer: If <= 1, plot this percent highest absolute values, otherwise plot this many top values.
i.e.: \code{plot.top = .2} will print the top 20% highest values, and \code{plot.top = 20} will plot the top 20
highest values}

\item{xnames}{Character, vector: Names of \code{x}}

\item{main}{Character: Main title}

\item{col}{Color, vector: Lollipop color}

\item{cex}{Float: Character expansion factor for points. Default = 1.2}

\item{matching.segment.col}{Logical: If TRUE, color line segments using \code{col}, i.e. same as}

\item{segment.alpha}{Float: Transparency for line segments. Default = .5
points. Default = FALSE, in which case they are colored with \code{theme$fg}}

\item{lty}{Integer: Line type for segment segments. See \code{par("lty")} Default = 1}

\item{lwd}{Float: Width for segment segments. See \code{par("lty")} Default = 1}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'darkgrid')}

\item{palette}{Vector of colors, or Character defining a builtin palette - get options with
\code{rtPalette()}}

\item{autolabel}{Vector to be used to generate autolabels when using \link{rtlayout} with
\code{autolabel = TRUE}. Default = \code{letters}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting. Default = TRUE}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename} is set). Default = 6}

\item{pdf.height}{Float: Height in inches for pdf output. Default = 6}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills graphics device size. Default = "m"
(See \code{par("pty")})}

\item{pch}{Integer: Point character. Default = 16}

\item{x.axis.at}{Float, vector: x coordinates to place tick marks. Default = NULL, determined by
\code{graphics::axis} aautomatically}

\item{y.axis.at}{As \code{x.axis.at} for y-axis}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4% on either end; "i": Use exact x-axis limits.
Default = "r"}

\item{yaxs}{Character: as \code{xaxs} for the y-axis. Default = "r"}

\item{xlab.adj}{Float: \code{adj} for \code{xlab} (See \code{par("adj")})}

\item{ylab.adj}{Float: \code{adj} for \code{ylab} (See \code{par("adj")})}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{...}{Additional arguments to be passed to theme function}
}
\description{
\code{mplot3} Lollipop Plot
}
\examples{
\dontrun{
x <- rnorm(12)
mplot3.lolli(x)
# a "rounded" barplot
mplot3.lolli(x, segments = T, points = F,
             lty = 1, matching.segment.col = T,
             lwd = 10, segment.alpha = 1)
}
}
\author{
E.D. Gennatas
}
