% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.marginal.R
\name{mplot3.marginal}
\alias{mplot3.marginal}
\title{Marginal Effects}
\usage{
mplot3.marginal(mod, x.name = "x", x = -10:10, y.name = "y",
  fix.at = 1, print.plot = TRUE, plot.type = "l", ...)
}
\arguments{
\item{mod}{\pkg{rtemis} model to use for prediction}

\item{x.name}{String: Predictor name as used in \code{mod} whose marginal effects you wish to estimate
Predictor names are stored in \code{mod$xnames}}

\item{x}{Numeric vector: Values at which to evaluate marginal effects. Define them based on the range of
the values used to train the model, or estimates will be unreliable, potentially entirely off.}

\item{fix.at}{Numeric: Assign this constant to all features other than \code{x.name}. Defaults to 1}

\item{print.plot}{Logical: if TRUE, draw plot using \link{mplot3}}

\item{plot.type}{"p" for points, "l" for line}

\item{...}{Additional arguments to be passed to \link{mplot3.xy}}
}
\value{
Returns predicted values invisibly
}
\description{
Plots and returns invisibly the estimated outcome by varying one predictor and keeping others fixed.
  Works on any \pkg{rtemis} model trained on multiple features
}
\author{
Efstathios D. Gennatas
}
