% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.R
\name{mplot3}
\alias{mplot3}
\title{\pkg{rtemis} static plotting}
\usage{
mplot3(x, y = NULL, type = NULL, fit = NULL, ...)
}
\arguments{
\item{x}{Numeric vector or list of vectors}

\item{y}{Numeric vector or list of vectors}

\item{type}{String. "scatter", "density", "index", "heatmap"}

\item{fit}{String. Any `rtemis` model, best to use "lm" for linear fits and "gam" for non-linear fits}

\item{...}{Additional arguments to be passed to}
}
\description{
Draw quality plots using fast, base graphics
}
\author{
Efstathios D. Gennatas
}
