% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.res.R
\name{mplot3.res}
\alias{mplot3.res}
\title{\code{mplot3} Plot \code{resample}}
\usage{
mplot3.res(res, col = NULL, theme = "white", ...)
}
\description{
Visualizes resampling output using \link{mplot3.img}
}
\details{
For resampling with no replacement where each case may be selected 0 or 1 time,
0 is white and 1 is teal For resampling with replacement, 0 is white, 1 is blue, 2 is teal
}
\author{
Efstathios D. Gennatas
}
