% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.roc.R
\name{mplot3.roc}
\alias{mplot3.roc}
\title{\code{mplot3} ROC curves}
\usage{
mplot3.roc(prob, labels, method = c("pROC", "rt"), type = "TPR.FPR",
  balanced.accuracy = FALSE, main = "", col = ucsfPalette,
  cex = 1.2, lwd = 2.5, diagonal = TRUE, diagonal.lwd = 2.5,
  diagonal.lty = 1, group.legend = FALSE, annotation = TRUE,
  annotation.col = col, annot.line = NULL, annot.adj = 1,
  annot.font = 1, mar = c(2.5, 3, 2.5, 1),
  theme = getOption("rt.theme", "lightgrid"), par.reset = TRUE,
  filename = NULL, pdf.width = 5, pdf.height = 5, ...)
}
\arguments{
\item{prob}{Vector, Float [0, 1]: Predicted probabilities (i.e. c(.1, .8, .2, .9))}

\item{labels}{Vector, Integer {0, 1}: True labels (i.e. c(0, 1, 0, 1))}

\item{...}{Additional parameters to pass to \link{mplot3.xy}}
}
\description{
Plot ROC curve for a binary classifier
}
\author{
Efstathios D. Gennatas
}
