% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.surv.R
\name{mplot3.surv}
\alias{mplot3.surv}
\title{\code{mplot3}: Survival Plots}
\usage{
mplot3.surv(
  x,
  lty = 1,
  lwd = 2,
  alpha = 1,
  col = NULL,
  normalize.time = TRUE,
  cex = 1.2,
  xlab = "Time",
  ylab = "Survival",
  main = "Kaplan-Meier estimate",
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  plot.error = FALSE,
  error.lty = 2,
  error.alpha = 0.5,
  group.legend = NULL,
  group.title = "",
  group.names = NULL,
  group.side = 3,
  group.adj = 0.98,
  group.padj = 2,
  group.at = NA,
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{Survival object / list of Survival objects created using \code{survival::Surv}}

\item{lty}{Integer: Line type. Default = 1. See \code{par("lty")}}

\item{lwd}{Float: Line width. Default = 2}

\item{alpha}{Float: Alpha for lines. Default = 1}

\item{normalize.time}{Logical: If TRUE, convert each input's time to 0-1 range. This is useful
when survival estimates are not provided in original time scale. Default = FALSE.}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: Plot title}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{palette}{Vector of colors, or Character defining a builtin palette - get all options with \code{rtPalette()}}

\item{group.legend}{Logical: If TRUE, place \code{group.names} in a legend}

\item{group.title}{Character: Group title, shown above group names. e.g. if group names are
c("San Francisco", "Philadelphia"), \code{group.title} can be "Place of residence"}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting. Default = TRUE}

\item{...}{Additional arguments to pass to \link{mplot3.xy}}
}
\description{
Plots survival step functions using \link{mplot3.xy}
}
\author{
E.D. Gennatas
}
