% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.varimp.R
\name{mplot3.varimp}
\alias{mplot3.varimp}
\title{\code{mplot3}: Variable Importance}
\usage{
mplot3.varimp(x, error = NULL, names = NULL, names.pad = 0.02,
  plot.top = 1, labelify = TRUE, col = colorGrad(length(x), lo =
  pennCol$lightestBlue, mid = pennCol$blue, hi = pennCol$lighterRed),
  alpha = 0.8, error.col = "white", error.lwd = 2, beside = TRUE,
  border = NA, width = 1, space = 0.2, xlim = NULL, ylim = NULL,
  xlab = "Variable Importance", xlab.line = 1.3, ylab = NULL,
  ylab.line = 1.5, main = NULL, main.line = 0.5, main.adj = 0,
  main.col = NULL, main.font = 2, main.family = "",
  names.arg = NULL, axisnames = FALSE, sidelabels = NULL,
  mar = NULL, pty = "m", cex = 1.2, cex.axis = cex,
  cex.names = cex, bg = NULL, plot.bg = NULL, barplot.axes = FALSE,
  xaxis = TRUE, x.axis.padj = -1.2, tck = -0.015, tick.col = NULL,
  theme = getOption("rt.theme", "light"), axes.col = NULL,
  labs.col = NULL, grid = FALSE, grid.lty = NULL, grid.lwd = NULL,
  grid.col = NULL, grid.alpha = 1, par.reset = TRUE,
  pdf.width = NULL, pdf.height = NULL, trace = 0, filename = NULL,
  ...)
}
\arguments{
\item{x}{Vector, numeric: Input}

\item{error}{Vector, numeric; length = length(x): Plot error bars with given error.}

\item{names}{Vector, string; optional: Names of variables in \code{x}}

\item{plot.top}{Float or Integer: If <= 1, plot this percent highest values, otherwise plot this many top values.
i.e.: \code{plot.top = .2} will print the top 20% highest values, and \code{plot.top = 20} will plot the top 20
highest values}

\item{labelify}{Logical: If TRUE convert \code{names(x)} using \link{labelify}. Default = TRUE}

\item{col}{Colors: Gradient to use for barplot fill. Default = colorGrad(length(x),
lo = ucsfCol$blue, mid = ucsfCol$red, hi = ucsfCol$yellow)}

\item{alpha}{Float (0, 1): Alpha for \code{col}}

\item{error.col}{Color: For error bars}

\item{trace}{Integer: If \code{trace > 0} prints out the automatically set \code{mar} (so you can adjust if needed)
\code{names} provided}
}
\value{
Position of bar centers (invisibly)
}
\description{
Draw horizontal barplots for variable importance
}
\author{
Efstathios D. Gennatas
}
