% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.x.R
\name{mplot3.x}
\alias{mplot3.x}
\title{\code{mplot3}: Univariate plots: index, histogram, density, QQ-line}
\usage{
mplot3.x(x, type = c("density", "histogram", "index", "line", "qqline"),
  group = NULL, data = NULL, xlab = NULL, ylab = NULL,
  main = NULL, main.col = NULL, main.adj = 0, main.line = 0.5,
  main.font = 2, main.family = "", xlim = NULL, ylim = NULL,
  index.ypad = 0.1, axes = TRUE, axes.swap = FALSE,
  axes.col = NULL, tick.col = NULL, cex = 1.2, col = NULL,
  alpha = 0.75, index.type = c("p", "l"), hist.breaks = "Sturges",
  hist.type = c("bars", "lines"), hist.lwd = 3, density.line = FALSE,
  density.shade = TRUE, density.legend.side = 3,
  density.legend.adj = 0.98, density.params = list(),
  qqline.col = NULL, qqline.alpha = 0.66, pch = 16,
  point.col = NULL, point.cex = 1, point.bg.col = NULL,
  point.alpha = 0.66, hline = NULL, vline = NULL, diagonal = FALSE,
  diagonal.col = NULL, grid = FALSE, grid.col = NULL,
  grid.alpha = 0.5, grid.lty = 3, grid.lwd = 1, bg = NULL,
  plot.bg = NULL, annotation = NULL, annot.col = NULL,
  group.legend = NULL, group.title = "", group.names = NULL,
  group.side = 3, group.adj = 0.02, group.at = NA, text.xy = NULL,
  text.x = NULL, text.y = NULL, text.xy.cex = 1,
  text.xy.col = "white", line.col = "#008E00", tck = 0.02,
  x.axis.padj = -1.1, xlab.line = 1.3, y.axis.padj = 0.9,
  ylab.line = 1.6, labs.col = NULL, bty = "o", box = NULL,
  box.col = NULL, box.alpha = 1, box.lty = 1, box.lwd = 2,
  lab.adj = 0.5, density.mean = ifelse(type == "density", TRUE, FALSE),
  hline.col = "black", hline.lwd = 1, hline.lty = 1,
  vline.col = "black", vline.lwd = 1, vline.lty = 1, lty = 1,
  lwd = 2, qqline.lwd = lwd, density.lwd = lwd,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"), zero.lines = NULL,
  zero.col = NULL, zero.alpha = 0.66, zero.lty = 1, zero.lwd = 1.5,
  pty = "s", mar = c(2.5, 2.5, 1.5, 1), xaxs = "r", yaxs = "r",
  autolabel = letters, new = FALSE, alpha.off = FALSE,
  na.rm = TRUE, par.reset = TRUE, export.par.orig = FALSE,
  filename = NULL, pdf.width = 6, pdf.height = 6, ...)
}
\arguments{
\item{x}{Numeric vector or list of vectors, one for each group.
If \code{data} is provided, x is name of variable in \code{data}}

\item{type}{Character: "density", "histogram",  "index", "line", "qqline"
Case-insensitive and supports partial matching: e.g. \code{mplot3.x(x, "H")} gives histogram}

\item{group}{Vector denoting group membership. Will be converted to factor.
If \code{data} is provided, \code{group} is name of variable if \code{data}}

\item{data}{Optional data frame containing x data}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: Plot title}

\item{main.col}{Color for title. Defaults to black for theme "light" and "box", otherwise white}

\item{main.adj}{Float: Plot title alignment (0 = left, .5 = center, 1 = right)}

\item{main.line}{Float: \code{mtext line} argument for the title. Higher numbers move title upwards.
Default = .5}

\item{main.font}{Integer: 1: regular, 2: bold, 3: italic. See \code{par("font")} for more}

\item{main.family}{Character: Font family to use. See \code{par("family")}}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{index.ypad}{Float: Expand ylim by this much for plot type "index" Default = .1
(Stops points being cut off)}

\item{axes}{Logical: Should the axes be drawn? Defaults to TRUE}

\item{axes.col}{Character: Color for axes values (box color set with \code{box.col})}

\item{tick.col}{Color: Tick color. Default = NULL, which results in theme-appropriate color}

\item{cex}{Float: Character expansion factor (See \code{?par})}

\item{index.type}{Character: "p" for points (Default), "l" for lines (timeseries)}

\item{hist.breaks}{See \code{histogram("breaks")}}

\item{hist.type}{Character: "bars" or "lines". Default = "bars"}

\item{hist.lwd}{Float: Line width for \code{type = "histogram"}; \code{hist.type = "lines"}}

\item{density.line}{Logical: If TRUE, draw line for \code{type = "density"}. Default = FALSE}

\item{density.shade}{Logical: If TRUE, draw shaded polygon for \code{type = "density"}. Default = TRUE}

\item{qqline.col}{Color for Q-Q line}

\item{qqline.alpha}{Float: Alpha for Q-Q line}

\item{grid}{Logical: Should grid be drawn?}

\item{grid.col}{Grid color}

\item{grid.alpha}{Alpha for \code{grid.col}}

\item{grid.lty}{Grid line type (See \code{par("lty")})}

\item{grid.lwd}{Grid line width}

\item{bg}{Background color. Defaults to white for themes "light" and "box", black otherwise.}

\item{plot.bg}{Color: Background color. Default = NULL, which results in theme-appropriate color}

\item{annotation}{Character: Add annotation at the bottom right of the plot}

\item{group.legend}{Logical: If TRUE, include legend with group names}

\item{group.title}{Character: Title above group names}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{tck}{Float: Tick length. Can be negative (See \code{par("tck")})}

\item{x.axis.padj}{Adjustment for the x axis tick labels position}

\item{xlab.line}{Adjustment for the x axis label position (See code{line} in \code{?mtext})}

\item{y.axis.padj}{Similar to \code{x.axis.padj} for the y axis}

\item{ylab.line}{Similar to \code{xlab.line} for the y axis}

\item{labs.col}{Color for labels}

\item{bty}{Character: "o", "l", "7", "c", "u", or "]" result in a box resembling each character.
(See \code{par("bty")})}

\item{box}{Logical: Should a box be drawn around the plot? Defaults to TRUE for themes "box" and "darkbox"}

\item{box.col}{Box color}

\item{box.alpha}{Alpha for \code{box.col}}

\item{box.lty}{Box line type}

\item{box.lwd}{Float: Box line width. Default = 1.5}

\item{lab.adj}{Adjust the axes labels. 0 = left adjust; 1 = right adjust; .5 = center (Default)}

\item{density.mean}{Logical: If TRUE, print mean of \code{x} along plot. Default = TRUE, for \code{type = "density"}}

\item{lwd}{Integer: Line width. Used for \code{type = "line" or "density"}}

\item{theme}{Character: "light", "dark", "lightgrid", "darkgrid", "lightbox", "darkbox"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{zero.lines}{Logical: If TRUE, draw lines at x = 0 and y = 0}

\item{zero.col}{Color for \code{zero.lines}}

\item{zero.alpha}{Color alpha for \code{zero.lines}}

\item{zero.lty}{Zero line line type}

\item{zero.lwd}{Zero line width}

\item{pty}{"s" gives a square plot, "m" gives a plot that fills graphics device size. Defaults to "s"
(See \code{par("pty")})}

\item{mar}{Vector, length 4: Margins; see \code{par("mar")}}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4% on either end; "i": Use exact x-axis limits.
Default = "r"}

\item{yaxs}{Character: as \code{xaxs} for the y-axis. Default = "r"}

\item{new}{Logical: If TRUE, add plot to existing plot. See \code{par("new")}}

\item{na.rm}{Logical: Will be passed to all functions that support it. If set to FALSE,
input containing NA values will result in error, depending on the \code{type}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting. Default = TRUE}

\item{filename}{Path to file: If supplied, plot will be printed to file}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename} is set). Default = 6}

\item{pdf.height}{Float: Height in inches for pdf output. Default = 6}

\item{...}{Additional arguments to be passed to learner function}
}
\description{
Draw plots of 1-dimensional data: index, histogram, density, and Q-Q plots.
}
\details{
You can group data either by supplying x as a list where each element contains one vector per group, or
as a data frame where each column represents group,
or by providing a \code{group} variable, which will be converted to factor.
For bivariate plots, see \link{mplot3.xy} and \link{mplot3.xym}. For heatmaps, see \link{mplot3.heatmap}
To plot histograms of multiple groups, it's best to use \code{hist.type = "lines"}, which will use \link{mhist}
and space apart the breaks for each group
}
\seealso{
\link{mplot3}, \link{mplot3.xy}, \link{mplot3.xym}, \link{mplot3.heatmap}
}
\author{
Efstathios D. Gennatas
}
