% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.xym.R
\name{mplot3.xym}
\alias{mplot3.xym}
\title{\code{mplot3} Scatter plot with marginal density and/or histogram}
\usage{
mplot3.xym(x, y, margin = c("histogram", "density", "both"),
  fit = "gam", se.fit = TRUE, xlim = NULL, ylim = NULL,
  col = "#18A3AC", density.alpha = 0.66, hist.breaks = 30,
  hist.alpha = 0.66, hist.space = 0.05, hist.lwd = 3, lwd = 4,
  main = NULL, main.adj = 0, axes.density = FALSE, pty = "m",
  mar = c(3, 3, 0, 0), mar.mar = 0.2, axes = TRUE, xaxs = "r",
  yaxs = "r", theme = getOption("rt.theme", "lightgrid"),
  par.reset = TRUE, widths = c(5, 1), heights = c(1, 5),
  filename = NULL, pdf.width = 7, pdf.height = 7, ...)
}
\arguments{
\item{x}{x-data}

\item{y}{y-data}

\item{margin}{What type of marginal plots to draw. Options: "density", "histogram", or "both"}

\item{fit}{What model to use to draw \code{y ~ x}. Options: \code{mode.select()}}

\item{se.fit}{Logical: If TRUE: plot +/- 2 * Standard Error of fit}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{col}{Color for marginal plots}

\item{density.alpha}{Alpha for density plots}

\item{hist.breaks}{Integer. Number of histogram breaks}

\item{hist.alpha}{Alpha for barplots}

\item{hist.space}{Space between bars in barplots}

\item{lwd}{Float: Line width}

\item{main}{Character: Main title}

\item{main.adj}{Float: Plot title alignment (0 = left, .5 = center, 1 = right)}

\item{axes.density}{Logical: If TRUE, plot margin plot axes for density (debugging only)}

\item{pty}{"s" gives a square plot, "m" gives a plot that fills graphics device size. Defaults to "s"
(See \code{?par})}

\item{mar}{Vector, length 4: Margins; see \code{mar} in \code{?par}}

\item{axes}{Logical: Should the axes be drawn? Defaults to TRUE}

\item{theme}{Character: "light", "dark", "lightgrid", "darkgrid", "lightbox", "darkbox"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{par.reset}{Logical. Resest \code{par} to original settings}

\item{...}{Additional arguments to be passed to \code{mplot3.xy}}
}
\description{
Draw a scatter plot with fit line and marginal density and/or histogram
}
\details{
At the moment, \code{mplot3.xym} is the only \code{mplot3} function that does not support
}
\author{
Efstathios D Gennatas
}
