% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_adsr.R
\name{mplot3_adsr}
\alias{mplot3_adsr}
\title{\code{mplot3}: ADSR Plot}
\usage{
mplot3_adsr(
  Attack = 300,
  Decay = 160,
  Sustain = 40,
  Release = 500,
  Value = 80,
  I = 200,
  O = 1800,
  lty = 1,
  lwd = 4,
  main = "ADSR Envelope",
  main.line = 1.6,
  main.col = "white",
  Attack.col = "#44A6AC",
  Decay.col = "#F4A362",
  Sustain.col = "#3574A7",
  Release.col = "#C23A70",
  Decay.nl = FALSE,
  draw.poly = FALSE,
  poly.alpha = 0.15,
  draw.verticals = TRUE,
  v.lty = 1,
  v.lwd = 0.8,
  arrow.code = 2,
  arrow.length = 0.09,
  grid = FALSE,
  grid.lty = 1,
  grid.lwd = 0.4,
  grid.col = NULL,
  zerolines.col = "gray50",
  theme = "darkgray",
  labs.col = "gray70",
  tick.col = "gray70",
  on.lty = 2,
  on.lwd = 7,
  on.alpha = 0.4,
  on.col = "gray70",
  off.col = "gray70",
  pty = "m",
  mar = c(3, 3, 3.2, 0.5),
  xaxs = "i",
  yaxs = "i",
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{Attack}{Float: Attack time (in milliseconds)}

\item{Decay}{Float: Decay time (in milliseconds)}

\item{Sustain}{Float: Sustain Level (percent)}

\item{Release}{Float: Release time (in milliseconds)}

\item{Value}{Float: Value (percent)}

\item{I}{Float: Note on time (in milliseconds)}

\item{O}{Float: Note off time (in milliseconds)}

\item{theme}{Character: "light" or "dark" (Default)}

\item{...}{Additional arguments to pass to \link{mplot3_xy}}
}
\description{
Plot Attack Decay Sustain Release Envelope Generator using \code{mplot3}
}
\details{
Learn more:
(https://en.wikipedia.org/wiki/Synthesizer#Attack_Decay_Sustain_Release_.28ADSR.29_envelope "ADSR Wikipedia")
}
\examples{
\dontrun{
mplot3_adsr()
}
}
\author{
E.D. Gennatas
}
