% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_conf.R
\name{mplot3_conf}
\alias{mplot3_conf}
\title{Plot confusion matrix}
\usage{
mplot3_conf(
  object,
  main = "auto",
  xlab = "Reference",
  ylab = "Predicted",
  plot.metrics = TRUE,
  mod.name = NULL,
  oma = c(0, 0, 0, 0),
  dim.main = NULL,
  dim.lab = 1,
  dim.in = 4,
  dim.out = -1,
  font.in = 2,
  font.out = 1,
  cex.main = 1.2,
  cex.in = 1.2,
  cex.lab = 1.2,
  cex.lab2 = 1.2,
  cex.lab3 = 1,
  cex.out = 1,
  col.main = "auto",
  col.lab = "auto",
  col.text.out = "auto",
  col.bg = "auto",
  col.bg.out1 = "auto",
  col.bg.out2 = "auto",
  col.text.hi = "auto",
  col.text.lo = "auto",
  show.ba = TRUE,
  theme = getOption("rt.theme", "white"),
  mid.col = "auto",
  hi.color.pos = "#18A3AC",
  hi.color.neg = "#C23A70",
  autolabel = letters,
  par.reset = TRUE,
  pdf.width = 7,
  pdf.height = 7,
  filename = NULL,
  ...
)
}
\arguments{
\item{object}{Either a classification \code{rtMod}, or a table/matrix/data.frame confusion matrix
where rows are the reference classes and columns are the predicted classes.}

\item{main}{Character: Plot title.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{plot.metrics}{Logical: If TRUE, draw classification metrics next to confusion
matrix.}

\item{mod.name}{Character: Name of the algorithm used to make predictions. If NULL,
will look for \code{object$mod.name}.}

\item{oma}{Numeric, vector, length 4: Outer margins.}

\item{dim.main}{Numeric: Height for title.}

\item{dim.lab}{Numeric: Height for labels.}

\item{dim.in}{Numeric: Height/Width for confusion matrix cells.}

\item{dim.out}{Numeric: Height for metrics cells. Default = -1, which autoadjusts
depending on number of output classes.}

\item{font.in}{Integer: The \code{font} parameter for confusion matrix cells.}

\item{font.out}{Integer: The \code{font} parameter for metrics cells.}

\item{cex.main}{Numeric: The \code{cex} parameter for the main title.}

\item{cex.in}{Numeric: The \code{cex} parameter for confusion matrix cells.}

\item{cex.lab}{Numeric: The \code{cex} parameter for first line of label cells.}

\item{cex.lab2}{Numeric: The \code{cex} parameter for second line of label cells.}

\item{cex.lab3}{Numeric: The \code{cex} parameter for classification metrics.}

\item{cex.out}{Numeric: The \code{cex} parameter for metrics cells.}

\item{col.main}{Color for title. Default = "auto", determined by \code{theme}.}

\item{col.lab}{Color for labels. Default = "auto", determined by \code{theme}.}

\item{col.text.out}{Color for metrics cells' text. Default = "auto",
determined by \code{theme}.}

\item{col.bg}{Color for background. Default = "auto", determined by
\code{theme}.}

\item{col.bg.out1}{Color for metrics cells' background (row1).
Default = "auto", determined by \code{theme}.}

\item{col.bg.out2}{Color for metrics cells' background (row2).
Default = "auto", determined by \code{theme}.}

\item{col.text.hi}{Color for high confusion matrix values. Default = "auto",
determined by \code{theme}.}

\item{col.text.lo}{Color for low confusion matrix values. Default = "auto",
determined by \code{theme}.}

\item{show.ba}{Logical: If TRUE, show Balanced Accuracy at bottom right corner.}

\item{theme}{Character: "light", or "dark". Set to
\code{options("rt.theme")}, if set, otherwise "light"}

\item{mid.col}{Color: The mid color for the confusion matrix.
Default = "auto", determined by \code{theme}.}

\item{hi.color.pos}{Color: The hi color for correct classification.}

\item{hi.color.neg}{Color: The hi color for missclassification.}

\item{autolabel}{Character vector to be used to generate autolabels when using
\link{rtlayout} with \code{autolabel = TRUE}.}

\item{par.reset}{Logical: If TRUE, reset par before exit.}

\item{pdf.width}{Numeric: PDF width, if \code{filename} is set.}

\item{pdf.height}{Numeric: PDF height, if \code{filename} is set.}

\item{filename}{Character: If specified, save plot to this path.}

\item{...}{Additional arguments passed to \code{theme}.}
}
\value{
List of metrics, invisibly
}
\description{
Plots confusion matrix and classification metrics
}
\details{
This function uses its multiple cex args instead of the theme's \code{cex} parameter
}
\examples{
\dontrun{
true <- c("alpha", "alpha", "alpha", "alpha", "beta", "beta", "beta", "beta")
predicted <- c("alpha", "alpha", "alpha", "beta", "beta", "alpha", "alpha", "beta")
mplot3_conf(table(predicted, true))
}
}
\author{
E.D. Gennatas
}
