% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_confbin.R
\name{mplot3_confbin}
\alias{mplot3_confbin}
\title{Plot extended confusion matrix for binary classification}
\usage{
mplot3_confbin(
  object,
  main = NULL,
  xlab = "True",
  ylab = "Estimated",
  mod.name = NULL,
  mar = c(4, 5, 4, 3),
  dim.lab = 1,
  dim.in = 4,
  dim.out = 2,
  font.in = 2,
  font.out = 2,
  cex.in = 1.2,
  cex.lab = 1.2,
  cex.lab2 = 1,
  cex.out = 1,
  col.text.out = "white",
  col.bg.out = "gray50",
  theme = "light",
  mid.color = NULL,
  hi.color.pos = "#18A3AC",
  hi.color.neg = "#716FB2",
  par.reset = TRUE,
  pdf.width = 8.7,
  pdf.height = 8.7,
  filename = NULL,
  ...
)
}
\arguments{
\item{object}{Either 1. a classification \code{rtMod}, b. a \code{caret::confusionMatrix} object, or c. a matrix /
data.frame / table}

\item{main}{Character: Plot title}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{mod.name}{Character: Name of the algorithm used to make predictions. If NULL, will look for
\code{object$mod.name}. Default = NULL}

\item{mar}{Numeric, vector, length 4: Overall margins}

\item{dim.lab}{Float: Height for labels}

\item{dim.in}{Float: Width and height for confusion matrix cells}

\item{dim.out}{Float: Height for metrics cells}

\item{font.in}{Integer: The \code{font} parameter for confusion matrix cells}

\item{font.out}{Integer: The \code{font} parameter for metrics cells}

\item{cex.in}{Float: The \code{cex} parameter for confusion matrix cells}

\item{cex.lab}{Float: The \code{cex} parameter for first line of label cells}

\item{cex.lab2}{Float: The \code{cex} parameter for second line of label cells}

\item{cex.out}{Float: The \code{cex} parameter for metrics cells}

\item{col.text.out}{Color for metrics cells' text}

\item{col.bg.out}{Color for metrics cells' background}

\item{theme}{Character: "light", or "dark"}

\item{mid.color}{Color: The mid color for the confusion matrix. Default = "white" for theme = "light",
"black" for "dark"}

\item{hi.color.pos}{Color: The hi color for correct classification.}

\item{hi.color.neg}{Color: The hi color for missclassification}

\item{par.reset}{Logical: If TRUE, reset par before exit. Default = TRUE}

\item{pdf.width}{Float: PDF width, if \code{filename} is set}

\item{pdf.height}{Float: PDF height, if \code{filename} is set}

\item{filename}{Character: If specified, save plot to this path. Default = NULL}

\item{...}{Not used}
}
\value{
List of metrics, invisibly
}
\description{
Plots an extended confusion matrix using \link{mplot3_img}
}
\author{
E.D. Gennatas
}
