% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_decision.R
\name{mplot3_decision}
\alias{mplot3_decision}
\title{\code{mplot3}: Decision boundaries}
\usage{
mplot3_decision(
  rtmod,
  data,
  vars = c(1, 2),
  dots.per.axis = 100,
  bg.cex = 0.5,
  bg.alpha = 0.4,
  bg.pch = 15,
  par.reset = TRUE,
  theme = "white",
  col = c("#18A3AC", "#F48024"),
  contour.col = "black",
  contour.lwd = 0.1,
  point.pch = c(3, 4),
  point.alpha = 1
)
}
\arguments{
\item{rtmod}{rtemics trained model}

\item{data}{Matrix / data frame of features; last column is class}

\item{vars}{Integer vector, length 2: Index of features (columns of \code{x}) to use to draw decision
boundaries. Default = c(1, 2)}

\item{dots.per.axis}{Integer: Draw a grid with this many dots on each axis. Default = 100}

\item{bg.cex}{Float: Point cex for background / decision surface. Default = .5}

\item{bg.alpha}{Float: Point alpha for background / decision surface. Default = .2}

\item{bg.pch}{Integer vector: pch for background / decision surface. Default = c(3, 4)}

\item{par.reset}{Logical: If TRUE, reset \code{par} before exiting. Default = TRUE}

\item{theme}{Character: Theme for \link{mplot3_xy}, "light" or "dark". Default = "light'}

\item{col}{Color vector for classes. Default = \code{ucsfPalette}}

\item{contour.col}{Color for decision boundary. Default = "black"}

\item{contour.lwd}{Float: Line width for decision boundary. Default = .3}

\item{point.pch}{Integer: pch for data points. Default = c(3, 4)}

\item{point.alpha}{Float: Alpha for data points. Default = 1}
}
\value{
Predicted labels for background grid (invisibly)
}
\description{
Plot classification decision boundaries of rtemis models
}
\details{
If data has more than 2 variables, any variable not selected using \code{vars} will be fixed to their mean
Underlying model (e.g. \code{randomForest}, \code{rpart}, etc) must support standard R predict format for classification:
\code{predict(model, newdata, type = "class")}
}
\examples{
\dontrun{
dat <- as.data.frame(mlbench::mlbench.2dnormals(200))
mod.cart <- s_CART(dat)
mod.rf <- s_RF(dat)
mplot3_decision(mod.cart, dat)
mplot3_decision(mod.rf, dat)
}
}
\author{
E.D. Gennatas
}
