% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_xy.R
\name{mplot3_fit}
\alias{mplot3_fit}
\title{True vs. Fitted plot}
\usage{
mplot3_fit(
  x,
  y,
  fit = "lm",
  se.fit = TRUE,
  fit.error = TRUE,
  axes.equal = TRUE,
  diagonal = TRUE,
  theme = rtTheme,
  marker.col = NULL,
  fit.col = NULL,
  pty = "s",
  fit.legend = FALSE,
  mar = NULL,
  ...
)
}
\arguments{
\item{x}{Numeric vector: True values}

\item{y}{Numeric vector: Predicted values}

\item{fit}{Character: \pkg{rtemis} model to calculate \code{y ~ x} fit.
Options: see \code{select_learn()}
Can also be Logical, which will give a GAM fit if TRUE. If you specify
"NLA", the activation function should be passed as a string.}

\item{se.fit}{Logical: If TRUE, draw the standard error of the fit}

\item{fit.error}{Logical: If TRUE: draw fit error annotation.
Default = NULL, which results in TRUE, if fit is set}

\item{axes.equal}{Logical: Should axes be equal? Defaults to FALSE}

\item{diagonal}{Logical: If TRUE, draw diagonal line.}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{marker.col}{Color for marker}

\item{fit.col}{Color: Color of the fit line.}

\item{pty}{Character: "s" for square plot, "m" to fill device. Default = "s"}

\item{fit.legend}{Logical: If TRUE, show fit legend}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{...}{Additional argument to be passed to \link{mplot3_conf} (classification) or \link{mplot3_xy} (regression)}
}
\description{
An \code{mplot3_xy} wrapper with defaults for plotting a learner's performance
}
\author{
EDG
}
