% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_harmonograph.R
\name{mplot3_harmonograph}
\alias{mplot3_harmonograph}
\title{Plot a harmonograph}
\usage{
mplot3_harmonograph(
  steps = seq(1, 500, by = 0.01),
  seed = NULL,
  col = "white",
  alpha = 0.2,
  bg = "black",
  lwd = 1,
  text = NULL,
  text.side = 1,
  text.line = -1,
  text.adj = 0,
  text.padj = 0,
  text.col = NULL,
  mar = c(0, 0, 0, 0),
  oma = c(0, 0, 0, 0),
  xlim = NULL,
  ylim = NULL,
  new = FALSE,
  par.reset = TRUE
)
}
\arguments{
\item{steps}{Float, vector}

\item{seed}{Integer}

\item{col}{Line color. Default = "white"}

\item{alpha}{Alpha for line color \code{col}. Default = .2}

\item{bg}{Color for background. Default = "black"}

\item{lwd}{Float: Line width}

\item{text}{Character: Text you want printed along with the harmonograph. Default = NULL}

\item{text.side}{Integer {1, 2, 3, 4}: \code{side} argument for \code{mtext}}

\item{text.line}{Float: \code{line} argument for \code{mtext}}

\item{text.adj}{Float: \code{adj} argument for \code{mtext}}

\item{text.padj}{Float: \code{padj} argument for \code{mtext}}

\item{text.col}{Color: Text color. Default is same as \code{col}}

\item{mar}{Float vector, length 4: Plot margins. (\code{par}'s \code{mar} argument)}

\item{oma}{Float vector, length 4: Outer margins. (\code{par}'s \code{oma} argument)}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{new}{Logical. If TRUE, do not clear plot before drawing}

\item{par.reset}{Logical. If TRUE, reset par before exit}
}
\description{
Plot a \href{https://en.wikipedia.org/wiki/Harmonograph}{harmonograph}
}
\details{
Unless you define a seed, each graph will be random. Try different seeds if you want to reproduce your graphs.
Some seeds to try: 9, 17, 26, 202, 208, ...
}
\author{
E.D. Gennatas
}
