% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_img.R
\name{mplot3_img}
\alias{mplot3_img}
\title{Draw image (False color 2D)}
\usage{
mplot3_img(
  z,
  as.mat = TRUE,
  col = NULL,
  xnames = NULL,
  xnames.y = 0,
  ynames = NULL,
  main = NULL,
  main.adj = 0,
  x.axis.side = 3,
  y.axis.side = 2,
  x.axis.line = -0.5,
  y.axis.line = -0.5,
  x.axis.las = 0,
  y.axis.las = 1,
  x.tick.labs.adj = NULL,
  y.tick.labs.adj = NULL,
  x.axis.font = 1,
  y.axis.font = 1,
  xlab = NULL,
  ylab = NULL,
  xlab.adj = 0.5,
  ylab.adj = 0.5,
  xlab.line = 1.7,
  ylab.line = 1.7,
  xlab.padj = 0,
  ylab.padj = 0,
  xlab.side = 1,
  ylab.side = 2,
  main.col = NULL,
  axlab.col = NULL,
  axes.col = NULL,
  labs.col = NULL,
  tick.col = NULL,
  cell.lab.hi.col = NULL,
  cell.lab.lo.col = NULL,
  cex = 1.2,
  cex.ax = NULL,
  cex.x = NULL,
  cex.y = NULL,
  zlim = NULL,
  autorange = TRUE,
  pty = "m",
  mar = NULL,
  asp = NULL,
  ann = FALSE,
  axes = FALSE,
  cell.labs = NULL,
  cell.labs.col = NULL,
  cell.labs.autocol = TRUE,
  bg = NULL,
  theme = getOption("rt.theme", "white"),
  autolabel = letters,
  filename = NULL,
  file.width = NULL,
  file.height = NULL,
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{z}{Input matrix}

\item{as.mat}{Logical: If FALSE, rows and columns of z correspond to x and y coordinates accoridngly.
This is the \code{image} default.
If TRUE (default), resulting image's cells will be in the same order as values appear when
you print z in the console. This is \code{t(apply(z, 2, rev))}. In this case, you can think of z
as a table of values you want to pictures with colors. For example, you can convert a correlation table
to a figure. In this case, you might want to add \code{cell.labs} with the values. Consider first using
\link{ddSci}.}

\item{col}{Colors to use. Defaults to \code{colorGrad(100)}}

\item{cell.labs}{Matrix of same dimensions as z (Optional): Will be printed as strings over cells}

\item{cell.labs.col}{Color for \code{cell.labs}. If NULL, the upper and lower quartiles will be
set to "white", the rest "black".}

\item{bg}{Background color}

\item{filename}{String (Optional): Path to file where image should be saved. R-supported extensions:
".pdf", ".jpeg", ".png", ".tiff".}

\item{file.width}{Output Width in inches}

\item{file.height}{Output height in inches}

\item{par.reset}{Logical: If TRUE, par will be reset to original settings before exit. Default = TRUE}

\item{...}{Additional arguments to be passed to \code{graphics::image}}
}
\description{
Draw a bitmap from a matrix of values.
}
\details{
This is also a good way to plot a large heatmap.
This function calls \code{image} which is a lot faster than drawing heatmaps
}
\author{
E.D. Gennatas
}
