% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_lolli.R
\name{mplot3_lolli}
\alias{mplot3_lolli}
\title{\code{mplot3} Lollipop Plot}
\usage{
mplot3_lolli(
  x,
  order.on.x = TRUE,
  plot.top = 1,
  orientation = c("horizontal", "vertical"),
  xnames = NULL,
  points = TRUE,
  segments = TRUE,
  main = NULL,
  col = NULL,
  cex = 1.2,
  matching.segment.col = FALSE,
  segment.alpha = 0.333,
  lty = 3,
  lwd = 2,
  theme = rtTheme,
  palette = rtPalette,
  autolabel = letters,
  par.reset = TRUE,
  pdf.width = 6,
  pdf.height = 6,
  mar = c(2.5, 3, 2, 1),
  pty = "m",
  pch = 16,
  x.axis.at = NULL,
  y.axis.at = NULL,
  xlab = NULL,
  ylab = NULL,
  label.las = 1,
  label.padj = 0.5,
  xaxs = "r",
  yaxs = "r",
  xlab.adj = 0.5,
  ylab.adj = 0.5,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{Float, vector: Input data}

\item{order.on.x}{Logical: If TRUE, order by value of \code{x}. Default = TRUE}

\item{plot.top}{Float or Integer: If <= 1, plot this percent highest absolute values, otherwise plot this many top values.
i.e.: \code{plot.top = .2} will print the top 20\% highest values, and \code{plot.top = 20} will plot the top 20
highest values}

\item{xnames}{Character, vector: Names of \code{x}}

\item{main}{Character: Main title}

\item{col}{Color, vector: Lollipop color}

\item{cex}{Float: Character expansion factor for points. Default = 1.2}

\item{matching.segment.col}{Logical: If TRUE, color line segments using \code{col}, i.e. same as}

\item{segment.alpha}{Float: Transparency for line segments. Default = .5
points. Default = FALSE, in which case they are colored with \code{theme$fg}}

\item{lty}{Integer: Line type for segment segments. See \code{par("lty")} Default = 1}

\item{lwd}{Float: Width for segment segments. See \code{par("lty")} Default = 1}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{palette}{Vector of colors, or Character defining a builtin palette -
get options with \code{rtpalette()}}

\item{autolabel}{Character vector to be used to generate autolabels when using
\link{rtlayout} with \code{autolabel = TRUE}.}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename}
is set).}

\item{pdf.height}{Float: Height in inches for pdf output.}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills
graphics device size. Default = "m" (See \code{par("pty")})}

\item{pch}{Integer: Point character.}

\item{x.axis.at}{Float, vector: x coordinates to place tick marks.
Default = NULL, determined by \code{graphics::axis} aautomatically}

\item{y.axis.at}{As \code{x.axis.at} for y-axis}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4\% on either end;
"i": Use exact x-axis limits.}

\item{yaxs}{Character: as \code{xaxs} for the y-axis.}

\item{xlab.adj}{Float: \code{adj} for \code{xlab} (See \code{par("adj")})}

\item{ylab.adj}{Float: \code{adj} for \code{ylab} (See \code{par("adj")})}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{...}{Additional arguments to be passed to theme function}
}
\description{
\code{mplot3} Lollipop Plot
}
\examples{
\dontrun{
x <- rnorm(12)
mplot3_lolli(x)
# a "rounded" barplot
mplot3_lolli(x, segments = T, points = F,
             lty = 1, matching.segment.col = T,
             lwd = 10, segment.alpha = 1)
}
}
\author{
E.D. Gennatas
}
