% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_missing.R
\name{mplot3_missing}
\alias{mplot3_missing}
\title{Plot missingness}
\usage{
mplot3_missing(
  x,
  feat.names = NULL,
  case.names = NULL,
  main = NULL,
  col.missing = "#FE4AA3",
  show = c("percent", "total"),
  names.srt = 90,
  case.names.x = 0.25,
  case.names.every = NULL,
  theme = rtTheme,
  alpha = 1,
  mar = c(3, 3.5, 5.5, 1),
  oma = c(0.5, 0.5, 0.5, 0.5),
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{Data matrix or data.frame}

\item{feat.names}{Character: Feature names. Defaults to \code{colnames(x)}}

\item{case.names}{Character: Case names. Defaults to \code{rownames(x)}}

\item{main}{Character: Main title}

\item{col.missing}{Color for missing cases.}

\item{show}{Character: "percent" or "total". Show percent missing or total missing
per column on the x-axis}

\item{names.srt}{Numeric: Angle of feature names in degrees.}

\item{case.names.x}{Numeric: x position of case names}

\item{case.names.every}{Numeric: Show case names every this many cases}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{alpha}{Numeric: Multiply theme's \code{fg} color by this amount}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{oma}{Float, vector, length 4: Outer margins; see \code{par("oma")}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{...}{Additional arguments to be passed to theme function}
}
\description{
Plot missingness
}
\examples{
\dontrun{
dat <- iris
dat[c(1, 5, 17:20, 110, 115, 140), 1] <-
dat[c(12, 15, 55, 73, 100:103), 2] <-
dat[sample(1:150, 25), 4] <- NA
mplot_missing(dat)
}

}
