% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_res.R
\name{mplot3_res}
\alias{mplot3_res}
\title{\code{mplot3} Plot \code{resample}}
\usage{
mplot3_res(res, col = NULL, mar = NULL, theme = getOption("rt.theme"), ...)
}
\description{
Visualizes resampling output using \link{mplot3_img}
}
\details{
For resampling with no replacement where each case may be selected 0 or 1 time,
0 is white and 1 is teal For resampling with replacement, 0 is white, 1 is blue, 2 is teal
}
\examples{
\dontrun{
x <- rnorm(500)
res <- resample(x)
mplot3_res(res)
}
}
\author{
E.D. Gennatas
}
