% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_roc.R
\name{mplot3_roc}
\alias{mplot3_roc}
\title{\code{mplot3} ROC curves}
\usage{
mplot3_roc(
  prob,
  labels,
  method = c("rt", "pROC"),
  type = "TPR.FPR",
  balanced.accuracy = FALSE,
  main = "",
  col = NULL,
  cex = 1.2,
  lwd = 2.5,
  diagonal = TRUE,
  diagonal.lwd = 2.5,
  diagonal.lty = 3,
  group.legend = FALSE,
  annotation = TRUE,
  annotation.col = col,
  annot.line = NULL,
  annot.adj = 1,
  annot.font = 1,
  pty = "s",
  mar = c(2.5, 3, 2, 1),
  theme = getOption("rt.theme", "whitegrid"),
  palette = rtPalette,
  verbose = TRUE,
  par.reset = TRUE,
  filename = NULL,
  pdf.width = 5,
  pdf.height = 5,
  ...
)
}
\arguments{
\item{prob}{Vector, Float [0, 1]: Predicted probabilities (i.e. c(.1, .8, .2, .9))}

\item{labels}{Vector, Integer {0, 1}: True labels (i.e. c(0, 1, 0, 1))}

\item{method}{Character: "rt" or "pROC" will use \link{rtROC} and \code{pROC::roc} respectively
to get points of the ROC. Default = "rt"}

\item{type}{Character: "TPR.FPR" or "Sens.Spec". Only changes the x and y labels. True positive rate vs.
False positive rate and Sensitivity vs. Specificity. Default = "TPR.FPR"}

\item{balanced.accuracy}{Logical: If TRUE, annotate the point of maximal Balanced Accuracy. Default = FALSE}

\item{main}{Character: Plot title. Default = ""}

\item{col}{Color, vector: Colors to use for ROC curve(s)}

\item{cex}{Float: Character expansion factor. Default = 1.2}

\item{lwd}{Float: Line width. Default = 2.5}

\item{diagonal}{Logical: If TRUE, draw diagonal. Default = TRUE}

\item{diagonal.lwd}{Float: Line width for diagonal. Default = 2.5}

\item{diagonal.lty}{Integer: Line type for diagonal. Default = 1}

\item{group.legend}{Logical}

\item{annotation}{Character: Add annotation at the bottom right of the plot}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills 
graphics device size. Default = "m" (See \code{par("pty")})}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{palette}{Vector of colors, or Character defining a builtin palette - 
get options with \code{rtpalette()}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{filename}{Path to file: If supplied, plot will be printed to file}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename} 
is set).}

\item{pdf.height}{Float: Height in inches for pdf output.}

\item{...}{Additional parameters to pass to \link{mplot3_xy}}
}
\description{
Plot ROC curve for a binary classifier
}
\author{
E.D. Gennatas
}
