% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_surv.R
\name{mplot3_surv}
\alias{mplot3_surv}
\title{\code{mplot3}: Survival Plots}
\usage{
mplot3_surv(
  x,
  lty = 1,
  lwd = 2,
  alpha = 1,
  col = NULL,
  mark.censored = TRUE,
  normalize.time = FALSE,
  cex = 1.2,
  xlab = NULL,
  ylab = "Survival",
  main = "Kaplan-Meier curve",
  theme = rtTheme,
  palette = rtPalette,
  plot.error = FALSE,
  error.lty = 2,
  error.alpha = 0.5,
  group.legend = NULL,
  group.title = "",
  group.names = NULL,
  group.side = 3,
  group.adj = 0.98,
  group.padj = 2,
  group.at = NA,
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{Survival object / list of Survival objects created using \code{survival::Surv}}

\item{lty}{Integer: Line type. Default = 1. See \code{par("lty")}}

\item{lwd}{Float: Line width. Default = 2}

\item{alpha}{Float: Alpha for lines. Default = 1}

\item{normalize.time}{Logical: If TRUE, convert each input's time to 0-1 range. This is useful
when survival estimates are not provided in original time scale. Default = FALSE.}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: Plot title}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{palette}{Vector of colors, or Character defining a builtin palette -
get options with \code{rtpalette()}}

\item{group.legend}{Logical: If TRUE, place \code{group.names} in a legend}

\item{group.title}{Character: Group title, shown above group names. e.g.
if group names are c("San Francisco", "Philadelphia"), \code{group.title}
can be "City"}

\item{group.names}{(Optional) If multiple groups are plotted, use these
names if \code{group.title = TRUE}}

\item{group.side}{Integer: Side to show group legend}

\item{group.adj}{Float: \code{adj} for group legend. See \code{mtext("adj")}}

\item{group.padj}{Float: \code{padj} for group legend See \code{mtext("padj")}}

\item{group.at}{Float: location for group legend. See \code{mtext("at")}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{...}{Additional arguments to pass to \link{mplot3_xy}}
}
\description{
Plots survival step functions using \link{mplot3_xy}
}
\examples{
\dontrun{
library(survival)
mplot3_surv(Surv(time = lung$time, event = lung$status))
}
}
\author{
E.D. Gennatas
}
