% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_survfit.R
\name{mplot3_survfit}
\alias{mplot3_survfit}
\title{\code{mplot3}: Plot \code{survfit} objects}
\usage{
mplot3_survfit(
  x,
  lty = 1,
  lwd = 1.5,
  alpha = 1,
  col = NULL,
  plot.median = FALSE,
  group.median = FALSE,
  median.lty = 3,
  median.lwd = 2,
  median.col = theme$fg,
  median.alpha = 0.5,
  censor.mark = TRUE,
  censor.col = NULL,
  censor.alpha = 0.4,
  censor.pch = "I",
  censor.cex = 0.8,
  mark.censored = FALSE,
  nrisk.table = FALSE,
  nrisk.pos = "below",
  nrisk.spacing = 0.9,
  table.font = 1,
  time.at = NULL,
  time.by = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "Time",
  ylab = "Survival",
  main = "",
  theme = rtTheme,
  palette = rtPalette,
  plot.error = FALSE,
  error.alpha = 0.33,
  autonames = TRUE,
  group.legend = NULL,
  group.legend.type = c("legend", "mtext"),
  group.names = NULL,
  group.title = NULL,
  group.line = NULL,
  group.side = NULL,
  legend.x = NULL,
  mar = c(2.5, 3, 2, 1),
  oma = NULL,
  par.reset = TRUE,
  pdf.width = 6,
  pdf.height = 6,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{survfit object (output of \code{survival::survfit})}

\item{lty}{Integer: Line type. See \code{par("lty")}}

\item{lwd}{Float: Line width.}

\item{alpha}{Float: Alpha for lines.}

\item{col}{Color, vector: Color(s) to use for survival curves and annotations. If NULL,
taken from \code{palette}}

\item{plot.median}{Logical: If TRUE, draw lines at 50 percent median survival.}

\item{group.median}{Logical: If TRUE, include median survival times with group legend}

\item{median.lty}{Integer: Median survival line type}

\item{median.lwd}{Float: Median line width.}

\item{median.col}{Color for median survival lines}

\item{median.alpha}{Float, (0, 1): Transparency for median survival lines.}

\item{censor.mark}{Logical: If TRUE, mark each censored case.}

\item{censor.col}{Color to mark censored cases if \code{censor.mark = TRUE}}

\item{censor.alpha}{Transparency for \code{censor.col}.}

\item{censor.pch}{Character: Point character for censored marks.}

\item{censor.cex}{Float: Character expansion factor for censor marks.}

\item{mark.censored}{Logical: This is an alternative to \code{censor.mark} which whill mark
censored cases using the same color as the survival curve. It can be harder to distinguish the
censoring marks from the curve itself, therefore not preferred.}

\item{nrisk.table}{Logical: If TRUE, print Number at risk table.}

\item{nrisk.pos}{Character: "above" or "below": where to place \code{nrisk.table}}

\item{nrisk.spacing}{Float: Determines spacing between \code{nrisk.table} rows.}

\item{table.font}{Integer: 1: regular font, 2: bold.}

\item{time.at}{Float, vector: x-axis positions to place tickmarks and labels as well as n at risk
values if \code{nrisk.table = TRUE}}

\item{time.by}{Float: Divide time by this amount to determine placing of tickmarks}

\item{xlim}{Float, vector, length 2: x-axis limits}

\item{ylim}{Float, vector, length 2: y-axis limits}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: main title}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{palette}{Vector of colors, or Character defining a builtin palette - get options with
\code{rtpalette()}}

\item{autonames}{Logical: If TRUE, extract grouping variable names and level labels from \code{x}
and use for legend. It is best to give informative level labels, like female, male instead of
0, 1 when using this.}

\item{group.legend}{Logical: If TRUE, include group legend}

\item{group.names}{Character, vector: Group names to use. If NULL, extracted from \code{x}}

\item{group.title}{Character: Group legend title}

\item{group.line}{Float, vector: Lines to print group legend using \code{mtext}}

\item{group.side}{Integer: Side to print group legend. Default is determined by survival curves,
to avoid overlap of legend with curves.}

\item{mar}{Float, vector, length 4: Margins. See \code{par("mar")}}

\item{oma}{Float, vector, length 4: Outer margins. See \code{par("oma")}}

\item{par.reset}{Logical: If TRUE, reset par to initial values before exit}

\item{...}{Additional arguments to pass to theme}
}
\description{
Plots survival step functions using \link{mplot3_xy}
}
\examples{
\dontrun{
# Get the lung dataset
data(cancer, package = "survival")
sf1 <- survival::survfit(survival::Surv(time, status) ~ 1, data = lung)
mplot3_survfit(sf1)
sf2 <- survival::survfit(survival::Surv(time, status) ~ sex, data = lung)
mplot3_survfit(sf2)
# with N at risk table
mplot3_survfit(sf2, nrisk.table = TRUE)
}
}
\author{
E.D. Gennatas
}
