% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_x.R
\name{mplot3_x}
\alias{mplot3_x}
\title{\code{mplot3}: Univariate plots: index, histogram, density, QQ-line}
\usage{
mplot3_x(
  x,
  type = c("density", "histogram", "hd", "lhist", "index", "ts", "qqline"),
  group = NULL,
  data = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  index.ypad = 0.1,
  axes.swap = FALSE,
  axes.col = NULL,
  tick.col = NULL,
  cex = 1.2,
  col = NULL,
  alpha = 0.75,
  index.type = c("p", "l"),
  hist.breaks = "Sturges",
  hist.type = c("bars", "lines"),
  hist.probability = FALSE,
  hist.lwd = 3,
  density.line = FALSE,
  density.shade = TRUE,
  density.legend.side = 3,
  density.legend.adj = 0.98,
  density.bw = "nrd0",
  density.kernel = "gaussian",
  density.params = list(na.rm = na.rm),
  qqline.col = "#18A3AC",
  qqline.alpha = 1,
  pch = 16,
  point.col = NULL,
  point.cex = 1,
  point.bg.col = NULL,
  point.alpha = 0.66,
  hline = NULL,
  vline = NULL,
  diagonal = FALSE,
  grid = FALSE,
  grid.col = NULL,
  grid.alpha = 0.5,
  grid.lty = 3,
  grid.lwd = 1,
  annotation = NULL,
  annot.col = NULL,
  group.legend = NULL,
  group.title = "",
  group.names = NULL,
  group.side = 3,
  group.adj = 0.02,
  group.at = NA,
  text.xy = NULL,
  text.x = NULL,
  text.y = NULL,
  text.xy.cex = 1,
  text.xy.col = "white",
  line.col = "#008E00",
  x.axis.padj = -1.1,
  y.axis.padj = 0.9,
  labs.col = NULL,
  lab.adj = 0.5,
  density.avg = ifelse(type == "density", TRUE, FALSE),
  density.avg.fn = c("median", "mean"),
  density.avg.line = FALSE,
  density.avg.lwd = 1.5,
  density.avg.lty = 3,
  hline.col = "black",
  hline.lwd = 1,
  hline.lty = 1,
  vline.col = "black",
  vline.lwd = 1,
  vline.lty = 1,
  lty = 1,
  lwd = 2,
  qqline.lwd = lwd,
  density.lwd = lwd,
  theme = rtTheme,
  palette = rtPalette,
  pty = "m",
  mar = NULL,
  oma = rep(0, 4),
  xaxs = "r",
  yaxs = "r",
  autolabel = letters,
  new = FALSE,
  alpha.off = FALSE,
  na.rm = TRUE,
  par.reset = TRUE,
  filename = NULL,
  pdf.width = 6,
  pdf.height = 6,
  ...
)
}
\arguments{
\item{x}{Numeric vector or list of vectors, one for each group.
If \code{data} is provided, x is name of variable in \code{data}}

\item{type}{Character: "density", "histogram", "hd" (histogram bars & density lines),
"lhist" (line histogram like \link{mhist}; same as \verb{type = "hist", hist.type = "lines"}),
"index", "ts", "qqline"
Case-insensitive and supports partial matching: e.g. \code{mplot3_x(x, "H")} gives histogram}

\item{group}{Vector denoting group membership. Will be converted to factor.
If \code{data} is provided, \code{group} is name of variable if \code{data}}

\item{data}{Optional data frame containing x data}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: Plot title}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{index.ypad}{Float: Expand ylim by this much for plot type "index" Default = .1
(Stops points being cut off)}

\item{index.type}{Character: "p" for points (Default), "l" for lines (timeseries)}

\item{hist.breaks}{See \code{histogram("breaks")}}

\item{hist.type}{Character: "bars" or "lines". Default = "bars"}

\item{hist.lwd}{Float: Line width for \code{type = "histogram"}; \code{hist.type = "lines"}}

\item{density.line}{Logical: If TRUE, draw line for \code{type = "density"}. Default = FALSE}

\item{density.shade}{Logical: If TRUE, draw shaded polygon for \code{type = "density"}.
Default = TRUE}

\item{qqline.col}{Color for Q-Q line}

\item{qqline.alpha}{Float: Alpha for Q-Q line}

\item{pch}{Integer: Point character.}

\item{point.cex}{Float: Character expansion for points.}

\item{point.bg.col}{Color: point background}

\item{hline}{Vector: y-value(s) for horizontal lines.}

\item{vline}{Vector: x-value(s) for vertical lines.}

\item{diagonal}{Logical: If TRUE, draw diagonal line.}

\item{annotation}{Character: Add annotation at the bottom right of the plot}

\item{group.legend}{Logical: If TRUE, include legend with group names}

\item{group.title}{Character: Title above group names}

\item{group.names}{(Optional) If multiple groups are plotted, use these
names if \code{group.title = TRUE}}

\item{group.side}{Integer: Side to show group legend}

\item{group.adj}{Float: \code{adj} for group legend. See \code{mtext("adj")}}

\item{group.at}{Float: location for group legend. See \code{mtext("at")}}

\item{line.col}{Color for lines}

\item{labs.col}{Color for labels}

\item{lab.adj}{Adjust the axes labels. 0 = left adjust; 1 = right adjust; .5 = center (Default)}

\item{density.avg}{Logical: If TRUE, print mean of \code{x} along plot. Default = TRUE,
for \code{type = "density"}}

\item{density.avg.fn}{Character: "median" or "mean". Function to use if
\code{density.avg = TRUE}. Default = "median"}

\item{density.avg.line}{Logical: If TRUE, draw vertical lines at the density average x-value}

\item{density.avg.lwd}{Float: Line width for \code{density.avg.line}. Default = 1.5}

\item{density.avg.lty}{Integer: Line type for \code{density.avg.line}. Default = 3}

\item{hline.col}{Color for horizontal line(s)}

\item{hline.lwd}{Float: Width for horizontal line(s)}

\item{hline.lty}{Integer: Line type for horizontal line(s)}

\item{vline.col}{Color for vertical lines}

\item{vline.lwd}{Float: Width for vertical lines}

\item{vline.lty}{Integer: Line type for vertical lines}

\item{lty}{Integer: Line type. See \code{par("lty")}}

\item{lwd}{Integer: Line width. Used for \verb{type = "ts" or "density"}}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{palette}{Vector of colors, or Character defining a builtin palette -
get options with \code{rtpalette()}}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills
graphics device size. Default = "m" (See \code{par("pty")})}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{oma}{Float, vector, length 4: Outer margins; see \code{par("oma")}}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4\% on either end;
"i": Use exact x-axis limits.}

\item{yaxs}{Character: as \code{xaxs} for the y-axis.}

\item{autolabel}{Character vector to be used to generate autolabels when using
\link{rtlayout} with \code{autolabel = TRUE}.}

\item{new}{Logical: If TRUE, add plot to existing plot. See \code{par("new")}}

\item{na.rm}{Logical: Will be passed to all functions that support it. If set to FALSE,
input containing NA values will result in error, depending on the \code{type}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{filename}{Path to file: If supplied, plot will be printed to file}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename}
is set).}

\item{pdf.height}{Float: Height in inches for pdf output.}

\item{...}{Additional arguments to be passed to theme function}
}
\value{
Invisibly returns the output of \code{density}, \code{hist}, \code{qqnorm}, or NULL
}
\description{
Draw plots of 1-dimensional data: index, histogram, density, and Q-Q plots.
}
\details{
You can group data either by supplying x as a list where each element contains one vector per
group, or as a data frame where each column represents group,
or by providing a \code{group} variable, which will be converted to factor.
For bivariate plots, see \link{mplot3_xy} and \link{mplot3_xym}. For heatmaps, see
\link{mplot3_heatmap}
To plot histograms of multiple groups, it's best to use \code{hist.type = "lines"}, which will
use \link{mhist}
and space apart the breaks for each group
}
\examples{
\dontrun{
mplot3_x(iris)
mplot3_x(split(iris$Sepal.Length, iris$Species), xlab = "Sepal Length")
}

}
\seealso{
\link{mplot3_xy}, \link{mplot3_xym}, \link{mplot3_heatmap}
}
\author{
E.D. Gennatas
}
