% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_xy.R
\name{mplot3_xy}
\alias{mplot3_xy}
\title{\code{mplot3}: XY Scatter and line plots}
\usage{
mplot3_xy(
  x,
  y = NULL,
  fit = NULL,
  formula = NULL,
  se.fit = FALSE,
  fit.params = NULL,
  error.x = NULL,
  error.y = NULL,
  cluster = NULL,
  cluster.params = list(),
  data = NULL,
  type = "p",
  group = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  xpd = TRUE,
  xaxs = "r",
  yaxs = "r",
  log = "",
  rsq = NULL,
  rsq.pval = FALSE,
  rsq.side = 1,
  rsq.adj = 0.98,
  rsq.col = NULL,
  rsq.line = NULL,
  fit.error = FALSE,
  fit.error.side = 1,
  fit.error.padj = NA,
  xaxp = NULL,
  yaxp = NULL,
  scatter = TRUE,
  axes.equal = FALSE,
  pty = "m",
  annotation = NULL,
  annotation.col = NULL,
  x.axis.at = NULL,
  x.axis.labs = TRUE,
  y.axis.at = NULL,
  y.axis.labs = TRUE,
  xlab.adj = 0.5,
  ylab.adj = 0.5,
  mar = NULL,
  oma = rep(0, 4),
  point.cex = 1,
  point.bg.col = NULL,
  pch = ifelse(is.null(point.bg.col), 16, 21),
  line.col = NULL,
  line.alpha = 0.66,
  lwd = 1,
  lty = 1,
  marker.col = NULL,
  marker.alpha = NULL,
  error.x.col = NULL,
  error.y.col = NULL,
  error.x.lty = 1,
  error.y.lty = 1,
  error.x.lwd = 1,
  error.y.lwd = 1,
  error.arrow.code = 3,
  fit.col = NULL,
  fit.lwd = 2.5,
  fit.alpha = 1,
  fit.legend = ifelse(is.null(fit), FALSE, TRUE),
  se.lty = "poly",
  se.lwd = 1,
  se.col = NULL,
  se.alpha = 0.5,
  se.times = 1.96,
  se.border = FALSE,
  se.density = NULL,
  hline = NULL,
  hline.col = NULL,
  hline.lwd = 1.5,
  hline.lty = 3,
  vline = NULL,
  vline.lwd = 1.5,
  vline.col = "blue",
  vline.lty = 3,
  diagonal = FALSE,
  diagonal.inv = FALSE,
  diagonal.lwd = 1.5,
  diagonal.lty = 1,
  diagonal.col = "gray50",
  diagonal.alpha = 1,
  group.legend = NULL,
  group.title = NULL,
  group.names = NULL,
  group.side = 3,
  group.adj = 0.02,
  group.padj = 2,
  group.at = NA,
  fit.legend.col = NULL,
  fit.legend.side = 3,
  fit.legend.adj = 0.02,
  fit.legend.padj = 2,
  fit.legend.at = NA,
  rm.na = TRUE,
  theme = rtTheme,
  palette = rtPalette,
  order.on.x = NULL,
  autolabel = letters,
  new = FALSE,
  par.reset = TRUE,
  return.lims = FALSE,
  pdf.width = 6,
  pdf.height = 6,
  trace = 0,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{Numeric vector or list of vectors for x-axis.
If \code{data} is provided, name of variable, unquoted.}

\item{y}{Numeric vector of list of vectors for y-axis
If \code{data} is provided, name of variable, unquoted.}

\item{fit}{Character: \pkg{rtemis} model to calculate \code{y ~ x} fit.
Options: see \code{select_learn()}
Can also be Logical, which will give a GAM fit if TRUE. If you specify
"NLA", the activation function should be passed as a string.}

\item{formula}{Formula: Provide a formula to be solved using \link{s_NLS}.
If provided, \code{fit} is forced to \code{'nls'}. e.g. y ~ b * m ^ x for a
power curve. Note: \code{nls} is powerful but is prone to errors
and warnings. Use single letters for parameter names, no numbers.}

\item{se.fit}{Logical: If TRUE, draw the standard error of the fit}

\item{fit.params}{List: Arguments for learner defined by \code{fit}.
Default = NULL, i.e. use default learner parameters}

\item{error.x}{Vector, float: Error in \code{x} (e.g. standard deviation)
will be plotted as bars around point}

\item{error.y}{Vector, float: Error in \code{y} (e.g. standard deviation)
will be plotted as bars around point}

\item{cluster}{Character: Clusterer name. Will cluster
\code{data.frame(x, y)} and pass result to \code{group}.
Run \link{select_clust} for options}

\item{cluster.params}{List: Names list of parameters to pass to the
\code{cluster} function}

\item{data}{(Optional) data frame, where \code{x} and \code{y} are defined}

\item{type}{Character: "p" for points, "l" for lines, "s" for steps.
Default = "p". If \code{x} and/or \code{y} contains multiple
vectors, \code{type} can be a vector, e.g. \code{c("p", "l", "l")} will
give a set of points and two sets of lines. Otherwise, \code{type} is
recycled to length of x}

\item{group}{Vector: will be converted to factor.
If \code{data} is provided, name of variable, unquoted.}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: Plot title}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{xpd}{Logical or NA: FALSE: plotting clipped to plot region; TRUE:
plotting clipped to figure region; NA: plotting clipped to device region.}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4\% on either end;
"i": Use exact x-axis limits.}

\item{yaxs}{Character: as \code{xaxs} for the y-axis.}

\item{log}{Character: "x", "y", or "xy", defines if either or both axes should
be log-transformed.}

\item{rsq}{Logical: If TRUE, add legend with R-squared (if fit is not NULL)}

\item{rsq.pval}{Logical: If TRUE, add legend with R-squared and its p-value,
if fit is not NULL}

\item{rsq.side}{Integer: [1:4] Where to place the \code{rsq} annotation.
Default = 1 (i.e. bottom)}

\item{rsq.adj}{Float: Adjust \code{rsq} annotation. See \verb{mtext "adj"}}

\item{rsq.col}{Color: Color for \code{rsq} annotation. Default = NULL, which
results in \code{fit.col}}

\item{rsq.line}{Numeric: Passed to \verb{mtext "line"} to place R-squared annotation.}

\item{fit.error}{Logical: If TRUE: draw fit error annotation.
Default = NULL, which results in TRUE, if fit is set}

\item{fit.error.side}{Integer [1:4]: Which side to draw \code{fit.error} on.}

\item{fit.error.padj}{Float: See \code{mtext:padg} Default = NA}

\item{xaxp}{See \code{par("xaxp")}}

\item{yaxp}{See \code{par("yaxp")}}

\item{scatter}{Logical: If TRUE, plot (x, y) scatter points.}

\item{axes.equal}{Logical: Should axes be equal? Defaults to FALSE}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills
graphics device size. Default = "m" (See \code{par("pty")})}

\item{annotation}{Character: Add annotation at the bottom right of the plot}

\item{annotation.col}{Color for annotation}

\item{x.axis.at}{Float, vector: x coordinates to place tick marks.
Default = NULL, determined by \code{graphics::axis} aautomatically}

\item{x.axis.labs}{See \code{axis("labels")}}

\item{y.axis.at}{As \code{x.axis.at} for y-axis}

\item{y.axis.labs}{See \code{axis("labels")}}

\item{xlab.adj}{Float: \code{adj} for \code{xlab} (See \code{par("adj")})}

\item{ylab.adj}{Float: \code{adj} for \code{ylab} (See \code{par("adj")})}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{oma}{Float, vector, length 4: Outer margins; see \code{par("oma")}}

\item{point.cex}{Float: Character expansion for points.}

\item{point.bg.col}{Color: point background}

\item{pch}{Integer: Point character.}

\item{line.col}{Color for lines}

\item{line.alpha}{Float [0, 1]: Transparency for lines}

\item{lwd}{Float: Line width}

\item{lty}{Integer: Line type. See \code{par("lty")}}

\item{marker.col}{Color for marker}

\item{marker.alpha}{Float [0, 1]: Transparency for markers}

\item{error.x.col}{Color for x-axis error bars}

\item{error.y.col}{Color for y-axis error bars}

\item{error.x.lty}{Integer: line type for x-axis error bars}

\item{error.y.lty}{Integer: line type for y-axis error bars}

\item{error.x.lwd}{Float: Line width for x-axis error bars}

\item{error.y.lwd}{Float: Line width for y-axis error bars}

\item{error.arrow.code}{Integer: Type of arrow to draw for error bars.
See \code{arrows("code")}}

\item{fit.col}{Color: Color of the fit line.}

\item{fit.lwd}{Float: Fit line width}

\item{fit.alpha}{Float [0, 1]: Transparency for fit line}

\item{fit.legend}{Logical: If TRUE, show fit legend}

\item{se.lty}{How to draw the \code{se.fit} "poly" draws a polygon around
the fit line, otherwise an integer defines the lty (line type) for lines to
be drawn}

\item{se.lwd}{Float: Line width for standard error bounds}

\item{se.col}{Color for \code{se.fit}}

\item{se.alpha}{Alpha for \code{se.fit}}

\item{se.times}{Draw polygon or lines at +/- \code{se.times} * standard
error of fit. Defaults to 1.96, which corresponds to 95 percent confidence interval}

\item{se.border}{Define border of polygon for \code{se.fit}.
See \code{border} in \code{graphics::polygon}}

\item{se.density}{Density of shading line of polygon for \code{se.fit}.
See \code{density} in \code{graphics::polygon}}

\item{hline}{Vector: y-value(s) for horizontal lines.}

\item{hline.col}{Color for horizontal line(s)}

\item{hline.lwd}{Float: Width for horizontal line(s)}

\item{hline.lty}{Integer: Line type for horizontal line(s)}

\item{vline}{Vector: x-value(s) for vertical lines.}

\item{vline.lwd}{Float: Width for vertical lines}

\item{vline.col}{Color for vertical lines}

\item{vline.lty}{Integer: Line type for vertical lines}

\item{diagonal}{Logical: If TRUE, draw diagonal line.}

\item{diagonal.inv}{Logical: If TRUE, draw inverse diagonal line. Will use
\code{diagonal.lwd}, \code{diagonal.lty}, \code{diagonal.col}, \code{diagonal.alpha} (Note: it only
makes sense to use only one of \code{diagonal} or \code{diagonal.inv})}

\item{diagonal.lwd}{Float: Line width for \code{diagonal}.}

\item{diagonal.lty}{Integer: Line type for \code{diagonal}.}

\item{diagonal.col}{Color: Color for \code{diagonal}.}

\item{diagonal.alpha}{Float: Alpha for \code{diagonal}}

\item{group.legend}{Logical: If TRUE, place \code{group.names} in a legend}

\item{group.title}{Character: Group title, shown above group names. e.g.
if group names are c("San Francisco", "Philadelphia"), \code{group.title}
can be "City"}

\item{group.names}{(Optional) If multiple groups are plotted, use these
names if \code{group.title = TRUE}}

\item{group.side}{Integer: Side to show group legend}

\item{group.adj}{Float: \code{adj} for group legend. See \code{mtext("adj")}}

\item{group.padj}{Float: \code{padj} for group legend See \code{mtext("padj")}}

\item{group.at}{Float: location for group legend. See \code{mtext("at")}}

\item{fit.legend.col}{Color for fit legend}

\item{fit.legend.side}{Integer: Side for fit legend}

\item{fit.legend.adj}{Float: \code{adj} for fit legend}

\item{fit.legend.padj}{Float: \code{padj} for fit legend}

\item{fit.legend.at}{Float: location for fit legend. See \code{mtext("at")}}

\item{rm.na}{Logical: If TRUE, remove all NA values pairwise between x and y.
Set to FALSE if you know your data has no missing values.}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{palette}{Vector of colors, or Character defining a builtin palette -
get options with \code{rtpalette()}}

\item{order.on.x}{Logical: If TRUE, order (x, y) by increasing x.
Default = NULL: will be set to TRUE if fit is set, otherwise FALSE}

\item{autolabel}{Character vector to be used to generate autolabels when using
\link{rtlayout} with \code{autolabel = TRUE}.}

\item{new}{Logical: If TRUE, add plot to existing plot. See \code{par("new")}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{return.lims}{Logical: If TRUE, return xlim and ylim.}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename}
is set).}

\item{pdf.height}{Float: Height in inches for pdf output.}

\item{trace}{Integer: If > 0, pass \code{verbose = TRUE} to the cluster and
fit functions, if used.}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{...}{Additional arguments to be passed to theme function}
}
\description{
Plot points and lines with optional fits and standard error bands
}
\examples{
\dontrun{
set.seed(1999)
x <- rnorm(500)
ycu <- x^3 + 12 + rnorm(500)
mplot3_xy(x, ycu)
mplot3_xy(x, ycu, fit = "gam")
ysq <- x^2 + 3 + rnorm(500)
mplot3_xy(x, list(squared = ysq, cubed = ycu), fit = "gam")
}
}
\author{
E.D. Gennatas
}
