% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot_AGGTEobj.R
\name{mplot_AGGTEobj}
\alias{mplot_AGGTEobj}
\title{Plot AGGTEobj object}
\usage{
mplot_AGGTEobj(
  x,
  x.factor = 1,
  y.factor = 1,
  error = c("se", "95\%ci"),
  main = "Average Effect by Length of Exposure",
  legend.title = "",
  group.names = c("Pre", "Post"),
  xlab = NULL,
  ylab = NULL,
  mar = c(2.5, 3.5, 2, 7),
  theme = rtTheme,
  font.family = "Helvetica",
  col = c("#EC1848", "#18A3AC"),
  filename = NULL,
  file.width = 6.5,
  file.height = 5.5,
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{AGGTEobj object}

\item{main}{Character: Plot title}

\item{group.names}{(Optional) If multiple groups are plotted, use these
names if \code{group.title = TRUE}}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{...}{Additional arguments to be passed to theme function}
}
\description{
Plot AGGTEobj object from the \pkg{did} package.
}
\author{
E.D. Gennatas
}
