% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot_raster.R
\name{mplot_raster}
\alias{mplot_raster}
\title{Plot Array as Raster Image}
\usage{
mplot_raster(
  x,
  max.value = max(x),
  mar = NULL,
  main = NULL,
  main.line = 0,
  main.side = 3,
  main.col = "#ffffff",
  main.adj = 0,
  main.font = 2,
  mono = FALSE,
  mono.fn = mean,
  bg = "gray10",
  par.set = TRUE,
  par.reset = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{Array, 2D or 3D: Input describing grayscale or color image in RGB space}

\item{mono}{Logical: If TRUE, plot as grayscale using \code{mono.fn} to convert RGB to grayscale. Default = FALSE}

\item{mono.fn}{Function: Apply this function to the array to convert to 2D for grayscale plotting. Default = mean}

\item{bg}{Color: Background color (around the plotted image when window proportions do not match image). Default = "gray10"}

\item{par.reset}{Logical: If TRUE, reset par settings before exiting. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Plots 2D (grayscale) or 3D (color) array as Raster Image
}
\examples{
\dontrun{
img <- imager::load.image("https://www.r-project.org/logo/Rlogo.png")
mplot_raster(img)
}
}
\author{
E.D. Gennatas
}
