% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\alias{msg0}
\title{Message with provenance}
\usage{
msg(
  ...,
  date = TRUE,
  caller = NULL,
  call.depth = 1,
  caller.id = 1,
  newline = TRUE,
  extraline = FALSE,
  newline.pre = FALSE,
  as.message = FALSE,
  color = NULL,
  sep = " "
)

msg0(
  ...,
  date = TRUE,
  caller = NULL,
  call.depth = 1,
  caller.id = 1,
  newline = TRUE,
  extraline = FALSE,
  newline.pre = FALSE,
  as.message = FALSE,
  color = NULL,
  sep = ""
)
}
\arguments{
\item{...}{Message to print}

\item{date}{Logical: if TRUE, include date and time in the prefix}

\item{caller}{Character: Name of calling function}

\item{call.depth}{Integer: Print the system call path of this depth. Default = NULL}

\item{caller.id}{Integer: Which function in the call stack to print}

\item{newline}{Logical: If TRUE end with a new line. Default = FALSE}

\item{newline.pre}{Logical: If TRUE begin with a new line. Default = FALSE}

\item{as.message}{Logical: if TRUE, print using \code{message()}}

\item{color}{Crayon color for message e.g. \code{crayon::red}}

\item{sep}{Character: Use to separate objects in \code{...}}
}
\value{
Invisibly: List with call, message, and date
}
\description{
Print message to output with a prefix including data and time, and calling function or full
call stack
}
\details{
If \code{msg} is called directly from the console, it will print \code{[interactive>]} in place of
  the call stack.
\code{msg0}, similar to \code{paste0}, is \code{msg(..., sep = "")}
}
\examples{
\dontrun{
msg("Your dataset has", nrow(iris), "rows and", ncol(iris), "columns")
}

}
\author{
E.D. Gennatas
}
