% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg2}
\alias{msg2}
\title{msg2}
\usage{
msg2(
  ...,
  date = rtDate,
  caller = NULL,
  call.depth = 1,
  caller.id = 1,
  newline.pre = FALSE,
  newline = TRUE,
  sep = " "
)
}
\arguments{
\item{...}{Message to print}

\item{date}{Logical: if TRUE, include date and time in the prefix}

\item{caller}{Character: Name of calling function}

\item{call.depth}{Integer: Print the system call path of this depth. Default = NULL}

\item{caller.id}{Integer: Which function in the call stack to print}

\item{newline.pre}{Logical: If TRUE begin with a new line. Default = FALSE}

\item{newline}{Logical: If TRUE end with a new line. Default = FALSE}

\item{sep}{Character: Use to separate objects in \code{...}}
}
\description{
msg2
}
\keyword{internal}
