% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg2done}
\alias{msg2done}
\title{msg2done}
\usage{
msg2done(caller = NULL, call.depth = 1, caller.id = 1, sep = " ")
}
\arguments{
\item{caller}{Character: Name of calling function}

\item{call.depth}{Integer: Print the system call path of this depth. Default = NULL}

\item{caller.id}{Integer: Which function in the call stack to print}

\item{sep}{Character: Use to separate objects in \code{...}}
}
\description{
msg2done
}
\keyword{internal}
