% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg2start}
\alias{msg2start}
\title{msg2}
\usage{
msg2start(..., newline.pre = FALSE, sep = " ")
}
\arguments{
\item{...}{Message to print}

\item{newline.pre}{Logical: If TRUE begin with a new line. Default = FALSE}

\item{sep}{Character: Use to separate objects in \code{...}}
}
\description{
msg2
}
