% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCr.R
\name{nCr}
\alias{nCr}
\title{n Choose r}
\usage{
nCr(n, r)
}
\arguments{
\item{n}{Integer: Total number of items}

\item{r}{Integer: Number of items in each combination}
}
\value{
Integer: Number of combinations
}
\description{
Calculate number of combinations
}
\details{
In plain language:
You have \code{n} items. How many different cobinations of \code{r} items can you make?
}
\author{
E.D. Gennatas
}
