% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlareg.R
\name{nlareg}
\alias{nlareg}
\title{\pkg{rtemis} internal: NonLinear Activation regression (NLAreg)}
\usage{
nlareg(
  x,
  y,
  b_o = mean(y),
  W_o = 1,
  b_h = 0,
  W_h = 0,
  activation = softplus,
  optim.method = "BFGS",
  control = list(),
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{b_o}{Float, vector (length y): Output bias. Defaults to \code{mean(y)}}

\item{W_o}{Float: Output weight. Defaults to 1}

\item{b_h}{Float: Hidden layer bias. Defaults to 0}

\item{W_h}{Float, vector (length \code{NCOL(x)}): Hidden layer weights. Defaults to 0}

\item{activation}{String or Function: Activation function to use: provide its name or the function itself.
Default = link{softplus}}

\item{optim.method}{Character: Optimization method to use: "Nelder-Mead", "BFGS", "CG", "L-BFGS-B",
"SANN", "Brent". See \code{stats::optim} for more details. Default = \code{"BFGS"}}

\item{control}{List: Control parameters passed to \code{stats::optim}}

\item{...}{Additional arguments}
}
\value{
\code{nlareg} object
}
\description{
\pkg{rtemis} internal: NonLinear Activation regression (NLAreg)
}
\author{
E.D. Gennatas
}
\keyword{internal}
