% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{nunique_perfeat}
\alias{nunique_perfeat}
\title{Number of unique values per feature}
\usage{
nunique_perfeat(x, excludeNA = FALSE)
}
\arguments{
\item{x}{data.table}

\item{excludeNA}{Logical: If TRUE, exclude NA values.}
}
\value{
Integer vector of length \code{NCOL(x)} with number of unique values per column/feature.
}
\description{
Number of unique values per feature
}
\author{
EDG
}
