% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcat.R
\name{objcat}
\alias{objcat}
\title{`rtemis-internal`: Object cat}
\usage{
objcat(x)
}
\arguments{
\item{x}{Character: Object description}
}
\description{
`rtemis-internal`: Object cat
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
