% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratiotable.R
\name{oddsratiotable}
\alias{oddsratiotable}
\title{Odds ratio table from logistic regression}
\usage{
oddsratiotable(x, confint.method = c("default", "profilelikelihood"))
}
\arguments{
\item{x}{\link{glm} object fit with \code{family = binomial}}

\item{confint.method}{"default" or "profilelikelihood"}
}
\value{
matrix with 4 columns: OR, 2.5\% & 97.5\% CI, p_val
}
\description{
Odds ratio table from logistic regression
}
\examples{
\dontrun{
ir2 <- iris[51:150, ]
ir2$Species <- factor(ir2$Species)
ir.fit <- glm(Species ~ ., data = ir2, family = binomial)
oddsratiotable(ir.fit)
}
}
\author{
E.D. Gennatas
}
