% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneHot.R, R/oneHot.data.frame.R,
%   R/oneHot.default.R
\name{oneHot}
\alias{oneHot}
\alias{oneHot.data.frame}
\alias{oneHot.default}
\title{One hot encoding}
\usage{
oneHot(x, verbose = TRUE)

\method{oneHot}{data.frame}(x, verbose = TRUE)

\method{oneHot}{default}(x, verbose = TRUE)
}
\arguments{
\item{x}{Vector or data.frame}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\value{
For vector input, a one-hot-encoded matrix, for data.frame frame input, a data.frame where all factors are
one-hot encoded
}
\description{
One hot encode a vector or all factors in a data.frame
}
\author{
Efstathios D Gennatas
}
