% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{order_colors}
\alias{order_colors}
\title{Order colors}
\usage{
order_colors(x, start_with = 1, order_by = "similarity")
}
\arguments{
\item{x}{Vector of colors}

\item{start_with}{Integer: Which color to output in first position}

\item{order_by}{Character: "similarity" or "dissimilarity"}
}
\description{
Order colors by RGB distance
}
\examples{

}
\author{
E.D. Gennatas
}
