% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_init.R
\name{outcome_name}
\alias{outcome_name}
\title{Get the name of the last column}
\usage{
outcome_name(x)
}
\arguments{
\item{x}{tabular data.}
}
\value{
Name of the last column.
}
\description{
Get the name of the last column
}
\details{
This applied to tabular datasets used for supervised learning in rtemis,
where, by convention, the last column is the outcome variable and all other columns
are features.
}
\examples{
outcome_name(iris)
}
\author{
EDG
}
