% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{palettize}
\alias{palettize}
\title{Palettize colors}
\usage{
palettize(
  x,
  grayscale_hicut = 0.8,
  start_with = "#16A0AC",
  order_by = c("separation", "dissimilarity", "similarity")
)
}
\arguments{
\item{x}{Color vector}

\item{grayscale_hicut}{Numeric: exclude colors whose grayscale equivalent
is greater than this value}

\item{start_with}{Integer or color: For integer, start with this color out
of \code{x}, otherwise find color \code{x} closer to this color and place it
first}

\item{order_by}{Character: "separation", "dissimilarity", "similarity"}
}
\description{
Filter and order a set of colors to produce a palette suitable for
multicolor plots
}
\author{
E.D. Gennatas
}
