% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemisInit.R
\name{parameterSummary}
\alias{parameterSummary}
\title{\code{rtemis-internals}: \code{parameterSummary}}
\usage{
parameterSummary(..., title = "Parameters", pad = 0,
  newline.pre = TRUE, newline = FALSE)
}
\arguments{
\item{...}{Variables to print}

\item{title}{String: title}

\item{pad}{Integer: Pad for \link{printls}}
}
\description{
Pretty print a collection of objects
}
\details{
Any number of R objects can be supplied. Their names and corresponding values will be printed line by line.
If input is a single data frame or list, its elements names and values will be printed.
If input includes multiple objects, any list's or data frame's elements' names will be printed with the
name of the list or data frame prepended.
}
\author{
Efstathios D. Gennatas
}
