% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytreew.R
\name{partLmw}
\alias{partLmw}
\title{\pkg{rtemis} internal: Ridge and Stump}
\usage{
partLmw(
  x1,
  y1,
  weights,
  .env,
  alpha = 1,
  lambda = 0.1,
  lambda.seq = NULL,
  part.minsplit = 2,
  part.xval = 0,
  part.max.depth = 1,
  part.cp = 0,
  part.minbucket = 5,
  minobsinnode.lin = 5,
  lin.type = "glmnet",
  verbose = TRUE,
  trace = 0
)
}
\description{
Fit a linear model on (x, y) and a tree on the residual yhat - y
}
