% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{pcat}
\alias{pcat}
\title{Pad-cat}
\usage{
pcat(left, right, pad = 17, newline = TRUE)
}
\description{
Pad-cat
}
\examples{
\dontrun{
{
  msg2("Hello")
  pcat("super", "wow")
  pcat(NULL, "oooo")
}
}
}
\keyword{internal}
