% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penn.heat.R
\name{penn.heat}
\alias{penn.heat}
\title{Create a color gradient}
\usage{
penn.heat(
  n = 11,
  space = "Lab",
  demo = FALSE,
  colorbar = FALSE,
  bar.min = -1,
  bar.mid = 0,
  bar.max = 1,
  cex = 1.2,
  filename = NULL,
  col.text = "black"
)
}
\arguments{
\item{n}{Integer. How many distinct colors you want. Defaults to 11}

\item{space}{String. Which colorspace to use. Option: "RGB", "Lab". Defaults to "Lab}

\item{demo}{Logical. Plot the colors horizontally}

\item{colorbar}{Logical. Create a vertical colorbar}

\item{bar.min}{Numeric. Lowest value in colorbar}

\item{bar.mid}{Numeric. Middle value in colorbar}

\item{bar.max}{Numeric. Max value in colorbar}

\item{filename}{Optional. Path to file to save colorbar}

\item{col.test}{Color for colorbar text}
}
\description{
Creates a gradient of colors ranging from light Penn blue (Penn blue with increased Value in HSV)
to Penn blue, to Penn red, to light Penn yellow.
It is best to provide an odd number, so that there is always an equal number of colors on either side
of the midpoint, which will be Penn blue.
For example, if you want a gradient from -1 to 1 or equivalent, an n = 11, will give 5 colors on either
side of 0, each representing a 20% change from the next.
}
