% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{pfread}
\alias{pfread}
\title{Read delimited file in parts}
\usage{
pfread(
  x,
  part_nrows,
  nrows = NULL,
  header = TRUE,
  sep = "auto",
  verbosity = 1L,
  stringsAsFactors = TRUE,
  ...
)
}
\arguments{
\item{x}{Character: Path to delimited file}

\item{part_nrows}{Integer: Number of rows to read in each part}

\item{nrows}{Integer: Number of rows in the file}

\item{header}{Logical: If TRUE, the file is assumed to include a header row}

\item{sep}{Character: Delimiter}

\item{verbosity}{Integer: Verbosity level.}

\item{stringsAsFactors}{Logical: If TRUE, characters will be converted to
factors}

\item{...}{Additional arguments to pass to \code{data.table::fread()}}
}
\value{
\code{data.table}.
}
\description{
Read delimited file in parts using \code{data.table::fread()}.
}
\author{
EDG
}
