% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_Supervised.R
\name{plot.Classification}
\alias{plot.Classification}
\title{Plot Classification}
\usage{
\method{plot}{Classification}(x, what = NULL, theme = "darkgraygrid", ...)
}
\arguments{
\item{x}{Classification object.}

\item{what}{Character vector: What to plot. "training", "validation", "test"}

\item{theme}{Character or Theme: Theme to use for the plot.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot Classification
}
\author{
EDG
}
