% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLAM.R
\name{plot.massGAM}
\alias{plot.massGAM}
\title{Plot \code{massGAM} object}
\usage{
\method{plot}{massGAM}(
  x,
  predictor = NULL,
  main = NULL,
  what = "pvals",
  p.adjust.method = "none",
  p.transform = function(x) -log10(x),
  show = c("all", "signif"),
  pval.hline = c(0.05, 0.001),
  hline.col = NULL,
  hline.dash = "dash",
  hline.annotate = as.character(pval.hline),
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  group = NULL,
  grouped.nonsig.alpha = 0.5,
  order.by.group = TRUE,
  palette = rtPalette,
  col.sig = "#43A4AC",
  col.ns = "#7f7f7f",
  theme = rtTheme,
  alpha = NULL,
  margin = NULL,
  displayModeBar = FALSE,
  trace = 0,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{\code{massGAM} object}

\item{xlab}{Character: x-axis label for volcano plot}
}
\description{
Plots a \code{massGAM} object using \link{dplot3_bar}
}
\author{
E.D. Gennatas
}
