% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/210_MassUni.R
\name{plot.MassGLM}
\alias{plot.MassGLM}
\title{Plot MassGLM using volcano plot}
\usage{
\method{plot}{MassGLM}(
  x,
  coefname = NULL,
  p_adjust_method = "holm",
  p_transform = function(x) -log10(x),
  xlab = "Coefficient",
  ylab = NULL,
  theme = choose_theme(getOption("rtemis_theme")),
  verbosity = 1L,
  ...
)
}
\arguments{
\item{x}{MassGLM object trained using \link{massGLM}.}

\item{coefname}{Character: Name of coefficient to plot. If \code{NULL}, the first coefficient is used.}

\item{p_adjust_method}{Character: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none" -
p-value adjustment method.}

\item{p_transform}{Function to transform p-values for plotting.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{theme}{\code{Theme} object. Create using one of the \code{theme_} functions, e.g.
\code{theme_whitegrid()}.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments passed to \link{draw_volcano}.}
}
\value{
plotly object with volcano plot.
}
\description{
Plot MassGLM using volcano plot
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
set.seed(2019)
y <- rnormmat(500, 500, return_df = TRUE)
x <- data.frame(x = y[, 3] + y[, 5] - y[, 9] + y[, 15] + rnorm(500))
mod <- massGLM(x, y)
plot(mod)
\dontshow{\}) # examplesIf}
}
\author{
EDG
}
