% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot.Regression}
\alias{plot.Regression}
\title{Plot Regression}
\arguments{
\item{x}{Regression object.}

\item{what}{Character vector: What to plot. Can include "training", "validation", "testing", or
"all", which will plot all available of the above.}
}
\description{
Plot Regression
}
\author{
EDG
}
