% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_cv.R
\name{plot.rtModCVCalibration}
\alias{plot.rtModCVCalibration}
\title{Plot \code{rtModCVCalibration} object}
\usage{
\method{plot}{rtModCVCalibration}(
  x,
  what = c("calibration", "brier"),
  type = c("aggregate.all", "aggregate.by.resample"),
  which.resample = 1,
  ...
)
}
\arguments{
\item{x}{\code{rtModCVCalibration} object}

\item{...}{Additional arguments}
}
\value{
plotly object
}
\description{
Plot \code{rtModCVCalibration} object
}
\author{
E.D. Gennatas
}
