% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{plot.rtTest}
\alias{plot.rtTest}
\title{Plot \code{rtTest} object}
\usage{
\method{plot}{rtTest}(
  x,
  main = NULL,
  mar = NULL,
  uni.type = c("density", "histogram", "hd"),
  boxplot.xlab = FALSE,
  theme = rtTheme,
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{rtTest} object}

\item{main}{Character: Main title}

\item{theme}{Character: Run \code{themes()} for available themes}
}
\description{
Plot \code{rtTest} object
}
\author{
E.D. Gennatas
}
