% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot.Supervised}
\alias{plot.Supervised}
\title{Plot Supervised}
\usage{
\method{plot}{Supervised}(x, ...)
}
\arguments{
\item{x}{Supervised object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot Supervised
}
\author{
EDG
}
