% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot_metric.SupervisedRes}
\alias{plot_metric.SupervisedRes}
\title{Plot Metric SupervisedRes}
\usage{
plot_metric.SupervisedRes(
  x,
  what = c("training", "test"),
  metric = NULL,
  ylab = labelify(metric),
  boxpoints = "all",
  theme = choose_theme(),
  ...
)
}
\arguments{
\item{x}{SupervisedRes object.}

\item{what}{Character vector: "training", "test". What to print. Default is to print both.}

\item{metric}{Character: Metric to plot.}

\item{ylab}{Character: Label for the y-axis.}

\item{boxpoints}{Character:"all", "outliers" - How to display points in the boxplot.}

\item{theme}{Theme object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot boxplot of performance metrics across resamples.
}
\author{
EDG
}
