% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot_roc.ClassificationRes}
\alias{plot_roc.ClassificationRes}
\title{Plot ROC for ClassificationRes}
\usage{
plot_roc.ClassificationRes(
  x,
  what = "all",
  theme = choose_theme(),
  col = rtpalette(rtemis_palette)[1:2],
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{ClassificationRes object.}

\item{what}{Character vector: "all", "training", "test". Which set(s) to plot.}

\item{theme}{Theme object.}

\item{col}{Character vector: Colors to use for the ROC curves.}

\item{filename}{Character: Filename to save the plot to.}

\item{...}{Additional arguments passed to \link{draw_roc}.}
}
\value{
plotly object.
}
\description{
Plot ROC for ClassificationRes
}
\author{
EDG
}
