% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot_true_pred.Classification}
\alias{plot_true_pred.Classification}
\title{Plot True vs. Predicted for Classification}
\usage{
plot_true_pred.Classification(x, what = NULL, theme = choose_theme(), ...)
}
\arguments{
\item{x}{Classification object.}

\item{what}{Character vector: What to plot. "training", "validation", "test"}

\item{theme}{Theme object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot True vs. Predicted for Classification
}
\author{
EDG
}
