% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot_true_pred.ClassificationRes}
\alias{plot_true_pred.ClassificationRes}
\title{Plot True vs. Predicted for ClassificationRes}
\usage{
plot_true_pred.ClassificationRes(x, what = "all", theme = choose_theme(), ...)
}
\arguments{
\item{x}{ClassificationRes object.}

\item{what}{Character vector: "all", "training", "test". Which set(s) to plot.}

\item{theme}{Theme object.}

\item{...}{Additional arguments passed to \link{draw_confusion}.}
}
\description{
Plot True vs. Predicted for ClassificationRes
}
\author{
EDG
}
