% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot_true_pred.Regression}
\alias{plot_true_pred.Regression}
\title{Plot True vs. Predicted for Regression}
\usage{
plot_true_pred.Regression(
  x,
  what = "all",
  fit = "glm",
  theme = choose_theme(),
  labelify = TRUE,
  ...
)
}
\arguments{
\item{x}{Regression object.}

\item{what}{Character vector: What to plot. Can include "training", "validation", "test", or
"all", which will plot all available.}

\item{fit}{Character: Algorithm to use to draw fit line.}

\item{theme}{Theme object.}

\item{labelify}{Logical: If TRUE, labelify the axis labels.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot True vs. Predicted for Regression
}
\author{
EDG
}
