% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_Supervised.R
\name{plot_true_pred.RegressionRes}
\alias{plot_true_pred.RegressionRes}
\title{Plot True vs. Predicted for RegressionRes}
\usage{
plot_true_pred.RegressionRes(
  x,
  what = "all",
  fit = "glm",
  theme = choose_theme(),
  labelify = TRUE,
  ...
)
}
\arguments{
\item{x}{RegressionRes object.}

\item{what}{Character vector: "all", "training", "test". Which set(s) to plot.}

\item{fit}{Character: Algorithm to use to draw fit line.}

\item{theme}{Theme object.}

\item{labelify}{Logical: If TRUE, labelify the axis labels.}

\item{...}{Additional arguments passed to \link{draw_fit}.}
}
\description{
Plot True vs. Predicted for RegressionRes
}
\author{
EDG
}
