% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.heat.R
\name{plotly.heat}
\alias{plotly.heat}
\title{Heatmap with \code{plotly}}
\usage{
plotly.heat(
  z,
  x = NULL,
  y = NULL,
  title = NULL,
  col = penn.heat(21),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  transpose = TRUE
)
}
\arguments{
\item{z}{Input matrix}

\item{x, y}{Vectors for x, y axes}

\item{title}{Plot title}

\item{col}{Set of colors to make gradient from}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{zlab}{z value label}

\item{transpose}{Logical: If TRUE, transpose matrix}
}
\description{
Draw a heatmap using \code{plotly}
}
\author{
E.D. Gennatas
}
