% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision}
\alias{precision}
\title{Precision (aka PPV)}
\usage{
precision(true, predicted, harmonize = FALSE, verbosity = 1L)
}
\arguments{
\item{true}{Factor: True labels}

\item{predicted}{Factor: predicted labels}

\item{harmonize}{Logical: If TRUE, run \link{factor_harmonize} first}

\item{verbosity}{Integer: If > 0, print messages to console.}
}
\description{
The first factor level is considered the positive case.
}
\keyword{internal}
