% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtboost.R
\name{predict.addtboost}
\alias{predict.addtboost}
\title{Predict method for \code{addtboost} object}
\usage{
\method{predict}{addtboost}(object, newdata = NULL,
  n.feat = NCOL(newdata), n.iter = NULL, fixed.cxr = NULL,
  as.matrix = FALSE, n.cores = 1, verbose = FALSE, ...)
}
\arguments{
\item{object}{\link{addtboost} object}
}
\description{
Predict method for \code{addtboost} object
}
\author{
Efstathios D. Gennatas
}
