% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.ADDT.R
\name{predict.addTree}
\alias{predict.addTree}
\title{Predict method for \code{addTree} object}
\usage{
\method{predict}{addTree}(object, newdata = NULL, learning.rate = NULL,
  n.feat = NULL, verbose = FALSE, cxrcoef = FALSE, trace = 0, ...)
}
\description{
Predict method for \code{addTree} object
}
\author{
Efstathios D. Gennatas
}
