% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{predict.boost}
\alias{predict.boost}
\title{Predict method for \code{boost} object}
\usage{
\method{predict}{boost}(
  object,
  newdata = NULL,
  n.feat = NCOL(newdata),
  n.iter = NULL,
  as.matrix = FALSE,
  verbose = FALSE,
  n.cores = rtCores,
  ...
)
}
\description{
Predict method for \code{boost} object
}
\author{
E.D. Gennatas
}
