% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartLinBoostTV.R
\name{predict.cartLinBoostTV}
\alias{predict.cartLinBoostTV}
\title{Predict method for \code{cartLinBoostTV} object}
\usage{
\method{predict}{cartLinBoostTV}(object, newdata = NULL,
  n.feat = NCOL(newdata), n.iter = NULL, as.matrix = FALSE,
  verbose = FALSE, n.cores = rtCores, ...)
}
\arguments{
\item{object}{\code{cartLinBoostTV} object}

\item{newdata}{Set of predictors}

\item{n.feat}{Integer: N of features to use. Default = NCOL(newdata)}

\item{n.iter}{Integer: N of iterations to predict from. Default = (all available)}

\item{as.matrix}{Logical: If TRUE, return predictions from each iterations. Default = FALSE}

\item{verbose}{Logical: If TRUE, print messages to console. Default = FALSE}

\item{n.cores}{Integer: Number of cores to use. Default = \code{rtCores}}
}
\description{
Predict method for \code{cartLinBoostTV} object
}
\author{
Efstathios D. Gennatas
}
