% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartLite.R
\name{predict.cartLite}
\alias{predict.cartLite}
\title{Predict method for \code{cartLite} object}
\usage{
\method{predict}{cartLite}(object, newdata, verbose = FALSE, ...)
}
\arguments{
\item{object}{\code{cartLite} object}

\item{newdata}{Data frame of predictors}

\item{verbose}{Logical: If TRUE, print messages to console.}

\item{...}{Unused}
}
\description{
Predict method for \code{cartLite} object
}
\author{
E.D. Gennatas
}
