% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamselx2.R
\name{predict.gamselx2}
\alias{predict.gamselx2}
\title{Predict Method for gamselx2 Fits}
\usage{
\method{predict}{gamselx2}(object, newdata = NULL, ...)
}
\description{
Obtains predictions from a fitted gamselx2 model object
}
\author{
Efstathios D. Gennatas
}
