% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmLite.R
\name{predict.glmLite}
\alias{predict.glmLite}
\title{Predict method for \code{glmLite} object}
\usage{
\method{predict}{glmLite}(object, newdata, verbose = FALSE, ...)
}
\arguments{
\item{object}{\link{glmLite} object}

\item{newdata}{Data frame of predictors}

\item{verbose}{Logical: If TRUE, print messages to console. Default = FALSE}

\item{...}{Unused}
}
\description{
Predict method for \code{glmLite} object
}
\author{
E.D. Gennatas
}
