% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{predict.hytboost}
\alias{predict.hytboost}
\title{Predict method for \code{hytboost} object}
\usage{
\method{predict}{hytboost}(
  object,
  newdata = NULL,
  n.feat = NCOL(newdata),
  n.iter = NULL,
  fixed.cxr = NULL,
  as.matrix = FALSE,
  n.cores = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{\link{hytboost} object}
}
\description{
Predict method for \code{hytboost} object
}
\author{
Efstathios D. Gennatas
}
