% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboostnow.R
\name{predict.hytboostnow}
\alias{predict.hytboostnow}
\title{Predict method for \code{hytboostnow} object}
\usage{
\method{predict}{hytboostnow}(
  object,
  newdata = NULL,
  n.feat = NCOL(newdata),
  n.iter = NULL,
  fixed.cxr = NULL,
  as.matrix = FALSE,
  n.cores = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{\code{hytboostnow} object}
}
\description{
Predict method for \code{hytboostnow} object
}
\author{
E.D. Gennatas
}
