% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JL_splitlin.R
\name{predict.JLsplitlin}
\alias{predict.JLsplitlin}
\title{predict JLsplitlin}
\usage{
\method{predict}{JLsplitlin}(object, newdata, ...)
}
\description{
predict JLsplitlin
}
\author{
E.D. Gennatas
}
\keyword{internal}
