% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LightRuleFit .R
\name{predict.LightRuleFit}
\alias{predict.LightRuleFit}
\title{\code{predict} method for \code{LightRuleFit} object}
\usage{
\method{predict}{LightRuleFit}(object, newdata = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{LightRuleFit} object}

\item{newdata}{Feature matrix / data.frame: will be converted to \code{data.table}}

\item{verbose}{Logical: If TRUE, print messages during execution. Default = TRUE}

\item{...}{Ignored}
}
\value{
Vector of estimated values
}
\description{
\code{predict} method for \code{LightRuleFit} object
}
