% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LIHAD.R
\name{predict.lihad}
\alias{predict.lihad}
\title{Predict method for \code{lihad} object}
\usage{
\method{predict}{lihad}(
  object,
  newdata = NULL,
  learning.rate = NULL,
  n.feat = NULL,
  verbose = FALSE,
  cxrcoef = FALSE,
  ...
)
}
\arguments{
\item{object}{an \code{rtMod} trained with \link{s_LIHAD} or an \code{lihad} object}

\item{newdata}{data frame of predictor features}

\item{learning.rate}{Float: learning rate if \code{object} was \code{lihad}}

\item{n.feat}{Integer: internal use only}

\item{verbose}{Logical: If TRUE, print messages to console.}

\item{cxrcoef}{Logical: If TRUE, return matrix of cases by coefficients along with
predictions.}

\item{...}{Not used}
}
\description{
Predict method for \code{lihad} object
}
\author{
E.D. Gennatas
}
