% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linadleaves.R
\name{predict.linadleaves}
\alias{predict.linadleaves}
\title{Predict method for \code{linadleaves} object}
\usage{
\method{predict}{linadleaves}(
  object,
  newdata,
  type = c("response", "probability", "all", "step"),
  n.leaves = NULL,
  fixed.cxr = NULL,
  cxr.newdata = NULL,
  cxr = FALSE,
  cxrcoef = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{\code{shytreeRaw}}

\item{newdata}{Data frame of predictors}

\item{type}{Character: "response", "probability", "all", "step"}

\item{n.leaves}{Integer: Use the first \code{n.leaves} of the tree for
prediction}

\item{fixed.cxr}{(internal use) Matrix: Cases by rules to use instead of matching cases to rules using
\code{newdata}}

\item{cxr.newdata}{(internal use) Data frame: Use these values to match cases by rules}

\item{cxr}{Logical: If TRUE, return list which includes cases-by-rules matrix along with predicted values}

\item{cxrcoef}{Logical: If TRUE, return cases-by-rules * coefficients matrix along with predicted values}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{...}{Not used}
}
\description{
Predict method for \code{linadleaves} object
}
\author{
E.D. Gennatas
}
