% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlareg.R
\name{predict.nlareg}
\alias{predict.nlareg}
\title{Predict method for \code{nlareg} object}
\usage{
\method{predict}{nlareg}(object, newdata, ...)
}
\arguments{
\item{object}{\link{nlareg} object}

\item{newdata}{Data frame of predictors}

\item{...}{Unused}
}
\description{
Predict method for \code{nlareg} object
}
\author{
E.D. Gennatas
}
