% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_supervised.R
\name{predict.nullmod}
\alias{predict.nullmod}
\title{\pkg{rtemis} internal: predict for an object of class \code{nullmod}}
\usage{
\method{predict}{nullmod}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{\code{nullmod} object.}

\item{newdata}{Not used.}

\item{...}{Not used.}
}
\description{
\pkg{rtemis} internal: predict for an object of class \code{nullmod}
}
