% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_cv.R
\name{predict.rtModCVCalibration}
\alias{predict.rtModCVCalibration}
\title{Predict using calibrated model}
\usage{
\method{predict}{rtModCVCalibration}(mod_cal, mod, newdata, which.repeat = 1)
}
\arguments{
\item{mod_cal}{\code{rtModCVCalibration} object returned by \link{calibrate_cv}}

\item{mod}{\code{rtModCV} object returned by \link{train_cv}}

\item{newdata}{Data frame: New data to predict on}

\item{which.repeat}{Integer: Which repeat to use for prediction}
}
\description{
Predict using a calibrated model returned by \link{calibrate_cv}
}
\author{
EDG
}
