% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_TLS.R
\name{predict.rtTLS}
\alias{predict.rtTLS}
\title{\code{predict.rtTLS}: \code{predict} method for \code{rtTLS} object}
\usage{
\method{predict}{rtTLS}(object, newdata, ...)
}
\arguments{
\item{object}{\code{rtTLS} object created by \link{s_TLS}}

\item{newdata}{\code{data.frame} of new data.}

\item{...}{Not used.}
}
\description{
\code{predict.rtTLS}: \code{predict} method for \code{rtTLS} object
}
