% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.RULEFEAT.R
\name{predict.ruleFeat}
\alias{predict.ruleFeat}
\title{\code{predict} method for \code{ruleFeat} object}
\usage{
\method{predict}{ruleFeat}(object, newdata = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{ruleFeat} object}

\item{newdata}{Feature matrix / data.frame: will be converted to \code{data.table}}

\item{verbose}{Logical: If TRUE, print messages during execution. Default = TRUE}

\item{...}{Ignored}
}
\value{
Vector of estimated values
}
\description{
\code{predict} method for \code{ruleFeat} object
}
