% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_CART.R
\name{predict_CART}
\alias{predict_CART}
\title{Predict from rpart model}
\usage{
predict_CART(model, newdata, type)
}
\arguments{
\item{model}{rpart model.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from rpart model
}
\keyword{internal}
