% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_GLM.R
\name{predict_GLM}
\alias{predict_GLM}
\title{Predict from GLM model}
\usage{
predict_GLM(model, newdata, type)
}
\arguments{
\item{model}{GLM model.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from GLM model
}
\keyword{internal}
