% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_GLMNET.R
\name{predict_GLMNET}
\alias{predict_GLMNET}
\title{Predict from GLMNET model}
\usage{
predict_GLMNET(model, newdata, type = NULL)
}
\arguments{
\item{model}{glmnet model.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from GLMNET model
}
\keyword{internal}
