% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_Isotonic.R
\name{predict_Isotonic}
\alias{predict_Isotonic}
\title{Predict from Isotonic model}
\usage{
predict_Isotonic(model, newdata, type)
}
\arguments{
\item{model}{Isotonic model.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from Isotonic model
}
\keyword{internal}
