% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_LightCART.R
\name{predict_LightCART}
\alias{predict_LightCART}
\title{Predict from LightCART LightGBM model}
\usage{
predict_LightCART(model, newdata, type, verbosity = 0L)
}
\arguments{
\item{model}{lgb.Booster object trained using \code{train_LightCART}.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from LightCART LightGBM model
}
\keyword{internal}
