% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_LightGBM.R
\name{predict_LightGBM}
\alias{predict_LightGBM}
\title{Predict from LightGBM LightGBM model}
\usage{
predict_LightGBM(model, newdata, type, verbosity = 0L)
}
\arguments{
\item{model}{lgb.Booster object trained using \code{train_LightGBM}.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from LightGBM LightGBM model
}
\keyword{internal}
