% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_LightRF.R
\name{predict_LightRF}
\alias{predict_LightRF}
\title{Predict from LightRF LightGBM model}
\usage{
predict_LightRF(model, newdata, type, verbosity = 0L)
}
\arguments{
\item{model}{lgb.Booster object trained using \code{train_LightRF}.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from LightRF LightGBM model
}
\keyword{internal}
