% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_LightRuleFit.R
\name{predict_LightRuleFit}
\alias{predict_LightRuleFit}
\title{Predict from LightRuleFit LightGBM model}
\usage{
predict_LightRuleFit(model, newdata, type, verbosity = 1L)
}
\arguments{
\item{model}{lgb.Booster object trained using \code{train_LightRuleFit}.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from LightRuleFit LightGBM model
}
\keyword{internal}
