% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_SVM.R
\name{predict_SVM}
\alias{predict_SVM}
\title{Predict from SVM model}
\usage{
predict_SVM(model, newdata, type, verbosity = 0L)
}
\arguments{
\item{model}{SVM model.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from SVM model
}
\keyword{internal}
