% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_TabNet.R
\name{predict_TabNet}
\alias{predict_TabNet}
\title{Predict from TabNet model}
\usage{
predict_TabNet(model, newdata, type)
}
\arguments{
\item{model}{TabNet model.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Predict from TabNet model
}
\keyword{internal}
