% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_ops.R
\name{preorderlgb}
\alias{preorderlgb}
\title{Preorder Traversal of LightGBM Tree}
\usage{
preorderlgb(
  tree,
  node,
  rule = "TRUE",
  left = "left_child",
  right = "right_child",
  split_feature = "split_feature",
  threshold = "threshold",
  xnames,
  factor_levels,
  trace = 0
)
}
\arguments{
\item{tree}{A LightGBM tree}
}
\description{
Called by \code{lgbtree2rules} and operates on \code{tree} environment in place.
}
\keyword{internal}
