% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_gridsearch.R
\name{present_gridsearch}
\alias{present_gridsearch}
\title{Present gridsearch results}
\usage{
present_gridsearch(x, rtModCV.repeat = 1, ...)
}
\arguments{
\item{x}{rtMod or rtModCV objects}

\item{rtModCV.repeat}{Integer: Which repeat to use, when x is rtModCV object}

\item{...}{Additional arguments to pass to \code{knitr::kable} used to print
the best tune table}
}
\description{
Present gridsearch results
}
\examples{
\dontrun{
mod <- s_CART(iris, cp = c(0, .1), maxdepth = c(3, 5))
mod_10ss <- elevate(iris, mod = "cart", cp = c(0, .1), maxdepth = c(3, 5))
present_gridsearch(mod)
present_gridsearch(mod_10ss)
}
}
\author{
E.D. Gennatas
}
