% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewcolor.R
\name{previewcolor}
\alias{previewcolor}
\title{Preview color v2.0}
\usage{
previewcolor(
  x,
  bg = "#333333",
  main = NULL,
  main.col = "#FFFFFF",
  width = NULL,
  ylim = c(0.7, 2),
  asp = 1,
  labels.y = 1.55,
  label.cex = NULL
)
}
\arguments{
\item{x}{Color, vector: One or more colors that R understands}

\item{bg}{Background color. Default = "#333333" (dark gray)}

\item{width}{Float: Plot width. Default = NULL, i.e. set automatically}

\item{asp}{Float: Plot aspect ratio. Default = 1}

\item{labels.y}{Float: y coord for labels. Default = 1.55 (rhombi are fixed and range y .5 - 1.5)}

\item{label.cex}{Float: Character expansion for labels. Default = NULL, and is calculated automatically based on
length of \code{x} in order to look reasonable in your RStudio plot panel.}
}
\value{
Nothing, prints plot
}
\description{
Preview one or multiple colors using little rhombi with their little labels up top
}
\examples{
colors <- colorgradient.x(seq(-5, 5))
previewcolor(colors)
}
