% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtboost.R
\name{print.addtboost}
\alias{print.addtboost}
\title{Print method for \link{boost} object}
\usage{
\method{print}{addtboost}(x, ...)
}
\description{
Print method for \link{boost} object
}
\author{
Efstathios D. Gennatas
}
