% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.ADDT.R
\name{print.addTree}
\alias{print.addTree}
\title{Print method for \code{addTree} object}
\usage{
\method{print}{addTree}(x, ...)
}
\description{
Print method for \code{addTree} object
}
\author{
Efstathios D. Gennatas
}
