% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTreeLeaves.R
\name{print.addTreeLeavesRC}
\alias{print.addTreeLeavesRC}
\title{Print method for \code{addTreeLeavesRC} object}
\usage{
\method{print}{addTree}(x, ...)
}
\arguments{
\item{x}{\code{addTreeLeavesRC} object}
}
\description{
Print method for \code{addTreeLeavesRC} object
}
\author{
Efstathios D. Gennatas
}
