% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{print.boost}
\alias{print.boost}
\title{Print method for \link{boost} object}
\usage{
\method{print}{boost}(x, ...)
}
\arguments{
\item{x}{\link{boost} object}

\item{...}{Not used}
}
\description{
Print method for \link{boost} object
}
\author{
E.D. Gennatas
}
