% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{print.CheckData}
\alias{print.CheckData}
\title{Describe generic}
\usage{
\method{print}{CheckData}(
  x,
  type = c("plaintext", "html"),
  name = NULL,
  check_integers = FALSE,
  css = list(font.family = "Helvetica", color = "#fff", background.color = "#242424"),
  ...
)
}
\arguments{
\item{x}{\code{CheckData} object}
}
\description{
Describe \code{CheckData} object
}
\details{
Print \code{CheckData} object
}
\author{
E.D. Gennatas
}
