% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_error.R
\name{print.class_error}
\alias{print.class_error}
\title{Print \link{class_error}}
\usage{
\method{print}{class_error}(x, decimal.places = 4, ...)
}
\arguments{
\item{x}{Object of type \link{class_error}}

\item{decimal.places}{Integer: Number of decimal places to print}

\item{...}{Not used}
}
\description{
Print \link{class_error}
}
\author{
E.D. Gennatas
}
