% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classError.R
\name{print.classError}
\alias{print.classError}
\title{Print \link{classError}}
\usage{
\method{print}{classError}(x, decimal.places = 4, ...)
}
\arguments{
\item{x}{Object of type \link{classError}}

\item{...}{Not used}
}
\description{
Print \link{classError}
}
\author{
Efstathios D. Gennatas
}
