% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmLiteBoostTV.R
\name{print.glmLiteBoostTV}
\alias{print.glmLiteBoostTV}
\title{Print method for \link{boost} object}
\usage{
\method{print}{boost}(x, ...)
}
\description{
Print method for \link{boost} object
}
\author{
Efstathios D. Gennatas
}
