% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{print.Hyperparameters}
\alias{print.Hyperparameters}
\title{Print Hyperparameters}
\usage{
\method{print}{Hyperparameters}(x, ...)
}
\arguments{
\item{x}{Hyperparameters object.}

\item{...}{Not used.}
}
\description{
Print method for Hyperparameters object.
}
\author{
EDG
}
