% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboostnow.R
\name{print.hytboostnow}
\alias{print.hytboostnow}
\title{Print method for \link{boost} object}
\usage{
\method{print}{hytboostnow}(x, ...)
}
\description{
Print method for \link{boost} object
}
\author{
Efstathios D. Gennatas
}
