% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.HYTREE.R
\name{print.hytree}
\alias{print.hytree}
\title{Print method for \code{hytree} object}
\usage{
\method{print}{hytree}(x, ...)
}
\description{
Print method for \code{hytree} object
}
\author{
Efstathios D. Gennatas
}
