% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LIHAD.R
\name{print.lihad}
\alias{print.lihad}
\title{Print method for \code{lihad} object}
\usage{
\method{print}{lihad}(x, ...)
}
\arguments{
\item{x}{\code{lihad} object}

\item{...}{Not used}
}
\description{
Print method for \code{lihad} object
}
\author{
E.D. Gennatas
}
