% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linadleaves.R
\name{print.linadleaves}
\alias{print.linadleaves}
\title{Print method for \code{linadleaves} object}
\usage{
\method{print}{linadleaves}(x, ...)
}
\arguments{
\item{x}{\code{linadleaves} object}

\item{...}{Not used}
}
\description{
Print method for \code{linadleaves} object
}
\author{
E.D. Gennatas
}
