% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R, R/massGLM2.R
\name{print.massGLM}
\alias{print.massGLM}
\title{\code{print}\link{massGLM} object}
\usage{
\method{print}{massGLM}(x, ...)

\method{print}{massGLM}(x, ...)
}
\arguments{
\item{x}{\link{massGLM} object}
}
\description{
\code{print}\link{massGLM} object

\code{print}\link{massGLM} object
}
\author{
E.D. Gennatas

E.D. Gennatas
}
