% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modError.R
\name{print.regError}
\alias{print.regError}
\title{Print \code{regError} object}
\usage{
\method{print}{regError}(x, ...)
}
\arguments{
\item{x}{\code{regError} object}
}
\description{
\code{print} \code{regError} object
}
\author{
E.D. Gennatas
}
