% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{print.resample}
\alias{print.resample}
\title{\code{print} method for \link{resample} object}
\usage{
\method{print}{resample}(x, ...)
}
\arguments{
\item{x}{\link{resample} object}

\item{...}{Not used}
}
\description{
Print resample information
}
\author{
E.D. Gennatas
}
