% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_variance.R
\name{print.rtBiasVariance}
\alias{print.rtBiasVariance}
\title{Print method for \link{bias_variance}}
\usage{
\method{print}{rtBiasVariance}(x, ...)
}
\arguments{
\item{x}{Output of \link{bias_variance}}

\item{...}{Not used}
}
\description{
Print method for \link{bias_variance}
}
\author{
E.D. Gennatas
}
