% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_TLS.R
\name{print.rtTLS}
\alias{print.rtTLS}
\title{\code{print.rtTLS}: \code{print} method for \code{rtTLS} object}
\usage{
\method{print}{rtTLS}(x, ...)
}
\arguments{
\item{x}{\code{rtTLS} object created by \link{s_TLS}}

\item{...}{Not used.}
}
\description{
\code{print.rtTLS}: \code{print} method for \code{rtTLS} object
}
