% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shyoptleaves.R
\name{print.shyoptleaves}
\alias{print.shyoptleaves}
\title{Print method for \code{shyoptleaves} object}
\usage{
\method{print}{shyoptleaves}(x, ...)
}
\arguments{
\item{x}{\code{shyoptleaves} object}
}
\description{
Print method for \code{shyoptleaves} object
}
\author{
Efstathios D. Gennatas
}
