% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shytreegamleaves.R
\name{print.shytreegamleaves}
\alias{print.shytreegamleaves}
\title{Print method for \code{shytreegamleaves} object}
\usage{
\method{print}{shytreegamleaves}(x, ...)
}
\arguments{
\item{x}{\code{shytreegamleaves} object}
}
\description{
Print method for \code{shytreegamleaves} object
}
\author{
Efstathios D. Gennatas
}
