% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shytreeLeaves.R
\name{print.shytreeLeavesRC}
\alias{print.shytreeLeavesRC}
\title{Print method for \code{shytreeLeavesRC} object}
\usage{
\method{print}{shytreeLeavesRC}(x, ...)
}
\arguments{
\item{x}{\code{shytreeLeavesRC} object}
}
\description{
Print method for \code{shytreeLeavesRC} object
}
\author{
Efstathios D. Gennatas
}
