% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survError.R
\name{print.surv_error}
\alias{print.surv_error}
\title{Print \link{surv_error}}
\usage{
\method{print}{surv_error}(x, decimal.places = 4, ...)
}
\arguments{
\item{x}{Object of type \link{surv_error}}

\item{decimal.places}{Integer: Number of decimal places to print. Default = 4}

\item{...}{Not used}
}
\description{
Print \link{surv_error}
}
\author{
E.D. Gennatas
}
