% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modError.R
\name{print.survError}
\alias{print.survError}
\title{Print \code{survError} object}
\usage{
\method{print}{survError}(x, ...)
}
\arguments{
\item{x}{\code{survError} object}
}
\description{
\code{print} \code{survError} object
}
\author{
E.D. Gennatas
}
