% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printls.R
\name{printdf1}
\alias{printdf1}
\title{Print 1 x N data frame}
\usage{
printdf1(x, pad = 2)
}
\arguments{
\item{x}{data frame}

\item{pad}{Integer: Pad output with this many spaces. Default = 2}
}
\description{
Pretty print a data frame containing 1 row of data with named columns as a vertical list
  of "       name : value"
     " other.name : other.value"
}
\author{
E.D. Gennatas
}
