% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.addtree.R
\name{prune.addtree}
\alias{prune.addtree}
\title{Prune AddTree tree}
\usage{
prune.addtree(
  addtree,
  prune.empty.leaves = TRUE,
  remove.bad.parents = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{addtree}{rtMod trained with \link{s_AddTree}}

\item{prune.empty.leaves}{Logical: If TRUE, remove leaves with 0 cases.}

\item{remove.bad.parents}{Logical: If TRUE, remove nodes with no siblings but children and
give their children to their parent.}

\item{verbose}{Logical: If TRUE, print messages to console.}
}
\description{
Prune an AddTree tree in Node format using \code{data.tree} to remove sister nodes with same
class estimate.
}
\author{
E.D. Gennatas
}
