% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.rpart.rt.R
\name{prune.rpart.rt}
\alias{prune.rpart.rt}
\title{\code{prune.rpart} experimental replacement}
\usage{
prune.rpart.rt(tree, cp, ...)
}
\description{
\code{prune.rpart} experimental replacement
}
\author{
E.D. Gennatas
}
