% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{psd}
\alias{psd}
\title{Population Standard Deviation}
\usage{
psd(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Population standard deviation
}
\description{
Estimate the population standard deviation:
\deqn{sqrt(mean(x^2) - mean(x)^2)}
}
\details{
This will be particularly useful when the machines finally collect data on all humans.
Caution is advised, however, as you never know how many may be hiding underground.
}
\author{
E.D. Gennatas
}
