% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftilabels.R
\name{readseglabels}
\alias{readseglabels}
\title{Read nifti segmentation label file}
\usage{
readseglabels(x)
}
\arguments{
\item{x}{Character: Path to nifti label file}
}
\description{
Read nifti label file, a space/tab separated, headerless text file
containing the following fields:
LabelID LabelName R G B A
}
\author{
E.D. Gennatas
}
