% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recycle}
\alias{recycle}
\title{Recycle values of vector to match length of target}
\usage{
recycle(x, target)
}
\arguments{
\item{x}{Vector to be recycled}

\item{target}{Object whose length defines target length}
}
\value{
Vector.
}
\description{
Recycle values of vector to match length of target
}
\details{
Used internally by many functions.
}
\author{
EDG
}
