% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{reg_error}
\alias{reg_error}
\title{Regression Error Metrics}
\usage{
reg_error(
  x,
  y,
  rho = FALSE,
  tau = FALSE,
  pct.red = FALSE,
  na.rm = FALSE,
  verbosity = 0
)
}
\arguments{
\item{x}{Numeric vector: True values}

\item{y}{Numeric vector: Predicted values}

\item{rho}{Logical: If TRUE, calculate Spearman's rho}

\item{tau}{Logical: If TRUE, calculate Kendall's tau}

\item{pct.red}{Logical: If TRUE, calculate percent reduction in error}

\item{na.rm}{Logical: If TRUE, remove NA values before computation}

\item{verbosity}{Integer: If > 0, print messages to console}
}
\value{
Object of class \code{regError}
}
\description{
Calculate error metrics for regression
}
\author{
E.D. Gennatas
}
