% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{regression_metrics}
\alias{regression_metrics}
\title{Regression Metrics}
\usage{
regression_metrics(true, predicted, na.rm = TRUE, sample = NULL)
}
\arguments{
\item{true}{Numeric vector: True values.}

\item{predicted}{Numeric vector: Predicted values.}

\item{na.rm}{Logical: If TRUE, remove NA values before computation.}

\item{sample}{Optional Character \{"Training", "Validation", "Test",
"Calibrated Training", "Calibrated Validation", "Calibrated Test"\}: Sample
the metrics describe.}
}
\value{
\code{RegressionMetrics} object.
}
\description{
Regression Metrics
}
\examples{
true <- rnorm(100)
predicted <- true + rnorm(100, sd = 0.5)
regression_metrics(true, predicted)
}
\author{
EDG
}
