% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resLearn_future.R
\name{resLearn_future}
\alias{resLearn_future}
\title{\pkg{rtemis} internal: Resample Learn}
\usage{
resLearn_future(
  x,
  y,
  mod,
  resample.rtset = rtset.cv.resample(),
  weights = NULL,
  params = list(),
  mtry = NULL,
  .preprocess = NULL,
  verbose = TRUE,
  res.verbose = FALSE,
  trace = 0,
  save.mods = TRUE,
  outdir = NULL,
  n.workers = 1,
  parallel.type = "nobodycares"
)
}
\arguments{
\item{x}{features - training set}

\item{y}{outcome - training set}

\item{mod}{Character: \pkg{rtemis} model. See \code{learnSelect} gives available models}

\item{resample.rtset}{List: output of \link{rtset} (or a list of same structure)}

\item{params}{List of named elements, each is a single value}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{res.verbose}{Logical: Will be passed to each \code{mod}'s \code{verbose} argument}

\item{save.mods}{Logical: If TRUE, save all models, otherwise discard after training.
Use with \link{train} when training a large number of resamples. Default = TRUE}

\item{outdir}{Character: Path to save output. Default = NULL}

\item{n.workers}{Integer: Number of cores to use.}
}
\description{
Train an \pkg{rtemis} learner on a set of resamples
}
\details{
Input: features (x) and outcome (y)
Procedure: \link{resample}, train learners
Output: trained learners
This is used internally by \link{train} and for bagging, when the \code{bag.resampler}
argument is set in a learner.
}
\author{
E.D. Gennatas
}
