% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseLevels.R
\name{reverseLevels}
\alias{reverseLevels}
\title{Reverse factor levels}
\usage{
reverseLevels(x)
}
\arguments{
\item{x}{Factor}
}
\description{
Reverse the order of a factor's levels
}
\author{
E.D. Gennatas
}
