% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revfactorlevels.R
\name{revfactorlevels}
\alias{revfactorlevels}
\title{Reverse factor level order}
\usage{
revfactorlevels(x)
}
\arguments{
\item{x}{factor}
}
\description{
Reverse factor level order
}
\author{
E.D. Gennatas
}
