% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge regression coefficients}
\usage{
ridge(x, y, lambda = 0, method = c("glmnet", "MASS", "R"))
}
\arguments{
\item{x}{Features}

\item{y}{Outcome}

\item{lambda}{Float: The lambda parameter}

\item{method}{String: "glmnet", "MASS", or "R": which method to use to estimate ridge coefficients.}
}
\description{
Ridge regression coefficients
}
\author{
Efstathios D. Gennatas
}
