% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rnormmat}
\alias{rnormmat}
\title{Random Normal Matrix}
\usage{
rnormmat(
  nrow = 10,
  ncol = 10,
  mean = 0,
  sd = 1,
  return_df = FALSE,
  seed = NULL
)
}
\arguments{
\item{nrow}{Integer: Number of rows.}

\item{ncol}{Integer: Number of columns.}

\item{mean}{Float: Mean.}

\item{sd}{Float: Standard deviation.}

\item{return_df}{Logical: If TRUE, return data.frame, otherwise matrix.}

\item{seed}{Integer: Set seed for \code{rnorm}.}
}
\description{
Create a matrix or data frame of defined dimensions, whose columns are random normal vectors
}
\author{
EDG
}
