% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{rowMax}
\alias{rowMax}
\title{Collapse data.frame to vector by getting row max}
\usage{
rowMax(x, na.rm = TRUE)
}
\description{
Collapse data.frame to vector by getting row max
}
\author{
Efstathios D. Gennatas
}
