% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd.R
\name{rsd}
\alias{rsd}
\title{Coefficient of Variation (Relative standard deviation)}
\usage{
rsd(x, as.percentage = TRUE, na.rm = TRUE, adjust = FALSE, adjust.lo = 1)
}
\arguments{
\item{x}{Numeric: Input}

\item{as.percentage}{Logical: If TRUE, multiply by 100}

\item{na.rm}{Logical: If TRUE, remove missing values before computation}

\item{adjust}{Logical: If TRUE, if \code{x} contains values < \code{adjust.lo}, x will be shifted up
by adding its minimum}

\item{adjust.lo}{Float: Threshold to be used if \code{adjust = TRUE}}
}
\description{
Calculates the coefficient of variation, also known as relative standard deviation, which is given by
\deqn{sd(x)/mean(x)}
}
\details{
This is not meaningful if mean is close to 0. For such cases, set \code{adjust = TRUE}.
This will add \code{min(x)} to x
}
\examples{
\dontrun{
mplot3_x(sapply(1:100, function(x) cov(rnorm(100))), 'd', xlab = 'rnorm(100) x 100 times')
# cov of rnorm without adjustment is all over the place
mplot3_x(sapply(1:100, function(x) cov(rnorm(100), adjust = T)), 'd',
xlab = 'rnorm(100) x 100 times')
# COV after shifting above 1 is what you probably want
}
}
