% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_theme_rtemis.R
\name{rstudio_theme_rtemis}
\alias{rstudio_theme_rtemis}
\title{Apply rtemis theme for RStudio}
\usage{
rstudio_theme_rtemis()
}
\description{
Apply the rtemis RStudio theme, a slight adaptation of the excellent rscodeio theme
(https://github.com/anthonynorth/rscodeio)
Recommended to use the Fira Code font with the theme
(https://fonts.google.com/specimen/Fira+Code?query=fira+code)
}
