% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot3.map.R
\name{rt_gtheme_map}
\alias{rt_gtheme_map}
\title{**ggplot2** map theme}
\usage{
rt_gtheme_map(
  base_size = 9,
  base_family = "Helvetica",
  legend.position = c(0.9, 0),
  plot.margin = ggplot2::unit(c(0.02, 0.1, 0.02, 0.02), "npc")
)
}
\arguments{
\item{base_size}{Float: Character \code{base_size.} Default = 9}

\item{base_family}{Character: Font family. Default = "Lato"}
}
\description{
**ggplot2** map theme for use with \link{gplot3.map}
based on \code{usmap:::theme_map()}
}
\author{
E.D. Gennatas
}
