% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_lle_calc_k.R
\name{rt_lle_calc_k}
\alias{rt_lle_calc_k}
\title{\pkg{rtemis} internal: \code{lle::calc_k} function adapted to work with \code{pbapply}}
\usage{
rt_lle_calc_k(X, m, kmin = 1, kmax = 20, plotres = TRUE,
  n.cores = rtCores, iLLE = FALSE)
}
\description{
\pkg{rtemis} internal: \code{lle::calc_k} function adapted to work with \code{pbapply}
}
\keyword{internal}
