% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_io.R
\name{rt_save}
\alias{rt_save}
\title{Write \pkg{rtemis} model to RDS file}
\usage{
rt_save(rtmod, outdir, file.prefix = "s_", verbose = TRUE)
}
\arguments{
\item{rtmod}{\pkg{rtemis} model}

\item{outdir}{Path to output directory}

\item{file.prefix}{Character: Prefix for filename}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Write \pkg{rtemis} model to RDS file
}
\author{
E.D. Gennatas
}
