% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtClust.R
\docType{class}
\name{rtClust-class}
\alias{rtClust-class}
\alias{rtClust}
\title{R6 class for \pkg{rtemis} clustering}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rtClust
}
\description{
\pkg{rtemis} clustering object
}
\section{Fields}{

\describe{
\item{\code{clust.name}}{Character: Name of clustering algorithm}

\item{\code{call}}{Originating call}

\item{\code{xnames}}{Column names of x}

\item{\code{clust}}{Clustering algorithm output}

\item{\code{clusters.train}}{Cluster assignment for training set}

\item{\code{clusters.test}}{Cluster assignment for testing set (if supported by algorithm and test set provided)}

\item{\code{extra}}{List: Algorithm-specific output}
}}

\author{
Efstathios D. Gennatas
}
\keyword{datasets}
