% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtClust.R
\name{rtClust-methods}
\alias{rtClust-methods}
\alias{print.rtClust}
\title{rtClust S3 methods}
\usage{
\method{print}{rtClust}(x, ...)
}
\arguments{
\item{x}{\code{rtClust} object}

\item{...}{Not used}
}
\description{
S3 methods for \code{rtClust} class.
}
