% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDecom.R
\docType{class}
\name{rtDecom-class}
\alias{rtDecom-class}
\alias{rtDecom}
\title{R6 Class for \pkg{rtemis} Decompositions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rtDecom
}
\description{
\pkg{rtemis} decomposition R6 object
}
\section{Fields}{

\describe{
\item{\code{decom.name}}{String: Name of decomposition algorithm}

\item{\code{call}}{Originating call}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{decom}}{Decomposition model output}

\item{\code{projections.train}}{Input data projected on new axes / basis}

\item{\code{projections.test}}{Input test data projected on new axes / basis}

\item{\code{extra}}{List: Algorithm-specific output}
}}

\author{
Efstathios D Gennatas
}
\keyword{datasets}
