% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDecom.R
\name{print.rtDecom}
\alias{print.rtDecom}
\title{\code{print.rtDecom}: \code{print} method for \code{rtDecom} object}
\usage{
\method{print}{rtDecom}(x, ...)
}
\arguments{
\item{x}{\code{rtDecom} object}

\item{...}{Not used}
}
\description{
\code{print.rtDecom}: \code{print} method for \code{rtDecom} object
}
