% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDecom.R
\docType{class}
\name{rtDecomLearn-class}
\alias{rtDecomLearn-class}
\alias{rtDecomLearn}
\title{R6 class for \pkg{rtemis} decomLearn}
\usage{
# rtDecom$new(decom.name = ...) # initialize new object
}
\description{
\pkg{rtemis} decomLearn R6 object
}
\author{
Efstathios D Gennatas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{decom.name}}{Character: Name of decomposition algorithm}

\item{\code{call}}{Originating call}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{decom}}{Decomposition model output}

\item{\code{projections.train}}{Input data projected on new axes / basis}

\item{\code{projections.test}}{Input test data projected on new axes / basis}

\item{\code{extra}}{List: Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{call}}{Originating call}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtDecomLearn$new()}}
\item \href{#method-clone}{\code{rtDecomLearn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtDecomLearn$new(
  mod.name = character(),
  decom.name = character(),
  decom.params = list(),
  decom.param.grid = list(),
  tuner.name = character(),
  tuner.params = list(),
  best.tune = list(),
  learner.name = character(),
  learner.params = list(),
  xnames = character(),
  decom = list(),
  mod = list(),
  projections.train = numeric(),
  projections.test = numeric(),
  extra = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtDecomLearn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
