% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis-package.R
\docType{package}
\name{rtemis-package}
\alias{rtemis}
\alias{rtemis-package}
\title{\pkg{rtemis}: Advanced Machine Learning and Visualization}
\description{
Schema-driven machine learning for validated, observable, auditable, and reproducible
machine learning workflows.
}
\section{Online Documentation and Vignettes}{

\url{https://docs.rtemis.org/r/ml}
}

\section{System Setup}{

There are some options you can define in your .Rprofile (usually found in your home directory),
so you do not have to define each time you execute a function.
\describe{
\item{rtemis_theme}{General plotting theme; set to e.g. "whiteigrid" or "darkgraygrid"}
\item{rtemis_font}{Font family to use in plots.}
\item{rtemis_palette}{Name of default palette to use in plots. See options by running \code{get_palette()}}
}
}

\section{Visualization}{

Graphics are handled using the \code{draw} family, which produces interactive plots primarily using
\code{plotly} and other packages.
}

\section{Supervised Learning}{

By convention, the last column of the data is the outcome variable, and all other columns are
predictors. Convenience function \link{set_outcome} can be used to move a specified column to the
end of the data.
Regression and Classification is performed using \code{train()}.
This function allows you to preprocess, train, tune, and test models on multiple resamples.
Use \link{available_supervised} to get a list of available algorithms
}

\section{Classification}{

For training of binary classification models, the outcome should be provided as a factor,
with the \emph{second} level of the factor being the 'positive' class.
}

\section{Clustering}{

Clustering is performed using \code{cluster()}.
Use \link{available_clustering} to get a list of available algorithms.
}

\section{Decomposition}{

Decomposition is performed using \code{decomp()}.
Use \link{available_decomposition} to get a list of available algorithms.
}

\section{Type Documentation}{

Function documentation includes input type (e.g. "Character", "Integer",
"Float"/"Numeric", etc).
When applicable, value ranges are provided in interval notation. For example, Float: [0, 1)
means floats between 0 and 1 including 0, but excluding 1.
Categorical variables may include set of allowed values using curly braces.
For example, Character: \{"future", "mirai", "none"\}.
}

\section{Tabular Data}{

\pkg{rtemis} internally uses methods for efficient handling of tabular data, with support for
\code{data.frame}, \code{data.table}, and \code{tibble}. If a function is documented as accepting
"tabular data", it should work with any of these data structures. If a function is documented
as accepting only one of these, then it should only be used with that structure.
For example, some optimized \code{data.table} operations that perform in-place modifications only
work with \code{data.table} objects.
}

\section{Error Conditions}{

Errors are signalled as classed conditions that can be handled with
\code{\link[=tryCatch]{tryCatch()}}. See \link{rtemis_conditions} for the class hierarchy.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.rtemis.org}
  \item \url{https://docs.rtemis.org/r/ml}
  \item \url{https://docs.rtemis.org/r/ml-api/}
  \item Report bugs at \url{https://github.com/rtemis-org/rtemis/issues}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID}) [copyright holder]
}

}
