% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rt.R
\name{rtemis_init}
\alias{rtemis_init}
\title{Initialize parallel processing and progress reporting}
\usage{
rtemis_init(n_cores = 1, context = NULL, verbosity = 1L)
}
\description{
Initialize parallel processing and progress reporting
}
\keyword{internal}
