% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rtemismlv-package}
\alias{rtemismlv}
\alias{rtemismlv-package}
\title{\pkg{rtemis}: Machine Learning and Visualization}
\description{
Advanced Machine Learning made easy, efficient, reproducible
}
\section{Online Documentation and Vignettes}{

\url{https://rtemis.org}
}

\section{System Setup}{

There are some options you can define in your .Rprofile (usually found in your home directory),
so you do not have to define each time you execute a function.
\describe{
\item{rt.theme}{General plotting theme; set to e.g. "whiteigrid" or "darkgraygrid"}
\item{rt.palette}{Name of default palette to use in plots. See options by running \code{rtpalette()}}
\item{rt.font}{Font family to use in plots.}
\item{rt.cores}{Number of cores to use. By default, rtemis will use available cores reported by
future::availableCores(). In shared systems, you should limit this as appropriate.}
\item{future.plan}{Default plan to use for parallel processing.}
}
}

\section{Visualization}{

Graphics are handled using the \code{draw} family, which is based on \code{plotly}.
Base graphics family \code{mplot3} is aavailable as a separate package.
}

\section{Supervised Learning}{

Regression and Classification is performed using \code{train()}.
This function allows you to preprocess, train, tune, and crossvalited models.
Run \link{available_supervised} to get a list of available algorithms
}

\section{Clustering}{

Clustering is performed using \code{cluster()}.
Run \link{available_clustering} to get a list of available algorithms.
}

\section{Decomposition}{

Decomposition is performed using \code{decompose()}.
Run \link{available_decomposition} to get a list of available algorithms.
}

\section{Notes}{

Function documentation includes input type (e.g. "String", "Integer",
"Float"/"Numeric", etc) and
range in interval notation where applicable. For example, Float: [0, 1)"
means floats between 0 and 1 including 0, but excluding 1

For all classification models, the outcome should be provided as a factor,
with the \emph{second} level of the factor being the 'positive' class.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.rtemis.org}
  \item Report bugs at \url{https://github.com/rtemis-org/rtemis/issues}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID})

}
