% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{rtenv}
\alias{rtenv}
\title{\pkg{rtemis} internal: rtemis environment}
\format{An object of class \code{environment} of length 1.}
\usage{
rtenv
}
\description{
\pkg{rtemis} internal: rtemis environment
}
\keyword{internal}
