% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtInitProjectDir.R
\name{rtInitProjectDir}
\alias{rtInitProjectDir}
\title{Initialize Project Directory}
\usage{
rtInitProjectDir(verbose = TRUE)
}
\description{
Initializes Directory Structure: "R", "Data", "Results"
}
\author{
Efstathios D Gennatas
}
