% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtlayout.R
\name{rtlayout}
\alias{rtlayout}
\title{Layout for plotting on multiple panels}
\usage{
rtlayout(nrows = NULL, ncols = NULL, byrow = FALSE,
  autolabel = FALSE, pdf.width = NULL, pdf.height = NULL,
  filename = NULL)
}
\arguments{
\item{nrows}{Integer: N of rows}

\item{ncols}{Integer: N of columns}

\item{byrow}{Logical: If TRUE, draw add plots by row Default = FALSE}

\item{pdf.width}{Float: Width of PDF to save, if \code{filename} is provided.
Default = \code{ncols * 4.5}
#' @param pdf.height Float: Height of PDF to save, if \code{filename} is provided.
Default = \code{nrows * 4.5}}

\item{filename}{String, optional: Save multiplot to file. Default = NULL}
}
\description{
Set layout for plotting multiple plots on the same view
}
\author{
Efstathios D. Gennatas
}
