% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_ops.R
\name{rtlgb2rules}
\alias{rtlgb2rules}
\title{Extract set of rules from LightGBM rtMod}
\usage{
rtlgb2rules(
  rtmod,
  dat,
  n_iter = NULL,
  right_cat_type = "in",
  return_unique = TRUE
)
}
\arguments{
\item{rtmod}{LightGBM rtMod object}

\item{dat}{Data frame used to train rtmod}

\item{n_iter}{Integer: Number of trees to convert to rules}
}
\value{
Character vector of rules
}
\description{
Extract set of rules from LightGBM rtMod
}
\keyword{internal}
