% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtMeta-class}
\alias{rtMeta-class}
\alias{rtMeta}
\title{\pkg{rtemis} Meta Model Class}
\description{
\pkg{rtemis} Meta Model Class

\pkg{rtemis} Meta Model Class
}
\details{
R6 Class for \pkg{rtemis} Meta Models
}
\section{Super class}{
\code{\link[rtemis:rtMod]{rtemis::rtMod}} -> \code{rtMeta}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{grid}}{Grid of algorithm names and resamples IDs}

\item{\code{base.resample.rtset}}{List of resampling parameters for base learner training}

\item{\code{base.mod.names}}{Character vector: names of base learner algorithms}

\item{\code{base.params}}{Named list of lists with base model parameters}

\item{\code{base.res.y.test}}{Base resamples' testing set outcomes}

\item{\code{base.res.predicted}}{Base resamples' predicted valuess}

\item{\code{base.mods}}{Base learners}

\item{\code{base.mods.error.train}}{Base learners' training error}

\item{\code{base.mods.error.test}}{Base learners' testing error}

\item{\code{meta.mod.name}}{Name of meta algorithm}

\item{\code{meta.params}}{List of meta model parameters}

\item{\code{meta.mod}}{Meta model}

\item{\code{sessionInfo}}{R session info at training time}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rtMeta-new}{\code{rtMeta$new()}}
\item \href{#method-rtMeta-print}{\code{rtMeta$print()}}
\item \href{#method-rtMeta-clone}{\code{rtMeta$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="describe"><a href='../../rtemis/html/rtMod.html#method-rtMod-describe'><code>rtemis::rtMod$describe()</code></a></li>
<li><span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plot"><a href='../../rtemis/html/rtMod.html#method-rtMod-plot'><code>rtemis::rtMod$plot()</code></a></li>
<li><span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFitted"><a href='../../rtemis/html/rtMod.html#method-rtMod-plotFitted'><code>rtemis::rtMod$plotFitted()</code></a></li>
<li><span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFittedPredicted"><a href='../../rtemis/html/rtMod.html#method-rtMod-plotFittedPredicted'><code>rtemis::rtMod$plotFittedPredicted()</code></a></li>
<li><span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotPredicted"><a href='../../rtemis/html/rtMod.html#method-rtMod-plotPredicted'><code>rtemis::rtMod$plotPredicted()</code></a></li>
<li><span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotVarImp"><a href='../../rtemis/html/rtMod.html#method-rtMod-plotVarImp'><code>rtemis::rtMod$plotVarImp()</code></a></li>
<li><span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="summary"><a href='../../rtemis/html/rtMod.html#method-rtMod-summary'><code>rtemis::rtMod$summary()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rtMeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-rtMeta-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{rtMeta} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMeta$new(
  mod.name = character(),
  y.train = numeric(),
  y.test = numeric(),
  x.name = character(),
  y.name = character(),
  xnames = character(),
  grid = data.frame(),
  base.resample.rtset = list(),
  base.mod.names = character(),
  base.params = list(),
  base.res.y.test = numeric(),
  base.res.predicted = numeric(),
  base.mods = list(),
  base.mods.error.train = list(),
  base.mods.error.test = list(),
  meta.mod.name = character(),
  meta.params = list(),
  meta.mod = list(),
  type = character(),
  fitted = numeric(),
  se.fit = numeric(),
  error.train = list(),
  predicted = numeric(),
  se.prediction = numeric(),
  error.test = list(),
  question = character(),
  extra = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod.name}}{Character: Algorithm name}

\item{\code{y.train}}{Training set output}

\item{\code{y.test}}{Testing set output}

\item{\code{x.name}}{Character: Feature set name}

\item{\code{y.name}}{Character: Output name}

\item{\code{xnames}}{Character vector: Feature names}

\item{\code{grid}}{Data frame with combinations of algorithm
names and resample IDs}

\item{\code{base.resample.rtset}}{List with resampling 
parameters for base learners}

\item{\code{base.mod.names}}{Base learner algorithm names}

\item{\code{base.params}}{Named list of lists with base model parameters}

\item{\code{base.res.y.test}}{Base learners' predicted values}

\item{\code{base.res.predicted}}{Base resamples' predicted valuess}

\item{\code{base.mods}}{Base learners}

\item{\code{base.mods.error.train}}{Base learners' training error}

\item{\code{base.mods.error.test}}{Base learners' testing error}

\item{\code{meta.mod.name}}{Meta model algorithm name}

\item{\code{meta.params}}{Meta model hyperparameters}

\item{\code{meta.mod}}{Trained meta model}

\item{\code{type}}{Character: Type of model (Regression, Classification, Survival)}

\item{\code{fitted}}{Fitted values (training set predictions)}

\item{\code{se.fit}}{Standard error of fitted values}

\item{\code{error.train}}{Training set error}

\item{\code{predicted}}{Predicted values (Testing set predictions)}

\item{\code{se.prediction}}{Standard error of predicted values}

\item{\code{error.test}}{Testing set error}

\item{\code{question}}{Question the model is trying to answer}

\item{\code{extra}}{List of extra model info}

\item{\code{base.res.mods}}{Base learner fitted models}

\item{\code{sessionInfo}}{R session info at time of training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rtMeta-print"></a>}}
\if{latex}{\out{\hypertarget{method-rtMeta-print}{}}}
\subsection{Method \code{print()}}{
Plot method for \code{rtMeta} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMeta$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rtMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rtMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
