% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtMeta-methods}
\alias{rtMeta-methods}
\alias{predict.rtMeta}
\title{rtMeta S3 methods}
\usage{
\method{predict}{rtMeta}(object, newdata, fn = median, ...)
}
\arguments{
\item{newdata}{Testing set features}
}
\description{
S3 methods for \code{rtMeta} class that differ from those of the \code{rtMod} superclass
}
