% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtMod-class}
\alias{rtMod-class}
\alias{rtMod}
\title{\pkg{rtemis} Supervised Model Class}
\description{
\pkg{rtemis} Supervised Model Class

\pkg{rtemis} Supervised Model Class
}
\details{
R6 Class for \pkg{rtemis} Supervised Models

Note on terminology:
The training set ('x.train') is used to build a model and calculate training error
(rtMod$error.train). The testing set ('x.test') is not seen during training and
only used to create predictions (rtMod$predicted) using the trained model and
calculate error metrics (rtMod$error.test).
This reflects generalizability of the model and is the error we care the most about.
The validation set is used during model tuning. Within rtemis, validation sets are
created automatically by \link{resample} when appropriate, they are not generally
input by the user, with few exceptions, as documented in individual functions.
}
\author{
E.D. Gennatas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mod.name}}{Learner algorithm name}

\item{\code{y.train}}{Training y data}

\item{\code{y.test}}{Testing y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{mod}}{Trained model}

\item{\code{type}}{Classification, Regression, or Survival}

\item{\code{gridsearch}}{Grid search output}

\item{\code{parameters}}{List of hyperparameters used when building model}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{varimp}}{Variable Importance. Note that this is a general concept and different ML algorithms provide very
different approaches}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{extra}}{Algorithm-specific output}

\item{\code{sessionInfo}}{The output of \code{sessionInfo()} at the time the model was trained}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtMod$new()}}
\item \href{#method-print}{\code{rtMod$print()}}
\item \href{#method-plot}{\code{rtMod$plot()}}
\item \href{#method-plotFitted}{\code{rtMod$plotFitted()}}
\item \href{#method-plotPredicted}{\code{rtMod$plotPredicted()}}
\item \href{#method-plotFittedPredicted}{\code{rtMod$plotFittedPredicted()}}
\item \href{#method-plotVarImp}{\code{rtMod$plotVarImp()}}
\item \href{#method-summary}{\code{rtMod$summary()}}
\item \href{#method-describe}{\code{rtMod$describe()}}
\item \href{#method-clone}{\code{rtMod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{rtMod} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$new(
  mod.name = character(),
  y.train = numeric(),
  y.test = numeric(),
  x.name = character(),
  y.name = character(),
  xnames = character(),
  mod = list(),
  type = character(),
  gridsearch = list(),
  parameters = list(),
  fitted = numeric(),
  se.fit = NULL,
  error.train = list(),
  predicted = NULL,
  se.prediction = NULL,
  error.test = NULL,
  varimp = NULL,
  question = character(),
  extra = list(),
  sessionInfo = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod.name}}{Character: Algorithm name}

\item{\code{y.train}}{Training set output}

\item{\code{y.test}}{Testing set output}

\item{\code{x.name}}{Character: Feature set name}

\item{\code{y.name}}{Character: Output name}

\item{\code{xnames}}{Character vector: Feature names}

\item{\code{mod}}{Trained model}

\item{\code{type}}{Character: Type of model (Regression, Classification, Survival)}

\item{\code{gridsearch}}{Grid search output}

\item{\code{parameters}}{List of training parameters}

\item{\code{fitted}}{Fitted values (training set predictions)}

\item{\code{se.fit}}{Standard error of fitted values}

\item{\code{error.train}}{Training set error}

\item{\code{predicted}}{Predicted values (Testing set predictions)}

\item{\code{se.prediction}}{Standard error of predicted values}

\item{\code{error.test}}{Testing set error}

\item{\code{varimp}}{Variable importance}

\item{\code{question}}{Question the model is trying to answer}

\item{\code{extra}}{List of extra model info}

\item{\code{sessionInfo}}{R session info at time of training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
code{print} method for \code{rtMod} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\code{rtMod} plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plot(
  estimate = NULL,
  theme = getOption("rt.theme"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate}}{Character: "fitted" or "predicted"}

\item{\code{theme}}{Theme to pass to plotting function}

\item{\code{filename}}{Character: Path to file to save plot}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotFitted}{}}}
\subsection{Method \code{plotFitted()}}{
Plot fitted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotFitted(
  print.plot = TRUE,
  theme = getOption("rt.theme"),
  main = NULL,
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print.plot}}{Logical: If TRUE show plot}

\item{\code{theme}}{Theme to be passed on to plotting function}

\item{\code{main}}{Character: main title}

\item{\code{filename}}{Character: path to file to save plot}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPredicted}{}}}
\subsection{Method \code{plotPredicted()}}{
Plot predicted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotPredicted(
  print.plot = TRUE,
  theme = getOption("rt.theme"),
  main = NULL,
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print.plot}}{Logical: If TRUE show plot}

\item{\code{theme}}{Theme to be passed on to plotting function}

\item{\code{main}}{Character: main title}

\item{\code{filename}}{Character: path to file to save plot}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFittedPredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotFittedPredicted}{}}}
\subsection{Method \code{plotFittedPredicted()}}{
Plot fitted and predicted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotFittedPredicted(
  print.plot = TRUE,
  theme = getOption("rt.theme"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print.plot}}{Logical: If TRUE show plot}

\item{\code{theme}}{Theme to be passed on to plotting function}

\item{\code{filename}}{Character: path to file to save plot}

\item{\code{...}}{Additional arguments passed to \code{mplot3_conf}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotVarImp"></a>}}
\if{latex}{\out{\hypertarget{method-plotVarImp}{}}}
\subsection{Method \code{plotVarImp()}}{
Plot variable importance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotVarImp(
  plot.top = 12,
  type = c("barplot", "lollipop"),
  theme = getOption("rt.theme"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot.top}}{Integer: Plot this many top features}

\item{\code{type}}{Character: "barplot" or "lollipop"}

\item{\code{theme}}{Theme to be passed on to plotting function}

\item{\code{...}}{Not used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method for \code{rtMod} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$summary(
  plots = TRUE,
  cex = 1,
  fit.true.line = "lm",
  resid.fit.line = "gam",
  fit.legend = TRUE,
  se.fit = TRUE,
  single.fig = TRUE,
  theme = getOption("rt.theme"),
  title.col = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plots}}{Logical: If TRUE show plots}

\item{\code{cex}}{Numeric: Character expansion factor}

\item{\code{fit.true.line}}{Character: algorithm to use to fit
true vs predicted curve}

\item{\code{resid.fit.line}}{Character: algorithm to use to
fit residuals plot}

\item{\code{fit.legend}}{Logical: If TRUE, show legend in
fit plot}

\item{\code{se.fit}}{Logical: If TRUE, include standard error
band in fit plot}

\item{\code{single.fig}}{Logical: If TRUE, combine in 
single plot}

\item{\code{theme}}{Theme to pass to plotting functions}

\item{\code{title.col}}{Title color}

\item{\code{...}}{Extra argument to pass to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe"></a>}}
\if{latex}{\out{\hypertarget{method-describe}{}}}
\subsection{Method \code{describe()}}{
Describe model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
