% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtMod-methods}
\alias{rtMod-methods}
\alias{print.rtMod}
\alias{fitted.rtMod}
\alias{predict.rtMod}
\alias{residuals.rtMod}
\alias{plot.rtMod}
\alias{summary.rtMod}
\alias{coef.rtMod}
\alias{predict.rtModLite}
\title{\code{rtMod} S3 methods}
\usage{
\method{print}{rtMod}(x, ...)

\method{fitted}{rtMod}(object, ...)

\method{predict}{rtMod}(object, newdata, trace = 0, verbose = TRUE, ...)

\method{residuals}{rtMod}(object, ...)

\method{plot}{rtMod}(x, estimate = NULL, theme = getOption("rt.theme"), filename = NULL, ...)

\method{summary}{rtMod}(
  object,
  plots = TRUE,
  cex = 1,
  fit.true.line = "lm",
  resid.fit.line = "gam",
  fit.legend = TRUE,
  se.fit = TRUE,
  single.fig = TRUE,
  summary = TRUE,
  theme = getOption("rt.theme"),
  title.col = NULL,
  ...
)

\method{coef}{rtMod}(object, verbose = TRUE, ...)

\method{predict}{rtModLite}(object, newdata, ...)
}
\arguments{
\item{x}{\code{rtMod} object}

\item{...}{Additional arguments to be passed to \link{mplot3_xy}}

\item{object}{Trained model of class \code{rtMod}}

\item{newdata}{Testing set features}

\item{trace}{Integer: Set trace level}

\item{verbose}{Logical: If TRUE, output messages to console}

\item{estimate}{Character: "fitted" or "predicted"}

\item{theme}{Character: theme to use. Options: "box", "darkbox", "light", "dark"}

\item{filename}{Character: Path to file to save plot}

\item{plots}{Logical: If TRUE, print plots. Default = TRUE}

\item{cex}{Float: Character expansion factor}

\item{fit.true.line}{\pkg{rtemis} algorithm to use for fitted vs. true line
Options: \code{modSelect()}}

\item{resid.fit.line}{\pkg{rtemis} algorithm to use for residuals vs. fitted line.
Options: \code{modSelect()}}

\item{fit.legend}{Logical: If TRUE, print fit legend. Default  = TRUE}

\item{se.fit}{Logical: If TRUE, plot 2 * standard error bands. Default = TRUE}

\item{single.fig}{Logical: If TRUE, draw all plots in a single figure. Default = TRUE}

\item{summary}{Logical: If TRUE, print summary. Default = TRUE}

\item{title.col}{Color for main title}
}
\description{
S3 methods for \code{rtMod} class.
Excludes functions \code{print} and \code{plot} defined within the \link{rtMod} class itself.
}
\author{
E.D. Gennatas

E.D. Gennatas
}
