% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtMod.out}
\alias{rtMod.out}
\title{rtMod.out}
\usage{
rtMod.out(
  rt,
  print.plot,
  plot.fitted,
  plot.predicted,
  y.test,
  mod.name,
  outdir,
  save.mod,
  verbose,
  theme,
  ...
)
}
\description{
rtMod.out
}
\keyword{internal}
