% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtModBag-methods}
\alias{rtModBag-methods}
\alias{predict.rtModBag}
\title{rtModBag S3 methods}
\usage{
\method{predict}{rtModBag}(object, newdata, aggr.fn = NULL,
  n.cores = 1, verbose = FALSE, ...)
}
\arguments{
\item{newdata}{Testing set features}

\item{aggr.fn}{Character: Function to aggregate models' prediction. Default = "median"}

\item{...}{Not used}
}
\description{
S3 methods for \code{rtModBag} class that differ from those of the \code{rtMod} superclass
}
