% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtModClass-class}
\alias{rtModClass-class}
\alias{rtModClass}
\title{\pkg{rtemis} Classification Model Class}
\description{
\pkg{rtemis} Classification Model Class

\pkg{rtemis} Classification Model Class
}
\details{
R6 Class for \pkg{rtemis} Classification Models
}
\section{Super class}{
\code{\link[rtemis:rtMod]{rtemis::rtMod}} -> \code{rtModClass}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fitted.prob}}{Training set probability estimates}

\item{\code{predicted.prob}}{Testing set probability estimates}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModClass$new()}}
\item \href{#method-plotROC}{\code{rtModClass$plotROC()}}
\item \href{#method-plotROCfitted}{\code{rtModClass$plotROCfitted()}}
\item \href{#method-plotROCpredicted}{\code{rtModClass$plotROCpredicted()}}
\item \href{#method-plotPR}{\code{rtModClass$plotPR()}}
\item \href{#method-plotPRfitted}{\code{rtModClass$plotPRfitted()}}
\item \href{#method-plotPRpredicted}{\code{rtModClass$plotPRpredicted()}}
\item \href{#method-clone}{\code{rtModClass$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="describe">}\href{../../rtemis/html/rtMod.html#method-describe}{\code{rtemis::rtMod$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plot">}\href{../../rtemis/html/rtMod.html#method-plot}{\code{rtemis::rtMod$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFitted">}\href{../../rtemis/html/rtMod.html#method-plotFitted}{\code{rtemis::rtMod$plotFitted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFittedPredicted">}\href{../../rtemis/html/rtMod.html#method-plotFittedPredicted}{\code{rtemis::rtMod$plotFittedPredicted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotPredicted">}\href{../../rtemis/html/rtMod.html#method-plotPredicted}{\code{rtemis::rtMod$plotPredicted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotVarImp">}\href{../../rtemis/html/rtMod.html#method-plotVarImp}{\code{rtemis::rtMod$plotVarImp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="print">}\href{../../rtemis/html/rtMod.html#method-print}{\code{rtemis::rtMod$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="summary">}\href{../../rtemis/html/rtMod.html#method-summary}{\code{rtemis::rtMod$summary()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{rtModClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$new(
  mod.name = character(),
  y.train = numeric(),
  y.test = numeric(),
  x.name = character(),
  y.name = character(),
  xnames = character(),
  mod = list(),
  type = character(),
  gridsearch = list(),
  parameters = list(),
  fitted = numeric(),
  fitted.prob = numeric(),
  se.fit = numeric(),
  error.train = list(),
  predicted = NULL,
  predicted.prob = NULL,
  se.prediction = NULL,
  error.test = NULL,
  varimp = NULL,
  question = character(),
  extra = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod.name}}{Character: Algorithm name}

\item{\code{y.train}}{Training set output}

\item{\code{y.test}}{Testing set output}

\item{\code{x.name}}{Character: Feature set name}

\item{\code{y.name}}{Character: Output name}

\item{\code{xnames}}{Character vector: Feature names}

\item{\code{mod}}{Trained model}

\item{\code{type}}{Character: Type of model (Regression, Classification, Survival)}

\item{\code{gridsearch}}{Grid search output}

\item{\code{parameters}}{List of training parameters}

\item{\code{fitted}}{Fitted values (training set predictions)}

\item{\code{fitted.prob}}{Training set probability estimates}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training set error}

\item{\code{predicted}}{Predicted values (Testing set predictions)}

\item{\code{predicted.prob}}{Testing set probability estimates}

\item{\code{se.prediction}}{}

\item{\code{error.test}}{Testing set error}

\item{\code{varimp}}{Variable importance}

\item{\code{question}}{Question the model is trying to answer}

\item{\code{extra}}{List of extra model info}

\item{\code{sessionInfo}}{R session info at time of training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROC"></a>}}
\if{latex}{\out{\hypertarget{method-plotROC}{}}}
\subsection{Method \code{plotROC()}}{
plot ROC. Uses testing set if available,
otherwise training
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotROC(theme = rtTheme, filename = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{Theme to pass to plotting function}

\item{\code{filename}}{Character: Path to file to save plot}

\item{\code{...}}{Extra arguments to pass to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROCfitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotROCfitted}{}}}
\subsection{Method \code{plotROCfitted()}}{
Plot training set ROC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotROCfitted(
  main = "ROC Training",
  theme = rtTheme,
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{main}}{Character: Main title}

\item{\code{theme}}{Theme to pass to plotting function}

\item{\code{filename}}{Character: Path to file to save plot}

\item{\code{...}}{Extra arguments to pass to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROCpredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotROCpredicted}{}}}
\subsection{Method \code{plotROCpredicted()}}{
plot testing set ROC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotROCpredicted(
  main = "ROC Testing",
  theme = rtTheme,
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{main}}{Character: Main title}

\item{\code{theme}}{Theme to pass to plotting function}

\item{\code{filename}}{Character: Path to file to save plot}

\item{\code{...}}{Extra arguments to pass to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPR"></a>}}
\if{latex}{\out{\hypertarget{method-plotPR}{}}}
\subsection{Method \code{plotPR()}}{
plot Precision-Recall curve. Uses testing set 
if available, otherwise training
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotPR(theme = rtTheme, filename = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{Theme to pass to plotting function}

\item{\code{filename}}{Character: Path to file to save plot}

\item{\code{...}}{Extra arguments to pass to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPRfitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPRfitted}{}}}
\subsection{Method \code{plotPRfitted()}}{
Plot training set Precision-Recall curve.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotPRfitted(
  main = "P-R Training",
  theme = rtTheme,
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{main}}{Character: Main title}

\item{\code{theme}}{Theme to pass to plotting function}

\item{\code{filename}}{Character: Path to file to save plot}

\item{\code{...}}{Extra arguments to pass to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPRpredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPRpredicted}{}}}
\subsection{Method \code{plotPRpredicted()}}{
plot testing set Precision-Recall curve.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotPRpredicted(
  main = "P-R Testing",
  theme = rtTheme,
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{main}}{Character: Main title}

\item{\code{theme}}{Theme to pass to plotting function}

\item{\code{filename}}{Character: Path to file to save plot}

\item{\code{...}}{Extra arguments to pass to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
