% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{data}
\name{rtModClass}
\alias{rtModClass}
\title{\pkg{rtemis} Classification Model Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
rtModClass
}
\description{
R6 Class for \pkg{rtemis} Classification Models
}
\keyword{datasets}
