% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtModCVclass-class}
\alias{rtModCVclass-class}
\alias{rtModCVclass}
\title{\pkg{rtemis} Cross-Validated Classification Model Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rtModCVclass
}
\description{
R6 Class for \pkg{rtemis} Cross-Validated Classification Models
}
\keyword{datasets}
