% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtModLite-methods}
\alias{rtModLite-methods}
\alias{print.rtModLite}
\title{rtModLite S3 methods}
\usage{
\method{print}{rtModLite}(x, ...)
}
\arguments{
\item{x}{\code{rtModLite} object}

\item{...}{Not used}
}
\description{
S3 methods for \code{rtModLite} class.
}
