% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModLog.R
\docType{class}
\name{rtModLogger-class}
\alias{rtModLogger-class}
\alias{rtModLogger}
\title{\pkg{rtemis} model logger}
\description{
\pkg{rtemis} model logger

\pkg{rtemis} model logger
}
\details{
R6 class to save trained models' parameters and performance. 
Keep your experiment results tidy in one place, with an option to write out 
to a multi-sheet Excel file.
}
\author{
E.D. Gennatas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mods}}{List of trained models}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModLogger$new()}}
\item \href{#method-print}{\code{rtModLogger$print()}}
\item \href{#method-add}{\code{rtModLogger$add()}}
\item \href{#method-summarize}{\code{rtModLogger$summarize()}}
\item \href{#method-summary}{\code{rtModLogger$summary()}}
\item \href{#method-tabulate}{\code{rtModLogger$tabulate()}}
\item \href{#method-plot}{\code{rtModLogger$plot()}}
\item \href{#method-clone}{\code{rtModLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{rtModLogger} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$new(mods = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mods}}{List of trained models}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{rtModLogger} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Add model to logger
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$add(mod, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{Model to add}

\item{\code{verbose}}{Logical: If TRUE, print messages to console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Summary method for \code{rtModLogger}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$summarize(
  class.metric = "Balanced Accuracy",
  reg.metric = "Rsq",
  surv.metric = "Coherence",
  decimal.places = 3,
  print.metric = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class.metric}}{Character: Metric to use for 
Classification models}

\item{\code{reg.metric}}{Character: Metric to use for 
Regression models}

\item{\code{surv.metric}}{Character: Metric to use for
Survival models}

\item{\code{decimal.places}}{Integer: Number of decimal
places to display}

\item{\code{print.metric}}{Logical: If TRUE, print metric name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method for \code{rtModLogger}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$summary(
  class.metric = "Balanced Accuracy",
  reg.metric = "Rsq",
  surv.metric = "Coherence"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{class.metric}}{Character: Metric to use for
Classification models}

\item{\code{reg.metric}}{Character: Metric to use for
Regression models}

\item{\code{surv.metric}}{Character: Metric to use for
Survival models}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate}{}}}
\subsection{Method \code{tabulate()}}{
Tabulate models' parameters and performance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$tabulate(filename = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Character: Path to file to save
parameters and performance - will be saved as .xlsx file
with multiple sheets}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method for \code{rtModLogger}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$plot(
  names = NULL,
  col = unlist(getOption("rt.palette", "rtCol1")),
  mar = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Character: Model names}

\item{\code{col}}{Colors to use}

\item{\code{mar}}{Float, vector: plot margins}

\item{\code{...}}{Additional arguments to pass to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
