% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModLog.R
\docType{data}
\name{rtModLogger}
\alias{rtModLogger}
\title{\pkg{rtemis} model logger}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
rtModLogger
}
\description{
R6 class to save trained models' parameters and performance. Keep your experiment results tidy in
one place, with an option to write out to a multi-sheet Excel file.
}
\author{
Efstathios D. Gennatas
}
\keyword{datasets}
