% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModSet.R
\name{rtModSet}
\alias{rtModSet}
\title{\pkg{rtemis} model}
\usage{
rtModSet(rtclass = "rtMod", mod = list(), mod.name = character(),
  type = character(), parameters = list(), call = "",
  y.train = numeric(), y.test = numeric(), x.name = character(),
  y.name = character(), xnames = character(),
  bag.resample.rtset = list(), fitted.bag = numeric(),
  fitted = numeric(), fitted.prob = numeric(),
  se.fit.bag = numeric(), se.fit = numeric(), error.train = list(),
  predicted.bag = numeric(), predicted = numeric(),
  predicted.prob = numeric(), se.prediction.bag = numeric(),
  se.prediction = numeric(), error.test = list(), varimp = numeric(),
  question = character(), extra = list())
}
\arguments{
\item{rtclass}{Character: Which class system to use. Previously "S3", "S4", "RC", "R6".
Currently, only "R6", aka "rtMod" is supported.}

\item{mod}{Trained model}

\item{mod.name}{Model (algorithm) name}

\item{type}{"Regression" or "Classification"}

\item{parameters}{List of model hyperparameters}

\item{call}{Originating call}

\item{y.train}{Training y data}

\item{y.test}{Validation y data}

\item{x.name}{Name of x data}

\item{y.name}{Name of y data}

\item{xnames}{Column names of x}

\item{fitted}{Fitted values}

\item{fitted.prob}{Fitted probabilities (Classification only)}

\item{se.fit}{Standard error of the fit}

\item{error.train}{Training error}

\item{predicted}{Predicted values}

\item{predicted.prob}{Predicted probabilities (Classification only)}

\item{se.prediction}{Standard error of the prediction}

\item{error.test}{Validation error}

\item{varimp}{Variable importance}

\item{question}{Question the model is attempting to answer}

\item{extra}{List with algorithm-specific extra fields}
}
\description{
Creates an \pkg{rtemis} object of your choice of class. All available class systems are supported,
but at this point \code{R6} is the default and most widely tested.
}
\author{
Efstathios D. Gennatas
}
