% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\name{rtpalette}
\alias{rtpalette}
\title{Color Palettes}
\usage{
rtpalette(palette = NULL, verbosity = 1L)
}
\arguments{
\item{palette}{Character: Name of palette to return. Default = NULL: available palette
names are printed and no palette is returned.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
A list of available palettes, invisibly.
}
\description{
\code{rtpalette()} prints names of available color palettes
Each palette is a named list of hexadecimal color definitions which can be used with
any graphics function.
\code{rtpalette(palette_name)} returns a list of colors for a given palette.
}
\examples{
\dontrun{
rtpalette("imperial")
rtpalette()
}
}
\author{
EDG
}
