% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtPalette.R
\docType{data}
\name{rtPalettes}
\alias{rtPalettes}
\alias{pennCol}
\alias{pennPalette}
\alias{pennLightPalette}
\alias{imperialCol}
\alias{ucsfCol}
\alias{ucsfPalette}
\alias{berkeleyCol}
\alias{stanfordCol}
\alias{usfCol}
\alias{ucsdCol}
\alias{uclaCol}
\alias{ucCol}
\alias{uwCol}
\alias{nihCol}
\alias{appleCol}
\alias{googleCol}
\alias{amazonCol}
\alias{microsoftCol}
\title{rtemis Color Palettes}
\format{An object of class \code{list} of length 30.}
\usage{
pennCol

pennPalette

pennLightPalette

imperialCol

ucsfCol

ucsfPalette

berkeleyCol

stanfordCol

usfCol

ucsdCol

uclaCol

ucCol

uwCol

nihCol

appleCol

googleCol

amazonCol

microsoftCol
}
\description{
\code{pennCol}: Penn color palette (http://www.upenn.edu/about/styleguide-color-type)

\code{imperialCol}: Imperial College London color palette (https://www.imperial.ac.uk/brand-style-guide/visual-identity/brand-colours/)

\code{ucsfCol}: UCSF color palette (http://identity.ucsf.edu/color)

\code{ucsfPalette}: Subset of \code{ucsfCol} for use with \link{mplot3}, etc

\code{berkeleyCol}: Berkeley color palette (https://brand.berkeley.edu/colors/)

\code{stanfordCol}: Stanford color palette (https://identity.stanford.edu/color.html#digital-color)

\code{usfCol}: USF color palette (https://myusf.usfca.edu/marketing-communications/resources/graphics-resources/brand-standards/color-palette)
Color conversions performed using https://www.pantone.com/color-finder/

\code{ucsdCol}: UC San Diego color palette (https://ucpa.ucsd.edu/brand/elements/color-palette/)

\code{uclaCol}: UCLA color palette (http://brand.ucla.edu/identity/colors)

\code{ucCol}: University of California color palette
(http://brand.universityofcalifornia.edu/guidelines/color.html#!primary-colors)

\code{uwCol}: University of Washington color palette
(http://www.washington.edu/brand/graphic-elements/primary-color-palette/)

\code{nihCol}: NIH color palette (https://www.nlm.nih.gov/about/nlm_logo_guidelines_030414_508.pdf)

\code{appleCol}: Apple Human Interface Guidelines color palette
(https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/color/)

\code{googleCol}: Google brand palette (https://brandpalettes.com/google-colors/)

\code{amazonCol}: Amazon brand palette
(https://images-na.ssl-images-amazon.com/images/G/01/AdvertisingSite/pdfs/AmazonBrandUsageGuidelines.pdf)

\code{microsoftCol}: Microsoft brand palette
(https://brandcolors.net/b/microsoft)
}
\keyword{datasets}
