% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtROC.R
\name{rtROC}
\alias{rtROC}
\title{Build an ROC curve}
\usage{
rtROC(
  true.labels,
  predicted.probabilities,
  thresholds = NULL,
  plot = FALSE,
  theme = getOption("rt.theme"),
  verbose = TRUE
)
}
\description{
Calculate the points of an ROC curve and the AUC
}
\details{
\code{true.labels} should be a factor (will be coerced to one) where the first level is the
"positive" case. \code{predicted.probabilities} should be a vector of floats {0, 1} where \code{[0, .5)}
corresponds to the first level and \code{[.5, 1]} corresponds to the second level.
predicted.probabilities
}
\author{
E.D. Gennatas
}
