% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtSave.R
\name{rtSave}
\alias{rtSave}
\title{Write \pkg{rtemis} model to RDS file}
\usage{
rtSave(rtmod, outdir, file.prefix = "s.", verbose = TRUE)
}
\arguments{
\item{rtmod}{\pkg{rtemis} model}

\item{outdir}{Path to output directory}

\item{file.prefix}{Character: Prefix for filename}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Write \pkg{rtemis} model to RDS file
}
\author{
Efstathios D. Gennatas
}
