% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.ADDT}
\alias{rtset.ADDT}
\title{\code{rtset.ADDT}: Set parameters for \link{s.ADDT}}
\usage{
rtset.ADDT(max.depth = 2, learning.rate = 1, lin.type = "glmnet",
  alpha = 0, lambda = 0.1, minobsinnode = 2, minobsinnode.lin = 20,
  ...)
}
\arguments{
\item{max.depth}{Integer: Max depth of additive tree}

\item{lambda}{Float: lambda parameter for \code{MASS::lm.ridge} Default = .01}

\item{minobsinnode}{Integer: Minimum N observations needed in node, before considering splitting}
}
\description{
\code{rtset.ADDT}: Set parameters for \link{s.ADDT}
}
