% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.cluster}
\alias{rtset.cluster}
\title{\code{rtset.cluster}: Set up parallel processing using forking (Linux, macOS) or PSOCK cluster (macOS, Linux, Windows)
Defaults to type = "fork", which would call \code{mclapply}
Some functions fail to work correctly with \code{mclapply}, like \code{nnet::multinom}, in those cases
we use PSOCK cluster}
\usage{
rtset.cluster(type = "fork", hosts = NULL, n.cores = rtCores, ...)
}
\arguments{
\item{type}{String: "fork", "psock"}

\item{hosts}{Vector of strings: For type = "psock": Host names on which to run (macOS, Linux, Windows)}

\item{n.cores}{Integer: Number of cores to use on \code{localhost} for type = "fork" (macOS, Linux only)}

\item{...}{\code{rtset.cluster}: Additional argument to be passed to \code{parallel::makePSOCKcluster}}
}
\value{
List with parameters
}
\description{
\code{rtset.cluster}: Set up parallel processing using forking (Linux, macOS) or PSOCK cluster (macOS, Linux, Windows)
Defaults to type = "fork", which would call \code{mclapply}
Some functions fail to work correctly with \code{mclapply}, like \code{nnet::multinom}, in those cases
we use PSOCK cluster
}
