% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.cv.resample}
\alias{rtset.cv.resample}
\title{\code{rtset.cv.resample}: \link{resample} defaults for cross-validation}
\usage{
rtset.cv.resample(
  resampler = "kfold",
  n.resamples = 10,
  stratify.var = NULL,
  train.p = 0.75,
  strat.n.bins = 4,
  target.length = NULL,
  verbose = TRUE
)
}
\arguments{
\item{resampler}{Character: Type of resampling to perform: "strat.sub" (Stratified subsampling), 
"strat.boot" (Stratified bootstrap), "kfold", "bootstrap", "loocv" (Leave-one-out).
Default = "strat.sub"}

\item{n.resamples}{Integer: Number of training/testing sets required}

\item{stratify.var}{Numeric vector (optional): Variable used for stratification. Defaults to \code{y}}

\item{train.p}{Float (0, 1): Fraction of cases to assign to traininig set for \code{resampler = "strat.sub"}}

\item{strat.n.bins}{Integer: Number of groups to use for stratification for
\code{resampler = "strat.sub" / "strat.boot"}}

\item{target.length}{Integer: Number of cases for training set for \code{resampler = "strat.boot"}.
Default = \code{length(y)}}

\item{verbose}{Logical: If TRUE, print messages to screen}
}
\description{
\code{rtset.cv.resample}: \link{resample} defaults for cross-validation
}
