% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.DN}
\alias{rtset.DN}
\title{\code{rtset.DN}: Set parameters for \link{s_DN}}
\usage{
rtset.DN(
  hidden = 1,
  activation = NULL,
  learning.rate = 0.8,
  momentum = 0.5,
  learningrate_scale = 1,
  output = NULL,
  numepochs = 100,
  batchsize = NULL,
  hidden_dropout = 0,
  visible_dropout = 0,
  ...
)
}
\description{
\code{rtset.DN}: Set parameters for \link{s_DN}
}
