% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.earlystop}
\alias{rtset.earlystop}
\title{\code{rtset.earlystop}: Set parameters for \link{earlystop}}
\usage{
rtset.earlystop(
  window = 150,
  window_decrease_pct_min = 0.01,
  total_decrease_pct_max = NULL
)
}
\arguments{
\item{window}{Integer: Number of steps to consider}

\item{window_decrease_pct_min}{Float: Stop if improvemenet is less than this percent over last `window` steps}

\item{total_decrease_pct_max}{Float: Stop if improvement from first to last step exceeds this percent. If defined, overrides `window_decrease_pct_min`}
}
\description{
\code{rtset.earlystop}: Set parameters for \link{earlystop}
}
