% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.HYTREE}
\alias{rtset.HYTREE}
\title{\code{rtset.HYTREE}: Set parameters for \link{s.HYTREE}}
\usage{
rtset.HYTREE(
  max.depth = 2,
  learning.rate = 1,
  lincoef.params = rtset.lincoef("glmnet"),
  alpha = 0,
  lambda = 0.1,
  minobsinnode = 2,
  minobsinnode.lin = 20,
  ...
)
}
\arguments{
\item{max.depth}{[gS] Integer: Max depth of additive tree. Default = 3}

\item{learning.rate}{[gS] Float (0, 1): Learning rate. Default = 1}

\item{lincoef.params}{Named List: Output of \link{rtset.lincoef}}

\item{alpha}{[gS] Float: \code{lincoef} alpha Overrides \code{lincoef.params} alpha}

\item{lambda}{[gS] Float: \code{lincoef} lambda. Overrides \code{lincoef.params} lambda}

\item{minobsinnode}{[gS] Integer: Minimum N observations needed in node, before considering splitting}
}
\description{
\code{rtset.HYTREE}: Set parameters for \link{s.HYTREE}
}
