% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.LightRuleFit}
\alias{rtset.LightRuleFit}
\title{Set \link{s_LightRuleFit} parameters}
\usage{
rtset.LightRuleFit(
  num_leaves = 32L,
  max_depth = 3,
  learning_rate = 0.1,
  subsample = 0.666,
  subsample_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  objective = NULL,
  extra.lgbm.params = NULL,
  lightgbm.ifw = TRUE,
  glmnet.ifw = TRUE,
  importance = FALSE,
  alpha = 1,
  lambda = NULL
)
}
\arguments{
\item{lightgbm.ifw}{Logical: Passed to \link{s_LightGBM}'s \code{ifw} argument}

\item{glmnet.ifw}{Logical: Passed to \link{s_GLMNET}'s \code{ifw} argument}
}
\description{
Sets parameters for the GBM and LASSO steps of \link{s_LightRuleFit}
}
\author{
ED Gennatas
}
