% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.LIHAD}
\alias{rtset.LIHAD}
\title{Set \link{s_LIHAD} parameters}
\usage{
rtset.LIHAD(
  max.depth = 2,
  learning.rate = 1,
  lincoef.params = rtset.lincoef("glmnet"),
  alpha = 0,
  lambda = 0.1,
  minobsinnode = 2,
  minobsinnode.lin = 20,
  ...
)
}
\arguments{
\item{max.depth}{[gS] Integer: Max depth of additive tree. Default = 3}

\item{learning.rate}{[gS] Float (0, 1): Learning rate.}

\item{lincoef.params}{Named List: Output of \link{rtset.lincoef}}

\item{alpha}{[gS] Float: \code{lincoef} alpha Overrides \code{lincoef.params} alpha}

\item{lambda}{[gS] Float: \code{lincoef} lambda. Overrides \code{lincoef.params} lambda}

\item{minobsinnode}{[gS] Integer: Minimum N observations needed in node, before
considering splitting}

\item{minobsinnode.lin}{Integer: Minimum N observations needed in node in order to
train linear model.}

\item{...}{Additional arguments}
}
\description{
Set \link{s_LIHAD} parameters
}
