% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.preprocess}
\alias{rtset.preprocess}
\title{\code{rtset.preprocess}: Set \link{preprocess} parameters for \link{elevate}'s \code{.preprocess} argument}
\usage{
rtset.preprocess(completeCases = FALSE, impute = FALSE,
  impute.type = "missForest", impute.niter = 10, impute.ntree = 500,
  impute.discrete = getMode, impute.numeric = mean,
  removeCases.thres = NULL, removeFeatures.thres = NULL,
  integer2factor = FALSE, nonzeroFactors = FALSE, scale = FALSE,
  center = FALSE, removeConstant = TRUE, oneHot = FALSE)
}
\arguments{
\item{decom}{String: Name of decomposer to use. Default = "ICA"}

\item{k}{Integer: Number of dimensions to project to. Default = 2}
}
\description{
\code{rtset.preprocess}: Set \link{preprocess} parameters for \link{elevate}'s \code{.preprocess} argument
}
