% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_TLS.R
\name{predict.rtTLS}
\alias{predict.rtTLS}
\alias{print.rtTLS}
\title{\code{predict.rtTLS}: \code{predict} method for \code{rtTLS} object}
\usage{
\method{predict}{rtTLS}(object, newdata, ...)

\method{print}{rtTLS}(x, ...)
}
\description{
\code{predict.rtTLS}: \code{predict} method for \code{rtTLS} object

\code{print.rtTLS}: \code{print} method for \code{rtTLS} object
}
