% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_dist.R
\name{rule_dist}
\alias{rule_dist}
\title{Rule distance}
\usage{
rule_dist(
  x,
  rules1,
  rules2 = NULL,
  print_plot = TRUE,
  plot_type = c("static", "interactive"),
  heat_lo = "black",
  heat_mid = NA,
  heat_hi = "#F48024",
  verbosity = 1L
)
}
\arguments{
\item{x}{Data frame / matrix: Input features (cases by features).}

\item{rules1}{Character, vector: Rules as combination of conditions on the features of \code{x}.}

\item{rules2}{String, vector, Optional: Rules as combination of conditions on the features of
\code{x}.}

\item{print_plot}{Logical: If TRUE, plot heatmap for calculated distance.}

\item{plot_type}{Character: "static", "interactive": type of graphics to use, base or plotly,
respectively.}

\item{heat_lo}{Color: Heatmap low color.}

\item{heat_mid}{Color: Heatmap mid color. Default = NA (i.e. create gradient from \code{heat_lo} to
\code{heat_hi})}

\item{heat_hi}{Colo: Heatmap hi colo.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{rtrule_dist} object.
}
\description{
Calculate pairwise distance among a set of rules or between two sets of rules,
where each rule defines a subpopulation.
}
\details{
If only rules1 is provided, computes pairwise distance among rules1, otherwise computes pairwise
distance between rules1 and rules2.
}
\author{
EDG
}
