% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleDist.R
\name{ruleDist}
\alias{ruleDist}
\title{Rule distance}
\usage{
ruleDist(
  x,
  rules1,
  rules2 = NULL,
  print.plot = TRUE,
  plot.type = c("static", "interactive"),
  heat.lo = "black",
  heat.mid = NA,
  heat.hi = "#F48024",
  verbose = TRUE
)
}
\arguments{
\item{x}{Data frame / matrix: Input features (cases by features)}

\item{rules1}{Character, vector: Rules as combination of conditions on the features of \code{x}}

\item{rules2}{String, vector, Optional: Rules as combination of conditions on the features of \code{x}}

\item{print.plot}{Logical: If TRUE, plot heatmap for calculated distance}

\item{plot.type}{Character: "static", "interactive": type of graphics to use, base or plotly, respectively. Default =
"static"}

\item{heat.lo}{Color: Heatmap low color. Default = "black"}

\item{heat.mid}{Color: Heatmap mid color. Default = NA (i.e. create gradient from `heat.lo` to `heat.hi`)}

\item{heat.hi}{Colo: Heatmap hi colo. Default = "#F48024" (orange)}

\item{verbose}{Logical: If TRUE, print console messages. Default = TRUE}
}
\description{
Calculate pairwise distance among a set of rules or between two sets of rules,
where each rule defines a subpopulation
}
\details{
If only rules1 is provided, computes pairwise distance among rules1, otherwise computes pairwise distance between
rules1 and rules2
}
\author{
E.D. Gennatas
}
