% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleFeat.R
\name{ruleFeat}
\alias{ruleFeat}
\title{ruleFeat [C]}
\usage{
ruleFeat(x, y = NULL, x.test = NULL, y.test = NULL, n.trees = 500,
  gbm.params = list(bag.fraction = 0.5, shrinkage = 0.001,
  interaction.depth = 5, ipw = TRUE), glmnet.select.params = list(alpha =
  1, ipw = TRUE), cases.by.rules = NULL, x.name = NULL,
  y.name = NULL, question = NULL, verbose = TRUE, n.cores = 1,
  outdir = NULL, save.mod = if (!is.null(outdir)) TRUE else FALSE)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{n.trees}{Integer: Initial number of trees to fit}

\item{gbm.params}{Named list: Parameters for \link{s.GBM}}

\item{glmnet.select.params}{Named list: Parameters for \link{s.GLMNET} for the feature
selection step}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{question}{String: the question you are attempting to answer with this model, in plain language.}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{outdir}{String: If defined, save log, 'plot.all' plots (see above) and RDS file of complete output}

\item{save.mod}{Logical. If TRUE, save all output as RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{glmnet.final.params}{Named list: Parameters for \link{s.GLMNET} for the final model
building}
}
\value{
\link{rtMod} object
}
\description{
Train a gradient boosting model, extract rules,
and fit using LASSO
}
\details{
Based on "Predictive Learning via Rule Ensembles"
by Friedman and Popescu
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\references{
Friedman JH, Popescu BE, "Predictive Learning via Rule Ensembles",
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\author{
Efstathios D. Gennatas
}
