% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfn.R
\name{runifmat}
\alias{runifmat}
\title{Random Uniform Matrix}
\usage{
runifmat(
  nrow = 10,
  ncol = 10,
  min = 0,
  max = 1,
  return.df = FALSE,
  seed = NULL
)
}
\arguments{
\item{nrow}{Integer: Number of rows. Default = 10}

\item{ncol}{Integer: Number of columns. Default = 10}

\item{min}{Float: Min Default = 0}

\item{max}{Float: Max. Default = 1}

\item{return.df}{Logical: If TRUE, return data.frame, otherwise matrix. Default = TRUE}

\item{seed}{Integer: Set seed for \code{rnorm}. Default = NULL}
}
\description{
Create a matrix or data frame of defined dimensions, whose columns are random uniform vectors
}
\author{
E.D. Gennatas
}
