% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.BRUTO.R
\name{s.BRUTO}
\alias{s.BRUTO}
\title{Projection Pursuit Regression (BRUTO) [R]}
\usage{
s.BRUTO(x, y = NULL, x.test = NULL, y.test = NULL, x.name = NULL,
  y.name = NULL, grid.resample.rtset = rtset.grid.resample(),
  weights = NULL, weights.col = NULL, dfmax = 6, cost = 2,
  maxit.select = 20, maxit.backfit = 20, thresh = 1e-04,
  start.linear = TRUE, n.cores = rtCores, print.plot = TRUE,
  plot.fitted = NULL, plot.predicted = NULL,
  plot.theme = getOption("rt.fit.theme", "lightgrid"), question = NULL,
  rtclass = NULL, verbose = TRUE, trace = 0, outdir = NULL,
  save.mod = ifelse(!is.null(outdir), TRUE, FALSE), ...)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{(Optional) Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{(Optional) Numeric vector of testing set outcome}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{String: "zero", "dark", "box", "darkbox"}

\item{question}{String: the question you are attempting to answer with this model, in plain language.}

\item{rtclass}{String: Class type to use. "S3", "S4", "RC", "R6"}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If higher than 0, will print more information to the console. Default = 0}

\item{outdir}{Path to output directory.
If defined, will save Predicted vs. True plot, if available,
as well as full model output, if \code{save.mod} is TRUE}

\item{save.mod}{Logical. If TRUE, save all output as RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{...}{Additional arguments to be passed to \code{mda::bruto}}
}
\value{
Object of class \pkg{rtemis}
}
\description{
Trains a BRUTO model and validates it
}
\seealso{
\link{elevate} for external cross-validation

Other Supervised Learning: \code{\link{s.ADABOOST}},
  \code{\link{s.ADDTREE}}, \code{\link{s.BART}},
  \code{\link{s.BAYESGLM}}, \code{\link{s.C50}},
  \code{\link{s.CART}}, \code{\link{s.CTREE}},
  \code{\link{s.DA}}, \code{\link{s.ET}},
  \code{\link{s.EVTREE}}, \code{\link{s.GAM.default}},
  \code{\link{s.GAM.formula}}, \code{\link{s.GAMSEL}},
  \code{\link{s.GAM}}, \code{\link{s.GBM3}},
  \code{\link{s.GBM}}, \code{\link{s.GLMNET}},
  \code{\link{s.GLM}}, \code{\link{s.GLS}},
  \code{\link{s.H2ODL}}, \code{\link{s.H2OGBM}},
  \code{\link{s.H2ORF}}, \code{\link{s.IRF}},
  \code{\link{s.KNN}}, \code{\link{s.LDA}},
  \code{\link{s.LM}}, \code{\link{s.MARS}},
  \code{\link{s.MLRF}}, \code{\link{s.MXN}},
  \code{\link{s.NBAYES}}, \code{\link{s.NLA}},
  \code{\link{s.NLS}}, \code{\link{s.NW}},
  \code{\link{s.POLYMARS}}, \code{\link{s.PPR}},
  \code{\link{s.PPTREE}}, \code{\link{s.QDA}},
  \code{\link{s.QRNN}}, \code{\link{s.RANGER}},
  \code{\link{s.RFSRC}}, \code{\link{s.RF}},
  \code{\link{s.SGD}}, \code{\link{s.SPLS}},
  \code{\link{s.SVM}}, \code{\link{s.TFN}},
  \code{\link{s.XGBLIN}}, \code{\link{s.XGB}}
}
\author{
Efstathios D. Gennatas
}
\concept{Supervised Learning}
