% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.CTREE.R
\name{s.CTREE}
\alias{s.CTREE}
\title{Conditional Inference Trees [C, R, S]}
\usage{
s.CTREE(x, y = NULL, x.test = NULL, y.test = NULL, weights = NULL,
  control = partykit::ctree_control(), ipw = TRUE, ipw.type = 2,
  upsample = FALSE, downsample = FALSE, resample.seed = NULL,
  x.name = NULL, y.name = NULL, print.plot = TRUE,
  plot.fitted = NULL, plot.predicted = NULL,
  plot.theme = getOption("rt.fit.theme", "lightgrid"), question = NULL,
  verbose = TRUE, outdir = NULL, save.mod = ifelse(!is.null(outdir),
  TRUE, FALSE), ...)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{control}{List of parameters for the CTREE algorithms. Set using
\code{partykit::ctree_control}}

\item{ipw}{Logical: If TRUE, apply inverse probability weighting (for Classification only).
Note: If \code{weights} are provided, \code{ipw} is not used. Default = TRUE}

\item{ipw.type}{Integer {0, 1, 2}
1: class.weights as in 0, divided by max(class.weights)
2: class.weights as in 0, divided by min(class.weights)
Default = 2}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Caution: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}. Default = TRUE}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{outdir}{Path to output directory.
If defined, will save Predicted vs. True plot, if available,
as well as full model output, if \code{save.mod} is TRUE}

\item{save.mod}{Logical. If TRUE, save all output as RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{...}{Additional arguments}
}
\value{
\link{rtMod} object
}
\description{
Train a conditional inference tree using {partykit::ctree}
}
\seealso{
\link{elevate}

Other Supervised Learning: \code{\link{s.ADABOOST}},
  \code{\link{s.ADDTREE}}, \code{\link{s.BART}},
  \code{\link{s.BAYESGLM}}, \code{\link{s.BRUTO}},
  \code{\link{s.C50}}, \code{\link{s.CART}},
  \code{\link{s.DA}}, \code{\link{s.ET}},
  \code{\link{s.EVTREE}}, \code{\link{s.GAM.default}},
  \code{\link{s.GAM.formula}}, \code{\link{s.GAMSEL}},
  \code{\link{s.GAM}}, \code{\link{s.GBM3}},
  \code{\link{s.GBM}}, \code{\link{s.GLMNET}},
  \code{\link{s.GLM}}, \code{\link{s.GLS}},
  \code{\link{s.H2ODL}}, \code{\link{s.H2OGBM}},
  \code{\link{s.H2ORF}}, \code{\link{s.IRF}},
  \code{\link{s.KNN}}, \code{\link{s.LDA}},
  \code{\link{s.LM}}, \code{\link{s.MARS}},
  \code{\link{s.MLRF}}, \code{\link{s.MXN}},
  \code{\link{s.NBAYES}}, \code{\link{s.NLA}},
  \code{\link{s.NLS}}, \code{\link{s.NW}},
  \code{\link{s.POLYMARS}}, \code{\link{s.PPR}},
  \code{\link{s.PPTREE}}, \code{\link{s.QDA}},
  \code{\link{s.QRNN}}, \code{\link{s.RANGER}},
  \code{\link{s.RFSRC}}, \code{\link{s.RF}},
  \code{\link{s.SGD}}, \code{\link{s.SPLS}},
  \code{\link{s.SVM}}, \code{\link{s.TFN}},
  \code{\link{s.XGBLIN}}, \code{\link{s.XGB}}

Other Tree-based methods: \code{\link{s.ADABOOST}},
  \code{\link{s.ADDTREE}}, \code{\link{s.BART}},
  \code{\link{s.C50}}, \code{\link{s.CART}},
  \code{\link{s.ET}}, \code{\link{s.EVTREE}},
  \code{\link{s.GBM3}}, \code{\link{s.GBM}},
  \code{\link{s.H2OGBM}}, \code{\link{s.H2ORF}},
  \code{\link{s.IRF}}, \code{\link{s.MLRF}},
  \code{\link{s.PPTREE}}, \code{\link{s.RANGER}},
  \code{\link{s.RFSRC}}, \code{\link{s.RF}},
  \code{\link{s.XGB}}
}
\author{
Efstathios D. Gennatas
}
\concept{Supervised Learning}
\concept{Tree-based methods}
