% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.KNN.R
\name{s.KNN}
\alias{s.KNN}
\title{k-Nearest Neighbors Classification and Regression [C, R]}
\usage{
s.KNN(x, y = NULL, x.test = NULL, y.test = NULL, x.name = NULL,
  y.name = NULL, k = 3, algorithm = "kd_tree", print.plot = TRUE,
  plot.fitted = NULL, plot.predicted = NULL,
  plot.theme = getOption("rt.fit.theme", "lightgrid"), question = NULL,
  rtclass = NULL, verbose = TRUE, outdir = NULL,
  save.mod = ifelse(!is.null(outdir), TRUE, FALSE), ...)
}
\arguments{
\item{x}{Numeric vector or matrix of features, i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{(Optional) Numeric vector or matrix of validation set features
must have set of columns as \code{x}}

\item{y.test}{(Optional) Numeric vector of validation set outcomes}

\item{k}{Integer: Number of neighbors considered}

\item{algorithm}{Character: Algorithm to use. Options: "kd_tree", "cover_tree", "brute"}

\item{outdir}{Optional. Path to directory to save output}
}
\value{
Object of class \link{rtMod}
}
\description{
Train a k-Nearest Neighbors learner for regression or classification using \code{FNN}
}
\details{
Note: FNN's KNN does not have a predict function
}
\seealso{
\link{elevate} for external cross-validation

Other Supervised Learning: \code{\link{s.ADABOOST}},
  \code{\link{s.ADDTREE}}, \code{\link{s.BART}},
  \code{\link{s.BAYESGLM}}, \code{\link{s.BRUTO}},
  \code{\link{s.C50}}, \code{\link{s.CART}},
  \code{\link{s.CTREE}}, \code{\link{s.DA}},
  \code{\link{s.ET}}, \code{\link{s.EVTREE}},
  \code{\link{s.GAM.default}}, \code{\link{s.GAM.formula}},
  \code{\link{s.GAMSEL}}, \code{\link{s.GAM}},
  \code{\link{s.GBM3}}, \code{\link{s.GBM}},
  \code{\link{s.GLMNET}}, \code{\link{s.GLM}},
  \code{\link{s.GLS}}, \code{\link{s.H2ODL}},
  \code{\link{s.H2OGBM}}, \code{\link{s.H2ORF}},
  \code{\link{s.IRF}}, \code{\link{s.LDA}},
  \code{\link{s.LM}}, \code{\link{s.MARS}},
  \code{\link{s.MLRF}}, \code{\link{s.MXN}},
  \code{\link{s.NBAYES}}, \code{\link{s.NLA}},
  \code{\link{s.NLS}}, \code{\link{s.NW}},
  \code{\link{s.POLYMARS}}, \code{\link{s.PPR}},
  \code{\link{s.PPTREE}}, \code{\link{s.QDA}},
  \code{\link{s.QRNN}}, \code{\link{s.RANGER}},
  \code{\link{s.RFSRC}}, \code{\link{s.RF}},
  \code{\link{s.SGD}}, \code{\link{s.SPLS}},
  \code{\link{s.SVM}}, \code{\link{s.TFN}},
  \code{\link{s.XGBLIN}}, \code{\link{s.XGB}}
}
\author{
Efstathios D. Gennatas
}
\concept{Supervised Learning}
