% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.GLM.R
\name{s.LOGISTIC}
\alias{s.LOGISTIC}
\title{Logistic Regression}
\usage{
s.LOGISTIC(x, y, x.test = NULL, y.test = NULL, family = binomial(link
  = "logit"), ...)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{family}{Error distribution and link function. See \code{stats::family}}

\item{...}{Additional arguments}
}
\description{
Convenience alias for \code{s.GLM(family = binomial(link = "logit"))}.
}
