% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.MLRF.R
\name{s.MLRF}
\alias{s.MLRF}
\title{Spark MLlib Random Forest [C, R]}
\usage{
s.MLRF(x, y = NULL, x.test = NULL, y.test = NULL, upsample = FALSE,
  upsample.seed = NULL, n.trees = 500, max.depth = 30L,
  subsampling.rate = 1, min.instances.per.node = 1,
  feature.subset.strategy = "auto", max.bins = 32L, x.name = NULL,
  y.name = NULL, spark.master = "local", print.plot = TRUE,
  plot.fitted = NULL, plot.predicted = NULL,
  plot.theme = getOption("rt.fit.theme", "lightgrid"), question = NULL,
  verbose = TRUE, trace = 0, outdir = NULL,
  save.mod = ifelse(!is.null(outdir), TRUE, FALSE), ...)
}
\arguments{
\item{x}{vector, matrix or dataframe of training set features}

\item{y}{vector of outcomes}

\item{x.test}{vector, matrix or dataframe of testing set features}

\item{y.test}{vector of testing set outcomes}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Caution: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{upsample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{n.trees}{Integer. Number of trees to train}

\item{max.depth}{Integer. Max depth of each tree}

\item{max.bins}{Integer. Max N of bins used for discretizing continuous features and for
choosing how to split on features at each node. More bins give higher granularity.}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{spark.master}{Spark cluster URL or "local"}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{String: "zero", "dark", "box", "darkbox"}

\item{question}{String: the question you are attempting to answer with this model, in plain language.}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If higher than 0, will print more information to the console. Default = 0}

\item{outdir}{Path to output directory.
If defined, will save Predicted vs. True plot, if available,
as well as full model output, if \code{save.mod} is TRUE}

\item{save.mod}{Logical. If TRUE, save all output as RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{...}{Additional arguments}

\item{type}{"regression" for continuous outcome; "classification" for categorical outcome.
"auto" will result in regression for numeric \code{y} and classification otherwise}
}
\value{
\link{rtMod} object
}
\description{
Train an MLlib Random Forest model on Spark
}
\details{
The overhead incurred by Spark means this should be used only for really large datasets on
a Spark cluster, not on a regular local machine.
}
\seealso{
\link{elevate} for external cross-validation

Other Supervised Learning: \code{\link{s.ADABOOST}},
  \code{\link{s.ADDTREE}}, \code{\link{s.BART}},
  \code{\link{s.BAYESGLM}}, \code{\link{s.BRUTO}},
  \code{\link{s.C50}}, \code{\link{s.CART}},
  \code{\link{s.CTREE}}, \code{\link{s.DA}},
  \code{\link{s.ET}}, \code{\link{s.EVTREE}},
  \code{\link{s.GAM.default}}, \code{\link{s.GAM.formula}},
  \code{\link{s.GAMSEL}}, \code{\link{s.GAM}},
  \code{\link{s.GBM3}}, \code{\link{s.GBM}},
  \code{\link{s.GLMNET}}, \code{\link{s.GLM}},
  \code{\link{s.GLS}}, \code{\link{s.H2ODL}},
  \code{\link{s.H2OGBM}}, \code{\link{s.H2ORF}},
  \code{\link{s.IRF}}, \code{\link{s.KNN}},
  \code{\link{s.LDA}}, \code{\link{s.LGB}},
  \code{\link{s.LM}}, \code{\link{s.MARS}},
  \code{\link{s.MXN}}, \code{\link{s.NBAYES}},
  \code{\link{s.NLA}}, \code{\link{s.NLS}},
  \code{\link{s.NW}}, \code{\link{s.POLYMARS}},
  \code{\link{s.PPR}}, \code{\link{s.PPTREE}},
  \code{\link{s.QDA}}, \code{\link{s.QRNN}},
  \code{\link{s.RANGER}}, \code{\link{s.RFSRC}},
  \code{\link{s.RF}}, \code{\link{s.SGD}},
  \code{\link{s.SPLS}}, \code{\link{s.SVM}},
  \code{\link{s.TFN}}, \code{\link{s.XGBLIN}},
  \code{\link{s.XGB}}

Other Tree-based methods: \code{\link{s.ADABOOST}},
  \code{\link{s.ADDTREE}}, \code{\link{s.BART}},
  \code{\link{s.C50}}, \code{\link{s.CART}},
  \code{\link{s.CTREE}}, \code{\link{s.ET}},
  \code{\link{s.EVTREE}}, \code{\link{s.GBM3}},
  \code{\link{s.GBM}}, \code{\link{s.H2OGBM}},
  \code{\link{s.H2ORF}}, \code{\link{s.IRF}},
  \code{\link{s.LGB}}, \code{\link{s.PPTREE}},
  \code{\link{s.RANGER}}, \code{\link{s.RFSRC}},
  \code{\link{s.RF}}, \code{\link{s.XGB}}
}
\author{
Efstathios D. Gennatas
}
\concept{Supervised Learning}
\concept{Tree-based methods}
