% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.PSURV.R
\name{s.PSURV}
\alias{s.PSURV}
\title{Parametric Survival Regression [S]}
\usage{
s.PSURV(
  x,
  y,
  x.test = NULL,
  y.test = NULL,
  x.name = NULL,
  y.name = NULL,
  weights = NULL,
  dist = "weibull",
  control = survival::survreg.control(),
  print.plot = TRUE,
  plot.fitted = NULL,
  plot.predicted = NULL,
  plot.theme = getOption("rt.theme"),
  question = NULL,
  verbose = TRUE,
  trace = 0,
  save.mod = FALSE,
  outdir = NULL,
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix of features, i.e. independent variables}

\item{y}{Object of class "Surv" created using \code{survival::Surv}}

\item{x.test}{(Optional) Numeric vector or matrix of testing set features
must have set of columns as \code{x}}

\item{y.test}{(Optional) Object of class "Surv" created using \code{survival::Surv}}

\item{weights}{Float: Vector of case weights}

\item{...}{Additional parameters to pass to \code{survival::survreg}}
}
\value{
Object of class \link{rtMod}
}
\description{
Fit a parametric survival regression model using \code{survival::survreg}
}
\seealso{
\link{elevate} for external cross-validation
}
\author{
E.D. Gennatas
}
\concept{Survival Regression}
