% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.SHYTREE.R
\name{s.SHYTREE}
\alias{s.SHYTREE}
\title{Stepwise Hybrid Tree [C, R]}
\usage{
s.SHYTREE(x, y = NULL, x.test = NULL, y.test = NULL,
  weights = NULL, ipw = TRUE, ipw.type = 2, upsample = FALSE,
  downsample = FALSE, resample.seed = NULL, max.leaves = 6,
  force.max.leaves = NULL, gamma = 0.1, alpha = 0, lambda = 0.05,
  lambda.seq = NULL, minobsinnode = 2, minobsinnode.lin = 10,
  learning.rate = 1, part.minsplit = 2, part.xval = 0,
  part.max.depth = 1, part.cp = 0, .rho = TRUE, rho.max = 1000,
  init = NULL, lin.type = c("glmnet", "cv.glmnet", "lm.ridge",
  "allSubsets", "forwardStepwise", "backwardStepwise", "glm", "solve"),
  cv.glmnet.nfolds = 5, cv.glmnet.lambda = "lambda.min",
  metric = "auto", .gs = FALSE, maximize = NULL,
  grid.resample.rtset = rtset.grid.resample(),
  grid.search.type = "exhaustive", save.gridrun = FALSE,
  grid.verbose = TRUE, keep.x = FALSE, simplify = TRUE,
  cxrcoef = FALSE, n.cores = rtCores, .preprocess = NULL,
  verbose = TRUE, plot.tune.error = FALSE, verbose.predict = FALSE,
  trace = 0, x.name = NULL, y.name = NULL, question = NULL,
  outdir = NULL, print.plot = TRUE, plot.fitted = NULL,
  plot.predicted = NULL, plot.theme = getOption("rt.fit.theme",
  "lightgrid"), save.mod = FALSE)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{ipw}{Logical: If TRUE, apply inverse probability weighting (for Classification only).
Note: If \code{weights} are provided, \code{ipw} is not used. Default = TRUE}

\item{ipw.type}{Integer {0, 1, 2}
1: class.weights as in 0, divided by max(class.weights)
2: class.weights as in 0, divided by min(class.weights)
Default = 2}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Caution: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{max.leaves}{Integer: Maximum number of terminal nodes to grow}

\item{lambda}{Float: lambda parameter for \code{MASS::lm.ridge} Default = .01}

\item{minobsinnode}{Integer: Minimum N observations needed in node, before considering splitting}

\item{part.max.depth}{Integer: Max depth for each tree model within the additive tree}

\item{init}{Initial value. Default = \code{mean(y)}}

\item{.gs}{internal use only}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{plot.tune.error}{Logical: If TRUE, plot validation error during gridsearch}

\item{trace}{Integer: If higher than 0, will print more information to the console. Default = 0}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{outdir}{Path to output directory.
If defined, will save Predicted vs. True plot, if available,
as well as full model output, if \code{save.mod} is TRUE}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}. Default = TRUE}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{save.mod}{Logical. If TRUE, save all output as RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}
}
\description{
Train a Stepwise Hybrid Tree for Regression or Binary Classification
}
\details{
The Stepwise Hybrid Tree grows a tree using a sequence of regularized linear models and tree stumps
We specify an upper threshold of leaves using \code{max.leaves} instead of directly defining a number,
because depending on the other parameters and the datasets, splitting may stop early.
}
\author{
Efstathios D. Gennatas
}
