% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_GBM3.R
\name{s_GBM3}
\alias{s_GBM3}
\title{Gradient Boosting Machine \link{C, R, S}}
\usage{
s_GBM3(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  weights = NULL,
  ifw = TRUE,
  ifw.type = 2,
  upsample = FALSE,
  downsample = FALSE,
  resample.seed = NULL,
  distribution = NULL,
  interaction.depth = 3,
  shrinkage = 0.01,
  bag.fraction = 0.9,
  mFeatures = NULL,
  n.minobsinnode = 5,
  n.trees = 2000,
  max.trees = 5000,
  force.n.trees = NULL,
  n.tree.window = 0,
  gbm.select.smooth = FALSE,
  smoother = c("loess", "supsmu"),
  n.new.trees = 500,
  min.trees = 50,
  failsafe.trees = 1000,
  imetrics = FALSE,
  .gs = FALSE,
  grid.resample.rtset = rtset.resample("kfold", 5),
  grid.search.type = c("exhaustive", "randomized"),
  grid.randomized.p = 0.1,
  metric = NULL,
  maximize = NULL,
  plot.tune.error = FALSE,
  exclude.test.lt.train = FALSE,
  exclude.lt.min.trees = FALSE,
  res.fail.thres = 0.99,
  n.extra.trees = 0,
  n.cores = rtCores,
  gbm.cores = 1,
  relInf = TRUE,
  varImp = FALSE,
  offset = NULL,
  var.monotone = NULL,
  keep.data = TRUE,
  var.names = NULL,
  response.name = "y",
  group = NULL,
  plot.perf = FALSE,
  plot.res = ifelse(!is.null(outdir), TRUE, FALSE),
  plot.fitted = NULL,
  plot.predicted = NULL,
  plotRelInf = FALSE,
  plotVarImp = FALSE,
  print.plot = FALSE,
  plot.theme = rtTheme,
  x.name = NULL,
  y.name = NULL,
  question = NULL,
  verbose = TRUE,
  trace = 0,
  grid.verbose = verbose,
  gbm.fit.verbose = FALSE,
  outdir = NULL,
  save.gridrun = FALSE,
  save.error.diagnostics = FALSE,
  save.rds = TRUE,
  save.res = FALSE,
  save.res.mod = FALSE,
  save.mod = ifelse(!is.null(outdir), TRUE, FALSE),
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ifw}, therefore set \code{weights = NULL} if using \code{ifw}.
Note: If \code{weight} are provided, \code{ifw} is not used. Leave NULL if setting \code{ifw = TRUE}.}

\item{ifw}{Logical: If TRUE, apply inverse frequency weighting
(for Classification only).
Note: If \code{weights} are provided, \code{ifw} is not used.}

\item{ifw.type}{Integer {0, 1, 2}
1: class.weights as in 0, divided by min(class.weights)
2: class.weights as in 0, divided by max(class.weights)}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Note: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{downsample}{Logical: If TRUE, downsample majority class to match size of minority class}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{interaction.depth}{\link{gS} Integer: Interaction depth. Default = 3}

\item{shrinkage}{\link{gS} Float: Shrinkage (learning rate). Default = .01}

\item{bag.fraction}{\link{gS} Float (0, 1): Fraction of cases to use to train each tree.
Helps avoid overfitting. Default = .9}

\item{mFeatures}{\link{gS} Integer: Number of features to randomly choose from all available features to train at each
step. Default = NULL which results in using all features.}

\item{n.minobsinnode}{\link{gS} Integer: Minimum number of observation allowed in node}

\item{n.trees}{Integer: Initial number of trees to fit. Default = 2000}

\item{max.trees}{Integer: Maximum number of trees to fit. Default = 5000}

\item{grid.resample.rtset}{List: Output of \link{rtset.resample} defining
\link{gridSearchLearn} parameters.}

\item{grid.search.type}{Character: Type of grid search to perform:
"exhaustive" or "randomized".}

\item{grid.randomized.p}{Float (0, 1): If
\code{grid.search.type = "randomized"}, randomly test this proportion of
combinations.}

\item{metric}{Character: Metric to minimize, or maximize if
\code{maximize = TRUE} during grid search. Default = NULL, which results in
"Balanced Accuracy" for Classification,
"MSE" for Regression, and "Coherence" for Survival Analysis.}

\item{maximize}{Logical: If TRUE, \code{metric} will be maximized if grid
search is run.}

\item{n.cores}{Integer: Number of cores to use.}

\item{relInf}{Logical: If TRUE (Default), estimate variables' relative influence.}

\item{varImp}{Logical: If TRUE, estimate variable importance by permutation (as in random forests;
noted as experimental in gbm). Takes longer than (default) relative influence.
The two measures are highly correlated.}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}.}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If higher than 0, will print more information to the console.}

\item{grid.verbose}{Logical: Passed to \link{gridSearchLearn}}

\item{outdir}{Character: If defined, save log, 'plot.all' plots (see above) and RDS file of complete output}

\item{save.rds}{Logical: If outdir is defined, should all data be saved in RDS file? s.SVDnetGBM will save
mod.gbm, so no need to save again.}

\item{save.res.mod}{Logical: If TRUE, save gbm model for each grid run. For diagnostic purposes only:
Object size adds up quickly}

\item{save.mod}{Logical: If TRUE, save all output to an RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{...}{Additional arguments}

\item{stratify.var}{If resampling is stratified, stratify against this variable. Defaults to outcome}
}
\description{
Train a GBM model using \code{gbm-developers/gbm3}
}
\details{
Early stopping is implemented by fitting \code{n.trees} initially, checking the (smoothed) validation
error curve, and adding \code{n.new.trees} if needed, until error does not reduce or \code{max.trees} is
reached.
\link{gS} in the argument description indicates that multiple values can be passed, in
which case tuning will be performed using grid search. gS is supported for:
interaction.depth, shrinkage, bag.fraction, mFeatures, and n.minobsinnode
This function includes a workaround for when \code{gbm.fit} fails.
If an error is detected, \code{gbm.fit} is rerun until successful and the procedure continues normally
}
\seealso{
\link{train} for external cross-validation

Other Supervised Learning: 
\code{\link{s_AdaBoost}()},
\code{\link{s_AddTree}()},
\code{\link{s_BART}()},
\code{\link{s_BRUTO}()},
\code{\link{s_BayesGLM}()},
\code{\link{s_C50}()},
\code{\link{s_CART}()},
\code{\link{s_CTree}()},
\code{\link{s_ET}()},
\code{\link{s_EVTree}()},
\code{\link{s_GAM.default}()},
\code{\link{s_GAM.formula}()},
\code{\link{s_GAM}()},
\code{\link{s_GBM}()},
\code{\link{s_GLMNET}()},
\code{\link{s_GLMTree}()},
\code{\link{s_GLM}()},
\code{\link{s_GLS}()},
\code{\link{s_H2ODL}()},
\code{\link{s_H2OGBM}()},
\code{\link{s_H2ORF}()},
\code{\link{s_HAL}()},
\code{\link{s_IRF}()},
\code{\link{s_KNN}()},
\code{\link{s_LDA}()},
\code{\link{s_LMTree}()},
\code{\link{s_LM}()},
\code{\link{s_LightGBM}()},
\code{\link{s_MARS}()},
\code{\link{s_MLRF}()},
\code{\link{s_NBayes}()},
\code{\link{s_NLA}()},
\code{\link{s_NLS}()},
\code{\link{s_NW}()},
\code{\link{s_POLYMARS}()},
\code{\link{s_PPR}()},
\code{\link{s_PPTree}()},
\code{\link{s_QDA}()},
\code{\link{s_QRNN}()},
\code{\link{s_RFSRC}()},
\code{\link{s_RF}()},
\code{\link{s_Ranger}()},
\code{\link{s_SDA}()},
\code{\link{s_SGD}()},
\code{\link{s_SPLS}()},
\code{\link{s_SVM}()},
\code{\link{s_TFN}()},
\code{\link{s_XGBoost}()},
\code{\link{s_XRF}()}

Other Tree-based methods: 
\code{\link{s_AdaBoost}()},
\code{\link{s_AddTree}()},
\code{\link{s_BART}()},
\code{\link{s_C50}()},
\code{\link{s_CART}()},
\code{\link{s_CTree}()},
\code{\link{s_ET}()},
\code{\link{s_EVTree}()},
\code{\link{s_GBM}()},
\code{\link{s_GLMTree}()},
\code{\link{s_H2OGBM}()},
\code{\link{s_H2ORF}()},
\code{\link{s_IRF}()},
\code{\link{s_LMTree}()},
\code{\link{s_LightGBM}()},
\code{\link{s_MLRF}()},
\code{\link{s_PPTree}()},
\code{\link{s_RFSRC}()},
\code{\link{s_RF}()},
\code{\link{s_Ranger}()},
\code{\link{s_XGBoost}()},
\code{\link{s_XRF}()}

Other Ensembles: 
\code{\link{s_AdaBoost}()},
\code{\link{s_GBM}()},
\code{\link{s_RF}()},
\code{\link{s_Ranger}()}
}
\author{
E.D. Gennatas
}
\concept{Ensembles}
\concept{Supervised Learning}
\concept{Tree-based methods}
