% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_GLM.R
\name{s_MULTINOM}
\alias{s_MULTINOM}
\title{Multinomial Logistic Regression}
\usage{
s_MULTINOM(x, y, x.test = NULL, y.test = NULL, class.method = "multinom", ...)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{class.method}{Character: Define "logistic" or "multinom" for classification. The only purpose
of this is so you can try \code{nnet::multinom} instead of glm for binary classification}

\item{...}{Additional arguments}
}
\description{
Convenience alias for \code{s_GLM(class.method = "multinom")}.
}
