% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_GLM.R
\name{s_POLY}
\alias{s_POLY}
\title{Polynomial Regression}
\usage{
s_POLY(x, y, x.test = NULL, y.test = NULL, poly.d = 3, poly.raw = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{poly.d}{Integer: degree of polynomial(s) to use}

\item{poly.raw}{Logical: if TRUE, use raw polynomials. Defaults to FALSE, resulting in
orthogonal polynomials. See \code{stats::poly}}

\item{...}{Additional arguments}
}
\description{
Convenience alias for \code{s_GLM(polynomial = T)}.
Substitutes all features with \code{poly(x, poly.d)}
}
