% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LM.R
\name{s_RLM}
\alias{s_RLM}
\title{Robust linear model}
\usage{
s_RLM(x, y, x.test = NULL, y.test = NULL, ...)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{...}{Additional parameters to be passed to \code{MASS::rlm}}
}
\description{
Convenience alias for \code{s_LM(robust = T)}. Uses \code{MASS::rlm}
}
