% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.rds.R
\name{save.rds}
\alias{save.rds}
\title{Write data to RDS file}
\usage{
save.rds(object, outdir, verbose = TRUE)
}
\arguments{
\item{object}{\code{R} object}

\item{outdir}{Path to output directory}

\item{verbose}{Logical}
}
\description{
Write data to RDS file
}
\author{
Efstathios D. Gennatas
}
