% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePMML.R
\name{savePMML}
\alias{savePMML}
\title{Save rtemis model to PMML file}
\usage{
savePMML(
  x,
  filename,
  transforms = NULL,
  model_name = NULL,
  model_version = NULL,
  description = NULL,
  copyright = NULL,
  ...
)
}
\arguments{
\item{x}{rtemis model}

\item{filename}{Character: path to file}
}
\description{
Save rtemis model to PMML file
}
\author{
E.D. Gennatas
}
