% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Extract standard error of fit from rtemis model}
\usage{
se(x)
}
\arguments{
\item{x}{An \code{rtMod} object}
}
\value{
Standard error of fitted values of \code{mod}
}
\description{
Returns mod$se.fit
}
\author{
E.D. Gennatas
}
