% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_GLM.R
\name{se_GLM}
\alias{se_GLM}
\title{Get Standard Errors from GLM model}
\usage{
se_GLM(model, newdata)
}
\arguments{
\item{model}{GLM model.}

\item{newdata}{data.frame or similar: Data to predict on.}
}
\description{
Get Standard Errors from GLM model
}
