% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftilabels.R
\name{seglabels2itksnap}
\alias{seglabels2itksnap}
\title{Convert nifti segmentation label file to itksnap-compatible label file}
\usage{
seglabels2itksnap(
  x,
  visible = NULL,
  mesh = NULL,
  filename = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Character: path to nifti label file (read with \link{readseglabels})}

\item{visible}{Integer, vector, length = number of labels in \code{x}: 1: visible; 0: invisible.
Normally, the background label is set to invisible}

\item{mesh}{Same as \code{visible} but for mesh renderings}

\item{filename}{Character: Path to filename to write to}
}
\description{
Convert nifti segmentation label file to itksnap-compatible label file
}
\author{
E.D. Gennatas
}
