% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_clust.R
\name{select_clust}
\alias{select_clust}
\title{Select \pkg{rtemis} Clusterer}
\usage{
select_clust(clust, fn = FALSE, desc = FALSE)
}
\arguments{
\item{clust}{Character: Clustering algorithm name. Case insensitive, supports
partial matching. e.g. "hop" for HOPACH}

\item{fn}{Logical: If TRUE, return function, otherwise name of function.}

\item{desc}{Logical: If TRUE, return full name of algorithm \code{clust}}
}
\value{
Name of function (Default) or function (\code{fn=TRUE}) or full name
of algorithm (\code{desc=TRUE})
}
\description{
Accepts clusterer name (supports abbreviations) and returns \pkg{rtemis}
function name or the function itself.
If run with no parameters, prints list of available algorithms.
}
\author{
E.D. Gennatas
}
