% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_decom.R
\name{select_decom}
\alias{select_decom}
\title{Select \pkg{rtemis} Decomposer}
\usage{
select_decom(decom, fn = FALSE, desc = FALSE)
}
\arguments{
\item{decom}{Character: Decomposition name. Case insensitive. e.g. "iso" for isomap}

\item{fn}{Logical: If TRUE, return function, otherwise name of function. Defaults to FALSE}

\item{desc}{Logical: If TRUE, return full name of algorithm \code{decom}}
}
\value{
Function or name of function (see param \code{fn}) or full name of algorithm (\code{desc})
}
\description{
Accepts decomposer name (supports abbreviations) and returns \pkg{rtemis} function name or
the function itself.
If run with no parameters, prints list of available algorithms.
}
\author{
E.D. Gennatas
}
