% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiter.R
\name{selectiter}
\alias{selectiter}
\title{Select N of learning iterations based on loss}
\usage{
selectiter(
  loss.valid,
  loss.train,
  smooth = TRUE,
  plot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{loss.valid}{Float, vector: Validation loss. Can be NULL}

\item{loss.train}{Float, vector: Training loss}

\item{smooth}{Logical: If TRUE, smooth loss before finding minimum.}

\item{plot}{Logical: If TRUE, plot loss curve.}

\item{verbose}{Logical: If TRUE, print messages to console.}
}
\description{
Select N of learning iterations based on loss
}
\author{
E.D. Gennatas
}
