% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linadleaves.R
\name{selectleaves}
\alias{selectleaves}
\title{Select leaves}
\usage{
selectleaves(
  object,
  x,
  y,
  x.valid,
  y.valid,
  smooth = FALSE,
  print.plot = TRUE,
  verbose = TRUE,
  trace = 0
)
}
\description{
Select number of leaves based on validation error
}
\keyword{internal}
