% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_supervised.R
\name{set_outcome}
\alias{set_outcome}
\title{Move outcome to last column}
\usage{
set_outcome(dat, outcome_column)
}
\arguments{
\item{dat}{data.frame or similar.}

\item{outcome_column}{Character: Name of outcome column.}
}
\value{
object of same class as \code{data}
}
\description{
Move outcome to last column
}
\author{
EDG
}
