% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_supervised.R
\name{set_positive_class}
\alias{set_positive_class}
\title{Set positive class}
\usage{
set_positive_class(x, positive_class)
}
\arguments{
\item{x}{Tabular data, i.e. data.frame, data.table, or tbl_df (tibble): Input dataset. The last column will be treated as the outcome variable.}

\item{positive_class}{Character scalar: The name of the positive class. Must be one of the levels of the outcome variable.}
}
\value{
Tabular data of the same class as \code{x}, with the positive class set as the second level of the outcome factor.
}
\description{
Checks and sets the positive class for binary classification in a tabular dataset.
}
\examples{
dat <- data.frame(x = rnorm(100), y = factor(sample(c("Case", "Control"), 100, replace = TRUE)))
levels(dat[["y"]])
dat <- set_positive_class(dat, "Case")
levels(dat[["y"]])
}
\author{
EDG
}
