% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiffsym.R
\name{setdiffsym}
\alias{setdiffsym}
\title{Symmetric Set Difference}
\usage{
setdiffsym(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector of same type as \code{x}}
}
\description{
Symmetric Set Difference
}
\examples{
setdiff(1:10, 1:5)
setdiff(1:5, 1:10)
setdiffsym(1:10, 1:5)
setdiffsym(1:5, 1:10)
}
\author{
E.D. Gennatas
}
