% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.color}
\alias{setup.color}
\title{Set \link{colorGrad} parameters}
\usage{
setup.color(
  n = 101,
  colors = NULL,
  space = "rgb",
  lo = "#01256E",
  lomid = NULL,
  mid = "white",
  midhi = NULL,
  hi = "#95001A",
  colorbar = FALSE,
  cb.mar = c(1, 1, 1, 1),
  ...
)
}
\arguments{
\item{n}{Integer: How many distinct colors you want. If not odd, converted to \code{n + 1}
Defaults to 21}

\item{colors}{Character: Acts as a shortcut to defining \code{lo}, \code{mid}, etc for a number of defaults:
"french", "penn", "grnblkred",}

\item{space}{Character: Which colorspace to use. Option: "rgb", or "Lab". Default = "rgb".
Recommendation: If \code{mid} is "white" or "black" (default), use "rgb", otherwise "Lab"}

\item{lo}{Color for low end}

\item{lomid}{Color for low-mid}

\item{mid}{Color for middle of the range or "mean", which will result in \code{colorOp(c(lo, hi), "mean")}.
If \code{mid = NA}, then only \code{lo} and \code{hi} are used to create the color gradient.}

\item{midhi}{Color for middle-high}

\item{hi}{Color for high end}

\item{colorbar}{Logical: Create a vertical colorbar}

\item{cb.mar}{Vector, length 4: Colorbar margins. Default: c(1, 1, 1, 1)}

\item{...}{Additional arguments}
}
\description{
Set \link{colorGrad} parameters
}
