% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.cv.resample}
\alias{setup.cv.resample}
\title{\code{setup.cv.resample}: \link{resample} defaults for cross-validation}
\usage{
setup.cv.resample(
  resampler = "strat.sub",
  n.resamples = 10,
  stratify.var = NULL,
  train.p = 0.8,
  strat.n.bins = 4,
  target.length = NULL,
  id.strat = NULL,
  verbosity = 1
)
}
\arguments{
\item{resampler}{Character: Type of resampling to perform: "bootstrap", "kfold",
"strat.boot", "strat.sub".}

\item{n.resamples}{Integer: Number of training/testing sets required}

\item{stratify.var}{Numeric vector (optional): Variable used for stratification.}

\item{train.p}{Float (0, 1): Fraction of cases to assign to traininig set for
\code{resampler = "strat.sub"}}

\item{strat.n.bins}{Integer: Number of groups to use for stratification for
\code{resampler = "strat.sub" / "strat.boot"}}

\item{target.length}{Integer: Number of cases for training set for
\code{resampler = "strat.boot"}.}

\item{id.strat}{Vector of IDs which may be replicated: resampling should
force replicates of each ID to only appear in the training or testing.}

\item{verbosity}{Logical: If TRUE, print messages to console}
}
\description{
\code{setup.cv.resample}: \link{resample} defaults for cross-validation
}
