% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.decompose}
\alias{setup.decompose}
\title{Set decomposition parameters for \link{train_cv} \code{.decompose} argument}
\usage{
setup.decompose(decom = "ICA", k = 2, ...)
}
\arguments{
\item{decom}{Character: Name of decomposer to use.}

\item{k}{Integer: Number of dimensions to project to.}

\item{...}{Additional arguments to be passed to decomposer}
}
\description{
Set decomposition parameters for \link{train_cv} \code{.decompose} argument
}
