% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.earlystop}
\alias{setup.earlystop}
\title{Set \link{earlystop} parameters}
\usage{
setup.earlystop(
  window = 150,
  window_decrease_pct_min = 0.01,
  total_decrease_pct_max = NULL
)
}
\arguments{
\item{window}{Integer: Number of steps to consider}

\item{window_decrease_pct_min}{Float: Stop if improvement is less than this percent over last \code{window} steps}

\item{total_decrease_pct_max}{Float: Stop if improvement from first to last step exceeds this percent. If defined, overrides \code{window_decrease_pct_min}}
}
\description{
Set \link{earlystop} parameters
}
