% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.GBM}
\alias{setup.GBM}
\title{Set \link{s_GBM} parameters}
\usage{
setup.GBM(
  interaction.depth = 2,
  shrinkage = 0.001,
  max.trees = 5000,
  min.trees = 100,
  bag.fraction = 0.9,
  n.minobsinnode = 5,
  grid.resample.params = setup.resample("kfold", 5),
  ifw = TRUE,
  upsample = FALSE,
  downsample = FALSE,
  resample.seed = NULL,
  ...
)
}
\arguments{
\item{interaction.depth}{[gS] Integer: Interaction depth.}

\item{shrinkage}{[gS] Float: Shrinkage (learning rate).}

\item{max.trees}{Integer: Maximum number of trees to fit}

\item{min.trees}{Integer: Minimum number of trees to fit.}

\item{bag.fraction}{[gS] Float (0, 1): Fraction of cases to use to train each tree.
Helps avoid overfitting.}

\item{n.minobsinnode}{[gS] Integer: Minimum number of observation allowed in node.}

\item{grid.resample.params}{List: Output of \link{setup.resample} defining
grid search parameters.}

\item{ifw}{Logical: If TRUE, apply inverse frequency weighting
(for Classification only).
Note: If \code{weights} are provided, \code{ifw} is not used.}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Note: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{downsample}{Logical: If TRUE, downsample majority class to match size of minority class}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{...}{Additional arguments}
}
\description{
Set \link{s_GBM} parameters
}
