% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.LightRuleFit}
\alias{setup.LightRuleFit}
\title{Set \link{s_LightRuleFit} parameters}
\usage{
setup.LightRuleFit(
  n_trees = 200,
  num_leaves = 32L,
  max_depth = 3,
  learning_rate = 0.1,
  subsample = 0.666,
  subsample_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  objective = NULL,
  extra.lgbm.params = NULL,
  lightgbm.ifw = TRUE,
  lightgbm.resample.params = setup.resample("kfold", 5),
  glmnet.ifw = TRUE,
  importance = FALSE,
  alpha = 1,
  lambda = NULL,
  glmnet.resample.params = setup.resample("kfold", 5)
)
}
\arguments{
\item{lightgbm.ifw}{Logical: Passed to \link{s_LightGBM}'s \code{ifw} argument}

\item{glmnet.ifw}{Logical: Passed to \link{s_GLMNET}'s \code{ifw} argument}
}
\description{
Sets parameters for the GBM and LASSO steps of \link{s_LightRuleFit}
}
\author{
ED Gennatas
}
