% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.lincoef}
\alias{setup.lincoef}
\title{Set \link{lincoef} parameters}
\usage{
setup.lincoef(
  method = c("glmnet", "cv.glmnet", "lm.ridge", "allSubsets", "forwardStepwise",
    "backwardStepwise", "glm", "sgd", "solve"),
  alpha = 0,
  lambda = 0.01,
  lambda.seq = NULL,
  cv.glmnet.nfolds = 5,
  which.cv.glmnet.lambda = c("lambda.min", "lambda.1se"),
  nbest = 1,
  nvmax = 8,
  sgd.model = "glm",
  sgd.model.control = list(lambda1 = 0, lambda2 = 0),
  sgd.control = list(method = "ai-sgd")
)
}
\arguments{
\item{method}{Character: Method to use:
\itemize{
\item{"glm"}{: uses \code{stats::lm.wfit}}
\item{"glmnet"}{: uses \code{glmnet::glmnet}}
\item{"cv.glmnet"}{: uses \code{glmnet:cv.glmnet}}
\item{"lm.ridge"}{: uses \code{MASS::lm.ridge}}
\item{"allsubsets"}{: uses \code{leaps::regsubsets} with \code{method = "exhaustive"}}
\item{"forwardStepwise"}{: uses \code{leaps::regsubsets} with \code{method = "forward"}}
\item{"backwardStepwise"}{: uses \code{leaps::regsubsets} with \code{method = "backward"}}
\item{"sgd"}{: uses \code{sgd::sgd}}
\item{"solve"}{: uses \code{base::solve}}
\item{"none"}{: fits no model and returns all zeroes, for programming convenience in special cases}
}}

\item{alpha}{Float: \code{alpha} for method = \code{glmnet} or
\code{cv.glmnet}.}

\item{lambda}{Float: The lambda value for \code{glmnet}, \code{cv.glmnet},
\code{lm.ridge}
Note: For \code{glmnet} and \code{cv.glmnet}, this is the lambda used for
prediction. Training uses \code{lambda.seq}.}

\item{lambda.seq}{Float, vector: lambda sequence for \code{glmnet} and
\code{cv.glmnet}.}

\item{cv.glmnet.nfolds}{Integer: Number of folds for \code{cv.glmnet}}

\item{which.cv.glmnet.lambda}{Character: Whitch lambda to pick from
cv.glmnet: "lambda.min": Lambda that gives minimum cross-validated error;}

\item{nbest}{Integer: For \code{method = "allSubsets"}, number of subsets of
each size to record. Default = 1}

\item{nvmax}{Integer: For \code{method = "allSubsets"}, maximum number of
subsets to examine.}

\item{sgd.model}{Character: Model to use for \code{method = "sgd"}.}

\item{sgd.model.control}{List: \code{model.control} list to pass to
\code{sgd::sgd}}

\item{sgd.control}{List: \code{sgd.control} list to pass to \code{sgd::sgd}
"lambda.1se": Largest lambda such that error is within 1 s.e. of the minimum.}
}
\description{
Set \link{lincoef} parameters
}
