% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.MARS}
\alias{setup.MARS}
\title{Set \link{s_MARS} parameters}
\usage{
setup.MARS(
  hidden = 1,
  activation = NULL,
  learning.rate = 0.8,
  momentum = 0.5,
  learningrate_scale = 1,
  output = NULL,
  numepochs = 100,
  batchsize = NULL,
  hidden_dropout = 0,
  visible_dropout = 0,
  ...
)
}
\arguments{
\item{...}{Additional parameters to pass to \code{earth::earth}}
}
\description{
Set \link{s_MARS} parameters
}
