% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.meta.resample}
\alias{setup.meta.resample}
\title{Set \link{resample} parameters for meta model training}
\usage{
setup.meta.resample(
  resampler = "strat.sub",
  n.resamples = 4,
  stratify.var = NULL,
  train.p = 0.75,
  strat.n.bins = 4,
  target.length = NULL,
  verbosity = TRUE
)
}
\arguments{
\item{resampler}{Character: Type of resampling to perform: "bootstrap", "kfold",
"strat.boot", "strat.sub".}

\item{n.resamples}{Integer: Number of training/testing sets required}

\item{stratify.var}{Numeric vector (optional): Variable used for stratification.}

\item{train.p}{Float (0, 1): Fraction of cases to assign to traininig set for
\code{resampler = "strat.sub"}}

\item{strat.n.bins}{Integer: Number of groups to use for stratification for
\code{resampler = "strat.sub" / "strat.boot"}}

\item{target.length}{Integer: Number of cases for training set for
\code{resampler = "strat.boot"}.}

\item{verbosity}{Logical: If TRUE, print messages to console}
}
\description{
Set \link{resample} parameters for meta model training
}
