% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup.Ranger}
\alias{setup.Ranger}
\title{Set \link{s_Ranger} parameters}
\usage{
setup.Ranger(
  n.trees = 1000,
  min.node.size = 1,
  mtry = NULL,
  grid.resample.params = setup.resample("kfold", 5),
  ifw = TRUE,
  upsample = FALSE,
  downsample = FALSE,
  resample.seed = NULL,
  ...
)
}
\arguments{
\item{n.trees}{Integer: Number of trees to grow. Default = 1000}

\item{min.node.size}{[gS] Integer: Minimum node size}

\item{mtry}{[gS] Integer: Number of features sampled randomly at each split.
Defaults to square root of n of
features for classification, and a third of n of features for regression.}

\item{grid.resample.params}{List: Output of \link{setup.resample} defining
grid search parameters.}

\item{ifw}{Logical: If TRUE, apply inverse frequency weighting
(for Classification only).
Note: If \code{weights} are provided, \code{ifw} is not used.}

\item{upsample}{Logical: If TRUE, upsample training set cases not belonging in majority outcome
group}

\item{downsample}{Logical: If TRUE, downsample majority class to match size of minority class}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{...}{Additional arguments to be passed to \code{ranger::ranger}}
}
\description{
Set \link{s_Ranger} parameters
}
