% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_Resampler.R
\name{setup_Custom}
\alias{setup_Custom}
\title{Setup Custom Resampler}
\usage{
setup_Custom(resamples = NULL)
}
\arguments{
\item{resamples}{Optional List: Resamples to use, one integer vector of training-case indices
per resample.}
}
\value{
\code{CustomConfig} object.
}
\description{
Setup Custom Resampler
}
\examples{
custom_resampler <- setup_Custom(resamples = list(1:3, 2:4))
custom_resampler
}
\author{
EDG
}
