% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_Hyperparameters.R
\name{setup_GLM}
\alias{setup_GLM}
\title{Setup GLM Hyperparameters}
\usage{
setup_GLM(ifw = FALSE)
}
\arguments{
\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
GLMHyperparameters object.
}
\description{
Setup hyperparameters for GLM training.
}
\examples{
glm_hyperparams <- setup_GLM(ifw = TRUE)
glm_hyperparams
}
\author{
EDG
}
