% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_Resampler.R
\name{setup_LOOCV}
\alias{setup_LOOCV}
\title{Setup LOOCV Resampler}
\usage{
setup_LOOCV()
}
\value{
\code{LOOCVConfig} object.
}
\description{
Leave-one-out cross-validation: one resample per case. \code{n_resamples} is not
a parameter here because only the data can say how many cases there are --
\code{resample()} fills it in.
}
\examples{
loocv_resampler <- setup_LOOCV()
loocv_resampler
}
\author{
EDG
}
