% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120_ClusteringConfig.R
\name{setup_NeuralGas}
\alias{setup_NeuralGas}
\title{Setup NeuralGasConfig}
\usage{
setup_NeuralGas(k = 3L, dist = "euclidean", features = NULL)
}
\arguments{
\item{k}{Integer [1, Inf): Number of clusters.}

\item{dist}{Character \{"euclidean", "manhattan"\}: Distance measure to use.}

\item{features}{Optional Character vector: Names of at least 2 distinct
feature columns to cluster on. \code{NULL} clusters on all numeric features.}
}
\value{
NeuralGasConfig object.
}
\description{
Setup NeuralGasConfig
}
\examples{
neuralgas_config <- setup_NeuralGas(k = 4L, dist = "euclidean")
neuralgas_config
}
\author{
EDG
}
