% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150_DecompositionConfig.R
\name{setup_NMF}
\alias{setup_NMF}
\title{Setup NMF config.}
\usage{
setup_NMF(k = 2L, method = "brunet", nrun = 1L, features = NULL)
}
\arguments{
\item{k}{Integer [1, Inf): Number of components.}

\item{method}{Character: NMF method. See \code{NMF::nmf}.}

\item{nrun}{Integer [1, Inf): Number of runs to perform.}

\item{features}{Optional Character vector: Names of at least 2 distinct
feature columns to decompose. \code{NULL} decomposes all numeric features.}
}
\value{
NMFConfig object.
}
\description{
Setup NMF config.
}
\examples{
nmf_config <- setup_NMF(k = 3L)
nmf_config
}
\author{
EDG
}
