% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120_ClusteringConfig.R
\name{setup_Nystrom}
\alias{setup_Nystrom}
\title{Setup NystromConfig}
\usage{
setup_Nystrom(sample = NULL)
}
\arguments{
\item{sample}{Optional Integer [1, Inf): Number of cases the approximation
samples. Must be fewer than the number of cases. Unset samples one sixth of
them.}
}
\value{
\code{NystromConfig} object.
}
\description{
Setup the Nystrom approximation for \link{setup_SpectralRBF} or
\link{setup_SpectralLaplace}. Passing one turns the approximation on; its
\code{sample} says how many cases it samples.
}
\details{
\code{sample} maps to \code{kernlab::specc()}'s \code{nystrom.sample}; passing the config
at all sets \code{nystrom.red = TRUE}.
}
\examples{
setup_Nystrom(sample = 200L)
}
\author{
EDG
}
