% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_submit.R
\name{sge_submit}
\alias{sge_submit}
\title{Submit expression to SGE grid}
\usage{
sge_submit(
  expr,
  obj_names = NULL,
  packages = NULL,
  queue = NULL,
  n_threads = 4,
  sge_out = file.path(getwd(), "./sge_out"),
  sge_error = sge_out,
  sge_env = "#! /usr/bin/env bash",
  sge_opts = "#$ -cwd",
  R_command = NULL,
  system_command = NULL,
  h_rt = "00:25:00",
  mem_free = NULL,
  temp_dir = file.path(getwd(), ".sge_tempdir"),
  verbose = TRUE,
  trace = 1
)
}
\arguments{
\item{expr}{R expression}

\item{obj_names}{Character vector: Names of objects to copy to cluster R session}

\item{packages}{Character vector: Names of packages to load in cluster R session}

\item{queue}{Characer: Name of SGE queue to submit to}

\item{n_threads}{Integer: Number of threads to request from scheduler}

\item{sge_out}{Character: Path to directory to write standard out message files}

\item{sge_error}{Character: Path to directory to write error message files}

\item{sge_env}{Character: Shell environment for script to be submitted to SGE}

\item{sge_opts}{Character: SGE options that will be written in shell script. Default = "#$ -cwd"}

\item{system_command}{Character: system command to be run by shell script before executing R code.
For example a command that export the R executable to use}

\item{h_rt}{Character: Max time to request. Default = "00:25:00", i.e. 25 minutes}

\item{mem_free}{Character: Amount of memory to request from the scheduler}

\item{temp_dir}{Character: Temporary directory that is accessible to all execution nodes.
Default = \code{file.path(getwd(), ".sge_tempdir")}}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}

\item{trace}{Integer: If > 0 print diagnostic messages to console. Default = 1}
}
\description{
Submit expression to SGE grid
}
\author{
E.D. Gennatas
}
