% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis_color_system.R
\name{show_color_key}
\alias{show_color_key}
\title{Show rtemis color key}
\usage{
show_color_key()
}
\value{
Used for side-effects: prints color key to console. Returns invisible NULL.
}
\description{
Show rtemis color key
}
\examples{
show_color_key()
}
\author{
EDG
}
