% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shytreeleaves.R
\name{shytreeLeavesRC}
\alias{shytreeLeavesRC}
\title{\pkg{rtemis internal}: Low-level Stepwise Hybrid Tree procedure}
\usage{
shytreeLeavesRC(
  x,
  y,
  x.valid = NULL,
  y.valid = NULL,
  lookback = FALSE,
  weights = NULL,
  max.leaves = 5,
  nvmax = 2,
  select.leaves.smooth = TRUE,
  gamma = 0.1,
  alpha = 0,
  lambda = 0.01,
  lambda.seq = NULL,
  minobsinnode = 2,
  minobsinnode.lin = 10,
  learning.rate = 1,
  part.minsplit = 2,
  part.xval = 0,
  part.max.depth = 1,
  part.cp = 0,
  part.minbucket = 5,
  .rho = TRUE,
  rho.max = 1000,
  lin.type = "forwardStepwise",
  cv.glmnet.nfolds = 5,
  which.cv.glmnet.lambda = "lambda.min",
  loss.fn = if (is.factor(y)) class.loss else mse,
  verbose = TRUE,
  plot.tuning = TRUE,
  trace = 0
)
}
\arguments{
\item{x}{Data frame}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{lookback}{Logical: If TRUE, check validation error to decide best number of leaves to use.
Default = TRUE}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{max.leaves}{Integer: Total number of terminal nodes to reach. 1 is a special case where no split is performed and a linear
model is trained. Otherwise, this should be an even number as each split introduces two children nodes.
Note: this is total N of nodes in the tree, with the root uncounted, not the number of terminal nodes.}

\item{nvmax}{[gS] Integer: Number of max features to use for lin.type "allSubsets", "forwardStepwise", or
"backwardStepwise". If values greater than n of features in \code{x} are provided, they will be excluded}

\item{lambda}{Float: lambda parameter for \code{MASS::lm.ridge} Default = .01}

\item{minobsinnode}{Integer: Minimum N observations needed in node, before considering splitting}

\item{part.max.depth}{Integer: Max depth for each tree model within the additive tree}

\item{loss.fn}{Function with arguments y, Fval, weights. Allows you to define a custom loss function.}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{plot.tuning}{Logical: If TRUE, plot validation error during gridsearch}

\item{trace}{Integer: If higher than 0, will print more information to the console. Default = 0}
}
\description{
Train a Stepwise Hybrid Tree for Classification & Regression
}
\details{
With \code{max.nodes = 0}, the output is a linear model trained according to \code{lin.type}
Note that lambda is treated differently by \code{glmnet::glmnet} and \code{MASS::lm.ridge}
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
