% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortedlines.R
\name{sortedlines}
\alias{sortedlines}
\title{lines, but sorted}
\usage{
sortedlines(x, y, col = "red", ...)
}
\arguments{
\item{x}{Input vector}

\item{y}{Input vector}

\item{col}{Line color. Default = "red"}

\item{...}{Extra params to pass to \code{lines}}
}
\description{
lines, but sorted
}
\author{
E.D. Gennatas
}
