% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseVectorSummary.R
\name{sparseVectorSummary}
\alias{sparseVectorSummary}
\title{Sparseness and pairwise correlation of vectors}
\usage{
sparseVectorSummary(vectors, y = NULL)
}
\arguments{
\item{vectors}{Matrix of column vectors}

\item{y}{Optional numeric vector.}
}
\description{
Get sparseness measure on a matrix of vectors
}
