% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsify.R
\name{sparsify}
\alias{sparsify}
\title{Sparsify a vector}
\usage{
sparsify(x, sparseness)
}
\arguments{
\item{x}{Input vector}

\item{sparseness}{Percent of values of \code{x} to keep. The rest will be set to zero.}
}
\description{
Keep top x\% of values of a vector
}
\author{
E.D. Gennatas
}
