% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(true, predicted, harmonize = FALSE, verbosity = 1L)
}
\arguments{
\item{true}{True labels}

\item{predicted}{predicted labels}

\item{harmonize}{Logical: If TRUE, run \link{factor_harmonize} first}

\item{verbosity}{Integer: If > 0, print messages to console.}
}
\description{
The first factor level is considered the positive case.
}
\keyword{internal}
