% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shyoptleaves.R
\name{splitlin_}
\alias{splitlin_}
\title{\code{rtemis} internal: Ridge and Stump}
\usage{
splitlin_(
  g,
  type,
  node.index,
  gamma,
  n.quantiles,
  minobsinnode,
  minbucket,
  lin.type,
  alpha,
  lambda,
  lambda.seq,
  cv.glmnet.nfolds,
  which.cv.glmnet.lambda,
  nbest,
  nvmax,
  n.cores,
  trace
)
}
\arguments{
\item{node.index}{Open nodes to work on

Fit a linear model on (x, y) and split on the gradient
Input: environment holding tree and index of node
Output: None; Expands tree within environment g by splitting indexed node}

\item{tree}{Node within tree environment}
}
\description{
Edits environment 'g' in-place (no output)
}
