% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitline.R
\name{splitline}
\alias{splitline}
\title{Splitline}
\usage{
splitline(
  x,
  y,
  caseweights = rep(1, length(y)),
  gamma = 0.1,
  search = "quantile",
  n_quantiles = 20,
  minbucket = 5,
  alpha = 1,
  lambda = 0.1,
  trace = 0
)
}
\arguments{
\item{search}{String: "quantile" or "exhaustive". Default = "quantile"}

\item{n_quantiles}{Integer: Number of quantiles to use if \code{search = "quantile"}}

\item{trace}{Integer: If greater than 0, print diagnostic messages to console}
}
\description{
Perform split in a feature of x to minimize sum of group loss in y
}
\details{
This is an R port of splitline.jl from Rtemis.jl. It could be rewritten in a more
Rtistic style.
}
\examples{
\dontrun{
x = Array(rnormmat(1000, 20, seed = 2020))
y = x[:, 3] .+ x[:, 5] .^ 2 .+ randn(1000)
xysplit = splitline(x, y)
xysplit = splitline(x, y, lambda = [0.1])
}
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
