% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{sqcoldist}
\alias{sqcoldist}
\title{Squared Color Distance}
\usage{
sqcoldist(x, y)
}
\arguments{
\item{x}{Color}

\item{y}{Color}
}
\description{
Get the squared RGB distance between two colors
}
\examples{
sqcoldist("red", "green")
sqcoldist("#16A0AC", "#FA6E1E")
}
\author{
E.D. Gennatas
}
