% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderror.R
\name{stderror}
\alias{stderror}
\title{Standard Error of the Mean}
\usage{
stderror(x)
}
\arguments{
\item{x}{Vector, numeric: Input data}
}
\description{
Calculate the standard error of the mean, which is equal to the standard deviation divided by
the square root of the sample size. NA values are automatically removed
}
\author{
E.D. Gennatas
}
