% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata2factor.R
\name{strata2factor}
\alias{strata2factor}
\title{Convert \code{survfit} object's strata to a factor}
\usage{
strata2factor(x)
}
\arguments{
\item{x}{\code{survfit} object}
}
\value{
factor
}
\description{
Convert \code{survfit} object's strata to a factor
}
\author{
E.D. Gennatas
}
