% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict.R
\name{strict}
\alias{strict}
\title{Strict assignment by class or type}
\usage{
strict(x, accept.class = NULL, accept.type = NULL)
}
\arguments{
\item{x}{Value to be assigned if \code{type} matches}

\item{accept.class}{Required class of value}

\item{accept.type}{Required type of value}
}
\description{
Allow assignment only if input is of correct class and/or type
}
\author{
Efstathios D. Gennatas
}
