% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strng.R
\name{bold}
\alias{bold}
\alias{hilite}
\alias{red}
\alias{green}
\alias{orange}
\alias{cyan}
\alias{magenta}
\alias{gray}
\alias{reset}
\title{String formatting utilities}
\usage{
bold(...)

hilite(..., bold = TRUE)

red(..., bold = FALSE)

green(..., bold = FALSE)

orange(..., bold = FALSE)

cyan(..., bold = FALSE)

magenta(..., bold = FALSE)

gray(..., bold = FALSE)

reset(...)
}
\description{
String formatting utilities
}
