% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize numeric variables}
\usage{
summarize(
  x,
  varname,
  group_by = NULL,
  type = c("median-range", "mean-sd"),
  na.rm = TRUE
)
}
\arguments{
\item{x}{data.frame or data.table (will be coerced to data.table)}

\item{varname}{Character, vector: Variable name(s) to summarize. Must be column names in \code{x}
of type numeric.}

\item{group_by}{Character, vector: Variable name(s) of factors to group by. Must be column names
in \code{x}. Default = NULL}

\item{type}{Character: "median-range" or "mean-sd". Default = "median-range"}

\item{na.rm}{Logical: Passed to \code{median} and \code{mean}. Default = TRUE}
}
\value{
\code{data.table} with summary
}
\description{
Summarize numeric variables
}
\author{
E.D. Gennatas
}
