% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLAM.R
\name{summary.massGAM}
\alias{summary.massGAM}
\title{\code{massGAM} object summary}
\usage{
\method{summary}{massGAM}(object, ...)
}
\arguments{
\item{object}{A \code{massGAM} object created by \link{massGLAM}}

\item{...}{Not used}
}
\description{
\code{massGAM} object summary
}
\author{
E.D. Gennatas
}
