% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R, R/massGLM2.R
\name{summary.massGLM}
\alias{summary.massGLM}
\title{\code{massGLM} object summary}
\usage{
\method{summary}{massGLM}(object, ...)

\method{summary}{massGLM}(object, ...)
}
\arguments{
\item{object}{An object created by \link{massGLM}}

\item{...}{Not used}
}
\description{
\code{massGLM} object summary

\code{massGLM} object summary
}
\author{
E.D. Gennatas

E.D. Gennatas
}
