% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survError.R
\name{surv_error}
\alias{surv_error}
\title{Survival Analysis Metrics}
\usage{
surv_error(true, estimated)
}
\arguments{
\item{true}{Vector, numeric: True survival times}

\item{estimated}{Vector, numeric: Estimated survival times}
}
\description{
Survival Analysis Metrics
}
\author{
E.D. Gennatas
}
