% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd1.R
\name{svd1}
\alias{svd1}
\title{\code{rtemis-internals} Project Variables to First Eigenvector}
\usage{
svd1(x, x.test = NULL)
}
\arguments{
\item{x}{Input matrix / data frame}

\item{x.test}{Optional test matrix / data frame}
}
\description{
Convenience function for SVD k = 1
}
\author{
E.D. Gennatas
}
