% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data.R
\name{synth_multimodal}
\alias{synth_multimodal}
\title{Create "Multimodal" Synthetic Data}
\usage{
synth_multimodal(
  n_cases = 10000,
  init_fn = "runifmat",
  init_fn_params = list(min = -10, max = 10),
  n_groups = 4,
  n_feat_per_group = round(seq(10, 300, length.out = n_groups)),
  contrib_p = 0.33,
  linear_p = 0.66,
  square_p = 0.1,
  atan_p = 0.1,
  pair_multiply_p = 0.05,
  pair_square_p = 0.05,
  pair_atan_p = 0.05,
  verbosity = 1L,
  seed = NULL,
  filename = NULL
)
}
\arguments{
\item{n_cases}{Integer: Number of cases to create.}

\item{init_fn}{Character: "runifmat" or "rnormmat". Use the respective functions to
generate features as random uniform and random normal variables, respectively.}

\item{init_fn_params}{Named list with arguments "min", "max" for "runifmat" and
"mean", "sd" for "rnormmat".}

\item{n_groups}{Integer: Number of feature groups / modalities to create.}

\item{n_feat_per_group}{Integer, vector, length \code{n_groups}: Number of features per group to create.}

\item{contrib_p}{Float (0, 1]: Ratio of features contributing to outcome per group.
a third of the features in each group will be used to produce the outcome y}

\item{linear_p}{Float [0, 1]: Ratio of contributing features to be included linearly.
features in each group will be included}

\item{square_p}{Float [0, 1]: Ratio of contributing features to be squared.
in each group will be squared}

\item{atan_p}{Float [0, 1]: Ratio of contributing features whose \code{atan} will be used. These will be selected
from the features that were NOT sampled for squaring.
i.e. .1 of .33 of features in each group will be transformed using \code{atan}, but given these features were not
already picked to be squared (see \code{square_p})}

\item{pair_multiply_p}{Float [0, 1] Ratio of features will be divided into pairs and multiplied.}

\item{pair_square_p}{Float [0, 1] Ratio of features which will be divided into pairs, multiplied and squared.}

\item{pair_atan_p}{Float [0, 1] Ratio of features which will be divided into pairs, multiplied and transformed using
\code{atan}.}

\item{verbosity}{Integer: Verbosity level.}

\item{seed}{Integer: If set, pass to \code{set.seed} for reproducibility}

\item{filename}{Character: Path to file to save output.}
}
\value{
List with elements x, y, index_square, index_atan, index_pair_square
}
\description{
Create "Multimodal" Synthetic Data using squares and arctangents
}
\details{
There are no checks yet for compatibility among inputs and certain combinations may not work.
}
\examples{
\dontrun{
xmm <- synth_multimodal(
  n_cases = 10000,
  init_fn = "runifmat",
  init_fn_params = list(min = -10, max = 10),
  n_groups = 5,
  n_feat_per_group = c(20, 50, 100, 200, 300),
  contrib_p = .33,
  linear_p = .66,
  square_p = .1,
  atan_p = .1,
  pair_multiply_p = .1,
  pair_square_p = .1,
  pair_atan_p = .1,
  seed = 2019
)
}
}
\author{
EDG
}
