% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data.R
\name{synth_reg_data}
\alias{synth_reg_data}
\title{Synthesize Simple Regression Data}
\usage{
synth_reg_data(
  nrow = 500,
  ncol = 50,
  noise_sd_factor = 1,
  resampler_parameters = setup_Resampler(),
  seed = NULL,
  verbosity = 0L
)
}
\arguments{
\item{nrow}{Integer: Number of rows.}

\item{ncol}{Integer: Number of columns.}

\item{noise_sd_factor}{Numeric: Add rnorm(nrow, sd = noise_sd_factor * sd(y)).}

\item{resampler_parameters}{Output of \link{setup_Resampler} defining training/testing split. The first resulting resample
will be used to create \code{dat_training} and \code{dat_testing} output; all resample output under \code{resamples}}

\item{seed}{Integer: Seed for random number generator.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
List with elements \verb{dat, dat_training, dat_testing, resamples, w, seed}
}
\description{
Synthesize Simple Regression Data
}
\author{
EDG
}
