% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Table 1}
\usage{
table1(
  x,
  summaryFn1 = mean,
  summaryFn2 = sd,
  summaryFn1.extraArgs = list(na.rm = TRUE),
  summaryFn2.extraArgs = list(na.rm = TRUE),
  labelify = TRUE,
  verbose = TRUE,
  filename = NULL
)
}
\arguments{
\item{x}{data.frame or matrix: Input data, cases by features}

\item{summaryFn1}{Function: Summary function 1. Default = \code{mean}. See Details}

\item{summaryFn2}{Function: Summary function 2. Default = \code{sd}. See Details}

\item{labelify}{Logical: If TRUE, apply \link{labelify} to column names of  \code{x}}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\value{
A data.frame, invisibly, with two columns: "Feature", "Value mean (sd) | N"
}
\description{
Build Table 1. Subject characteristics
}
\details{
The output will look like "summaryFn1 (summaryFn2)".
Using defaults this will be "mean (sd)"
}
\author{
Efstathios D. Gennatas
}
