% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_blackgrid}
\alias{theme_blackgrid}
\title{Set theme for \code{mplot3} functions}
\usage{
theme_blackgrid(
  bg = "#000000",
  plot.bg = NA,
  fg = "#ffffff",
  pch = 16,
  cex = 1.2,
  lwd = 2,
  bty = "n",
  box.col = fg,
  box.alpha = 1,
  box.lty = 1,
  box.lwd = 0.5,
  grid = TRUE,
  grid.nx = NULL,
  grid.ny = NULL,
  grid.col = fg,
  grid.alpha = 0.2,
  grid.lty = 1,
  grid.lwd = 1,
  axes = TRUE,
  axes.col = NA,
  tick.col = NA,
  tick.alpha = 0,
  tick.labels.col = fg,
  x.axis.side = 1,
  y.axis.side = 2,
  labs.col = fg,
  zerolines = TRUE,
  zerolines.col = fg,
  zerolines.alpha = 0.5,
  zerolines.lty = 1,
  zerolines.lwd = 1,
  main.line = 0.5,
  main.adj = 0,
  main.font = 1,
  main.col = fg,
  font.family = ""
)
}
\description{
Set theme for \code{mplot3} functions
}
