% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeProc.R
\name{timeProc}
\alias{timeProc}
\title{Time a process}
\usage{
timeProc(...)
}
\arguments{
\item{...}{Command to be timed. (Will be converted using \code{as.expression})}
}
\description{
\code{timeProc} measures how long it takes for a process to run
}
\author{
Efstathios D. Gennatas
}
