% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{tohtml}
\alias{tohtml}
\title{Describe generic}
\usage{
tohtml(
  x,
  name = NULL,
  check_integers = FALSE,
  css = list(font.family = "Helvetica", color = "#fff", background.color = "#242424")
)
}
\description{
Describe \code{CheckData} object
}
