% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{toml_meta}
\alias{toml_meta}
\title{Write TOML metadata}
\usage{
toml_meta(x, schema_version = "1.0")
}
\arguments{
\item{x}{Object to create metadata for. Class name will be included in metadata.}
}
\description{
Creates named list which will become first TOML table in the following format:

\if{html}{\out{<div class="sourceCode toml">}}\preformatted{[_meta]
package = "rtemis"
package_version = "0.4.2"
schema_version = "1.0"
object_type = "SuperConfig"
created_at = 2026-2-11T22:45:00Z
}\if{html}{\out{</div>}}
}
