% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_SVM.R
\name{train_SVM}
\alias{train_SVM}
\title{Train a SVM model}
\usage{
train_SVM(
  x,
  weights = NULL,
  hyperparameters = NULL,
  tuner_parameters = NULL,
  verbosity = 1L
)
}
\arguments{
\item{x}{data.frame or similar: Training set.}

\item{weights}{Numeric vector: Case weights.}

\item{hyperparameters}{\code{SVMHyperparameters} object: make using \link{setup_RadialSVM}.}

\item{tuner_parameters}{\code{TunerParameters} object: make using \link{setup_GridSearch}.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Object of class \code{svm}.
}
\description{
Train a SVM model using \code{SVM}.
}
\details{
SVM does not work in the presence of missing values.
}
\author{
EDG
}
\keyword{internal}
