% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_TabNet.R
\name{train_TabNet}
\alias{train_TabNet}
\title{Train a TabNet model}
\usage{
train_TabNet(
  x,
  weights = NULL,
  hyperparameters = NULL,
  tuner_parameters = NULL,
  verbosity = 1L
)
}
\arguments{
\item{x}{data.frame or similar: Training set.}

\item{weights}{Numeric vector: Case weights.}

\item{hyperparameters}{\code{TabNetHyperparameters} object: make using \link{setup_TabNet}.}

\item{tuner_parameters}{\code{TunerParameters} object: make using \link{setup_GridSearch}.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Object of class \code{TabNet}.
}
\description{
Train a TabNet model using \code{TabNet}.
}
\details{
TabNet does not work in the presence of missing values.
}
\author{
EDG
}
\keyword{internal}
