% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modSelect.R
\name{tunable}
\alias{tunable}
\title{Print tunable hyperparameters for a supervised learning algorithm}
\usage{
tunable(
  alg = c("glmnet", "svm", "cart", "ranger", "gbm", "xgboost", "lightgbm")
)
}
\description{
Print tunable hyperparameters for a supervised learning algorithm
}
