% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{type_inspect}
\alias{type_inspect}
\title{Inspect character vector}
\usage{
type_inspect(x, xname = NULL, verbose = TRUE, thresh = 0.5)
}
\arguments{
\item{x}{Character vector}
}
\description{
Checks character vector to determine whether it might be best to convert to numeric.
}
\examples{
\dontrun{
x <- c("3", "5", "undefined", "21", "4", NA)
type_inspect(x)
z <- c("mango", "banana", "tangerine", NA)
type_inspect(z)
}
}
\author{
E.D. Gennatas
}
