% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.CMEANS.R
\name{u.CMEANS}
\alias{u.CMEANS}
\title{Fuzzy C-means Clustering}
\usage{
u.CMEANS(x, k = 2, iter.max = 100, dist = "euclidean",
  method = "cmeans", m = 2, rate.par = NULL, weights = 1,
  control = list(), verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{iter.max}{Integer: Maximum number of iterations}

\item{dist}{String: Distance measure to use: 'euclidean' or 'manhattan'}

\item{method}{String: "cmeans" - fuzzy c-means clustering; "ufcl": on-line update}

\item{m}{Float (>1): Degree of fuzzification. Default = 2}

\item{rate.par}{Float (0, 1): Learning rate for the online variant. (Default = .3)}

\item{weights}{Float (>0): Case weights}

\item{control}{List of control parameters. See \code{e1071::cmeans}}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{e1071::cmeans}}
}
\value{
\link{rtClust} object
}
\description{
Perform fuzzy C-means clustering
}
\seealso{
Other Clustering: \code{\link{u.EMC}},
  \code{\link{u.H2OKMEANS}}, \code{\link{u.HARDCL}},
  \code{\link{u.HOPACH}}, \code{\link{u.KMEANS}},
  \code{\link{u.NGAS}}, \code{\link{u.PAMK}},
  \code{\link{u.PAM}}, \code{\link{u.SPEC}}
}
\author{
Efstathios D. Gennatas
}
\concept{Clustering}
