% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.DBSCAN.R
\name{u.DBSCAN}
\alias{u.DBSCAN}
\title{Density-based spatial clustering of applications with noise}
\usage{
u.DBSCAN(
  x,
  x.test = NULL,
  eps = 1,
  minPts = NCOL(x) + 1,
  weights = NULL,
  borderPoints = TRUE,
  search = c("kdtree", "linear", "dist"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform DBSCAN clustering
}
\details{
See \code{dbscan::dbscan} for info on how to choose \code{eps} and 
\code{minPts}
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.EMC}()},
\code{\link{u.H2OKMEANS}()},
\code{\link{u.HARDCL}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.MEANSHIFT}()},
\code{\link{u.NGAS}()},
\code{\link{u.PAMK}()},
\code{\link{u.PAM}()},
\code{\link{u.SPEC}()}
}
\author{
Efstathios D. Gennatas
}
\concept{Clustering}
