% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.EMC.R
\name{u.EMC}
\alias{u.EMC}
\title{Expectation Maximization Clustering}
\usage{
u.EMC(
  x,
  x.test = NULL,
  k = 2,
  lab = NULL,
  EMC = EMCluster::.EMC,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{lab}{Vector, length \code{NROW(x)}: Labels for semi-supervised clustering}

\item{EMC}{List of control parameters for \code{EMCluster::emcluster}. Default = \code{EMCluster::.EMC}}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{EMCluster::emcluster}}
}
\description{
Perform EM Clustering using \code{EMCluster::emcluster}
}
\details{
First, \code{EMCluster::simple.init(x, nclass = k)} is run,
followed by \code{EMCluster::emcluster(x, emobj = emobj, assign.class = TRUE, ...)}

This can be very slow.
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.H2OKMEANS}()},
\code{\link{u.HARDCL}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.NGAS}()},
\code{\link{u.PAMK}()},
\code{\link{u.PAM}()},
\code{\link{u.SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
