% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.H2OKMEANS.R
\name{u.H2OKMEANS}
\alias{u.H2OKMEANS}
\title{K-Means Clustering on H2O}
\usage{
u.H2OKMEANS(
  x,
  x.test = NULL,
  k = 2,
  estimate.k = FALSE,
  nfolds = 0,
  max.iterations = 10,
  ip = "localhost",
  port = 54321,
  n.cores = rtCores,
  seed = -1,
  init = c("Furthest", "Random", "PlusPlus", "User"),
  categorical.encoding = c("AUTO", "Enum", "OneHotInternal", "OneHotExplicit",
    "Binary", "Eigen", "LabelEncoder", "SortByResponse", "EnumLimited"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{estimate.k}{Logical: if TRUE, estimate k up to a maximum set by the \code{k} argument}

\item{nfolds}{Integer: Number of cross-validation folds}

\item{max.iterations}{Integer: Maximum number of iterations}

\item{ip}{Character: IP address of H2O server. Default = "localhost"}

\item{port}{Integer: Port number of H2O server. Default = 54321}

\item{n.cores}{Integer: Number of cores to use}

\item{seed}{Integer: Seed for H2O's random number generator. Default = -1 (time-based ranodm number)}

\item{init}{Character: Initialization mode: "Furthest", "Random", "PlusPlus", "User".
Default = "Furthest"}

\item{categorical.encoding}{Character: How to encode categorical variables: "AUTO", "Enum", "OneHotInternal",
"OneHotExplicit", "Binary", "Eigen", "LabelEncoder", "SortByResponse", "EnumLimited".
Default = "AUTO"}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional arguments to pass to \code{h2p::h2o.kmeans}}
}
\value{
\link{rtMod} object
}
\description{
K-Means clustering using \code{h2o::h2o.kmeans}
Check out the H2O Flow at \code{[ip]:[port]}, Default IP:port is "localhost:54321"
e.g. if running on localhost, point your web browser to \code{localhost:54321}
}
\details{
For additional information, see help on \code{h2o::h2o.kmeans}
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.EMC}()},
\code{\link{u.HARDCL}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.NGAS}()},
\code{\link{u.PAMK}()},
\code{\link{u.PAM}()},
\code{\link{u.SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
