% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.HARDCL.R
\name{u.HARDCL}
\alias{u.HARDCL}
\title{Clustering by Hard Competitive Learning}
\usage{
u.HARDCL(x, x.test = NULL, k = 2, dist = "euclidean", verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform clustering by Hard Competitive Learning using \code{flexclust::cclust}
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.EMC}()},
\code{\link{u.H2OKMEANS}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.NGAS}()},
\code{\link{u.PAMK}()},
\code{\link{u.PAM}()},
\code{\link{u.SPEC}()}
}
\author{
Efstathios D. Gennatas
}
\concept{Clustering}
