% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.HOPACH.R
\name{u.HOPACH}
\alias{u.HOPACH}
\title{Hieararchical Ordered Partitioning and Collapsing Hybrid}
\usage{
u.HOPACH(x, dmat = NULL, metric = c("cosangle", "abscosangle",
  "euclid", "abseuclid", "cor", "abscor"), K = 15, kmax = 9,
  khigh = 9, verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{dmat}{Matrix (numeric, no missing values) or \code{hdist} object of pairwise distances.
If NULL}

\item{metric}{Character: Dissimilarity metric to be used. Options: 'euclidean', 'manhattan'}

\item{K}{Integer, (0:15]: Maximum number of levels}

\item{kmax}{Integer, [1:9]: Maximum number of children at each node in the tree}

\item{khigh}{Integer, [1:9]: Maximum number of children at each nod ein the tree when computing the
the Mean/Median Split Silhouette. Usually same as \code{kmax}}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{cluster::hopach}}
}
\description{
Perform HOPACH clustering using \code{hopach::hopach}
}
\seealso{
Other Clustering: \code{\link{u.CMEANS}},
  \code{\link{u.EMC}}, \code{\link{u.H2OKMEANS}},
  \code{\link{u.HARDCL}}, \code{\link{u.KMEANS}},
  \code{\link{u.NGAS}}, \code{\link{u.PAMK}},
  \code{\link{u.PAM}}, \code{\link{u.SPEC}}
}
\author{
Efstathios D. Gennatas
}
\concept{Clustering}
