% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.MEANSHIFT.R
\name{u.MEANSHIFT}
\alias{u.MEANSHIFT}
\title{Mean Shift Clustering}
\usage{
u.MEANSHIFT(
  x,
  nNeighbors = NROW(x),
  algorithm = c("LINEAR", "KDTREE"),
  kernelType = c("NORMAL", "EPANECHNIKOV", "BIWEIGHT"),
  bandwidth = rep(1, NCOL(x)),
  alpha = 0,
  iterations = 10,
  epsilon = 1e-08,
  epsilonCluster = 1e-04,
  parameters = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform Mean Shift clustering using \code{meanShiftR::meanShift}
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.DBSCAN}()},
\code{\link{u.EMC}()},
\code{\link{u.H2OKMEANS}()},
\code{\link{u.HARDCL}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.NGAS}()},
\code{\link{u.PAMK}()},
\code{\link{u.PAM}()},
\code{\link{u.SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
