% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.NGAS.R
\name{u.NGAS}
\alias{u.NGAS}
\title{Neural Gas Clustering}
\usage{
u.NGAS(x, x.test = NULL, k = 2, dist = "euclidean", verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Testing set matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\value{
\link{rtClust} object
}
\description{
Perform Neural Gas clustering using \code{flexclust::cclust}
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.EMC}()},
\code{\link{u.H2OKMEANS}()},
\code{\link{u.HARDCL}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.PAMK}()},
\code{\link{u.PAM}()},
\code{\link{u.SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
