% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.PAM.R
\name{u.PAM}
\alias{u.PAM}
\title{Partitioning Around Medoids}
\usage{
u.PAM(
  x,
  k = 2,
  diss = FALSE,
  metric = "euclidean",
  do.swap = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{k}{Integer: Number of clusters to get}

\item{diss}{Logical: If TRUE, \code{x} should be a \code{dist} or dissimilarity matrix.
Otherwise, \code{x} should be a matrix of cases by features. Default = FALSE}

\item{metric}{Character: Dissimilarity metric to be used. Options: 'euclidean', 'manhattan'}

\item{do.swap}{Logical: If TRUE, perform the swap phase (See \code{cluster::pam}), as in the
original PAM algorithm. This is computationally intensive and can be skipped. Default = TRUE}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{cluster::pam}}
}
\description{
Perform PAM clustering using \code{cluster::pam}
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.DBSCAN}()},
\code{\link{u.EMC}()},
\code{\link{u.H2OKMEANS}()},
\code{\link{u.HARDCL}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.MEANSHIFT}()},
\code{\link{u.NGAS}()},
\code{\link{u.PAMK}()},
\code{\link{u.SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
