% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.PAMK.R
\name{u.PAMK}
\alias{u.PAMK}
\title{Partitioning Around Medoids with k Estimation}
\usage{
u.PAMK(
  x,
  krange = 2:10,
  criterion = "asw",
  usepam = ifelse(nrow(x) < 2000, TRUE, FALSE),
  scaling = TRUE,
  diss = inherits(data, "dist"),
  metric = "euclidean",
  do.swap = TRUE,
  trace = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{krange}{Integer vector: Range of k values to try}

\item{criterion}{Character: Criterion to use for selecting k: "asw", "multiasw" or "ch". See \code{fpc::pamk}}

\item{usepam}{Logical: If TRUE, use \code{cluster::pam}, otherwise use \code{cluster::clara}.
Default = TRUE}

\item{scaling}{Logical or Numeric vector: If TRUE, scale input. If numeric vector of length equal to number of
features, the features are divided by the corresponding value. Default = TRUE}

\item{diss}{Logical: If TRUE, treat \code{x} as a dissimilarity matrix, otherwise as a matrix of
cases by features. Default = TRUE, if x inherits from class \code{dist}, FALSE otherwise.}

\item{metric}{Character: Dissimilarity metric to be used. Options: 'euclidean', 'manhattan'}

\item{do.swap}{Logical: If TRUE, perform the swap phase (See \code{cluster::pam}), as in the
original PAM algorithm. This is computationally intensive and can be skipped. Default = TRUE}

\item{trace}{Integer [0, 3]: Trace level for \code{fpc::pamk}}

\item{verbose}{Logical: If TRUE, print messages to screen}

\item{...}{Additional parameters to be passed to \code{fpc::pamk} and/or \code{cluster::pam}}
}
\value{
\link{rtClust} object
}
\description{
Estimate PAM clustering solution and optimal k using \code{fpc::pamk}
}
\seealso{
Other Clustering: 
\code{\link{u.CMEANS}()},
\code{\link{u.EMC}()},
\code{\link{u.H2OKMEANS}()},
\code{\link{u.HARDCL}()},
\code{\link{u.HOPACH}()},
\code{\link{u.KMEANS}()},
\code{\link{u.NGAS}()},
\code{\link{u.PAM}()},
\code{\link{u.SPEC}()}
}
\author{
Efstathios D. Gennatas
}
\concept{Clustering}
