% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uci_heart_failure}
\alias{uci_heart_failure}
\title{UCI Heart Failure Data}
\format{
\subsection{\code{uci_heart_failure}}{

A data frame with 299 rows and 13 columns:
\describe{
\item{DEATH_EVENT}{Boolean: If the patient died during the follow-up period}
}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/519/heart+failure+clinical+records}
}
\usage{
uci_heart_failure
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
