% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_ops.R
\name{uniquevalsperfeat}
\alias{uniquevalsperfeat}
\title{Unique values per feature}
\usage{
uniquevalsperfeat(x, excludeNA = FALSE)
}
\arguments{
\item{x}{matrix or data frame input}

\item{excludeNA}{Logical: If TRUE, exclude NA values from unique count.}
}
\value{
Vector, integer of length \code{NCOL(x)} with number of unique
values per column/feature
}
\description{
Get number of unique values per features
}
\examples{
\dontrun{
uniquevalsperfeat(iris)
}
}
\author{
E.D. Gennatas
}
