% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtboost.R
\name{update.addtboost}
\alias{update.addtboost}
\title{Update \link{boost} object's fitted values}
\usage{
\method{update}{addtboost}(object, x, y = NULL, trace = 0, ...)
}
\arguments{
\item{object}{\link{addtboost} object}
}
\value{
\link{addtboost} object
}
\description{
Calculate new fitted values for a \link{boost}.
Advanced use only: run after updating learning.rate
}
\details{
All this will eventually be automated using an R6 object, maybe
}
\author{
Efstathios D. Gennatas
}
