% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmLiteBoostTV.R
\name{update.glmLiteBoostTV}
\alias{update.glmLiteBoostTV}
\title{\pkg{rtemis} internals: Update \link{glmLiteBoostTV} object's fitted values}
\usage{
\method{update}{glmLiteBoostTV}(object, x = NULL, x.valid = NULL,
  trace = 0, last.step.only = FALSE, n.cores = rtCores, ...)
}
\arguments{
\item{object}{\link{glmLiteBoostTV} object}

\item{x}{Data frame: Features}

\item{last.step.only}{Logicall: If TRUE, \code{x} must be provided and only the last meta model will be updated
using this \code{x}}
}
\value{
\link{glmLiteBoostTV} object

Nothing; updates \code{object} in-place
}
\description{
Calculate new fitted values for a \link{glmLiteBoostTV} object.
Advanced use only: run with new \code{x} or after updating learning.rate in object
}
\author{
Efstathios D. Gennatas
}
