% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboostnow.R
\name{update.hytboostnow}
\alias{update.hytboostnow}
\title{Update \link{boost} object's fitted values}
\usage{
\method{update}{hytboostnow}(object, x, y = NULL, trace = 0, ...)
}
\arguments{
\item{object}{\link{hytboostnow} object}
}
\value{
\link{hytboostnow} object
}
\description{
Calculate new fitted values for a \link{boost}.
Advanced use only: run after updating learning.rate
}
\details{
All this will eventually be automated using an R6 object, maybe
}
\author{
E.D. Gennatas
}
