% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{update.rtMod.boost}
\alias{update.rtMod.boost}
\title{\pkg{rtemis} internals: Update \code{rtMod} \link{boost} object's fitted values in-place}
\usage{
\method{update}{rtMod.boost}(object, x = NULL, x.test = NULL,
  trace = 0, last.step.only = FALSE, n.cores = rtCores, ...)
}
\arguments{
\item{object}{\link{boost} object}

\item{x}{Data frame: Features}

\item{last.step.only}{Logical: If TRUE, \code{x} must be provided and only the last meta model will be updated
using this \code{x}}
}
\value{
\link{boost} object

Nothing; updates \code{object} in-place
}
\description{
Calculate new fitted values for a \link{boost} object.
Advanced use only: run with new \code{x} or after updating learning.rate in object
}
\author{
Efstathios D. Gennatas
}
