% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_SVM.R
\name{varimp_SVM}
\alias{varimp_SVM}
\title{Get coefficients from SVM model}
\usage{
varimp_SVM(model)
}
\arguments{
\item{model}{SVM model.}
}
\description{
Get coefficients from SVM model
}
\keyword{internal}
