% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_TabNet.R
\name{varimp_TabNet}
\alias{varimp_TabNet}
\title{Get coefficients from TabNet model}
\usage{
varimp_TabNet(model)
}
\arguments{
\item{model}{TabNet model.}
}
\description{
Get coefficients from TabNet model
}
\keyword{internal}
