% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelect.R
\name{varSelect}
\alias{varSelect}
\title{Variable Selection by Variable Importace}
\usage{
varSelect(x, y, method = c("RANGER", "XGBLIN"), xgb.params = list(alpha
  = 0.1, lambda = 0.1), p = 0.2, print.plot = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{Matrix / Data Frame of Predictors}

\item{y}{Outcome vector}

\item{method}{Character: "RANGER", "XGBLIN": Learner to use for estimating variable importace. Default = "RANGER"}

\item{xgb.params}{List of parameters for \code{method = "XGBLIN"}}

\item{p}{Float (0, 1): Fraction of variables in x to select. \code{p * ncol(x)}. May help to set to a fraction twice
what you expect to be the true fraction of useful variables, to reduce false negatives at the expense of false
positives which can be dealt by an appropriate learning algorithm. (Default = .2)}

\item{print.plot}{Logical: If TRUE, print index plot of variable importance using \link{mplot3.x}}

\item{verbose}{Logical: If TRUE, print messages to screen}
}
\description{
Select important variables from a set of features based on RANGER- or XGBLIN-estimated variable importance
}
\author{
Efstathios D. Gennatas
}
