% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{vec2df}
\alias{vec2df}
\title{Vector to data.frame}
\usage{
vec2df(x, col_names = NULL)
}
\arguments{
\item{x}{Vector.}

\item{col_names}{Character: Name of the vector.}
}
\value{
data.frame.
}
\description{
Convert vector to 1-row data.frame, maintaining names if present
}
\author{
EDG
}
