% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize vector}
\usage{
winsorize(
  x,
  lo = NULL,
  hi = NULL,
  prob_lo = 0.025,
  prob_hi = 0.975,
  quantile_type = 7,
  verbosity = 1L
)
}
\arguments{
\item{x}{Numeric vector: Input data}

\item{lo}{Numeric: If not NULL, replace any values in \code{x} lower than
this with this.}

\item{hi}{Numeric: If not NULL, replace any values in \code{x} higher than
this with this.}

\item{prob_lo}{Numeric (0, 1): If not NULL and \code{lo = NULL}, find sample
quantile that corresponds to this probability and set as \code{lo}.}

\item{prob_hi}{Numeric (0, 1): If not NULL and \code{hi = NULL}, find sample
quantile that corresponds to this probability and set as \code{hi}.}

\item{quantile_type}{Integer: passed to \code{stats::quantile}}

\item{verbosity}{Integer: Verbosity level.}
}
\description{
Replace extreme values by absolute or quantile threshold
}
\details{
If both lo and prob_lo or both hi and prob_hi are NULL, cut-off is set to min(x) and max(x) respectively, i.e.
no values are changed
}
\examples{
\dontrun{
# Winsorize a normally distributed variable
x <- rnorm(500)
xw <- winsorize(x)
# Winsorize an exponentially distributed variable only on
# the top 5\% highest values
x <- rexp(500)
xw <- winsorize(x, prob_lo = NULL, prob_hi = .95)
}
}
\author{
EDG
}
