% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.CCA.R
\name{x.CCA.permute.both}
\alias{x.CCA.permute.both}
\title{modified PMA:::CCA.permute.both for parallel execution}
\usage{
x.CCA.permute.both(x, z, typex, typez, penaltyxs, penaltyzs, niter, v,
  trace, nperms, standardize, chromx, chromz, upos, uneg, vpos, vneg,
  outcome, y, cens, verbose = TRUE, n.cores = NULL,
  parallel.type = ifelse(.Platform$OS.type == "unix", "fork", "psock"))
}
\description{
Run PMA::CCA.permute.both permutations in parallel
}
\author{
adapted by Efstathios D Gennatas; original authors: Daniela M Witten, Robert Tibshirani
}
\keyword{internal}
