% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.CCA.R
\name{x.CCA.permute}
\alias{x.CCA.permute}
\title{modified PMA::CCA.permute for parallel execution}
\usage{
x.CCA.permute(x, z, typex = c("standard", "ordered"),
  typez = c("standard", "ordered"), penaltyxs = NULL,
  penaltyzs = NULL, niter = 3, v = NULL, trace = TRUE,
  nperms = 25, standardize = TRUE, chromx = NULL, chromz = NULL,
  upos = FALSE, uneg = FALSE, vpos = FALSE, vneg = FALSE,
  outcome = NULL, y = NULL, cens = NULL, verbose = TRUE,
  n.cores = rtCores)
}
\description{
Run PMA::CCA.permute permutation in parallel
}
\author{
adapted by Efstathios D Gennatas; original authors: Daniela M. Witten, Robert Tibshirani
}
\keyword{internal}
