% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_CCA.R
\name{x_CCA.permute.both}
\alias{x_CCA.permute.both}
\title{modified PMA:::CCA.permute.both for parallel execution}
\usage{
x_CCA.permute.both(
  x,
  z,
  typex,
  typez,
  penaltyxs,
  penaltyzs,
  niter,
  v,
  trace,
  nperms,
  standardize,
  chromx,
  chromz,
  upos,
  uneg,
  vpos,
  vneg,
  outcome,
  y,
  cens,
  verbose = TRUE,
  n.cores = rtCores,
  parallel.type = ifelse(.Platform$OS.type == "unix", "fork", "psock")
)
}
\description{
Run PMA::CCA.permute.both permutations in parallel
}
\author{
adapted by E.D. Gennatas; original authors: Daniela M Witten, Robert Tibshirani
}
\keyword{internal}
