% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_CCA.R
\name{x_CCA}
\alias{x_CCA}
\title{Sparse Canonical Correlation Analysis (CCA)}
\usage{
x_CCA(
  x,
  z,
  x.test = NULL,
  z.test = NULL,
  y = NULL,
  outcome = NULL,
  k = 3,
  niter = 20,
  nperms = 50,
  permute.niter = 15,
  typex = "standard",
  typez = "standard",
  penaltyx = NULL,
  penaltyz = NULL,
  standardize = TRUE,
  upos = FALSE,
  vpos = FALSE,
  verbose = TRUE,
  n.cores = rtCores,
  outdir = NULL,
  save.mod = ifelse(!is.null(outdir), TRUE, FALSE),
  ...
)
}
\arguments{
\item{x}{Matrix: Training x dataset}

\item{z}{Matrix: Training z dataset}

\item{x.test}{Matrix (Optional): Testing x set}

\item{z.test}{Matrix (Optional): Testing z set}

\item{y}{Outcome vector (Optional): If supplied, linear combinations of \code{x} and \code{z} need to be
additionaly correlated with this}

\item{outcome}{Character: Type of outcome \code{y}: "survival", "multiclass", "quantitative"}

\item{k}{Integer: Number of components}

\item{niter}{Integer: Number of iterations}

\item{nperms}{Integer: Number of permutations to run with \code{CCA.permute}
The higher, the better.}

\item{permute.niter}{Integer: Number of iterations to run for each permutation with \code{CCA.permute}}

\item{typex}{Character: "standard", "ordered". Use "standard" if columns of x are unordered; lasso
penalty is applied to enforce sparsity. Otherwise, use "ordered"; fused lasso penalty is applied,
to enforce both sparsity and smoothness.}

\item{typez}{Character: "standard", "ordered". Same as \code{typex} for z dataset}

\item{penaltyx}{Float: The penalty to be applied to the matrix x, i.e. the penalty that results
in the canonical vector u. If typex is "standard" then the L1 bound on u is
penaltyx*sqrt(ncol(x)). In this case penaltyx must be between 0 and 1 (larger L1 bound
corresponds to less penalization). If "ordered" then it's the fused lasso penalty lambda, which
must be non-negative (larger lambda corresponds to more penalization).}

\item{penaltyz}{Float: The penalty to be applied to the matrix z, i.e. the penalty that results
in the canonical vector v. If typez is "standard" then the L1 bound on v is
penaltyz*sqrt(ncol(z)). In this case penaltyz must be between 0 and 1 (larger L1 bound
corresponds to less penalization). If "ordered" then it's the fused lasso penalty lambda, which
must be non-negative (larger lambda corresponds to more penalization).}

\item{standardize}{Logical: If TRUE, center and scale columns of \code{x} and \code{z}}

\item{upos}{Logical: Require elements of u to be positive}

\item{vpos}{Logical: Require elements of v to be positive}

\item{verbose}{Logical: Print messages, including \code{trace} from \code{x_CCA.permute} and
\code{PMA::CCA}}

\item{n.cores}{Integer: Number of cores to use}

\item{outdir}{Path to output directory. Default = NULL}

\item{save.mod}{Logical: If TRUE, and \code{outdir} is defined, will save trained CCA model to \code{outdir}.
Default = TRUE if \code{outdir} is set, otherwise FALSE}

\item{...}{Additional arguments to be passed to \code{PMA::CCA}}
}
\description{
Run a sparse Canonical Correlation Analysis using the \code{PMA} package
}
\details{
#' \code{x_CCA} runs \code{PMA::CCA}. If penaltyx is NULL, penaltyx \emph{and} penaltyz will be estimated automatically
using x_CCA.permute (adapted to run in parallel)
}
\seealso{
Other Cross-Decomposition: 
\code{\link{xselect_decom}()}
}
\author{
E.D. Gennatas
}
\concept{Cross-Decomposition}
