% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdecomSelect.R
\name{xdecomSelect}
\alias{xdecomSelect}
\title{Select \pkg{rtemis} cross-decomposer}
\usage{
xdecomSelect(xdecom, fn = FALSE, desc = FALSE)
}
\arguments{
\item{xdecom}{Character: Cross-decomposition name; case insensitive}

\item{fn}{Logical: If TRUE, return function, otherwise return name of function. Default = FALSE}

\item{desc}{Logical: If TRUE, return full name of algorithm. Default = FALSE}
}
\value{
Function or name of function (see param \code{fn}) or full name of algorithm (\code{desc})
}
\description{
Accepts decomposer name (supports abbreviations) and returns \pkg{rtemis} function name or
  the function itself.
  If run with no parameters, prints list of available algorithms.
}
\seealso{
Other Cross-Decomposition: 
\code{\link{x_CCA}()}
}
\author{
E.D. Gennatas
}
\concept{Cross-Decomposition}
