% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx2list.R
\name{xlsx2list}
\alias{xlsx2list}
\title{Read all sheets of an XLSX file into a list}
\usage{
xlsx2list(
  x,
  sheet = NULL,
  startRow = 1,
  colNames = TRUE,
  na.strings = "NA",
  detectDates = TRUE,
  skipEmptyRows = TRUE,
  skipEmptyCols = TRUE
)
}
\arguments{
\item{x}{Character: path or URL to XLSX file}

\item{sheet}{Integer, vector: Sheet(s) to read. If NULL, will read all
sheets in \code{x}}

\item{startRow}{Integer, vector: First row to start reading. Will be
recycled as needed for all sheets}

\item{colNames}{Logical: If TRUE, use the first row of data}

\item{na.strings}{Character vector: stringd to be interpreted as NA}

\item{detectDates}{Logical: If TRUE, try to automatically detect dates}

\item{skipEmptyRows}{Logical: If TRUE, skip empty rows}

\item{skipEmptyCols}{Logical: If TRUE, skip empty columns}
}
\description{
Read all sheets of an XLSX file into a list
}
\author{
E.D. Gennatas
}
