% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtdescribe.R
\name{xtdescribe}
\alias{xtdescribe}
\title{Describe longitudinal dataset}
\usage{
xtdescribe(x, ID_col = 1, time_col = 2, n_patterns = 9)
}
\arguments{
\item{x}{data.frame with longitudinal data}

\item{ID_col}{Integer: The column position of the ID variable}

\item{time_col}{Integer: The column position of the time variable}

\item{n_patterns}{Integer: The number of patterns to display}
}
\description{
This is a test to emulate the \code{xtdescribe} function in Stata.
}
\author{
EDG
}
