% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotops.R
\name{yhi}
\alias{yhi}
\title{Get y above current plot area}
\usage{
yhi(pct_higher = 0.08)
}
\arguments{
\item{pct_higher}{Numeric: Get y this percent above top of plot}
}
\description{
Get y above current plot area
}
\author{
E.D. Gennatas
}
