% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ops.R
\name{ylo}
\alias{ylo}
\title{Get y below current plot area}
\usage{
ylo(pct_lower = 0.08)
}
\arguments{
\item{pct_lower}{Numeric: Get y this percent below bottom of plot}
}
\description{
Get y below current plot area
}
\author{
E.D. Gennatas
}
