% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_ops.R
\name{zip2longlat}
\alias{zip2longlat}
\title{Get Longitude and Lattitude for zip code(s)}
\usage{
zip2longlat(x, zipdt)
}
\arguments{
\item{x}{Character vector: Zip code(s)}

\item{zipdt}{data.table with "zip", "lng", and "lat" columns}
}
\description{
Returns a data.table of longitude and lattitude for one or more zip codes,
given an input dataset
}
