% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{any_constant}
\alias{any_constant}
\title{Check for constant columns}
\usage{
any_constant(x)
}
\arguments{
\item{x}{Input Data Frame}
}
\value{
Logical.
}
\description{
Checks if any column of a data frame have zero variance
}
\author{
EDG
}
