% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{any_constant}
\alias{any_constant}
\title{Check for constant columns}
\usage{
any_constant(x)
}
\arguments{
\item{x}{Input Data Frame}
}
\description{
Checks if any column of a data frame have zero variance
}
\author{
E.D. Gennatas
}
