% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_Decomposition.R
\name{apply_decomp}
\alias{apply_decomp}
\title{Apply Decomposition to New Data}
\usage{
apply_decomp(decom, new_data, verbosity = 1L)
}
\arguments{
\item{decom}{Decomposition object.}

\item{new_data}{Tabular data (data.frame, data.table, or tibble): New data to which the
decomposition will be applied.}

\item{verbosity}{Integer: Verbosity level}
}
\value{
Transformed data: a matrix of the learned components for \code{new_data}.
}
\description{
Apply a fitted decomposition to new data for algorithms that support this.
}
\author{
EDG
}
