% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{apply_preprocessor}
\alias{apply_preprocessor}
\title{Apply trained Preprocessor to new data}
\usage{
apply_preprocessor(preprocessor, new_data, verbosity = 1L)
}
\arguments{
\item{preprocessor}{\code{Preprocessor}: Trained preprocessor, i.e. the output of \link{preprocess}.}

\item{new_data}{Tabular data, i.e. data.frame, data.table, or tbl_df (tibble):
New data to preprocess.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Preprocessed data, in the same structure as \code{new_data}.
}
\description{
Apply a trained \code{Preprocessor} to new data, reusing the values learned from the training
data. For example, the same scale centers and coefficients, one-hot levels, and removed
features will be applied to the new data.
}
\examples{
res <- resample(iris, setup_KFold(seed = 2026))
iris_train <- iris[res[[1]], ]
iris_test <- iris[-res[[1]], ]

# Preprocess training data
iris_pre <- preprocess(iris_train, setup_Preprocessor(scale = TRUE, center = TRUE))

# Apply the same preprocessing to test data
iris_test_pre <- apply_preprocessor(iris_pre, iris_test)
}
\seealso{
\link{preprocess}, \link{setup_Preprocessor}
}
\author{
EDG
}
