% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linadleaves.R
\name{as.data.tree.linadleaves}
\alias{as.data.tree.linadleaves}
\title{Convert \code{linadleaves} to \code{data.tree} object}
\usage{
as.data.tree.linadleaves(object)
}
\arguments{
\item{object}{\code{linadleaves} object}
}
\description{
Convert \code{linadleaves} to \code{data.tree} object
}
