% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shyoptleaves.R
\name{as.data.tree.shyoptleaves}
\alias{as.data.tree.shyoptleaves}
\title{Convert \code{shyoptleaves} to \code{data.tree} object}
\usage{
as.data.tree.shyoptleaves(object)
}
\arguments{
\item{object}{\code{shyoptleaves} object}
}
\description{
Convert \code{shyoptleaves} to \code{data.tree} object
}
