% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rules.rpart.R
\name{as.rules.rpart}
\alias{as.rules.rpart}
\title{Convert \code{rpart} rules to \code{data.tree} object}
\usage{
as.rules.rpart(object, verbose = FALSE)
}
\arguments{
\item{object}{\code{rpart} object}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\value{
\code{data.tree} object
}
\description{
Convert an \code{rpart} object to a \code{data.tree} object, which can be plotted with
\link{dplot3.cart}
}
\author{
E.D. Gennatas
}
\keyword{internal}
