% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmDB.R
\name{available_supervised}
\alias{available_supervised}
\alias{available_algorithms}
\alias{available_clustering}
\alias{available_decomposition}
\title{Available Algorithms}
\usage{
available_supervised()

available_clustering()

available_decomposition()
}
\value{
Called for its side effect of printing available algorithms.
}
\description{
Print available algorithms for supervised learning, clustering, and decomposition.
}
\author{
EDG
}
