% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmDB.R
\name{available_supervised}
\alias{available_supervised}
\alias{available_algorithms}
\alias{available_clustering}
\alias{available_decomposition}
\title{Available Algorithms}
\usage{
available_supervised(verbosity = 1L)

available_clustering(verbosity = 1L)

available_decomposition(verbosity = 1L)
}
\arguments{
\item{verbosity}{Integer: Verbosity level.}
}
\value{
Named list of algorithm descriptions, invisibly.
}
\description{
Print available algorithms for supervised learning, clustering, and decomposition.
}
\examples{
available_supervised()
available_clustering()
available_decomposition()
}
\author{
EDG
}
