% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmDB.R
\name{available_draw}
\alias{available_draw}
\title{Available Draw Functions}
\usage{
available_draw(verbosity = 1L)
}
\arguments{
\item{verbosity}{Integer: Verbosity level.}
}
\value{
Named list of draw function descriptions, invisibly.
}
\description{
Print available draw functions for visualization.
}
\examples{
available_draw()
}
\author{
EDG
}
