% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{bacc}
\alias{bacc}
\title{Balanced Accuracy}
\usage{
bacc(true, predicted, harmonize = FALSE, verbosity = 1)
}
\arguments{
\item{true}{True labels}

\item{predicted}{Estimated labels}

\item{harmonize}{Logical: passed to \link{sensitivity} and \link{specificity}, which use \link{factor_harmonize}.
Default = FALSE}

\item{verbosity}{Integer: If > 0, print messages to console.}
}
\description{
Balanced Accuracy of a binary classifier
}
\details{
BAcc = .5 * (Sensitivity + Specificity)
}
