% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_variance.R
\name{bias_variance}
\alias{bias_variance}
\title{Bias-Variance Decomposition}
\usage{
bias_variance(
  x,
  y,
  mod,
  res1_train.p = 0.7,
  params = list(),
  resample.params = setup.resample(n.resamples = 100),
  seed = NULL,
  verbose = TRUE,
  res.verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Predictors}

\item{y}{Outcome}

\item{mod}{Character: rtemis learner}

\item{res1_train.p}{Numeric: Proportion of cases to use for training}

\item{params}{List of \code{mod} parameters}

\item{resample.params}{Output of \link{setup.resample}}

\item{seed}{Integer: Seed for initial train/test split}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{res.verbose}{Logical: passed to the learning function}

\item{...}{Additional arguments passed to \code{resLearn}}
}
\description{
Bias-Variance Decomposition
}
\author{
E.D. Gennatas
}
