% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcat.R
\name{boxcat}
\alias{boxcat}
\title{Box cat}
\usage{
boxcat(x, style = 2, col = NULL, newline.pre = TRUE, newline = FALSE, pad = 0)
}
\arguments{
\item{x}{Character: Text to be output to console}

\item{style}{Integer: {1, 2}: 1: vintage style, 2: modern style. Default = 2}

\item{col}{Color: Any color support by \pkg{crayon}}

\item{newline.pre}{Logical: If TRUE, start with a new line. Default = TRUE}

\item{newline}{Logical: If TRUE, end with a new (empty) line. Default = FALSE}

\item{pad}{Integer: Pad message with this many spaces on the left. Default = 0}
}
\description{
\code{cat} with a box; a boxed cat
}
\author{
Efstathios D. Gennatas
}
\keyword{internal}
