% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_HARDCL.R
\name{c_HARDCL}
\alias{c_HARDCL}
\title{Clustering by Hard Competitive Learning}
\usage{
c_HARDCL(x, x.test = NULL, k = 2, dist = "euclidean", verbose = TRUE, ...)
}
\arguments{
\item{x}{Input matrix / data.frame}

\item{x.test}{Optional test set data}

\item{k}{Integer: Number of clusters to get}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{...}{Additional parameters to be passed to \code{flexclust::cclust}}
}
\description{
Perform clustering by
\href{https://en.wikipedia.org/wiki/Competitive_learning}{Hard Competitive Learning}
using \code{flexclust::cclust}
}
\seealso{
Other Clustering: 
\code{\link{c_CMeans}()},
\code{\link{c_DBSCAN}()},
\code{\link{c_EMC}()},
\code{\link{c_H2OKMeans}()},
\code{\link{c_HOPACH}()},
\code{\link{c_KMeans}()},
\code{\link{c_MeanShift}()},
\code{\link{c_NGAS}()},
\code{\link{c_PAM}()},
\code{\link{c_PAMK}()},
\code{\link{c_SPEC}()}
}
\author{
E.D. Gennatas
}
\concept{Clustering}
