% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate.Classification}
\alias{calibrate.Classification}
\title{Calibrate Binary Classification Models}
\usage{
calibrate.Classification(
  x,
  predicted_probabilities,
  true_labels,
  algorithm = "isotonic",
  hyperparameters = NULL,
  verbosity = 1L,
  ...
)
}
\arguments{
\item{x}{\code{Classification} object.}

\item{predicted_probabilities}{Numeric vector: Predicted probabilities.}

\item{true_labels}{Factor: True class labels.}

\item{algorithm}{Character: Algorithm to use to train calibration model.}

\item{hyperparameters}{\code{Hyperparameters} object: Setup using one of \verb{setup_*} functions.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Not used}
}
\value{
\code{CalibratedClassification} object.
}
\description{
The goal of calibration is to adjust the predicted probabilities of a binary classification
model so that they better reflect the true probabilities (i.e. empirical risk) of the positive
class.
}
\details{
Important: The calibration model's training data should be different from the classification
model's training data.
}
\examples{
datc2 <- data.frame(
  gn = factor(sample(c("alpha", "beta", "gamma"), 100, replace = TRUE)),
  iris[51:150, ]
)
res <- resample(datc2)
datc2$Species <- factor(datc2$Species)
datc2_train <- datc2[res[[1]], ]
datc2_test <- datc2[-res[[1]], ]
mod_c_lightrf <- train(
  x = datc2_train,
  dat_test = datc2_test,
  hyperparameters = setup_LightRF(nrounds = 20L)
)
mod_c_lightrf_cal <- calibrate(
  mod_c_lightrf,
  predicted_probabilities = mod_c_lightrf$predicted_prob_training,
  true_labels = mod_c_lightrf$y_training
)
mod_c_lightrf_cal
# Note: In this trivial example, there may be no change in predicted probabilities.
}
\author{
EDG
}
