% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Binary Classification Models}
\usage{
calibrate(x, ...)
}
\arguments{
\item{x}{Classification or ClassificationCV object: The model will be returned with the calibration model and
calibration metrics added.}

\item{predicted_probabilities}{Numeric vector: Predicted probabilities.}

\item{true_labels}{Factor: True class labels.}

\item{algorithm}{Character: Algorithm to use to train calibration model.}

\item{hyperparameters}{Hyperparameters object: Setup using one of \verb{setup_*} functions.}
}
\description{
The goal of calibration is to adjust the predicted probabilities of a binary classification
model so that they better reflect the true probabilities (i.e. empirical risk) of the positive
class.
}
\details{
Important: The calibration model's training data should be different from the classification
model's training data.
}
\author{
EDG
}
