% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate predicted probabilities using GAM}
\usage{
calibrate(
  true.labels,
  est.prob,
  pos.class.idi = NULL,
  mod = c("gam", "glm"),
  k = 5,
  verbose = TRUE
)
}
\arguments{
\item{true.labels}{Factor with true class labels}

\item{est.prob}{Numeric vector with predicted probabilities}

\item{pos.class.idi}{Integer: Index of the positive class}

\item{mod}{Character: Model to use for calibration. Either "gam" or "glm"}

\item{k}{Integer: GAM degrees of freedom}

\item{verbose}{Logical: If TRUE, print messages to the console}
}
\value{
mod: fitted GAM model. Use \code{mod$fitted.values} to get calibrated
input probabilities; use \code{predict(mod, newdata = newdata, type = "response")}
to calibrate other estimated probabilities.
}
\description{
Calibrate predicted probabilities using a generalized additive model (GAM).
}
\examples{
\dontrun{
data(segment_logistic, package = "probably")

# Plot the calibration curve of the original predictions
dplot3_calibration(
  true.labels = segment_logistic$Class,
  est.prob = segment_logistic$.pred_poor,
  n_windows = 10,
  pos.class.idi = 2
)

# Plot the calibration curve of the calibrated predictions
dplot3_calibration(
  true.labels = segment_logistic$Class,
  est.prob = calibrate(
    segment_logistic$Class,
    segment_logistic$.pred_poor
  )$fitted.values,
  n_windows = 10,
  pos.class.idi = 2
)
}
}
\author{
EDG
}
