% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_init.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate \code{Classification} & \code{ClassificationRes} Models}
\usage{
calibrate(x, hyperparameters = setup_Isotonic(), verbosity = 1L, ...)
}
\arguments{
\item{x}{\code{Classification} or \code{ClassificationRes} object to calibrate.}

\item{hyperparameters}{\code{Hyperparameters} object: Setup using one of \verb{setup_*} functions.
Defines the algorithm used to train the calibration model.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Calibrated model object.
}
\description{
Generic function to calibrate binary classification models.
}
\details{
The goal of calibration is to adjust the predicted probabilities of a binary classification
model so that they better reflect the true probabilities (i.e. empirical risk) of the positive
class.
}
\section{Method-specific parameters}{


\strong{For \code{Classification} objects:}
\itemize{
\item \code{predicted_probabilities}: Numeric vector of the positive class's
predicted probabilities, one per case
\item \code{true_labels}: Factor of true class labels
}

\strong{For \code{ClassificationRes} objects:}
\itemize{
\item \code{resampler_config}: \code{ResamplerConfig} object for calibration training
\item \code{train_verbosity}: Integer controlling calibration model training output
}
}

\examples{
# --- Calibrate Classification ---
dat <- iris[51:150, ]
res <- resample(dat)
dat$Species <- factor(dat$Species)
dat_train <- dat[res[[1]], ]
dat_test <- dat[-res[[1]], ]

# Train GLM on a training/test split
mod_c_glm <- train(
  x = dat_train,
  dat_test = dat_test,
  hyperparameters = setup_GLM()
)

# Calibrate the `Classification` by defining `predicted_probabilities` and `true_labels`,
# in this case using the training data, but it could be a separate calibration dataset.
mod_c_glm_cal <- calibrate(
  mod_c_glm,
  predicted_probabilities = mod_c_glm$predicted_prob_training[, 1L],
  true_labels = mod_c_glm$y_training
)
mod_c_glm_cal

# --- Calibrate ClassificationRes ---

# Train GLM with cross-validation
resmod_c_glm <- train(
  x = dat,
  hyperparameters = setup_GLM(),
  outer_resampling_config = setup_Resampler(n_resamples = 3L, type = "KFold")
)

# Calibrate the `ClassificationRes` using the same resampling configuration as used for training.
resmod_c_glm_cal <- calibrate(resmod_c_glm)
resmod_c_glm_cal
}
\author{
EDG
}
