% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Classification & ClassificationRes Models}
\usage{
calibrate(
  x,
  algorithm = "isotonic",
  hyperparameters = NULL,
  verbosity = 1L,
  ...
)
}
\arguments{
\item{x}{\code{Classification} or \code{ClassificationRes} object to calibrate.}

\item{algorithm}{Character: Algorithm to use to train calibration model.}

\item{hyperparameters}{\code{Hyperparameters} object: Setup using one of \verb{setup_*} functions.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Calibrated model object.
}
\description{
Generic function to calibrate binary classification models.
}
\details{
The goal of calibration is to adjust the predicted probabilities of a binary classification
model so that they better reflect the true probabilities (i.e. empirical risk) of the positive
class.
}
\author{
EDG
}
