% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_cv.R
\name{calibrate_cv}
\alias{calibrate_cv}
\title{Calibrate cross-validated model}
\usage{
calibrate_cv(
  mod,
  alg = "gam",
  learn_params = list(),
  which.repeat = 1,
  verbosity = 1
)
}
\arguments{
\item{mod}{List: Model object returned by \link{train_cv}}

\item{alg}{Character: "gam" or "glm", algorithm to use for calibration}

\item{learn_params}{List: List of parameters to pass to the learning algorithm}

\item{which.repeat}{Integer: Which repeat to use for calibration}
}
\value{
List: Calibrated models, calibrated probabilities, and test set metrics
}
\description{
Calibrate cross-validated model trained using \link{train_cv}
}
\details{
This is a work in progress to be potentially incorporated into \link{train_cv}
}
\author{
E.D. Gennatas
}
