% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_cv.R
\name{calibrate_cv}
\alias{calibrate_cv}
\title{Calibrate cross-validated model}
\usage{
calibrate_cv(
  mod,
  alg = "isotonic",
  learn.params = list(),
  resample.params = setup.resample(resampler = "kfold", n.resamples = 5, seed = NULL),
  which.repeat = 1,
  verbosity = 1,
  debug = FALSE
)
}
\arguments{
\item{mod}{\code{rtModCV} object returned by \link{train_cv}}

\item{alg}{Character: Algorithm to use for calibration}

\item{learn.params}{List: List of parameters to pass to the learning algorithm}

\item{resample.params}{List of parameters to pass to the resampling algorithm.
Build using \link{setup.resample}}

\item{which.repeat}{Integer: Which repeat to use for calibration}

\item{verbosity}{Integer: 0: silent, > 0: print messages}

\item{debug}{Logical: If TRUE, run without parallel processing, to allow better
debugging.}
}
\value{
List: Calibrated models, test-set labels, test set performance metrics,
estimated probabilities (uncalibrated), calibrated probabilities,
}
\description{
Calibrate cross-validated model trained using \link{train_cv}
}
\details{
This is a work in progress to be potentially incorporated into \link{train_cv}
You start by training a cross-validated model using \link{train_cv}, then this function
can be used to calibrate the model. In order to use all available data, each outer
resample from the input \code{mod} is resampled (using 5-fold CV by default) to train
and test calibration models. This allows using the original label-based metrics
of \code{mod} and also extract calibration metrics based on the same data, after
aggregating the test set predictions of the calibration models.
}
\author{
E.D. Gennatas
}
