% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{catrange}
\alias{catrange}
\title{Print range of continuous variable}
\usage{
catrange(x, ddSci = TRUE, decimal.places = 1, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{ddSci}{Logical: If TRUE, use \link{ddSci} or range. Default = TRUE}

\item{decimal.places}{Integer: Number of decimal place to use if \code{ddSci = TRUE}. Default = 1}

\item{na.rm}{Logical: passed to \code{base::range}}
}
\description{
Print range of continuous variable
}
\author{
E.D. Gennatas
}
