% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataops.R
\name{catsize}
\alias{catsize}
\title{Size}
\usage{
catsize(x, verbose = TRUE)
}
\arguments{
\item{x}{R object (usually that inherits from matrix or data.frame)}

\item{verbose}{Logical: If TRUE, print NROW and NCOL to console. Default = TRUE}
}
\value{
vector of NROW, NCOL invisibly
}
\description{
Get \code{NCOL(x)} and \code{NROW{x}}
}
\author{
E.D. Gennatas
}
