% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check Data}
\usage{
check_data(
  x,
  name = NULL,
  get_duplicates = TRUE,
  get_na_case_pct = FALSE,
  get_na_feature_pct = FALSE
)
}
\arguments{
\item{x}{data.frame, data.table or similar structure}

\item{name}{Character: Name of dataset}

\item{get_duplicates}{Logical: If TRUE, check for duplicate cases}

\item{get_na_case_pct}{Logical: If TRUE, calculate percent of NA values per
case}

\item{get_na_feature_pct}{Logical: If TRUE, calculate percent of NA values
per feature}
}
\description{
Check Data
}
\examples{
\dontrun{
n <- 1000
x <- rnormmat(n, 50, return.df = TRUE)
x$char1 <- sample(letters, n, TRUE)
x$char2 <- sample(letters, n, TRUE)
x$fct <- factor(sample(letters, n, TRUE))
x <- rbind(x, x[1, ])
x$const <- 99L
x[sample(nrow(x), 20), 3] <- NA
x[sample(nrow(x), 20), 10] <- NA
x$fct[30:35] <- NA
check_data(x)
}

}
\author{
E.D. Gennatas
}
