% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classError.R
\name{classError}
\alias{classError}
\title{Classification Error}
\usage{
classError(
  true,
  estimated,
  estimated.prob = NULL,
  calc.auc = TRUE,
  auc.method = "pROC",
  trace = 0
)
}
\arguments{
\item{true}{Factor: True labels}

\item{estimated}{Factor: Estimated values}

\item{estimated.prob}{Numeric vector: Estimated probabilities}

\item{calc.auc}{Logical: If TRUE, calculate AUC.}

\item{trace}{Integer: If > 0, print  diagnostic messages. Default = 0}
}
\value{
S3 object of type "classError"
}
\description{
Calculates Classification Metrics
}
\details{
Note that auc.method = "pROC" is the only one that will output an AUC even if
one or more estimated probabilities are NA.
}
\examples{
\dontrun{
true <- factor(c("a", "a", "a", "b", "b", "b", "b"))
estimated <- c("a", "a", "b", "b", "a", "a", "b")
estimated.prob <- c(0.7, 0.55, 0.45, 0.25, 0.6, 0.7, 0.2)

classError(true, estimated, estimated.prob, auc.method = "pROC")
classError(true, estimated, estimated.prob, auc.method = "ROCR")
classError(true, estimated, estimated.prob, auc.method = "auc_pairs")
}
}
\author{
E.D. Gennatas
}
