% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classError.R
\name{classError}
\alias{classError}
\title{Classification Error}
\usage{
classError(true, estimated, estimated.prob = NULL, trace = 0)
}
\arguments{
\item{true}{Vector, factor: True values}

\item{estimated}{Vector, factor: Estimated probabilities}

\item{trace}{Integer: If > 0, print  diagnostic messages. Default = 0}
}
\value{
S3 object of type  "classError"
}
\description{
Calculates Classification Metrics
}
\author{
Efstathios D. Gennatas
}
