% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classError.R
\name{classError}
\alias{classError}
\title{Classification Error}
\usage{
classError(true, estimated, estimated.prob = NULL, calc.auc = TRUE, trace = 0)
}
\arguments{
\item{true}{Factor: True labels}

\item{estimated}{Factor: Estimated values}

\item{estimated.prob}{Numeric vector: Estimated probabilities}

\item{calc.auc}{Logical: If TRUE, calculate AUC.}

\item{trace}{Integer: If > 0, print  diagnostic messages. Default = 0}
}
\value{
S3 object of type "classError"
}
\description{
Calculates Classification Metrics
}
\author{
E.D. Gennatas
}
