% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelify.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Clean column names}
\usage{
clean_colnames(x)
}
\arguments{
\item{x}{Character, vector}
}
\description{
Clean column names by replacing all spaces and punctuation with a single underscore
}
\author{
E.D. Gennatas
}
