% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelify.R
\name{clean_names}
\alias{clean_names}
\title{Clean names}
\usage{
clean_names(x)
}
\description{
Replace all symbols and combinations of symbols in a character vector
with underscores
}
\examples{
clean_names(colnames(iris))
}
\author{
E.D. Gennatas
}
