% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Perform Clustering}
\usage{
cluster(x, algorithm = "KMeans", parameters = NULL, verbosity = 1L)
}
\arguments{
\item{x}{Matrix or data.frame: Data to cluster. Rows are cases to be clustered.}

\item{algorithm}{Character: Clustering algorithm.}

\item{parameters}{List: Algorithm-specific parameters.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Clustering object.
}
\description{
Perform clustering on the rows (usually cases) of a dataset.
}
\author{
EDG
}
