% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Perform Clustering}
\usage{
cluster(x, algorithm = "KMeans", config = NULL, verbosity = 1L)
}
\arguments{
\item{x}{Matrix or data.frame: Data to cluster. Rows are cases to be clustered.}

\item{algorithm}{Character: Clustering algorithm.}

\item{config}{List: Algorithm-specific config.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{Clustering} object.
}
\description{
Perform clustering on the rows (usually cases) of a dataset.
}
\details{
See \href{https://rdocs.rtemis.org/cluster}{rdocs.rtemis.org/cluster} for detailed documentation.
}
\examples{
iris_km <- cluster(exc(iris, "Species"), algorithm = "KMeans")
}
\author{
EDG
}
