% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{col2hex}
\alias{col2hex}
\title{Convert R color to hexadecimal code}
\usage{
col2hex(color)
}
\arguments{
\item{color}{Color(s) that R understands}
}
\value{
Character vector of hexadecimal codes.
}
\description{
Convert a color that R understands into the corresponding hexadecimal code
}
\examples{
\dontrun{
col2hex(c("gray50", "skyblue"))
}
}
\author{
EDG
}
