% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{color_fade}
\alias{color_fade}
\title{Fade color towards target}
\usage{
color_fade(x, to = "#000000", pct = 0.5)
}
\arguments{
\item{x}{Color source}

\item{to}{Target color}

\item{pct}{Numeric (0, 1) fraction of the distance in RGBA space between
\code{x} and \code{to} to move. e.g. .5 gets the mean RGBA value of the two}
}
\value{
Color in hex notation
}
\description{
Fade color towards target
}
\author{
E.D. Gennatas
}
