% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorgradient.x.R
\name{colorgradient.x}
\alias{colorgradient.x}
\title{Color gradient for continuous variable}
\usage{
colorgradient.x(
  x,
  symmetric = FALSE,
  lo.col = "#0290EE",
  mid.col = "#1A1A1A",
  hi.col = "#FFBD4F",
  space = "Lab"
)
}
\arguments{
\item{x}{Float, vector}

\item{space}{Character: "rgb" or "Lab". Default = "Lab"}

\item{color}{Color, vector, length 2}
}
\description{
Color gradient for continuous variable
}
\examples{
\dontrun{
x <- seq(-10, 10, length.out = 51)
previewcolor(colorgradient.x(x))
x <- sort(rnorm(40))
previewcolor(colorgradient.x(x, mid.col = "white"))
# Notice how most values are near zero therefore almost white
}
}
\author{
E.D. Gennatas
}
