% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorMix.R
\name{colorMix}
\alias{colorMix}
\title{Create an alternating sequence of graded colors}
\usage{
colorMix(color, n = 4)
}
\arguments{
\item{color}{List: List of two or more elements, each containing two colors.
A gradient will be created from the first to the second color of each element}

\item{n}{Integer: Number of steps in each gradient.}
}
\description{
Create an alternating sequence of graded colors
}
\examples{
color <- list(blue = c("#82afd3", "#000f3a"),
              gray = c("gray10", "gray85"))
previewcolor(desaturate(colorMix(color, 6), .3))

color <- list(blue = c("#82afd3", "#57000a"),
              gray = c("gray10", "gray85"))
previewcolor(desaturate(colorMix(color, 6), .3))

color <- list(blue = c("#82afd3", "#000f3a"),
              purple = c("#23001f", "#c480c1"))
previewcolor(desaturate(colorMix(color, 5), .3))
}
\author{
E.D. Gennatas
}
