% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorOps.R
\name{colorOp}
\alias{colorOp}
\title{Simple Color Operations}
\usage{
colorOp(col, fn = c("invert", "mean"), space = c("HSV", "RGB"))
}
\arguments{
\item{col}{Input color(s)}

\item{fn}{Character: "invert", "mean": Function to perform}

\item{space}{Character: "HSV", "RGB": Colorspace to operate in - for
averaging only}
}
\value{
Color
}
\description{
Invert a color or calculate the mean of two colors in HSV or RGB space.
This may be useful in creating colors for plots
}
\details{
The average of two colors in RGB space will often pass through gray,
which is likely undesirable. Averaging in HSV space, better for most applications.
}
\author{
E.D. Gennatas
}
