% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_LLE.R
\name{d_LLE}
\alias{d_LLE}
\title{Locally Linear Embedding}
\usage{
d_LLE(x, k = 2, nn = 6, verbose = TRUE)
}
\arguments{
\item{x}{Input data}

\item{k}{Integer: dimensionality of the embedding}

\item{nn}{Integer: Number of neighbors.}

\item{verbose}{Logical: If TRUE, print messages to screen. Default = TRUE}
}
\value{
\code{rtDecom} object
}
\description{
Perform LLE decomposition using \code{RDRToolbox::lle}
}
\details{
Project scaled variables to LLE components
Input must be n by p,
where n represents number of cases,
and p represents number of features.
}
\seealso{
Other Decomposition: 
\code{\link{d_H2OAE}()},
\code{\link{d_H2OGLRM}()},
\code{\link{d_ICA}()},
\code{\link{d_Isomap}()},
\code{\link{d_KPCA}()},
\code{\link{d_MDS}()},
\code{\link{d_NMF}()},
\code{\link{d_PCA}()},
\code{\link{d_SPCA}()},
\code{\link{d_SVD}()},
\code{\link{d_TSNE}()},
\code{\link{d_UMAP}()}
}
\author{
E.D. Gennatas
}
\concept{Decomposition}
