% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{dataPrepare}
\alias{dataPrepare}
\title{\code{rtemis-internals}: \code{dataPrepare}}
\usage{
dataPrepare(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  x.valid = NULL,
  y.valid = NULL,
  ipw = FALSE,
  ipw.type = 2,
  upsample = FALSE,
  downsample = FALSE,
  resample.seed = NULL,
  removeDots = FALSE,
  .preprocess = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{x.valid}{Matrix / Data frame: Validation set features}

\item{y.valid}{Vector: Validation outcome}

\item{ipw}{Logical: If TRUE, return class weights for inverse probability weighting
(for Classification)}

\item{ipw.type}{{1, 2}: 1:}

\item{upsample}{Logical: If TRUE, downsample majority class to match size of minority class}

\item{downsample}{Logical: If TRUE, downsample majority class to match size of minority class}

\item{resample.seed}{Integer: If set, use \code{set.seed} for reproducibility. Default = NULL}

\item{removeDots}{Logical: If TRUE, replace dots in variable names with underscores.
Some algorithms do not work with variable names containing dots (SparkML)}

\item{.preprocess}{List: Preprocessing settings. Set with \link{rtset.preprocess}}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\description{
Prepare data for \pkg{rtemis} modeling
}
