% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decom.R
\name{decom}
\alias{decom}
\title{Matrix Decomposition with \pkg{rtemis}}
\usage{
decom(x, decom = "ICA", verbose = TRUE, ...)
}
\arguments{
\item{x}{Numeric matrix / data frame: Input data}

\item{decom}{Character: Decomposer name. See ]link{decomSelect}. Default = "ICA"}

\item{verbose}{Logical: if TRUE, print messages to screen}

\item{...}{Additional arguments to be passed to \code{decom}}
}
\value{
\link{rtDecom} object
}
\description{
Convenience function to perform any \pkg{rtemis} decomposition
}
\details{
\code{decom} returns an R6 class object \code{rtDecom}
}
\author{
E.D. Gennatas
}
