% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{decomp}
\alias{decomp}
\title{Perform Data Decomposition}
\usage{
decomp(x, algorithm = "ICA", config = NULL, verbosity = 1L)
}
\arguments{
\item{x}{Matrix or data frame: Input data.}

\item{algorithm}{Character: Decomposition algorithm.}

\item{config}{DecompositionConfig: Algorithm-specific config.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{Decomposition} object.
}
\description{
Perform linear or non-linear decomposition of numeric data.
}
\details{
See \href{https://rdocs.rtemis.org/decomp}{rdocs.rtemis.org/decomp} for detailed documentation.
}
\examples{
iris_pca <- decomp(exc(iris, "Species"), algorithm = "PCA")
}
\author{
EDG
}
