% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depCheck.R
\name{depCheck}
\alias{depCheck}
\title{\pkg{rtemis} internal: Dependencies check}
\usage{
depCheck(..., verbose = TRUE)
}
\arguments{
\item{...}{List or vector of strings defining namespaces to be checked}

\item{verbose}{Logical. If TRUE, print messages to consolde. Note: messages are always printed
if dependencies are missing. Setting this to FALSE stops it from printing
"Dependencies check passed". Default = TRUE}
}
\value{
TRUE if dependencies can be loaded, otherwise FALSE
}
\description{
Checks if dependencies can be loaded; names missing dependencies if not.
}
\author{
E.D. Gennatas
}
