% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{describe.factor}
\alias{describe.factor}
\title{Describe factor}
\arguments{
\item{x}{factor.}

\item{max_n}{Integer: Return counts for up to this many levels.}

\item{return_ordered}{Logical: If TRUE, return levels ordered by count, otherwise
return in level order.}
}
\value{
Character with level counts.
}
\description{
Outputs a single character with names and counts of each level of the input factor.
}
\examples{
# Small number of levels
describe(iris[["Species"]])

# Large number of levels: show top n by count
x <- factor(sample(letters, 1000, TRUE))
describe(x)
describe(x, 3)
describe(x, 3, return_ordered = FALSE)
}
\author{
EDG
}
