% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{describe}
\alias{describe}
\title{Describe rtemis object}
\usage{
describe(x, ...)
}
\arguments{
\item{x}{An rtemis object.}
}
\value{
A character string describing the object.
}
\description{
This generic is used to provide a description of an rtemis object in plain language.
}
\author{
EDG
}
