% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{describe}
\alias{describe}
\title{Describe generic}
\usage{
describe(object, ...)
}
\arguments{
\item{object}{object to describe}

\item{...}{Additional arguments passed to \code{describe}}
}
\description{
Describe generic
}
\author{
E.D. Gennatas
}
