% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_init.R
\name{describe}
\alias{describe}
\title{Describe object}
\usage{
describe(x, verbosity = 1L, ...)
}
\arguments{
\item{x}{R object to describe. See method documentation for supported classes.}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character, invisibly.
}
\description{
Describe object
}
\details{
Extra arguments for \code{factor} method:
\itemize{
\item \code{max_n}: Integer: Return counts for up to this many levels.
\item \code{return_ordered}: Logical: If TRUE, return levels ordered by count, otherwise return in level order.
\item \code{verbosity}: Integer: Verbosity level.
}
}
\examples{
# --- For `Supervised` objects ---
species_lightrf <- train(iris, hyperparameters = setup_LightRF())
describe(species_lightrf)

# --- For `SupervisedRes` objects ---
mod <- train(iris, hyperparameters = setup_CART(), outer_resampling_config = setup_Resampler())
describe(mod)

# --- For factors ---
# Small number of levels
describe(iris[["Species"]])

# Large number of levels: show top n by count
x <- factor(sample(letters, 1000, TRUE))
describe(x)
describe(x, 3)
describe(x, 3, return_ordered = FALSE)
}
\author{
EDG
}
