% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.addtree.R
\name{dplot3.addtree}
\alias{dplot3.addtree}
\title{Plot ADDTREE trees}
\usage{
dplot3.addtree(
  addtree,
  col.positive = "#F48024DD",
  col.negative = "#18A3ACDD",
  node.col = "#666666",
  node.shape = "none",
  node.labels = TRUE,
  node.labels.pct.pos = NULL,
  pos.name = NULL,
  edge.col = "#999999",
  edge.width = 2,
  edge.labels = FALSE,
  arrowhead = "vee",
  layout = "dot",
  drop.leaves = FALSE,
  rankdir = "TB",
  splines = "polyline",
  fontname = "helvetica",
  bg.color = "#ffffff",
  overlap = "false",
  prune = NULL,
  prune.empty.leaves = TRUE,
  remove.bad.parents = FALSE,
  filename = NULL
)
}
\arguments{
\item{col.positive}{Color for outcome positive. Default = "#F48024DD"}

\item{col.negative}{Color for negative outcome. Default = "#18A3ACDD"}

\item{node.col}{Color for non-terminal leaves. Default = "#666666"}

\item{edge.col}{Color for edges. Default = "#999999"}

\item{bg.color}{Background color. Default = "#ffffff"}

\item{filename}{String, Optional: Save PDF to this file. Requires packages \code{DiagrammeRsvg}
and \code{rsvg}. Default = NULL}
}
\description{
Plot ADDTREE trees trained with \link{s.ADDTREE} using \code{data.tree::plot.Node}
}
\details{
Edge info and styles have been removed because of problems with \code{DiagrammeR}
}
\author{
E.D. Gennatas
}
