% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.bar.R
\name{dplot3.bar}
\alias{dplot3.bar}
\title{Dynamic Barplots}
\usage{
dplot3.bar(x, main = NULL, xlab = NULL, ylab = NULL, col = NULL,
  alpha = 0.8, bg = "white", plot.bg = "gray95",
  palette = getOption("rt.palette", "rtCol1"), group.names = NULL,
  feature.names = NULL, font.size = 16, font.alpha = 0.8,
  font.col = "black", font.family = "Helvetica Neue",
  grid.col = "gray85", grid.alpha = 0.8, margin = list(t = 35))
}
\arguments{
\item{x}{data.frame: Input where rows are groups (can be 1 row), columns are features}

\item{main}{Character: Plot title. Default = NULL}

\item{xlab}{Character: x-axis label. Default = NULL}

\item{ylab}{Character: y-axis label. Default = NULL}

\item{col}{Color, vector: Color for bars. Defaults to colors defined by \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for bar colors. Default = .8}

\item{bg}{Color: Background color. Default = "white"}

\item{plot.bg}{Color: Background color for plot area. Default = "white"}

\item{palette}{Character: Name of `rtemis` palette to use. Default = "rtCol1". Only used if \code{col = NULL}}

\item{group.names}{Character, vector, length = NROW(x): Group names. Default = NULL, which uses \code{rownames(x)}}

\item{feature.names}{Character, vector, length = NCOL(x): Feature names. Default = NULL, which uses
\code{colnames(x)}}

\item{font.size}{Float: Font size for all labels excluding main. Default = 16}

\item{font.alpha}{Float (0, 1]: Transparency for fonts. Default = .8}

\item{font.col}{Color: Font color. Default = "black"}

\item{font.family}{String: Font family to use. Default = "Helvetica Neue"}

\item{grid.col}{Color: Grid color. Default = "gray85"}

\item{grid.alpha}{Float (0, 1]: Transparency for \code{grid.col}}

\item{margin}{Named list: plot margins. Default = \code{list(t = 35)}}
}
\description{
Draw a dynamic barplot using \code{plotly}
}
\details{
This is an early release, part of an ongoing effort to complete graphic functions.
Will be adding theme support, among other options.
}
\examples{
\dontrun{
dplot3.bar(VADeaths)
}
}
\author{
Efstathios D. Gennatas
}
