% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.box.R
\name{dplot3.box}
\alias{dplot3.box}
\title{Interactive Boxplots & Violin plots}
\usage{
dplot3.box(
  x,
  time = NULL,
  time.bin = c("year", "quarter", "month", "day"),
  type = c("box", "violin"),
  group = NULL,
  main = NULL,
  xlab = "",
  ylab = NULL,
  col = NULL,
  alpha = 0.6,
  bg = NULL,
  plot.bg = NULL,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  boxpoints = "outliers",
  quartilemethod = "linear",
  violin.box = TRUE,
  orientation = "v",
  annotate_n = FALSE,
  annotate_n_y = 1,
  annotate.col = theme$labs.col,
  xnames = NULL,
  labelify = TRUE,
  order.by.fn = NULL,
  font.size = 16,
  legend = NULL,
  legend.col = NULL,
  legend.xy = NULL,
  legend.orientation = "v",
  legend.xanchor = "auto",
  legend.yanchor = "auto",
  xaxis.type = "category",
  margin = list(t = 35, pad = 0),
  automargin.x = TRUE,
  automargin.y = TRUE,
  boxgap = 0,
  boxgroupgap = NULL,
  hovertext = NULL,
  show_n = FALSE,
  boxmode = NULL,
  use_plotly_group = FALSE,
  displayModeBar = TRUE,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  print.plot = TRUE,
  ...
)
}
\arguments{
\item{x}{Vector or List of vectors: Input}

\item{main}{Character: Plot title. Default = NULL}

\item{xlab}{Character: x-axis label. Default = NULL}

\item{ylab}{Character: y-axis label. Default = NULL}

\item{col}{Color, vector: Color for boxes. Default NULL, which will draw colors from \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for box colors. Default = .8}

\item{bg}{Color: Background color. Default = "white"}

\item{plot.bg}{Color: Background color for plot area. Default = "white"}

\item{theme}{Character: THeme to use: "light", "dark", "lightgrid", "darkgrid". Default = "lightgrid"}

\item{palette}{Character: Name of \pkg{rtemis} palette to use. Default = "rtCol1". Only used if \code{col = NULL}}

\item{boxpoints}{Character or FALSE: "all", "suspectedoutliers", "outliers"
See \url{https://plotly.com/r/box-plots/#choosing-the-algorithm-for-computing-quartiles}}

\item{quartilemethod}{Character: "linear", "exclusive", "inclusive"}

\item{xnames}{Character, vector, length = NROW(x): x-axis names. Default = NULL, which
tries to set names appropriately}

\item{order.by.fn}{Function: If defined, order boxes by increasing value of this function
(e.g. median). Default = NULL}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{legend}{Logical: If TRUE, draw legend. Default = TRUE}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by theme}

\item{legend.xy}{Float, vector, length 2: Relative x, y position for legend. Default = NULL, which places
the legend top right beside the plot area. For example, c(0, 1) places the legend top left within the plot area}

\item{xaxis.type}{Character: "linear", "log", "date", "category", "multicategory"
Default = "category"}

\item{margin}{Named list: plot margins. Default = \code{list(t = 35)}}

\item{feature.names}{Character, vector, length = NCOL(x): Feature names. Default = NULL, which uses
\code{colnames(x)}}
}
\description{
Draw interactive boxplots or violin plots using \pkg{plotly}
}
\details{
For multiple box plots, the recommendation is:
- `x=dat[, columnindex]` for multiple variables of a data.frame
- `x=list(a=..., b=..., etc.)` for multiple variables of potentially different length
- `x=split(var, group)` for one variable with multiple groups: group names appear below boxplots
- `x=dat[, columnindex], group = factor` for grouping multiple variables: group names appear in legend

If \code{orientation = "h"}, \code{xlab} is applied to y-axis and vice versa. Similarly,
\code{x.axist.type} applies to y-axis - this defaults to "category" and would not normally need
changing.
}
\examples{
\dontrun{
# A.1 Box plot of 4 variables
dplot3.box(iris[, 1:4])
# A.2 Grouped Box plot
dplot3.box(iris[, 1:4], group = iris$Species)
dplot3.box(iris[, 1:4], group = iris$Species, annotate_n = TRUE)
# B. Boxplot binned by time periods
# Synthetic data with an instantenous shift in distributions
set.seed(2021)
dat1 <- data.frame(alpha = rnorm(200, 0), beta = rnorm(200, 2), gamma = rnorm(200, 3))
dat2 <- data.frame(alpha = rnorm(200, 5), beta = rnorm(200, 8), gamma = rnorm(200, -3))
x <- rbind(dat1, dat2)
startDate <- as.Date("2019-12-04")
endDate <- as.Date("2021-03-31")
time <- seq(startDate, endDate, length.out = 400)
dplot3.box(x[, 1], time, "year", ylab = "alpha")
dplot3.box(x, time, "year", legend.xy = c(0, 1))
dplot3.box(x, time, "quarter", legend.xy = c(0, 1))
dplot3.box(x, time, "month",
           legend.orientation = "h",
           legend.xy = c(0, 1),
           legend.yanchor = "bottom")
# (Note how the boxplots widen when the period includes data from both dat1 and dat2)
}
}
\author{
E.D. Gennatas
}
