% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.heatmap.R
\name{dplot3.heatmap}
\alias{dplot3.heatmap}
\title{Interactive Heatmaps}
\usage{
dplot3.heatmap(
  z,
  Rowv = TRUE,
  Colv = TRUE,
  cluster = FALSE,
  symm = FALSE,
  cellnote = NULL,
  colorGrad.n = 101,
  colors = NULL,
  space = "rgb",
  lo = "#18A3AC",
  lomid = NULL,
  mid = NULL,
  midhi = NULL,
  hi = "#F48024",
  k_row = 1,
  k_col = 1,
  grid.gap = 0,
  limits = NULL,
  margins = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  key.title = NULL,
  showticklabels = NULL,
  colorbar_len = 0.7,
  plot_method = "plotly",
  theme = getOption("rt.theme", "white"),
  row_side_colors,
  row_side_palette = NULL,
  col_side_colors,
  col_side_palette = NULL,
  font.size = NULL,
  padding = 0,
  filename = NULL,
  ...
)
}
\arguments{
\item{z}{Input matrix}

\item{Rowv}{Logical or dendrogram.
If Logical: Compute dendrogram and reorder rows. Defaults to FALSE
If dendrogram: use as is, without reordering
See more at \code{heatmaply::heatmaply("Rowv")}}

\item{Colv}{Logical or dendrogram.
If Logical: Compute dendrogram and reorder columns. Defaults to FALSE
If dendrogram: use as is, without reordering
See more at \code{heatmaply::heatmaply("Colv")}}

\item{cluster}{Logical: If TRUE, set \code{Rowv} and \code{Colv} to TRUE}

\item{symm}{Logical: If TRUE, treat \code{x} symmetrically - \code{x} must be a square matrix. Default = FALSE}

\item{cellnote}{Matrix with values to be desplayed on hover. Defaults to \code{ddSci(z)}}

\item{colorGrad.n}{Integer: Number of distinct colors to generate using \link{colorGrad}. Default = 101}

\item{colors}{Character: Acts as a shortcut to defining \code{lo}, \code{mid}, etc for a number of defaults:
"french", "penn", "grnblkred",}

\item{space}{Character: Which colorspace to use. Option: "rgb", or "Lab". Default = "rgb".
Recommendation: If \code{mid} is "white" or "black" (default), use "rgb", otherwise "Lab"}

\item{lo}{Color for low end}

\item{lomid}{Color for low-mid}

\item{mid}{Color for middle of the range or "mean", which will result in \code{colorOp(c(lo, hi), "mean")}.
If \code{mid = NA}, then only \code{lo} and \code{hi} are used to create the color gradient.}

\item{midhi}{Color for middle-high}

\item{hi}{Color for high end}

\item{k_row}{Integer: Number of desired number of groups by which to color dendrogram branches in the rows.
Default = NA (determined automatically). See \code{heatmaply::heatmaply("k_row")}}

\item{k_col}{Integer: Number of desired number of groups by which to color dendrogram branches in the columns.
Default = NA (determined automatically). See \code{heatmaply::heatmaply("k_col")}}

\item{grid.gap}{Integer: Space between cells. Default = 0 (no space)}

\item{limits}{Float, length 2: Determine color range. Default = NULL, which automatically centers values around 0}

\item{margins}{Float, length 4: Heatmap margins. Default = c(30, 30, 30, 30)}

\item{main}{Character: Plot title}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{key.title}{Character: Title for the color key. Default = NULL (no title)}

\item{plot_method}{Character: Update February 2021: "ggplot" causes R session to hand
on MacOS but plotly" seems to work}

\item{theme}{Character: "light", "dark"}

\item{filename}{String (Optional: Path to file to save colorbar}

\item{...}{Additional arguments to be passed to \code{heatmaply::heatmaply}}
}
\description{
Draw interactive heatmaps using \code{heatmaply}
}
\examples{
\dontrun{
x <- rnormmat(200, 20)
xcor <- cor(x)
dplot3.heatmap(xcor)
}
}
\author{
E.D. Gennatas
}
