% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.ts.R
\name{dplot3.ts}
\alias{dplot3.ts}
\title{Interactive Timeseries Plots}
\usage{
dplot3.ts(
  x,
  time,
  window = 7,
  roll.fn = c("mean", "median", "max", "sum"),
  roll.col = NULL,
  roll.alpha = 1,
  roll.lwd = 2,
  roll.name = NULL,
  alpha = NULL,
  align = "center",
  use = "data.table",
  xlab = "Time",
  n.xticks = 12,
  tickmode = "array",
  scatter.type = "scattergl",
  legend = TRUE,
  theme = getOption("rt.theme"),
  palette = getOption("rt.palette", "rtCol1"),
  ...
)
}
\arguments{
\item{x}{Numeric vector of values to plot}

\item{time}{Numeric or Date vector of time corresponding to values of \code{x}}

\item{roll.fn}{Character: "mean", "median", "max", or "sum": Function to apply on rolling windows of \code{x}}

\item{roll.col}{Color for rolling line}

\item{roll.alpha}{Numeric: transparency for rolling line}

\item{roll.lwd}{Numeric: width of rolling line}

\item{align}{Character: "center", "right", or "left"}

\item{use}{Character: "data.table" or "zoo": which package to use to apply
rolling function}

\item{xlab}{Character: x-axis label}

\item{n.xticks}{Integer: number of x-axis ticks to use (approximately)}

\item{theme}{Character: theme name or list of theme parameters}

\item{palette}{Character: palette name, or list of colors}
}
\description{
Draw interactive timeseries plots using \code{plotly}
}
\author{
E.D. Gennatas
}
