% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.ts.R
\name{dplot3.ts}
\alias{dplot3.ts}
\title{Interactive Timeseries Plots}
\usage{
dplot3.ts(
  x,
  time,
  window = 7L,
  group = NULL,
  roll.fn = c("mean", "median", "max", "none"),
  roll.col = NULL,
  roll.alpha = 1,
  roll.lwd = 2,
  roll.name = NULL,
  alpha = NULL,
  align = "center",
  group.names = NULL,
  xlab = "Time",
  n.xticks = 12,
  tickmode = "array",
  scatter.type = "scatter",
  legend = TRUE,
  x.showspikes = TRUE,
  y.showspikes = FALSE,
  spikedash = "solid",
  spikemode = "across",
  spikesnap = "hovered data",
  spikecolor = NULL,
  spikethickness = 1,
  displayModeBar = TRUE,
  theme = getOption("rt.theme"),
  palette = getOption("rt.palette", "rtCol1"),
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{Numeric vector of values to plot or list of vectors}

\item{time}{Numeric or Date vector of time corresponding to values of \code{x}}

\item{window}{Integer: apply \code{roll.fn} over this many units of time}

\item{roll.fn}{Character: "mean", "median", "max", or "sum": Function to apply on rolling windows of \code{x}}

\item{roll.col}{Color for rolling line}

\item{roll.alpha}{Numeric: transparency for rolling line}

\item{roll.lwd}{Numeric: width of rolling line}

\item{align}{Character: "center", "right", or "left"}

\item{xlab}{Character: x-axis label}

\item{n.xticks}{Integer: number of x-axis ticks to use (approximately)}

\item{x.showspikes}{Logical: If TRUE, show x-axis spikes on hover}

\item{displayModeBar}{Logical: If TRUE, display plotly's modebar}

\item{theme}{Character: theme name or list of theme parameters}

\item{palette}{Character: palette name, or list of colors}

\item{filename}{Character: filename to save plot to}

\item{use}{Character: "data.table" or "zoo": which package to use to apply
rolling function}
}
\description{
Draw interactive timeseries plots using \code{plotly}
}
\examples{
\dontrun{
time <- sample(seq(as.Date("2020-03-01"), as.Date("2020-09-23"), length.out = 140))
x1 <- rnorm(140)
x2 <- rnorm(140, 1, 1.2)
# Single timeseries
dplot3.ts(x1, time)
# Multiple timeseries input as list
dplot3.ts(list(Alpha = x1, Beta = x2), time)
# Multiple timeseries grouped by group, different lengths
time1 <- sample(seq(as.Date("2020-03-01"), as.Date("2020-07-23"), length.out = 100))
time2 <- sample(seq(as.Date("2020-05-01"), as.Date("2020-09-23"), length.out = 140))
time <- c(time1, time2)
x <- c(rnorm(100), rnorm(140, 1, 1.5))
group <- c(rep("Alpha", 100), rep("Beta", 140))
dplot3.ts(x, time, 7, group)
}

}
\author{
E.D. Gennatas
}
