% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.volcano.R
\name{dplot3.volcano}
\alias{dplot3.volcano}
\title{Volcano Plot}
\usage{
dplot3.volcano(
  x,
  pvals,
  xnames = NULL,
  x.thresh = 0,
  p.thresh = 0.05,
  p.transform = c("-log10", "none"),
  p.adjust.method = "holm",
  label.lo = "<b>Low</b>",
  label.hi = "<b>High</b>",
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  alpha = 0.5,
  theme = getOption("rt.theme"),
  font.size = 16,
  palette = list("#F48024", "#7f7f7f", "#18A3AC"),
  legend.x = c(-0.2, 0.2),
  legend.y = 0.97,
  annotate.n = 7,
  ay.lo = NULL,
  ay.hi = NULL,
  annotate = TRUE,
  annotate.alpha = 0.7,
  hovertext = NULL,
  displayModeBar = FALSE,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  verbose = TRUE,
  trace = 0,
  ...
)
}
\arguments{
\item{x}{Numeric vector: Input values, e.g. log2 fold change, coefficients, etc.}

\item{pvals}{Numeric vector: p-values}

\item{p.thresh}{Numeric: p-value threshold of significance. Default = .05}
}
\description{
Volcano Plot
}
\author{
E.D. Gennatas
}
