% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.volcano.R
\name{dplot3.volcano}
\alias{dplot3.volcano}
\title{Volcano Plot}
\usage{
dplot3.volcano(
  x,
  pvals,
  xnames = NULL,
  x.thresh = 0,
  p.thresh = 0.05,
  p.transform = c("-log10", "none"),
  p.adjust.method = "holm",
  label.lo = NULL,
  label.hi = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  alpha = 0.5,
  theme = getOption("rt.theme"),
  font.size = 16,
  palette = list("#18A3AC", "#7f7f7f", "#F48024"),
  legend.x.lo = NULL,
  legend.x.hi = NULL,
  legend.y = 0.97,
  annotate.n = 7,
  ay.lo = NULL,
  ay.hi = NULL,
  annotate = TRUE,
  annotate.alpha = 0.7,
  hovertext = NULL,
  displayModeBar = FALSE,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector: Input values, e.g. log2 fold change, coefficients, etc.}

\item{pvals}{Numeric vector: p-values}

\item{p.thresh}{Numeric: p-value threshold of significance. Default = .05}
}
\description{
Volcano Plot
}
\examples{
\dontrun{
set.seed(2019)
x <- rnormmat(500, 500)
y <- x[, 3] + x[, 5] + x[, 9] + x[, 15] + rnorm(500)
mod <- massGLM(y, x)
dplot3.volcano(mod$summary$`Coefficient y`, mod$summary$`p_value y`)
}
}
\author{
E.D. Gennatas
}
