% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.xy.R
\name{dplot3.xy}
\alias{dplot3.xy}
\title{Interactive Scatter Plots}
\usage{
dplot3.xy(x, y, fit = NULL, se.fit = FALSE, se.times = 1.96,
  cluster = NULL, cluster.params = list(k = 2), group = NULL,
  formula = NULL, rsq = NULL, rsq.pval = FALSE, mode = c("scatter",
  "line"), type = c("p", "l"), main = NULL, xlab = NULL,
  ylab = NULL, col = NULL, alpha = 0.66, bg = NULL,
  plot.bg = NULL, theme = getOption("rt.theme", "light"),
  palette = getOption("rt.palette", "rtCol1"), axes.square = FALSE,
  zero.lines = TRUE, group.names = NULL, font.size = 16,
  font.alpha = 0.8, font.col = NULL, font.family = "Helvetica Neue",
  marker.col = NULL, fit.col = NULL, fit.alpha = 0.8,
  fit.lwd = 2.5, se.col = NULL, se.alpha = 0.4, main.col = NULL,
  axes.col = NULL, labs.col = NULL, grid.col = NULL, grid.lwd = 1,
  grid.alpha = 0.8, tick.col = NULL, legend = TRUE,
  legend.xy = c(0, 1), legend.col = NULL, legend.bg = "#FFFFFF00",
  legend.border.col = "#FFFFFF00", legend.group.gap = 0,
  margin = list(t = 35), zerolines = TRUE, mod.params = list(),
  width = NULL, height = NULL, padding = 0, trace = 0,
  filename = NULL, file.width = 500, file.height = 500, ...)
}
\arguments{
\item{x}{Numeric, vector: Input}

\item{main}{Character: Plot title. Default = NULL}

\item{xlab}{Character: x-axis label. Default = NULL}

\item{ylab}{Character: y-axis label. Default = NULL}

\item{col}{Color, vector: Color for bars. Default NULL, which will draw colors from \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for bar colors. Default = .8}

\item{bg}{Color: Background color. Default = "white"}

\item{plot.bg}{Color: Background color for plot area. Default = "white"}

\item{theme}{Character: THeme to use: "light", "dark", "lightgrid", "darkgrid". Default = "lightgrid"}

\item{palette}{Character: Name of `rtemis` palette to use. Default = "rtCol1". Only used if \code{col = NULL}}

\item{axes.square}{Logical: If TRUE: draw a square plot to fill the graphic device.
Default = FALSE. Note: If TRUE, the device size at time of call is captured and height and width are set so as
to draw the largest square available. This means that resizing the device window will not automatically resize the
plot.}

\item{group.names}{Character, vector, length = NROW(x): Group names. Default = NULL, which uses \code{rownames(x)}}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{font.alpha}{Float (0, 1]: Transparency for fonts. Default = .8}

\item{font.col}{Color: Font color. Default = "black"}

\item{font.family}{String: Font family to use. Default = "Helvetica Neue"}

\item{main.col}{Color: Title color. Default = NULL, determined by theme}

\item{axes.col}{Color: Axes color. Default = NULL, determined, by theme}

\item{labs.col}{Color: Labels' color. Default = NULL, determined by theme}

\item{grid.col}{Color: Grid color. Default = "gray85"}

\item{grid.lwd}{Float: Grid line width. Default = 1}

\item{grid.alpha}{Float (0, 1]: Transparency for \code{grid.col}. Default = .8}

\item{tick.col}{Color: Color for ticks and tick labels. Default = NULL, determined, by theme}

\item{legend}{Logical: If TRUE, draw legend. Default = NULL, which will be set to TRUE if x is a list of more than
1 element}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by theme}

\item{margin}{Named list: plot margins. Default = \code{list(t = 35)}}

\item{zerolines}{Logical: If TRUE: draw lines at y = 0. Default = FALSE}

\item{width}{Float: Force plot size to this width. Default = NULL, i.e. fill available space}

\item{height}{Float: Force plot size to this height. Default = NULL, i.e. fill available space}

\item{filename}{Character: Path to file to save static plot. Default = NULL}

\item{file.width}{Integer: File width in pixels for when \code{filename} is set. Default = 500}

\item{file.height}{Integer: File height in pixels for when \code{filename} is set. Default = 500}

\item{histnorm}{Character: NULL, "percent", "probability", "density", "probability density"}

\item{histfunc}{Character: "count", "sum", "avg", "min", "max". Default = "count"}

\item{hist.n.bins}{Integer: Number of bins to use if type = "histogram". Default = 20}
}
\description{
Draw interactive scatter plots using \code{plotly}
}
\examples{
\dontrun{
dplot3.xy(iris$Sepal.Length, iris$Petal.Length, fit = "gam", se.fit = TRUE, group = iris$Species)
}
}
\author{
Efstathios D. Gennatas
}
