% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3.xy.R
\name{dplot3.xy}
\alias{dplot3.xy}
\title{Interactive Scatter Plots}
\usage{
dplot3.xy(x, y, fit = NULL, se.fit = FALSE, se.times = 1.96,
  cluster = NULL, cluster.params = list(k = 2), group = NULL,
  formula = NULL, rsq = TRUE, mode = "markers", order.on.x = NULL,
  main = NULL, xlab = NULL, ylab = NULL, col = NULL,
  alpha = 0.66, bg = NULL, plot.bg = NULL,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"), axes.square = FALSE,
  zero.lines = TRUE, group.names = NULL, font.size = 16,
  font.alpha = 0.8, font.col = NULL, font.family = "Helvetica Neue",
  marker.col = NULL, fit.col = NULL, fit.alpha = 0.8,
  fit.lwd = 2.5, se.col = NULL, se.alpha = 0.4, main.col = NULL,
  axes.col = NULL, labs.col = NULL, grid.col = NULL, grid.lwd = 1,
  grid.alpha = 0.8, tick.col = NULL, legend = TRUE,
  legend.xy = c(0, 1), legend.xanchor = "left",
  legend.yanchor = "auto", legend.orientation = "v",
  legend.col = NULL, legend.bg = "#FFFFFF00",
  legend.border.col = "#FFFFFF00", legend.borderwidth = 0,
  legend.group.gap = 0, margin = list(t = 35), zerolines = TRUE,
  mod.params = list(), width = NULL, height = NULL, padding = 0,
  trace = 0, filename = NULL, file.width = 500, file.height = 500,
  ...)
}
\arguments{
\item{x}{Numeric, vector: Input}

\item{y}{Numeric vector of list of vectors for y-axis
If \code{data} is provided, name of variable, unquoted.}

\item{fit}{Character: \pkg{rtemis} model to calculate y ~ x fit. Options: see \code{modSelect}
Can also be Logical, which will give a GAM fit if TRUE. If you specify "NLA", the activation function should
be passed as a string.}

\item{se.fit}{Logical: If TRUE, draw the standard error of the fit}

\item{se.times}{Draw polygon or lines at +/- \code{se.times} * standard error of fit. Defaults to 2
(1.96 * standard error corresponds to 95\% confidence interval)}

\item{cluster}{Character: Clusterer name. Will cluster \code{data.frame(x, y)} and
pass result to \code{group}. Run \link{clustSelect} for options}

\item{cluster.params}{List: Names list of parameters to pass to the \code{cluster} function}

\item{group}{Vector: will be converted to factor.
If \code{data} is provided, name of variable, unquoted.}

\item{formula}{Formula: Provide a formula to be solved using \link{s.NLS}. If provided, \code{fit} is
forced to \code{'nls'}. e.g. y ~ b * m ^ x for a power curve. Note: \code{nls} is prone to errors
and warnings, but is powerful. Use single letters for parameter names, no numbers.}

\item{rsq}{Logical: If TRUE, print R-squared values in legend if \code{fit} is set}

\item{mode}{Character, vector: "markers", "lines", "markers+lines". Default = "markers"}

\item{order.on.x}{Logical: If TRUE, order \code{x} and \code{y} on \code{x}. Default = NULL, which becomes
\code{TRUE} if \code{mode} includes lines.}

\item{main}{Character: Plot title. Default = NULL}

\item{xlab}{Character: x-axis label. Default = NULL}

\item{ylab}{Character: y-axis label. Default = NULL}

\item{col}{Color, vector: Color for bars. Default NULL, which will draw colors from \code{palette}}

\item{alpha}{Float (0, 1]: Transparency for bar colors. Default = .8}

\item{bg}{Color: Background color. Default = "white"}

\item{plot.bg}{Color: Background color for plot area. Default = "white"}

\item{theme}{Character: THeme to use: "light", "dark", "lightgrid", "darkgrid". Default = "lightgrid"}

\item{palette}{Character: Name of `rtemis` palette to use. Default = "rtCol1". Only used if \code{col = NULL}}

\item{axes.square}{Logical: If TRUE: draw a square plot to fill the graphic device.
Default = FALSE. Note: If TRUE, the device size at time of call is captured and height and width are set so as
to draw the largest square available. This means that resizing the device window will not automatically resize the
plot.}

\item{zero.lines}{Logical: If TRUE, draw lines at x = 0 and y = 0}

\item{group.names}{Character, vector, length = NROW(x): Group names. Default = NULL, which uses \code{rownames(x)}}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{font.alpha}{Float (0, 1]: Transparency for fonts. Default = .8}

\item{font.col}{Color: Font color. Default = "black"}

\item{font.family}{String: Font family to use. Default = "Helvetica Neue"}

\item{fit.lwd}{Float: Fit line width}

\item{se.col}{Color for \code{se.fit}}

\item{se.alpha}{Alpha for \code{se.fit}}

\item{main.col}{Color: Title color. Default = NULL, determined by theme}

\item{axes.col}{Color: Axes color. Default = NULL, determined, by theme}

\item{labs.col}{Color: Labels' color. Default = NULL, determined by theme}

\item{grid.col}{Color: Grid color. Default = "gray85"}

\item{grid.lwd}{Float: Grid line width. Default = 1}

\item{grid.alpha}{Float (0, 1]: Transparency for \code{grid.col}. Default = .8}

\item{tick.col}{Color: Color for ticks and tick labels. Default = NULL, determined, by theme}

\item{legend}{Logical: If TRUE, draw legend. Default = NULL, which will be set to TRUE if x is a list of more than
1 element}

\item{legend.orientation}{Character: "v": vertical, or "h": horizontal. Default = "v"}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by theme}

\item{margin}{Named list: plot margins. Default = \code{list(b = 50, l = 50, t = 50, r = 20)}}

\item{zerolines}{Logical: If TRUE: draw lines at y = 0. Default = FALSE}

\item{mod.params}{List: Arguments for learner defined by \code{fit}. Default = NULL, i.e. use default learner
parameters}

\item{width}{Float: Force plot size to this width. Default = NULL, i.e. fill available space}

\item{height}{Float: Force plot size to this height. Default = NULL, i.e. fill available space}

\item{padding}{Integer: N pixels to pad plot. Default = 0}

\item{trace}{Integer: If > 0, pass \code{verbose = TRUE} to the cluster and fit functions, if used. Default = 0}

\item{filename}{Character: Path to file to save static plot. Default = NULL}

\item{file.width}{Integer: File width in pixels for when \code{filename} is set. Default = 500}

\item{file.height}{Integer: File height in pixels for when \code{filename} is set. Default = 500}

\item{...}{Additional arguments to be passed to learner function}

\item{histnorm}{Character: NULL, "percent", "probability", "density", "probability density"}

\item{histfunc}{Character: "count", "sum", "avg", "min", "max". Default = "count"}

\item{hist.n.bins}{Integer: Number of bins to use if type = "histogram". Default = 20}
}
\description{
Draw interactive scatter plots using \code{plotly}
}
\examples{
\dontrun{
dplot3.xy(iris$Sepal.Length, iris$Petal.Length, fit = "gam", se.fit = TRUE, group = iris$Species)
}
}
\author{
Efstathios D. Gennatas
}
