% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_addtree.R
\name{dplot3_addtree}
\alias{dplot3_addtree}
\title{Plot AddTree trees}
\usage{
dplot3_addtree(
  addtree,
  col.positive = "#F48024DD",
  col.negative = "#18A3ACDD",
  node.col = "#666666",
  node.shape = "none",
  node.labels = TRUE,
  node.labels.pct.pos = NULL,
  pos.name = NULL,
  edge.col = "#999999",
  layout = "dot",
  rankdir = "TB",
  splines = "polyline",
  fontname = "helvetica",
  bg.color = "#ffffff",
  overlap = "false",
  prune = NULL,
  prune.empty.leaves = TRUE,
  remove.bad.parents = FALSE
)
}
\arguments{
\item{addtree}{Additive Tree object created by \link{s_AddTree}}

\item{col.positive}{Color for outcome positive.}

\item{col.negative}{Color for negative outcome.}

\item{node.col}{Color for non-terminal leaves.}

\item{node.shape}{Character: Node shape, passed to \code{data.tree::SetNodeStyle}}

\item{node.labels}{Logical: If \code{TRUE}, show node labels.}

\item{node.labels.pct.pos}{Logical: If \code{TRUE}, show \% positive cases in node labels.}

\item{pos.name}{Character: Name for "positive" outcome.}

\item{edge.col}{Color for edges.}

\item{layout}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{rankdir}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{splines}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{fontname}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{bg.color}{Background color.}

\item{overlap}{Character: Passed to \code{data.tree::SetGraphStyle}}

\item{prune}{Logical: If \code{TRUE}, prune AddTree.}

\item{prune.empty.leaves}{Logical: If \code{TRUE}, prune empty leaves.}

\item{remove.bad.parents}{Logical: If TRUE, remove nodes with no siblings but children and
give their children to their parent.}
}
\description{
Plot AddTree trees trained with \link{s_AddTree} using \code{data.tree::plot.Node}
}
\details{
Edge info and styles have been removed because of problems with \code{DiagrammeR}
}
\author{
E.D. Gennatas
}
