% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_calibration.R
\name{dplot3_calibration}
\alias{dplot3_calibration}
\title{Draw calibration plot}
\usage{
dplot3_calibration(
  true.labels,
  est.prob,
  n.windows = 10,
  bin.method = c("quantile", "equal"),
  pos.class.idi = 1,
  xlab = "Mean estimated probability",
  ylab = "Empirical risk",
  mode = "markers+lines",
  ...
)
}
\arguments{
\item{true.labels}{Factor with true class labels}

\item{est.prob}{Numeric vector with predicted probabilities}

\item{n.windows}{Integer: Number of windows to split the data into}

\item{bin.method}{Character: "quantile" or "equal": Method to bin the estimated
probabilities.}

\item{pos.class.idi}{Integer: Index of the positive class}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{mode}{Character: Plot mode}

\item{...}{Additional arguments passed to \link{dplot3_xy}}
}
\description{
Draw calibration plot
}
\examples{
\dontrun{
data(segment_logistic, package = "probably")

# Plot the calibration curve of the original predictions
dplot3_calibration(
  true.labels = segment_logistic$Class,
  est.prob = segment_logistic$.pred_poor,
  n.windows = 10,
  pos.class.idi = 2
)

# Plot the calibration curve of the calibrated predictions
dplot3_calibration(
  true.labels = segment_logistic$Class,
  est.prob = calibrate(
    segment_logistic$Class,
    segment_logistic$.pred_poor
  )$fitted.values,
  n.windows = 10,
  pos.class.idi = 2
)
}
}
\author{
EDG
}
