% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_calibration.R
\name{dplot3_calibration}
\alias{dplot3_calibration}
\title{Draw calibration plot}
\usage{
dplot3_calibration(
  true.labels,
  predicted.prob,
  n.bins = 10,
  bin.method = c("quantile", "equidistant"),
  pos.class = NULL,
  main = NULL,
  subtitle = NULL,
  xlab = "Mean predicted probability",
  ylab = "Empirical risk",
  show.marginal.x = TRUE,
  marginal.x.y = -0.02,
  marginal.col = NULL,
  marginal.size = 10,
  mode = "markers+lines",
  show.brier = TRUE,
  theme = rtTheme,
  filename = NULL,
  ...
)
}
\arguments{
\item{true.labels}{Factor or list of factors with true class labels}

\item{predicted.prob}{Numeric vector or list of numeric vectors with predicted probabilities}

\item{n.bins}{Integer: Number of windows to split the data into}

\item{bin.method}{Character: "quantile" or "equidistant": Method to bin the estimated
probabilities.}

\item{pos.class}{Integer: Index of the positive class}

\item{main}{Character: Main title}

\item{subtitle}{Character: Subtitle, placed bottom right of plot}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{show.marginal.x}{Logical: Add marginal plot of distribution of estimated probabilities}

\item{marginal.x.y}{Numeric: Y position of marginal markers on x-axis}

\item{marginal.col}{Color for marginal markers}

\item{marginal.size}{Numeric: Size of marginal markers}

\item{mode}{Character: "lines", "markers", "lines+markers": How to plot.}

\item{show.brier}{Logical: If TRUE, add Brier scores to trace names.}

\item{theme}{List or Character: Either the output of a \verb{theme_*()} function or the name of  a
theme. Use \code{themes()} to get available theme names. Theme functions are of the form
\verb{theme_<name>}.}

\item{filename}{Character: Path to save output.}

\item{...}{Additional arguments passed to \link{dplot3_xy}}
}
\description{
Draw calibration plot
}
\examples{
\dontrun{
data(segment_logistic, package = "probably")

# Plot the calibration curve of the original predictions
dplot3_calibration(
  true.labels = segment_logistic$Class,
  predicted.prob = segment_logistic$.pred_poor,
  n.bins = 10,
  pos.class = 2
)

# Plot the calibration curve of the calibrated predictions
dplot3_calibration(
  true.labels = segment_logistic$Class,
  predicted.prob = calibrate(
    segment_logistic$Class,
    segment_logistic$.pred_poor
  )$fitted.values,
  n.bins = 10,
  pos.class = 2
)
}
}
\author{
EDG
}
