% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_conf.R
\name{dplot3_conf}
\alias{dplot3_conf}
\title{Plot confusion matrix}
\usage{
dplot3_conf(
  x,
  true.col = "#72CDF4",
  false.col = "#FEB2E0",
  font.size = 18,
  main = NULL,
  main.y = 1,
  main.yanchor = "bottom",
  theme = rtTheme,
  margin = list(l = 75, r = 75, b = 75, t = 75),
  ...
)
}
\arguments{
\item{x}{Confusion matrix}

\item{true.col}{Color for true positives & true negatives}

\item{false.col}{Color for false positives & false negatives}

\item{font.size}{Integer: font size}

\item{main}{Character: plot title}

\item{main.y}{Numeric: y position of the title}

\item{main.yanchor}{Character: y anchor of the title}

\item{theme}{List or Character: Either the output of a \verb{theme_*()} function or the name of  a
theme. Use \code{themes()} to get available theme names. Theme functions are of the form
\verb{theme_<name>}.}

\item{margin}{List: Plot margins}

\item{...}{Additional arguments passed to theme}
}
\value{
A plotly object
}
\description{
Plot confusion matrix
}
\examples{
\dontrun{
true_labels <- factor(c("a", "b", "a", "a", "b", "b", "b", "b"))
predicted_labels <- factor(c("a", "a", "b", "a", "b", "b", "a", "a"))
error <- mod_error(true_labels, predicted_labels)
dplot3_conf(error$ConfusionMatrix)
}
}
\author{
EDG
}
