% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_protein.R
\name{dplot3_protein}
\alias{dplot3_protein}
\title{Plot the primary amino acid sequence of a protein}
\usage{
dplot3_protein(
  x,
  site = NULL,
  region = NULL,
  ptm = NULL,
  variant = NULL,
  disease.variants = NULL,
  n.per.row = NULL,
  main = NULL,
  main.xy = c(0.055, 0.975),
  main.xref = "paper",
  main.yref = "paper",
  main.xanchor = "middle",
  main.yanchor = "top",
  layout = c("simple", "grid", "1curve", "2curve"),
  show.markers = TRUE,
  show.labels = TRUE,
  font.size = 18,
  label.col = NULL,
  scatter.mode = "markers+lines",
  marker.size = 28,
  marker.col = NULL,
  marker.alpha = 1,
  marker.symbol = "circle",
  marker.hoverinfo = "text",
  line.col = NULL,
  line.alpha = 1,
  line.width = 2,
  show.full.names = TRUE,
  region.scatter.mode = "markers+lines",
  region.style = 3,
  region.marker.size = marker.size,
  region.marker.alpha = 0.6,
  region.marker.symbol = "circle",
  region.line.dash = "solid",
  region.line.shape = "line",
  region.line.smoothing = 1,
  region.line.width = 1,
  region.line.alpha = 0.6,
  theme = rtTheme,
  region.palette = rtPalette,
  region.outline.only = FALSE,
  region.outline.pad = 2,
  region.pad = 0.35,
  region.fill.alpha = 0.1666666,
  region.fill.shape = "line",
  region.fill.smoothing = 1,
  bpadcx = 0.5,
  bpadcy = 0.5,
  site.marker.size = marker.size,
  site.marker.symbol = marker.symbol,
  site.marker.alpha = 1,
  site.border.width = 1.5,
  site.palette = rtPalette,
  variant.col = "#FA6E1E",
  disease.variant.col = "#E266AE",
  showlegend.ptm = TRUE,
  ptm.col = 2:10,
  ptm.symbol = "circle",
  ptm.offset = 0.12,
  ptm.pad = 0.35,
  ptm.marker.size = marker.size/4.5,
  annotate.position.every = 10,
  annotate.position.alpha = 0.5,
  annotate.position.ay = -0.4 * marker.size,
  position.font.size = font.size - 6,
  legend.xy = c(0.97, 0.954),
  legend.xanchor = "left",
  legend.yanchor = "top",
  legend.orientation = "v",
  legend.col = NULL,
  legend.bg = "#FFFFFF00",
  legend.border.col = "#FFFFFF00",
  legend.borderwidth = 0,
  legend.group.gap = 0,
  margin = list(b = 0, l = 0, t = 0, r = 0, pad = 0),
  showgrid.x = FALSE,
  showgrid.y = FALSE,
  automargin.x = TRUE,
  automargin.y = TRUE,
  xaxis.autorange = TRUE,
  yaxis.autorange = "reversed",
  scaleanchor.y = "x",
  scaleratio.y = 1,
  hoverlabel.align = "left",
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  scrollZoom = TRUE,
  modebar.bg = "transparent",
  filename = NULL,
  file.width = 1320,
  file.height = 990,
  file.scale = 1,
  width = NULL,
  height = NULL,
  verbose = TRUE,
  trace = 0,
  ...
)
}
\arguments{
\item{x}{Character vector of amino acid sequence (1-letter abbreviations)}

\item{site}{Named list of lists with indices of sites. These will be
highlighted by coloring the border of markers}

\item{region}{Named list of lists with indices of regions. These will be
highlighted by coloring the markers and lines of regions using the
\code{palette} colors}

\item{ptm}{List of post-translational modifications}

\item{variant}{List of variant information}

\item{disease.variants}{List of disease variant information}

\item{n.per.row}{Integer: Number of amino acids to show per row}

\item{main}{Character: Main title}

\item{main.xy}{Numeric vector, length 2: x and y coordinates for title.
e.g. if \code{main.xref} and \code{main.yref} are \code{"paper"}:
\code{c(0.055, .975)} is top left, \code{c(.5, .975)} is top and
middle}

\item{main.xref}{Character: xref for title}

\item{main.yref}{Character: yref for title}

\item{main.xanchor}{Character: xanchor for title}

\item{main.yanchor}{Character: yanchor for title}

\item{layout}{Character: "1curve", "grid": type of layout to use}

\item{show.markers}{Logical: If TRUE, show amino acid markers}

\item{show.labels}{Logical: If TRUE, annotate amino acids with elements}

\item{modebar.bg}{Color for modebar background
from input \code{x}}

\item{...}{Additional arguments to pass to the theme function}
}
\description{
Plot the primary amino acid sequence of a protein
}
\examples{
\dontrun{
tau <- seqinr::read.fasta("https://rest.uniprot.org/uniprotkb/P10636.fasta",
    seqtype = "AA"
)
dplot3_protein(as.character(tau[[1]]))

# or directly using the UniProt accession number:
dplot3_protein("P10636")
}
}
\author{
E.D. Gennatas
}
