% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_protein.R
\name{dplot3_protein}
\alias{dplot3_protein}
\title{Plot the amino acid sequence with annotations}
\usage{
dplot3_protein(
  x,
  site = NULL,
  region = NULL,
  ptm = NULL,
  clv = NULL,
  variant = NULL,
  disease.variants = NULL,
  n.per.row = NULL,
  main = NULL,
  main.xy = c(0.055, 0.975),
  main.xref = "paper",
  main.yref = "paper",
  main.xanchor = "middle",
  main.yanchor = "top",
  layout = c("simple", "grid", "1curve", "2curve"),
  show.markers = TRUE,
  show.labels = TRUE,
  font.size = 18,
  label.col = NULL,
  scatter.mode = "markers+lines",
  marker.size = 28,
  marker.col = NULL,
  marker.alpha = 1,
  marker.symbol = "circle",
  line.col = NULL,
  line.alpha = 1,
  line.width = 2,
  show.full.names = TRUE,
  region.scatter.mode = "markers+lines",
  region.style = 3,
  region.marker.size = marker.size,
  region.marker.alpha = 0.6,
  region.marker.symbol = "circle",
  region.line.dash = "solid",
  region.line.shape = "line",
  region.line.smoothing = 1,
  region.line.width = 1,
  region.line.alpha = 0.6,
  theme = rtTheme,
  region.palette = rtPalette,
  region.outline.only = FALSE,
  region.outline.pad = 2,
  region.pad = 0.35,
  region.fill.alpha = 0.1666666,
  region.fill.shape = "line",
  region.fill.smoothing = 1,
  bpadcx = 0.5,
  bpadcy = 0.5,
  site.marker.size = marker.size,
  site.marker.symbol = marker.symbol,
  site.marker.alpha = 1,
  site.border.width = 1.5,
  site.palette = rtPalette,
  variant.col = "#FA6E1E",
  disease.variant.col = "#E266AE",
  showlegend.ptm = TRUE,
  ptm.col = NULL,
  ptm.symbol = "circle",
  ptm.offset = 0.12,
  ptm.pad = 0.35,
  ptm.marker.size = marker.size/4.5,
  clv.col = NULL,
  clv.symbol = "triangle-down",
  clv.offset = 0.12,
  clv.pad = 0.35,
  clv.marker.size = marker.size/4,
  annotate.position.every = 10,
  annotate.position.alpha = 0.5,
  annotate.position.ay = -0.4 * marker.size,
  position.font.size = font.size - 6,
  legend.xy = c(0.97, 0.954),
  legend.xanchor = "left",
  legend.yanchor = "top",
  legend.orientation = "v",
  legend.col = NULL,
  legend.bg = "#FFFFFF00",
  legend.border.col = "#FFFFFF00",
  legend.borderwidth = 0,
  legend.group.gap = 0,
  margin = list(b = 0, l = 0, t = 0, r = 0, pad = 0),
  showgrid.x = FALSE,
  showgrid.y = FALSE,
  automargin.x = TRUE,
  automargin.y = TRUE,
  xaxis.autorange = TRUE,
  yaxis.autorange = "reversed",
  scaleanchor.y = "x",
  scaleratio.y = 1,
  hoverlabel.align = "left",
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  scrollZoom = TRUE,
  filename = NULL,
  file.width = 1320,
  file.height = 990,
  file.scale = 1,
  width = NULL,
  height = NULL,
  verbosity = 1,
  ...
)
}
\arguments{
\item{x}{Character vector: amino acid sequence (1-letter abbreviations) OR
\code{a3} object OR Character: path to JSON file OR Character: UniProt accession number}

\item{site}{Named list of lists with indices of sites. These will be
highlighted by coloring the border of markers}

\item{region}{Named list of lists with indices of regions. These will be
highlighted by coloring the markers and lines of regions using the
\code{palette} colors}

\item{ptm}{List of post-translational modifications}

\item{clv}{List of cleavage sites}

\item{variant}{List of variant information}

\item{disease.variants}{List of disease variant information}

\item{n.per.row}{Integer: Number of amino acids to show per row}

\item{main}{Character: Main title}

\item{main.xy}{Numeric vector, length 2: x and y coordinates for title.
e.g. if \code{main.xref} and \code{main.yref} are \code{"paper"}:
\code{c(0.055, .975)} is top left, \code{c(.5, .975)} is top and
middle}

\item{main.xref}{Character: xref for title}

\item{main.yref}{Character: yref for title}

\item{main.xanchor}{Character: xanchor for title}

\item{main.yanchor}{Character: yanchor for title}

\item{layout}{Character: "1curve", "grid": type of layout to use}

\item{show.markers}{Logical: If TRUE, show amino acid markers}

\item{show.labels}{Logical: If TRUE, annotate amino acids with elements}

\item{font.size}{Integer: Font size for labels}

\item{label.col}{Color for labels}

\item{scatter.mode}{Character: Mode for scatter plot}

\item{marker.size}{Integer: Size of markers}

\item{marker.col}{Color for markers}

\item{marker.alpha}{Numeric: Alpha for markers}

\item{marker.symbol}{Character: Symbol for markers}

\item{line.col}{Color for lines}

\item{line.alpha}{Numeric: Alpha for lines}

\item{line.width}{Numeric: Width for lines}

\item{show.full.names}{Logical: If TRUE, show full names of amino acids}

\item{region.scatter.mode}{Character: Mode for scatter plot}

\item{region.style}{Integer: Style for regions}

\item{region.marker.size}{Integer: Size of region markers}

\item{region.marker.alpha}{Numeric: Alpha for region markers}

\item{region.marker.symbol}{Character: Symbol for region markers}

\item{region.line.dash}{Character: Dash for region lines}

\item{region.line.shape}{Character: Shape for region lines}

\item{region.line.smoothing}{Numeric: Smoothing for region lines}

\item{region.line.width}{Numeric: Width for region lines}

\item{region.line.alpha}{Numeric: Alpha for region lines}

\item{theme}{Character: Theme to use: Run \code{themes()} for available themes}

\item{region.palette}{Named list of colors for regions}

\item{region.outline.only}{Logical: If TRUE, only show outline of regions}

\item{region.outline.pad}{Numeric: Padding for region outline}

\item{region.pad}{Numeric: Padding for region}

\item{region.fill.alpha}{Numeric: Alpha for region fill}

\item{region.fill.shape}{Character: Shape for region fill}

\item{region.fill.smoothing}{Numeric: Smoothing for region fill}

\item{bpadcx}{Numeric: Padding for region border}

\item{bpadcy}{Numeric: Padding for region border}

\item{site.marker.size}{Integer: Size of site markers}

\item{site.marker.symbol}{Character: Symbol for site markers}

\item{site.marker.alpha}{Numeric: Alpha for site markers}

\item{site.border.width}{Numeric: Width for site borders}

\item{site.palette}{Named list of colors for sites}

\item{variant.col}{Color for variants}

\item{disease.variant.col}{Color for disease variants}

\item{showlegend.ptm}{Logical: If TRUE, show legend for PTMs}

\item{ptm.col}{Named list of colors for PTMs}

\item{ptm.symbol}{Character: Symbol for PTMs}

\item{ptm.offset}{Numeric: Offset for PTMs}

\item{ptm.pad}{Numeric: Padding for PTMs}

\item{ptm.marker.size}{Integer: Size of PTM markers}

\item{clv.col}{Color for cleavage site annotations}

\item{clv.symbol}{Character: Symbol for cleavage site annotations}

\item{clv.offset}{Numeric: Offset for cleavage site annotations}

\item{clv.pad}{Numeric: Padding for cleavage site annotations}

\item{clv.marker.size}{Integer: Size of cleavage site annotation markers}

\item{annotate.position.every}{Integer: Annotate every nth position}

\item{annotate.position.alpha}{Numeric: Alpha for position annotations}

\item{annotate.position.ay}{Numeric: Y offset for position annotations}

\item{position.font.size}{Integer: Font size for position annotations}

\item{legend.xy}{Numeric vector, length 2: x and y coordinates for legend}

\item{legend.xanchor}{Character: xanchor for legend}

\item{legend.yanchor}{Character: yanchor for legend}

\item{legend.orientation}{Character: Orientation for legend}

\item{legend.col}{Color for legend}

\item{legend.bg}{Color for legend background}

\item{legend.border.col}{Color for legend border}

\item{legend.borderwidth}{Numeric: Width for legend border}

\item{legend.group.gap}{Numeric: Gap between legend groups}

\item{margin}{List: Margin settings}

\item{showgrid.x}{Logical: If TRUE, show x grid}

\item{showgrid.y}{Logical: If TRUE, show y grid}

\item{automargin.x}{Logical: If TRUE, use automatic margin for x axis}

\item{automargin.y}{Logical: If TRUE, use automatic margin for y axis}

\item{xaxis.autorange}{Logical: If TRUE, use automatic range for x axis}

\item{yaxis.autorange}{Character: If TRUE, use automatic range for y axis}

\item{scaleanchor.y}{Character: Scale anchor for y axis}

\item{scaleratio.y}{Numeric: Scale ratio for y axis}

\item{hoverlabel.align}{Character: Alignment for hover label}

\item{displayModeBar}{Logical: If TRUE, display mode bar}

\item{modeBar.file.format}{Character: File format for mode bar}

\item{scrollZoom}{Logical: If TRUE, enable scroll zoom}

\item{filename}{Character: File name to save plot}

\item{file.width}{Integer: Width for saved file}

\item{file.height}{Integer: Height for saved file}

\item{file.scale}{Numeric: Scale for saved file}

\item{width}{Integer: Width for plot}

\item{height}{Integer: Height for plot}

\item{verbosity}{Integer: If > 0, print messages to console. If > 1, print
trace messages}

\item{...}{Additional arguments to pass to the theme function}
}
\value{
A plotly object
}
\description{
Plot the amino acid sequence with annotations
}
\examples{
\dontrun{
tau <- seqinr::read.fasta("https://rest.uniprot.org/uniprotkb/P10636.fasta",
  seqtype = "AA"
)
dplot3_protein(as.character(tau[[1]]))

# or directly using the UniProt accession number:
dplot3_protein("P10636")
}
}
\author{
E.D. Gennatas
}
