% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_volcano.R
\name{dplot3_volcano}
\alias{dplot3_volcano}
\title{Volcano Plot}
\usage{
dplot3_volcano(
  x,
  pvals,
  xnames = NULL,
  group = NULL,
  x.thresh = 0,
  p.thresh = 0.05,
  p.transform = function(x) -log10(x),
  p.adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  legend = NULL,
  legend.lo = NULL,
  legend.hi = NULL,
  label.lo = "Low",
  label.hi = "High",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  margin = list(b = 65, l = 65, t = 50, r = 10, pad = 0),
  xlim = NULL,
  ylim = NULL,
  alpha = NULL,
  hline = NULL,
  hline.col = NULL,
  hline.width = 1,
  hline.dash = "solid",
  hline.annotate = NULL,
  hline.annotation.x = 1,
  annotate = TRUE,
  annotate.col = theme$labs.col,
  theme = rtTheme,
  font.size = 16,
  palette = NULL,
  legend.x.lo = NULL,
  legend.x.hi = NULL,
  legend.y = 0.97,
  annotate.n = 7,
  ax.lo = NULL,
  ay.lo = NULL,
  ax.hi = NULL,
  ay.hi = NULL,
  annotate.alpha = 0.7,
  hovertext = NULL,
  displayModeBar = FALSE,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  file.scale = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector: Input values, e.g. log2 fold change, coefficients, etc.}

\item{pvals}{Numeric vector: p-values}

\item{xnames}{Character vector: \code{x} names}

\item{group}{Factor: Used to color code points. If NULL, significant points
below \code{x.thresh}, non-significant points, and significant points
above \code{x.thresh} will be plotted with the first, second and third
color fo \code{palette}}

\item{x.thresh}{Numeric x-axis threshold separating low from high}

\item{p.thresh}{Numeric: p-value threshold of significance. Default = .05}

\item{p.transform}{function. Default = \verb{\\(x) -log10(x)}}

\item{p.adjust.method}{Character: p-value adjustment method.
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"
Default = "holm". Use "none" for raw p-values.}

\item{legend}{Logical: If TRUE, show legend. Will default to FALSE, if
\code{group = NULL}, otherwise to TRUE}

\item{legend.lo}{Character: Legend to annotate significant points below the
\code{x.thresh}}

\item{legend.hi}{Character: Legend to annotate significant points above the
\code{x.thresh}}

\item{label.lo}{Character: label for low values}

\item{label.hi}{Character: label for high values}

\item{main}{Character: Main plot title.}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{margin}{Named list of plot margins.
Default = \code{list(b = 65, l = 65, t = 50, r = 10, pad = 0)}}

\item{xlim}{Numeric vector, length 2: x-axis limits}

\item{ylim}{Numeric vector, length 2: y-axis limits}

\item{alpha}{Numeric: point transparency}

\item{hline}{Numeric: If defined, draw a horizontal line at this y value.}

\item{hline.col}{Color for \code{hline}. Default = "#ff0000" (red)}

\item{hline.width}{Numeric: Width for \code{hline}. Default = 1}

\item{hline.dash}{Character: Type of line to draw: "solid", "dot", "dash",
"longdash", "dashdot",
or "longdashdot"}

\item{hline.annotate}{Character: Text of horizontal line annotation if
\code{hline} is set}

\item{hline.annotation.x}{Numeric: x position to place annotation with paper
as reference. 0: to the left of the plot area; 1: to the right of the plot area}

\item{annotate}{Logical: If TRUE, annotate significant points}

\item{annotate.col}{Color for annotations}

\item{theme}{List or Character: Either the output of a \verb{theme_*()} function or the name of  a
theme. Use \code{themes()} to get available theme names. Theme functions are of the form
\verb{theme_<name>}.}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{palette}{Character: Name of \pkg{rtemis} palette to use.
Default = "rtCol1". Only used if \code{col = NULL}}

\item{legend.x.lo}{Numeric: x position of \code{legend.lo}}

\item{legend.x.hi}{Numeric: x position of \code{legend.hi}}

\item{legend.y}{Numeric: y position for \code{legend.lo} and \code{legend.hi}}

\item{annotate.n}{Integer: Number of significant points to annotate}

\item{ax.lo}{Numeric: Sets the x component of the arrow tail about the arrow head for
significant points below \code{x.thresh}}

\item{ay.lo}{Numeric: Sets the y component of the arrow tail about the arrow head for
significant points below \code{x.thresh}}

\item{ax.hi}{Numeric: Sets the x component of the arrow tail about the arrow head for
significant points above \code{x.thresh}}

\item{ay.hi}{Numeric: Sets the y component of the arrow tail about the arrow head for
significant points above \code{x.thresh}}

\item{annotate.alpha}{Numeric: Transparency for annotations}

\item{hovertext}{List of character vectors with hovertext to include for each group
of markers}

\item{displayModeBar}{Logical: If TRUE, show plotly's modebar}

\item{filename}{Character: Path to file to save static plot. Default = NULL}

\item{file.width}{Integer: File width in pixels for when \code{filename} is
set.}

\item{file.height}{Integer: File height in pixels for when \code{filename}
is set.}

\item{file.scale}{Numeric: If saving to file, scale plot by this number}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{...}{Additional parameters passed to \link{dplot3_xy}}
}
\description{
Volcano Plot
}
\examples{
\dontrun{
set.seed(2019)
x <- rnormmat(500, 500)
y <- x[, 3] + x[, 5] - x[, 9] + x[, 15] + rnorm(500)
mod <- massGLM(y, x)
dplot3_volcano(mod$summary$`Coefficient y`, mod$summary$`p_value y`)
}

}
\author{
E.D. Gennatas
}
