% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_xt.R
\name{dplot3_xt}
\alias{dplot3_xt}
\title{Plot timeseries data}
\usage{
dplot3_xt(
  x,
  y,
  y2 = NULL,
  xname = NULL,
  yname = NULL,
  y2name = NULL,
  main = NULL,
  main.y = 1,
  main.yanchor = "bottom",
  show.rangeslider = FALSE,
  slider.start = NULL,
  slider.end = NULL,
  theme = rtTheme,
  font.size = 16,
  yfill = "tozeroy",
  y2fill = "none",
  line1.col = "#16A0AC",
  line1.width = 2,
  line1.fill.col = NULL,
  line2.col = "#FA6E1E",
  line2.fill.col = NULL,
  line2.width = 2,
  x.showspikes = TRUE,
  spike.dash = "solid",
  spike.col = NULL,
  x.spikethickness = -2,
  tickfont.size = 16,
  legend.x = 1,
  legend.y = 1.1,
  legend.xanchor = "right",
  legend.yanchor = "top",
  legend.orientation = "v",
  margin = list(l = 75, r = 75, b = 75, t = 75),
  x.standoff = 20L,
  y.standoff = 20L,
  y2.standoff = 20L,
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  scrollZoom = TRUE,
  filename = NULL,
  file.width = 960,
  file.height = 500,
  file.scale = 1,
  ...
)
}
\arguments{
\item{x}{Datetime vector}

\item{y}{Numeric vector}

\item{y2}{Numeric vector. If provided, a second y-axis will be added to the right
side of the plot}
}
\value{
A plotly object
}
\description{
Plot timeseries data
}
\author{
EDG
}
