% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_xt.R
\name{dplot3_xt}
\alias{dplot3_xt}
\title{Plot timeseries data}
\usage{
dplot3_xt(
  x,
  y = NULL,
  x2 = NULL,
  y2 = NULL,
  which.xy = NULL,
  which.xy2 = NULL,
  shade.bin = NULL,
  shade.interval = NULL,
  shade.col = NULL,
  shade.x = NULL,
  shade.name = "",
  shade.showlegend = FALSE,
  ynames = NULL,
  y2names = NULL,
  xlab = NULL,
  ylab = NULL,
  y2lab = NULL,
  xunits = NULL,
  yunits = NULL,
  y2units = NULL,
  yunits.col = NULL,
  y2units.col = NULL,
  zt = NULL,
  show.zt = TRUE,
  show.zt.every = NULL,
  zt.nticks = 18L,
  main = NULL,
  main.y = 1,
  main.yanchor = "bottom",
  x.nticks = 0,
  y.nticks = 0,
  show.rangeslider = NULL,
  slider.start = NULL,
  slider.end = NULL,
  theme = rtTheme,
  palette = rtpalette(rtPalette),
  font.size = 16,
  yfill = "none",
  y2fill = "none",
  fill.alpha = 0.2,
  yline.width = 2,
  y2line.width = 2,
  x.showspikes = TRUE,
  spike.dash = "solid",
  spike.col = NULL,
  x.spike.thickness = -2,
  tickfont.size = 16,
  x.tickmode = "auto",
  x.tickvals = NULL,
  x.ticktext = NULL,
  x.tickangle = NULL,
  legend.x = 0,
  legend.y = 1.1,
  legend.xanchor = "left",
  legend.yanchor = "top",
  legend.orientation = "h",
  margin = list(l = 75, r = 75, b = 75, t = 75),
  x.standoff = 20L,
  y.standoff = 20L,
  y2.standoff = 20L,
  hovermode = "x",
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  scrollZoom = TRUE,
  filename = NULL,
  file.width = 960,
  file.height = 500,
  file.scale = 1,
  ...
)
}
\arguments{
\item{x}{Datetime vector or list of vectors OR object of class \code{xt}. If \code{xt}, \code{x2}, \code{y}, \code{y2},
\code{xunits}, \code{yunits}, and \code{y2units} will be extracted from the object and the corresponding
arguments will be ignored.}

\item{y}{Numeric vector or named list of vectors: y-axis data.}

\item{x2}{Datetime vector or list of vectors, optional: must be provided if \code{y2} does not
correspond to values in \code{x}. A single x-axis will be drawn for all values in \code{x} and \code{x2}.}

\item{y2}{Numeric vector, optional: If provided, a second y-axis will be added to the right
side of the plot}

\item{which.xy}{Integer vector: Indices of \code{x} and \code{y} to plot.
If not provided, will select up to the first two x-y traces.}

\item{which.xy2}{Integer vector: Indices of \code{x2} and \code{y2} to plot.
If not provided, will select up to the first two x2-y2 traces.}

\item{shade.bin}{Integer vector {0, 1}: Time points in \code{x} to shade on the plot. For example,
if there are 10 time points in \code{x}, and you want to shade time points 3 to 7,
\code{shade.bin = c(0, 0, 1, 1, 1, 1, 1, 0, 0, 0)}. Only set \code{shade.bin} or \code{shade.interval}, not
both.}

\item{shade.interval}{List of numeric vectors: Intervals to shade on the plot. Only set
\code{shade.bin} or \code{shade.interval}, not both.}

\item{shade.col}{Color: Color to shade intervals.}

\item{shade.x}{Numeric vector: x-values to use for shading.}

\item{shade.name}{Character: Name for shaded intervals.}

\item{shade.showlegend}{Logical: If TRUE, show legend for shaded intervals.}

\item{ynames}{Character vector, optional: Names for each vector in \code{y}.}

\item{y2names}{Character vector, optional: Names for each vector in \code{y2}.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{y2lab}{Character: y2-axis label.}

\item{xunits}{Character: x-axis units.}

\item{yunits}{Character: y-axis units.}

\item{y2units}{Character: y2-axis units.}

\item{yunits.col}{Color for y-axis units.}

\item{y2units.col}{Color for y2-axis units.}

\item{zt}{Numeric vector: Zeitgeber time. If provided, will be shown on the x-axis instead of
\code{x}. To be used only with a single \code{x} vector and no \code{x2}.}

\item{show.zt}{Logical: If TRUE, show zt on x-axis, if zt is provided.}

\item{show.zt.every}{Integer: Show zt every \code{show.zt.every} ticks. If NULL, will be calculated
to be \code{x.nticks} +/- 1 if \code{x.nticks} is not 0, otherwise 12 +/- 1.}

\item{zt.nticks}{Integer: Number of zt ticks to show. Only used if \code{show.zt.every} is NULL.
The actual number of ticks shown will depend on the periodicity of zt, so that zt = 0 is always
included.}

\item{main}{Character: Main title.}

\item{main.y}{Numeric: Y position of main title.}

\item{main.yanchor}{Character: "top", "middle", "bottom".}

\item{x.nticks}{Integer: Number of ticks on x-axis.}

\item{y.nticks}{Integer: Number of ticks on y-axis.}

\item{show.rangeslider}{Logical: If TRUE, show a range slider.}

\item{slider.start}{Numeric: Start of range slider.}

\item{slider.end}{Numeric: End of range slider.}

\item{theme}{Character or list: Name of theme or list of plot parameters.}

\item{palette}{Color list: will be used to draw each vector in \code{y} and \code{y2}, in order.}

\item{font.size}{Numeric: Font size for text.}

\item{yfill}{Character: Fill type for y-axis: "none", "tozeroy", "tonexty"}

\item{y2fill}{Character: Fill type for y2-axis: "none", "tozeroy", "tonexty"}

\item{fill.alpha}{Numeric: Fill opacity for y-axis.}

\item{yline.width}{Numeric: Line width for y-axis lines.}

\item{y2line.width}{Numeric: Line width for y2-axis lines.}

\item{x.showspikes}{Logical: If TRUE, show spikes on x-axis.}

\item{spike.dash}{Character: Dash type for spikes: "solid", "dot", "dash", "longdash",
"dashdot", "longdashdot".}

\item{spike.col}{Color for spikes.}

\item{x.spike.thickness}{Numeric: Thickness of spikes. \code{-2} avoids drawing border around spikes.}

\item{tickfont.size}{Numeric: Font size for tick labels.}

\item{x.tickmode}{Character: "auto", "linear", "array".}

\item{x.tickvals}{Numeric vector: Tick positions.}

\item{x.ticktext}{Character vector: Tick labels.}

\item{x.tickangle}{Numeric: Angle of tick labels.}

\item{legend.x}{Numeric: X position of legend.}

\item{legend.y}{Numeric: Y position of legend.}

\item{legend.xanchor}{Character: "left", "center", "right".}

\item{legend.yanchor}{Character: "top", "middle", "bottom".}

\item{legend.orientation}{Character: "v" for vertical, "h" for horizontal.}

\item{margin}{Named list with 4 numeric values: "l", "r", "t", "b" for left, right, top, bottom
margins.}

\item{x.standoff}{Numeric: Distance from x-axis to x-axis label.}

\item{y.standoff}{Numeric: Distance from y-axis to y-axis label.}

\item{y2.standoff}{Numeric: Distance from y2-axis to y2-axis label.}

\item{hovermode}{Character: "closest", "x", "x unified"}

\item{displayModeBar}{Logical: If TRUE, display plotly mode bar.}

\item{modeBar.file.format}{Character: "png", "svg", "jpeg", "webp", "pdf": file format for mode
bar image export.}

\item{scrollZoom}{Logical: If TRUE, enable zooming by scrolling.}

\item{filename}{Character: Path to file to save static plot.}

\item{file.width}{Integer: File width in pixels for when \code{filename} is
set.}

\item{file.height}{Integer: File height in pixels for when \code{filename}
is set.}

\item{file.scale}{Numeric: If saving to file, scale plot by this number}

\item{...}{Additional theme arguments.}
}
\value{
A plotly object
}
\description{
Plot timeseries data
}
\details{
We are switching to \code{palette} being a color vector instead of the name of a built-in palette.
}
\author{
EDG
}
