% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot3_xyz.R
\name{dplot3_xyz}
\alias{dplot3_xyz}
\title{Interactive 3D Plots}
\usage{
dplot3_xyz(
  x,
  y = NULL,
  z = NULL,
  fit = NULL,
  cluster = NULL,
  cluster.params = list(k = 2),
  group = NULL,
  formula = NULL,
  rsq = TRUE,
  mode = "markers",
  order.on.x = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  col = NULL,
  alpha = 0.8,
  bg = NULL,
  plot.bg = NULL,
  theme = rtTheme,
  palette = rtPalette,
  axes.square = FALSE,
  group.names = NULL,
  font.size = 16,
  marker.col = NULL,
  marker.size = 8,
  fit.col = NULL,
  fit.alpha = 0.7,
  fit.lwd = 2.5,
  tick.font.size = 12,
  spike.col = NULL,
  legend = NULL,
  legend.xy = c(0, 1),
  legend.xanchor = "left",
  legend.yanchor = "auto",
  legend.orientation = "v",
  legend.col = NULL,
  legend.bg = "#FFFFFF00",
  legend.border.col = "#FFFFFF00",
  legend.borderwidth = 0,
  legend.group.gap = 0,
  margin = list(t = 30, b = 0, l = 0, r = 0),
  fit.params = list(),
  width = NULL,
  height = NULL,
  padding = 0,
  displayModeBar = TRUE,
  modeBar.file.format = "svg",
  trace = 0,
  filename = NULL,
  file.width = 500,
  file.height = 500,
  file.scale = 1,
  ...
)
}
\arguments{
\item{x}{Numeric, vector/data.frame/list: x-axis data. If y is NULL and
\code{NCOL(x) > 1}, first two columns used as \code{x} and \code{y}, respectively}

\item{y}{Numeric, vector/data.frame/list: y-axis data}

\item{z}{Numeric, vector/data.frame/list: z-axis data}

\item{fit}{Character: \pkg{rtemis} model to calculate \code{y ~ x} fit.
Options: see \code{select_learn()}
Can also be Logical, which will give a GAM fit if TRUE. If you specify
"NLA", the activation function should be passed as a string.}

\item{cluster}{Character: Clusterer name. Will cluster
\code{data.frame(x, y)} and pass result to \code{group}.
Run \link{select_clust} for options}

\item{cluster.params}{List: Names list of parameters to pass to the
\code{cluster} function}

\item{group}{Vector: Will be converted to factor; levels define group
members. Default = NULL}

\item{formula}{Formula: Provide a formula to be solved using \link{s_NLS}.
If provided, \code{fit} is forced to \code{'nls'}. e.g. y ~ b * m ^ x for a
power curve. Note: \code{nls} is powerful but is prone to errors
and warnings. Use single letters for parameter names, no numbers.}

\item{rsq}{Logical: If TRUE, print R-squared values in legend if \code{fit} is set}

\item{mode}{Character, vector: "markers", "lines", "markers+lines".}

\item{order.on.x}{Logical: If TRUE, order \code{x} and \code{y} on \code{x}. Default = NULL,
which becomes \code{TRUE} if \code{mode} includes lines.}

\item{main}{Character: Main plot title.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{zlab}{Character: z-axis label}

\item{col}{Color for markers. Default=NULL, which will draw colors from palette}

\item{alpha}{Float (0, 1]: Transparency for bar colors. Default = .8}

\item{bg}{Background color}

\item{plot.bg}{Plot background color}

\item{theme}{List or Character: Either the output of a \verb{theme_*()} function or the name of  a
theme. Use \code{themes()} to get available theme names. Theme functions are of the form
\verb{theme_<name>}.}

\item{palette}{Character: Name of \pkg{rtemis} palette to use.
Default = "rtCol1". Only used if \code{col = NULL}}

\item{axes.square}{Logical: If TRUE: draw a square plot to fill the graphic device.
Default = FALSE. Note: If TRUE, the device size at time of call is captured and
height and width are set so as to draw the largest square available. This means that
resizing the device window will not automatically resize the plot.}

\item{group.names}{Character, vector, length = NROW(x): Group names.
Default = NULL, which uses \code{rownames(x)}}

\item{font.size}{Float: Font size for all labels. Default = 16}

\item{marker.col}{Color for marker}

\item{marker.size}{Numeric: Marker size.}

\item{fit.col}{Color: Color of the fit line.}

\item{fit.alpha}{Float [0, 1]: Transparency for fit line}

\item{fit.lwd}{Float: Fit line width}

\item{tick.font.size}{Numeric: Tick font size}

\item{spike.col}{Spike lines color}

\item{legend}{Logical: If TRUE, draw legend. Default = NULL, which will be set to
TRUE if there are more than 1 groups, or \code{fit} is set}

\item{legend.xy}{Numeric, vector, length 2: x and y for plotly's legend}

\item{legend.xanchor}{Character: Legend's x anchor: "left", "center",
"right", "auto"}

\item{legend.yanchor}{Character: Legend's y anchor: "top", "middle",
"bottom", "auto"}

\item{legend.orientation}{"v" or "h" for vertical or horizontal}

\item{legend.col}{Color: Legend text color. Default = NULL, determined by
theme}

\item{legend.bg}{Color: Background color for legend}

\item{legend.border.col}{Color: Border color for legend}

\item{legend.borderwidth}{Numeric: Border width for legend}

\item{legend.group.gap}{Numeric: Gap between legend groups}

\item{margin}{Numeric, named list: Margins for top, bottom, left, right.
Default = \verb{list(t = 30, b = 0, l = 0, r = 0}}

\item{fit.params}{List: Arguments for learner defined by \code{fit}.
Default = NULL, i.e. use default learner parameters}

\item{width}{Numeric: Force plot size to this width. Default = NULL, i.e. fill
available space}

\item{height}{Numeric: Force plot size to this height. Default = NULL, i.e. fill
available space}

\item{padding}{Numeric: Graph padding.}

\item{displayModeBar}{Logical: If TRUE, show plotly's modebar}

\item{modeBar.file.format}{Character: "svg", "png", "jpeg", "pdf" / any
output file type supported by plotly and your system}

\item{trace}{Integer: The height the number the more diagnostic info is
printed to the console}

\item{filename}{Character: Path to file to save static plot. Default = NULL}

\item{file.width}{Integer: File width in pixels for when \code{filename} is
set.}

\item{file.height}{Integer: File height in pixels for when \code{filename}
is set.}

\item{file.scale}{Numeric: If saving to file, scale plot by this number}

\item{...}{Additional arguments passed to theme}
}
\description{
Draw interactive 3D plots using \code{plotly}
}
\details{
Note that \code{dplot3_xyz} uses the theme's \code{plot.bg} as \code{grid.col}
}
\examples{
\dontrun{
dplot3_xyz(iris, group = iris$Species, theme = "darkgrid")
}
}
\author{
E.D. Gennatas
}
