% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_3Dscatter.R
\name{draw_3Dscatter}
\alias{draw_3Dscatter}
\title{Interactive 3D Scatter Plots}
\usage{
draw_3Dscatter(
  x,
  y = NULL,
  z = NULL,
  fit = NULL,
  cluster = NULL,
  cluster_params = list(k = 2),
  group = NULL,
  formula = NULL,
  rsq = TRUE,
  mode = "markers",
  order_on_x = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  col = NULL,
  alpha = 0.8,
  bg = NULL,
  plot_bg = NULL,
  theme = choose_theme(),
  palette = rtemis_palette,
  axes_square = FALSE,
  group_names = NULL,
  font_size = 16,
  marker_col = NULL,
  marker_size = 8,
  fit_col = NULL,
  fit_alpha = 0.7,
  fit_lwd = 2.5,
  tick_font_size = 12,
  spike_col = NULL,
  legend = NULL,
  legend_xy = c(0, 1),
  legend_xanchor = "left",
  legend_yanchor = "auto",
  legend_orientation = "v",
  legend_col = NULL,
  legend_bg = "#FFFFFF00",
  legend_border_col = "#FFFFFF00",
  legend_borderwidth = 0,
  legend_group_gap = 0,
  margin = list(t = 30, b = 0, l = 0, r = 0),
  fit_params = NULL,
  width = NULL,
  height = NULL,
  padding = 0,
  displayModeBar = TRUE,
  modeBar_file_format = "svg",
  verbosity = 0L,
  filename = NULL,
  file_width = 500,
  file_height = 500,
  file_scale = 1
)
}
\arguments{
\item{x}{Numeric, vector/data.frame/list: x-axis data.}

\item{y}{Numeric, vector/data.frame/list: y-axis data.}

\item{z}{Numeric, vector/data.frame/list: z-axis data.}

\item{fit}{Character: Fit method.}

\item{cluster}{Character: Clustering method.}

\item{cluster_params}{List: Parameters for clustering.}

\item{group}{Factor: Grouping variable.}

\item{formula}{Formula: Formula for non-linear least squares fit.}

\item{rsq}{Logical: If TRUE, print R-squared values in legend if \code{fit} is set.}

\item{mode}{Character, vector: "markers", "lines", "markers+lines".}

\item{order_on_x}{Logical: If TRUE, order \code{x} and \code{y} on \code{x}.}

\item{main}{Character: Main title.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{zlab}{Character: z-axis label.}

\item{col}{Color for markers.}

\item{alpha}{Numeric: Alpha for markers.}

\item{bg}{Background color.}

\item{plot_bg}{Plot background color.}

\item{theme}{Theme object.}

\item{palette}{Character: Color palette.}

\item{axes_square}{Logical: If TRUE, draw a square plot.}

\item{group_names}{Character: Names for groups.}

\item{font_size}{Numeric: Font size.}

\item{marker_col}{Color for markers.}

\item{marker_size}{Numeric: Marker size.}

\item{fit_col}{Color for fit line.}

\item{fit_alpha}{Numeric: Alpha for fit line.}

\item{fit_lwd}{Numeric: Line width for fit line.}

\item{tick_font_size}{Numeric: Tick font size.}

\item{spike_col}{Spike lines color.}

\item{legend}{Logical: If TRUE, draw legend.}

\item{legend_xy}{Numeric: Position of legend.}

\item{legend_xanchor}{Character: X anchor for legend.}

\item{legend_yanchor}{Character: Y anchor for legend.}

\item{legend_orientation}{Character: Orientation of legend.}

\item{legend_col}{Color for legend text.}

\item{legend_bg}{Color for legend background.}

\item{legend_border_col}{Color for legend border.}

\item{legend_borderwidth}{Numeric: Border width for legend.}

\item{legend_group_gap}{Numeric: Gap between legend groups.}

\item{margin}{Numeric, named list: Margins for top, bottom, left, right.}

\item{fit_params}{Hyperparameters: Parameters for fit.}

\item{width}{Numeric: Width of plot.}

\item{height}{Numeric: Height of plot.}

\item{padding}{Numeric: Graph padding.}

\item{displayModeBar}{Logical: If TRUE, display mode bar.}

\item{modeBar_file_format}{Character: File format for mode bar.}

\item{verbosity}{Integer: Verbosity level.}

\item{filename}{Character: Filename to save plot.}

\item{file_width}{Numeric: Width of saved file.}

\item{file_height}{Numeric: Height of saved file.}

\item{file_scale}{Numeric: Scale of saved file.}
}
\value{
A \code{plotly} object.
}
\description{
Draw interactive 3D scatter plots using \code{plotly}.
}
\details{
Note that \code{draw_3Dscatter} uses the theme's \code{plot_bg} as \code{grid_col}.
}
\examples{
\dontrun{
draw_3Dscatter(iris, group = iris$Species, theme = theme_blackigrid())
}
}
\author{
EDG
}
