% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heat.R
\name{draw_heat}
\alias{draw_heat}
\title{Heatmap with \code{plotly}}
\usage{
draw_heat(
  z,
  x = NULL,
  y = NULL,
  title = NULL,
  palette = penn_heat(21),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  transpose = TRUE
)
}
\arguments{
\item{z}{Input matrix.}

\item{x}{Vectors for x-axis.}

\item{y}{Vectors for y-axis.}

\item{title}{Plot title.}

\item{palette}{Character vector: Colors to use.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{zlab}{z value label.}

\item{transpose}{Logical: If TRUE, transpose matrix.}
}
\value{
\code{plotly} object.
}
\description{
Draw a heatmap using \code{plotly}.
}
\details{
See \href{https://rdocs.rtemis.org/draw}{rdocs.rtemis.org/draw} for detailed documentation.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
mat <- matrix(rnorm(100), nrow = 10)
draw_heat(mat, x = 1:10, y = 1:10, title = "Heatmap Example")
\dontshow{\}) # examplesIf}
}
\author{
EDG
}
