% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_roc.R
\name{draw_roc}
\alias{draw_roc}
\title{Draw ROC curve}
\usage{
draw_roc(
  true_labels,
  predicted_prob,
  multiclass_fill_labels = TRUE,
  main = NULL,
  theme = choose_theme(),
  palette = rtpalette(getOption("rtemis_palette", "rtms")),
  legend = TRUE,
  legend_title = "Group (AUC)",
  legend_xy = c(1, 0),
  legend_xanchor = "right",
  legend_yanchor = "bottom",
  auc_dp = 3L,
  xlim = c(-0.05, 1.05),
  ylim = c(-0.05, 1.05),
  diagonal = TRUE,
  diagonal_col = NULL,
  axes_square = TRUE,
  filename = NULL,
  ...
)
}
\arguments{
\item{true_labels}{Factor: True outcome labels.}

\item{predicted_prob}{Numeric vector [0, 1]: Predicted probabilities for the positive class (i.e. second level of outcome).
Or, for multiclass, a matrix of predicted probabilities with one column per class.
Or, a list of such vectors/matrices to draw multiple ROC curves on the same plot.}

\item{multiclass_fill_labels}{Logical: If TRUE, fill in labels for multiclass ROC curves.
If FALSE, column names of \code{predicted_prob} must match levels of \code{true_labels}.}

\item{main}{Character: Main title for the plot.}

\item{theme}{\code{Theme} object.}

\item{palette}{Character vector: Colors to use.}

\item{legend}{Logical: If TRUE, draw legend.}

\item{legend_title}{Character: Title for the legend.}

\item{legend_xy}{Numeric vector: Position of the legend in the form c(x, y).}

\item{legend_xanchor}{Character: X anchor for the legend.}

\item{legend_yanchor}{Character: Y anchor for the legend.}

\item{auc_dp}{Integer: Number of decimal places for AUC values.}

\item{xlim}{Numeric vector: Limits for the x-axis.}

\item{ylim}{Numeric vector: Limits for the y-axis.}

\item{diagonal}{Logical: If TRUE, draw diagonal line.}

\item{diagonal_col}{Character: Color for the diagonal line.}

\item{axes_square}{Logical: If TRUE, make axes square.}

\item{filename}{Character: If provided, save the plot to this file.}

\item{...}{Additional arguments passed to \link{draw_scatter}.}
}
\value{
\code{plotly} object.
}
\description{
Draw ROC curve
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Binary classification
true_labels <- factor(c("A", "B", "A", "A", "B", "A", "B", "B", "A", "B"))
predicted_prob <- c(0.1, 0.4, 0.35, 0.8, 0.65, 0.2, 0.9, 0.55, 0.3, 0.7)
draw_roc(true_labels, predicted_prob)
\dontshow{\}) # examplesIf}
}
\author{
EDG
}
