% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_get_column_attr}
\alias{dt_get_column_attr}
\title{Tabulate column attributes}
\usage{
dt_get_column_attr(x, attr = "source", useNA = "always")
}
\arguments{
\item{x}{data.table}

\item{attr}{Character: Attribute to get}

\item{useNA}{Character: Passed to \code{table}}
}
\description{
Tabulate column attributes
}
\author{
E.D. Gennatas
}
