% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_get_factor_levels}
\alias{dt_get_factor_levels}
\title{Get factor levels from data.table}
\usage{
dt_get_factor_levels(dat)
}
\arguments{
\item{dat}{data.table

Note: If \code{dat} is not a data.table, it will be converted in place, i.e. the original
object will be modified.}
}
\value{
Named list of factor levels. Names correspond to column names.
}
\description{
Get factor levels from data.table
}
