% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_icd10_comorbidities}
\alias{dt_icd10_comorbidities}
\title{Calculate ICD10 comorbidities using \code{icd10} package}
\usage{
dt_icd10_comorbidities(
  x,
  score = c("ahrq", "elix", "quan_elix", "charlson", "ccs", "pccc")
)
}
\arguments{
\item{x}{data.frame-compatible input with first column being IDs and second column
ICD10 codes}

\item{score}{Character: Comorbidity scores to calculate.}
}
\value{
data.table with IDs & columns for each score
}
\description{
Calculates multiple comorbidity scores using the \code{icd10} package and merges
into a \code{data.table}.
}
\examples{
\dontrun{
library(icd10)
uranium_comorb <- dt_icd10_comorbidities(uranium_pathology)
uranium_comorb
}
}
\author{
E.D. Gennatas
}
