% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_inspect_type}
\alias{dt_inspect_type}
\title{Inspect column types}
\usage{
dt_inspect_type(x, cols = NULL, verbose = TRUE)
}
\arguments{
\item{x}{data.table}

\item{cols}{Character vector: columns to inspect.}

\item{verbose}{Logical: If TRUE, print messages to console.}
}
\description{
Will attempt to identify columns that should be numeric but are either character or
factor by running \link{inspect_type} on each column.
}
\author{
E.D. Gennatas
}
