% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_names_by_attr}
\alias{dt_names_by_attr}
\title{List column names by attribute}
\usage{
dt_names_by_attr(x, which, exact = TRUE, sorted = TRUE)
}
\arguments{
\item{x}{data.table}

\item{which}{Character: name of attribute}

\item{exact}{Logical: If TRUE, use exact matching}

\item{sorted}{Logical: If TRUE, sort the output}
}
\description{
List column names by attribute
}
\author{
E.D. Gennatas
}
