% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_names_by_class}
\alias{dt_names_by_class}
\title{List column names by class}
\usage{
dt_names_by_class(x, sorted = TRUE, item.format = hilite, maxlength = 24)
}
\arguments{
\item{x}{data.table}

\item{sorted}{Logical: If TRUE, sort the output}

\item{item.format}{Function: Function to format each item}

\item{maxlength}{Integer: Maximum number of items to print}
}
\description{
List column names by class
}
\author{
E.D. Gennatas
}
