% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{dt_pctmatch}
\alias{dt_pctmatch}
\title{Get N and percent match of values between two columns of two data.tables}
\usage{
dt_pctmatch(x, y, on = NULL, left_on = NULL, right_on = NULL, verbose = TRUE)
}
\arguments{
\item{x}{data.table}

\item{y}{data.table}

\item{on}{Integer or character: column to read in \code{x} and \code{y}, if it is the
same}

\item{left_on}{Integer or character: column to read in \code{x}}

\item{right_on}{Integer or character: column to read in \code{y}}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\description{
Get N and percent match of values between two columns of two data.tables
}
\author{
E.D. Gennatas
}
