% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eightBall.R
\name{eightBall}
\alias{eightBall}
\title{Magic 8-Ball}
\usage{
eightBall(question = NULL)
}
\arguments{
\item{question}{String: Your question for the magic 8-ball}
}
\description{
Magic 8-Ball
}
\author{
Efstathios D. Gennatas
}
