% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eightBall.R
\name{eightball}
\alias{eightball}
\title{Magic 8-Ball}
\usage{
eightball(question = NULL)
}
\arguments{
\item{question}{Character: Your question for the magic 8-ball}
}
\description{
Magic 8-Ball
}
\author{
E.D. Gennatas
}
