% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{mse}
\alias{mse}
\alias{msew}
\alias{rmse}
\alias{mae}
\title{Error functions}
\usage{
mse(x, y, na.rm = TRUE)

msew(x, y, weights = rep(1, length(y)), na.rm = TRUE)

rmse(x, y, na.rm = TRUE)

mae(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of True values}

\item{y}{Vector of Estimated values}

\item{na.rm}{Logical: If TRUE, remove NA values before computation. Default = TRUE}

\item{weights}{Float, vector: Case weights}
}
\description{
Convenience functions for calculating loss.
These can be passed as arguments to learners that support custom loss functions.
}
