% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_init.R
\name{errorSummary}
\alias{errorSummary}
\title{\code{rtemis-internals}: \code{errorSummary}}
\usage{
errorSummary(error, mod.name = NULL, pre = NULL)
}
\description{
Print Fit and Validation modError
}
\keyword{internal}
