% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{exc}
\alias{exc}
\title{Exclude columns by character or numeric vector.}
\usage{
exc(x, idx)
}
\arguments{
\item{x}{data.frame or similar.}

\item{idx}{Character or numeric vector: Column names or indices to exclude.}
}
\value{
data.frame, tibble, or data.table.
}
\description{
Exclude columns by character or numeric vector.
}
\examples{
exc(iris, "Species") |> head()
exc(iris, c(1, 3)) |> head()
}
\author{
EDG
}
