% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{expand.hytboost}
\alias{expand.hytboost}
\title{Expand boosting series}
\usage{
expand.hytboost(
  object,
  x,
  y = NULL,
  x.valid = NULL,
  y.valid = NULL,
  resid = NULL,
  mod.params = NULL,
  max.iter = 10,
  learning.rate = NULL,
  case.p = 1,
  cxrcoef = FALSE,
  prefix = NULL,
  verbose = TRUE,
  trace = 0,
  print.error.plot = "final"
)
}
\arguments{
\item{object}{\link{boost} object}

\item{x}{Data frame: Input features}

\item{y}{Vector: Output}

\item{mod.params}{Named list of arguments for \code{mod}}

\item{max.iter}{Integer: Maximum number of iterations (additive steps) to perform. Default = 10}

\item{learning.rate}{Float (0, 1] Learning rate for the additive steps}

\item{cxrcoef}{Logical: If TRUE, pass \code{cxr = TRUE, cxrcoef = TRUE} to \link{predict.hytreeRaw}}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If > 0, print diagnostic info to console}

\item{print.error.plot}{String or Integer: "final" plots a training and validation (if available) error curve at the
end of training. If integer, plot training and validation error curve every this many iterations
during training
for each base learner}
}
\description{
Add iterations to a \link{boost} object
}
\author{
Efstathios D. Gennatas
}
