% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{explain}
\alias{explain}
\title{Explain Supervised}
\usage{
explain(model, x, dat_training = NULL, method = NULL)
}
\arguments{
\item{model}{Supervised object.}

\item{x}{data.frame or similar: Data to explain.}

\item{dat_training}{data.frame or similar: Training data.}

\item{method}{Character: Method to use.}
}
\value{
Object depending on model: list, shapr, or other.
}
\description{
Explain Supervised Learning Model
}
\details{
The explain API is under development.
Different models require different inputs.
Currently, different explain methods output different objects.
This will likely be replaced in the future using a custom S7 class.
}
