% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain individual-level model predictions}
\usage{
explain(mod, x, digits = 2, top = NULL, trace = 0)
}
\arguments{
\item{mod}{\code{rtMod} object.}

\item{x}{Single-row data.frame of predictors.}

\item{digits}{Integer: Number of digits to round coefficients to.}

\item{top}{Integer: Number of top rules to show by absolute coefficient.}

\item{trace}{Integer: If > 0, print more messages to output.}
}
\description{
Explain individual-level model predictions
}
\author{
ED Gennatas
}
