% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_error.R
\name{f1}
\alias{f1}
\title{F1 score}
\usage{
f1(precision, recall)
}
\arguments{
\item{precision}{Float [0, 1]: Precision a.k.a. Positive Predictive Value}

\item{recall}{Float [0, 1]: Recall a.k.a. Sensitivity}
}
\description{
Calculate the F1 score for classification:
}
\details{
\deqn{F1 = 2 \frac{Recall \cdot Precision}{Recall + Precision}}{F1 = 2 * (Recall * Precision)/(Recall + Precision)}
}
\author{
E.D. Gennatas
}
