% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{factor_harmonize}
\alias{factor_harmonize}
\title{Factor harmonize}
\usage{
factor_harmonize(reference, x, verbose = TRUE)
}
\arguments{
\item{reference}{Reference factor}

\item{x}{Input factor}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Factor harmonize
}
