% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{factor_harmonize}
\alias{factor_harmonize}
\title{Factor harmonize}
\usage{
factor_harmonize(reference, x, verbosity = 1)
}
\arguments{
\item{reference}{Reference factor}

\item{x}{Input factor}

\item{verbosity}{Integer: If > 0, print messages to console.}
}
\description{
Factor harmonize
}
