% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{features}
\alias{features}
\title{Get features (all columns except the last one)}
\usage{
features(x, ...)
}
\arguments{
\item{x}{data.frame or similar.}
}
\value{
object same as input, after removing the last column.
}
\description{
Get features (all columns except the last one)
}
\author{
EDG
}
