% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{features}
\alias{features}
\title{Get features}
\usage{
features(x)
}
\arguments{
\item{x}{tabular data.}
}
\value{
object same as input, after removing the last column.
}
\description{
Returns all columns except the last one
}
\details{
This applied to tabular datasets used for supervised learning in rtemis,
where, by convention, the last column is the outcome variable and all other columns
are features.
}
\examples{
features(iris) |> head()
}
\author{
EDG
}
