% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7init.R
\name{features}
\alias{features}
\title{Get features from tabular data}
\usage{
features(x)
}
\arguments{
\item{x}{tabular data: Input data to get features from.}
}
\value{
Object of the same class as the input, after removing the last column.
}
\description{
Returns all columns except the last one.
}
\details{
This can be applied to tabular datasets used for supervised learning in \pkg{rtemis},
where, by convention, the last column is the outcome variable and all other columns
are features.
}
\examples{
features(iris) |> head()
}
\author{
EDG
}
