% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{filter_order}
\alias{filter_order}
\title{Filter order}
\usage{
filter_order(x, idl, decreasing = FALSE)
}
\description{
Filter order
}
\examples{
x <- rnorm(10)
x
x[filter_order(x, x < 0)]
}
\author{
E.D. Gennatas
}
\keyword{internal}
