% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_ops.R
\name{fmt_thresh}
\alias{fmt_thresh}
\title{Format rule thresholds for categorical splits}
\usage{
fmt_thresh(catsplit, feature, threshold, factor_levels)
}
\arguments{
\item{feature}{Character: feature name}

\item{threshold}{Character: threshold as reported by lightgbm}

\item{factor_levels}{Named list of factor levels. Names should correspond to training
set column names.}
}
\description{
Format rule thresholds for categorical splits
}
