% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm_ops.R
\name{fmt_thresh}
\alias{fmt_thresh}
\title{Format rules to convert integer levels to factor levels}
\usage{
fmt_thresh(feature, threshold, factor_levels)
}
\arguments{
\item{feature}{Character: feature name}

\item{threshold}{Character: threshold as reported by lightgbm}

\item{factor_levels}{Named list of factor levels. Names should correspond to training
set column names.}

\item{rules}{Character vector of rules}
}
\description{
Format rules to convert integer levels to factor levels
}
\keyword{Format}
\keyword{categorical}
\keyword{for}
\keyword{internal}
\keyword{rule}
\keyword{splits}
\keyword{thresholds}
