% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.call.R
\name{format.call}
\alias{format.call}
\title{Format method for \code{call} objects}
\usage{
\method{format}{call}(x, as.html = FALSE, class = "rtcode", ...)
}
\arguments{
\item{x}{\code{call} object}

\item{as.html}{Logical: If TRUE, output HTML span element}

\item{class}{Character: CSS class to assign to span containing code}

\item{...}{Not used}
}
\description{
Format method for \code{call} objects
}
\examples{
\dontrun{
irmod <- elevate(iris,
  mod = "cart",
  maxdepth = 2:3,
  n.resamples = 9,
  train.p = .85
)
format(irmod$call) |> cat()
}
}
\author{
E.D. Gennatas
}
