% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read data from file}
\usage{
get_data(
  filename,
  datadir = NULL,
  make.unique = TRUE,
  character2factor = FALSE,
  clean.colnames = TRUE,
  reader = c("data.table", "arrow", "vroom"),
  sep = NULL,
  verbose = TRUE,
  fread_verbose = FALSE,
  timed = verbose,
  ...
)
}
\arguments{
\item{filename}{Character: filename or full path if \code{datadir = NULL}}

\item{datadir}{Character: Optional path to directory where \code{filename} 
is located. If not specified, \code{filename} must be the full path.}

\item{make.unique}{Logical: If TRUE, keep unique rows only}

\item{character2factor}{Logical: If TRUE, convert character variables to 
factors}

\item{clean.colnames}{Logical: If TRUE, clean columns names using 
\link{clean_colnames}}

\item{reader}{Character: "data.table" or "arrow", to use 
\code{data.table::fread()} or \code{arrow::read_delim_arrow()}, respectively,
to read \code{filename}}

\item{sep}{Single character: field separator. If \code{reader = "fread"}
and \code{sep = NULL}, this defaults to "auto", otherwise defaults to ","}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{fread_verbose}{Logical: Passed to \code{data.table::fread}}

\item{timed}{Logical: If TRUE, time the process and print to console}

\item{...}{Additional parameters to pass to \code{data.table::fread}, 
\code{arrow::read_delim_arrow()} or \code{vroom::vroom()}}
}
\description{
Convenience function to read data into a data.table using either
\code{data.table:fread()} or \code{arrow:read_delim_arrow()}
}
\examples{
\dontrun{
datadir <- "~/icloud/Data"
dat <- get_data("iris.csv", datadir)
}
}
\author{
E.D. Gennatas
}
