% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read data from file}
\usage{
get_data(filename, datadir = NULL, make_unique = TRUE, verbose = TRUE)
}
\arguments{
\item{filename}{Character: filename or full path if \code{datadir = NULL}}

\item{datadir}{Character: path to directory where \code{filename} is located}

\item{make_unique}{Logical: If TRUE, keep unique rows only}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\description{
Convenience function to read data into a data.table
}
\examples{
\dontrun{
datadir <- "~/icloud/Data"
dat <- get_data("iris.csv", datadir)
}
}
\author{
E.D. Gennatas
}
