% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loaded_pkg_version.R
\name{get_loaded_pkg_version}
\alias{get_loaded_pkg_version}
\title{Get version of all loaded packages (namespaces)}
\usage{
get_loaded_pkg_version()
}
\value{
Data frame with columns "Package_Name" and "Version"
}
\description{
Get version of all loaded packages (namespaces)
}
\author{
E.D. Gennatas
}
