% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{get_mode}
\alias{get_mode}
\title{Get the mode of a factor or integer}
\usage{
get_mode(x, na.exclude = TRUE, getlast = TRUE, retain.class = TRUE)
}
\arguments{
\item{x}{Vector, factor or integer: Input data}

\item{na.exclude}{Logical: If TRUE, exclude NAs}

\item{getlast}{Logical: If TRUE, get}

\item{retain.class}{Logical: If TRUE, output is always same class as input}
}
\value{
The mode of \code{x}
}
\description{
Returns the mode of a factor or integer
}
\examples{
x <- c(9, 3, 4, 4, 0, 2, 2, NA)
get_mode(x)
x <- c(9, 3, 2, 2, 0, 4, 4, NA)
get_mode(x)
get_mode(x, getlast = FALSE)

}
\author{
E.D. Gennatas
}
