% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_init.R
\name{get_varimp}
\alias{get_varimp}
\title{Get variable importance}
\usage{
get_varimp(x, ...)
}
\arguments{
\item{x}{\code{Supervised} or \code{SupervisedRes} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
\code{VariableImportance} object or list of \code{VariableImportance} objects.
}
\description{
Get variable importance
}
\examples{
mod <- train(iris, hyperparameters = setup_LightRF())
get_varimp(mod)
}
\author{
EDG
}
