% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7init.R
\name{get_varimp}
\alias{get_varimp}
\title{Get variable importance}
\usage{
get_varimp(x, ...)
}
\arguments{
\item{x}{\code{Supervised} or \code{SupervisedRes} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Numeric vector or data.table of variable importance scores.
}
\description{
Get variable importance
}
\examples{
mod <- train(iris, algorithm = "LightRF")
get_varimp(mod)
}
\author{
EDG
}
