% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_light.R
\name{ggtheme_light}
\alias{ggtheme_light}
\title{\pkg{rtemis} \code{ggplot2} light theme}
\usage{
ggtheme_light(
  base_size = 14,
  base_family = "Helvetica Neue",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  axis.text.size.rel = 1,
  legend.key.fill = NA,
  legend.text.size.rel = 1,
  legend.position = "right",
  strip.background.fill = "grey85"
)
}
\arguments{
\item{base_size}{Float: Base font size. Default = 14}

\item{base_family}{Character: Font family. Default = "Helvetica Neue"}

\item{base_line_size}{Float: Line size. Default = base_size/22}

\item{base_rect_size}{Float: Size for rect elements. Default = base_size/22}

\item{axis.text.size.rel}{Float: Relative size for axis text. Default = 1}

\item{legend.key.fill}{Color: Fill color for legend. Default = NA (no color)}

\item{legend.text.size.rel}{Float: Relative size for legend text. Default = 1}

\item{legend.position}{Character: Legend position, "top", "bottom", "right", "left" Default  = "right"}

\item{strip.background.fill}{Color: Fill color from facet labels. Default = "grey85"}
}
\description{
\pkg{rtemis} \code{ggplot2} light theme
}
\examples{
\dontrun{
(p <- ggplot(iris, aes(Sepal.Length, Petal.Length, color = Species)) +
geom_point() +
ggtheme_light())
}
}
\author{
E.D. Gennatas
}
