% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm2table.R
\name{glm2table}
\alias{glm2table}
\title{Collect summary table from list of massGLMs with same predictors, different outcome ("massy")}
\usage{
glm2table(x, xnames = NULL, include_anova_pvals = NA)
}
\arguments{
\item{x}{list of \link{glm} models}

\item{xnames}{Character, vector: names of models}
}
\value{
\code{data.table} with glm summaries
}
\description{
Collect summary table from list of massGLMs with same predictors, different outcome ("massy")
}
\author{
E.D. Gennatas
}
