% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot2factor.R
\name{\%BC\%}
\alias{\%BC\%}
\title{Binary matrix times character vector}
\usage{
x \%BC\% labels
}
\arguments{
\item{x}{A binary matrix or data.frame}

\item{labels}{Character vector length equal to \code{ncol(x)}}
}
\value{
a character vector
}
\description{
Binary matrix times character vector
}
\author{
E.D. Gennatas
}
