% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{hytboost}
\alias{hytboost}
\title{Boosting of Hybrid Trees}
\usage{
hytboost(
  x,
  y,
  x.valid = NULL,
  y.valid = NULL,
  resid = NULL,
  boost.obj = NULL,
  case.p = 0.9,
  learning.rate = 0.1,
  max.iter = 10,
  max.depth = 5,
  gamma = 0.1,
  alpha = 0,
  lambda = 1,
  lambda.seq = NULL,
  minobsinnode = 2,
  minobsinnode.lin = 10,
  shrinkage = 1,
  part.minsplit = 2,
  part.xval = 0,
  part.max.depth = 1,
  part.cp = 0,
  part.minbucket = 5,
  lin.type = "glmnet",
  cv.glmnet.nfolds = 5,
  which.cv.glmnet.lambda = "lambda.min",
  earlystop.params = rtset.earlystop(),
  earlystop.using = "train",
  init = mean(y),
  cxrcoef = FALSE,
  print.progress.every = 5,
  print.error.plot = "final",
  base.verbose = FALSE,
  verbose = TRUE,
  trace = 0,
  prefix = NULL,
  print.plot = TRUE,
  plot.theme = "darkgrid",
  plot.type = "l"
)
}
\arguments{
\item{x}{Data frame: Input features}

\item{y}{Vector: Output}

\item{learning.rate}{Float (0, 1] Learning rate for the additive steps}

\item{max.iter}{Integer: Maximum number of iterations (additive steps) to perform.}

\item{init}{Float: Initial value for prediction. Default = mean(y)}

\item{cxrcoef}{Logical: If TRUE, pass \verb{cxr = TRUE, cxrcoef = TRUE} to
\link{predict.hytreew}}

\item{print.error.plot}{String or Integer: "final" plots a training and validation
(if available) error curve at the end of training. If integer, plot training and
validation error curve every this many iterations during training for each base
learner}

\item{base.verbose}{Logical: \code{verbose} argument passed to learner}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: If > 0, print diagnostic info to console}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}.}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{mod}{Algorithm to boost, for options, see \link{learnSelect}}

\item{mod.params}{Named list of arguments for \code{mod}}

\item{tolerance}{Float: If training error <= this value, training stops}

\item{tolerance.valid}{Float: If validation error <= this value, training stops}

\item{...}{Additional parameters to be passed to learner}
}
\value{
\code{hytboost} object
}
\description{
Boosted Hybrid Trees. This is a lower-level training function
}
\author{
E.D. Gennatas
}
\keyword{internal}
