% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7init.R
\name{inc}
\alias{inc}
\title{Select (include) columns by character or numeric vector.}
\usage{
inc(x, idx)
}
\arguments{
\item{x}{tabular data.}

\item{idx}{Character or numeric vector: Column names or indices to include.}
}
\value{
data.frame, tibble, or data.table.
}
\description{
Select (include) columns by character or numeric vector.
}
\examples{
inc(iris, c(3, 4)) |> head()
inc(iris, c("Sepal.Length", "Species")) |> head()
}
\author{
EDG
}
