% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_check.R
\name{inherits_test}
\alias{inherits_test}
\title{Check class of object}
\usage{
inherits_test(x, cl)
}
\arguments{
\item{x}{Object to check}

\item{cl}{Character: class to check against}
}
\value{
Logical
}
\description{
Check class of object
}
\examples{
\dontrun{
inherits_check("papaya", "character") # TRUE
inherits_check(c(1, 2.5, 3.2), "integer") # FALSE
inherits_check(iris, "list") # FALSE, compare to is_check(iris, is.list)
}
}
\author{
EDG
}
\keyword{internal}
