% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_project_dir}
\alias{init_project_dir}
\title{Initialize Project Directory}
\usage{
init_project_dir(verbosity = 1L)
}
\arguments{
\item{verbosity}{Integer: Verbosity level.}
}
\value{
Character: the working directory path, invisibly.
}
\description{
Initializes Directory Structure: "R", "Data", "Results"
}
\author{
EDG
}
