% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{inspect}
\alias{inspect}
\title{Inspect rtemis object}
\usage{
inspect(x)
}
\arguments{
\item{x}{R object to inspect.}
}
\value{
Called for side effect of printing information to console; returns character string
invisibly.
}
\description{
Inspect rtemis object
}
\examples{
inspect(iris)
}
\author{
EDG
}
