% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{inspect_type}
\alias{inspect_type}
\title{Inspect character and factor vector}
\usage{
inspect_type(x, xname = NULL, verbose = TRUE, thresh = 0.5, na.omit = TRUE)
}
\arguments{
\item{x}{Character or factor vector.}

\item{xname}{Character: Name of input vector \code{x}.}

\item{verbose}{Logical: If TRUE, print messages to console.}

\item{thresh}{Numeric: Threshold for determining whether to convert to numeric.}

\item{na.omit}{Logical: If TRUE, remove NA values before checking.}
}
\description{
Checks character or factor vector to determine whether it might be best to convert to
numeric.
}
\examples{
\dontrun{
x <- c("3", "5", "undefined", "21", "4", NA)
inspect_type(x)
z <- c("mango", "banana", "tangerine", NA)
inspect_type(z)
}
}
\author{
E.D. Gennatas
}
