% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{is_discrete}
\alias{is_discrete}
\title{Check if variable is discrete (factor or integer)}
\usage{
is_discrete(x)
}
\arguments{
\item{x}{Input}
}
\description{
Check if variable is discrete (factor or integer)
}
\author{
E.D. Gennatas
}
