% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelify.R
\name{labelify}
\alias{labelify}
\title{Convert text for label printing}
\usage{
labelify(x, underscoresToSpaces = TRUE, dotsToSpaces = TRUE,
  toTitleCase = TRUE, capitalize.strings = c("id"),
  stringsToSpaces = c("\\\\$"))
}
\arguments{
\item{x}{Character: Input}

\item{underscoresToSpaces}{Logical: If TRUE, convert underscores to spaces. Default = TRUE}

\item{dotsToSpaces}{Logical: If TRUE, convert dots to spaces. Default = TRUE}

\item{toTitleCase}{Logical: If TRUE, convert to Title Case. Default = TRUE}

\item{capitalize.strings}{String, vector: Always capitalize these strings, if present. Default = "id"}

\item{stringsToSpaces}{String, vector: Replace these strings with spaces. Escape as needed for \code{gsub}.
Default = "\\$", which formats common input of the type \code{data.frame$variable}}
}
\description{
Convert text for label printing
}
\author{
Efstathios D. Gennatas
}
