% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelify.R
\name{labelify}
\alias{labelify}
\title{Convert text for label printing}
\usage{
labelify(
  x,
  underscoresToSpaces = TRUE,
  dotsToSpaces = TRUE,
  toLower = FALSE,
  toTitleCase = TRUE,
  capitalize.strings = c("id"),
  stringsToSpaces = c("\\\\$")
)
}
\arguments{
\item{x}{Character: Input}

\item{underscoresToSpaces}{Logical: If TRUE, convert underscores to spaces. Default = TRUE}

\item{dotsToSpaces}{Logical: If TRUE, convert dots to spaces. Default = TRUE}

\item{toLower}{Logical: If TRUE, convert to lowercase (precedes \code{toTitleCase}).
Default = FALSE (Good for getting all-caps words converted to title case, bad for abbreviations
you want to keep all-caps)}

\item{toTitleCase}{Logical: If TRUE, convert to Title Case. Default = TRUE (This does not change
all-caps words, set \code{toLower} to TRUE if desired)}

\item{capitalize.strings}{Character, vector: Always capitalize these strings, if present. Default = "id"}

\item{stringsToSpaces}{Character, vector: Replace these strings with spaces. Escape as needed for \code{gsub}.
Default = "\\$", which formats common input of the type \code{data.frame$variable}}
}
\description{
Convert text for label printing
}
\author{
E.D. Gennatas
}
