% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels2nii.R
\name{labels2nii}
\alias{labels2nii}
\title{Write Data to a Nifti File}
\usage{
labels2nii(label.vals, labeledNifti, prefix, datatype = "auto", verbose = TRUE)
}
\arguments{
\item{label.vals}{Vector: The index of this vector should correspond to label numbers.
Its values will replace the corresponding label numbers to form the new nifti file.
This means that the values of this vector need to be in the same order as the labels.
For example if \code{labelefNifti} has four labels: 0 (background), 10, 11, 20, then your
\code{label.vals} should should have three values and they will be assigned to labels
10, 11, and 20.}

\item{labeledNifti}{Character: Path to the labeled file whose labels corresponds to the values in
\code{label.vals}}

\item{prefix}{Character: Prefix of the output file. ".nii.gz" will be added automatically}

\item{datatype}{Integer or "auto": Defines the datatype of the output. Options:
2: uint8, 4: int16, 8: int32, 16: float (Default), 64: double. Default = "auto"}

\item{verbose}{Logical: If TRUE, print messages to output}
}
\description{
Writes ROI values to nifti
}
\author{
E.D. Gennatas
}
