% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotri2edgeList.R
\name{lotri2edgeList}
\alias{lotri2edgeList}
\title{Connectivity Matrix to Edge List}
\usage{
lotri2edgeList(A, filename = NULL, verbose = TRUE)
}
\arguments{
\item{A}{Square matrix}

\item{filename}{Character: Path for csv file. Defaults to "conmat2edgelist.csv"}

\item{verbose}{Logical: If TRUE, print messages to console}
}
\description{
Turn the lower triangle of a connectivity matrix (e.g. correlation matrix or similar)
to an edge list of the form:
Source, Target, Weight
}
\details{
The output can be read, for example, into gephi
}
\author{
E.D. Gennatas
}
