% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{make_SupervisedCV}
\alias{make_SupervisedCV}
\title{Make SupervisedCV}
\usage{
make_SupervisedCV(
  algorithm,
  type,
  models,
  preprocessor,
  hyperparameters,
  tuner_parameters,
  crossvalidation_resampler,
  y_training,
  y_test,
  predicted_training,
  predicted_test,
  predicted_prob_training,
  predicted_prob_test,
  se_training = NULL,
  se_test = NULL,
  xnames = character(),
  varimp = NULL,
  question = character(),
  extra = NULL
)
}
\description{
Make SupervisedCV
}
\author{
EDG
}
\keyword{internal}
