% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGAM.R
\name{massGAM}
\alias{massGAM}
\title{Mass-univariate GAM Analysis}
\usage{
massGAM(
  x,
  y,
  covariates = NULL,
  x.name = NULL,
  y.name = NULL,
  k = NULL,
  family = gaussian(),
  weights = NULL,
  method = "REML",
  n.cores = rtCores,
  save.mods = FALSE,
  save.summary = TRUE,
  print.plots = FALSE,
  outdir = NULL,
  labeledNifti = NULL,
  save.plots = FALSE,
  new.x.breaks = 9
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Float, Matrix / data frame: Outcomes}

\item{x.name}{Character: Name of the predictor}

\item{y.name}{Character, vector: Names of the outcomes}

\item{k}{Integer. Number of bases for smoothing spline}

\item{family}{\code{family} argument for \code{mgcv::gam}}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{method}{Estimation method for GAM}

\item{n.cores}{Integer. Number of cores to use}

\item{save.mods}{Logical. Should models be saved}

\item{save.summary}{Logical. Should model summary be saved}

\item{print.plots}{Logical Should plots be shown}

\item{outdir}{Path to save output}

\item{labeledNifti}{String. Path to labeled nifti file.}

\item{save.plots}{Logical. Should plots be saved}

\item{new.x.breaks}{Integer. Number of splits in the range of x
to form vector of features for estimation of fitted values}

\item{weight}{Vector. Weights for GAM}
}
\description{
Fits a GAM for each of multiple outcomes using a fixed set of features (many y's, one X).
}
\details{
NA in the input will be kept as NA in the results, maintaining n of cases.
}
\author{
Eftsthios D. Gennatas
}
