% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R
\name{massGLM}
\alias{massGLM}
\title{Mass-univariate GLM Analysis}
\usage{
massGLM(
  x,
  y,
  type = NULL,
  xnames = NULL,
  ynames = NULL,
  save.mods = TRUE,
  print.plot = FALSE,
  verbose = TRUE,
  trace = 0,
  n.cores = rtCores
)
}
\arguments{
\item{x}{Matrix / data frame of features}

\item{y}{Matrix / data frame of outcomes}

\item{type}{Character: "massx" or "massy". Default = NULL,
where if (NCOL(x) > NCOL(y)) "massx" else "massy"}

\item{xnames}{Character vector: names of \code{x} feature(s)}

\item{ynames}{Character vector: names of \code{y} feature(s)}

\item{save.mods}{Logical: If TRUE, save models. Default = TRUE}

\item{print.plot}{Logical: If TRUE, print plot. Default = FALSE (best to choose which p-values
you want to plot directly)}

\item{verbose}{Logical: If TRUE, print messages during run}

\item{trace}{Integer: If > 0, print more verbose output to console. Default = 0}

\item{n.cores}{Integer: Number of cores to use}
}
\description{
Run a mass-univariate analysis with either:
a) single outome (y) and multiple predictors (x), one at a time, with optional a common set of
covariates in each model - "massx"
b) multiple different outcomes (y) with a fixed set of predictors (x) - "massy"
Therefore the term mass-univariate refers to looking at one variable of interest (with
potential covariates of no interest) at a time
}
\author{
E.D. Gennatas
}
