% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R
\name{massGLM}
\alias{massGLM}
\title{Mass-univariate GLM Analysis}
\usage{
massGLM(
  x,
  y,
  xnames = NULL,
  ynames = NULL,
  save.mods = FALSE,
  p.adjust.method = "holm",
  print.plot = TRUE,
  verbose = TRUE,
  trace = 0,
  n.cores = rtCores
)
}
\arguments{
\item{x}{Matrix / data frame of features}

\item{y}{Matrix / data frame of outcomes}

\item{xnames}{Character vector: names of \code{x} feature(s)}

\item{ynames}{Character vector: names of \code{y} feature(s)}

\item{verbose}{Logical: If TRUE, print messages during run}

\item{n.cores}{Integer: Number of cores to use}

\item{...}{Arguments to be passed to \code{mod}}
}
\description{
Run a mass-univariate analysis with either:
a) single outome (y) and multiple predictors (x), one at a time, with optional a common set of
covariates in each model - "massx"
b) multiple different outcomes (y) with a fixed set of predictors (x) - "massy"
Therefore the term mass-univariate refers to looking at one variable of interest (with
potential covariates of no interest) at a time
}
\author{
E.D. Gennatas
}
