% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R
\name{massGLM}
\alias{massGLM}
\title{Mass-univariate GLM Analysis}
\usage{
massGLM(x, y, scale_y = NULL, center_y = NULL, verbosity = 1L)
}
\arguments{
\item{x}{tabular data: Predictor variables. Usually a small number of covariates.}

\item{y}{data.frame or similar: Each column is a different outcome. The function will train one
GLM for each column of \code{y}. Usually a large number of features.}

\item{scale_y}{Logical: If TRUE, scale each column of \code{y} to have mean 0 and sd 1. If \code{NULL},
defaults to TRUE if \code{y} is numeric, FALSE otherwise.}

\item{center_y}{Logical: If TRUE, center each column of \code{y} to have mean 0. If \code{NULL}, defaults
to TRUE if \code{scale_y} is TRUE, FALSE otherwise.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{MassGLM} object.
}
\description{
Mass-univariate GLM Analysis
}
\examples{
set.seed(2022)
y <- rnormmat(500, 40, return_df = TRUE)
x <- data.frame(
  x1 = y[[3]] - y[[5]] + y[[14]] + rnorm(500),
  x2 = y[[21]] + rnorm(500)
)
massmod <- massGLM(x, y)
# Print table of coefficients, p-values, etc. for all models
summary(massmod)
}
\author{
EDG
}
