% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{matchcases}
\alias{matchcases}
\title{Match cases by covariates}
\usage{
matchcases(
  target,
  pool,
  n_matches = 1,
  target_id = NULL,
  pool_id = NULL,
  exactmatch_factors = TRUE,
  exactmatch_cols = NULL,
  distmatch_cols = NULL,
  norepeats = TRUE,
  ignore_na = FALSE,
  verbosity = 1L
)
}
\arguments{
\item{target}{data.frame you are matching against}

\item{pool}{data.frame you are looking for matches from}

\item{n_matches}{Integer: Number of matches to return}

\item{target_id}{Character: Column name in \code{target} that holds unique
cases IDs. Default = NULL, in which case integer case numbers will be used}

\item{pool_id}{Character: Same as \code{target_id} for \code{pool}}

\item{exactmatch_factors}{Logical: If TRUE, selected cases will have to
exactly match factors
available in \code{target}}

\item{exactmatch_cols}{Character: Names of columns that should be matched
exactly}

\item{distmatch_cols}{Character: Names of columns that should be
distance-matched}

\item{norepeats}{Logical: If TRUE, cases in \code{pool} can only be chosen
once.}

\item{ignore_na}{Logical: If TRUE, ignore NA values during exact matching.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
data.frame
}
\description{
Find one or more cases from a \code{pool} data.frame that match cases in a target
data.frame. Match exactly and/or by distance (sum of squared distances).
}
\examples{
\dontrun{
set.seed(2021)
cases <- data.frame(
  PID = paste0("PID", seq(4)),
  Sex = factor(c(1, 1, 0, 0)),
  Handedness = factor(c(1, 1, 0, 1)),
  Age = c(21, 27, 39, 24),
  Var = c(.7, .8, .9, .6),
  Varx = rnorm(4)
)
controls <- data.frame(
  CID = paste0("CID", seq(50)),
  Sex = factor(sample(c(0, 1), 50, TRUE)),
  Handedness = factor(sample(c(0, 1), 50, TRUE, c(.1, .9))),
  Age = sample(16:42, 50, TRUE),
  Var = rnorm(50),
  Vary = rnorm(50)
)

mc <- matchcases(cases, controls, 2, "PID", "CID")
}
}
\author{
EDG
}
