% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlegend.R
\name{mlegend}
\alias{mlegend}
\title{Add legend to \code{mplot3} plot}
\usage{
mlegend(
  lims,
  title = NULL,
  group.names,
  title.col = "black",
  col = rtPalette("rtCol1"),
  horiz.pad = 0.04,
  footer = NULL,
  font = 1,
  font.family = "Helvetica Neue",
  ...
)
}
\arguments{
\item{lims}{List with plot limits, as returned e.g. by \code{mplot3.xy}}
}
\description{
Add legend to \code{mplot3} plot
}
\author{
E.D. Gennatas
}
