% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlegend.R
\name{mlegend}
\alias{mlegend}
\title{Add legend to \code{mplot3} plot}
\usage{
mlegend(
  lims,
  title = NULL,
  group.names,
  title.col = "black",
  col = rtpalette("rtCol1"),
  horiz.pad = 0.04,
  footer = NULL,
  font = 1,
  font.family = "Helvetica Neue"
)
}
\arguments{
\item{lims}{List with plot limits in the form list(xlim = xlim, ylim = ylim)
e.g. as returned by \link{mplot3_xy}}

\item{title}{Character: Legend title}

\item{group.names}{Character: group names}

\item{title.col}{Title color}

\item{col}{Color vector}

\item{horiz.pad}{Numeric: Proportion of plot width to pad by}

\item{footer}{Character: Footer annotation}

\item{font}{1 or 2 for regular and bold}

\item{font.family}{Character: Font family to use}
}
\description{
Add legend to \code{mplot3} plot
}
\author{
E.D. Gennatas
}
