% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.bar.R
\name{mplot3.bar}
\alias{mplot3.bar}
\title{\code{mplot3}: Barplot}
\usage{
mplot3.bar(x, error = NULL, col = NULL, error.col = "white",
  error.lwd = 2, alpha = 1, border = NA, width = 1, space = NULL,
  xlim = NULL, ylim = NULL, xlab = NULL, xlab.line = 1.5,
  ylab = NULL, ylab.line = 1.5, main = NULL, main.line = 0.5,
  main.adj = 0, main.col = NULL, main.font = 2, main.family = "",
  names.arg = NULL, axisnames = FALSE, las = 1.5,
  group.names = NULL, group.names.srt = 0,
  group.names.adj = ifelse(group.names.srt == 0, 0.5, 1),
  group.names.line = 0.5, group.names.font = 1, group.names.cex = 1,
  group.names.y.pad = 0.08, group.names.at = NULL,
  color.bygroup = FALSE, legend = FALSE, legend.names = NULL,
  legend.position = "topright", legend.inset = c(0, 0),
  toplabels = NULL, mar = c(3, 2.5, 2.5, 1), pty = "m", cex = 1.2,
  cex.axis = cex, cex.names = 1, bg = NULL, plot.bg = NULL,
  barplot.axes = FALSE, yaxis = TRUE, ylim.pad = 0.04,
  y.axis.padj = 1, tck = -0.015, tick.col = NULL,
  theme = getOption("rt.theme", "light"),
  palette = getOption("rt.palette", "rtCol1"), axes.col = NULL,
  labs.col = NULL, grid = FALSE, grid.ny = NULL, grid.lty = NULL,
  grid.lwd = NULL, grid.col = NULL, grid.alpha = 1,
  par.reset = TRUE, pdf.width = 6, pdf.height = 6, filename = NULL,
  ...)
}
\arguments{
\item{x}{Vector or Matrix: If Vector, each value will be drawn as a bar.
If Matrix, each column is a vector, so multiple columns signify a different group.
e.g. Columns could be months and rows could be N days sunshine, N days rainfall, N days snow, etc.}

\item{col}{Vector of colors to use}

\item{alpha}{Float: Alpha to be applied to \code{col}}

\item{border}{Color if you wish to draw border around bars, NA for no borders (Default)}

\item{space}{Float: Space left free on either side of the bars, as a fraction of bar width. A single number or a
vector, one value per bar. If \code{x} is a matrix, space can be length 2 vector, signifying space between bars
within group and between groups. Default = c(0, 1) if x is matrix and \code{beside = TRUE}, otherwise Default = .2}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{xlab}{String: x-axis label}

\item{xlab.line}{Adjustment for the x axis label position (See code{line} in \code{?mtext})}

\item{ylab}{String: y-axis label}

\item{ylab.line}{Similar to \code{xlab.line} for the y axis}

\item{main}{String: Plot title}

\item{main.line}{Float: \code{mtext line} argument for the title. Higher numbers move title upwards.
Default = .5}

\item{main.adj}{Float: Plot title alignment (0 = left, .5 = center, 1 = right)}

\item{main.col}{Color for title. Defaults to black for theme "light" and "box", otherwise white}

\item{main.font}{Integer: 1: regular, 2: bold, 3: italic. See \code{par("font")} for more}

\item{main.family}{String: Font family to use. See \code{par("family")}}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{color.bygroup}{Logical: If TRUE, and input is a matrix, each group's bars will be given the same color,
otherwise bars across groups will be given the same sequence of colors. Default = FALSE}

\item{legend}{Logical: If TRUE, and input is matrix, draw legend for each case. Note: you may need to adjust
\code{mar} and \code{legend.inset} if you want to place the legend outside the plot
(can use e.g.\code{legend.inset = c(-.5, 0)})}

\item{mar}{Vector, length 4: Margins; see \code{mar} in \code{?par}}

\item{pty}{"s" gives a square plot, "m" gives a plot that fills graphics device size. Defaults to "s"
(See \code{?par})}

\item{cex}{Float: Character expansion factor (See \code{?par})}

\item{bg}{Background color. Defaults to white for themes "light" and "box", black otherwise.}

\item{y.axis.padj}{Similar to \code{x.axis.padj} for the y axis}

\item{tck}{Float: Tick length. Can be negative (See \code{?par("tck")})}

\item{theme}{String: "light", "dark", "lightgrid", "darkgrid", "lightbox", "darkbox"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{axes.col}{String: Color for axes values (box color set with \code{box.col})}

\item{grid}{Logical: Should grid be drawn?}

\item{grid.ny}{Integer: N of grid lines on the y axis
Use with caution, probably along with \code{yaxs = "i"}}

\item{grid.lty}{Grid line type (See \code{?par("lty")})}

\item{grid.lwd}{Grid line width}

\item{grid.col}{Grid color}

\item{grid.alpha}{Alpha for \code{grid.col}}

\item{...}{Additional arguments to \code{graphics::barplot}}
}
\description{
Draw barplots
}
\author{
Efstathios D. Gennatas
}
