% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.box.R
\name{mplot3.box}
\alias{mplot3.box}
\title{\code{mplot3}: Boxplot}
\usage{
mplot3.box(
  x,
  col = NULL,
  alpha = 0.66,
  border = NULL,
  border.alpha = 1,
  space = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  xlab.line = 1.5,
  ylab = NULL,
  ylab.line = 1.5,
  main = NULL,
  names.arg = NULL,
  axisnames = FALSE,
  group.names = NULL,
  group.names.at = NULL,
  group.names.y = NULL,
  group.names.font = 1,
  group.names.adj = 0.5,
  group.names.srt = 0,
  legend = FALSE,
  legend.names = NULL,
  legend.position = "topright",
  legend.inset = c(0, 0),
  mar = c(4, 2, 2.5, 0.5),
  pty = "m",
  cex.axis = cex,
  cex.names = cex,
  boxplot.axes = FALSE,
  yaxis = TRUE,
  ylim.pad = 0,
  y.axis.padj = 1.7,
  tck = -0.015,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  par.reset = TRUE,
  pdf.width = 6,
  pdf.height = 6,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{Matrix: Each column will be drawn as a box}

\item{col}{Vector of colors to use}

\item{alpha}{Float: Alpha to be applied to \code{col}}

\item{border}{Color for lines around boxes}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{xlab}{Character: x-axis label}

\item{xlab.line}{Float: Adjustment for the x axis label position (See code{line} in \code{?mtext})}

\item{ylab}{Character: y-axis label}

\item{ylab.line}{Float: Similar to \code{xlab.line} for the y axis}

\item{main}{Character: Plot title}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills graphics device size. Default = "m"
(See \code{par("pty")})}

\item{y.axis.padj}{Float: Similar to \code{x.axis.padj} for the y axis}

\item{tck}{Float: Tick length. Can be negative (See \code{par("tck")})}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{palette}{Vector of colors, or Character defining a builtin palette - get all options with \code{rtPalette()}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting. Default = TRUE}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename} is set). Default = 6}

\item{pdf.height}{Float: Height in inches for pdf output. Default = 6}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{...}{Additional arguments to \code{graphics::boxplot}}
}
\description{
Draw boxplots
}
\author{
Efstathios D. Gennatas
}
