% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.graph.R
\name{mplot3.graph}
\alias{mplot3.graph}
\title{Plot graph using \pkg{igraph}}
\usage{
mplot3.graph(
  x,
  node.col = "#18A3ACff",
  edge.col = "ffffff44",
  theme = getOption("rt.theme", "darkgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  node.size = 8,
  node.border.color = NA,
  node.label = NA,
  edge.curved = 0.5,
  layout = igraph::layout_with_fr,
  margin = rep(-0.3, 4),
  ...
)
}
\arguments{
\item{x}{\pkg{igraph} graph}
}
\description{
Plot graph using \pkg{igraph}
}
\author{
E.D. Gennatas
}
