% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.surv.R
\name{mplot3.surv}
\alias{mplot3.surv}
\title{\code{mplot3}: Survival Plots}
\usage{
mplot3.surv(x, lty = 1, lwd = 2, alpha = 1, col = NULL,
  normalize = TRUE, cex = 1.2, xlab = "Time", ylab = "Survival",
  main = "Kaplan-Meier estimate with 95\% CI", group.legend = NULL,
  group.title = "", group.names = NULL, group.side = 3,
  group.adj = 0.98, group.padj = 2, group.at = NA,
  par.reset = TRUE, ...)
}
\arguments{
\item{x}{Survival object / list of Survival objects created using \code{survival::Surv}}

\item{lty}{Integer: Line type. Default = 1. See \code{par("lty")}}

\item{lwd}{Float: Line width. Default = 2}

\item{alpha}{Float: Alpha for lines. Default = 1}

\item{normalize}{Logical: If TRUE: \link{drange} \code{x} to 0:100}

\item{cex}{Float: Character expansion factor (See \code{?par})}

\item{xlab}{String: x-axis label}

\item{ylab}{String: y-axis label}

\item{main}{String: Plot title}

\item{group.legend}{Logical: If TRUE, place \code{group.names} in a legend}

\item{group.title}{String: Group title, shown above group names. e.g. if group names are
c("San Francisco", "Philadelphia"), \code{group.title} can be "Place of residence"}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{...}{Additional arguments to pass to \link{mplot3.xy}}
}
\description{
Plots survival step functions using \link{mplot3.xy}
}
\author{
Efstathios D. Gennatas
}
