% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.survfit.R
\name{mplot3.survfit}
\alias{mplot3.survfit}
\title{\code{mplot3}: Survival Plots}
\usage{
mplot3.survfit(
  x,
  lty = 1,
  lwd = 1.5,
  alpha = 1,
  col = NULL,
  mark.censored = FALSE,
  draw.median = FALSE,
  median.lty = 3,
  median.col = theme$fg,
  median.alpha = 0.5,
  censor.mark = TRUE,
  censor.col = NULL,
  censor.alpha = 0.4,
  censor.pch = "I",
  censor.cex = 0.8,
  nrisk.table = FALSE,
  nrisk.pos = "below",
  nrisk.spacing = 0.9,
  table.font = 1,
  time.at = NULL,
  time.by = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "Time",
  ylab = "Survival",
  main = "Kaplan-Meier Estimate",
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  error.lty = 2,
  error.alpha = 0.5,
  autonames = TRUE,
  group.legend = NULL,
  group.names = NULL,
  group.title = NULL,
  group.line = NULL,
  group.side = NULL,
  group.adj = 0.98,
  group.at = NA,
  mar = c(2.5, 3, 2, 1),
  oma = NULL,
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{survfit object (output of \code{survival::survfit}}

\item{lty}{Integer: Line type. Default = 1. See \code{par("lty")}}

\item{lwd}{Float: Line width. Default = 2}

\item{alpha}{Float: Alpha for lines. Default = 1}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: Plot title}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'darkgrid')}

\item{palette}{Vector of colors, or Character defining a builtin palette - get options with
\code{rtPalette()}}

\item{group.legend}{Logical: If TRUE, place \code{group.names} in a legend}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{group.title}{Character: Group title, shown above group names. e.g. if group names are
c("San Francisco", "Philadelphia"), \code{group.title} can be "Place of residence"}

\item{group.side}{Integer: Side to show group legend}

\item{group.adj}{Float: \code{adj} for group legend. See \code{mtext("adj")}}

\item{group.at}{Float: location for group legend. See \code{mtext("at")}}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{oma}{Float, vector, length 4: Outer margins; see \code{par("oma")}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting. Default = TRUE}

\item{...}{Additional arguments to pass to \link{mplot3.xy}}
}
\description{
Plots survival step functions using \link{mplot3.xy}
}
\examples{
\dontrun{
# Get the lung dataset
data(cancer, package = "survival")
sf1 <- survival::survfit(survival::Surv(time, status) ~ 1, data = lung)
mplot3.survfit(sf1)
sf2 <- survival::survfit(survival::Surv(time, status) ~ sex, data = lung)
mplot3.survfit(sf2)
# with N at risk table
mplot3.survfit(sf2, nrisk.table = TRUE)
}
}
\author{
E.D. Gennatas
}
