% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.varimp.R
\name{mplot3.varimp}
\alias{mplot3.varimp}
\title{\code{mplot3}: Variable Importance}
\usage{
mplot3.varimp(
  x,
  error = NULL,
  names = NULL,
  names.pad = 0.02,
  plot.top = 1,
  labelify = TRUE,
  col = NULL,
  palette = NULL,
  alpha = 1,
  error.col = theme$fg,
  error.lwd = 2,
  beside = TRUE,
  border = NA,
  width = 1,
  space = 0.75,
  xlim = NULL,
  ylim = NULL,
  xlab = "Variable Importance",
  xlab.line = 1.3,
  ylab = NULL,
  ylab.line = 1.5,
  main = NULL,
  names.arg = NULL,
  axisnames = FALSE,
  sidelabels = NULL,
  mar = NULL,
  pty = "m",
  barplot.axes = FALSE,
  xaxis = TRUE,
  x.axis.padj = -1.2,
  tck = -0.015,
  theme = getOption("rt.theme", "lightgrid"),
  zerolines = FALSE,
  par.reset = TRUE,
  pdf.width = NULL,
  pdf.height = NULL,
  trace = 0,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{Vector, numeric: Input}

\item{error}{Vector, numeric; length = length(x): Plot error bars with given error.}

\item{names}{Vector, string; optional: Names of variables in \code{x}}

\item{plot.top}{Float or Integer: If <= 1, plot this percent highest absolute values, otherwise plot this many top values.
i.e.: \code{plot.top = .2} will print the top 20% highest values, and \code{plot.top = 20} will plot the top 20
highest values}

\item{labelify}{Logical: If TRUE convert \code{names(x)} using \link{labelify}. Default = TRUE}

\item{col}{Colors: Gradient to use for barplot fill.}

\item{alpha}{Float (0, 1): Alpha for \code{col}}

\item{error.col}{Color: For error bars}

\item{trace}{Integer: If \code{trace > 0} prints out the automatically set \code{mar} (so you can adjust if needed)
\code{names} provided}
}
\value{
Position of bar centers (invisibly)
}
\description{
Draw horizontal barplots for variable importance
}
\author{
E.D. Gennatas
}
