% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3.x.R
\name{mplot3.x}
\alias{mplot3.x}
\title{\code{mplot3}: Univariate plots: index, histogram, density, QQ-line}
\usage{
mplot3.x(
  x,
  type = c("density", "histogram", "index", "ts", "qqline"),
  group = NULL,
  data = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  index.ypad = 0.1,
  axes.swap = FALSE,
  axes.col = NULL,
  tick.col = NULL,
  cex = 1.2,
  col = NULL,
  alpha = 0.75,
  index.type = c("p", "l"),
  hist.breaks = "Sturges",
  hist.type = c("bars", "lines"),
  hist.lwd = 3,
  density.line = FALSE,
  density.shade = TRUE,
  density.legend.side = 3,
  density.legend.adj = 0.98,
  density.params = list(),
  qqline.col = NULL,
  qqline.alpha = 0.66,
  pch = 16,
  point.col = NULL,
  point.cex = 1,
  point.bg.col = NULL,
  point.alpha = 0.66,
  hline = NULL,
  vline = NULL,
  diagonal = FALSE,
  grid = FALSE,
  grid.col = NULL,
  grid.alpha = 0.5,
  grid.lty = 3,
  grid.lwd = 1,
  annotation = NULL,
  annot.col = NULL,
  group.legend = NULL,
  group.title = "",
  group.names = NULL,
  group.side = 3,
  group.adj = 0.02,
  group.at = NA,
  text.xy = NULL,
  text.x = NULL,
  text.y = NULL,
  text.xy.cex = 1,
  text.xy.col = "white",
  line.col = "#008E00",
  x.axis.padj = -1.1,
  y.axis.padj = 0.9,
  labs.col = NULL,
  lab.adj = 0.5,
  density.mean = ifelse(type == "density", TRUE, FALSE),
  hline.col = "black",
  hline.lwd = 1,
  hline.lty = 1,
  vline.col = "black",
  vline.lwd = 1,
  vline.lty = 1,
  lty = 1,
  lwd = 2,
  qqline.lwd = lwd,
  density.lwd = lwd,
  theme = getOption("rt.theme", "lightgrid"),
  palette = getOption("rt.palette", "rtCol1"),
  pty = "m",
  mar = c(2.5, 3, 2, 1),
  xaxs = "r",
  yaxs = "r",
  autolabel = letters,
  new = FALSE,
  alpha.off = FALSE,
  na.rm = TRUE,
  par.reset = TRUE,
  filename = NULL,
  pdf.width = 6,
  pdf.height = 6,
  ...
)
}
\arguments{
\item{x}{Numeric vector or list of vectors, one for each group.
If \code{data} is provided, x is name of variable in \code{data}}

\item{type}{Character: "density", "histogram",  "index", "ts", "qqline"
Case-insensitive and supports partial matching: e.g. \code{mplot3.x(x, "H")} gives histogram}

\item{group}{Vector denoting group membership. Will be converted to factor.
If \code{data} is provided, \code{group} is name of variable if \code{data}}

\item{data}{Optional data frame containing x data}

\item{xlab}{Character: x-axis label}

\item{ylab}{Character: y-axis label}

\item{main}{Character: Plot title}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{index.ypad}{Float: Expand ylim by this much for plot type "index" Default = .1
(Stops points being cut off)}

\item{axes.col}{Character: Color for axes values (box color set with \code{box.col})}

\item{index.type}{Character: "p" for points (Default), "l" for lines (timeseries)}

\item{hist.breaks}{See \code{histogram("breaks")}}

\item{hist.type}{Character: "bars" or "lines". Default = "bars"}

\item{hist.lwd}{Float: Line width for \code{type = "histogram"}; \code{hist.type = "lines"}}

\item{density.line}{Logical: If TRUE, draw line for \code{type = "density"}. Default = FALSE}

\item{density.shade}{Logical: If TRUE, draw shaded polygon for \code{type = "density"}. Default = TRUE}

\item{qqline.col}{Color for Q-Q line}

\item{qqline.alpha}{Float: Alpha for Q-Q line}

\item{point.cex}{Float: Character expansion for points. Default = .85}

\item{hline}{Vector: y-value(s) for horizontal lines. Default = NULL}

\item{vline}{Vector: x-value(s) for vertical lines. Default = NULL}

\item{diagonal}{Logical: If TRUE, draw diagonal line. Default = FALSE}

\item{annotation}{Character: Add annotation at the bottom right of the plot}

\item{group.legend}{Logical: If TRUE, include legend with group names}

\item{group.title}{Character: Title above group names}

\item{group.names}{(Optional) If multiple groups are plotted, use these names if \code{group.title = TRUE}}

\item{x.axis.padj}{Float: Adjustment for the x axis tick labels position}

\item{y.axis.padj}{Float: Similar to \code{x.axis.padj} for the y axis}

\item{labs.col}{Color for labels}

\item{lab.adj}{Adjust the axes labels. 0 = left adjust; 1 = right adjust; .5 = center (Default)}

\item{density.mean}{Logical: If TRUE, print mean of \code{x} along plot. Default = TRUE, for \code{type = "density"}}

\item{hline.col}{Color for horizontal line(s)}

\item{hline.lwd}{Float: Width for horizontal line(s)}

\item{hline.lty}{Integer: Line type for horizontal line(s)}

\item{vline.col}{Color for vertical lines}

\item{vline.lwd}{Float: Width for vertical lines}

\item{vline.lty}{Integer: Line type for vertical lines}

\item{lwd}{Integer: Line width. Used for \code{type = "ts" or "density"}}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'lightgrid')}

\item{palette}{Vector of colors, or Character defining a builtin palette - get all options with \code{rtPalette()}}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills graphics device size. Default = "m"
(See \code{par("pty")})}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4% on either end; "i": Use exact x-axis limits.
Default = "r"}

\item{yaxs}{Character: as \code{xaxs} for the y-axis. Default = "r"}

\item{new}{Logical: If TRUE, add plot to existing plot. See \code{par("new")}}

\item{na.rm}{Logical: Will be passed to all functions that support it. If set to FALSE,
input containing NA values will result in error, depending on the \code{type}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting. Default = TRUE}

\item{filename}{Path to file: If supplied, plot will be printed to file}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename} is set). Default = 6}

\item{pdf.height}{Float: Height in inches for pdf output. Default = 6}

\item{...}{Additional arguments to be passed to theme function}
}
\value{
Invisibly returns the output of \code{density}, \code{hist}, \code{qqnorm}, or NULL
}
\description{
Draw plots of 1-dimensional data: index, histogram, density, and Q-Q plots.
}
\details{
You can group data either by supplying x as a list where each element contains one vector per group, or
as a data frame where each column represents group,
or by providing a \code{group} variable, which will be converted to factor.
For bivariate plots, see \link{mplot3.xy} and \link{mplot3.xym}. For heatmaps, see \link{mplot3.heatmap}
To plot histograms of multiple groups, it's best to use \code{hist.type = "lines"}, which will use \link{mhist}
and space apart the breaks for each group
}
\seealso{
\link{mplot3}, \link{mplot3.xy}, \link{mplot3.xym}, \link{mplot3.heatmap}
}
\author{
Efstathios D. Gennatas
}
