% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_adsr.R
\name{mplot3_adsr}
\alias{mplot3_adsr}
\title{\code{mplot3}: ADSR Plot}
\usage{
mplot3_adsr(
  Attack = 300,
  Decay = 160,
  Sustain = 40,
  Release = 500,
  Value = 80,
  I = 200,
  O = 1800,
  lty = 1,
  lwd = 4,
  main = "ADSR Envelope",
  main.line = 1.6,
  main.col = "white",
  Attack.col = "#44A6AC",
  Decay.col = "#F4A362",
  Sustain.col = "#3574A7",
  Release.col = "#C23A70",
  draw.poly = FALSE,
  poly.alpha = 0.15,
  draw.verticals = TRUE,
  v.lty = 1,
  v.lwd = 0.8,
  arrow.code = 2,
  arrow.length = 0.09,
  grid = FALSE,
  grid.lty = 1,
  grid.lwd = 0.4,
  grid.col = NULL,
  zerolines.col = "gray50",
  theme = "darkgray",
  labs.col = "gray70",
  tick.col = "gray70",
  on.col = "gray70",
  off.col = "gray70",
  pty = "m",
  mar = c(3, 3, 3.2, 0.5),
  xaxs = "i",
  yaxs = "i",
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{Attack}{Numeric: Attack time (in milliseconds)}

\item{Decay}{Numeric: Decay time (in milliseconds)}

\item{Sustain}{Numeric: Sustain Level (percent)}

\item{Release}{Numeric: Release time (in milliseconds)}

\item{Value}{Numeric: Value (percent)}

\item{I}{Numeric: Note on time (in milliseconds)}

\item{O}{Numeric: Note off time (in milliseconds)}

\item{lty}{Integer: Line type}

\item{lwd}{Numeric: Line width}

\item{main}{Character: Main title}

\item{main.line}{Numeric: Main title line height}

\item{main.col}{Main title color}

\item{Attack.col}{Attack color}

\item{Decay.col}{Decay color}

\item{Sustain.col}{Sustain color}

\item{Release.col}{Release color}

\item{draw.poly}{Logical: If TRUE, draw polygons for each segment}

\item{poly.alpha}{Numeric: Polygon alpha}

\item{draw.verticals}{Logical: If TRUE, draw vertical lines}

\item{v.lty}{Integer: Vertical line type}

\item{v.lwd}{Numeric: Vertical line width}

\item{arrow.code}{Integer: Arrow code}

\item{arrow.length}{Numeric: Arrow length}

\item{grid}{Logical: If TRUE, draw grid}

\item{grid.lty}{Integer: Grid line type}

\item{grid.lwd}{Numeric: Grid line width}

\item{grid.col}{Grid line color}

\item{zerolines.col}{Color for zero lines}

\item{theme}{Character: "light" or "dark" (Default)}

\item{labs.col}{Color for axis labels}

\item{tick.col}{Color for axis ticks}

\item{on.col}{Color for "on" line}

\item{off.col}{Color for "off" line}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills
graphics device size. Default = "m" (See \code{par("pty")})}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4\% on either end;
"i": Use exact x-axis limits.}

\item{yaxs}{Character: as \code{xaxs} for the y-axis.}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{...}{Additional arguments to pass to \link{mplot3_xy}}
}
\description{
Plot Attack Decay Sustain Release Envelope Generator using \link{mplot3_xy}
}
\details{
Learn more:
(https://en.wikipedia.org/wiki/Synthesizer#Attack_Decay_Sustain_Release_.28ADSR.29_envelope "ADSR Wikipedia")
}
\examples{
\dontrun{
mplot3_adsr()
}
}
\author{
E.D. Gennatas
}
