% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_pr.R
\name{mplot3_pr}
\alias{mplot3_pr}
\title{\code{mplot3} Precision Recall curves}
\usage{
mplot3_pr(
  prob,
  labels,
  f1 = FALSE,
  main = "",
  col = NULL,
  cex = 1.2,
  lwd = 2.5,
  diagonal = FALSE,
  hline.lty = 1,
  hline.lwd = 1,
  hline.col = "red",
  diagonal.lwd = 2.5,
  diagonal.lty = 3,
  group.legend = FALSE,
  annotation = TRUE,
  annotation.side = 3,
  annotation.col = col,
  annot.line = NULL,
  annot.adj = 1,
  annot.font = 1,
  mar = c(2.5, 3, 2.5, 1),
  theme = rtTheme,
  palette = rtPalette,
  par.reset = TRUE,
  verbose = TRUE,
  filename = NULL,
  pdf.width = 5,
  pdf.height = 5
)
}
\arguments{
\item{prob}{Vector, Float [0, 1]: Predicted probabilities (i.e. c(.1, .8, .2, .9))}

\item{labels}{Vector, Integer {0, 1}: True labels (i.e. c(0, 1, 0, 1))}

\item{f1}{Logical: If TRUE, annotate the point of maximal F1 score.}

\item{main}{Character: Plot title.}

\item{col}{Color, vector: Colors to use for ROC curve(s)}

\item{cex}{Float: Character expansion factor.}

\item{lwd}{Float: Line width.}

\item{diagonal}{Logical: If TRUE, draw diagonal.}

\item{hline.lty}{Integer: Line type for horizontal line(s)}

\item{hline.lwd}{Float: Width for horizontal line(s)}

\item{hline.col}{Color for horizontal line(s)}

\item{diagonal.lwd}{Float: Line width for diagonal.}

\item{diagonal.lty}{Integer: Line type for diagonal.}

\item{group.legend}{Logical}

\item{annotation}{Character: Add annotation at the bottom right of the plot}

\item{annotation.side}{Integer: Side of plot to place annotation.}

\item{annotation.col}{Color: Color of annotation.}

\item{annot.line}{Numeric: Line number for annotation.}

\item{annot.adj}{Numeric: Adjustment for annotation.}

\item{annot.font}{Integer: Font for annotation.}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{palette}{Vector of colors, or Character defining a builtin palette -
get options with \code{rtpalette()}}

\item{par.reset}{Logical: If TRUE, reset \code{par} setting before exiting.}

\item{verbose}{Logical: If TRUE, print messages to console.}

\item{filename}{Path to file: If supplied, plot will be printed to file}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename}
is set).}

\item{pdf.height}{Float: Height in inches for pdf output.}
}
\value{
List with Precision, Recall, and Threshold values, invisibly
}
\description{
Plot Precision Recall curve for a binary classifier
}
\author{
E.D. Gennatas
}
