% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot3_xym.R
\name{mplot3_xym}
\alias{mplot3_xym}
\title{Scatter plot with marginal density and/or histogram}
\usage{
mplot3_xym(
  x,
  y,
  margin = c("histogram", "density", "both"),
  fit = "gam",
  se.fit = TRUE,
  xlim = NULL,
  ylim = NULL,
  col = "#18A3AC",
  density.alpha = 0.66,
  hist.breaks = 30,
  hist.alpha = 0.66,
  hist.space = 0.05,
  hist.lwd = 3,
  lwd = 4,
  main = NULL,
  main.adj = 0,
  axes.density = FALSE,
  pty = "m",
  mar = c(3, 3, 0, 0),
  margin.mar = 0.2,
  xaxs = "r",
  yaxs = "r",
  theme = rtTheme,
  par.reset = TRUE,
  widths = NULL,
  heights = NULL,
  filename = NULL,
  pdf.width = 7,
  pdf.height = 7,
  ...
)
}
\arguments{
\item{x}{Numeric vector: x-axis data}

\item{y}{Numeric vector: y-axis data}

\item{margin}{Character: "density", "histogram", or "both". Type of marginal plots to draw.}

\item{fit}{Character: Algorithm to use to draw \code{y ~ x}.}

\item{se.fit}{Logical: If TRUE: plot +/- 2 * Standard Error of fit}

\item{xlim}{Float vector, length 2: x-axis limits}

\item{ylim}{Float vector, length 2: y-axis limits}

\item{col}{Color for marginal plots}

\item{density.alpha}{Numeric: Alpha for density plots}

\item{hist.breaks}{Integer: Number of histogram breaks}

\item{hist.alpha}{Numeric: Alpha for barplots}

\item{hist.space}{Numeric: Space between bars in barplots}

\item{hist.lwd}{Numeric: Line width for barplots}

\item{lwd}{Numeric: Line width}

\item{main}{Character: Main title}

\item{main.adj}{Numeric: Main title adjustment}

\item{axes.density}{Logical: If TRUE, plot margin plot axes for density (debugging only)}

\item{pty}{Character: "s" gives a square plot; "m" gives a plot that fills
graphics device size. Default = "m" (See \code{par("pty")})}

\item{mar}{Float, vector, length 4: Margins; see \code{par("mar")}}

\item{margin.mar}{Numeric: Margin for marginal plots}

\item{xaxs}{Character: "r": Extend plot x-axis limits by 4\% on either end;
"i": Use exact x-axis limits.}

\item{yaxs}{Character: as \code{xaxs} for the y-axis.}

\item{theme}{Character: Run \code{themes()} for available themes}

\item{par.reset}{Logical: Resest \code{par} to original settings}

\item{filename}{Character: Path to file to save plot. Default = NULL}

\item{pdf.width}{Float: Width in inches for pdf output (if \code{filename}
is set).}

\item{pdf.height}{Float: Height in inches for pdf output.}

\item{...}{Additional arguments to passed to \link{mplot3_xy}}
}
\description{
Draw a scatter plot with fit line and marginal density and/or histogram
}
\details{
To make wide plot, change \code{widths}: e.g. widths = c(7, 1)
}
\examples{
\dontrun{
x <- rnorm(500)
y <- x^3 + 12 + rnorm(500)
mplot3_xym(x, y)
}
}
\author{
E.D. Gennatas
}
