% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot_hsv.R
\name{mplot_hsv}
\alias{mplot_hsv}
\title{Plot HSV color range}
\usage{
mplot_hsv(
  h.steps = seq(0, 1, 0.025),
  s.steps = seq(0, 1, 0.05),
  v = 1,
  alpha = 1,
  pch = 16,
  bg = "black",
  axes = TRUE,
  pty = "s",
  cex = 2,
  mar = c(3, 3, 2, 0.5),
  lab.col = NULL,
  type = c("radial", "square"),
  line.col = "gray50",
  show.grid = TRUE,
  show.radial.grid = FALSE,
  show.grid.labels = 1,
  cex.axis = 1,
  cex.lab = 1,
  par.reset = TRUE
)
}
\arguments{
\item{h.steps}{Float, vector: Hue values to plot.
Default = \code{seq(0, 1, .0125)}}

\item{s.steps}{Float, vector: Saturation values to plot. Default = same as
\code{h.steps}}

\item{v}{Float: Value.}

\item{alpha}{Float: Alpha.}

\item{pch}{Integer: pch plot parameter. Default = 15 (square)}

\item{bg}{Color: Background color. Default = "black"}

\item{axes}{Logical: for \code{type = "square"}: If TRUE, draw axes.}

\item{pty}{Character: for \code{type = "square"}: "s", "r", par's pty
argument. Default = "s" (square plot)}

\item{cex}{Float: \code{par/plot}'s cex argument.}

\item{mar}{Float, vector: for \code{type = "square"}: \code{par}'s mar
argument.}

\item{lab.col}{Color: Color for axes and labels. Defaults to inverse of
\code{bg}, i.e. white if bg is black}

\item{type}{Character: "square" for square plot, "radial" for radial plot.}

\item{show.grid}{Logical: if TRUE, show grid then type is "radial"}

\item{par.reset}{Logical: If TRUE, reset \code{par} before exit}
}
\description{
Plot HSV color range
}
\examples{
\dontrun{
mplot_hsv()
}
}
\author{
E.D. Gennatas
}
