% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigplot.R
\name{multigplot}
\alias{multigplot}
\title{Multipanel \strong{ggplot2} plots}
\usage{
multigplot(plots = NULL, nrows = NULL, byrow = TRUE)
}
\arguments{
\item{plots}{List of ggplot2 plots}

\item{nrows}{Integer: number of rows for panel arrangement. Defaults to number of rows required
to plot 2 plots per row}

\item{byrow}{Logical: If TRUE, draw plots in order provided by row, otherwise by column.
Default = TRUE}
}
\description{
Plot a panel of \strong{gplot2} plots
}
\author{
E.D. Gennatas
}
