% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ops.R
\name{nunique_perfeat}
\alias{nunique_perfeat}
\title{Number of unique values per feature}
\usage{
nunique_perfeat(x, excludeNA = FALSE)
}
\arguments{
\item{x}{data.table}

\item{excludeNA}{Logical: If TRUE, exclude NA values}
}
\description{
Number of unique values per feature
}
\author{
E.D. Gennatas
}
