% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneHot.R
\name{oneHot}
\alias{oneHot}
\title{One hot encoding}
\usage{
oneHot(x, verbose = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\value{
data.frame where all factors are one hot encoded
}
\description{
One hot encode all factors in a data.frame
}
\author{
Efstathios D Gennatas
}
