% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneHot.R, R/oneHot.data.frame.R,
%   R/oneHot.default.R
\name{oneHot}
\alias{oneHot}
\alias{oneHot.data.frame}
\alias{oneHot.default}
\title{One hot encoding}
\usage{
oneHot(x, verbose = FALSE)

\method{oneHot}{data.frame}(x, verbose = TRUE)

\method{oneHot}{default}(x, verbose = TRUE)
}
\arguments{
\item{x}{Vector or data.frame}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\value{
For vector input, a one-hot-encoded matrix, for data.frame frame input, an expanded data.frame where all
factors are one-hot encoded
}
\description{
One hot encode a vector or factors in a data.frame
}
\details{
A vector input will be one-hot encoded regardless of type by looking at all unique values. With data.frame input,
only column of type factor will be one-hot encoded. This function is used by \link{preprocess}
}
\author{
E.D. Gennatas
}
