% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{outcome}
\alias{outcome}
\title{Get last column as a vector}
\usage{
outcome(x, ...)
}
\arguments{
\item{x}{data.frame or similar.}

\item{...}{Not used.}
}
\value{
Vector containing the last column of \code{x}.
}
\description{
Get last column as a vector
}
\author{
EDG
}
