% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{outcome}
\alias{outcome}
\title{Get the outcome as a vector}
\usage{
outcome(x)
}
\arguments{
\item{x}{data.frame or similar.}
}
\value{
Vector containing the last column of \code{x}.
}
\description{
Returns the last column of \code{x}, which is by convention the outcome variable.
}
\details{
This applied to tabular datasets used for supervised learning in rtemis,
where, by convention, the last column is the outcome variable and all other columns
are features.
}
\examples{
outcome(iris)
}
\author{
EDG
}
