% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_init.R
\name{outro}
\alias{outro}
\title{\code{rtemis-internals}: \code{outro}}
\usage{
outro(
  start.time,
  message = NULL,
  verbose = TRUE,
  sinkOff = FALSE,
  newline.pre = FALSE
)
}
\description{
Outro
}
\details{
Stops function execution timer and closes log file.

Second part to \code{intro}
}
\keyword{internal}
