% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.MXINCEPTION.R
\name{p.MXINCEPTION}
\alias{p.MXINCEPTION}
\title{Classify Images using pre-trained Inception network with \code{mxnet} [C]}
\usage{
p.MXINCEPTION(img, img.path = NULL, iteration = 39,
  plot.input = FALSE, plot.normed = FALSE, plot.graphviz = FALSE,
  verbose = TRUE)
}
\description{
Predict using Inception
}
\seealso{
Other Deep Learning: \code{\link{d.H2OAE}},
  \code{\link{s.H2ODL}}, \code{\link{s.MXN}},
  \code{\link{s.TFN}}
}
\author{
Efstathios D. Gennatas
}
\concept{Deep Learning}
\concept{Pretrained Models}
