% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytreew.R
\name{partLmw}
\alias{partLmw}
\title{\pkg{rtemis} internal: Ridge and Stump}
\usage{
partLmw(
  x1,
  y1,
  weights,
  .env,
  minobsinnode.lin,
  lin.type,
  alpha,
  lambda,
  lambda.seq,
  cv.glmnet.nfolds,
  cv.glmnet.lambda,
  part.minsplit,
  part.xval,
  part.max.depth,
  part.cp,
  part.minbucket,
  verbose,
  trace
)
}
\description{
Fit a linear model on (x, y) and a tree on the residual yhat - y
}
