% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfread.R
\name{pfread}
\alias{pfread}
\title{fread delimited file in parts}
\usage{
pfread(
  x,
  part_nrows,
  nrows = NULL,
  header = TRUE,
  sep = "auto",
  verbose = TRUE,
  stringsAsFactors = TRUE,
  ...
)
}
\arguments{
\item{x}{Character: Path to delimited file}

\item{part_nrows}{Integer: Number of rows to read in each part}

\item{nrows}{Integer: Number of rows in the file}

\item{header}{Logical: If TRUE, the file is assumed to include a header row}

\item{sep}{Character: Delimiter}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{stringsAsFactors}{Logical: If TRUE, characters will be converted to
factors}

\item{...}{Additional arguments to pass to \code{data.table::fread()}}
}
\description{
fread delimited file in parts
}
\author{
E.D. Gennatas
}
