% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_S7_Supervised.R
\name{plot.Classification}
\alias{plot.Classification}
\title{Plot Classification}
\usage{
\method{plot}{Classification}(x, what = NULL, theme = choose_theme(), ...)
}
\arguments{
\item{x}{Classification object.}

\item{what}{Character vector: What to plot. Can include "training", "validation", "test", or
"all", which will plot all available.}

\item{theme}{Theme object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot Classification
}
\author{
EDG
}
