% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_MassUni.R
\name{plot.MassGLM}
\alias{plot.MassGLM}
\title{Plot MassGLM using volcano plot}
\usage{
\method{plot}{MassGLM}(x, xname = NULL, p_adjust = "holm", theme = choose_theme(), ...)
}
\arguments{
\item{x}{MassGLM object}

\item{xname}{Character: Name of covariate to get data for. If \code{NULL}, the first covariate is used.}

\item{p_adjust}{Character: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none" -
p-value adjustment method.}

\item{theme}{Theme object}

\item{...}{Additional arguments passed to \link{draw_volcano} or \link{draw_bar}}
}
\description{
Plot MassGLM using volcano plot
}
\author{
EDG
}
