% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R
\name{plot.massGLM}
\alias{plot.massGLM}
\title{Plot \code{massGLM} object}
\usage{
\method{plot}{massGLM}(
  x,
  predictor = NULL,
  main = NULL,
  what = c("coefs", "pvals", "volcano"),
  p.adjust.method = "none",
  p.transform = function(x) -log10(x),
  pval.hline = c(0.05, 0.001),
  hline.col = "#ffffff",
  hline.dash = "dash",
  hline.annotate = as.character(pval.hline),
  ylim = NULL,
  ylab = NULL,
  col.neg = "#43A4AC",
  col.pos = "#FA9860",
  col.ns = "#7f7f7f",
  theme = getOption("rt.theme"),
  volcano.annotate = TRUE,
  volcano.annotate.n = 7,
  volcano.p.transform = "-log10",
  margin = NULL,
  displayModeBar = FALSE,
  trace = 0,
  ...
)
}
\arguments{
\item{x}{\code{massGLM} object}

\item{what}{Character: "adjusted" or "raw" p-values to plot}
}
\description{
Plots a \code{massGLM} object using \link{dplot3_bar}
}
\author{
E.D. Gennatas
}
