% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_MassUni.R
\name{plot.MassGLM}
\alias{plot.MassGLM}
\title{Plot MassGLM using volcano plot}
\usage{
\method{plot}{MassGLM}(
  x,
  coefname = NULL,
  p_adjust_method = "holm",
  p_transform = function(x) -log10(x),
  xlab = "Coefficient",
  ylab = NULL,
  theme = choose_theme(),
  verbosity = 1L,
  ...
)
}
\arguments{
\item{x}{MassGLM object}

\item{coefname}{Character: Name of coefficient to plot. If \code{NULL}, the first coefficient is used.}

\item{p_adjust_method}{Character: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none" -
p-value adjustment method.}

\item{p_transform}{Function to transform p-values for plotting. Default is \code{function(x) -log10(x)}.}

\item{xlab}{Character: x-axis label.}

\item{ylab}{Character: y-axis label.}

\item{theme}{Theme object}

\item{verbosity}{Integer: Verbosity level.}

\item{...}{Additional arguments passed to \link{draw_volcano} or \link{draw_bar}}
}
\description{
Plot MassGLM using volcano plot
}
\author{
EDG
}
