% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R
\name{plot.massGLM}
\alias{plot.massGLM}
\title{Plot \code{massGLM} object}
\usage{
\method{plot}{resample}(
  x,
  what = c("Adjusted_pvalue", "pvalue"),
  pval.hline = 0.05,
  hline.col = "#FE4AA3",
  hline.dash = "dash",
  ...
)
}
\arguments{
\item{x}{\code{massGLM} object}

\item{what}{Character:}
}
\description{
Plots a \code{massGLM} object using \link{dplot3.bar}
}
\author{
E.D. Gennatas
}
