% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLM.R
\name{plot.massGLM}
\alias{plot.massGLM}
\title{Plot \code{massGLM} object}
\usage{
\method{plot}{massGLM}(
  x,
  predictor = NULL,
  main = NULL,
  what = c("volcano", "coefs", "pvals"),
  p.adjust.method = "holm",
  p.transform = function(x) -log10(x),
  show = c("all", "signif"),
  xnames = NULL,
  pval.hline = c(0.05, 0.001),
  hline.col = "#ffffff",
  hline.dash = "dash",
  hline.annotate = as.character(pval.hline),
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  group = NULL,
  col.neg = "#43A4AC",
  col.pos = "#FA9860",
  col.ns = "#7f7f7f",
  theme = rtTheme,
  alpha = NULL,
  volcano.annotate = TRUE,
  volcano.annotate.n = 7,
  volcano.hline = NULL,
  volcano.hline.dash = "dot",
  volcano.hline.annotate = NULL,
  volcano.p.transform = function(x) -log10(x),
  margin = NULL,
  displayModeBar = TRUE,
  trace = 0,
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{\code{massGLM} object}

\item{what}{Character: "adjusted" or "raw" p-values to plot}

\item{xlab}{Character: x-axis label for volcano plot}
}
\description{
Plots a \code{massGLM} object using \link{dplot3_bar}
}
\author{
E.D. Gennatas
}
