% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_Supervised.R
\name{plot.Regression}
\alias{plot.Regression}
\title{Plot Regression}
\usage{
\method{plot}{Regression}(x, what = "all", fit = "glm", theme = rtemis_theme, labelify = TRUE, ...)
}
\arguments{
\item{x}{Regression object.}

\item{what}{Character vector: What to plot. Can include "training", "validation", "test", or
"all", which will plot all available.}

\item{fit}{Character: Algorithm to use to draw fit line.}

\item{theme}{Character or Theme: Theme to use for the plot.}

\item{labelify}{Logical: If TRUE, labelify the axis labels.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot Regression
}
\author{
EDG
}
