% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{plot.resample}
\alias{plot.resample}
\title{\code{plot} method for \code{resample} object}
\usage{
\method{plot}{resample}(x, col = NULL, ...)
}
\arguments{
\item{x}{Vector; numeric or factor: Outcome used for resampling}

\item{col}{Vector, color}

\item{res}{\link{resample} object}
}
\description{
\code{plot} method for \code{resample} object
}
\author{
Efstathios D. Gennatas
}
