% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{plot.resample}
\alias{plot.resample}
\title{\code{plot} method for \code{resample} object}
\usage{
\method{plot}{resample}(x, col = NULL, ...)
}
\arguments{
\item{x}{Vector; numeric or factor: Outcome used for resampling}

\item{col}{Vector, color}

\item{...}{Additional arguments passed to \link{mplot3_res}}
}
\description{
Run \link{mplot3_res} on a \link{resample} object
}
\author{
E.D. Gennatas
}
