% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_cv.R
\name{plot.rtModCVCalibration}
\alias{plot.rtModCVCalibration}
\title{Plot \code{rtModCVCalibration} object}
\usage{
\method{plot}{rtModCVCalibration}(
  x,
  what = c("calibration", "brier"),
  type = c("aggregate.all", "aggregate.by.resample"),
  bin.method = c("quantile", "equidistant"),
  filename = NULL,
  ...
)
}
\arguments{
\item{x}{\code{rtModCVCalibration} object}

\item{what}{Character: "calibration" or "brier"}

\item{type}{Character: "aggregate.all" or "aggregate.by.resample"}

\item{bin.method}{Character: "quantile" or "equidistant"}

\item{filename}{Character: Path to save plot as pdf}

\item{...}{Additional arguments}
}
\value{
plotly object
}
\description{
Plot Calibration plots and Brier score boxplots for \code{rtModCVCalibration} object
}
\details{
For calibration plots, \code{type = "aggregate.all"} is likely the more informative one.
It shows calibrations curves before and after calibration by aggregating across all
outer test sets. The \code{type = "aggregate.by.resample"} option shows the calibration
curves after calibration for each outer resample.

For Brier boxplots, \code{type = "aggregate.all"} shows 1 score per outer resample prior
to calibration and multiple (equal n.resamples used in \link{calibrate_cv}) brier scores
per outer resample after calibration. This is for diagnostic purposes mainly.
For presentation, the \code{type = "aggregate.by.resample"} option shows the mean Brier
score per outer resample prior to calibration and after calibration and makes more
sense. The uncalibrated estimates could be resampled using the calibration model
resamples to produce a more comparable boxplot of Brier scores when using the
\code{type = "aggregate.all"} option, but that seems artifactual.

More options are certainly possibly, e.g. an "aggregate.none" that would show all
calibration resamples for all outer ersamples, and can be added in the future if
needed.
}
\author{
E.D. Gennatas
}
