% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{plot.rtTest}
\alias{plot.rtTest}
\title{Plot \code{rtTest} object}
\usage{
\method{plot}{rtTest}(
  x,
  main = NULL,
  mar = NULL,
  uni.type = c("density", "histogram", "hd"),
  boxplot.xlab = FALSE,
  theme = getOption("rt.theme", "lightgrid"),
  par.reset = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{rtTest} object}

\item{main}{Character: Main title}

\item{theme}{Character: "black", "blackgrid", "darkgrid", "white", "whitegrid", "lightgrid"
Default = "lightgrid" if no default \code{"rt.fit"} is set using \code{options}.
You can set a system-wide default in your \code{.Rprofile} by including a line like
options(rt.theme = 'darkgrid')}
}
\description{
Plot \code{rtTest} object
}
\author{
E.D. Gennatas
}
