% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_MassUni.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot for MassGLM}
\arguments{
\item{x}{MassGLM object.}

\item{xname}{Character: Name of covariate to get data for. If \code{NULL}, the first covariate is used.}

\item{p_adjust_method}{Character: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none" -
p-value adjustment method.}

\item{p_transform}{Function to transform p-values for plotting. Default is \code{function(x) -log10(x)}.}

\item{ylab}{Character: y-axis label.}

\item{theme}{Theme object.}

\item{col_pos}{Character: Color for positive significant coefficients.}

\item{col_neg}{Character: Color for negative significant coefficients.}

\item{alpha}{Numeric: Transparency level for the bars.}

\item{...}{Additional arguments passed to \link{draw_bar}.}
}
\description{
Create a Manhattan plot for MassGLM objects created with \link{massGLM}.
}
\author{
EDG
}
