% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R, R/07_S7_MassUni.R
\name{plot_manhattan}
\alias{plot_manhattan}
\alias{plot_manhattan.MassGLM}
\title{Manhattan plot}
\usage{
plot_manhattan(x, ...)

plot_manhattan.MassGLM(
  x,
  coefname = NULL,
  p_adjust_method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  p_transform = function(x) -log10(x),
  ylab = NULL,
  theme = choose_theme(),
  col_pos = "#43A4AC",
  col_neg = "#FA9860",
  alpha = 0.8,
  ...
)
}
\arguments{
\item{x}{MassGLM object.}

\item{...}{Additional arguments passed to \link{draw_bar}.}

\item{coefname}{Character: Name of coefficient to plot. If \code{NULL}, the first coefficient is used.}

\item{p_adjust_method}{Character: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none" -
p-value adjustment method.}

\item{p_transform}{Function to transform p-values for plotting. Default is \code{function(x) -log10(x)}.}

\item{ylab}{Character: y-axis label.}

\item{theme}{Theme object.}

\item{col_pos}{Character: Color for positive significant coefficients.}

\item{col_neg}{Character: Color for negative significant coefficients.}

\item{alpha}{Numeric: Transparency level for the bars.}
}
\description{
Draw a Manhattan plot for MassGLM objects created with \link{massGLM}.

Create a Manhattan plot for MassGLM objects created with \link{massGLM}.
}
\author{
EDG
}
