% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot Metric}
\usage{
plot_metric(x, ...)
}
\arguments{
\item{x}{Supervised or SupervisedRes object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
plotly object
}
\description{
Plot metric for Supervised or SupervisedRes objects.
}
\author{
EDG
}
