% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curve}
\usage{
plot_roc(x, ...)
}
\arguments{
\item{x}{\code{Classification} or \code{ClassificationRes} object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A plotly object containing the ROC curve.
}
\description{
This generic is used to plot the ROC curve for a model.
}
\examples{
\dontrun{
x <- as.data.table(iris[51:150, ])
x[, Species := factor(Species)]
species_glm <- train(x, algorithm = "GLM")
plot_roc(species_glm)
}
}
\author{
EDG
}
