% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curve}
\usage{
plot_roc(x, ...)
}
\arguments{
\item{x}{\code{Classification} or \code{ClassificationRes} object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A plotly object containing the ROC curve.
}
\description{
This generic is used to plot the ROC curve for a model.
}
\examples{
ir <- iris[51:150, ]
ir[["Species"]] <- factor(ir[["Species"]])
species_glm <- train(ir, algorithm = "GLM")
plot_roc(species_glm)
}
\author{
EDG
}
