% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{plot_true_pred}
\alias{plot_true_pred}
\title{Plot True vs. Predicted Values}
\usage{
plot_true_pred(x, ...)
}
\arguments{
\item{x}{Supervised or SupervisedRes object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
plotly object.
}
\description{
Plot True vs. Predicted Values for Supervised objects.
For classification, it plots a confusion matrix.
For regression, it plots a scatter plot of true vs. predicted values.
}
\author{
EDG
}
