% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_init.R
\name{plot_true_pred}
\alias{plot_true_pred}
\title{Plot True vs. Predicted Values}
\usage{
plot_true_pred(x, ...)
}
\arguments{
\item{x}{\code{Supervised} or \code{SupervisedRes} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
plotly object.
}
\description{
Plot True vs. Predicted Values for Supervised objects.
For classification, it plots a confusion matrix.
For regression, it plots a scatter plot of true vs. predicted values.
}
\examples{
x <- set_outcome(iris, "Sepal.Length")
sepallength_glm <- train(x, hyperparameters = setup_GLM())
plot_true_pred(sepallength_glm)
}
\author{
EDG
}
