% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_Supervised.R
\name{plot_varimp}
\alias{plot_varimp}
\title{Plot Variable Importance}
\usage{
plot_varimp(x, ...)
}
\arguments{
\item{x}{Supervised object.}

\item{...}{Additional arguments passed to the plotting function.}
}
\description{
Plot Variable Importance for Supervised objects.
}
\author{
EDG
}
