% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_Supervised.R
\name{plot_varimp}
\alias{plot_varimp}
\title{Plot Variable Importance}
\usage{
plot_varimp(x, ...)
}
\arguments{
\item{x}{Supervised object.}

\item{...}{Additional arguments passed to the plotting function.}

\item{theme}{Theme to use for the plot.}

\item{filename}{Character: Filename to save the plot to. If NULL, the plot is not saved.}
}
\description{
Plot Variable Importance for Supervised objects.
}
\author{
EDG
}
