% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modError.R
\name{precision}
\alias{precision}
\title{Precision (aka PPV)}
\usage{
precision(true, estimated, harmonize = FALSE, verbose = TRUE)
}
\arguments{
\item{true}{Factor: True labels}

\item{estimated}{Factor: Estimated labels}

\item{harmonize}{Logical: If TRUE, run \link{factorHarmonize} first}

\item{verbose}{Logical: If TRUE, print messages to output. Default = TRUE}
}
\description{
The first factor level is considered the positive case.
}
