% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error.R
\name{precision}
\alias{precision}
\title{Precision (aka PPV)}
\usage{
precision(true, estimated, harmonize = FALSE, verbosity = 1)
}
\arguments{
\item{true}{Factor: True labels}

\item{estimated}{Factor: Estimated labels}

\item{harmonize}{Logical: If TRUE, run \link{factor_harmonize} first}

\item{verbosity}{Integer: If > 0, print messages to console.}
}
\description{
The first factor level is considered the positive case.
}
