% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtree.R
\name{predict.addtree}
\alias{predict.addtree}
\title{Predict Method for MediBoost Model}
\usage{
\method{predict}{addtree}(object, newdata, verbose = FALSE, ...)
}
\arguments{
\item{object}{A trained model of class \code{"addtree"}}

\item{newdata}{Optional: a matrix / data.frame of features with which to predict}

\item{verbose}{Logical: If TRUE, print messages to output}

\item{...}{Not used}
}
\description{
Obtains predictions from a trained MediBoost model
}
\author{
E.D. Gennatas
}
