% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\name{predict.boost}
\alias{predict.boost}
\title{Predict method for \code{boost} object}
\usage{
\method{predict}{boost}(
  object,
  newdata = NULL,
  n.feat = NCOL(newdata),
  n.iter = NULL,
  as.matrix = FALSE,
  verbose = FALSE,
  n.cores = rtCores,
  ...
)
}
\arguments{
\item{object}{\link{boost} object}

\item{newdata}{data.frame: New data to predict on}

\item{n.feat}{Integer: Number of features to use from \code{newdata}}

\item{n.iter}{Integer: Number of iterations to use}

\item{as.matrix}{Logical: If TRUE, return matrix of predictions for each iteration,
otherwise return vector}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{n.cores}{Integer: Number of cores to use}

\item{...}{Not used}
}
\description{
Predict method for \code{boost} object
}
\author{
E.D. Gennatas
}
