% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmLiteBoostTV.R
\name{predict.glmLiteBoostTV}
\alias{predict.glmLiteBoostTV}
\title{Predict method for \code{glmLiteBoostTV} object}
\usage{
\method{predict}{glmLiteBoostTV}(
  object,
  newdata = NULL,
  n.feat = NCOL(newdata),
  n.iter = NULL,
  as.matrix = FALSE,
  verbose = FALSE,
  n.cores = rtCores,
  ...
)
}
\arguments{
\item{object}{\code{glmLiteBoostTV} object}

\item{newdata}{Set of predictors}

\item{n.feat}{Integer: N of features to use.}

\item{n.iter}{Integer: N of iterations to predict from.}

\item{as.matrix}{Logical: If TRUE, return predictions from each iteration.}

\item{verbose}{Logical: If TRUE, print messages to console.}

\item{n.cores}{Integer: Number of cores to use.}

\item{...}{Unused}
}
\description{
Predict method for \code{glmLiteBoostTV} object
}
\author{
E.D. Gennatas
}
