% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{predict.hytboost}
\alias{predict.hytboost}
\title{Predict method for \code{hytboost} object}
\usage{
\method{predict}{hytboost}(
  object,
  newdata = NULL,
  n.iter = NULL,
  fixed.cxr = NULL,
  as.matrix = FALSE,
  n.cores = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{\code{hytboost} object}

\item{newdata}{data.frame of predictors}

\item{n.iter}{Integer: Use the first so many trees for prediction}

\item{fixed.cxr}{(internal use) Matrix: Cases by rules to use instead of matching
cases to rules using \code{newdata}}

\item{as.matrix}{Logical: If TRUE, output}

\item{n.cores}{Integer: Number of cores to use}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{...}{Not used}
}
\description{
Predict method for \code{hytboost} object
}
\author{
E.D. Gennatas
}
