% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LightRuleFit.R
\name{predict.LightRuleFit}
\alias{predict.LightRuleFit}
\title{\code{predict} method for \code{LightRuleFit} object}
\usage{
\method{predict}{LightRuleFit}(
  object,
  newdata = NULL,
  return.cases.by.rules = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{\code{LightRuleFit} object}

\item{newdata}{Feature matrix / data.frame: will be converted to \code{data.table}}

\item{return.cases.by.rules}{Logical: If TRUE, return cases by rules matrix}

\item{verbose}{Logical: If TRUE, print messages during execution. Default = TRUE}

\item{...}{Ignored}
}
\value{
Vector of estimated values
}
\description{
\code{predict} method for \code{LightRuleFit} object
}
