% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2bsplinemat.R
\name{predict.rtBSplines}
\alias{predict.rtBSplines}
\title{Predict S3 method for \code{rtBSplines}}
\usage{
\method{predict}{rtBSplines}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{\code{rtBSplines} object created by \link{dat2bsplinemat}}

\item{newdata}{\code{data.frame} of new data.}

\item{...}{Not used.}
}
\description{
Predict S3 method for \code{rtBSplines}
}
\author{
E.D. Gennatas
}
