% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_cv.R
\name{predict.rtModCVCalibration}
\alias{predict.rtModCVCalibration}
\title{Predict using calibrated model}
\usage{
\method{predict}{rtModCVCalibration}(object, mod, newdata, ...)
}
\arguments{
\item{object}{\code{rtModCVCalibration} object returned by \link{calibrate_cv}}

\item{mod}{\code{rtModCV} object returned by \link{train_cv}}

\item{newdata}{Data frame: New data to predict on}

\item{...}{Additional arguments - Use to define \code{which.repeat}, which should be an integer
defining which repeat to use for prediction. Defaults to 1 if not specified.}
}
\description{
Predict using a calibrated model returned by \link{calibrate_cv}
}
\author{
EDG
}
