% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtree.R
\name{preorderTree.addtree}
\alias{preorderTree.addtree}
\title{\code{rtemis-internal} Traverse ADDTREE tree by preorder}
\usage{
preorderTree.addtree(rt, x, verbose = FALSE)
}
\arguments{
\item{rt}{rt Object from \link{s.ADDTREE}}

\item{x}{Features}

\item{verbose}{Logical: If TRUE, print messages to stdout}
}
\description{
Recursively Traverses ADDTREE tree by preorder function and builds data frame representation
}
\author{
E.D. Gennatas
}
