% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Data}
\usage{
## S7 generic
preprocess(x, parameters, ...)
}
\arguments{
\item{x}{data.frame or similar: Data to be preprocessed.}

\item{parameters}{PreprocessorParameters or Preprocessor: PreprocessorParameters when
preprocessing training set data. Setup using \link{setup_Preprocessor}.
Preprocessor when preprocessing validation and test set data.}

\item{...}{Used to pass \code{dat_validation} and \code{dat_test} to the method for Preprocessor.}
}
\value{
Preprocessor object.
}
\description{
Preprocess data for analysis and visualization.
}
\details{
Methods are provided for preprocessing training set data, which accepts a PreprocessorParameters
object, and for preprocessing validation and test set data, which accept a Preprocessor
object.

Order of operations:
\itemize{
\item keep complete cases only
\item remove constants
\item remove duplicates
\item remove cases by missingness threshold
\item remove features by missingness threshold
\item integer to factor
\item integer to numeric
\item logical to factor
\item logical to numeric
\item numeric to factor
\item cut numeric to n bins
\item cut numeric to n quantiles
\item numeric with less than N unique values to factor
\item character to factor
\item factor NA to named level
\item add missingness column
\item impute
\item scale and/or center
\item one-hot encoding
}
}
\author{
EDG
}
