% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R, R/preprocess.data.table.R,
%   R/preprocess.default.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess.data.table}
\alias{preprocess.default}
\title{Data preprocessing}
\usage{
preprocess(x, ...)

\method{preprocess}{data.table}(x, completeCases = FALSE,
  removeCases.thres = NULL, removeFeatures.thres = NULL,
  impute = FALSE, impute.discrete = getMode, impute.numeric = mean,
  integer2factor = FALSE, integer2numeric = FALSE,
  logical2factor = FALSE, logical2numeric = FALSE,
  nonzeroFactors = FALSE, scale = FALSE, center = FALSE,
  removeConstant = TRUE, verbose = TRUE, ...)

\method{preprocess}{default}(x, y = NULL, completeCases = FALSE,
  removeCases.thres = NULL, removeFeatures.thres = NULL,
  impute = FALSE, impute.type = c("missForest", "rfImpute",
  "meanMode"), impute.niter = 10, impute.ntree = 500,
  missForest.parallelize = c("no", "variables", "forests"),
  impute.discrete = getMode, impute.numeric = mean,
  integer2factor = FALSE, integer2numeric = FALSE,
  logical2factor = FALSE, logical2numeric = FALSE,
  numeric2factor = FALSE, numeric2factor.levels = NULL,
  nonzeroFactors = FALSE, scale = FALSE, center = FALSE,
  removeConstant = TRUE, oneHot = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{Input}

\item{completeCases}{Logical: If TRUE, only retain complete cases (no missing data).
Default = FALSE}

\item{removeCases.thres}{Float: Remove cases with >= to this fraction of missing features.
Default = NULL}

\item{removeFeatures.thres}{Float: Remove features with missing values in >= to this fraction of
cases. Default = NULL}

\item{impute}{Logical: If TRUE, impute missing cases. See \code{impute.discrete} and
\code{impute.numeric} for how}

\item{impute.discrete}{Function that returns single value: How to impute discrete variables for 
\code{impute.type = "meanMode"}. Default = \link{getMode}}

\item{impute.numeric}{Function that returns single value: How to impute continuous variables for 
\code{impute.type = "meanMode"}.
Default = \code{mean}}

\item{integer2factor}{Logical: If TRUE, convert all integers to factors}

\item{integer2numeric}{Logical: If TRUE, convert all integers to numeric (will only work 
if \code{integer2factor = FALSE})}

\item{nonzeroFactors}{Logical: Shift factor values to exclude zeros. Default = FALSE}

\item{scale}{Logical: If TRUE, scale columns of \code{x}}

\item{center}{Logical: If TRUE, center columns of \code{x}}

\item{removeConstant}{Logical: If TRUE, remove all columns with zero variance. Default = TRUE}

\item{verbose}{Logical: If TRUE, write messages to console. Default = TRUE}

\item{impute.type}{String: How to impute data: "missForest" uses the package of the same name to impute by iterative
random forest regression. "rfImpute" uses \code{randomForest::rfImpute} (see its documentation), "meanMode" will use
mean and mode by default or any custom function defined in \code{impute.discrete} and \code{impute.numeric}}
}
\description{
Prepare data for data analysis
}
\details{
By default, only removes constant features, everything else can be specified.

Order of operations:
  * completeCases
  * removeCases.thres
  * removeFeatures.thres
  * integer2factor
  * nonzeroFactors
  * impute
  * scale/center
  * removeConstant
}
\author{
Efstathios D. Gennatas
}
