% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_init.R, R/preprocess.R
\name{preprocess}
\alias{preprocess}
\alias{preprocess,data.frame,rtemis::PreprocessorConfig-method}
\alias{preprocess,data.table,rtemis::PreprocessorConfig-method}
\alias{preprocess,data.frame,rtemis::Preprocessor-method}
\alias{preprocess,data.table,rtemis::Preprocessor-method}
\title{Preprocess Data}
\usage{
preprocess(x, config, ...)

## S7 method for classes <data.frame>/<data.table>, <rtemis::PreprocessorConfig>
preprocess(x, config, dat_validation = NULL, dat_test = NULL, verbosity = 1L)

## S7 method for classes <data.frame>/<data.table>, <rtemis::Preprocessor>
preprocess(x, config, verbosity = 1L)
}
\arguments{
\item{x}{data.frame, data.table, tbl_df (tabular data): Data to be preprocessed.}

\item{config}{\code{PreprocessorConfig}: Setup using \link{setup_Preprocessor} OR \code{Preprocessor} object:
Output of previous run of \code{preprocess}. This allows, for example, applying preprocessing to a
validation or test set using the same parameters as were used for the training set. In
particular, the same scale centers and coefficients will be applied to the new data.}

\item{...}{Not used.}

\item{dat_validation}{tabular data: Validation set data.}

\item{dat_test}{tabular data: Test set data.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{Preprocessor} object.
}
\description{
Preprocess data for analysis and visualization.
}
\details{
Methods are provided for preprocessing training set data, which accepts a \code{PreprocessorConfig}
object, and for preprocessing validation and test set data, which accept a \code{Preprocessor}
object.
}
\examples{
# Setup a `Preprocessor`: this outputs a `PreprocessorConfig` object.
prp <- setup_Preprocessor(remove_duplicates = TRUE, scale = TRUE, center = TRUE)

# Includes a long list of parameters
prp

# Resample iris to get train and test data
res <- resample(iris, setup_Resampler(seed = 2026))
iris_train <- iris[res[[1]], ]
iris_test <- iris[-res[[1]], ]

# Preprocess training data
iris_pre <- preprocess(iris_train, prp)

# Access preprocessd training data with `preprocessed()`
preprocessed(iris_pre)

# Apply the same preprocessing to test data
# In this case, the scale and center values from training data will be used.
# Note how `preprocess()` accepts either a `PreprocessorConfig` or `Preprocessor` object for
# this reason.
iris_test_pre <- preprocess(iris_test, iris_pre)

# Access preprocessed test data
preprocessed(iris_test_pre)
}
\author{
EDG
}
