% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Data}
\usage{
## S7 generic
preprocess(x, config, ...)
}
\arguments{
\item{x}{data.frame or similar: Data to be preprocessed.}

\item{config}{\code{PreprocessorConfig} or \code{Preprocessor}: PreprocessorConfig when
preprocessing training set data. Setup using \link{setup_Preprocessor}.
Preprocessor when preprocessing validation and test set data.}

\item{...}{Used to pass \code{dat_validation} and \code{dat_test} to the method for Preprocessor.}
}
\value{
\code{Preprocessor} object.
}
\description{
Preprocess data for analysis and visualization.
}
\details{
Methods are provided for preprocessing training set data, which accepts a PreprocessorConfig
object, and for preprocessing validation and test set data, which accept a Preprocessor
object.

Order of operations:
\itemize{
\item keep complete cases only
\item remove constants
\item remove duplicates
\item remove cases by missingness threshold
\item remove features by missingness threshold
\item integer to factor
\item integer to numeric
\item logical to factor
\item logical to numeric
\item numeric to factor
\item cut numeric to n bins
\item cut numeric to n quantiles
\item numeric with less than N unique values to factor
\item character to factor
\item factor NA to named level
\item add missingness column
\item impute
\item scale and/or center
\item one-hot encoding
}
}
\examples{
# Setup a `Preprocessor`: this outputs a `PreprocessorConfig` object.
prp <- setup_Preprocessor(remove_duplicates = TRUE, scale = TRUE, center = TRUE)

# Includes a long list of parameters
prp

# Resample iris to get train and test data
res <- resample(iris, setup_Resampler(seed = 2026))
iris_train <- iris[res[[1]], ]
iris_test <- iris[-res[[1]], ]

# Preprocess training data
iris_pre <- preprocess(iris_train, prp)

# Access preprocessd training data with `preprocessed()`
preprocessed(iris_pre)

# Apply the same preprocessing to test data
# In this case, the scale and center values from training data will be used.
# Note how `preprocess()` accepts either a `PreprocessorConfig` or `Preprocessor` object for
# this reason.
iris_test_pre <- preprocess(iris_test, iris_pre)

# Access preprocessed test data
preprocessed(iris_test_pre)
}
\author{
EDG
}
