% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{preprocessed}
\alias{preprocessed}
\title{Get preprocessed data from Preprocessor}
\usage{
preprocessed(x, ...)
}
\arguments{
\item{x}{Preprocessor: A Preprocessor object.}
}
\value{
data.frame: The preprocessed data.
}
\description{
Get preprocessed data from Preprocessor
}
