% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{preprocessed}
\alias{preprocessed}
\title{Get preprocessed data from \code{Preprocessor}.}
\usage{
preprocessed(x)
}
\arguments{
\item{x}{\code{Preprocessor}: A \code{Preprocessor} object.}
}
\value{
data.frame: The preprocessed data.
}
\description{
Returns the preprocessed data from a \code{Preprocessor} object.
}
