% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{present.list}
\alias{present.list}
\title{Present list of Supervised or SupervisedRes objects}
\usage{
present.list(
  x,
  metric = NULL,
  model_names = NULL,
  ylim = NULL,
  theme = choose_theme(),
  boxpoints = "all"
)
}
\arguments{
\item{x}{List of Supervised or SupervisedRes objects.}

\item{metric}{Character: Metric to plot.}

\item{model_names}{Character: Names of models being plotted.}

\item{ylim}{Numeric vector of length 2: y-axis limits for the boxplots.}

\item{theme}{Theme object.}

\item{boxpoints}{Character: "all", "outliers", or "suspectedoutliers". Determines how points are
displayed in the boxplot.}
}
\value{
plotly object
}
\description{
Plot training and testing performance boxplots of multiple \code{Supervised} or \code{SupervisedRes} objects
}
\author{
EDG
}
