% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{present.list}
\alias{present.list}
\title{Present list of Supervised or SupervisedRes objects}
\usage{
present.list(
  x,
  metric = NULL,
  model_names = NULL,
  ylim = NULL,
  theme = choose_theme(getOption("rtemis_theme")),
  boxpoints = "all",
  filename = NULL,
  file_width = 800,
  file_height = 600,
  file_scale = 1,
  verbosity = 1L
)
}
\arguments{
\item{x}{List of \code{Supervised} or \code{SupervisedRes} objects.}

\item{metric}{Character: Metric to plot.}

\item{model_names}{Character: Names of models being plotted.}

\item{ylim}{Numeric vector of length 2: y-axis limits for the boxplots.}

\item{theme}{\code{Theme} object.}

\item{boxpoints}{Character: "all", "outliers", or "suspectedoutliers". Determines how points are
displayed in the boxplot.}

\item{filename}{Character: Filename to save the plot to.}

\item{file_width}{Numeric: Width of the exported image in pixels.}

\item{file_height}{Numeric: Height of the exported image in pixels.}

\item{file_scale}{Numeric: Scale factor for the exported image.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
plotly object
}
\description{
Plot training and testing performance boxplots of multiple \code{Supervised} or \code{SupervisedRes} objects
}
\examples{
\dontrun{
iris_lightrf <- train(
  iris,
  algorithm = "lightrf",
  outer_resampling_config = setup_Resampler(seed = 2026)
)
iris_rsvm <- train(
  iris,
  algorithm = "radialsvm",
  outer_resampling_config = setup_Resampler(seed = 2026)
)
present(list(iris_lightrf, iris_rsvm), metric = "Balanced_Accuracy")
}
}
\author{
EDG
}
