% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R, R/present.R
\name{present}
\alias{present}
\title{Present rtemis object}
\usage{
present(x, ...)
}
\arguments{
\item{x}{List of SupervisedRes objects.}

\item{...}{Additional arguments passed to the plotting function.}

\item{metric}{Character: Metric to plot.}

\item{model_names}{Character: Names of models being plotted.}

\item{ylim}{Numeric vector of length 2: y-axis limits for the boxplots.}

\item{theme}{Theme object.}

\item{boxpoints}{Character: "all", "outliers", or "suspectedoutliers". Determines how points are
displayed in the boxplot.}
}
\value{
A plotly object along with console output.

plotly object
}
\description{
This generic is used to present an rtemis object using plots and text.

Plot training and testing performance boxplots of multiple \code{SupervisedRes} objects
}
\author{
EDG
}
