% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{present}
\alias{present}
\title{Present elevate models}
\usage{
present(
  ...,
  which.repeat = 1,
  metric = NULL,
  main = NULL,
  htest = "wilcox.test",
  htest.annotate.y = -0.05
)
}
\arguments{
\item{...}{rtModCV objects created with \link{elevate}}
}
\description{
Present elevate models
}
\author{
E.D. Gennatas
}
