% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{present}
\alias{present}
\title{Present elevate models}
\usage{
present(
  ...,
  mod.names = NULL,
  which.repeat = 1,
  metric = NULL,
  plot.train = TRUE,
  plot.test = TRUE,
  boxpoints = "all",
  annotate_meansd = TRUE,
  main = NULL,
  ylim = NULL,
  htest = "none",
  htest.annotate.y = NULL,
  col = NULL,
  theme = rtTheme,
  margin = list(b = 65, l = 100, t = 60, r = 18, pad = 0),
  subplot.margin = 0.0666,
  filename = NULL,
  file.width = 500,
  file.height = 550,
  file.scale = 1
)
}
\arguments{
\item{...}{rtModCV objects created with \link{train_cv}}

\item{mod.names}{Character: Names of models being plotted.}

\item{which.repeat}{Integer: which \code{rtModCV} repeat to plot.}

\item{metric}{Character: which metric to plot.}

\item{plot.train}{Logical: If TRUE, plot training performance.}

\item{plot.test}{Logical: If TRUE, plot testing performance.}

\item{boxpoints}{Character or FALSE: "all", "suspectedoutliers", "outliers"
See \url{https://plotly.com/r/box-plots/#choosing-the-algorithm-for-computing-quartiles}}

\item{annotate_meansd}{Logical: If TRUE, annotate with mean (SD) of each box}

\item{main}{Character: Plot title.}

\item{ylim}{Numeric vector: y-axis limits}

\item{htest}{Character: e.g. "t.test", "wilcox.test" to compare each box to
the \emph{first} box. If grouped, compare within each group to the first box.
If p-value of test is less than \code{htest.thresh}, add asterisk above/
to the side of each box}

\item{htest.annotate.y}{Numeric: y-axis paper coordinate for htest annotation}

\item{col}{Color, vector: Color for boxes. If NULL, which will draw
colors from \code{palette}}

\item{theme}{Character: Theme to use: Run \code{themes()} for available themes}

\item{margin}{Named list: plot margins.
Default = \code{list(b = 65, l = 65, t = 50, r = 10, pad = 0)}}

\item{subplot.margin}{Numeric: margin between subplots.}

\item{filename}{Character: Path to file to save static plot.}

\item{file.width}{Integer: File width in pixels for when \code{filename} is
set.}

\item{file.height}{Integer: File height in pixels for when \code{filename}
is set.}

\item{file.scale}{Numeric: If saving to file, scale plot by this number}
}
\description{
Plot training and testing performance boxplots of multiple `rtModCV``
objects created by \link{train_cv} using \link{dplot3_box}
}
\author{
E.D. Gennatas
}
