% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_init.R
\name{present}
\alias{present}
\title{Present rtemis object}
\usage{
present(x, ...)
}
\arguments{
\item{x}{\code{Supervised} or \code{SupervisedRes} object or list of such objects.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A plotly object.
}
\description{
This generic is used to present an rtemis object by printing to console and drawing plots.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
ir <- set_outcome(iris, "Sepal.Length")
seplen_lightrf <- train(ir, hyperparameters = setup_LightRF())
present(seplen_lightrf)
\dontshow{\}) # examplesIf}
}
\author{
EDG
}
