% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{present}
\alias{present}
\title{Present rtemis object}
\usage{
present(x, ...)
}
\arguments{
\item{x}{\code{Supervised} or \code{SupervisedRes} object or list of such objects.}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
A plotly object along with console output.
}
\description{
This generic is used to present an rtemis object using plots and text.
}
\examples{
\dontrun{
species_lightrf <- train(iris, algorithm = "lightrf")
present(species_lightrf)
# LightGBM Random Forest was used for classification.
# Balanced accuracy was 0.95 on the training set.
}
}
\author{
EDG
}
