% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartLiteBoostTV.R
\name{print.cartLiteBoostTV}
\alias{print.cartLiteBoostTV}
\title{Print method for \link{boost} object}
\usage{
\method{print}{boost}(x, ...)
}
\arguments{
\item{x}{\code{cartLiteBoostTV} object}

\item{...}{Additional arguments}
}
\description{
Print method for \link{boost} object
}
\author{
E.D. Gennatas
}
