% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{print.CheckData}
\alias{print.CheckData}
\title{Print \code{CheckData} object}
\usage{
\method{print}{CheckData}(
  x,
  type = c("plaintext", "html"),
  name = NULL,
  check_integers = FALSE,
  css = list(font.family = "Helvetica", color = "#fff", background.color = "#242424"),
  ...
)
}
\arguments{
\item{x}{\code{CheckData} object.}

\item{type}{Character: Output type: "plaintext" or "html".}

\item{name}{Character: Dataset name.}

\item{check_integers}{Logical: If TRUE and there are integer features, prints a
message to consider converting to factors.}

\item{css}{List with \code{font.family}, \code{color}, and \code{background.color} elements.}

\item{...}{Not used.}
}
\description{
Print \code{CheckData} object
}
\author{
E.D. Gennatas
}
