% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmLiteBoostTV.R
\name{print.glmLiteBoostTV}
\alias{print.glmLiteBoostTV}
\title{Print method for \code{glmLiteBoostTV} object}
\usage{
\method{print}{glmLiteBoostTV}(x, ...)
}
\arguments{
\item{x}{\code{glmLiteBoostTV} object}

\item{...}{Not used}
}
\description{
Print method for \code{glmLiteBoostTV} object
}
\author{
E.D. Gennatas
}
