% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{print.hytboost}
\alias{print.hytboost}
\title{Print method for \link{boost} object}
\usage{
\method{print}{hytboost}(x, ...)
}
\description{
Print method for \link{boost} object
}
\author{
E.D. Gennatas
}
