% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboost.R
\name{print.hytboost}
\alias{print.hytboost}
\title{Print method for \code{hytboost} object}
\usage{
\method{print}{hytboost}(x, ...)
}
\arguments{
\item{x}{\code{hytboost} object}

\item{...}{Not used}
}
\description{
Print method for \code{hytboost} object
}
\author{
E.D. Gennatas
}
