% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytboostnow.R
\name{print.hytboostnow}
\alias{print.hytboostnow}
\title{Print method for \code{boost} object}
\usage{
\method{print}{hytboostnow}(x, ...)
}
\arguments{
\item{x}{\code{hytboostnow} object}

\item{...}{Not used}
}
\description{
Print method for \code{boost} object
}
\author{
E.D. Gennatas
}
