% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massGLAM.R
\name{print.massGAM}
\alias{print.massGAM}
\title{\code{print}\link{massGAM} object}
\usage{
\method{print}{massGAM}(x, ...)
}
\arguments{
\item{x}{\link{massGAM} object}

\item{...}{Not used}
}
\description{
\code{print}\link{massGAM} object
}
\author{
E.D. Gennatas
}
