% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survError.R
\name{print.survError}
\alias{print.survError}
\title{Print \link{survError}}
\usage{
\method{print}{survError}(x, decimal.places = 4, ...)
}
\arguments{
\item{x}{Object of type \link{survError}}

\item{...}{Not used}

\item{decimal.place}{Integer: Number of decimal places to print. Default = 4}
}
\description{
Print \link{survError}
}
\author{
E.D. Gennatas
}
