% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_print.R
\name{printls}
\alias{printls}
\title{Pretty print list}
\usage{
printls(
  x,
  prefix = "",
  pad = 2L,
  item_format = bold,
  maxlength = 4L,
  center_title = TRUE,
  title = NULL,
  title_newline = TRUE,
  newline_pre = FALSE,
  format_fn_rhs = ddSci,
  print_class = TRUE,
  abbrev_class_n = 3L,
  print_df = FALSE,
  print_S4 = FALSE,
  limit = 12L
)
}
\arguments{
\item{x}{list or object that will be converted to a list.}

\item{prefix}{Character: Optional prefix for names.}

\item{pad}{Integer: Pad output with this many spaces.}

\item{center_title}{Logical: If TRUE, autopad title for centering, if present.}

\item{print_class}{Logical: If TRUE, print abbreviated class of object.}

\item{abbrev_class_n}{Integer: Number of characters to abbreviate class names to.}

\item{print_df}{Logical: If TRUE, print data frame contents, otherwise print n rows and columns.}

\item{print_S4}{Logical: If TRUE, print S4 object contents, otherwise print class name.}

\item{format_fn}{Formatting function.}

\item{output_type}{Character: One of "console", "ansi", "html" for output formatting.}
}
\description{
Pretty print a list (or data frame) recursively
}
\details{
Data frames in R began life as lists
}
\author{
EDG
}
\keyword{internal}
