% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.addtree.R
\name{prune.addtree}
\alias{prune.addtree}
\title{Prune ADDTREE tree}
\usage{
prune.addtree(
  addtree,
  prune.empty.leaves = TRUE,
  remove.bad.parents = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{addtree}{rtMod trained with \link{s.ADDTREE}}

\item{prune.empty.leaves}{Logical: If TRUE, remove leaves with 0 cases. Default = TRUE}

\item{remove.bad.parents}{Logical: If TRUE, remove nodes with no siblings but children and
give their children to their parent. Default = TRUE}
}
\description{
Prune an ADDTREE tree in Node format using \code{data.tree} to remove sister nodes with same
class estimate.
}
\author{
E.D. Gennatasd
}
