% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read delimited file or XLSX into a data.table}
\usage{
read(
  filename,
  datadir = NULL,
  make.unique = TRUE,
  character2factor = FALSE,
  clean.colnames = TRUE,
  delim.reader = c("data.table", "vroom", "duckdb", "arrow", "rpolars"),
  xlsx.sheet = 1,
  sep = NULL,
  quote = "\\"",
  na.strings = c(""),
  rpolars_ignore_errors = TRUE,
  rpolars_infer_schema_length = 100,
  rpolars_encoding = "utf8-lossy",
  rpolars_parse_dates = TRUE,
  output = c("data.table", "default"),
  attr = NULL,
  value = NULL,
  verbose = TRUE,
  fread_verbose = FALSE,
  timed = verbose,
  ...
)
}
\arguments{
\item{filename}{Character: filename or full path if \code{datadir = NULL}}

\item{datadir}{Character: Optional path to directory where \code{filename}
is located. If not specified, \code{filename} must be the full path.}

\item{make.unique}{Logical: If TRUE, keep unique rows only}

\item{character2factor}{Logical: If TRUE, convert character variables to
factors}

\item{clean.colnames}{Logical: If TRUE, clean columns names using
\link{clean_colnames}}

\item{delim.reader}{Character: "data.table" or "arrow", to use
\code{data.table::fread()} or \code{arrow::read_delim_arrow()}, respectively,
to read \code{filename}}

\item{xlsx.sheet}{Integer or character: Name or number of XLSX sheet to read}

\item{sep}{Single character: field separator. If \code{delim.reader = "fread"}
and \code{sep = NULL}, this defaults to "auto", otherwise defaults to ","}

\item{quote}{Single character: quote character}

\item{na.strings}{Character vector: Strings to be interpreted as NA values.
For \code{delim.reader = "duckdb"}, this must be a single string.
For \code{delim.reader = "rpolars"}, this must be a single string, otherwise, if an
unnamed character vector, it maps each string to each column. If named, the names 
should match columns. See \code{?rpolars::csv_reader} for more details.}

\item{output}{Character: "default" or "data.table", If default, return the delim.reader's
default data structure, otherwise convert to data.table}

\item{verbose}{Logical: If TRUE, print messages to console}

\item{fread_verbose}{Logical: Passed to \code{data.table::fread}}

\item{timed}{Logical: If TRUE, time the process and print to console}

\item{...}{Additional parameters to pass to \code{data.table::fread},
\code{arrow::read_delim_arrow()}, \code{vroom::vroom()}, 
or \code{openxlsx::read.xlsx()}}
}
\description{
Convenience function to read delimited data using
\code{data.table:fread()}, \code{arrow:read_delim_arrow()},
\code{vroom::vroom()}, or \code{duckdb::duckdb_read_csv()}.
Read XLSX files using \code{readxl::read_excel()}.
}
\examples{
\dontrun{
datadir <- "~/icloud/Data"
dat <- read("iris.csv", datadir)
}
}
\author{
E.D. Gennatas
}
