% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read rtemis configuration file}
\usage{
read_config(config.path, verbose = TRUE)
}
\arguments{
\item{file}{Character: Path to configuration file created by \link{create_config}.}
}
\value{
List.
}
\description{
Reads rtemis configuration file.
}
\author{
EDG
}
