% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_SuperConfig.R
\name{read_config}
\alias{read_config}
\title{Read \code{SuperConfig} from TOML file}
\usage{
read_config(file)
}
\arguments{
\item{file}{Character: Path to input TOML file.}
}
\value{
\code{SuperConfig} object.
}
\description{
Read \code{SuperConfig} object from TOML file that was written with \code{write_toml()}.
}
\author{
EDG
}
