% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{recycle}
\alias{recycle}
\title{Recycle values of vector to match length of target}
\usage{
recycle(x, target)
}
\arguments{
\item{x}{Vector to be recycled}

\item{target}{Object whose length defines target length}
}
\description{
Recycle values of vector to match length of target
}
\author{
E.D. Gennatas
}
