% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ddSci}
\alias{labelify}
\alias{clean_names}
\alias{clean_colnames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rtemis.core}{\code{\link[rtemis.core:clean_colnames]{clean_colnames()}}, \code{\link[rtemis.core:clean_names]{clean_names()}}, \code{\link[rtemis.core:ddSci]{ddSci()}}, \code{\link[rtemis.core:labelify]{labelify()}}}
}}

