% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{regression_metrics}
\alias{regression_metrics}
\title{Regression Metrics}
\usage{
regression_metrics(true, predicted, na.rm = TRUE, sample = character())
}
\arguments{
\item{true}{Numeric vector: True values.}

\item{predicted}{Numeric vector: Predicted values.}

\item{na.rm}{Logical: If TRUE, remove NA values before computation.}

\item{sample}{Character: Sample name (e.g. "training", "testing").}
}
\value{
RegressionMetrics object
}
\description{
Regression Metrics
}
\author{
EDG
}
