% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample data}
\usage{
resample(x, parameters = setup_Resampler(), verbosity = 1L)
}
\arguments{
\item{x}{Vector or data.frame: Usually the outcome; \code{NROW(y)} defines sample size}

\item{parameters}{Resampler object created by \link{setup_Resampler}.}

\item{verbosity}{Logical: If TRUE, print messages to console.}
}
\description{
Create resamples of your data, e.g. for model building or validation.
"KFold" creates stratified folds, , "StratSub" creates stratified subsamples,
"Bootstrap" gives the standard bootstrap, i.e. random sampling with replacement,
while "StratBoot" uses StratSub and then randomly duplicates some of the training cases to
reach original length of input (default) or length defined by \code{target_length}.
}
\details{
Note that option 'KFold' may result in resamples of slightly different length. Avoid all
operations which rely on equal-length vectors. For example, you can't place resamples in a
data.frame, but must use a list instead.
}
\examples{
\dontrun{
y <- rnorm(200)
# 10-fold (stratified)
res <- resample(y, 10, "KFold")
# 25 stratified subsamples
res <- resample(y, 25, "StratSub")
# 100 stratified bootstraps
res <- resample(y, 100, "StratBoot")
}
}
\author{
EDG
}
