% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resampling methods}
\usage{
resample(y = NULL, n.resamples = 10, resampler = c("strat.sub",
  "strat.boot", "kfold", "bootstrap", "loocv"), index = NULL,
  group = NULL, stratify.var = y, cv.p = 0.75, cv.groups = 4,
  target.length = NROW(y), rtset = NULL, seed = NULL,
  verbose = FALSE)
}
\arguments{
\item{y}{Numeric vector. Usually the outcome; \code{length(y)} defines sample size}

\item{n.resamples}{Integer: Number of training/testing sets required}

\item{resampler}{String: Type of resampling to perform: "bootstrap", "kfold", "strat.boot", "strat.sub".
Default = "strat.boot" for \code{length(y) < 200}, otherwise "strat.sub"}

\item{index}{List where each element is a vector of training set indices. Use this for manual or precalculated
train/test splits}

\item{group}{Integer, vector, length = \code{length(y)}: Integer vector, where numbers define fold membership.
e.g. for 10-fold on a dataset with 1000 cases, you could use \code{group = rep(1:10, each = 100)}}

\item{stratify.var}{Numeric vector (optional): Variable used for stratification. Defaults to \code{y}}

\item{cv.p}{Float (0, 1): Fraction of cases to assign to traininig set for \code{resampler = "strat.sub"}}

\item{cv.groups}{Integer: Number of groups to use for stratification for
\code{resampler = "strat.sub" / "strat.boot"}}

\item{target.length}{Integer: Number of cases for training set for \code{resampler = "strat.boot"}.
Default = \code{length(y)}}

\item{rtset}{List: Output of an \link{rtset.resample} (or named list with same structure).
NOTE: Overrides all other arguments. Default = NULL}

\item{seed}{Integer: (Optional) Set seed for random number generator, in order to make output reproducible.
See \code{?base::set.seed}}

\item{verbose}{Logical: If TRUE, print messages to screen}
}
\description{
Create resamples of your data, e.g. for model building or validation.
"bootstrap" gives the standard bootstrap, i.e. random sampling with replacement, using \code{caret::createResample},
"strat.sub" creates stratified subsamples using \code{caret::createDataPartition}, while "strat.boot"
runs \code{caret::createDataPartition} and then randomly duplicates some of the training cases to reach original
length of input (default) or length defined by \code{target.length}.
}
\details{
\code{resample} is used by multiple \pkg{rtemis} learners and functions. Note that option 'kfold', which uses \code{caret::createFolds} results in resamples of slightly
  different length for y of small length, so avoid all operations which rely on equal-length vectors.
  For example, you can't place resamples in a data.frame, but must use a list instead.
}
\seealso{
\link{elevate}
}
\author{
Efstathios D. Gennatas
}
