% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample data}
\usage{
resample(x, config = setup_Resampler(), verbosity = 1L)
}
\arguments{
\item{x}{Vector or data.frame: Usually the outcome; \code{NROW(x)} defines the sample size. A
config naming a column -- \code{stratify_var} or \code{id_strat} -- needs the data frame that column
lives in, since a name cannot be looked up in a bare vector.}

\item{config}{Resampler object created by \link{setup_Resampler}.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{Resampler} object.
}
\description{
Create resamples of your data, e.g. for model building or validation.
"KFold" creates stratified folds, , "StratSub" creates stratified subsamples,
"Bootstrap" gives the standard bootstrap, i.e. random sampling with replacement,
while "StratBoot" uses StratSub and then randomly duplicates some of the training cases to
reach the original length of the input, or the length defined by \code{target_length}.
}
\details{
Note that option 'KFold' may result in resamples of slightly different length. Avoid all
operations which rely on equal-length vectors. For example, you can't place resamples in a
data.frame, but must use a list instead.
}
\examples{
y <- rnorm(200)
# 10-fold (stratified)
y_10fold <- resample(y, setup_Resampler(10L, "kfold"))
y_10fold
# 25 stratified subsamples
y_25strat <- resample(y, setup_Resampler(25L, "stratsub"))
y_25strat
# 100 stratified bootstraps
y_100strat <- resample(y, setup_Resampler(100L, "stratboot"))
y_100strat
# LOOCV
y_loocv <- resample(y, setup_Resampler(type = "LOOCV"))
y_loocv
# User-supplied resamples
y_custom <- resample(
  y,
  setup_Resampler(type = "Custom", resamples = list(1:150, 51:200))
)
y_custom
}
\author{
EDG
}
