% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{rowMax}
\alias{rowMax}
\title{Collapse data.frame to vector by getting row max}
\usage{
rowMax(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input vector}

\item{na.rm}{Logical. If TRUE, missing values are not considered.}
}
\description{
Collapse data.frame to vector by getting row max
}
\author{
E.D. Gennatas
}
