% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_reactable.R
\name{rt_reactable}
\alias{rt_reactable}
\title{View table using reactable}
\usage{
rt_reactable(
  x,
  datatypes = NULL,
  lightsout = TRUE,
  bg = "#121212",
  pagination = TRUE,
  searchable = TRUE,
  bordered = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame, data.table or similar}

\item{lightsout}{Logical: If TRUE, use dark theme}
}
\description{
View table using reactable
}
\author{
E D Gennatas
}
