% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_reactable.R
\name{rt_reactable}
\alias{rt_reactable}
\title{View table using reactable}
\usage{
rt_reactable(
  x,
  datatypes = NULL,
  lightsout = TRUE,
  bg = "#121212",
  pagination = TRUE,
  searchable = TRUE,
  bordered = TRUE,
  ...
)
}
\arguments{
\item{x}{data.frame, data.table or similar}

\item{datatypes}{Character vector: Data types of columns in x,
e.g. \code{c("numeric", "factor", "character")}}

\item{lightsout}{Logical: If TRUE, use dark theme.}

\item{bg}{Background color.}

\item{pagination}{Logical: If TRUE, paginate table.}

\item{searchable}{Logical: If TRUE, add search box.}

\item{bordered}{Logical: If TRUE, add border.}

\item{...}{Additional arguments passed to \code{reactable::reactable}}
}
\description{
View table using reactable
}
\author{
E D Gennatas
}
