% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtClust.R
\docType{class}
\name{rtClust-class}
\alias{rtClust-class}
\alias{rtClust}
\title{R6 class for \pkg{rtemis} clustering}
\description{
\pkg{rtemis} clustering object
}
\author{
E.D. Gennatas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{clust.name}}{Character: Name of clustering algorithm}

\item{\code{call}}{Originating call}

\item{\code{xnames}}{Column names of x}

\item{\code{clust}}{Clustering algorithm output}

\item{\code{clusters.train}}{Cluster assignment for training set}

\item{\code{clusters.test}}{Cluster assignment for testing set (if supported by algorithm and test set provided)}

\item{\code{extra}}{List: Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{call}}{Originating call}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtClust$new()}}
\item \href{#method-print}{\code{rtClust$print()}}
\item \href{#method-clone}{\code{rtClust$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtClust$new(
  clust.name = character(),
  k = NULL,
  xnames = character(),
  clust = list(),
  clusters.train = numeric(),
  clusters.test = numeric(),
  parameters = list(),
  extra = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtClust$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtClust$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
