% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDecom.R
\docType{class}
\name{rtDecom-class}
\alias{rtDecom-class}
\alias{rtDecom}
\title{R6 Class for \pkg{rtemis} Decompositions}
\description{
\pkg{rtemis} decomposition R6 object
}
\author{
E.D. Gennatas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{decom.name}}{Character: Name of decomposition algorithm}

\item{\code{call}}{Originating call}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{decom}}{Decomposition model output}

\item{\code{parameters}}{List of decompotision parameters}

\item{\code{center}}{Numeric vector of column means if centering was applied using \code{scale()}
prior to decomposition}

\item{\code{scale}}{Numeric vector of column scale factor if scaling was applied using \code{scale()}
prior to decomposition}

\item{\code{projections.train}}{Input data projected on new axes / basis}

\item{\code{projections.test}}{Input test data projected on new axes / basis}

\item{\code{extra}}{List: Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{call}}{Originating call}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtDecom$new()}}
\item \href{#method-print}{\code{rtDecom$print()}}
\item \href{#method-clone}{\code{rtDecom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtDecom$new(
  decom.name = character(),
  xnames = character(),
  decom = list(),
  parameters = list(),
  center = numeric(),
  scale = numeric(),
  projections.train = numeric(),
  projections.test = numeric(),
  extra = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtDecom$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtDecom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
