% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_init.R
\name{rtemis_init}
\alias{rtemis_init}
\title{Initialize parallel processing and progress reporting}
\usage{
rtemis_init(n.cores = 1, context = NULL, verbose = TRUE)
}
\description{
Initialize parallel processing and progress reporting
}
\keyword{internal}
