% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rtemis_palette}
\alias{rtemis_palette}
\title{Access rtemis palette colors}
\usage{
rtemis_palette(n, palette = rtPalette)
}
\arguments{
\item{n}{Integer: Number of colors to output}

\item{palette}{Character: Palette to use. See available options with \code{rtpalette()}.
Default = \code{rtPalette}}
}
\description{
Allows you to get \code{n} colors of a defined palette, useful for passing to other functions, like ggplot
}
\examples{
\dontrun{
rtemis_palette(3)
}
}
\author{
E.D. Gennatas
}
