% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtInitProjectDir.R
\name{rtInitProjectDir}
\alias{rtInitProjectDir}
\title{Initialize Project Directory}
\usage{
rtInitProjectDir(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, If TRUE, print messages to console}
}
\description{
Initializes Directory Structure: "R", "Data", "Results"
}
\author{
E.D. Gennatas
}
