% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtlayout.R
\name{rtlayout}
\alias{rtlayout}
\title{Create multipanel plots with the \code{mplot3} family}
\usage{
rtlayout(
  nrows = NULL,
  ncols = NULL,
  byrow = FALSE,
  autolabel = FALSE,
  pdf.width = NULL,
  pdf.height = NULL,
  filename = NULL
)
}
\arguments{
\item{nrows}{Integer: N of rows}

\item{ncols}{Integer: N of columns}

\item{byrow}{Logical: If TRUE, draw add plots by row Default = FALSE}

\item{autolabel}{Logical: If TRUE, place letter labels on the top left corner of each figure. Default = FALSE}

\item{pdf.width}{Float: Width of PDF to save, if \code{filename} is provided.
Default = \code{ncols * 4.5}}

\item{pdf.height}{Float: Height of PDF to save, if \code{filename} is provided.
Default = \code{nrows * 4.5}}

\item{filename}{String, optional: Save multiplot to file. Default = NULL}
}
\description{
Set layout for drawing multiple plots in the same view
}
\author{
E.D. Gennatas
}
