% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtMeta-methods}
\alias{rtMeta-methods}
\alias{predict.rtMeta}
\title{rtMeta S3 methods}
\usage{
\method{predict}{rtMeta}(object, newdata, fn = median, ...)
}
\arguments{
\item{object}{\code{rtMeta} object}

\item{newdata}{Testing set features}

\item{fn}{Function to average predictions}

\item{...}{Additional arguments passed to \code{predict(object)}}
}
\description{
S3 methods for \code{rtMeta} class that differ from those of the \code{rtMod} superclass
}
