% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtMod-class}
\alias{rtMod-class}
\alias{rtMod}
\title{\pkg{rtemis} Supervised Model Class}
\description{
R6 Class for \pkg{rtemis} Supervised Models
}
\details{
Note on terminology:
The training set ('x.train') is used to build a model and calculate training error
(rtMod$error.train). The testing set ('x.test') is not seen during training and
only used to create predictions (rtMod$predicted) using the trained model and
calculate error metrics (rtMod$error.test).
This reflects generalizability of the model and is the error we care the most about.
The validation set is used during model tuning. Within rtemis, validation sets are
created automatically by \link{resample} when appropriate, they are not generally
input by the user, with few exceptions, as documented in individual functions.
}
\author{
Efstathios D. Gennatas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mod.name}}{Learner algorithm name}

\item{\code{y.train}}{Training y data}

\item{\code{y.test}}{Testing y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{mod}}{Trained model}

\item{\code{type}}{Classification, Regression, or Survival}

\item{\code{parameters}}{List of hyperparameters used when building model}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{varimp}}{Variable Importance. Note that this is a general concept and different ML algorithms provide very
different approaches}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{extra}}{Algorithm-specific output}

\item{\code{sessionInfo}}{The output of \code{sessionInfo()} at the time the model was trained}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtMod$new()}}
\item \href{#method-print}{\code{rtMod$print()}}
\item \href{#method-plot}{\code{rtMod$plot()}}
\item \href{#method-plotFitted}{\code{rtMod$plotFitted()}}
\item \href{#method-plotPredicted}{\code{rtMod$plotPredicted()}}
\item \href{#method-plotFittedPredicted}{\code{rtMod$plotFittedPredicted()}}
\item \href{#method-plotVarImp}{\code{rtMod$plotVarImp()}}
\item \href{#method-summary}{\code{rtMod$summary()}}
\item \href{#method-describe}{\code{rtMod$describe()}}
\item \href{#method-clone}{\code{rtMod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$new(
  mod.name = character(),
  y.train = numeric(),
  y.test = numeric(),
  x.name = character(),
  y.name = character(),
  xnames = character(),
  mod = list(),
  type = character(),
  parameters = list(),
  fitted = numeric(),
  se.fit = NULL,
  error.train = list(),
  predicted = NULL,
  se.prediction = NULL,
  error.test = NULL,
  varimp = NULL,
  question = character(),
  extra = list(),
  sessionInfo = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plot(
  estimate = NULL,
  filename = NULL,
  theme = getOption("rt.fit.theme", "lightgrid"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotFitted}{}}}
\subsection{Method \code{plotFitted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotFitted(
  print.plot = TRUE,
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPredicted}{}}}
\subsection{Method \code{plotPredicted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotPredicted(
  print.plot = TRUE,
  filename = NULL,
  theme = getOption("rt.fit.theme", "lightgrid"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFittedPredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotFittedPredicted}{}}}
\subsection{Method \code{plotFittedPredicted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotFittedPredicted(
  print.plot = TRUE,
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotVarImp"></a>}}
\if{latex}{\out{\hypertarget{method-plotVarImp}{}}}
\subsection{Method \code{plotVarImp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$plotVarImp(
  plot.top = 12,
  theme = getOption("rt.fit.theme", "lightgrid"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$summary(
  plots = TRUE,
  cex = 1,
  fit.true.line = "lm",
  resid.fit.line = "gam",
  fit.legend = TRUE,
  se.fit = TRUE,
  single.fig = TRUE,
  summary = TRUE,
  theme = getOption("rt.fit.theme", "lightgrid"),
  title.col = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe"></a>}}
\if{latex}{\out{\hypertarget{method-describe}{}}}
\subsection{Method \code{describe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtMod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
