% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtMod-class}
\alias{rtMod-class}
\alias{rtMod}
\title{\pkg{rtemis} Supervised Model Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rtMod
}
\description{
R6 Class for \pkg{rtemis} Supervised Models
}
\details{
Note on terminology:
The training set ('x.train') is used to build a model and calculate training error
(rtMod$error.train). The testing set ('x.test') is not seen during training and
only used to create predictions (rtMod$predicted) using the trained model and
calculate error metrics (rtMod$error.test).
This reflects generalizability of the model and is the error we care the most about.
The validation set is used during model tuning. Within rtemis, validation sets are
created automatically by \link{resample} when appropriate, they are not generally
input by the user, with few exceptions, as documented in individual functions.
}
\section{Fields}{

\describe{
\item{\code{mod.name}}{Learner algorithm name}

\item{\code{y.train}}{Training y data}

\item{\code{y.test}}{Testing y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{mod}}{Trained model}

\item{\code{type}}{Classification, Regression, or Survival}

\item{\code{parameters}}{List of hyperparameters used when building model}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{extra}}{Algorithm-specific output}

\item{\code{sessionInfo}}{The output of \code{sessionInfo()} at the time the model was trained}
}}

\author{
Efstathios D. Gennatas
}
\keyword{datasets}
