% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtMod-methods}
\alias{rtMod-methods}
\alias{print.rtMod}
\alias{fitted.rtMod}
\alias{predict.rtMod}
\alias{residuals.rtMod}
\alias{plot.rtMod}
\alias{summary.rtMod}
\alias{coef.rtMod}
\alias{predict.rtModLite}
\title{\code{rtMod} S3 methods}
\usage{
\method{print}{rtMod}(x, ...)

\method{fitted}{rtMod}(object, ...)

\method{predict}{rtMod}(object, newdata, trace = 0, verbose = TRUE, ...)

\method{residuals}{rtMod}(object, ...)

\method{plot}{rtMod}(
  x,
  estimate = NULL,
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)

\method{summary}{rtMod}(
  object,
  plots = TRUE,
  cex = 1,
  fit.true.line = "lm",
  resid.fit.line = "gam",
  fit.legend = TRUE,
  se.fit = TRUE,
  single.fig = TRUE,
  summary = TRUE,
  theme = getOption("rt.fit.theme", "lightgrid"),
  title.col = NULL,
  ...
)

\method{coef}{rtMod}(object, verbose = TRUE, ...)

\method{predict}{rtModLite}(object, newdata, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to \link{mplot3}}

\item{object}{Trained model of class \code{rtMod}}

\item{newdata}{Testing set features}

\item{theme}{String. \code{mplot3} theme to use. Options: "box", "darkbox", "light", "dark"}

\item{plots}{Logical. Should plots be printed?}

\item{cex}{Numeric. Character expansion factor}

\item{fit.true.line}{\pkg{rtemis} model to use for fitted vs. true line
Options: \code{modSelect()}}

\item{resid.fit.line}{\pkg{rtemis} model to use for residuals vs. fitted line.
Options: \code{modSelect()}}

\item{fit.legend}{Logical. Should fit legends be printed? Defaults to TRUE}

\item{se.fit}{Logical. Should 2 * standard error bands be plotted? Defaults to TRUE}

\item{single.fig}{Logical. Should all plots be in same figure? Defaults to TRUE}

\item{summary}{Logical. Should summary be printed? Defaults to TRUE}

\item{title.col}{Color for main title}
}
\description{
S3 methods for \code{rtMod} class.
Excludes functions \code{print} and \code{plot} defined within the \link{rtMod} class itself.
}
\author{
Efstathios D. Gennatas

Efstathios D. Gennatas
}
