% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtMod.out}
\alias{rtMod.out}
\title{rtMod.out}
\usage{
rtMod.out(
  rt,
  print.plot,
  plot.fitted,
  plot.predicted,
  y.test,
  mod.name,
  outdir,
  save.mod,
  verbose,
  theme,
  ...
)
}
\arguments{
\item{rt}{rtemid object}

\item{print.plot}{Logical: If TRUE, print plot}

\item{plot.fitted}{Logical: If TRUE, print fitted values plot}

\item{plot.predicted}{Logical: If TRUE, print predicted values plot}

\item{y.test}{Test set predictions}

\item{mod.name}{Character: name of algorithm used to train model}

\item{outdir}{Character: Path to directory to save model}

\item{save.mod}{Logical: if TRUE, save trained model to \code{outdir},
otherwise save only log, summary, and plots}

\item{verbose}{Logical: If TRUE, print messages in console}

\item{theme}{rtemis theme to use}

\item{...}{Additional arguments passed to \code{plot(rt)}}
}
\description{
rtMod.out
}
\author{
E.D. Gennatas
}
\keyword{internal}
