% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtModBag-class}
\alias{rtModBag-class}
\alias{rtModBag}
\title{\pkg{rtemis} Bagged Supervised Model Class}
\description{
R6 Class for \pkg{rtemis} Bagged Supervised Models
}
\section{Super class}{
\code{\link[rtemis:rtMod]{rtemis::rtMod}} -> \code{rtModBag}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mod.name}}{Model (algorithm) name}

\item{\code{y.train}}{Training y data}

\item{\code{y.test}}{Testing y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{bag.resample.rtset}}{List of settings for \link{resample}.
Set using \link{rtset.bag.resample}}

\item{\code{mod}}{Trained model}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{extra}}{Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mod.name}}{Model (algorithm) name}

\item{\code{y.train}}{Training y data}

\item{\code{y.test}}{Testing y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{mod}}{Trained model}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{extra}}{Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModBag$new()}}
\item \href{#method-print}{\code{rtModBag$print()}}
\item \href{#method-clone}{\code{rtModBag$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="describe">}\href{../../rtemis/html/rtMod.html#method-describe}{\code{rtemis::rtMod$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plot">}\href{../../rtemis/html/rtMod.html#method-plot}{\code{rtemis::rtMod$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFitted">}\href{../../rtemis/html/rtMod.html#method-plotFitted}{\code{rtemis::rtMod$plotFitted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFittedPredicted">}\href{../../rtemis/html/rtMod.html#method-plotFittedPredicted}{\code{rtemis::rtMod$plotFittedPredicted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotPredicted">}\href{../../rtemis/html/rtMod.html#method-plotPredicted}{\code{rtemis::rtMod$plotPredicted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotVarImp">}\href{../../rtemis/html/rtMod.html#method-plotVarImp}{\code{rtemis::rtMod$plotVarImp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="summary">}\href{../../rtemis/html/rtMod.html#method-summary}{\code{rtemis::rtMod$summary()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModBag$new(
  mod.name = character(),
  y.train = numeric(),
  y.test = numeric(),
  x.name = character(),
  y.name = character(),
  xnames = character(),
  bag.resample.rtset = list(),
  mod = list(),
  type = character(),
  parameters = list(),
  fitted.bag = numeric(),
  fitted = numeric(),
  se.fit.bag = numeric(),
  se.fit = numeric(),
  error.train = list(),
  predicted.bag = numaric(),
  predicted = numeric(),
  se.prediction.bag = numeric(),
  se.prediction = numeric(),
  aggr.fn = character(),
  error.test = list(),
  varimp = NULL,
  question = character(),
  extra = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModBag$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModBag$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
