% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtModBag-methods}
\alias{rtModBag-methods}
\alias{predict.rtModBag}
\title{rtModBag S3 methods}
\usage{
\method{predict}{rtModBag}(object, newdata, fn = "median",
  verbose = FALSE, ...)
}
\arguments{
\item{newdata}{Testing set features}
}
\description{
S3 methods for \code{rtModBag} class that differ from those of the \code{rtMod} superclass
}
