% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtModClass}
\alias{rtModClass}
\title{\pkg{rtemis} Classification Model Class}
\description{
R6 Class for \pkg{rtemis} Classification Models
}
\section{Super class}{
\code{\link[rtemis:rtMod]{rtemis::rtMod}} -> \code{rtModClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModClass$new()}}
\item \href{#method-plotROC}{\code{rtModClass$plotROC()}}
\item \href{#method-plotROCfitted}{\code{rtModClass$plotROCfitted()}}
\item \href{#method-plotROCpredicted}{\code{rtModClass$plotROCpredicted()}}
\item \href{#method-plotPR}{\code{rtModClass$plotPR()}}
\item \href{#method-plotPRfitted}{\code{rtModClass$plotPRfitted()}}
\item \href{#method-plotPRpredicted}{\code{rtModClass$plotPRpredicted()}}
\item \href{#method-clone}{\code{rtModClass$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="describe">}\href{../../rtemis/html/rtMod.html#method-describe}{\code{rtemis::rtMod$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plot">}\href{../../rtemis/html/rtMod.html#method-plot}{\code{rtemis::rtMod$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFitted">}\href{../../rtemis/html/rtMod.html#method-plotFitted}{\code{rtemis::rtMod$plotFitted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotFittedPredicted">}\href{../../rtemis/html/rtMod.html#method-plotFittedPredicted}{\code{rtemis::rtMod$plotFittedPredicted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotPredicted">}\href{../../rtemis/html/rtMod.html#method-plotPredicted}{\code{rtemis::rtMod$plotPredicted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="plotVarImp">}\href{../../rtemis/html/rtMod.html#method-plotVarImp}{\code{rtemis::rtMod$plotVarImp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="print">}\href{../../rtemis/html/rtMod.html#method-print}{\code{rtemis::rtMod$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtMod" data-id="summary">}\href{../../rtemis/html/rtMod.html#method-summary}{\code{rtemis::rtMod$summary()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$new(
  mod.name = character(),
  y.train = numeric(),
  y.test = numeric(),
  x.name = character(),
  y.name = character(),
  xnames = character(),
  mod = list(),
  type = character(),
  gridsearch = list(),
  parameters = list(),
  fitted = numeric(),
  fitted.prob = numeric(),
  se.fit = numeric(),
  error.train = list(),
  predicted = NULL,
  predicted.prob = NULL,
  se.prediction = NULL,
  error.test = NULL,
  varimp = NULL,
  question = character(),
  extra = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROC"></a>}}
\if{latex}{\out{\hypertarget{method-plotROC}{}}}
\subsection{Method \code{plotROC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotROC(
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROCfitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotROCfitted}{}}}
\subsection{Method \code{plotROCfitted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotROCfitted(
  main = "ROC Training",
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROCpredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotROCpredicted}{}}}
\subsection{Method \code{plotROCpredicted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotROCpredicted(
  main = "ROC Testing",
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPR"></a>}}
\if{latex}{\out{\hypertarget{method-plotPR}{}}}
\subsection{Method \code{plotPR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotPR(
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPRfitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPRfitted}{}}}
\subsection{Method \code{plotPRfitted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotPRfitted(
  main = "P-R Training",
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPRpredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPRpredicted}{}}}
\subsection{Method \code{plotPRpredicted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$plotPRpredicted(
  main = "P-R Testing",
  theme = getOption("rt.fit.theme", "lightgrid"),
  filename = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
