% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtModCV-class}
\alias{rtModCV-class}
\alias{rtModCV}
\title{\pkg{rtemis} Cross-Validated Supervised Model Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rtModCV
}
\description{
R6 Class for \pkg{rtemis} Cross-Validated Supervised Models
}
\section{Fields}{

\describe{
\item{\code{mod.name}}{Model (algorithm) name}

\item{\code{y.train}}{Training set y data}

\item{\code{y.test}}{Testing set y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{resampler}}{List of settings for \link{resample}. Set using \link{rtset.cv.resample}}

\item{\code{n.repeats}}{Integer: Number of repeats. This is the outermost iterator: i.e. You will run
\code{resampler} this many times.}

\item{\code{mod}}{Trained model}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{extra}}{Algorithm-specific output}
}}

\keyword{datasets}
