% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtModCV-class}
\alias{rtModCV-class}
\alias{rtModCV}
\title{\pkg{rtemis} Cross-Validated Supervised Model Class}
\description{
\pkg{rtemis} Cross-Validated Supervised Model Class

\pkg{rtemis} Cross-Validated Supervised Model Class
}
\details{
R6 Class for \pkg{rtemis} Cross-Validated Supervised Models
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mod}}{\code{rtModCV} object}

\item{\code{mod.name}}{Model (algorithm) name}

\item{\code{type}}{"Classification", "Regression", or "Survival"}

\item{\code{y.train}}{Training set y data}

\item{\code{y.test}}{Testing set y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{parameters}}{List of algorithm hyperparameters}

\item{\code{resampler}}{List of settings for \link{resample}. 
Set using \link{rtset.cv.resample}}

\item{\code{n.repeats}}{Integer: Number of repeats. This is the outermost iterator: i.e. You will run
\code{resampler} this many times.}

\item{\code{resampler.params}}{List of resampling parameters}

\item{\code{resamples}}{Resamples produced by \link{resample}}

\item{\code{y.train.res}}{Resamples' fitted values}

\item{\code{y.train.res.aggr}}{Aggregated training set outcomes}

\item{\code{fitted.res}}{Resamples' fitted values}

\item{\code{fitted.res.aggr}}{Aggregated fitted values}

\item{\code{error.train.res}}{Resamples' training error}

\item{\code{error.train.res.mean}}{Resamples' mean training error}

\item{\code{error.train.res.aggr}}{Error of aggregated fitted values}

\item{\code{error.train.repeats}}{Mean training error for each repeat}

\item{\code{error.train.repeats.mean}}{Mean error across all repeats}

\item{\code{error.train.repeats.sd}}{Standard deviation of error across all repeats}

\item{\code{y.test.res}}{Resamples' predicted values}

\item{\code{mod}}{Trained model}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{extra}}{Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{y.test}}{Testing set y data}

\item{\code{resampler}}{List of settings for \link{resample}. 
Set using \link{rtset.cv.resample}}

\item{\code{fitted}}{Fitted values}

\item{\code{se.fit}}{Standard error of the fit}

\item{\code{error.train}}{Training error}

\item{\code{predicted}}{Predicted values}

\item{\code{se.prediction}}{Standard error of the prediction}

\item{\code{error.test}}{Testing error}

\item{\code{extra}}{Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModCV$new()}}
\item \href{#method-print}{\code{rtModCV$print()}}
\item \href{#method-plot}{\code{rtModCV$plot()}}
\item \href{#method-plotPredicted}{\code{rtModCV$plotPredicted()}}
\item \href{#method-plotFitted}{\code{rtModCV$plotFitted()}}
\item \href{#method-plotVarImp}{\code{rtModCV$plotVarImp()}}
\item \href{#method-describe}{\code{rtModCV$describe()}}
\item \href{#method-clone}{\code{rtModCV$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{rtModCV} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$new(
  mod = NULL,
  mod.name = NULL,
  type = NULL,
  y.train = NULL,
  x.name = NULL,
  y.name = NULL,
  xnames = NULL,
  parameters = NULL,
  n.repeats = NULL,
  resampler.params = NULL,
  resamples = NULL,
  y.train.res = NULL,
  y.train.res.aggr = NULL,
  fitted.res = NULL,
  fitted.res.aggr = NULL,
  error.train.res = NULL,
  error.train.res.mean = NULL,
  error.train.res.aggr = NULL,
  error.train.repeats = NULL,
  error.train.repeats.mean = NULL,
  error.train.repeats.sd = NULL,
  y.test.res = NULL,
  y.test.res.aggr = NULL,
  predicted.res = NULL,
  predicted.res.aggr = NULL,
  error.test.res = NULL,
  error.test.res.mean = NULL,
  error.test.res.aggr = NULL,
  error.test.repeats = NULL,
  error.test.repeats.mean = NULL,
  error.test.repeats.sd = NULL,
  fitted.bag = NULL,
  error.bag = NULL,
  varimp = NULL,
  question = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{\code{rtModCV} object}

\item{\code{mod.name}}{Character: Algorithm name}

\item{\code{type}}{Character: Type of model (Regression, Classification, Survival)}

\item{\code{y.train}}{Training set output}

\item{\code{x.name}}{Character: Feature set name}

\item{\code{y.name}}{Character: Output name}

\item{\code{xnames}}{Character vector: Feature names}

\item{\code{parameters}}{List of training parameters}

\item{\code{n.repeats}}{Integer: Number of repeats. This is the outermost iterator: i.e. You will run
\code{resampler} this many times.}

\item{\code{varimp}}{Variable importance}

\item{\code{question}}{Question the model is trying to answer}

\item{\code{y.test}}{Testing set output}

\item{\code{gridsearch}}{Grid search output}

\item{\code{fitted}}{Fitted values (training set predictions)}

\item{\code{se.fit}}{Standard error of fitted values}

\item{\code{error.train}}{Training set error}

\item{\code{predicted}}{Predicted values (Testing set predictions)}

\item{\code{se.prediction}}{Standard error of predicted values}

\item{\code{error.test}}{Testing set error}

\item{\code{extra}}{List of extra model info}

\item{\code{sessionInfo}}{R session info at time of training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\code{print} method for \code{rtModCV} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\code{plot} method for \code{rtModCV} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plot(which.repeat = 1, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPredicted}{}}}
\subsection{Method \code{plotPredicted()}}{
Plot predicted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plotPredicted(
  which.repeat = 1,
  theme = getOption("rt.theme"),
  filename = NULL,
  mar = c(2.5, 3, 2.5, 1),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotFitted}{}}}
\subsection{Method \code{plotFitted()}}{
Plot fitted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plotFitted(
  which.repeat = 1,
  theme = getOption("rt.theme"),
  filename = NULL,
  mar = c(2.5, 3, 2.5, 1),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotVarImp"></a>}}
\if{latex}{\out{\hypertarget{method-plotVarImp}{}}}
\subsection{Method \code{plotVarImp()}}{
Plot variable importance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plotVarImp(
  which.repeat = 1,
  type = c("barplot", "lollipop"),
  plot.top = 12,
  theme = getOption("rt.theme"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe"></a>}}
\if{latex}{\out{\hypertarget{method-describe}{}}}
\subsection{Method \code{describe()}}{
Describe \code{rtModCV}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
