% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtModCV-class}
\alias{rtModCV-class}
\alias{rtModCV}
\title{\pkg{rtemis} Cross-Validated Supervised Model Class}
\description{
\pkg{rtemis} Cross-Validated Supervised Model Class

\pkg{rtemis} Cross-Validated Supervised Model Class
}
\details{
R6 Class for \pkg{rtemis} Cross-Validated Supervised Models
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mod}}{\code{rtModCV} object}

\item{\code{mod.name}}{Model (algorithm) name}

\item{\code{type}}{"Classification", "Regression", or "Survival"}

\item{\code{y.train}}{Training set y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{parameters}}{List of algorithm hyperparameters}

\item{\code{n.repeats}}{Integer: Number of repeats. This is the outermost iterator: i.e. You will run
\code{resampler} this many times.}

\item{\code{resampler.params}}{List of resampling parameters}

\item{\code{resamples}}{Resamples produced by \link{resample}}

\item{\code{y.train.res}}{Resamples' fitted values}

\item{\code{y.train.res.aggr}}{Aggregated training set outcomes}

\item{\code{fitted.res}}{Resamples' fitted values}

\item{\code{fitted.res.aggr}}{Aggregated fitted values}

\item{\code{error.train.res}}{Resamples' training error}

\item{\code{error.train.res.mean}}{Resamples' mean training error}

\item{\code{error.train.res.aggr}}{Error of aggregated fitted values}

\item{\code{error.train.repeats}}{Mean training error for each repeat}

\item{\code{error.train.repeats.mean}}{Mean training error across all repeats}

\item{\code{error.train.repeats.sd}}{Standard deviation of training error across all repeats}

\item{\code{y.test.res}}{Resamples' predicted values}

\item{\code{y.test.res.aggr}}{Aggregated testing set outcomes}

\item{\code{predicted.res}}{Resamples' predicted values}

\item{\code{predicted.res.aggr}}{Aggregated predicted values}

\item{\code{error.test.res}}{Resamples' testing error}

\item{\code{error.test.res.mean}}{Resamples' mean testing error}

\item{\code{error.test.res.aggr}}{Error of aggregated predicted values}

\item{\code{error.test.repeats}}{Mean testing error for each repeat}

\item{\code{error.test.repeats.mean}}{Mean testing error across all repeats}

\item{\code{error.test.repeats.sd}}{Standard deviation of testing error across all repeats}

\item{\code{fitted.bag}}{Bagged model's fitted values}

\item{\code{error.bag}}{Bagged model's error}

\item{\code{varimp}}{Resamples' variable importance}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{sessionInfo}}{R session info at time of training}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModCV$new()}}
\item \href{#method-print}{\code{rtModCV$print()}}
\item \href{#method-plot}{\code{rtModCV$plot()}}
\item \href{#method-plotPredicted}{\code{rtModCV$plotPredicted()}}
\item \href{#method-plotFitted}{\code{rtModCV$plotFitted()}}
\item \href{#method-plotVarImp}{\code{rtModCV$plotVarImp()}}
\item \href{#method-describe}{\code{rtModCV$describe()}}
\item \href{#method-clone}{\code{rtModCV$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{rtModCV} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$new(
  mod = NULL,
  mod.name = NULL,
  type = NULL,
  y.train = NULL,
  x.name = NULL,
  y.name = NULL,
  xnames = NULL,
  parameters = NULL,
  n.repeats = NULL,
  resampler.params = NULL,
  resamples = NULL,
  y.train.res = NULL,
  y.train.res.aggr = NULL,
  fitted.res = NULL,
  fitted.res.aggr = NULL,
  error.train.res = NULL,
  error.train.res.mean = NULL,
  error.train.res.aggr = NULL,
  error.train.repeats = NULL,
  error.train.repeats.mean = NULL,
  error.train.repeats.sd = NULL,
  y.test.res = NULL,
  y.test.res.aggr = NULL,
  predicted.res = NULL,
  predicted.res.aggr = NULL,
  error.test.res = NULL,
  error.test.res.mean = NULL,
  error.test.res.aggr = NULL,
  error.test.repeats = NULL,
  error.test.repeats.mean = NULL,
  error.test.repeats.sd = NULL,
  fitted.bag = NULL,
  error.bag = NULL,
  varimp = NULL,
  question = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{\code{rtModCV} object}

\item{\code{mod.name}}{Model (algorithm) name}

\item{\code{type}}{"Classification", "Regression", or "Survival"}

\item{\code{y.train}}{Training set y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{parameters}}{List of algorithm hyperparameters}

\item{\code{n.repeats}}{Integer: Number of repeats. This is the outermost iterator: i.e. You will run
\code{resampler} this many times.}

\item{\code{resampler.params}}{List of resampling parameters}

\item{\code{resamples}}{Resamples produced by \link{resample}}

\item{\code{y.train.res}}{Resamples' fitted values}

\item{\code{y.train.res.aggr}}{Aggregated training set outcomes}

\item{\code{fitted.res}}{Resamples' fitted values}

\item{\code{fitted.res.aggr}}{Aggregated fitted values}

\item{\code{error.train.res}}{Resamples' training error}

\item{\code{error.train.res.mean}}{Resamples' mean training error}

\item{\code{error.train.res.aggr}}{Error of aggregated fitted values}

\item{\code{error.train.repeats}}{Mean training error for each repeat}

\item{\code{error.train.repeats.mean}}{Mean training error across all repeats}

\item{\code{error.train.repeats.sd}}{Standard deviation of training error across all repeats}

\item{\code{y.test.res}}{Resamples' predicted values}

\item{\code{y.test.res.aggr}}{Aggregated testing set outcomes}

\item{\code{predicted.res}}{Resamples' predicted values}

\item{\code{predicted.res.aggr}}{Aggregated predicted values}

\item{\code{error.test.res}}{Resamples' testing error}

\item{\code{error.test.res.mean}}{Resamples' mean testing error}

\item{\code{error.test.res.aggr}}{Error of aggregated predicted values}

\item{\code{error.test.repeats}}{Mean testing error for each repeat}

\item{\code{error.test.repeats.mean}}{Mean testing error across all repeats}

\item{\code{error.test.repeats.sd}}{Standard deviation of testing error across all repeats}

\item{\code{fitted.bag}}{Bagged model's fitted values}

\item{\code{error.bag}}{Bagged model's error}

\item{\code{varimp}}{Resamples' variable importance}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{y.test}}{Testing set y data}

\item{\code{resampler}}{List of settings for \link{resample}.
Set using \link{rtset.cv.resample}}

\item{\code{sessionInfo}}{R session info at time of training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\code{print} method for \code{rtModCV} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\code{plot} method for \code{rtModCV} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plot(which.repeat = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: which repeat to plot}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPredicted}{}}}
\subsection{Method \code{plotPredicted()}}{
Plot predicted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plotPredicted(
  which.repeat = 1,
  theme = getOption("rt.theme"),
  filename = NULL,
  mar = c(2.5, 3, 2.5, 1),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: which repeat to plot}

\item{\code{theme}}{rtemis theme}

\item{\code{filename}}{Character: path to file to save plot}

\item{\code{mar}}{Numeric vector of plot margins}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotFitted}{}}}
\subsection{Method \code{plotFitted()}}{
Plot fitted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plotFitted(
  which.repeat = 1,
  theme = getOption("rt.theme"),
  filename = NULL,
  mar = c(2.5, 3, 2.5, 1),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: which repeat to plot}

\item{\code{theme}}{rtemis theme}

\item{\code{filename}}{Character: path to file to save plot}

\item{\code{mar}}{Numeric vector of plot margins}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotVarImp"></a>}}
\if{latex}{\out{\hypertarget{method-plotVarImp}{}}}
\subsection{Method \code{plotVarImp()}}{
Plot variable importance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$plotVarImp(
  which.repeat = 1,
  type = c("barplot", "lollipop"),
  plot.top = 12,
  theme = getOption("rt.theme"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: which repeat to plot}

\item{\code{type}}{Character: "barplot" or "lollipop"}

\item{\code{plot.top}}{Integer: Plotting this many top features}

\item{\code{theme}}{rtemis theme to use}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe"></a>}}
\if{latex}{\out{\hypertarget{method-describe}{}}}
\subsection{Method \code{describe()}}{
Describe \code{rtModCV}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
