% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtModCV-methods}
\alias{rtModCV-methods}
\alias{plot.rtModCV}
\alias{summary.rtModCV}
\alias{predict.rtModCV}
\title{S3 methods for \code{rtModCV} class that differ from those of the \code{rtMod} superclass}
\usage{
\method{plot}{rtModCV}(x, ...)

\method{summary}{rtModCV}(object, ...)

\method{predict}{rtModCV}(
  object,
  newdata,
  which.repeat = 1,
  bag.fn = mean,
  verbose = TRUE,
  n.cores = 1,
  ...
)
}
\arguments{
\item{x}{\code{rtModCV} object}

\item{...}{Additional arguments to pass to \code{mplot3.fit}}

\item{newdata}{Set of predictors to use}
}
\description{
S3 methods for \code{rtModCV} class that differ from those of the \code{rtMod} superclass

\code{plot.rtModCV}: \code{plot} method for \code{rtModCV} object

\code{summary.rtModCV}: \code{summary} method for \code{rtModCV} object

\code{predict.rtModCV}: \code{predict} method for \code{rtModCV} object
}
