% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\name{rtModCV-methods}
\alias{rtModCV-methods}
\alias{plot.rtModCV}
\alias{summary.rtModCV}
\alias{predict.rtModCV}
\title{S3 methods for \code{rtModCV} class that differ from those of the \code{rtMod} superclass}
\usage{
\method{plot}{rtModCV}(x, ...)

\method{summary}{rtModCV}(object, ...)

\method{predict}{rtModCV}(object, newdata, which.repeat = 1,
  bag.fn = mean, verbose = TRUE, n.cores = 1, ...)
}
\arguments{
\item{x}{\code{rtModCV} object}

\item{...}{Additional arguments to pass to \code{mplot3.fit}}

\item{newdata}{Set of predictors to use}
}
\description{
# special cases: GBM - include n.trees: can be supplied in '...'
  # BRUTO: newdata must be matrix
  if (verbose) msg("Calculating estimated values using", fn, "of", length(object$mod), "bag resamples")
  estimated.df <- sapply(object$mod, function(m) predict(m$mod1, newdata = newdata, ...))
  estimated <- apply(estimated.df, 1, fn)
  if (object$type == "Regression") {
    estimated <- as.numeric(estimated)
  } else {
    estimated <- levels(object$y.train)[estimated]
  }
  return(estimated)
}
