% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtMod.R
\docType{class}
\name{rtModCVClass-class}
\alias{rtModCVClass-class}
\alias{rtModCVClass}
\title{\pkg{rtemis} Cross-Validated Classification Model Class}
\description{
\pkg{rtemis} Cross-Validated Classification Model Class

\pkg{rtemis} Cross-Validated Classification Model Class
}
\details{
R6 Class for \pkg{rtemis} Cross-Validated Classification Models
}
\author{
E.D. Gennatas
}
\section{Super class}{
\code{\link[rtemis:rtModCV]{rtemis::rtModCV}} -> \code{rtModCVClass}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fitted.prob.aggr}}{Aggregated training set probability estimates}

\item{\code{predicted.prob.aggr}}{Aggregated testing set probability estimates}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModCVClass$new()}}
\item \href{#method-plotROC}{\code{rtModCVClass$plotROC()}}
\item \href{#method-plotROCfitted}{\code{rtModCVClass$plotROCfitted()}}
\item \href{#method-plotROCpredicted}{\code{rtModCVClass$plotROCpredicted()}}
\item \href{#method-plotPR}{\code{rtModCVClass$plotPR()}}
\item \href{#method-plotPRfitted}{\code{rtModCVClass$plotPRfitted()}}
\item \href{#method-plotPRpredicted}{\code{rtModCVClass$plotPRpredicted()}}
\item \href{#method-clone}{\code{rtModCVClass$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtModCV" data-id="describe">}\href{../../rtemis/html/rtModCV.html#method-describe}{\code{rtemis::rtModCV$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtModCV" data-id="plot">}\href{../../rtemis/html/rtModCV.html#method-plot}{\code{rtemis::rtModCV$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtModCV" data-id="plotFitted">}\href{../../rtemis/html/rtModCV.html#method-plotFitted}{\code{rtemis::rtModCV$plotFitted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtModCV" data-id="plotPredicted">}\href{../../rtemis/html/rtModCV.html#method-plotPredicted}{\code{rtemis::rtModCV$plotPredicted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtModCV" data-id="plotVarImp">}\href{../../rtemis/html/rtModCV.html#method-plotVarImp}{\code{rtemis::rtModCV$plotVarImp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rtemis" data-topic="rtModCV" data-id="print">}\href{../../rtemis/html/rtModCV.html#method-print}{\code{rtemis::rtModCV$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{rtModCVClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$new(
  mod = NULL,
  mod.name = NULL,
  type = NULL,
  y.train = NULL,
  x.name = NULL,
  y.name = NULL,
  xnames = NULL,
  parameters = NULL,
  n.repeats = NULL,
  resampler.params = NULL,
  resamples = NULL,
  y.train.res = NULL,
  y.train.res.aggr = NULL,
  fitted.res = NULL,
  fitted.res.aggr = NULL,
  fitted.prob.aggr = NULL,
  error.train.res = NULL,
  error.train.res.mean = NULL,
  error.train.res.aggr = NULL,
  error.train.repeats = NULL,
  error.train.repeats.mean = NULL,
  error.train.repeats.sd = NULL,
  y.test.res = NULL,
  y.test.res.aggr = NULL,
  predicted.res = NULL,
  predicted.res.aggr = NULL,
  predicted.prob.aggr = NULL,
  error.test.res = NULL,
  error.test.res.mean = NULL,
  error.test.res.aggr = NULL,
  error.test.repeats = NULL,
  error.test.repeats.mean = NULL,
  error.test.repeats.sd = NULL,
  fitted.bag = NULL,
  error.bag = NULL,
  varimp = NULL,
  question = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{\code{rtModCV} object}

\item{\code{mod.name}}{Model (algorithm) name}

\item{\code{type}}{"Classification", "Regression", or "Survival"}

\item{\code{y.train}}{Training set y data}

\item{\code{x.name}}{Name of x data}

\item{\code{y.name}}{Name of y data}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{parameters}}{List of algorithm hyperparameters}

\item{\code{n.repeats}}{Integer: Number of repeats. This is the outermost iterator: i.e. You will run
\code{resampler} this many times.}

\item{\code{resampler.params}}{List of resampling parameters}

\item{\code{resamples}}{Resamples produced by \link{resample}}

\item{\code{y.train.res}}{Resamples' fitted values}

\item{\code{y.train.res.aggr}}{Aggregated training set outcomes}

\item{\code{fitted.res}}{Resamples' fitted values}

\item{\code{fitted.res.aggr}}{Aggregated fitted values}

\item{\code{fitted.prob.aggr}}{Aggregated training set probability estimates}

\item{\code{error.train.res}}{Resamples' training error}

\item{\code{error.train.res.mean}}{Resamples' mean training error}

\item{\code{error.train.res.aggr}}{Error of aggregated fitted values}

\item{\code{error.train.repeats}}{Mean training error for each repeat}

\item{\code{error.train.repeats.mean}}{Mean training error across all repeats}

\item{\code{error.train.repeats.sd}}{Standard deviation of training error across all repeats}

\item{\code{y.test.res}}{Resamples' predicted values}

\item{\code{y.test.res.aggr}}{Aggregated testing set outcomes}

\item{\code{predicted.res}}{Resamples' predicted values}

\item{\code{predicted.res.aggr}}{Aggregated predicted values}

\item{\code{predicted.prob.aggr}}{Aggregated testing set probability estimates}

\item{\code{error.test.res}}{Resamples' testing error}

\item{\code{error.test.res.mean}}{Resamples' mean testing error}

\item{\code{error.test.res.aggr}}{Error of aggregated predicted values}

\item{\code{error.test.repeats}}{Mean testing error for each repeat}

\item{\code{error.test.repeats.mean}}{Mean testing error across all repeats}

\item{\code{error.test.repeats.sd}}{Standard deviation of testing error across all repeats}

\item{\code{fitted.bag}}{Bagged model's fitted values}

\item{\code{error.bag}}{Bagged model's error}

\item{\code{varimp}}{Resamples' variable importance}

\item{\code{question}}{Question the model is hoping to answer}

\item{\code{y.test}}{Testing set y data}

\item{\code{resampler}}{List of settings for \link{resample}.
Set using \link{rtset.cv.resample}}

\item{\code{sessionInfo}}{R session info at time of training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROC"></a>}}
\if{latex}{\out{\hypertarget{method-plotROC}{}}}
\subsection{Method \code{plotROC()}}{
Plot ROC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$plotROC(which.repeat = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: Which repeat to plot}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROCfitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotROCfitted}{}}}
\subsection{Method \code{plotROCfitted()}}{
Plot ROC plot for fitted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$plotROCfitted(which.repeat = 1, main = "ROC Training", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: Which repeat to plot}

\item{\code{main}}{Character: Main title}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotROCpredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotROCpredicted}{}}}
\subsection{Method \code{plotROCpredicted()}}{
Plot ROC plot for predicted values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$plotROCpredicted(which.repeat = 1, main = "ROC Testing", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: Which repeat to plot}

\item{\code{main}}{Character: Main title}

\item{\code{...}}{Additional arguments passed to plotting function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPR"></a>}}
\if{latex}{\out{\hypertarget{method-plotPR}{}}}
\subsection{Method \code{plotPR()}}{
Plot Precision-Recall curve
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$plotPR(which.repeat = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: Which repeat to use}

\item{\code{...}}{Additional arguments passed to
\code{plotPRpredicted} method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPRfitted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPRfitted}{}}}
\subsection{Method \code{plotPRfitted()}}{
Plot fitted Precision-Recall curve
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$plotPRfitted(which.repeat = 1, main = "P-R Training", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: Which repeat to use}

\item{\code{main}}{Character: main title}

\item{\code{...}}{Additional arguments passed to
\code{mplot3_pr}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPRpredicted"></a>}}
\if{latex}{\out{\hypertarget{method-plotPRpredicted}{}}}
\subsection{Method \code{plotPRpredicted()}}{
Plot predicted Precision-Recall curve
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$plotPRpredicted(which.repeat = 1, main = "P-R Testing", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which.repeat}}{Integer: Which repeat to use}

\item{\code{main}}{Character: main title}

\item{\code{...}}{Additional arguments passed to
\code{mplot3_pr}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModCVClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
