% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModLog.R
\docType{class}
\name{rtModLog-class}
\alias{rtModLog-class}
\alias{rtModLog}
\title{\pkg{rtemis} Supervised Model Log Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
rtModLog
}
\description{
\pkg{rtemis} Supervised Model Log Class
}
\section{Fields}{

\describe{
\item{\code{mod.name}}{Learner algorithm name}

\item{\code{parameters}}{List of hyperparameters used when building model}

\item{\code{error.train}}{Training error}

\item{\code{error.test}}{Testing error}

\item{\code{sessionInfo}}{The output of \code{sessionInfo()} at the time the model was trained}
}}

\author{
Efstathios D. Gennatas
}
\keyword{datasets}
