% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModLog.R
\name{rtModLogger}
\alias{rtModLogger}
\title{\pkg{rtemis} model logger}
\description{
R6 class to save trained models' parameters and performance. Keep your experiment results tidy in
one place, with an option to write out to a multi-sheet Excel file.
}
\author{
Efstathios D. Gennatas
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtModLogger$new()}}
\item \href{#method-print}{\code{rtModLogger$print()}}
\item \href{#method-add}{\code{rtModLogger$add()}}
\item \href{#method-summarize}{\code{rtModLogger$summarize()}}
\item \href{#method-summary}{\code{rtModLogger$summary()}}
\item \href{#method-tabulate}{\code{rtModLogger$tabulate()}}
\item \href{#method-plot}{\code{rtModLogger$plot()}}
\item \href{#method-clone}{\code{rtModLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$new(mods = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$add(mod, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\subsection{Method \code{summarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$summarize(
  class.metric = "Balanced Accuracy",
  reg.metric = "Rsq",
  surv.metric = "Coherence",
  decimal.places = 3,
  print.metric = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$summary(
  class.metric = "Balanced Accuracy",
  reg.metric = "Rsq",
  surv.metric = "Coherence"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate"></a>}}
\subsection{Method \code{tabulate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$tabulate(filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$plot(names = NULL, col = unlist(ucsfPalette), mar = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtModLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
