% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtPalette.R
\name{rtPalette}
\alias{rtPalette}
\title{\pkg{rtemis} Color Palettes}
\usage{
rtPalette(palette = NULL)
}
\arguments{
\item{palette}{Character: Name of palette to return. Default = NULL: available palette names
are printed and no palette is returned}
}
\value{
A list of available palettes, invisibly
}
\description{
\code{rtPalette} prints names of available color palettes
Each palette is a named list of hexadecimal color definitions which can be used with any
graphics function.
}
\examples{
rtPalette("imperial")
}
