% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rtpalette}
\alias{rtpalette}
\title{\pkg{rtemis} Color Palettes}
\usage{
rtpalette(palette = NULL, verbose = TRUE)
}
\arguments{
\item{palette}{Character: Name of palette to return. Default = NULL: available palette
names are printed and no palette is returned}

\item{verbose}{Logical: If \code{TRUE}, print messages to console}
}
\value{
A list of available palettes, invisibly
}
\description{
\code{rtpalette()} prints names of available color palettes
Each palette is a named list of hexadecimal color definitions which can be used with
any graphics function.
\code{rtpalette(palette_name)} returns a list of colors for a given palette.
}
\examples{
rtpalette("imperial")
}
