% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtPalette.R
\docType{data}
\name{rtPalettes}
\alias{rtPalettes}
\alias{pennCol}
\alias{pennPalette}
\alias{pennLightPalette}
\alias{imperialCol}
\alias{stanfordCol}
\alias{ucsfCol}
\alias{ucsfPalette}
\alias{ucdCol}
\alias{berkeleyCol}
\alias{ucscCol}
\alias{ucmercedCol}
\alias{ucsbCol}
\alias{uclaCol}
\alias{ucrColor}
\alias{uciCol}
\alias{ucsdCol}
\alias{californiaCol}
\alias{caltechCol}
\alias{cmuCol}
\alias{mitCol}
\alias{princetonCol}
\alias{columbiaCol}
\alias{brownCol}
\alias{yaleCol}
\alias{cornellCol}
\alias{hmsCol}
\alias{dartmouthCol}
\alias{usfCol}
\alias{uwCol}
\alias{jhuCol}
\alias{nyuCol}
\alias{chicagoCol}
\alias{pennstateCol}
\alias{sfsuCol}
\alias{illinoisCol}
\alias{umdCol}
\alias{texasCol}
\alias{techCol}
\alias{jeffersonCol}
\alias{hawaiiCol}
\alias{nihCol}
\alias{ubcCol}
\alias{torontoCol}
\alias{mcgillCol}
\alias{uclCol}
\alias{oxfordCol}
\alias{nhsCol}
\alias{ethCol}
\alias{rwthCol}
\alias{mozillaCol}
\alias{firefoxCol}
\alias{appleCol}
\alias{googleCol}
\alias{amazonCol}
\alias{microsoftCol}
\title{rtemis Color Palettes}
\format{
An object of class \code{list} of length 30.

An object of class \code{list} of length 11.

An object of class \code{list} of length 5.

An object of class \code{list} of length 26.

An object of class \code{list} of length 27.

An object of class \code{list} of length 13.

An object of class \code{list} of length 8.

An object of class \code{list} of length 18.

An object of class \code{list} of length 18.

An object of class \code{list} of length 9.

An object of class \code{list} of length 7.

An object of class \code{list} of length 10.

An object of class \code{list} of length 13.

An object of class \code{list} of length 3.

An object of class \code{list} of length 8.

An object of class \code{list} of length 11.

An object of class \code{list} of length 17.

An object of class \code{list} of length 19.

An object of class \code{list} of length 8.

An object of class \code{list} of length 3.

An object of class \code{list} of length 2.

An object of class \code{list} of length 19.

An object of class \code{list} of length 8.

An object of class \code{list} of length 10.

An object of class \code{list} of length 14.

An object of class \code{list} of length 14.

An object of class \code{list} of length 14.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.

An object of class \code{list} of length 21.

An object of class \code{list} of length 17.

An object of class \code{list} of length 25.

An object of class \code{list} of length 20.

An object of class \code{list} of length 9.

An object of class \code{list} of length 15.

An object of class \code{list} of length 3.

An object of class \code{list} of length 10.

An object of class \code{list} of length 3.

An object of class \code{list} of length 8.

An object of class \code{list} of length 11.

An object of class \code{list} of length 2.

An object of class \code{list} of length 6.

An object of class \code{list} of length 2.

An object of class \code{list} of length 27.

An object of class \code{list} of length 23.

An object of class \code{list} of length 24.

An object of class \code{list} of length 21.

An object of class \code{list} of length 10.

An object of class \code{list} of length 60.

An object of class \code{list} of length 8.

An object of class \code{list} of length 8.

An object of class \code{list} of length 8.

An object of class \code{list} of length 4.

An object of class \code{list} of length 2.

An object of class \code{list} of length 4.
}
\usage{
pennCol

pennPalette

pennLightPalette

imperialCol

stanfordCol

ucsfCol

ucsfPalette

ucdCol

berkeleyCol

ucscCol

ucmercedCol

ucsbCol

uclaCol

ucrColor

uciCol

ucsdCol

californiaCol

caltechCol

cmuCol

mitCol

princetonCol

columbiaCol

brownCol

yaleCol

cornellCol

hmsCol

dartmouthCol

usfCol

uwCol

jhuCol

nyuCol

chicagoCol

pennstateCol

sfsuCol

illinoisCol

umdCol

texasCol

techCol

jeffersonCol

hawaiiCol

nihCol

ubcCol

torontoCol

mcgillCol

uclCol

oxfordCol

nhsCol

ethCol

rwthCol

mozillaCol

firefoxCol

appleCol

googleCol

amazonCol

microsoftCol
}
\description{
\code{pennCol}: Penn color palette
(http://www.upenn.edu/about/styleguide-color-type)

\code{imperialCol}: Imperial College London colour palette
(https://www.imperial.ac.uk/brand-style-guide/visual-identity/brand-colours/)

\code{stanfordCol}: Stanford color palette
(https://identity.stanford.edu/color.html#digital-color)

\code{ucsfCol}: UCSF color palette (http://identity.ucsf.edu/color)

\code{ucsfPalette}: Subset of \code{ucsfCol} for use with \link{mplot3}, etc

\code{ucdCol}: UC Davis color palette
(https://marketingtoolbox.ucdavis.edu/visual-identity/color.html)

\code{berkeleyCol}: Berkeley color palette
(https://brand.berkeley.edu/colors/)

\code{ucscCol}: UC Santa Cruz color palette
(https://communications.ucsc.edu/visual-design/color/)

\code{ucmercedCol}: UC Merced color palette
(https://publicrelations.ucmerced.edu/color-guidelines)

\code{ucsbCol}: UC Santa Barbara color palette
(https://www.ucsb.edu/visual-identity/color)

\code{uclaCol}: UCLA color palette (http://brand.ucla.edu/identity/colors)

\code{ucrCol}: UC Riverside color palette (https://brand.ucr.edu/ucr-colors)

\code{uciCol}: UCI color palette (https://communications.uci.edu/campus-resources/graphic-standards/colors.php)

\code{ucsdCol}: UC San Diego color palette
(https://ucpa.ucsd.edu/brand/elements/color-palette/)

\code{californiaCol}: University of California color palette
(http://brand.universityofcalifornia.edu/guidelines/color.html#!primary-colors)

\code{caltechCol}: Caltech color palette (http://identity.caltech.edu/colors)

\code{cmuCol}: Carnegie Mellon color palette
(https://www.cmu.edu/marcom/brand-standards/web-standards.html#colors)

\code{mitCol}: MIT color palette
(http://web.mit.edu/graphicidentity/colors.html)

\code{princetonCol}: Princeton color palette
(https://communications.princeton.edu/guides-tools/logo-graphic-identity)

\code{columbiaCol}: Columbia color palette
(https://visualidentity.columbia.edu/content/web-0)

\code{brownCol}: Brown color palette
(https://www.brown.edu/university-identity/sites/university-identity/files/Brown_Visual_Identity_Policy_2016-07-22.pdf)

\code{yaleCol}: Yale color palette (https://yaleidentity.yale.edu/web)

\code{cornellCol}: Yale color palette
(https://brand.cornell.edu/design-center/colors/

\code{hmsCol}: Harvard Medical School color palette
(https://identityguide.hms.harvard.edu/color)

\code{dartmouthCol}: Dartmouth color palette
(https://communications.dartmouth.edu/visual-identity/design-elements/color-palette#web%20palette)

\code{usfCol}: USF color palette
(https://myusf.usfca.edu/marketing-communications/resources/graphics-resources/brand-standards/color-palette)
Color conversions performed using https://www.pantone.com/color-finder/

\code{uwCol}: University of Washington color palette
(http://www.washington.edu/brand/graphic-elements/primary-color-palette/)

\code{jhuCol}: Johns Hopkins University color palette
(https://brand.jhu.edu/color/)

\code{nyuCol}: NYU color palette
(https://www.nyu.edu/employees/resources-and-services/media-and-communications/styleguide/website/graphic-visual-design.html)

\code{chicagoCol}: University of Chicago color palette
(https://news.uchicago.edu/sites/default/files/attachments/_uchicago.identity.guidelines.pdf)

\code{texasCol}: Penn State color palette
(https://brand.psu.edu/design-essentials.html#color)

\code{sfsuCol}: SF State Colors

\code{illinoisCol}: University of Illinois color palette
(https://www.uillinois.edu/OUR/brand/color_palettes)

\code{umdCol}: University of Maryland color palette
(https://osc.umd.edu/licensing-trademarks/brand-standards/logos/#color)

\code{texasCol}: University of Texas color palette
(https://brand.utexas.edu/identity/color/)

\code{techCol}: Georgia Tech color palette
(http://www.licensing.gatech.edu/super-block/239)

\code{jeffersonCol}: Jefferson color palette (http://creative.jefferson.edu/downloads/Jefferson-Brand-Guidelines.pdf)

\code{hawaiiCol}: University of Hawaii color palette (https://www.hawaii.edu/offices/eaur/graphicsstandards.pdf)

\code{nihCol}: NIH color palette (https://www.nlm.nih.gov/about/nlm_logo_guidelines_030414_508.pdf)

\code{ubcCol}: UBC color palette (http://assets.brand.ubc.ca/downloads/ubc_colour_guide.pdf)

\code{torontoCol}: U Toronto color palette (https://trademarks.utoronto.ca/colors-fonts/)

\code{mcgillCol}: McGill color palette (https://www.mcgill.ca/visual-identity/visual-identity-guide)

\code{uclCol}: UCL colour palette (https://www.ucl.ac.uk/cam/brand/guidelines/colour)

\code{oxfordCol}: Oxford University colour palette (https://www.ox.ac.uk/sites/files/oxford/media_wysiwyg/Oxford%20Blue%20LR.pdf)

\code{nhsCol}: NHS colour palette (https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/)

\code{ethCol}: ETH color palette (https://ethz.ch/services/en/service/communication/corporate-design/colour.html)

\code{rwthCol}: RWTH Aachen color palette (http://www9.rwth-aachen.de/global/show_document.asp?id=aaaaaaaaaadpbhq)

\code{mozillaCol}: Mozilla design colors
(https://mozilla.design/mozilla/color/)

\code{firefoxCol}: Firefox design colors
(https://mozilla.design/firefox/color/)

\code{appleCol}: Apple Human Interface Guidelines color palette
(https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/color/)

\code{googleCol}: Google brand palette (https://brandpalettes.com/google-colors/)

\code{amazonCol}: Amazon brand palette
(https://images-na.ssl-images-amazon.com/images/G/01/AdvertisingSite/pdfs/AmazonBrandUsageGuidelines.pdf)

\code{microsoftCol}: Microsoft brand palette
(https://brandcolors.net/b/microsoft)
}
\details{
\code{sfsuCol}: SF State color palette 
(https://logo.sfsu.edu/color-system)
}
\keyword{datasets}
