% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtRandom.R
\name{rtrandom}
\alias{rtrandom}
\title{Random \pkg{rtemis} art}
\usage{
rtrandom(
  pch = sample(15:18, 1),
  col = rtCol,
  text = "rtemis",
  text.col = "gray50",
  text.as.legend = FALSE,
  legend.bg = NULL,
  legend.alpha = 1,
  random.bg = TRUE
)
}
\arguments{
\item{pch}{Point character to use}

\item{col}{Colors to use}

\item{text}{Text to print}

\item{text.col}{Color for text}
}
\description{
Draw random shapes and colors
}
\author{
E.D. Gennatas
}
