% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.decompose}
\alias{rtset.decompose}
\title{\code{rtset.decompose}: Set decomposition parameters for \code{elevate}'s \code{.decompose} argument}
\usage{
rtset.decompose(decom = "ICA", k = 2, ...)
}
\arguments{
\item{decom}{String: Name of decomposer to use. Default = "ICA"}

\item{k}{Integer: Number of dimensions to project to. Default = 2}
}
\description{
\code{rtset.decompose}: Set decomposition parameters for \code{elevate}'s \code{.decompose} argument
}
