% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.GBM}
\alias{rtset.GBM}
\title{\code{rtset.GBM}: Set parameters for \link{s_GBM}}
\usage{
rtset.GBM(
  interaction.depth = 2,
  shrinkage = 0.001,
  max.trees = 5000,
  min.trees = 100,
  bag.fraction = 0.9,
  n.minobsinnode = 5,
  grid.resample.rtset = rtset.resample("kfold", 5),
  ipw = TRUE,
  upsample = FALSE,
  downsample = FALSE,
  resample.seed = NULL,
  ...
)
}
\arguments{
\item{interaction.depth}{[gS] Integer: Interaction depth. Default = 2}

\item{shrinkage}{[gS] Float: Shrinkage (learning rate). Default = .01}

\item{bag.fraction}{[gS] Float (0, 1): Fraction of cases to use to train each tree.
Helps avoid overfitting. Default = .75}

\item{n.minobsinnode}{[gS] Integer: Minimum number of observation allowed in node. Default = 5}

\item{ipw}{Logical: If TRUE, apply inverse probability weighting (for Classification only).
Note: If \code{weights} are provided, \code{ipw} is not used. Default = TRUE}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Note: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{downsample}{Logical: If TRUE, downsample majority class to match size of minority class}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}
}
\description{
\code{rtset.GBM}: Set parameters for \link{s_GBM}
}
