% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.HYTREE}
\alias{rtset.HYTREE}
\title{\code{rtset.ADDT}: Set parameters for \link{s.HYTREE}}
\usage{
rtset.HYTREE(
  max.depth = 2,
  learning.rate = 1,
  lincoef.params = rtset.lincoef("glmnet"),
  alpha = 0,
  lambda = 0.1,
  minobsinnode = 2,
  minobsinnode.lin = 20,
  ...
)
}
\description{
\code{rtset.ADDT}: Set parameters for \link{s.HYTREE}
}
