% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.preprocess}
\alias{rtset.preprocess}
\title{\code{rtset.preprocess}: Set \link{preprocess} parameters for \link{train}'s \code{.preprocess} argument}
\usage{
rtset.preprocess(
  completeCases = FALSE,
  removeCases.thres = NULL,
  removeFeatures.thres = NULL,
  impute = FALSE,
  impute.type = "missRanger",
  impute.missRanger.params = list(pmm.k = 0, maxiter = 10),
  impute.discrete = get_mode,
  impute.numeric = mean,
  integer2factor = FALSE,
  integer2numeric = FALSE,
  logical2factor = FALSE,
  logical2numeric = FALSE,
  numeric2factor = FALSE,
  numeric2factor.levels = NULL,
  numeric.cut.n = 0,
  numeric.cut.labels = FALSE,
  numeric.quant.n = 0,
  character2factor = FALSE,
  scale = FALSE,
  center = FALSE,
  removeConstants = TRUE,
  oneHot = FALSE,
  exclude = NULL
)
}
\arguments{
\item{decom}{Character: Name of decomposer to use. Default = "ICA"}

\item{k}{Integer: Number of dimensions to project to. Default = 2}
}
\description{
\code{rtset.preprocess}: Set \link{preprocess} parameters for \link{train}'s \code{.preprocess} argument
}
