% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.preprocess}
\alias{rtset.preprocess}
\title{\code{rtset.preprocess}: Set \link{preprocess} parameters for \link{elevate}'s \code{.preprocess} argument}
\usage{
rtset.preprocess(completeCases = FALSE, removeCases.thres = NULL,
  removeFeatures.thres = NULL, impute = FALSE,
  impute.type = "missRanger", impute.missRanger.params = list(pmm.k =
  0, maxiter = 10), impute.missForest.params = list(maxiter = 10, ntree =
  500, parallelize = "no"), impute.rfImpute.params = list(niter = 10,
  ntree = 500), impute.discrete = getMode, impute.numeric = mean,
  integer2factor = FALSE, integer2numeric = FALSE,
  logical2factor = FALSE, logical2numeric = FALSE,
  numeric2factor = FALSE, numeric2factor.levels = NULL,
  character2factor = FALSE, nonzeroFactors = FALSE, scale = FALSE,
  center = FALSE, removeConstants = TRUE, oneHot = FALSE,
  exclude = NULL)
}
\arguments{
\item{decom}{Character: Name of decomposer to use. Default = "ICA"}

\item{k}{Integer: Number of dimensions to project to. Default = 2}
}
\description{
\code{rtset.preprocess}: Set \link{preprocess} parameters for \link{elevate}'s \code{.preprocess} argument
}
