% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtset.R
\name{rtset.RANGER}
\alias{rtset.RANGER}
\title{\code{rtset.RANGER}: Set parameters for \link{s.RANGER}}
\usage{
rtset.RANGER(n.trees = 1000, min.node.size = 1, mtry = NULL,
  grid.resample.rtset = rtset.resample("kfold", 5), ipw = TRUE,
  upsample = FALSE, upsample.seed = NULL, ...)
}
\arguments{
\item{n.trees}{Integer: Number of trees to grow. Default = 1000}

\item{min.node.size}{[gS] Integer: Minimum node size}

\item{mtry}{[gS] Integer: Number of features sampled randomly at each split. Defaults to square root of n of
features for classification, and a third of n of features for regression.}

\item{grid.resample.rtset}{List: Output of \link{rtset.resample} defining \link{gridSearchLearn} parameters.
Default = \code{rtset.resample("kfold", 5)}}

\item{ipw}{Logical: If TRUE, apply inverse probability weighting (for Classification only).
Note: If \code{weights} are provided, \code{ipw} is not used. Default = TRUE}

\item{upsample}{Logical: If TRUE, upsample training set cases not belonging in majority outcome
group}

\item{upsample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{...}{Additional arguments to be passed to \code{ranger::ranger}}
}
\description{
\code{rtset.RANGER}: Set parameters for \link{s.RANGER}
}
