% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{rtset}
\alias{rtset}
\title{\pkg{rtemis} default-setting functions}
\description{
These functions output lists of default settings for different \pkg{rtemis} functions.
This removes the need of passing named lists of arguments, and provides autocompletion,
making it easier to setup functions without having to refer to the manual.
}
\author{
E.D. Gennatas
}
