% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtXDecom.R
\docType{class}
\name{rtXDecom-class}
\alias{rtXDecom-class}
\alias{rtXDecom}
\title{R6 class for \pkg{rtemis} cross-decompositions}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
rtXDecom
}
\description{
\pkg{rtemis} cross-decomposition R6 object
}
\section{Fields}{

\describe{
\item{\code{xdecom.name}}{Character: Name of cross-decomposition algorithm}

\item{\code{k}}{Integer: Number of projections}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{xdecom}}{Cross-decomposition model output}

\item{\code{projections.train}}{Input data projected on new axes / basis}

\item{\code{projections.test}}{Input test data projected on new axes / basis}

\item{\code{extra}}{List: Algorithm-specific output}
}}

\author{
Efstathios D Gennatas
}
\keyword{datasets}
