% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtXDecom.R
\docType{class}
\name{rtXDecom-class}
\alias{rtXDecom-class}
\alias{rtXDecom}
\title{R6 class for \pkg{rtemis} cross-decompositions}
\description{
R6 class for \pkg{rtemis} cross-decompositions

R6 class for \pkg{rtemis} cross-decompositions
}
\details{
\pkg{rtemis} cross-decomposition R6 object
}
\author{
E.D. Gennatas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{xdecom.name}}{Character: Name of cross-decomposition algorithm}

\item{\code{k}}{Integer: Number of projections}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{znames}}{Character vector: Column names of z}

\item{\code{xdecom}}{Cross-decomposition model output}

\item{\code{xprojections.train}}{x data training set projections}

\item{\code{xprojections.test}}{x data test set data projections}

\item{\code{zprojections.train}}{z data training set projections}

\item{\code{zprojections.test}}{z data test set projections}

\item{\code{parameters}}{Cross-decomposition parameters}

\item{\code{extra}}{List: Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rtXDecom$new()}}
\item \href{#method-print}{\code{rtXDecom$print()}}
\item \href{#method-clone}{\code{rtXDecom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtXDecom$new(
  xdecom.name = character(),
  k = integer(),
  xnames = character(),
  znames = character(),
  xdecom = list(),
  xprojections.train = numeric(),
  xprojections.test = numeric(),
  zprojections.train = numeric(),
  zprojections.test = numeric(),
  parameters = list(),
  extra = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdecom.name}}{Character: Name of cross-decomposition algorithm}

\item{\code{k}}{Integer: Number of projections}

\item{\code{xnames}}{Character vector: Column names of x}

\item{\code{znames}}{Character vector: Column names of z}

\item{\code{xdecom}}{Cross-decomposition model output}

\item{\code{xprojections.train}}{x data training set projections}

\item{\code{xprojections.test}}{x data test set data projections}

\item{\code{zprojections.train}}{z data training set projections}

\item{\code{zprojections.test}}{z data test set projections}

\item{\code{parameters}}{Cross-decomposition parameters}

\item{\code{extra}}{List: Algorithm-specific output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{rtXDecom} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtXDecom$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rtXDecom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
