% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rules.R
\name{rules2medmod}
\alias{rules2medmod}
\title{Convert rules from cutoffs to median/mode and range}
\usage{
rules2medmod(rules, x, .ddSci = TRUE, verbosity = 1L)
}
\arguments{
\item{rules}{Character, vector: Input rules}

\item{x}{Data frame: Data to evaluate rules}

\item{.ddSci}{Logical: If TRUE, format all continuous variables using
\link{ddSci}, which will give either 2 decimal places, or scientific
notation if two decimal places result in 0.00}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
Character vector.
}
\description{
Convert rules from cutoffs to \code{median (range)} and \code{mode (range)} format
}
\author{
EDG
}
