% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules2medmod.R
\name{rules2medmod}
\alias{rules2medmod}
\title{\code{rtemis-internals}: Convert rules from cutoffs to median/mode and range}
\usage{
rules2medmod(rules, x, .ddSci = TRUE, verbose = TRUE, trace = 0)
}
\arguments{
\item{rules}{Character, vector: Input rules}

\item{x}{Data frame: Data to evaluate rules}

\item{.ddSci}{Logical: If TRUE, format all continuous variables using \link{ddSci}, which will
give either 2 decimal places, or scientific notation if two decimal places result in 0.00}

\item{verbose}{Logical: If TRUE, print messages to console. Default = TRUE}
}
\description{
Convert rules from cutoffs to \code{median (range)} and \code{mode (range)} format
}
\author{
E.D. Gennatas
}
