% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ops.R
\name{runifmat}
\alias{runifmat}
\title{Random Uniform Matrix}
\usage{
runifmat(
  nrow = 10,
  ncol = 10,
  min = 0,
  max = 1,
  return.df = FALSE,
  seed = NULL
)
}
\arguments{
\item{nrow}{Integer: Number of rows.}

\item{ncol}{Integer: Number of columns.}

\item{min}{Float: Min.}

\item{max}{Float: Max.}

\item{return.df}{Logical: If TRUE, return data.frame, otherwise matrix.}

\item{seed}{Integer: Set seed for \code{rnorm}.}
}
\description{
Create a matrix or data frame of defined dimensions, whose columns are random uniform vectors
}
\author{
E.D. Gennatas
}
