% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LIGHTGBM.R
\name{s_LIGHTRF}
\alias{s_LIGHTRF}
\title{Random Forest with LightGBM}
\usage{
s_LIGHTRF(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  x.name = NULL,
  y.name = NULL,
  weights = NULL,
  ipw = TRUE,
  ipw.type = 2,
  upsample = FALSE,
  downsample = FALSE,
  resample.seed = NULL,
  boosting = "rf",
  objective = NULL,
  max_nrounds = 500L,
  force_nrounds = NULL,
  early_stopping_rounds = 10L,
  nrounds_default = 100L,
  num_leaves = 31L,
  max_depth = -1L,
  learning_rate = 0.01,
  bagging_fraction = 0.623,
  bagging_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  linear_tree = FALSE,
  tree_learner = "data_parallel",
  .gs = FALSE,
  grid.resample.rtset = rtset.resample("kfold", 5),
  grid.search.type = "exhaustive",
  metric = NULL,
  maximize = NULL,
  importance = TRUE,
  print.plot = FALSE,
  plot.fitted = NULL,
  plot.predicted = NULL,
  plot.theme = rtTheme,
  question = NULL,
  rtclass = NULL,
  save.dump = FALSE,
  verbose = TRUE,
  grid.verbose = FALSE,
  lightgbm_verbose = -1,
  trace = 0,
  save.gridrun = FALSE,
  n.cores = 1,
  n_threads = rtCores,
  force_col_wise = FALSE,
  force_row_wise = FALSE,
  parallel.type = c("psock", "fork"),
  outdir = NULL,
  save.mod = ifelse(!is.null(outdir), TRUE, FALSE),
  ...
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ipw}, therefore set \code{weights = NULL} if using \code{ipw}.
Note: If \code{weight} are provided, \code{ipw} is not used. Leave NULL if setting \code{ipw = TRUE}. Default = NULL}

\item{ipw}{Logical: If TRUE, apply inverse probability weighting (for Classification only).
Note: If \code{weights} are provided, \code{ipw} is not used. Default = TRUE}

\item{ipw.type}{Integer {0, 1, 2}
1: class.weights as in 0, divided by max(class.weights)
2: class.weights as in 0, divided by min(class.weights)
Default = 2}

\item{upsample}{Logical: If TRUE, upsample cases to balance outcome classes (for Classification only)
Note: upsample will randomly sample with replacement if the length of the majority class is more than double
the length of the class you are upsampling, thereby introducing randomness}

\item{downsample}{Logical: If TRUE, downsample majority class to match size of minority class}

\item{resample.seed}{Integer: If provided, will be used to set the seed during upsampling.
Default = NULL (random seed)}

\item{boosting}{Character: [gS] "gbdt", "rf", "dart", "goss"}

\item{objective}{(Default = NULL)}

\item{max_nrounds}{Integer: Maximum number of rounds to run. Can be set to a high number
as early stopping will limit nrounds by monitoring inner CV error}

\item{force_nrounds}{Integer: Number of rounds to run if not estimating optimal number by CV}

\item{early_stopping_rounds}{Integer: Training on resamples of \code{x} (tuning) will
stop if performance does not improve for this many rounds}

\item{nrounds_default}{Integer: Default number of rounds to run if 
cross-validation fails - likely will never be used}

\item{num_leaves}{Integer: [gS] Maximum tree leaves for base learners.}

\item{max_depth}{Integer: [gS] Maximum tree depth for base learners, <=0 means no limit.}

\item{learning_rate}{Numeric: [gS] Boosting learning rate}

\item{bagging_fraction}{Numeric: [gS] Subsample ratio of the training instance}

\item{lambda_l1}{Numeric: [gS] L1 regularization term}

\item{lambda_l2}{Numeric: [gS] L2 regularization term}

\item{linear_tree}{Logical: [gS] If \code{TRUE}, use linear trees}

\item{tree_learner}{Character: [gS] "serial", "feature", "data", "voting"}

\item{grid.resample.rtset}{List: Output of \link{rtset.resample} defining 
\link{gridSearchLearn} parameters.}

\item{grid.search.type}{Character: Type of grid search to perform: 
"exhaustive" or "randomized".}

\item{metric}{Character: Metric to minimize, or maximize if 
\code{maximize = TRUE} during grid search. Default = NULL, which results in 
"Balanced Accuracy" for Classification,
"MSE" for Regression, and "Coherence" for Survival Analysis.}

\item{maximize}{Logical: If TRUE, \code{metric} will be maximized if grid 
search is run.}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}. Default = TRUE}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{grid.verbose}{Logical: Passed to \link{gridSearchLearn}}

\item{n.cores}{Integer: Number of cores to use.}

\item{n_threads}{Integer: Number of threads for lightgbm using OpenMP. Only parallelize
resamples using \code{n.cores} or the lightgbm execution using this setting.}

\item{outdir}{Path to output directory.
If defined, will save Predicted vs. True plot, if available,
as well as full model output, if \code{save.mod} is TRUE}

\item{save.mod}{Logical: If TRUE, save all output to an RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}
}
\description{
Random Forest with LightGBM
}
\examples{
\dontrun{
x <- rnormmat(500, 10)
y <- x[, 3] + .5 * x[, 5]^2 + rnorm(500)
dat <- data.frame(x, y)
mod <- s_LIGHTRF(dat)
}
}
\author{
ED Gennatas
}
