% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LightRuleFit.R
\name{s_LightRuleFit}
\alias{s_LightRuleFit}
\title{RuleFit with LightGBM (C, R)}
\usage{
s_LightRuleFit(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  params = setup.LightRuleFit(),
  lgbm.mod = NULL,
  empirical_risk = TRUE,
  cases_by_rules = NULL,
  save_cases_by_rules = FALSE,
  x.name = NULL,
  y.name = NULL,
  n.cores = rtCores,
  question = NULL,
  print.plot = FALSE,
  plot.fitted = NULL,
  plot.predicted = NULL,
  plot.theme = rtTheme,
  outdir = NULL,
  save.mod = if (!is.null(outdir)) TRUE else FALSE,
  verbose = TRUE,
  trace = 0
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{params}{Training parameters for GBM and LASSO steps, set using
\link{setup.LightRuleFit}.}

\item{lgbm.mod}{rtMod object created by \link{s_LightGBM}. If provided, the gradient
boosting step is skipped.}

\item{empirical_risk}{Logical: If TRUE, calculate empirical risk}

\item{cases_by_rules}{Matrix of cases by rules from a previoue rulefit run.
If provided, the GBM step is skipped.}

\item{save_cases_by_rules}{Logical: If TRUE, save cases_by_rules to object}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{n.cores}{Integer: Number of cores to use}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}.}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{outdir}{Path to output directory.
If defined, will save Predicted vs. True plot, if available,
as well as full model output, if \code{save.mod} is TRUE}

\item{save.mod}{Logical: If TRUE, save all output to an RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: Verbosity level}

\item{n.trees}{Integer: Number of trees to train. Passed to \link{s_LightGBM}
\code{force.n.trees}. If set to NULL, can set \code{max_nrounds} in \code{lgbm.params}, to perform
cross-validation to determine optimal number of trees.}
}
\value{
\code{rtMod} object
}
\description{
Train a LightGBM gradient boosting model, extract rules,
and fit using LASSO
}
\details{
Based on "Predictive Learning via Rule Ensembles"
by Friedman and Popescu
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\references{
Friedman JH, Popescu BE, "Predictive Learning via Rule Ensembles",
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\author{
E.D. Gennatas
}
