% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_LightRuleFit.R
\name{s_LightRuleFit}
\alias{s_LightRuleFit}
\title{RuleFit with LightGBM (C, R)}
\usage{
s_LightRuleFit(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  lgbm.mod = NULL,
  n_trees = 200,
  num_leaves = 32L,
  max_depth = 4,
  learning_rate = 0.1,
  subsample = 0.666,
  subsample_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  objective = NULL,
  importance = FALSE,
  lgbm.ifw = TRUE,
  lgbm.grid.resample.params = setup.resample(resampler = "kfold", n.resamples = 5),
  glmnet.ifw = TRUE,
  alpha = 1,
  lambda = NULL,
  glmnet.grid.resample.params = setup.resample(resampler = "kfold", n.resamples = 5),
  grid.resample.params = setup.resample("kfold", 5),
  gridsearch.type = "exhaustive",
  metric = NULL,
  maximize = NULL,
  grid.verbose = FALSE,
  save.gridrun = FALSE,
  weights = NULL,
  empirical_risk = TRUE,
  cases_by_rules = NULL,
  save_cases_by_rules = FALSE,
  x.name = NULL,
  y.name = NULL,
  n.cores = rtCores,
  question = NULL,
  print.plot = FALSE,
  plot.fitted = NULL,
  plot.predicted = NULL,
  plot.theme = rtTheme,
  outdir = NULL,
  save.mod = if (!is.null(outdir)) TRUE else FALSE,
  verbose = TRUE,
  trace = 0,
  .gs = FALSE
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{lgbm.mod}{rtMod object created by \link{s_LightGBM}. If provided, the gradient
boosting step is skipped.}

\item{num_leaves}{Integer: [gS] Maximum tree leaves for base learners.}

\item{max_depth}{Integer: [gS] Maximum tree depth for base learners, <=0 means no limit.}

\item{learning_rate}{Numeric: [gS] Boosting learning rate}

\item{subsample}{Numeric: [gS] Subsample ratio of the training set.}

\item{subsample_freq}{Integer: Subsample every this many iterations}

\item{lambda_l1}{Numeric: [gS] L1 regularization term}

\item{lambda_l2}{Numeric: [gS] L2 regularization term}

\item{objective}{(Default = NULL)}

\item{importance}{Logical: If \code{TRUE}, calculate variable importance}

\item{alpha}{[gS] Float [0, 1]: The elasticnet mixing parameter:
\code{a = 0} is the ridge penalty, \code{a = 1} is the lasso penalty}

\item{lambda}{[gS] Float vector: Best left to NULL, \code{cv.glmnet} will
compute its own lambda sequence}

\item{grid.resample.params}{List: Output of \link{setup.resample} defining
grid search parameters.}

\item{gridsearch.type}{Character: Type of grid search to perform:
"exhaustive" or "randomized".}

\item{metric}{Character: Metric to minimize, or maximize if
\code{maximize = TRUE} during grid search. Default = NULL, which results in
"Balanced Accuracy" for Classification,
"MSE" for Regression, and "Coherence" for Survival Analysis.}

\item{maximize}{Logical: If TRUE, \code{metric} will be maximized if grid
search is run.}

\item{grid.verbose}{Logical: Passed to \code{gridSearchLearn}}

\item{save.gridrun}{Logical: If \code{TRUE}, save all grid search models}

\item{weights}{Numeric vector: Weights for cases. For classification, \code{weights} takes precedence
over \code{ifw}, therefore set \code{weights = NULL} if using \code{ifw}.
Note: If \code{weight} are provided, \code{ifw} is not used. Leave NULL if setting \code{ifw = TRUE}.}

\item{empirical_risk}{Logical: If TRUE, calculate empirical risk}

\item{cases_by_rules}{Matrix of cases by rules from a previoue rulefit run.
If provided, the GBM step is skipped.}

\item{save_cases_by_rules}{Logical: If TRUE, save cases_by_rules to object}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{n.cores}{Integer: Number of cores to use}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}.}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{outdir}{Path to output directory.
If defined, will save Predicted vs. True plot, if available,
as well as full model output, if \code{save.mod} is TRUE}

\item{save.mod}{Logical: If TRUE, save all output to an RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{verbose}{Logical: If TRUE, print summary to screen.}

\item{trace}{Integer: Verbosity level}

\item{.gs}{(Internal use only)}
}
\value{
\code{rtMod} object
}
\description{
Train a LightGBM gradient boosting model, extract rules,
and fit using LASSO
}
\details{
Based on "Predictive Learning via Rule Ensembles"
by Friedman and Popescu
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\references{
Friedman JH, Popescu BE, "Predictive Learning via Rule Ensembles",
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\author{
E.D. Gennatas
}
