% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_RuleFit.R
\name{s_RuleFit}
\alias{s_RuleFit}
\title{Rulefit [C, R]}
\usage{
s_RuleFit(
  x,
  y = NULL,
  x.test = NULL,
  y.test = NULL,
  gbm.params = list(list(n.trees = 300, bag.fraction = 1, shrinkage = 0.1,
    interaction.depth = 3, ifw = TRUE)),
  meta.alpha = 1,
  meta.lambda = NULL,
  meta.extra.params = list(ifw = TRUE),
  cases.by.rules = NULL,
  x.name = NULL,
  y.name = NULL,
  n.cores = rtCores,
  question = NULL,
  print.plot = FALSE,
  plot.fitted = NULL,
  plot.predicted = NULL,
  plot.theme = rtTheme,
  outdir = NULL,
  save.mod = if (!is.null(outdir)) TRUE else FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{Numeric vector or matrix / data frame of features i.e. independent variables}

\item{y}{Numeric vector of outcome, i.e. dependent variable}

\item{x.test}{Numeric vector or matrix / data frame of testing set features
Columns must correspond to columns in \code{x}}

\item{y.test}{Numeric vector of testing set outcome}

\item{gbm.params}{List of named lists: A list, each element of which is a named list
of parameters for \link{s_GBM}. i.e. If you want to train a single GBM model, this could
be:
\code{gbm.params = list( list( n.trees = 300, bag.fraction = 1, shrinkage = .1, interaction.depth = 3, ifw = TRUE ) )}
if you wanted to train 2 GBM models, this could be:
\code{gbm.params = list( list( n.trees = 300, bag.fraction = 1, shrinkage = .1, interaction.depth = 3, ifw = TRUE ), list( n.trees = 500, bag.fraction = 1, shrinkage = .1, interaction.depth = 3, ifw = TRUE ) )}}

\item{meta.alpha}{Float [0, 1]: \code{alpha} for \link{s_GLMNET}}

\item{meta.lambda}{Float: \code{lambda} for \link{s_GLMNET}. Default = NULL (will be
determined automatically by crossvalidation)}

\item{meta.extra.params}{Named list: Parameters for \link{s_GLMNET} for the
feature selection step}

\item{cases.by.rules}{Matrix of cases by rules from a previoue rulefit run.
If provided, the GBM step is skipped. Default = NULL}

\item{x.name}{Character: Name for feature set}

\item{y.name}{Character: Name for outcome}

\item{n.cores}{Integer: Number of cores to use}

\item{question}{Character: the question you are attempting to answer with this model, in plain language.}

\item{print.plot}{Logical: if TRUE, produce plot using \code{mplot3}
Takes precedence over \code{plot.fitted} and \code{plot.predicted}.}

\item{plot.fitted}{Logical: if TRUE, plot True (y) vs Fitted}

\item{plot.predicted}{Logical: if TRUE, plot True (y.test) vs Predicted.
Requires \code{x.test} and \code{y.test}}

\item{plot.theme}{Character: "zero", "dark", "box", "darkbox"}

\item{outdir}{Character: If defined, save log, 'plot.all' plots (see above) and RDS file of complete output}

\item{save.mod}{Logical: If TRUE, save all output to an RDS file in \code{outdir}
\code{save.mod} is TRUE by default if an \code{outdir} is defined. If set to TRUE, and no \code{outdir}
is defined, outdir defaults to \code{paste0("./s.", mod.name)}}

\item{verbose}{Logical: If TRUE, print summary to screen.}
}
\value{
\code{rtMod} object
}
\description{
Train a gradient boosting model, extract rules,
and fit using LASSO
}
\details{
Based on "Predictive Learning via Rule Ensembles"
by Friedman and Popescu
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\references{
Friedman JH, Popescu BE, "Predictive Learning via Rule Ensembles",
http://statweb.stanford.edu/~jhf/ftp/RuleFit.pdf
}
\author{
E.D. Gennatas
}
