% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePMML.R
\name{savePMML}
\alias{savePMML}
\title{Save rtemis model to PMML file}
\usage{
savePMML(
  x,
  filename,
  transforms = NULL,
  model_name = NULL,
  model_version = NULL,
  description = NULL,
  copyright = NULL,
  ...
)
}
\arguments{
\item{x}{rtemis model}

\item{filename}{Character: path to file}

\item{transforms}{List of PMML transformations}

\item{model_name}{Character: name of the model}

\item{model_version}{Character: version of the model}

\item{description}{Character: description of the model}

\item{copyright}{Character: copyright information}

\item{...}{Additional arguments passed to pmml::pmml()}
}
\description{
Save rtemis model to PMML file
}
\author{
E.D. Gennatas
}
