% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity}
\usage{
sensitivity(true, predicted, harmonize = FALSE, verbosity = 1L)
}
\arguments{
\item{true}{Factor: True labels.}

\item{predicted}{Factor: Predicted labels.}

\item{harmonize}{Logical: If TRUE, run \code{factor_harmonize} first.}

\item{verbosity}{Integer: Verbosity level.}
}
\description{
The first factor level is considered the positive case.
}
\keyword{internal}
