% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seql.R
\name{seql}
\alias{seql}
\title{Sequence generation with automatic cycling}
\usage{
seql(x, target)
}
\arguments{
\item{x}{R object of some \code{length}}

\item{target}{R object of some \code{length}}
}
\description{
Sequence generation with automatic cycling
}
\examples{
color <- c("red", "blue")
target <- 1:5
color[seql(color, target)]
# "red"  "blue" "red"  "blue" "red"
color <- c("red", "green", "blue", "yellow", "orange")
target <- 1:3
color[seql(color, target)]
# "red"   "green" "blue"
}
\author{
E.D. Gennatas
}
