% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{setup_CART}
\alias{setup_CART}
\title{Setup CART Hyperparameters}
\usage{
setup_CART(
  cp = 0.01,
  maxdepth = 20L,
  minsplit = 2L,
  minbucket = 1L,
  prune_cp = NULL,
  method = "auto",
  model = TRUE,
  maxcompete = 4L,
  maxsurrogate = 5L,
  usesurrogate = 2L,
  surrogatestyle = 0L,
  xval = 0L,
  cost = NULL,
  ifw = FALSE
)
}
\arguments{
\item{cp}{(Tunable) Numeric: Complexity parameter.}

\item{maxdepth}{(Tunable) Integer: Maximum depth of tree.}

\item{minsplit}{(Tunable) Integer: Minimum number of observations in a node to split.}

\item{minbucket}{(Tunable) Integer: Minimum number of observations in a terminal node.}

\item{prune_cp}{(Tunable) Numeric: Complexity for cost-complexity pruning after tree is built}

\item{method}{String: Splitting method.}

\item{model}{Logical: If TRUE, return a model.}

\item{maxcompete}{Integer: Maximum number of competitive splits.}

\item{maxsurrogate}{Integer: Maximum number of surrogate splits.}

\item{usesurrogate}{Integer: Number of surrogate splits to use.}

\item{surrogatestyle}{Integer: Type of surrogate splits.}

\item{xval}{Integer: Number of cross-validation folds.}

\item{cost}{Numeric (>=0): One for each feature.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
CARTHyperparameters object.
}
\description{
Setup hyperparameters for CART training.
}
\details{
Get more information from \link[rpart:rpart]{rpart::rpart} and \link[rpart:rpart.control]{rpart::rpart.control}.
}
\author{
EDG
}
