% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_Hyperparameters.R
\name{setup_CART}
\alias{setup_CART}
\title{Setup CART Hyperparameters}
\usage{
setup_CART(
  cp = 0.01,
  maxdepth = 20L,
  minsplit = 2L,
  minbucket = 1L,
  prune_cp = NULL,
  method = "auto",
  model = TRUE,
  maxcompete = 4L,
  maxsurrogate = 5L,
  usesurrogate = 2L,
  surrogatestyle = 0L,
  xval = 0L,
  cost = NULL,
  ifw = FALSE
)
}
\arguments{
\item{cp}{(Tunable) Numeric [0, Inf): Complexity parameter.}

\item{maxdepth}{(Tunable) Integer [1, 30]: Maximum depth of tree.}

\item{minsplit}{(Tunable) Integer [1, Inf): Minimum number of observations in a node to split.}

\item{minbucket}{(Tunable) Integer [1, Inf): Minimum number of observations in a terminal node.}

\item{prune_cp}{(Tunable) Optional Numeric [0, Inf): Complexity for cost-complexity pruning after tree is built.}

\item{method}{Character \{"auto", "anova", "class", "poisson", "exp"\}: Splitting method.}

\item{model}{Logical: If TRUE, keep a copy of the model frame.}

\item{maxcompete}{Integer [0, Inf): Maximum number of competitive splits.}

\item{maxsurrogate}{Integer [0, Inf): Maximum number of surrogate splits.}

\item{usesurrogate}{Integer [0, 2]: Number of surrogate splits to use.}

\item{surrogatestyle}{Integer [0, 1]: Type of surrogate splits.}

\item{xval}{Integer [0, Inf): Number of cross-validation folds.}

\item{cost}{Optional Numeric (0, Inf) vector: One for each feature.}

\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
CARTHyperparameters object.
}
\description{
Setup hyperparameters for CART training.
}
\details{
Get more information from \link[rpart:rpart]{rpart::rpart} and \link[rpart:rpart.control]{rpart::rpart.control}.
}
\examples{
cart_hyperparams <- setup_CART(cp = 0.01, maxdepth = 10L, ifw = TRUE)
cart_hyperparams
}
\author{
EDG
}
