% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120_ClusteringConfig.R
\name{setup_CMeans}
\alias{setup_CMeans}
\title{Setup CMeansConfig}
\usage{
setup_CMeans(
  k = 2L,
  max_iter = 100L,
  dist = "euclidean",
  method = "cmeans",
  m = 2,
  rate_par = NULL,
  weights = 1,
  control = list()
)
}
\arguments{
\item{k}{Integer [1, Inf): Number of clusters.}

\item{max_iter}{Integer [1, Inf): Maximum number of iterations.}

\item{dist}{Character \{"euclidean", "manhattan"\}: Distance measure to use.}

\item{method}{Character \{"cmeans", "ufcl"\}: "cmeans" - fuzzy c-means clustering; "ufcl": on-line update.}

\item{m}{Numeric (1, Inf): Degree of fuzzification.}

\item{rate_par}{Optional Numeric [0, 1]: Learning rate for the online variant.}

\item{weights}{Numeric vector: Case weights. Either a scalar, applied to every case, or a vector with one value per case.}

\item{control}{List: Control config for clustering algorithm.}
}
\value{
CMeansConfig object.
}
\description{
Setup CMeansConfig
}
\examples{
cmeans_config <- setup_CMeans(k = 4L, dist = "euclidean")
cmeans_config
}
\author{
EDG
}
