% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_S7_ClusteringConfig.R
\name{setup_CMeans}
\alias{setup_CMeans}
\title{Setup CMeansConfig}
\usage{
setup_CMeans(
  k = 2L,
  max_iter = 100L,
  dist = c("euclidean", "manhattan"),
  method = c("cmeans", "ufcl"),
  m = 2,
  rate_par = NULL,
  weights = 1,
  control = list()
)
}
\arguments{
\item{k}{Integer: Number of clusters.}

\item{max_iter}{Integer: Maximum number of iterations.}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'.}

\item{method}{Character: "cmeans" - fuzzy c-means clustering; "ufcl": on-line update.}

\item{m}{Float (>1): Degree of fuzzification.}

\item{rate_par}{Float (0, 1): Learning rate for the online variant.}

\item{weights}{Float (>0): Case weights.}

\item{control}{List: Control config for clustering algorithm.}
}
\value{
CMeansConfig object.
}
\description{
Setup CMeansConfig
}
\examples{
cmeans_config <- setup_CMeans(k = 4L, dist = "euclidean")
cmeans_config
}
\author{
EDG
}
