% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_S7_ClusteringParameters.R
\name{setup_DBSCAN}
\alias{setup_DBSCAN}
\title{Setup DBSCANParameters}
\usage{
setup_DBSCAN(
  eps = 0.5,
  min_points = 5L,
  weights = NULL,
  border_points = TRUE,
  search = c("kdtree", "linear", "dist"),
  bucket_size = 100L,
  split_rule = c("suggest", "STD", "MIDPT", "FAIR", "SL_MIDPT", "SL_FAIR"),
  approx = FALSE
)
}
\arguments{
\item{eps}{Float: Radius of neighborhood.}

\item{min_points}{Integer: Minimum number of points in a neighborhood to form a cluster.}

\item{weights}{Numeric vector: Weights for data points.}

\item{border_points}{Logical: If TRUE, assign border points to clusters.}

\item{search}{Character: Nearest neighbor search strategy: "kdtree", "linear", or "dist".}

\item{bucket_size}{Integer: Size of buckets for k-dtree search.}

\item{split_rule}{Character: Rule for splitting clusters: "SUGGEST", "STD", "MIDPT", "FAIR", "SL_MIDPT", "SL_FAIR".}

\item{approx}{Logical: If TRUE, use approximate nearest neighbor search.}
}
\value{
DBSCANParameters object.
}
\description{
Setup DBSCANParameters
}
\author{
EDG
}
