% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120_ClusteringConfig.R
\name{setup_DBSCAN}
\alias{setup_DBSCAN}
\title{Setup DBSCANConfig}
\usage{
setup_DBSCAN(
  eps = 0.5,
  min_points = 5L,
  weights = NULL,
  border_points = TRUE,
  search = "kdtree",
  bucket_size = 100L,
  split_rule = "SUGGEST",
  approx = FALSE
)
}
\arguments{
\item{eps}{Numeric (0, Inf): Radius of neighborhood.}

\item{min_points}{Integer [1, Inf): Minimum number of points in a neighborhood to form a cluster.}

\item{weights}{Optional Numeric vector: Weights for data points.}

\item{border_points}{Logical: If TRUE, assign border points to clusters.}

\item{search}{Character \{"kdtree", "linear", "dist"\}: Nearest neighbor search strategy.}

\item{bucket_size}{Integer [1, Inf): Size of buckets for k-d tree search.}

\item{split_rule}{Character \{"SUGGEST", "STD", "MIDPT", "FAIR", "SL_MIDPT", "SL_FAIR"\}: Rule for splitting the k-d tree.}

\item{approx}{Logical: If TRUE, use approximate nearest neighbor search.}
}
\value{
DBSCANConfig object.
}
\description{
Setup DBSCANConfig
}
\examples{
dbscan_config <- setup_DBSCAN(eps = 0.5, min_points = 5L)
dbscan_config
}
\author{
EDG
}
