% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_ExecutionConfig.R
\name{setup_ExecutionConfig}
\alias{setup_ExecutionConfig}
\title{Setup Execution Configuration}
\usage{
setup_ExecutionConfig(
  backend = c("future", "mirai", "none"),
  n_workers = NULL,
  future_plan = NULL
)
}
\arguments{
\item{backend}{Character: Execution backend: "future", "mirai", or "none".}

\item{n_workers}{Integer: Number of workers for parallel execution. Only used if \verb{backend is "future"} or "mirai". Do not rely on the default value, set to an appropriate number depending
on your system.}

\item{future_plan}{Character: Future plan to use if \code{backend} is "future".}
}
\value{
\code{ExecutionConfig} object.
}
\description{
Setup Execution Configuration
}
\examples{
setup_ExecutionConfig(backend = "future", n_workers = 4L, future_plan = "multisession")
}
\author{
EDG
}
