% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{setup_GAM}
\alias{setup_GAM}
\title{Setup GAM Hyperparameters}
\usage{
setup_GAM(k = 5L, ifw = FALSE)
}
\arguments{
\item{k}{(Tunable) Integer: Number of knots.}

\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
GAMHyperparameters object.
}
\description{
Setup hyperparameters for GAM training.
}
\details{
Get more information from \link[mgcv:gam]{mgcv::gam}.
}
\author{
EDG
}
