% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{setup_GLMNET}
\alias{setup_GLMNET}
\title{Setup GLMNET Hyperparameters}
\usage{
setup_GLMNET(
  alpha = 1,
  family = NULL,
  offset = NULL,
  which_lambda_cv = "lambda.1se",
  nlambda = 100L,
  lambda = NULL,
  penalty_factor = NULL,
  standardize = TRUE,
  intercept = TRUE,
  ifw = TRUE
)
}
\arguments{
\item{alpha}{(Tunable) Numeric: Mixing parameter.}

\item{family}{Character: Family for GLMNET.}

\item{offset}{Numeric: Offset for GLMNET.}

\item{which_lambda_cv}{Character: Which lambda to use for prediction:
"lambda.1se" or "lambda.min"}

\item{nlambda}{Positive integer: Number of lambda values.}

\item{lambda}{Numeric: Lambda values.}

\item{penalty_factor}{Numeric: Penalty factor for each feature.}

\item{standardize}{Logical: If TRUE, standardize features.}

\item{intercept}{Logical: If TRUE, include intercept.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
GLMNETHyperparameters object.
}
\description{
Setup hyperparameters for GLMNET training.
}
\details{
Get more information from \link[glmnet:glmnet]{glmnet::glmnet}.
}
\author{
EDG
}
