% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_Hyperparameters.R
\name{setup_GLMNET}
\alias{setup_GLMNET}
\title{Setup GLMNET Hyperparameters}
\usage{
setup_GLMNET(
  alpha = 1,
  family = NULL,
  offset = NULL,
  which_lambda_cv = "lambda.1se",
  nlambda = 100L,
  lambda = NULL,
  penalty_factor = NULL,
  standardize = TRUE,
  intercept = TRUE,
  ifw = TRUE
)
}
\arguments{
\item{alpha}{(Tunable) Numeric [0, 1]: Elastic net mixing parameter.}

\item{family}{Optional Character \{"gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"\}: Family. NULL = set from outcome type.}

\item{offset}{Optional Numeric vector: Offset, one value per case.}

\item{which_lambda_cv}{Character \{"lambda.1se", "lambda.min"\}: Which lambda to use for prediction.}

\item{nlambda}{Integer [1, Inf): Number of lambda values.}

\item{lambda}{Optional Numeric [0, Inf) vector: Lambda values. NULL = determined by cv.glmnet during tuning.}

\item{penalty_factor}{Optional Numeric [0, Inf) vector: Penalty factor for each feature.}

\item{standardize}{Logical: If TRUE, standardize features.}

\item{intercept}{Logical: If TRUE, include intercept.}

\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
GLMNETHyperparameters object.
}
\description{
Setup hyperparameters for GLMNET training.
}
\details{
Get more information from \link[glmnet:glmnet]{glmnet::glmnet}.
}
\examples{
glm_hyperparams <- setup_GLMNET(alpha = 1, ifw = TRUE)
glm_hyperparams
}
\author{
EDG
}
