% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_S7_Tuner.R
\name{setup_GridSearch}
\alias{setup_GridSearch}
\title{Setup Grid Search Config}
\usage{
setup_GridSearch(
  resampler_config = setup_Resampler(n_resamples = 5L, type = "KFold"),
  search_type = "exhaustive",
  randomize_p = NULL,
  metrics_aggregate_fn = "mean",
  metric = NULL,
  maximize = NULL
)
}
\arguments{
\item{resampler_config}{\code{ResamplerConfig} set by \link{setup_Resampler}.}

\item{search_type}{Character: "exhaustive" or "randomized". Type of
grid search to use. Exhaustive search will try all combinations of
config. Randomized will try a random sample of size
\code{randomize_p} * \verb{N of total combinations}}

\item{randomize_p}{Float (0, 1): For \code{search_type == "randomized"},
randomly test this proportion of combinations.}

\item{metrics_aggregate_fn}{Character: Name of function to use to aggregate error metrics.}

\item{metric}{Character: Metric to minimize or maximize.}

\item{maximize}{Logical: If TRUE, maximize \code{metric}, otherwise minimize it.}
}
\value{
A \code{GridSearchConfig} object.
}
\description{
Create a \code{GridSearchConfig} object that can be passed to \link{train}.
}
\examples{
gridsearch_config <- setup_GridSearch(
  resampler_config = setup_Resampler(n_resamples = 5L, type = "KFold"),
  search_type = "exhaustive"
)
gridsearch_config
}
\author{
EDG
}
