% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_ClusteringParameters.R
\name{setup_HardCL}
\alias{setup_HardCL}
\title{Setup HardCLParameters}
\usage{
setup_HardCL(k = 3L, dist = c("euclidean", "manhattan"))
}
\arguments{
\item{k}{Number of clusters.}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'.}
}
\value{
HardCLParameters object.
}
\description{
Setup HardCLParameters
}
\author{
EDG
}
