% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ClusteringConfig.R
\name{setup_HardCL}
\alias{setup_HardCL}
\title{Setup HardCLConfig}
\usage{
setup_HardCL(k = 3L, dist = "euclidean")
}
\arguments{
\item{k}{Integer [1, Inf): Number of clusters.}

\item{dist}{Character \{"euclidean", "manhattan"\}: Distance measure to use.}
}
\value{
HardCLConfig object.
}
\description{
Setup HardCLConfig
}
\examples{
hardcl_config <- setup_HardCL(k = 4L, dist = "euclidean")
hardcl_config
}
\author{
EDG
}
