% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_S7_ClusteringConfig.R
\name{setup_HardCL}
\alias{setup_HardCL}
\title{Setup HardCLConfig}
\usage{
setup_HardCL(k = 3L, dist = c("euclidean", "manhattan"))
}
\arguments{
\item{k}{Number of clusters.}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'.}
}
\value{
HardCLConfig object.
}
\description{
Setup HardCLConfig
}
\examples{
hardcl_config <- setup_HardCL(k = 4L, dist = "euclidean")
hardcl_config
}
\author{
EDG
}
