% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_DecompositionConfig.R
\name{setup_ICA}
\alias{setup_ICA}
\title{setup_ICA}
\usage{
setup_ICA(
  k = 3L,
  type = "parallel",
  fun = "logcosh",
  alpha = 1,
  row_norm = TRUE,
  maxit = 100L,
  tol = 1e-04,
  features = NULL
)
}
\arguments{
\item{k}{Integer [1, Inf): Number of components.}

\item{type}{Character \{"parallel", "deflation"\}: Type of ICA.}

\item{fun}{Character \{"logcosh", "exp"\}: ICA function.}

\item{alpha}{Numeric [1, 2]: Used in approximation to neg-entropy with \code{fun = "logcosh"}.}

\item{row_norm}{Logical: If TRUE, normalize rows of \code{x} before ICA.}

\item{maxit}{Integer [1, Inf): Maximum number of iterations.}

\item{tol}{Numeric [0, Inf): Tolerance.}

\item{features}{Optional Character vector: Names of at least 2 distinct
feature columns to decompose. \code{NULL} decomposes all numeric features.}
}
\value{
ICAConfig object.
}
\description{
Setup ICA config.
}
\examples{
ica_config <- setup_ICA(k = 3L)
ica_config
}
\author{
EDG
}
