% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_S7_DecompositionConfig.R
\name{setup_ICA}
\alias{setup_ICA}
\title{setup_ICA}
\usage{
setup_ICA(
  k = 3L,
  type = c("parallel", "deflation"),
  fun = c("logcosh", "exp"),
  alpha = 1,
  row_norm = TRUE,
  maxit = 100L,
  tol = 1e-04
)
}
\arguments{
\item{k}{Integer: Number of components.}

\item{type}{Character: Type of ICA: "parallel" or "deflation".}

\item{fun}{Character: ICA function: "logcosh", "exp".}

\item{alpha}{Numeric [1, 2]: Used in approximation to neg-entropy with \code{fun = "logcosh"}.}

\item{row_norm}{Logical: If TRUE, normalize rows of \code{x} before ICA.}

\item{maxit}{Integer: Maximum number of iterations.}

\item{tol}{Numeric: Tolerance.}
}
\value{
ICAConfig object.
}
\description{
Setup ICA config.
}
\examples{
ica_config <- setup_ICA(k = 3L)
ica_config
}
\author{
EDG
}
