% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150_DecompositionConfig.R
\name{setup_Isomap}
\alias{setup_Isomap}
\title{Setup Isomap config.}
\usage{
setup_Isomap(
  k = 2L,
  dist_method = "euclidean",
  nsd = 0L,
  path = "shortest",
  features = NULL
)
}
\arguments{
\item{k}{Integer [1, Inf): Number of components.}

\item{dist_method}{Character \{"euclidean", "manhattan"\}: Distance method.}

\item{nsd}{Integer [0, Inf): Number of shortest dissimilarities retained.}

\item{path}{Character \{"shortest", "extended"\}: Path argument for \code{vegan::isomap}.}

\item{features}{Optional Character vector: Names of at least 2 distinct
feature columns to decompose. \code{NULL} decomposes all numeric features.}
}
\value{
IsomapConfig object.
}
\description{
Setup Isomap config.
}
\examples{
isomap_config <- setup_Isomap(k = 3L)
isomap_config
}
\author{
EDG
}
