% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_DecompositionConfig.R
\name{setup_Isomap}
\alias{setup_Isomap}
\title{Setup Isomap config.}
\usage{
setup_Isomap(
  k = 2L,
  dist_method = c("euclidean", "manhattan"),
  nsd = 0L,
  path = c("shortest", "extended")
)
}
\arguments{
\item{k}{Integer: Number of components.}

\item{dist_method}{Character: Distance method.}

\item{nsd}{Integer: Number of shortest dissimilarities retained.}

\item{path}{Character: Path argument for \code{vegan::isomap}.}
}
\value{
IsomapConfig object.
}
\description{
Setup Isomap config.
}
\examples{
isomap_config <- setup_Isomap(k = 3L)
isomap_config
}
\author{
EDG
}
