% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_Hyperparameters.R
\name{setup_Isotonic}
\alias{setup_Isotonic}
\title{Setup Isotonic Hyperparameters}
\usage{
setup_Isotonic(ifw = FALSE)
}
\arguments{
\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
IsotonicHyperparameters object.
}
\description{
Setup hyperparameters for Isotonic Regression.
}
\details{
There are not hyperparameters for this algorithm at this moment.
}
\examples{
isotonic_hyperparams <- setup_Isotonic(ifw = TRUE)
isotonic_hyperparams
}
\author{
EDG
}
