% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_S7_ClusteringParameters.R
\name{setup_KMeans}
\alias{setup_KMeans}
\title{Setup KMmeansParameters}
\usage{
setup_KMeans(k = 3L, dist = c("euclidean", "manhattan"))
}
\arguments{
\item{k}{Number of clusters.}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'.}
}
\value{
KMeansParameters object.
}
\description{
Setup KMmeansParameters
}
\author{
EDG
}
