% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ClusteringConfig.R
\name{setup_KMeans}
\alias{setup_KMeans}
\title{Setup KMeansConfig}
\usage{
setup_KMeans(k = 3L, dist = "euclidean")
}
\arguments{
\item{k}{Integer [1, Inf): Number of clusters.}

\item{dist}{Character \{"euclidean", "manhattan"\}: Distance measure to use.}
}
\value{
KMeansConfig object.
}
\description{
Setup KMeansConfig
}
\examples{
kmeans_config <- setup_KMeans(k = 4L, dist = "euclidean")
kmeans_config
}
\author{
EDG
}
