% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_ClusteringConfig.R
\name{setup_KMeans}
\alias{setup_KMeans}
\title{Setup KMeansConfig}
\usage{
setup_KMeans(k = 3L, dist = c("euclidean", "manhattan"))
}
\arguments{
\item{k}{Number of clusters.}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'.}
}
\value{
KMeansConfig object.
}
\description{
Setup KMeansConfig
}
\examples{
kmeans_config <- setup_KMeans(k = 4L, dist = "euclidean")
kmeans_config
}
\author{
EDG
}
