% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{setup_LightCART}
\alias{setup_LightCART}
\title{Setup LightCART Hyperparameters}
\usage{
setup_LightCART(
  num_leaves = 32L,
  max_depth = -1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  max_cat_threshold = 32L,
  min_data_per_group = 100L,
  linear_tree = FALSE,
  objective = NULL,
  ifw = FALSE
)
}
\arguments{
\item{num_leaves}{(Tunable) Positive integer: Maximum number of leaves in one tree.}

\item{max_depth}{(Tunable) Integer: Maximum depth of trees.}

\item{lambda_l1}{(Tunable) Numeric: L1 regularization.}

\item{lambda_l2}{(Tunable) Numeric: L2 regularization.}

\item{max_cat_threshold}{(Tunable) Positive integer: Maximum number of categories for categorical features.}

\item{min_data_per_group}{(Tunable) Positive integer: Minimum number of data per categorical group.}

\item{linear_tree}{(Tunable) Logical: If TRUE, use linear trees.}

\item{objective}{Character: Objective function.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
LightCARTHyperparameters object.
}
\description{
Setup hyperparameters for LightCART training.
}
\details{
Get more information from \link[lightgbm:lgb.train]{lightgbm::lgb.train}.
}
\author{
EDG
}
