% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_Hyperparameters.R
\name{setup_LightCART}
\alias{setup_LightCART}
\title{Setup LightCART Hyperparameters}
\usage{
setup_LightCART(
  num_leaves = 32L,
  max_depth = -1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  min_data_in_leaf = 20L,
  max_cat_threshold = 32L,
  min_data_per_group = 100L,
  linear_tree = FALSE,
  objective = NULL,
  ifw = FALSE
)
}
\arguments{
\item{num_leaves}{(Tunable) Integer [1, Inf): Maximum number of leaves in one tree.}

\item{max_depth}{(Tunable) Integer: Maximum depth of trees. -1 = no limit.}

\item{lambda_l1}{(Tunable) Numeric [0, Inf): L1 regularization.}

\item{lambda_l2}{(Tunable) Numeric [0, Inf): L2 regularization.}

\item{min_data_in_leaf}{(Tunable) Integer [1, Inf): Minimum number of data in a leaf.}

\item{max_cat_threshold}{(Tunable) Integer [1, Inf): Maximum number of categories for categorical features.}

\item{min_data_per_group}{(Tunable) Integer [1, Inf): Minimum number of observations per categorical group.}

\item{linear_tree}{(Tunable) Logical: If TRUE, use linear trees.}

\item{objective}{Optional Character: Objective function. NULL = set from outcome type.}

\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
LightCARTHyperparameters object.
}
\description{
Setup hyperparameters for LightCART training.
}
\details{
Get more information from \link[lightgbm:lgb.train]{lightgbm::lgb.train}.
}
\examples{
lightcart_hyperparams <- setup_LightCART(num_leaves = 32L, ifw = FALSE)
lightcart_hyperparams
}
\author{
EDG
}
