% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{setup_LightGBM}
\alias{setup_LightGBM}
\title{Setup LightGBM Hyperparameters}
\usage{
setup_LightGBM(
  max_nrounds = 1000L,
  force_nrounds = NULL,
  early_stopping_rounds = 10L,
  num_leaves = 8L,
  max_depth = -1L,
  learning_rate = 0.01,
  feature_fraction = 1,
  subsample = 1,
  subsample_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  max_cat_threshold = 32L,
  min_data_per_group = 32L,
  linear_tree = FALSE,
  objective = NULL,
  ifw = FALSE
)
}
\arguments{
\item{max_nrounds}{Positive integer: Maximum number of boosting rounds.}

\item{force_nrounds}{Positive integer: Use this many boosting rounds. Disable search for nrounds.}

\item{early_stopping_rounds}{Positive integer: Number of rounds without improvement to stop training.}

\item{num_leaves}{(Tunable) Positive integer: Maximum number of leaves in one tree.}

\item{max_depth}{(Tunable) Integer: Maximum depth of trees.}

\item{learning_rate}{(Tunable) Numeric: Learning rate.}

\item{feature_fraction}{(Tunable) Numeric: Fraction of features to use.}

\item{subsample}{(Tunable) Numeric: Fraction of data to use.}

\item{subsample_freq}{(Tunable) Positive integer: Frequency of subsample.}

\item{lambda_l1}{(Tunable) Numeric: L1 regularization.}

\item{lambda_l2}{(Tunable) Numeric: L2 regularization.}

\item{max_cat_threshold}{(Tunable) Positive integer: Maximum number of categories for categorical features.}

\item{min_data_per_group}{(Tunable) Positive integer: Minimum number of data per categorical group.}

\item{linear_tree}{Logical: If TRUE, use linear trees.}

\item{objective}{Character: Objective function.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\description{
Setup hyperparameters for LightGBM training.
}
\details{
Get more information from \link[lightgbm:lgb.train]{lightgbm::lgb.train}.
}
\author{
EDG
}
