% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Hyperparameters.R
\name{setup_LightGBM}
\alias{setup_LightGBM}
\title{Setup LightGBM Hyperparameters}
\usage{
setup_LightGBM(
  max_nrounds = 1000L,
  force_nrounds = NULL,
  early_stopping_rounds = 10L,
  num_leaves = 8L,
  max_depth = -1L,
  learning_rate = 0.01,
  feature_fraction = 1,
  subsample = 1,
  subsample_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  max_cat_threshold = 32L,
  min_data_per_group = 32L,
  linear_tree = FALSE,
  ifw = FALSE,
  objective = NULL,
  device_type = "cpu",
  tree_learner = "serial",
  force_col_wise = TRUE
)
}
\arguments{
\item{max_nrounds}{Integer [1, Inf): Maximum number of boosting rounds.}

\item{force_nrounds}{Optional Integer [1, Inf): Use this many boosting rounds. Disables search for nrounds.}

\item{early_stopping_rounds}{Integer [1, Inf): Number of rounds without improvement to stop training.}

\item{num_leaves}{(Tunable) Integer [1, Inf): Maximum number of leaves in one tree.}

\item{max_depth}{(Tunable) Integer: Maximum depth of trees. -1 = no limit.}

\item{learning_rate}{(Tunable) Numeric (0, 1]: Learning rate.}

\item{feature_fraction}{(Tunable) Numeric (0, 1]: Fraction of features to use.}

\item{subsample}{(Tunable) Numeric (0, 1]: Fraction of data to use.}

\item{subsample_freq}{(Tunable) Integer [1, Inf): Frequency of subsample.}

\item{lambda_l1}{(Tunable) Numeric [0, Inf): L1 regularization.}

\item{lambda_l2}{(Tunable) Numeric [0, Inf): L2 regularization.}

\item{max_cat_threshold}{(Tunable) Integer [1, Inf): Maximum number of categories for categorical features.}

\item{min_data_per_group}{(Tunable) Integer [1, Inf): Minimum number of observations per categorical group.}

\item{linear_tree}{(Tunable) Logical: If TRUE, use linear trees.}

\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}

\item{objective}{Optional Character: Objective function. NULL = set from outcome type.}

\item{device_type}{Character \{"cpu", "gpu", "cuda"\}: Compute device.}

\item{tree_learner}{Character \{"serial", "feature", "data", "voting"\}: Tree learner type.}

\item{force_col_wise}{Logical: Use only with CPU - If TRUE, force col-wise histogram building.}
}
\value{
LightGBMHyperparameters object.
}
\description{
Setup hyperparameters for LightGBM training.
}
\details{
Get more information from \link[lightgbm:lgb.train]{lightgbm::lgb.train}.
\code{nrounds} is auto-tuned using early stopping (up to \code{max_nrounds}) unless
\code{force_nrounds} is set.
}
\examples{
lightgbm_hyperparams <- setup_LightGBM(
  max_nrounds = 500L,
  learning_rate = c(0.001, 0.01, 0.05), ifw = TRUE
)
lightgbm_hyperparams
}
\author{
EDG
}
