% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{setup_LightRuleFit}
\alias{setup_LightRuleFit}
\title{Setup LightRuleFit Hyperparameters}
\usage{
setup_LightRuleFit(
  nrounds = 200L,
  num_leaves = 32L,
  max_depth = 4L,
  learning_rate = 0.1,
  subsample = 0.666,
  subsample_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  objective = NULL,
  ifw_lightgbm = FALSE,
  alpha = 1,
  lambda = NULL,
  ifw_glmnet = FALSE,
  ifw = FALSE
)
}
\arguments{
\item{nrounds}{(Tunable) Positive integer: Number of boosting rounds.}

\item{num_leaves}{(Tunable) Positive integer: Maximum number of leaves in one tree.}

\item{max_depth}{(Tunable) Integer: Maximum depth of trees.}

\item{learning_rate}{(Tunable) Numeric: Learning rate.}

\item{subsample}{(Tunable) Numeric: Fraction of data to use.}

\item{subsample_freq}{(Tunable) Positive integer: Frequency of subsample.}

\item{lambda_l1}{(Tunable) Numeric: L1 regularization.}

\item{lambda_l2}{(Tunable) Numeric: L2 regularization.}

\item{objective}{Character: Objective function.}

\item{ifw_lightgbm}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in the LightGBM
step.}

\item{alpha}{(Tunable) Numeric: Alpha for GLMNET.}

\item{lambda}{Numeric: Lambda for GLMNET.}

\item{ifw_glmnet}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in the GLMNET step.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification. This applies IFW
to both LightGBM and GLMNET.}
}
\value{
LightRuleFitHyperparameters object.
}
\description{
Setup hyperparameters for LightRuleFit training.
}
\details{
Get more information from \link[lightgbm:lgb.train]{lightgbm::lgb.train}.
}
\author{
EDG
}
