% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_Hyperparameters.R
\name{setup_LightRuleFit}
\alias{setup_LightRuleFit}
\title{Setup LightRuleFit Hyperparameters}
\usage{
setup_LightRuleFit(
  nrounds = 200L,
  num_leaves = 32L,
  max_depth = 4L,
  learning_rate = 0.1,
  subsample = 0.666,
  subsample_freq = 1L,
  lambda_l1 = 0,
  lambda_l2 = 0,
  objective = NULL,
  ifw_lightgbm = FALSE,
  alpha = 1,
  lambda = NULL,
  ifw_glmnet = FALSE,
  ifw = FALSE
)
}
\arguments{
\item{nrounds}{(Tunable) Integer [1, Inf): Number of boosting rounds.}

\item{num_leaves}{(Tunable) Integer [1, Inf): Maximum number of leaves in one tree.}

\item{max_depth}{(Tunable) Integer: Maximum depth of trees.}

\item{learning_rate}{(Tunable) Numeric (0, 1]: Learning rate.}

\item{subsample}{(Tunable) Numeric (0, 1]: Fraction of data to use.}

\item{subsample_freq}{(Tunable) Integer [1, Inf): Frequency of subsample.}

\item{lambda_l1}{(Tunable) Numeric [0, Inf): L1 regularization.}

\item{lambda_l2}{(Tunable) Numeric [0, Inf): L2 regularization.}

\item{objective}{Optional Character: Objective function. NULL = set from outcome type.}

\item{ifw_lightgbm}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in the LightGBM
step.}

\item{alpha}{(Tunable) Numeric [0, 1]: Alpha for GLMNET.}

\item{lambda}{Optional Numeric [0, Inf) vector: Lambda for GLMNET. NULL = determined by cv.glmnet.}

\item{ifw_glmnet}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in the GLMNET step.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification. This applies IFW
to both LightGBM and GLMNET; cannot be combined with \code{ifw_lightgbm} or \code{ifw_glmnet}.}
}
\value{
LightRuleFitHyperparameters object.
}
\description{
Setup hyperparameters for LightRuleFit training.
}
\details{
Get more information from \link[lightgbm:lgb.train]{lightgbm::lgb.train}.
}
\examples{
lightrulefit_hyperparams <- setup_LightRuleFit(nrounds = 300L, max_depth = 3L)
lightrulefit_hyperparams
}
\author{
EDG
}
