% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_S7_Hyperparameters.R
\name{setup_LinearSVM}
\alias{setup_LinearSVM}
\title{Setup LinearSVM Hyperparameters}
\usage{
setup_LinearSVM(cost = 1, ifw = FALSE)
}
\arguments{
\item{cost}{(Tunable) Numeric: Cost of constraints violation.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
LinearSVMHyperparameters object.
}
\description{
Setup hyperparameters for LinearSVM training.
}
\details{
Get more information from \link[e1071:svm]{e1071::svm}.
}
\author{
EDG
}
