% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Hyperparameters.R
\name{setup_LinearSVM}
\alias{setup_LinearSVM}
\title{Setup LinearSVM Hyperparameters}
\usage{
setup_LinearSVM(cost = 1, ifw = FALSE)
}
\arguments{
\item{cost}{(Tunable) Numeric (0, Inf): Cost of constraints violation.}

\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
LinearSVMHyperparameters object.
}
\description{
Setup hyperparameters for LinearSVM training.
}
\details{
Get more information from \link[e1071:svm]{e1071::svm}.
}
\examples{
linear_svm_hyperparams <- setup_LinearSVM(cost = 0.5, ifw = TRUE)
linear_svm_hyperparams
}
\author{
EDG
}
