% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_S7_ClusteringParameters.R
\name{setup_NeuralGas}
\alias{setup_NeuralGas}
\title{Setup NeuralGasParameters}
\usage{
setup_NeuralGas(k = 3L, dist = c("euclidean", "manhattan"))
}
\arguments{
\item{k}{Number of clusters.}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'.}
}
\value{
NeuralGasParameters object.
}
\description{
Setup NeuralGasParameters
}
\author{
EDG
}
