% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ClusteringConfig.R
\name{setup_NeuralGas}
\alias{setup_NeuralGas}
\title{Setup NeuralGasConfig}
\usage{
setup_NeuralGas(k = 3L, dist = "euclidean")
}
\arguments{
\item{k}{Integer [1, Inf): Number of clusters.}

\item{dist}{Character \{"euclidean", "manhattan"\}: Distance measure to use.}
}
\value{
NeuralGasConfig object.
}
\description{
Setup NeuralGasConfig
}
\examples{
neuralgas_config <- setup_NeuralGas(k = 4L, dist = "euclidean")
neuralgas_config
}
\author{
EDG
}
