% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_S7_ClusteringConfig.R
\name{setup_NeuralGas}
\alias{setup_NeuralGas}
\title{Setup NeuralGasConfig}
\usage{
setup_NeuralGas(k = 3L, dist = c("euclidean", "manhattan"))
}
\arguments{
\item{k}{Number of clusters.}

\item{dist}{Character: Distance measure to use: 'euclidean' or 'manhattan'.}
}
\value{
NeuralGasConfig object.
}
\description{
Setup NeuralGasConfig
}
\examples{
neuralgas_config <- setup_NeuralGas(k = 4L, dist = "euclidean")
neuralgas_config
}
\author{
EDG
}
