% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_S7_DecompositionConfig.R
\name{setup_NMF}
\alias{setup_NMF}
\title{Setup NMF config.}
\usage{
setup_NMF(k = 2L, method = "brunet", nrun = if (length(k) > 1L) 30L else 1L)
}
\arguments{
\item{k}{Integer: Number of components.}

\item{method}{Character: NMF method. See \code{NMF::nmf}.}

\item{nrun}{Integer: Number of runs to perform.}
}
\value{
NMFConfig object.
}
\description{
Setup NMF config.
}
\examples{
nmf_config <- setup_NMF(k = 3L)
nmf_config
}
\author{
EDG
}
