% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_S7_DecompositionParameters.R
\name{setup_PCA}
\alias{setup_PCA}
\title{Setup PCA parameters.}
\usage{
setup_PCA(k = 3L, center = TRUE, scale = TRUE, tol = NULL)
}
\arguments{
\item{k}{Integer: Number of components. (passed to \code{prcomp} \code{rank.})}

\item{center}{Logical: If TRUE, center the data.}

\item{scale}{Logical: If TRUE, scale the data.}

\item{tol}{Numeric: Tolerance.}
}
\value{
PCAParameters object.
}
\description{
Setup PCA parameters.
}
\author{
EDG
}
