% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_DecompositionConfig.R
\name{setup_PCA}
\alias{setup_PCA}
\title{Setup PCA config.}
\usage{
setup_PCA(k = 3L, center = TRUE, scale = TRUE, tol = NULL, features = NULL)
}
\arguments{
\item{k}{Integer [1, Inf): Number of components. (passed to \code{prcomp} \code{rank.})}

\item{center}{Logical: If TRUE, center the data.}

\item{scale}{Logical: If TRUE, scale the data.}

\item{tol}{Optional Numeric [0, Inf): Tolerance.}

\item{features}{Optional Character: Names of the feature columns to decompose.
\code{NULL} decomposes all numeric features.}
}
\value{
PCAConfig object.
}
\description{
Setup PCA config.
}
\examples{
pca_config <- setup_PCA(k = 3L)
pca_config
}
\author{
EDG
}
