% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Preprocessor.R
\name{setup_Preprocessor}
\alias{setup_Preprocessor}
\title{Setup Preprocessor}
\usage{
setup_Preprocessor(
  complete_cases = FALSE,
  remove_features_thres = NULL,
  remove_cases_thres = NULL,
  missingness = FALSE,
  impute = FALSE,
  impute_type = c("missRanger", "micePMM", "meanMode"),
  impute_missRanger_params = list(pmm.k = 3, maxiter = 10, num.trees = 500),
  impute_discrete = "get_mode",
  impute_continuous = "mean",
  integer2factor = FALSE,
  integer2numeric = FALSE,
  logical2factor = FALSE,
  logical2numeric = FALSE,
  numeric2factor = FALSE,
  numeric2factor_levels = NULL,
  numeric_cut_n = 0L,
  numeric_cut_labels = FALSE,
  numeric_quant_n = 0L,
  numeric_quant_NAonly = FALSE,
  unique_len2factor = 0L,
  character2factor = FALSE,
  factorNA2missing = FALSE,
  factorNA2missing_level = "missing",
  factor2integer = FALSE,
  factor2integer_startat0 = TRUE,
  scale = FALSE,
  center = scale,
  scale_centers = NULL,
  scale_coefficients = NULL,
  remove_constants = FALSE,
  remove_constants_skip_missing = TRUE,
  remove_features = NULL,
  remove_duplicates = FALSE,
  one_hot = FALSE,
  one_hot_levels = NULL,
  add_date_features = FALSE,
  date_features = c("weekday", "month", "year"),
  add_holidays = FALSE,
  exclude = NULL
)
}
\arguments{
\item{complete_cases}{Logical: If TRUE, only retain complete cases (no missing data).}

\item{remove_features_thres}{Optional Numeric (0, 1]: Remove features with
missing values in >= to this fraction of cases. 1 removes only fully-missing
features.}

\item{remove_cases_thres}{Optional Numeric (0, 1]: Remove cases with >= to
this fraction of missing features. 1 removes only fully-missing cases.}

\item{missingness}{Logical: If TRUE, generate new boolean columns for each
feature with missing values, indicating which cases were missing data.}

\item{impute}{Logical: If TRUE, impute missing cases. See \code{impute_discrete} and
\code{impute_continuous}.}

\item{impute_type}{Character \{"missRanger", "micePMM", "meanMode"\}: Package to use for
imputation.}

\item{impute_missRanger_params}{List: Parameters passed to
\code{missRanger::missRanger}, e.g. "pmm.k", "maxiter", "num.trees". \code{pmm.k}
greater than 0 results in predictive mean matching. Reduce \code{num.trees} for
faster imputation especially in large datasets. Set \code{pmm.k = 0} to
disable predictive mean matching.}

\item{impute_discrete}{Character: Name of function that returns single value: How to impute
discrete variables for \code{impute_type = "meanMode"}.}

\item{impute_continuous}{Character: Name of function that returns single value: How to impute
continuous variables for \code{impute_type = "meanMode"}.}

\item{integer2factor}{Logical: If TRUE, convert all integers to factors. This includes
\code{bit64::integer64} columns.}

\item{integer2numeric}{Logical: If TRUE, convert all integers to numeric
(will only work if \code{integer2factor = FALSE}). This includes
\code{bit64::integer64} columns.}

\item{logical2factor}{Logical: If TRUE, convert all logical variables to
factors.}

\item{logical2numeric}{Logical: If TRUE, convert all logical variables to
numeric.}

\item{numeric2factor}{Logical: If TRUE, convert all numeric variables to
factors.}

\item{numeric2factor_levels}{Optional Character vector: Will be passed to
\code{levels} arg of \code{factor()} if \code{numeric2factor = TRUE}. For advanced/
specific use cases; need to know unique values of numeric vector(s) and given all
numeric vars have same unique values.}

\item{numeric_cut_n}{Integer [0, Inf): If > 0, convert all numeric variables to factors by
binning using \code{base::cut} with \code{breaks} equal to this number.}

\item{numeric_cut_labels}{Logical: The \code{labels} argument of \link[base:cut]{base::cut}.}

\item{numeric_quant_n}{Integer [0, Inf): If > 0, convert all numeric variables to factors by
binning using \code{base::cut} with \code{breaks} equal to this number of quantiles.
produced using \code{stats::quantile}.}

\item{numeric_quant_NAonly}{Logical: If TRUE, only bin numeric variables with
missing values.}

\item{unique_len2factor}{Integer [0, Inf): Convert all variables with less
than or equal to this number of unique values to factors. 0 disables.
For example, if binary variables are encoded with 1, 2, you could use
\code{unique_len2factor = 2} to convert them to factors.}

\item{character2factor}{Logical: If TRUE, convert all character variables to
factors.}

\item{factorNA2missing}{Logical: If TRUE, make NA values in factors be of
level \code{factorNA2missing_level}. In many cases this is the preferred way
to handle missing data in categorical variables. Note that since this step
is performed before imputation, you can use this option to handle missing
data in categorical variables and impute numeric variables in the same
\code{preprocess} call.}

\item{factorNA2missing_level}{Character: Name of level if
\code{factorNA2missing = TRUE}.}

\item{factor2integer}{Logical: If TRUE, convert all factors to integers.}

\item{factor2integer_startat0}{Logical: If TRUE, start integer coding at 0.}

\item{scale}{Logical: If TRUE, scale columns of \code{x}.}

\item{center}{Logical: If TRUE, center columns of \code{x}. If unset, follows \code{scale}.}

\item{scale_centers}{Optional Named vector: Centering values for each feature.}

\item{scale_coefficients}{Optional Named vector: Scaling values for each feature.}

\item{remove_constants}{Logical: If TRUE, remove constant columns.}

\item{remove_constants_skip_missing}{Logical: If TRUE, skip missing values, before
checking if feature is constant.}

\item{remove_features}{Optional Character vector: Features to remove.}

\item{remove_duplicates}{Logical: If TRUE, remove duplicate cases.}

\item{one_hot}{Logical: If TRUE, convert all factors using one-hot encoding.}

\item{one_hot_levels}{Optional List: Named list of the form "feature_name" = "levels". Used when applying
one-hot encoding to validation or test data using \code{Preprocessor}.}

\item{add_date_features}{Logical: If TRUE, extract date features from date columns.}

\item{date_features}{Character \{"weekday", "month", "year"\} vector:
Features to extract from dates.}

\item{add_holidays}{Logical: If TRUE, extract holidays from date columns.}

\item{exclude}{Optional Integer vector: Exclude these columns from preprocessing.}
}
\value{
\code{PreprocessorConfig} object.
}
\description{
Creates a \code{PreprocessorConfig} object, which can be used in \link{preprocess}.
}
\section{Order of Operations}{

\itemize{
\item keep complete cases only
\item remove constants
\item remove duplicates
\item remove cases by missingness threshold
\item remove features by missingness threshold
\item integer to factor
\item integer to numeric
\item logical to factor
\item logical to numeric
\item numeric to factor
\item cut numeric to n bins
\item cut numeric to n quantiles
\item numeric with less than N unique values to factor
\item character to factor
\item factor NA to named level
\item add missingness column
\item impute
\item scale and/or center
\item one-hot encoding
}
}

\examples{
preproc_config <- setup_Preprocessor(factorNA2missing = TRUE)
preproc_config
}
\author{
EDG
}
