% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Hyperparameters.R
\name{setup_RadialSVM}
\alias{setup_RadialSVM}
\title{Setup RadialSVM Hyperparameters}
\usage{
setup_RadialSVM(cost = 1, gamma = 0.01, ifw = FALSE)
}
\arguments{
\item{cost}{(Tunable) Numeric: Cost of constraints violation.}

\item{gamma}{(Tunable) Numeric: Kernel coefficient.}

\item{ifw}{Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
RadialSVMHyperparameters object.
}
\description{
Setup hyperparameters for RadialSVM training.
}
\details{
Get more information from \link[e1071:svm]{e1071::svm}.
}
\examples{
radial_svm_hyperparams <- setup_RadialSVM(cost = 10, gamma = 0.1, ifw = TRUE)
radial_svm_hyperparams
}
\author{
EDG
}
