% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Resampler.R
\name{setup_Resampler}
\alias{setup_Resampler}
\title{Setup Resampler}
\usage{
setup_Resampler(
  n_resamples = 10L,
  type = c("KFold", "StratSub", "StratBoot", "Bootstrap", "LOOCV"),
  stratify_var = NULL,
  train_p = 0.75,
  strat_n_bins = 4L,
  target_length = NULL,
  id_strat = NULL,
  seed = NULL,
  verbosity = 1L
)
}
\arguments{
\item{n_resamples}{Optional Integer [1, Inf): Number of resamples to make.
Stored as the \code{n} property. LOOCV determines it from the data.}

\item{type}{Character \{"KFold", "StratSub", "StratBoot", "Bootstrap", "LOOCV", "Custom"\}:
Type of resampler.}

\item{stratify_var}{Optional Character: Variable to stratify by.}

\item{train_p}{Numeric (0, 1): Training set percentage.}

\item{strat_n_bins}{Integer [1, Inf): Number of bins to stratify by.}

\item{target_length}{Optional Integer [1, Inf): Target length for stratified bootstraps.}

\item{id_strat}{Integer: Vector of indices to stratify by. These may be, for example, case IDs
if your dataset contains repeated measurements. By specifying this vector, you can ensure that
each case can only be present in the training or test set, but not both.}

\item{seed}{Optional Integer [0, Inf): Random seed.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
ResamplerConfig object.
}
\description{
Setup Resampler
}
\examples{
tenfold_resampler <- setup_Resampler(n_resamples = 10L, type = "KFold", seed = 2026L)
tenfold_resampler
}
\author{
EDG
}
