% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_S7_SuperConfig.R
\name{setup_superconfig}
\alias{setup_superconfig}
\title{Setup SuperConfig}
\usage{
setup_superconfig(
  dat_training_path,
  dat_validation_path = NULL,
  dat_test_path = NULL,
  weights = NULL,
  preprocessor_config = NULL,
  algorithm = NULL,
  hyperparameters = NULL,
  tuner_config = NULL,
  outer_resampling_config = NULL,
  execution_config,
  question = NULL,
  outdir = "results/",
  verbosity = 1L
)
}
\arguments{
\item{dat_training_path}{Character: Path to training data file.}

\item{dat_validation_path}{Character: Path to validation data file.}

\item{dat_test_path}{Character: Path to test data file.}

\item{weights}{Optional Character: Column name in training data to use as observation weights. If NULL, no weights are used.}

\item{preprocessor_config}{\code{PreprocessorConfig} object: Configuration for data preprocessing.}

\item{algorithm}{Character: Algorithm to use for training.}

\item{hyperparameters}{\code{Hyperparameters} object: Configuration for model hyperparameters.}

\item{tuner_config}{\code{TunerConfig} object: Configuration for hyperparameter tuning.}

\item{outer_resampling_config}{\code{ResamplerConfig} object: Configuration for outer res
resampling during model training.}

\item{question}{Character: Question to answer with the supervised learning analysis.}

\item{outdir}{Character: Output directory for results.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{SuperConfig} object.
}
\description{
Setup \code{SuperConfig} object.
}
\author{
EDG
}
