% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_SuperConfig.R
\name{setup_SuperConfig}
\alias{setup_SuperConfig}
\title{Setup SuperConfig}
\usage{
setup_SuperConfig(
  dat_training_path,
  dat_validation_path = NULL,
  dat_test_path = NULL,
  weights = NULL,
  preprocessor_config = NULL,
  algorithm = NULL,
  hyperparameters = NULL,
  tuner_config = NULL,
  outer_resampling_config = NULL,
  execution_config = setup_ExecutionConfig(),
  question = NULL,
  outdir = "results/",
  verbosity = 1L
)
}
\arguments{
\item{dat_training_path}{Character: Path to training data file.}

\item{dat_validation_path}{Character: Path to validation data file.}

\item{dat_test_path}{Character: Path to test data file.}

\item{weights}{Optional Character: Column name in training data to use as observation weights. If NULL, no weights are used.}

\item{preprocessor_config}{\code{PreprocessorConfig} object: Configuration for data preprocessing.}

\item{algorithm}{Character: Algorithm to use for training.}

\item{hyperparameters}{\code{Hyperparameters} object: Configuration for model hyperparameters.}

\item{tuner_config}{\code{TunerConfig} object: Configuration for hyperparameter tuning.}

\item{outer_resampling_config}{\code{ResamplerConfig} object: Configuration for outer res
resampling during model training.}

\item{execution_config}{\code{ExecutionConfig} object: Configuration for execution settings. Setup
with \link{setup_ExecutionConfig}.}

\item{question}{Character: Question to answer with the supervised learning analysis.}

\item{outdir}{Character: Output directory for results.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{SuperConfig} object.
}
\description{
Setup \code{SuperConfig} object.
}
\examples{
sc <- setup_SuperConfig(
  dat_training_path = "train.csv",
  preprocessor_config = setup_Preprocessor(remove_duplicates = TRUE),
  algorithm = "LightRF",
  hyperparameters = setup_LightRF(),
  tuner_config = setup_GridSearch(),
  outer_resampling_config = setup_Resampler(),
  execution_config = setup_ExecutionConfig(),
  question = "Can we tell iris species apart given their measurements?",
  outdir = "models/"
)
}
\author{
EDG
}
