% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Hyperparameters.R
\name{setup_TabNet}
\alias{setup_TabNet}
\title{Setup TabNet Hyperparameters}
\usage{
setup_TabNet(
  batch_size = 1048576L,
  penalty = 0.001,
  clip_value = NULL,
  loss = "auto",
  epochs = 50L,
  drop_last = FALSE,
  decision_width = NULL,
  attention_width = NULL,
  num_steps = 3L,
  feature_reusage = 1.3,
  mask_type = "sparsemax",
  virtual_batch_size = 65536L,
  valid_split = 0,
  learn_rate = 0.02,
  optimizer = "adam",
  lr_scheduler = NULL,
  lr_decay = 0.1,
  step_size = 30L,
  checkpoint_epochs = 10L,
  cat_emb_dim = 1L,
  num_independent = 2L,
  num_shared = 2L,
  num_independent_decoder = 1L,
  num_shared_decoder = 1L,
  momentum = 0.02,
  pretraining_ratio = 0.5,
  device = "auto",
  importance_sample_size = NULL,
  early_stopping_monitor = "auto",
  early_stopping_tolerance = 0,
  early_stopping_patience = 0L,
  num_workers = 0L,
  skip_importance = FALSE,
  ifw = FALSE
)
}
\arguments{
\item{batch_size}{(Tunable) Integer [1, Inf): Batch size.}

\item{penalty}{(Tunable) Numeric [0, Inf): Regularization penalty.}

\item{clip_value}{(Tunable) Optional Numeric: Clip value.}

\item{loss}{(Tunable) Character: Loss function.}

\item{epochs}{(Tunable) Integer [1, Inf): Number of epochs.}

\item{drop_last}{(Tunable) Logical: If TRUE, drop last batch.}

\item{decision_width}{(Tunable) Optional Integer [1, Inf): Decision width.}

\item{attention_width}{(Tunable) Optional Integer [1, Inf): Attention width.}

\item{num_steps}{(Tunable) Integer [1, Inf): Number of steps.}

\item{feature_reusage}{(Tunable) Numeric [0, Inf): Feature reusage.}

\item{mask_type}{(Tunable) Character \{"sparsemax", "entmax"\}: Mask type.}

\item{virtual_batch_size}{(Tunable) Integer [1, Inf): Virtual batch size.}

\item{valid_split}{(Tunable) Numeric [0, 1): Validation split.}

\item{learn_rate}{(Tunable) Numeric (0, Inf): Learning rate.}

\item{optimizer}{Character or torch function: Optimizer.}

\item{lr_scheduler}{Optional Character or torch function: "step", "reduce_on_plateau".}

\item{lr_decay}{(Tunable) Numeric [0, 1]: Learning rate decay.}

\item{step_size}{(Tunable) Integer [1, Inf): Step size.}

\item{checkpoint_epochs}{(Tunable) Integer [1, Inf): Checkpoint epochs.}

\item{cat_emb_dim}{(Tunable) Integer [1, Inf): Categorical embedding dimension.}

\item{num_independent}{(Tunable) Integer [1, Inf): Number of independent Gated Linear Units (GLU)
at each step of the encoder.}

\item{num_shared}{(Tunable) Integer [1, Inf): Number of shared Gated Linear Units (GLU) at each
step of the encoder.}

\item{num_independent_decoder}{(Tunable) Integer [1, Inf): Number of independent GLU layers for
pretraining.}

\item{num_shared_decoder}{(Tunable) Integer [1, Inf): Number of shared GLU layers for
pretraining.}

\item{momentum}{(Tunable) Numeric [0, Inf): Momentum.}

\item{pretraining_ratio}{(Tunable) Numeric [0, 1]: Pretraining ratio.}

\item{device}{Character \{"auto", "cpu", "cuda"\}: Compute device.}

\item{importance_sample_size}{(Tunable) Optional Integer [1, Inf): Importance sample size.}

\item{early_stopping_monitor}{(Tunable) Character \{"auto", "valid_loss", "train_loss"\}: Early stopping monitor.}

\item{early_stopping_tolerance}{(Tunable) Numeric [0, Inf): Minimum relative improvement to reset the patience
counter.}

\item{early_stopping_patience}{(Tunable) Integer [0, Inf): Number of epochs without improving before
stopping.}

\item{num_workers}{Integer [0, Inf): Number of subprocesses for data loading.}

\item{skip_importance}{Logical: If TRUE, skip importance calculation.}

\item{ifw}{(Tunable) Logical: If TRUE, use Inverse Frequency Weighting in classification.}
}
\value{
TabNetHyperparameters object.
}
\description{
Setup hyperparameters for TabNet training.
}
\examples{
tabnet_hyperparams <- setup_TabNet(epochs = 100L, learn_rate = 0.01)
tabnet_hyperparams
}
\author{
EDG
}
