% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_S7_DecompositionParameters.R
\name{setup_tSNE}
\alias{setup_tSNE}
\title{Setup tSNE parameters.}
\usage{
setup_tSNE(
  k = 2L,
  initial_dims = 50L,
  perplexity = 30,
  theta = 0.5,
  check_duplicates = TRUE,
  pca = TRUE,
  partial_pca = FALSE,
  max_iter = 1000L,
  verbose = getOption("verbose", FALSE),
  is_distance = FALSE,
  Y_init = NULL,
  pca_center = TRUE,
  pca_scale = FALSE,
  normalize = TRUE,
  stop_lying_iter = ifelse(is.null(Y_init), 250L, 0L),
  mom_switch_iter = ifelse(is.null(Y_init), 250L, 0L),
  momentum = 0.5,
  final_momentum = 0.8,
  eta = 200,
  exaggeration_factor = 12,
  num_threads = 1L
)
}
\arguments{
\item{k}{Integer: Number of components.}

\item{initial_dims}{Integer: Initial dimensions.}

\item{perplexity}{Integer: Perplexity.}

\item{theta}{Float: Theta.}

\item{check_duplicates}{Logical: If TRUE, check for duplicates.}

\item{pca}{Logical: If TRUE, perform PCA.}

\item{partial_pca}{Logical: If TRUE, perform partial PCA.}

\item{max_iter}{Integer: Maximum number of iterations.}

\item{verbose}{Logical: If TRUE, print messages.}

\item{is_distance}{Logical: If TRUE, \code{x} is a distance matrix.}

\item{Y_init}{Matrix: Initial Y matrix.}

\item{pca_center}{Logical: If TRUE, center PCA.}

\item{pca_scale}{Logical: If TRUE, scale PCA.}

\item{normalize}{Logical: If TRUE, normalize.}

\item{stop_lying_iter}{Integer: Stop lying iterations.}

\item{mom_switch_iter}{Integer: Momentum switch iterations.}

\item{momentum}{Float: Momentum.}

\item{final_momentum}{Float: Final momentum.}

\item{eta}{Float: Eta.}

\item{exaggeration_factor}{Float: Exaggeration factor.}

\item{num_threads}{Integer: Number of threads.}
}
\value{
tSNEParameters object.
}
\description{
Setup tSNE parameters.
}
\details{
Get more information on the parameters by running \code{?Rtsne::Rtsne}.
}
\author{
EDG
}
