% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_DecompositionConfig.R
\name{setup_tSNE}
\alias{setup_tSNE}
\title{Setup tSNE config.}
\usage{
setup_tSNE(
  k = 2L,
  initial_dims = 50L,
  perplexity = 30,
  theta = 0.5,
  check_duplicates = TRUE,
  pca = TRUE,
  partial_pca = FALSE,
  max_iter = 1000L,
  verbose = getOption("verbose", FALSE),
  is_distance = FALSE,
  Y_init = NULL,
  pca_center = TRUE,
  pca_scale = FALSE,
  normalize = TRUE,
  stop_lying_iter = if (is.null(Y_init)) 250L else 0L,
  mom_switch_iter = if (is.null(Y_init)) 250L else 0L,
  momentum = 0.5,
  final_momentum = 0.8,
  eta = 200,
  exaggeration_factor = 12,
  num_threads = 1L
)
}
\arguments{
\item{k}{Integer [1, Inf): Number of components.}

\item{initial_dims}{Integer [1, Inf): Initial dimensions.}

\item{perplexity}{Numeric [0, Inf): Perplexity.}

\item{theta}{Numeric [0, 1]: Speed/accuracy trade-off.}

\item{check_duplicates}{Logical: If TRUE, check for duplicates.}

\item{pca}{Logical: If TRUE, perform PCA.}

\item{partial_pca}{Logical: If TRUE, perform partial PCA.}

\item{max_iter}{Integer [1, Inf): Maximum number of iterations.}

\item{verbose}{Logical: If TRUE, print messages.}

\item{is_distance}{Logical: If TRUE, \code{x} is a distance matrix.}

\item{Y_init}{Optional Matrix: Initial Y matrix.}

\item{pca_center}{Logical: If TRUE, center PCA.}

\item{pca_scale}{Logical: If TRUE, scale PCA.}

\item{normalize}{Logical: If TRUE, normalize.}

\item{stop_lying_iter}{Integer [0, Inf): Stop lying iterations.}

\item{mom_switch_iter}{Integer [0, Inf): Momentum switch iterations.}

\item{momentum}{Numeric: Momentum.}

\item{final_momentum}{Numeric: Final momentum.}

\item{eta}{Numeric: Eta.}

\item{exaggeration_factor}{Numeric: Exaggeration factor.}

\item{num_threads}{Integer [0, Inf): Number of threads.}
}
\value{
tSNEConfig object.
}
\description{
Setup tSNE config.
}
\details{
Get more information on the config by running \code{?Rtsne::Rtsne}.
}
\examples{
tSNE_config <- setup_tSNE(k = 3L)
tSNE_config
}
\author{
EDG
}
