% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_DecompositionConfig.R
\name{setup_UMAP}
\alias{setup_UMAP}
\title{Setup UMAP config.}
\usage{
setup_UMAP(
  k = 2L,
  n_neighbors = 15L,
  init = "spectral",
  metric = "euclidean",
  n_epochs = NULL,
  learning_rate = 1,
  scale = TRUE,
  features = NULL
)
}
\arguments{
\item{k}{Integer [1, Inf): Number of components.}

\item{n_neighbors}{Integer [1, Inf): Number of neighbors.}

\item{init}{Character: Initialization type. See \verb{uwot::umap "init"}.}

\item{metric}{Character \{"euclidean", "cosine", "manhattan", "hamming", "categorical"\}: Distance metric.}

\item{n_epochs}{Optional Integer [1, Inf): Number of epochs.}

\item{learning_rate}{Numeric [0, Inf): Learning rate.}

\item{scale}{Logical: If TRUE, scale input data before doing UMAP.}

\item{features}{Optional Character vector: Names of at least 2 distinct
feature columns to decompose. \code{NULL} decomposes all numeric features.}
}
\value{
UMAPConfig object.
}
\description{
Setup UMAP config.
}
\details{
A high \code{n_neighbors} value may give error in some systems:
"Error in irlba::irlba(L, nv = n, nu = 0, maxit = iters) :
function 'as_cholmod_sparse' not provided by package 'Matrix'"
}
\examples{
umap_config <- setup_UMAP(k = 3L)
umap_config
}
\author{
EDG
}
