% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_S7_DecompositionConfig.R
\name{setup_UMAP}
\alias{setup_UMAP}
\title{Setup UMAP config.}
\usage{
setup_UMAP(
  k = 2L,
  n_neighbors = 15L,
  init = "spectral",
  metric = c("euclidean", "cosine", "manhattan", "hamming", "categorical"),
  n_epochs = NULL,
  learning_rate = 1,
  scale = TRUE
)
}
\arguments{
\item{k}{Integer: Number of components.}

\item{n_neighbors}{Integer: Number of keighbors.}

\item{init}{Character: Initialization type. See \verb{uwot::umap "init"}.}

\item{metric}{Character: Distance metric to use: "euclidean", "cosine",
"manhattan", "hamming", "categorical".}

\item{n_epochs}{Integer: Number of epochs.}

\item{learning_rate}{Float: Learning rate.}

\item{scale}{Logical: If TRUE, scale input data before doing UMAP.}
}
\value{
UMAPConfig object.
}
\description{
Setup UMAP config.
}
\details{
A high \code{n_neighbors} value may give error in some systems:
"Error in irlba::irlba(L, nv = n, nu = 0, maxit = iters) :
function 'as_cholmod_sparse' not provided by package 'Matrix'"
}
\examples{
umap_config <- setup_UMAP(k = 3L)
umap_config
}
\author{
EDG
}
