% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsernorm.R
\name{sparsernorm}
\alias{sparsernorm}
\title{Sparse rnorm}
\usage{
sparsernorm(n, sparseness = 0.1, mean = 0, sd = 1)
}
\arguments{
\item{n}{Integer: Length of output vector}

\item{sparseness}{Float (0, 1): Fraction of required nonzero elements, i.e. output will have
\verb{round(sparseness * n) nonzero elements}.
If \code{sparseness = 0}, a vector of zeros length \code{n} is returned,
if \code{sparseness = 1}, \code{rnorm(n, mean, sd)} is returned.
Default = 0.1}

\item{mean}{Float: Target mean of nonzero elements, passed to \code{stats::rnorm} Default = 0}

\item{sd}{Float: Target sd of nonzero elements, passed to \code{stats::rnorm} Default = 1}
}
\description{
A sparse version of \code{stats::rnorm}
Outputs a vector where a fraction of values are zeros (determined by \code{sparseness})
and the rest are drawn from a random normal distribution using \code{stats::rnorm}
}
\author{
E.D. Gennatas
}
