% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modError.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(true, estimated, harmonize = FALSE, verbose = TRUE)
}
\arguments{
\item{true}{True labels}

\item{estimated}{Estimated labels}

\item{harmonize}{Logical: If TRUE, run \link{factorHarmonize} first}

\item{verbose}{Logical: If TRUE, print messages to output. Default = TRUE}
}
\description{
The first factor level is considered the positive case.
}
