% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shyoptleaves.R
\name{splitlin_}
\alias{splitlin_}
\title{\code{rtemis} internal: Ridge and Stump}
\usage{
splitlin_(
  g,
  type,
  node.index,
  alpha = 1,
  lambda = 0.05,
  lambda.seq = NULL,
  cv.glmnet.nfolds = 5,
  part.minsplit = 2,
  part.xval = 0,
  part.max.depth = 1,
  part.cp = 0,
  part.minbucket = 5,
  minobsinnode.lin = 5,
  lin.type,
  nvmax = nvmax,
  verbose = TRUE,
  trace = 0
)
}
\arguments{
\item{node.index}{Open nodes to work on

Fit a linear model on (x, y) and split on the gradient
Input: environment holding tree and index of node
Output: None; Expands tree within environment g by splitting indexed node}

\item{tree}{Node within tree environment}
}
\description{
Edits environment 'g' in-place (no output)
}
