% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shytreegamleaves.R
\name{splitlineRC}
\alias{splitlineRC}
\title{\code{rtemis} internal: Ridge and Stump}
\usage{
splitlineRC(
  g,
  type,
  node.index,
  gamma,
  gamma.on.lin,
  lin.type,
  alpha,
  lambda,
  lambda.seq,
  cv.glmnet.nfolds,
  which.cv.glmnet.lambda,
  part.minsplit,
  part.xval,
  part.max.depth,
  part.cp,
  part.minbucket,
  minobsinnode.lin,
  nvmax,
  verbose = TRUE,
  trace = 0
)
}
\arguments{
\item{node.index}{Open nodes to work on}

\item{tree}{Node within tree environment}
}
\description{
Edits environment 'g' in-place (no output)
}
\details{
Fit a linear model on (x, y) and split on the gradient
Input: environment holding tree and index of node
Output: None; Expands tree within environment g by splitting indexed node
}
\keyword{internal}
