% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{strat.boot}
\alias{strat.boot}
\title{Stratified Bootstrap Resampling}
\usage{
strat.boot(
  x,
  n.resamples = 10,
  train.p = 0.75,
  stratify.var = NULL,
  strat.n.bins = 4,
  target.length = NULL,
  seed = NULL,
  verbosity = TRUE
)
}
\arguments{
\item{x}{Input vector}

\item{n.resamples}{Integer: Number of training/testing sets required}

\item{train.p}{Float (0, 1): Fraction of cases to assign to traininig set for
\code{resampler = "strat.sub"}}

\item{stratify.var}{Numeric vector (optional): Variable used for stratification.}

\item{strat.n.bins}{Integer: Number of groups to use for stratification for
\code{resampler = "strat.sub" / "strat.boot"}}

\item{target.length}{Integer: Number of cases for training set for
\code{resampler = "strat.boot"}.}

\item{seed}{Integer: (Optional) Set seed for random number generator, in order to make
output reproducible. See \code{?base::set.seed}}

\item{verbosity}{Logical: If TRUE, print messages to console}
}
\description{
Stratified Bootstrap Resampling
}
\author{
E.D. Gennatas
}
