% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict.R, R/type_check.R
\name{strict}
\alias{strict}
\title{Strict assignment by class or type}
\usage{
strict(object, class, allow_null = TRUE)

strict(object, class, allow_null = TRUE)
}
\arguments{
\item{object}{Object to check and return}

\item{class}{Character vector: class(es) to check against}

\item{allow_null}{Logical: if TRUE, allows NULL objects}

\item{x}{Value to be assigned if class and/or type check passes}

\item{accept.class}{Required class of \code{x}}

\item{accept.type}{Required type of \code{x}}

\item{non.null.val}{Function which should have a non-null value when applied to \code{x}}
}
\value{
Object
}
\description{
Allow assignment only if input is of correct class and/or type
}
\examples{
\dontrun{
strict("papaya", "character") # "papaya"
strict(c(1, 2.5, 3.2), "integer") # Error
strict(iris, "list") # Error
}
}
\author{
E.D. Gennatas

EDG
}
\keyword{internal}
