% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strng.R
\name{bold}
\alias{bold}
\alias{italic}
\alias{underline}
\alias{hilite}
\alias{hilitebig}
\alias{red}
\alias{green}
\alias{orange}
\alias{cyan}
\alias{magenta}
\alias{gray}
\alias{reset}
\title{String formatting utilities}
\usage{
bold(...)

italic(...)

underline(...)

hilite(..., col = "69;1", bold = TRUE)

hilitebig(x)

red(..., bold = FALSE)

green(..., bold = FALSE)

orange(..., bold = FALSE)

cyan(..., bold = FALSE)

magenta(..., bold = FALSE)

gray(..., bold = FALSE, sep = " ")

reset(...)
}
\arguments{
\item{...}{Character objects to format}

\item{bold}{Logical: If TRUE, use bold font}

\item{x}{Numeric: Input}

\item{sep}{Character: Separator}
}
\description{
String formatting utilities
}
