% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_multimodal.R
\name{synth_multimodal}
\alias{synth_multimodal}
\title{Create "Multimodal" Synthetic Data}
\usage{
synth_multimodal(
  n.cases = 10000,
  init.fn = "runifmat",
  init.fn.params = list(min = -10, max = 10),
  n.groups = 4,
  n.feat.per.group = round(seq(10, 300, length.out = n.groups)),
  contrib.p = 0.33,
  linear.p = 0.66,
  square.p = 0.1,
  atan.p = 0.1,
  pair.multiply.p = 0.05,
  pair.square.p = 0.05,
  pair.atan.p = 0.05,
  verbose = TRUE,
  seed = NULL,
  filename = NULL
)
}
\arguments{
\item{n.cases}{Integer: Number of cases to create. Default = 10000}

\item{init.fn}{Character: "runifmat" or "rnormmat". Use the respective functions to
generate features as random uniform and random normal variables, respectively.
Default = "runifmat"}

\item{init.fn.params}{Named list with arguments "min", "max" for "runifmat" and
"mean", "sd" for "rnormmat". Default = \code{list(min = -10, max = 10)}}

\item{n.groups}{Integer: Number of feature groups / modalities to create. Default = 4}

\item{n.feat.per.group}{Integer, vector, length \code{n.groups}: Number of features per group to create.
Default = \code{c(50, 100, 200, 300)}}

\item{contrib.p}{Float (0, 1]: Ratio of features contributing to outcome per group. Default = .33, i.e.
a third of the features in each group will be used to produce the outcome y}

\item{linear.p}{Float [0, 1]: Ratio of contributing features to be included linearly. Default = .1, i.e. .1 of .33 of
features in each group will be included}

\item{square.p}{Float [0, 1]: Ratio of contributing features to be squared. Default = .1, i.e. .1 of .33 of features
in each group will be squared}

\item{atan.p}{Float [0, 1]: Ratio of contributing features whose \code{atan} will be used. These will be selected
from the features that were NOT sampled for squaring. Default = .1,
i.e. .1 of .33 of features in each group will be transformed using \code{atan}, but given these features were not
already picked to be squared (see \code{square.p})}

\item{pair.multiply.p}{Float [0, 1] Ratio of features will be divided into pairs and multiplied. Default = .05}

\item{pair.square.p}{Float [0, 1] Ratio of features which will be divided into pairs, multiplied and squared.}

\item{pair.atan.p}{Float [0, 1] Ratio of features which will be divided into pairs, multiplied and transformed using
\code{atan}.}

\item{verbose}{Logical: If TRUE, print messages to console.}

\item{seed}{Integer: If set, pass to \code{set.seed} for reproducibility}

\item{filename}{Character: Path to file to save output.}
}
\value{
List with elements x, y, index.square, index.atan, index.pair.square
}
\description{
Create "Multimodal" Synthetic Data using squares and arctangents
}
\details{
There are no checks yet for compatibility among inputs and certain combinations may not work.
}
\examples{
xmm <- synth_multimodal(
  n.cases = 10000,
  init.fn = "runifmat",
  init.fn.params = list(min = -10, max = 10),
  n.groups = 5,
  n.feat.per.group = c(20, 50, 100, 200, 300),
  contrib.p = .33,
  linear.p = .66,
  square.p = .1,
  atan.p = .1,
  pair.multiply.p = .1,
  pair.square.p = .1,
  pair.atan.p = .1,
  seed = 2019
)
}
\author{
E.D. Gennatas
}
