% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_reg_data.R
\name{synth_reg_data}
\alias{synth_reg_data}
\title{Synthesize Simple Regression Data}
\usage{
synth_reg_data(
  nrow = 500,
  ncol = 50,
  noise.sd.factor = 1,
  resample.params = setup.resample(),
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{nrow}{Integer: Number of rows. Default = 500}

\item{ncol}{Integer: Number of columns. Default = 50}

\item{noise.sd.factor}{Numeric: Add rnorm(nrow, sd = noise.sd.factor * sd(y)). Default = 2}

\item{resample.params}{Output of \link{setup.resample} defining training/testing split. The first resulting resample
will be used to create \code{dat.train} and \code{dat.test} output; all resample output under \code{resamples}}

\item{seed}{Integer: Seed for random number generator. Default = NULL}

\item{verbose}{Logical: If TRUE, print messages to console. Default = FALSE}
}
\value{
List with elements \verb{dat, dat.train, dat.test, resamples, w, seed}
}
\description{
Synthesize Simple Regression Data
}
\author{
E.D. Gennatas
}
