% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthRegData.R
\name{synthRegData}
\alias{synthRegData}
\title{Synthesize Simple Regression Data}
\usage{
synthRegData(nrow = 500, ncol = 50, seed = NULL)
}
\arguments{
\item{nrow}{Integer: Number of rows. Default = 500}

\item{ncol}{Integer: Number of columns. Default = 50}

\item{seed}{Integer: Seed for random number generator. Default = NULL}
}
\description{
Synthesize Simple Regression Data
}
\author{
Efstathios D. Gennatas
}
