% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table1}
\alias{table1}
\title{Table 1}
\usage{
table1(
  x,
  summaryFn1 = mean,
  summaryFn2 = sd,
  summaryFn1_extraArgs = list(na.rm = TRUE),
  summaryFn2_extraArgs = list(na.rm = TRUE),
  labelify = TRUE,
  verbosity = 1L,
  filename = NULL
)
}
\arguments{
\item{x}{data.frame or matrix: Input data, cases by features}

\item{summaryFn1}{Function: Summary function 1.}

\item{summaryFn2}{Function: Summary function 2.}

\item{summaryFn1_extraArgs}{List: Extra arguments for \code{summaryFn1}.}

\item{summaryFn2_extraArgs}{List: Extra arguments for \code{summaryFn2}.}

\item{labelify}{Logical: If TRUE, apply \link{labelify} to column names of  \code{x}}

\item{verbosity}{Integer: Verbosity level.}

\item{filename}{Character: Path to output CSV file to save table.}
}
\value{
A data.frame, invisibly, with two columns: "Feature", "Value mean (sd) | N"
}
\description{
Build Table 1. Subject characteristics
}
\details{
The output will look like "summaryFn1 (summaryFn2)".
Using defaults this will be "mean (sd)"
}
\examples{
\dontrun{
table1(iris)
}
}
\author{
EDG
}
