% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{tohtml}
\alias{tohtml}
\title{Generate \code{CheckData} object description in HTML}
\usage{
tohtml(
  x,
  name = NULL,
  css = list(font.family = "Helvetica", color = "#fff", background.color = "#242424")
)
}
\arguments{
\item{x}{\code{CheckData} object}

\item{name}{Character: Name of the data set}

\item{css}{List: CSS styles}
}
\description{
Generate \code{CheckData} object description in HTML
}
\author{
E.D. Gennatas
}
