% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_learn.R
\name{tunable}
\alias{tunable}
\title{Print tunable hyperparameters for a supervised learning algorithm}
\usage{
tunable(
  alg = c("glmnet", "svm", "cart", "ranger", "gbm", "xgboost", "lightgbm")
)
}
\arguments{
\item{alg}{Character string: Algorithm name.}
}
\value{
Prints tunable hyperparameters for the specified algorithm.
}
\description{
Print tunable hyperparameters for a supervised learning algorithm
}
\author{
EDG
}
