% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7init.R
\name{validate_hyperparameters}
\alias{validate_hyperparameters}
\title{Check hyperparameters given training data}
\usage{
validate_hyperparameters(x, hyperparameters)
}
\arguments{
\item{x}{tabular data: Training data.}

\item{hyperparameters}{\code{Hyperparameters} to check.}
}
\description{
Check hyperparameters given training data
}
