% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R, R/14_S7_SuperConfig.R
\name{write_toml}
\alias{write_toml}
\alias{write_toml.SuperConfig}
\title{Write to TOML file}
\usage{
write_toml(x, file, overwrite = FALSE, verbosity = 1L)

write_toml.SuperConfig(x, file, overwrite = FALSE, verbosity = 1L)
}
\arguments{
\item{x}{\code{SuperConfig} object.}

\item{file}{Character: Path to output TOML file.}

\item{overwrite}{Logical: If TRUE, overwrite existing file.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
\code{SuperConfig} object, invisibly.
}
\description{
Write to TOML file
}
\examples{
x <- setup_SuperConfig(
  dat_training_path = "~/Data/iris.csv",
  dat_validation_path = NULL,
  dat_test_path = NULL,
  weights = NULL,
  preprocessor_config = setup_Preprocessor(remove_duplicates = TRUE),
  algorithm = "LightRF",
  hyperparameters = setup_LightRF(),
  tuner_config = setup_GridSearch(),
  outer_resampling_config = setup_Resampler(),
  execution_config = setup_ExecutionConfig(),
  question = "Can we tell iris species apart given their measurements?",
  outdir = "models/",
  verbosity = 1L
)
tmpdir <- tempdir()
write_toml(x, file.path(tmpdir, "rtemis.toml"))
}
\author{
EDG
}
