% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S7_init.R
\name{write_toml}
\alias{write_toml}
\title{Write to TOML file}
\usage{
write_toml(x, file, overwrite = FALSE, verbosity = 1L)
}
\arguments{
\item{x}{\code{SuperConfig} object to write to TOML file.}

\item{file}{Character: Path to output TOML file.}
}
\value{
\code{SuperConfig} object, invisibly.
}
\description{
Write to TOML file
}
\author{
EDG
}
